Add distance information to resampling objects

Add distance information to resampling objects

add.distance(object, ...)

# S3 method for resampling
add.distance(object, data, coords = c("x", "y"), ...)

# S3 method for represampling
add.distance(object, ...)

Arguments

object
resampling or represampling object.

...
Additional arguments to dataset_distance and add.distance.resampling, respectively.

data
data.frame containing at least the columns specified by coords

coords
(ignored by partition_cv)

Value

A resampling or represampling object containing an additional. $distance component in each resampling object. The distance component is a single numeric value indicating, for each train / test pair, the (by default, mean) nearest-neighbour distance between the two sets.

Details

Nearest-neighbour distances are calculated for each sample in the test set. These nrow(???$test) nearest-neighbour distances are then averaged. Aggregation methods other than mean can be chosen using the fun argument, which will be passed on to dataset_distance.

See also

dataset_distance represampling resampling

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador
nsp.parti <- partition_cv(ecuador)
sp.parti <- partition_kmeans(ecuador)
nsp.parti <- add.distance(nsp.parti, ecuador)
sp.parti <- add.distance(sp.parti, ecuador)
# non-spatial partioning: very small test-training distance:
nsp.parti[[1]][[1]]$distance

#> [1] 53.79223

# spatial partitioning: more substantial distance, depending on number of
# folds etc.
sp.parti[[1]][[1]]$distance

#> [1] 390.1742