Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or boostrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the seleciton (see Examples).

Another example is bootstrap resampling. represampling.bootstrap with argument oob = TRUE generates [rep]resampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling.list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

represampling, partition.cv, partition.kmeans, represampling.bootstrap, etc.

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor( as.character( floor( ecuador$dem / 200 ) ) )
smp <- as.resampling(parti)
summary(smp)

#>    n.train n.test
#> 10    1657    343
#> 11    1555    445
#> 12    1728    272
#> 13    1801    199
#> 14    1938     62
#> 15    1985     15
#> 8     1885    115
#> 9     1451    549

# Compare:
summary(parti)

#>  10  11  12  13  14  15   8   9 
#> 343 445 272 199  62  15 115 549 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition.cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str( ecuador[ parti[[1]]$test , ] )

#> 'data.frame':	200 obs. of  13 variables:
#>  $ x             : num  714922 715072 715552 713542 712392 ...
#>  $ y             : num  9558862 9558242 9558812 9559972 9560162 ...
#>  $ dem           : num  2536 2745 2518 2184 2002 ...
#>  $ slope         : num  54.91 49.05 7.98 35.24 47.28 ...
#>  $ hcurv         : num  0.01305 0.00097 -0.01863 -0.00841 -0.00284 ...
#>  $ vcurv         : num  -0.00555 0.01483 -0.00507 -0.0001 0.01804 ...
#>  $ carea         : num  430 567 159065 1527 629 ...
#>  $ cslope        : num  46.3 32.2 27.1 20.1 28.5 ...
#>  $ distroad      : num  300 300 300 300 24.2 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 300 247 0 ...
#>  $ distslidespast: num  100 100 100 100 89 39 100 0 24 4 ...
#>  $ log.carea     : num  2.63 2.75 5.2 3.18 2.8 ...

# the corresponding training sample - larger:
str( ecuador[ parti[[1]]$train , ] )

#> 'data.frame':	1800 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715042 715382 ...
#>  $ y             : num  9560002 9559582 9560172 9559312 9560142 ...
#>  $ dem           : num  1912 2199 1989 2320 2021 ...
#>  $ slope         : num  25.6 23.2 40.5 42.9 42 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 -0.01106 0.00958 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 -0.04634 0.02642 ...
#>  $ carea         : num  5577 1399 351155 501 671 ...
#>  $ cslope        : num  34.4 30.7 32.8 33.9 41.6 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 9.15 300 300 300 0 ...
#>  $ distslidespast: num  9 21 40 100 21 2 100 100 41 5 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.7 2.83 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling.bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str( ecuador[ parti[[1]]$test , ] )

#> 'data.frame':	716 obs. of  13 variables:
#>  $ x             : num  712882 715232 712802 715272 713512 ...
#>  $ y             : num  9560002 9559582 9559952 9557702 9559092 ...
#>  $ dem           : num  1912 2199 1838 2813 2166 ...
#>  $ slope         : num  25.6 23.2 52.1 31 56 ...
#>  $ hcurv         : num  -0.00681 -0.00501 0.00183 -0.00123 0.02056 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.09203 0.00393 -0.06976 ...
#>  $ carea         : num  5577 1399 634 2081 301 ...
#>  $ cslope        : num  34.4 30.7 30.3 37.6 49.4 ...
#>  $ distroad      : num  300 300 300 300 300 30 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 9.15 300 300 ...
#>  $ distslidespast: num  9 21 2 100 41 20 100 100 26 31 ...
#>  $ log.carea     : num  3.75 3.15 2.8 3.32 2.48 ...

# bootstrap training sample: same size as nrow(ecuador):
str( ecuador[ parti[[1]]$train , ] )

#> 'data.frame':	2000 obs. of  13 variables:
#>  $ x             : num  713242 715012 712892 715022 713302 ...
#>  $ y             : num  9561232 9558742 9558862 9557712 9559172 ...
#>  $ dem           : num  1967 2582 2118 2736 2107 ...
#>  $ slope         : num  38.2 30.7 30.9 30.3 36 ...
#>  $ hcurv         : num  0.00236 0.00293 0.01191 0.00133 0.01354 ...
#>  $ vcurv         : num  0.00483 -0.00163 0.00699 0.00367 0.00245 ...
#>  $ carea         : num  1097 782 1849 806 630 ...
#>  $ cslope        : num  39.9 28.5 27.5 26.5 30 ...
#>  $ distroad      : num  80.7 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 1 2 2 2 2 2 1 1 1 2 ...
#>  $ distdeforest  : num  0 300 300 300 300 ...
#>  $ distslidespast: num  100 86.1 41.9 100 0 ...
#>  $ log.carea     : num  3.04 2.89 3.27 2.91 2.8 ...