Default error function

Calculate a variety of accuracy measures from observations and predictions of numerical and categorical response variables.

err.default(obs, pred)

Arguments

obs

factor, logical, or numeric vector with observations

pred

factor, logical, or numeric vector with predictions. Must be of same type as obs with the exception that pred may be numeric if obs is factor or logical ('soft' classification).

Value

A list with (currently) the following components, depending on the type of prediction problem:

Note

NA values are currently not handled by this function, i.e. they will result in an error.

See also

ROCR

Examples

obs <- rnorm(1000)
# Two mock (soft) classification examples:
err.default( obs > 0, rnorm(1000) ) # just noise

#> $auroc
#> [1] 0.4935626
#> 
#> $error
#> [1] 0.509
#> 
#> $accuracy
#> [1] 0.491
#> 
#> $sensitivity
#> [1] 0.3205374
#> 
#> $specificity
#> [1] 0.6764092
#> 
#> $fpr70
#> [1] 0.7181628
#> 
#> $fpr80
#> [1] 0.8434238
#> 
#> $fpr90
#> [1] 0.9060543
#> 
#> $tpr80
#> [1] 0.2207294
#> 
#> $tpr90
#> [1] 0.1362764
#> 
#> $tpr95
#> [1] 0.06142035
#> 
#> $events
#> [1] 521
#> 
#> $count
#> [1] 1000
#> 

err.default( obs > 0, obs + rnorm(1000) ) # some discrimination

#> $auroc
#> [1] 0.8417529
#> 
#> $error
#> [1] 0.277
#> 
#> $accuracy
#> [1] 0.723
#> 
#> $sensitivity
#> [1] 0.5738964
#> 
#> $specificity
#> [1] 0.8851775
#> 
#> $fpr70
#> [1] 0.1983299
#> 
#> $fpr80
#> [1] 0.2964509
#> 
#> $fpr90
#> [1] 0.4634656
#> 
#> $tpr80
#> [1] 0.7024952
#> 
#> $tpr90
#> [1] 0.5412668
#> 
#> $tpr95
#> [1] 0.4644914
#> 
#> $events
#> [1] 521
#> 
#> $count
#> [1] 1000
#> 

# Three mock regression examples:
err.default( obs, rnorm(1000) ) # just noise, but no bias

#> $bias
#> [1] 0.01665322
#> 
#> $stddev
#> [1] 1.388829
#> 
#> $mse
#> [1] 1.927196
#> 
#> $mad
#> [1] 1.392276
#> 
#> $median
#> [1] -0.01502064
#> 
#> $iqr
#> [1] 1.879338
#> 
#> $count
#> [1] 1000
#> 

err.default( obs, obs + rnorm(1000) ) # some association, no bias

#> $bias
#> [1] 0.002175692
#> 
#> $stddev
#> [1] 0.9893763
#> 
#> $mse
#> [1] 0.9778914
#> 
#> $mad
#> [1] 1.074292
#> 
#> $median
#> [1] -0.005034918
#> 
#> $iqr
#> [1] 1.439094
#> 
#> $count
#> [1] 1000
#> 

err.default( obs, obs + 1 ) # perfect correlation, but with bias

#> $bias
#> [1] -1
#> 
#> $stddev
#> [1] 6.419493e-17
#> 
#> $mse
#> [1] 1
#> 
#> $mad
#> [1] 0
#> 
#> $median
#> [1] -1
#> 
#> $iqr
#> [1] 0
#> 
#> $count
#> [1] 1000
#>