Identify small partitions that need to be fixed.

get.small.tiles identifies partitions (tiles) that are too small according to some defined criterion / criteria (minimum number of samples in tile and/or minimum fraction of entire dataset).

get.small.tiles(tile, min.n = NULL, min.frac = 0, ignore = c())

Arguments

tile

factor: tile/partition names for all samples; names must be coercible to class tilename(), i.e. of the form 'X4:Y2' etc.

min.n

integer (optional): minimum number of samples per partition.

min.frac

numeric >0, <1: minimum relative size of partition as percentage of sample.

ignore

character vector: names of tiles to be ignored, i.e. to be retained even if the inclusion criteria are not met.

Value

character vector: names of tiles that are considered 'small' according to these criteria

See also

partition.tiles(), tilename()

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador
# Rectangular partitioning without removal of small tiles:
parti = partition.tiles(ecuador, nsplit = c(10,10), reassign = FALSE)
summary(parti)

#> $`1`
#>        n.train n.test
#> X1:Y10    1983     12
#> X1:Y5     1978     17
#> X1:Y6     1983     12
#> X1:Y7     1964     31
#> X1:Y8     1938     57
#> X1:Y9     1987      8
#> X10:Y1    1984     11
#> X10:Y2    1972     23
#> X10:Y3    1972     23
#> X10:Y4    1980     15
#> X10:Y5    1981     14
#> X2:Y10    1986      9
#> X2:Y4     1977     18
#> X2:Y5     1971     24
#> X2:Y6     1981     14
#> X2:Y7     1942     53
#> X2:Y8     1956     39
#> X2:Y9     1954     41
#> X3:Y10    1958     37
#> X3:Y3     1983     12
#> X3:Y4     1972     23
#> X3:Y5     1959     36
#> X3:Y6     1971     24
#> X3:Y7     1982     13
#> X3:Y8     1985     10
#> X3:Y9     1937     58
#> X4:Y10    1958     37
#> X4:Y3     1978     17
#> X4:Y4     1963     32
#> X4:Y5     1978     17
#> X4:Y6     1978     17
#> X4:Y7     1971     24
#> X4:Y8     1981     14
#> X4:Y9     1960     35
#> X5:Y1     1987      8
#> X5:Y10    1947     48
#> X5:Y2     1974     21
#> X5:Y3     1970     25
#> X5:Y4     1954     41
#> X5:Y5     1981     14
#> X5:Y6     1975     20
#> X5:Y7     1971     24
#> X5:Y8     1981     14
#> X5:Y9     1981     14
#> X6:Y1     1978     17
#> X6:Y10    1956     39
#> X6:Y2     1977     18
#> X6:Y3     1970     25
#> X6:Y4     1975     20
#> X6:Y5     1965     30
#> X6:Y6     1977     18
#> X6:Y7     1975     20
#> X6:Y8     1973     22
#> X6:Y9     1980     15
#> X7:Y1     1975     20
#> X7:Y10    1922     73
#> X7:Y2     1969     26
#> X7:Y3     1980     15
#> X7:Y4     1969     26
#> X7:Y5     1957     38
#> X7:Y6     1967     28
#> X7:Y7     1978     17
#> X7:Y8     1975     20
#> X7:Y9     1971     24
#> X8:Y1     1969     26
#> X8:Y10    1979     16
#> X8:Y2     1969     26
#> X8:Y3     1964     31
#> X8:Y4     1967     28
#> X8:Y5     1965     30
#> X8:Y6     1944     51
#> X8:Y7     1974     21
#> X8:Y8     1978     17
#> X8:Y9     1976     19
#> X9:Y1     1981     14
#> X9:Y2     1978     17
#> X9:Y3     1974     21
#> X9:Y4     1963     32
#> X9:Y5     1969     26
#> X9:Y6     1969     26
#> X9:Y7     1985     10
#> X9:Y8     1983     12
#> X9:Y9     1990      5
#> 

length(parti[[1]])

#> [1] 83

# Same in factor format for the application of get.small.tiles:
parti.fac <- partition.tiles(ecuador, nsplit = c(10, 10), reassign = FALSE,
                             return.factor = TRUE)
get.small.tiles(parti.fac[[1]], min.n = 20) # tiles with less than 20 samples

#>  [1] X9:Y9  X1:Y9  X5:Y1  X2:Y10 X3:Y8  X9:Y7  X10:Y1 X1:Y10 X1:Y6  X3:Y3 
#> [11] X9:Y8  X3:Y7  X10:Y5 X2:Y6  X4:Y8  X5:Y5  X5:Y8  X5:Y9  X9:Y1  X10:Y4
#> [21] X6:Y9  X7:Y3  X8:Y10 X1:Y5  X4:Y3  X4:Y5  X4:Y6  X6:Y1  X7:Y7  X8:Y8 
#> [31] X9:Y2  X2:Y4  X6:Y2  X6:Y6  X8:Y9 
#> 83 Levels: X1:Y10 X1:Y5 X1:Y6 X1:Y7 X1:Y8 X1:Y9 X10:Y1 X10:Y2 X10:Y3 ... X9:Y9

parti2 <- partition.tiles(ecuador, nsplit = c(10, 10), reassign = TRUE,
                          min.n = 20, min.frac = 0)
length(parti2[[1]]) # < length(parti[[1]])

#> [1] 60