Perform spatial error estimation and variable importance assessment

parsperrorest is a flexible interface for multiple types of parallelized spatial and non-spatial cross-validation and bootstrap error estimation and parallelized permutation-based assessment of spatial variable importance.

parsperrorest(formula, data, coords = c("x", "y"), model.fun,
  model.args = list(), pred.fun = NULL, pred.args = list(),
  smp.fun = partition.loo, smp.args = list(), train.fun = NULL,
  train.param = NULL, test.fun = NULL, test.param = NULL,
  err.fun = err.default, error.fold = TRUE, error.rep = TRUE,
  err.train = TRUE, imp.variables = NULL, imp.permutations = 1000,
  importance = !is.null(imp.variables), distance = FALSE, do.gc = 1,
  do.try = FALSE, progress = 1, out.progress = "", par.args = list(),
  benchmark = FALSE, ...)

Arguments

formula

A formula specifying the variables used by the model. Only simple formulas without interactions or nonlinear terms should be used, e.g. y~x1+x2+x3 but not y~x1*x2+log(x3). Formulas involving interaction and nonlinear terms may possibly work for error estimation but not for variable importance assessment, but should be used with caution.

data

a data.frame with predictor and response variables. Training and test samples will be drawn from this data set by train.fun and test.fun, respectively.

coords

vector of length 2 defining the variables in data that contain the x and y coordinates of sample locations.

model.fun

Function that fits a predictive model, such as glm or rpart. The function must accept at least two arguments, the first one being a formula and the second a data.frame with the learning sample.

model.args

Arguments to be passed to model.fun (in addition to the formula and data argument, which are provided by sperrorest)

pred.fun

Prediction function for a fitted model object created by model. Must accept at least two arguments: the fitted object and a data.frame newdata with data on which to predict the outcome.

pred.args

(optional) Arguments to pred.fun (in addition to the fitted model object and the newdata argument, which are provided by sperrorest)

smp.fun

A function for sampling training and test sets from data. E.g., partition.kmeans() for spatial cross-validation using spatial k-means clustering.

smp.args

(optional) Arguments to be passed to est.fun

train.fun

(optional) A function for resampling or subsampling the training sample in order to achieve, e.g., uniform sample sizes on all training sets, or maintaining a certain ratio of positives and negatives in training sets. E.g., resample.uniform() or resample.strat.uniform()

train.param

(optional) Arguments to be passed to resample.fun

test.fun

(optional) Like train.fun but for the test set.

test.param

(optional) Arguments to be passed to test.fun

err.fun

A function that calculates selected error measures from the known responses in data and the model predictions delivered by pred.fun. E.g., err.default() (the default). See example and details below.

error.fold

logical (default: TRUE) if importance is TRUE, otherwise FALSE): calculate error measures on each fold within a resampling repetition.

error.rep

logical (default: TRUE): calculate error measures based on the pooled predictions of all folds within a resampling repetition.

err.train

logical (default: TRUE): calculate error measures on the training set (in addition to the test set estimation).

imp.variables

(optional; used if importance = TRUE) Variables for which permutation-based variable importance assessment is performed. If importance = TRUE and imp.variables is NULL, all variables in formula will be used.

imp.permutations

(optional; used if importance = TRUE) Number of permutations used for variable importance assessment.

importance

logical: perform permutation-based variable importance assessment?

distance

logical (default: FALSE): if TRUE, calculate mean nearest-neighbour distances from test samples to training samples using add.distance.represampling()

do.gc

numeric (default: 1): defines frequency of memory garbage collection by calling gc(); if <1, no garbage collection; if >=1, run a gc() after each repetition; if >=2, after each fold

do.try

logical (default: FALSE): if TRUE untested!!, use try() to robustify calls to model.fun and err.fun; use with caution!

progress

numeric (default: 1): Whether to show progress information. For par.mode = 1, information about elapsed time, estimated time remaining and a percentage indicator (0% - 100%) are shown. progress = 2 only applies to par.mode = 2 and shows repetition information only (instead of repetition and fold). Set to FALSE for no progress information.

out.progress

only used if par.mode = 2: Optional write progress output to a file instead of console output. The default ('') results in console output for Unix-systems and file output ('parsperrorest.progress.txt') in the current working directory for Windows-systems.

par.args

list of parallelization parameters: par.mode (the parallelization mode), par.units (the number of parallel processing units), par.libs (libraries to be loaded on cluster workers, character list). See Details for more information.

benchmark

(optional) logical (default: FALSE): if TRUE, perform benchmarking and return sperrorestbenchmarks object

...

currently not used

Value

A list (object of class sperrorest) with (up to) six components:

Details

Two par.mode options are availabe. The default mode is par.mode = 1. Here, pbapply::pblapply() is used which either calls parallel::mclapply() (on Unix-systems) or parallel::parApply() (on Windows-systems). par.mode = 2 uses foreach::foreach(). While this approach is not as efficient, it may work in cases in which par.mode = 1 fails.

par.libs only applies to par.mode = 1 on Windows-systems.

This parallelized version of sperrorest() may highly decrease computation time. However, please note that problems may occur depending on which function is used for cross-validation. While the rpart::rpart() example (see Examples) here works fine, you may encounter problems with other functions.

For par.mode = 2, you may encounter missing repetitions in the results if repetitions finish to quickly. In this case, consider using sperrorest::sperrorest()

Known problems when being parallized: randomForest::randomForest()

Note

(1) Optionally save fitted models, training and test samples in the results object; (2) Optionally save intermediate results in some file, and enable the function to continue an interrupted sperrorest call where it was interrupted. (3) Optionally have sperrorest dump the result of each repetition into a file, and to skip repetitions for which a file already exists.

References

Brenning, A. 2012. Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: the R package 'sperrorest'. 2012 IEEE International Geoscience and Remote Sensing Symposium (IGARSS), 23-27 July 2012, p. 5372-5375.

Brenning, A. 2005. Spatial prediction models for landslide hazards: review, comparison and evaluation. Natural Hazards and Earth System Sciences, 5(6): 853-862.

Brenning, A., S. Long & P. Fieguth. Forthcoming. Detecting rock glacier flow structures using Gabor filters and IKONOS imagery. Submitted to Remote Sensing of Environment.

Russ, G. & A. Brenning. 2010a. Data mining in precision agriculture: Management of spatial information. In 13th International Conference on Information Processing and Management of Uncertainty, IPMU 2010; Dortmund; 28 June - 2 July 2010. Lecture Notes in Computer Science, 6178 LNAI: 350-359.

Russ, G. & A. Brenning. 2010b. Spatial variable importance assessment for yield prediction in Precision Agriculture. In Advances in Intelligent Data Analysis IX, Proceedings, 9th International Symposium, IDA 2010, Tucson, AZ, USA, 19-21 May 2010. Lecture Notes in Computer Science, 6065 LNCS: 184-195.

See also

sperrorest::sperrorest()

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope

# Example of a classification tree fitted to this data:
library(rpart)
mypred.rpart <- function(object, newdata) predict(object, newdata)[, 2]
ctrl <- rpart.control(cp = 0.005) # show the effects of overfitting
fit <- rpart(fo, data = ecuador, control = ctrl)

# Non-spatial 5-repeated 10-fold cross-validation:
mypred.rpart <- function(object, newdata) predict(object, newdata)[,2]
par.nsp.res <- parsperrorest(data = ecuador, formula = fo,
                             model.fun = rpart, model.args = list(control = ctrl),
                             pred.fun = mypred.rpart,
                             progress = TRUE,
                             smp.fun = partition.cv, 
                             smp.args = list(repetition = 1:5, nfold = 15), 
                             par.args = list(par.units = 2, par.mode = 1),
                             error.rep = TRUE, error.fold = TRUE)
summary(par.nsp.res$error.rep)
summary(par.nsp.res$error.fold)
summary(par.nsp.res$represampling)
# plot(par.nsp.res$represampling, ecuador)

# Spatial 5-repeated 10-fold spatial cross-validation:
par.sp.res <- parsperrorest(data = ecuador, formula = fo,
                            model.fun = rpart, model.args = list(control = ctrl),
                            pred.fun = mypred.rpart,
                            progress = TRUE,
                            smp.fun = partition.kmeans, 
                            smp.args = list(repetition = 1:5, nfold = 15), 
                            par.args = list(par.units = 2, par.mode = 2),
                            error.rep = TRUE, error.fold = TRUE)
summary(par.sp.res$error.rep)
summary(par.sp.res$error.fold)
summary(par.sp.res$represampling)
# plot(par.sp.res$represampling, ecuador)
![](example-plot.jpg "Example Plot Title")

smry <- data.frame(
    nonspat.training = unlist(summary(par.nsp.res$error.rep, level = 1)$train.auroc),
    nonspat.test     = unlist(summary(par.nsp.res$error.rep, level = 1)$test.auroc),
    spatial.training = unlist(summary(par.sp.res$error.rep, level = 1)$train.auroc),
    spatial.test     = unlist(summary(par.sp.res$error.rep, level = 1)$test.auroc))
boxplot(smry, col = c('red','red','red','green'), 
    main = 'Training vs. test, nonspatial vs. spatial',
    ylab = 'Area under the ROC curve')
    

#> Error: <text>:38:2: unexpected '['
#> 37: # plot(par.sp.res$represampling, ecuador)
#> 38: ![
#>      ^