Partition the data for a (non-spatial) cross-validation

partition.cv creates a represampling() object for length(repetition)-repeated nfold-fold cross-validation.

partition.cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return.factor = FALSE)

Arguments

data
data.frame containing at least the columns specified by coords

coords
(ignored by partition.cv)

nfold
number of partitions (folds) in nfold-fold cross-validation partitioning

repetition
numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1
seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.factor
if FALSE (default), return a represampling() object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return.factor = FALSE (the default), a represampling() object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling() objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return.factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

sperrorest(), represampling()

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition.cv(ecuador, nfold = 5, repetition = 1:2)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>               x       y     dem     slope     hcurv      vcurv        carea
#> 36710  715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 46178  714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 26531  714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 43327  713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 43669  712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 49372  714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 39629  714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226    2325.2053
#> 38287  714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 21703  713842.5 9559672 2282.58 24.991846   0.01581    0.01189     287.5425
#> 8019   714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 33792  714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 2990   713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 43964  714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 43007  713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 40307  713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 38279  714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 23375  713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 31576  713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 26215  714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 4827   714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 45575  715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 34280  713652.5 9560702 1896.37 41.273588   0.00382    0.00088    2322.1746
#> 28230  715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 15390  713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 13847  715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 46125  713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 3466   713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 1635   713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 35142  714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 28577  712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 6557   713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 47011  713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 24100  715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 27084  713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517    1000.7119
#> 46201  714832.5 9561022 1845.94 34.573992  -0.00856    0.01796    4729.6382
#> 31316  715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 17423  715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 26567  714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 34707  713962.5 9561092 1819.49 65.686173  -0.04749    0.04779    2972.0193
#> 16015  712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 25244  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 835    715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 4617   714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 18892  712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 38862  712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 7466   712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 30406  715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 2595   714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 13029  713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 6317   715042.5 9557702 2745.36 27.873824   0.00090   -0.00299     688.0706
#> 7668   713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 35083  714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 20065  713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 31121  713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 24501  713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 28046  714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 45500  712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 17431  712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 40270  713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 7385   713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 47366  714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 18897  712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 39722  715022.5 9557712 2735.53 30.260766   0.00133    0.00367     806.0966
#> 6845   713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 7041   713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 1973   715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 32879  713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 10467  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 44767  714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 47718  712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 7      714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 18348  714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 42201  714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 20797  713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 15432  715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 29228  712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#>             cslope distroad slides distdeforest distslidespast log.carea
#> 36710  30.71569444   300.00   TRUE       300.00          21.00  3.145890
#> 46178  20.93129417   300.00   TRUE       300.00          31.00  2.440958
#> 26531  15.98781432   300.00   TRUE       300.00           2.00  2.552843
#> 43327  29.05870050   173.40   TRUE        33.81         100.00  3.516154
#> 43669  28.47657538    24.22   TRUE         0.00          89.00  2.798316
#> 49372  23.54054397   213.54   TRUE        20.21          68.00  3.430977
#> 39629  47.54862150   300.00   TRUE       300.00          85.00  3.366461
#> 38287  14.89690267    41.43   TRUE         1.90         100.00  3.349380
#> 21703  22.26915062   300.00   TRUE       300.00          34.00  2.458702
#> 8019   35.78006839   300.00   TRUE       300.00          61.00  2.928936
#> 33792  51.20810294   300.00   TRUE       300.00           5.00  2.596460
#> 2990   25.32129680   300.00   TRUE       123.56           5.00  2.520462
#> 43964  28.75732470   300.00   TRUE       300.00          46.00  2.952143
#> 43007  38.96170303   300.00   TRUE       300.00          37.00  2.473902
#> 40307  29.51362898    77.82   TRUE        76.41          54.00  3.524078
#> 38279  37.62843024   300.00   TRUE       300.00          24.00  5.142607
#> 23375  34.63988238   300.00   TRUE       195.00           2.00  3.224136
#> 31576  26.93531891   300.00   TRUE       300.00           4.00  3.635091
#> 26215  23.83160653   300.00   TRUE       300.00          10.00  3.268382
#> 4827   21.07396066   300.00   TRUE       300.00           0.00  2.559711
#> 45575  25.72809683   300.00   TRUE       300.00          59.00  2.562268
#> 34280  33.62116342   300.00   TRUE         0.00         100.00  3.365895
#> 28230  37.90173111   300.00   TRUE       300.00         100.00  2.977867
#> 15390  29.32226108    60.00   TRUE       118.92           2.00  3.372821
#> 13847  20.57090372   300.00   TRUE       300.00         100.00  3.062637
#> 46125  25.65705007   300.00   TRUE       300.00         100.00  3.982935
#> 3466   26.79494425   101.32   TRUE        95.02          19.00  3.724022
#> 1635   27.82684124   300.00   TRUE       300.00         100.00  3.303016
#> 35142  27.15304287   300.00   TRUE       300.00         100.00  2.176525
#> 28577  34.39121869   300.00   TRUE         0.56           0.00  3.781222
#> 6557   27.41946824   300.00   TRUE       300.00           5.00  3.147064
#> 47011  44.53429054   300.00   TRUE       300.00          10.00  3.652797
#> 24100  34.29782657   300.00   TRUE       300.00         100.00  5.558684
#> 27084  21.67270156   300.00   TRUE       300.00          81.00  3.000309
#> 46201  22.18377991   167.75   TRUE         0.00          55.00  3.674828
#> 31316  30.00981044   300.00   TRUE       300.00          85.00  2.740390
#> 17423  25.46052554   300.00   TRUE       300.00         100.00  2.868983
#> 26567  40.43706935   300.00   TRUE       300.00         100.00  3.023906
#> 34707  27.70250939   165.13   TRUE        39.56         100.00  3.473052
#> 16015  38.83278752   152.09   TRUE        90.23          45.00  2.971800
#> 25244  33.21894705   300.00   TRUE       300.00           6.00  5.916915
#> 835    36.65039128   300.00   TRUE       300.00         100.00  3.141386
#> 4617   27.95174604   300.00   TRUE       300.00           6.00  2.211304
#> 18892  27.06308850   235.02   TRUE        27.24           0.00  2.675293
#> 38862  34.90000522   300.00   TRUE       104.48          92.00  3.151879
#> 7466   28.79055625   255.23   TRUE         0.00          60.00  3.163437
#> 30406  43.10075014   300.00   TRUE       300.00          85.00  2.770965
#> 2595   30.43952878   300.00   TRUE       300.00         100.00  3.538794
#> 13029  34.18781868   300.00   TRUE        96.65           4.00  2.813330
#> 6317   27.32321133   300.00   TRUE       300.00         100.00  2.837633
#> 7668   36.32036759   300.00   TRUE       300.00           5.00  2.789721
#> 35083  43.54078173   300.00   TRUE       300.00          23.00  2.546735
#> 20065  30.38968145   300.00   TRUE       300.00           5.00  3.243142
#> 31121  38.16013507   300.00   TRUE       300.00          45.00  3.194113
#> 24501  27.15246991   300.00   TRUE       300.00           0.00  2.925363
#> 28046  19.59114589   300.00   TRUE       300.00          15.00  2.896560
#> 45500  34.95615508   300.00   TRUE         0.00           2.00  3.626296
#> 17431  25.32530750   300.00   TRUE         1.11           5.00  4.633173
#> 40270  27.59765812   300.00   TRUE       300.00           5.00  3.139431
#> 7385   28.71034216   300.00   TRUE       300.00         100.00  4.219794
#> 47366  28.72753089   300.00   TRUE       300.00         100.00  2.872349
#> 18897  31.78827143   300.00   TRUE        20.00           5.00  3.919162
#> 39722  26.54169690   300.00   TRUE       300.00         100.00  2.906387
#> 6845   44.59616998   300.00   TRUE       300.00          29.00  2.769292
#> 7041   33.21894705   300.00   TRUE       300.00           6.00  5.916915
#> 1973   19.32586643   300.00   TRUE       300.00         100.00  2.290100
#> 32879  29.73822844   300.00   TRUE       300.00           6.00  3.130450
#> 10467  38.99608049   300.00   TRUE       300.00          90.00  2.784369
#> 44767  17.29530400   300.00   TRUE       300.00          25.00  3.046207
#> 47718  32.32341401   300.00   TRUE        10.00           1.00  3.759272
#> 7      13.22042816   300.00   TRUE       300.00           0.00  2.263215
#> 18348  41.01060010   300.00   TRUE       300.00         100.00  3.055854
#> 42201  39.55872505   300.00   TRUE       300.00         100.00  3.989241
#> 20797  32.05698864   285.07   TRUE         0.00          96.00  3.663802
#> 15432  28.76763794   300.00   TRUE       300.00         100.00  2.880810
#> 29228  30.39942174   131.25   TRUE        80.55           2.00  3.618527
#>  [ reached getOption("max.print") -- omitted 324 rows ]