Leave-one-disc-out cross-validation and leave-one-out cross-validation

partition.disc partitions the sample into training and tests set by selecting circular test areas (possibly surrounded by an exclusion buffer) and using the remaining samples as training samples (leave-one-disc-out cross-validation). partition.loo creates training and test sets for leave-one-out cross-validation with (optional) buffer.

partition.disc(data, coords = c("x", "y"), radius, buffer = NULL,
  ndisc = nrow(data), seed1 = NULL, return.train = TRUE, prob = NULL,
  replace = FALSE, repetition = 1)

partition.loo(data, ndisc = nrow(data), replace = FALSE, ...)

Arguments

data

data.frame containing at least the columns specified by coords

coords

vector of length 2 defining the variables in data that contain the x and y coordinates of sample locations.

radius

radius of test area discs; performs leave-one-out resampling if radius <0.

buffer

radius of additional 'neutral area' around test area discs that is excluded from training and test sets; defaults to 0, i.e. all samples are either in the test area or in the training area.

ndisc

Number of discs to be randomly selected; each disc constitutes a separate test set. Defaults to nrow(data), i.e. one disc around each sample.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.train

If FALSE, returns only test sample; if TRUE, also the training area.

prob

optional argument to sample.

replace

optional argument to sample: sampling with or without replacement?

repetition

see partition.cv; however, see Note below: repetition should normally be =1 in this function.

...

arguments to be passed to partition.disc

Value

A represampling object. Contains length(repetition) resampling objects. Each of these contains ndisc lists with indices of test and (if return.train=TRUE) training sets.

Note

Test area discs are centered at (random) samples, not at general random locations. Test area discs may (and likely will) overlap independently of the value of replace. replace only controls the replacement of the center point of discs when drawing center points from the samples.

radius<0 does leave-one-out resampling with an optional buffer. radius=0 is similar except that samples with identical coordinates would fall within the test area disc.

References

Brenning, A. 2005. Spatial prediction models for landslide hazards: review, comparison and evaluation. Natural Hazards and Earth System Sciences, 5(6): 853-862.

See also

sperrorest, partition.cv, partition.kmeans

Examples

data(ecuador)
parti <- partition.disc(ecuador, radius = 200, buffer = 200,
                        ndisc = 5, repetition = 1:2)
# plot(parti,ecuador)
summary(parti)

#> $`1`
#>      n.train n.test
#> 1609    1901     22
#> 72      1886     32
#> 573     1937     15
#> 410     1966     14
#> 194     1938     24
#> 
#> $`2`
#>      n.train n.test
#> 913     1888     37
#> 1206    1919     26
#> 1060    1946     18
#> 1987    1897     18
#> 1372    1910     22
#> 

# leave-one-out with buffer:
parti.loo <- partition.loo(ecuador, buffer = 200)
summary(parti)

#> $`1`
#>      n.train n.test
#> 1609    1901     22
#> 72      1886     32
#> 573     1937     15
#> 410     1966     14
#> 194     1938     24
#> 
#> $`2`
#>      n.train n.test
#> 913     1888     37
#> 1206    1919     26
#> 1060    1946     18
#> 1987    1897     18
#> 1372    1910     22
#>