Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029    5577.3916
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487    1077.7585
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090   0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189     287.5425
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024     394.8751
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581    1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973     514.5488
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678    1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694    2123.4473
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150    1274.3594
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077    3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991    1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517    1000.7119
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178     800.1933
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199     196.4080
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310    2833.9001
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332    1418.6622
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009    1247.7061
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169     495.8840
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469    1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180    3904.5254
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058    1817.2371
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241    1752.9242
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520    3390.5708
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141    4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833     369.3771
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010    3467.2690
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380     601.0843
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526     706.1463
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610     454.8103
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093     118.1402
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105    4432.8760
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181    1286.3949
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116    1817.5222
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674  -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422     967.9738
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048    1550.6704
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528     355.3475
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268     252.0072
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836     250.5111
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849     764.8193
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624    1870.8307
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377    1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172  -0.00119     792.3544
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157     766.0220
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448     941.8503
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982