remove_missing_levels

Accounts for missing factor levels present only in test data but not in train data by setting values to NA

remove_missing_levels(fit, test_data)

Arguments

fit

fitted model on training data

test_data

data to make predictions for

Value

data.frame with matching factor levels to fitted model

Examples

foo <- data.frame(
  response = rnorm(3),
  predictor = as.factor(c("A", "B", "C"))
)
model <- lm(response ~ predictor, foo)
foo.new <- data.frame(predictor = as.factor(c("A", "B", "C", "D")))
predict(model, newdata = remove_missing_levels(
  fit = model,
  test_data = foo.new
))
#> 
#> 'sperrorest()': Setting missing levels in 'predictor', only present in test data but missing in train data, to 'NA'.
#>          1          2          3          4 
#> -1.1786522  0.1819376  0.1866193         NA