Spatial block bootstrap using rectangular blocks

represampling.tile.bootstrap performs a non-overlapping spatial block bootstrap by resampling at the level of rectangular partitions or 'tiles' generated by partition.tiles().

represampling.tile.bootstrap(data, coords = c("x", "y"), repetition = 1,
  nboot = -1, seed1 = NULL, oob = FALSE, ...)

Arguments

data
data.frame containing at least the columns specified by coords

coords
vector of length 2 defining the variables in data that contain the x and y coordinates of sample locations.

repetition
numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

nboot
see represampling.factor.bootstrap()

seed1
seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

oob
see represampling.factor.bootstrap()

...
additional arguments to be passed to partition.tiles()