Draw uniform random (sub)sample at the group level

resample.factor draws a random (sub)sample (with or without replacement) of the groups or clusters identified by the fac argument.

resample.factor(data, param = list(fac = "class", n = Inf, replace = FALSE))

Arguments

data

a data.frame, rows represent samples

param

a list with the following components: fac is a factor variable of length nrow(data) or the name of a factor variable in data; n is a numeric value specifying the size of the subsample (in terms of groups, not observations); replace determines if resampling of groups is to be done with or without replacement.

Value

a data.frame containing a subset of the rows of data.

Details

If param$replace=FALSE, a subsample of min(param$n,nlevel(data[,fac])) groups will be drawn from data. If param$replace=TRUE, the number of groups to be drawn is param$n.

See also

resample.strat.uniform(), sample()

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador
d <- resample.uniform(ecuador, param = list(strat = 'slides', n = 200))
nrow(d) # == 200

#> [1] 200

sum(d$slides == 'TRUE')

#> [1] 100