Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or bootstrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the selection (see Examples).

Another example is bootstrap resampling. represampling_bootstrap with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling_list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

Examples

# Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor(as.character(floor(ecuador$dem / 200)))
smp <- as.resampling(parti)
summary(smp)
#>    n.train n.test
#> 10     600    151
#> 11     585    166
#> 12     660     91
#> 13     641    110
#> 14     727     24
#> 15     747      4
#> 8      730     21
#> 9      567    184
# Compare:
summary(parti)
#>  10  11  12  13  14  15   8   9 
#> 151 166  91 110  24   4  21 184 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition_cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	75 obs. of  13 variables:
#>  $ x             : num  715302 715532 714902 714712 714292 ...
#>  $ y             : num  9557472 9558762 9559262 9561042 9559282 ...
#>  $ dem           : num  2857 2549 2363 1839 2334 ...
#>  $ slope         : num  39.5 38.1 50.7 44.1 32.9 ...
#>  $ hcurv         : num  -0.01021 -0.00373 -0.01407 -0.00223 0.00102 ...
#>  $ vcurv         : num  -0.01579 -0.00057 0.00547 0.01783 0.00388 ...
#>  $ carea         : num  1319 1630 519 378 1051 ...
#>  $ cslope        : num  36.3 31 35.3 16.7 29.3 ...
#>  $ distroad      : num  300 300 300 135 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 300 0 300 ...
#>  $ distslidespast: num  100 100 59 2 5 100 0 2 2 100 ...
#>  $ log.carea     : num  3.12 3.21 2.72 2.58 3.02 ...
# the corresponding training sample - larger:
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	676 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715042 715382 ...
#>  $ y             : num  9560002 9559582 9560172 9559312 9560142 ...
#>  $ dem           : num  1912 2199 1989 2320 2021 ...
#>  $ slope         : num  25.6 23.2 40.5 42.9 42 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 -0.01106 0.00958 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 -0.04634 0.02642 ...
#>  $ carea         : num  5577 1399 351155 501 671 ...
#>  $ cslope        : num  34.4 30.7 32.8 33.9 41.6 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 9.15 300 300 300 0 ...
#>  $ distslidespast: num  9 21 40 100 21 2 100 100 41 5 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.7 2.83 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling_bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	282 obs. of  13 variables:
#>  $ x             : num  712882 715382 712802 713512 714752 ...
#>  $ y             : num  9560002 9560142 9559952 9559092 9561022 ...
#>  $ dem           : num  1912 2021 1838 2166 1848 ...
#>  $ slope         : num  25.6 42 52.1 56 33.4 ...
#>  $ hcurv         : num  -0.00681 0.00958 0.00183 0.02056 -0.00347 ...
#>  $ vcurv         : num  -0.00029 0.02642 -0.09203 -0.06976 0.02357 ...
#>  $ carea         : num  5577 671 634 301 1752 ...
#>  $ cslope        : num  34.4 41.6 30.3 49.4 23.8 ...
#>  $ distroad      : num  300 300 300 300 159 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 9.15 300 0 ...
#>  $ distslidespast: num  9 21 2 41 5 35 100 6 11 4 ...
#>  $ log.carea     : num  3.75 2.83 2.8 2.48 3.24 ...
# bootstrap training sample: same size as nrow(ecuador):
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	751 obs. of  13 variables:
#>  $ x             : num  713592 714672 714262 715022 715332 ...
#>  $ y             : num  9560762 9559272 9561262 9559942 9560852 ...
#>  $ dem           : num  1833 2317 1792 2216 1873 ...
#>  $ slope         : num  36 40.4 12.2 42.1 30.3 ...
#>  $ hcurv         : num  0.00448 -0.01966 0.00161 -0.00613 -0.00161 ...
#>  $ vcurv         : num  -0.00328 -0.02234 0.00709 0.00903 0.00171 ...
#>  $ carea         : num  592 849 1181 2360 757 ...
#>  $ cslope        : num  19.942 35.78 0.142 28.611 31.48 ...
#>  $ distroad      : num  256 300 18 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 1 2 1 2 2 2 2 1 ...
#>  $ distdeforest  : num  0 300 140 300 117 ...
#>  $ distslidespast: num  100 61 100 57 100 100 100 100 43 100 ...
#>  $ log.carea     : num  2.77 2.93 3.07 3.37 2.88 ...