Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or boostrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object
depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...
currently not used.

x
object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the seleciton (see Examples).

Another example is bootstrap resampling. represampling.bootstrap() with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below). as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning. as.resampling.list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

represampling(), partition.cv(), partition.kmeans(), represampling.bootstrap(), etc.

Examples

data(ecuador) # Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor( as.character( floor( ecuador$dem / 200 ) ) )
smp <- as.resampling(parti)
summary(smp)

#>    n.train n.test
#> 10    1657    343
#> 11    1555    445
#> 12    1728    272
#> 13    1801    199
#> 14    1938     62
#> 15    1985     15
#> 8     1885    115
#> 9     1451    549

# Compare:
summary(parti)

#>  10  11  12  13  14  15   8   9 
#> 343 445 272 199  62  15 115 549 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition.cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str( ecuador[ parti[[1]]$test , ] )

#> 'data.frame':	200 obs. of  13 variables:
#>  $ x             : num  715272 714752 715192 714142 714602 ...
#>  $ y             : num  9557702 9561022 9559332 9558562 9560112 ...
#>  $ dem           : num  2813 1848 2266 2383 2143 ...
#>  $ slope         : num  31 33.4 41.2 35.6 51.1 ...
#>  $ hcurv         : num  -0.00123 -0.00347 -0.07664 0.01412 -0.00017 ...
#>  $ vcurv         : num  0.00393 0.02357 -0.05766 0.00898 -0.00032 ...
#>  $ carea         : num  2081 1752 140591 459 995 ...
#>  $ cslope        : num  37.6 23.8 28.5 35.4 42.5 ...
#>  $ distroad      : num  300 159 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 0 300 300 300 ...
#>  $ distslidespast: num  100 5 100 45 100 100 100 10 100 100 ...
#>  $ log.carea     : num  3.32 3.24 5.15 2.66 3 ...

# the corresponding training sample - larger:
str( ecuador[ parti[[1]]$train , ] )

#> 'data.frame':	1800 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715042 715382 ...
#>  $ y             : num  9560002 9559582 9560172 9559312 9560142 ...
#>  $ dem           : num  1912 2199 1989 2320 2021 ...
#>  $ slope         : num  25.6 23.2 40.5 42.9 42 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 -0.01106 0.00958 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 -0.04634 0.02642 ...
#>  $ carea         : num  5577 1399 351155 501 671 ...
#>  $ cslope        : num  34.4 30.7 32.8 33.9 41.6 ...
#>  $ distroad      : num  300 300 300 300 300 300 300 300 30 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 ...
#>  $ distslidespast: num  9 21 40 100 21 2 100 41 20 100 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.7 2.83 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling.bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str( ecuador[ parti[[1]]$test , ] )

#> 'data.frame':	761 obs. of  13 variables:
#>  $ x             : num  712882 715392 713512 714752 714922 ...
#>  $ y             : num  9560002 9560172 9559092 9561022 9558862 ...
#>  $ dem           : num  1912 1989 2166 1848 2536 ...
#>  $ slope         : num  25.6 40.5 56 33.4 54.9 ...
#>  $ hcurv         : num  -0.00681 -0.01919 0.02056 -0.00347 0.01305 ...
#>  $ vcurv         : num  -0.00029 -0.04051 -0.06976 0.02357 -0.00555 ...
#>  $ carea         : num  5577 351155 301 1752 430 ...
#>  $ cslope        : num  34.4 32.8 49.4 23.8 46.3 ...
#>  $ distroad      : num  300 300 300 159 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 0 300 300 300 300 300 300 ...
#>  $ distslidespast: num  9 40 41 5 100 100 2 100 6 11 ...
#>  $ log.carea     : num  3.75 5.55 2.48 3.24 2.63 ...

# bootstrap training sample: same size as nrow(ecuador):
str( ecuador[ parti[[1]]$train , ] )

#> 'data.frame':	2000 obs. of  13 variables:
#>  $ x             : num  713632 713432 714582 714572 715582 ...
#>  $ y             : num  9559632 9558862 9560992 9559302 9559482 ...
#>  $ dem           : num  2226 2237 1862 2276 2317 ...
#>  $ slope         : num  50.8 20.2 17.6 44 38.4 ...
#>  $ hcurv         : num  -0.01881 -0.01092 -0.00729 -0.01185 0.01984 ...
#>  $ vcurv         : num  0.00121 -0.00327 -0.00431 -0.00095 0.02797 ...
#>  $ carea         : num  1783 6915 5705 877 308 ...
#>  $ cslope        : num  40.5 25 28.1 33.9 29 ...
#>  $ distroad      : num  300 300 199 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 1 1 2 1 2 1 2 1 2 ...
#>  $ distdeforest  : num  300 300 16.4 300 300 ...
#>  $ distslidespast: num  4 96 100 93.5 100 7.85 100 4.48 100 0 ...
#>  $ log.carea     : num  3.25 3.84 3.76 2.94 2.49 ...