Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or bootstrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the selection (see Examples).

Another example is bootstrap resampling. represampling_bootstrap with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling_list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

Examples

# Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor(as.character(floor(ecuador$dem / 200)))
smp <- as.resampling(parti)
summary(smp)
#>    n.train n.test
#> 10     600    151
#> 11     585    166
#> 12     660     91
#> 13     641    110
#> 14     727     24
#> 15     747      4
#> 8      730     21
#> 9      567    184
# Compare:
summary(parti)
#>  10  11  12  13  14  15   8   9 
#> 151 166  91 110  24   4  21 184 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition_cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	75 obs. of  13 variables:
#>  $ x             : num  715232 714842 714752 715072 714892 ...
#>  $ y             : num  9559582 9558892 9561022 9558242 9559312 ...
#>  $ dem           : num  2199 2483 1848 2745 2380 ...
#>  $ slope         : num  23.2 68.8 33.4 49.1 32.8 ...
#>  $ hcurv         : num  -0.00501 -0.04921 -0.00347 0.00097 -0.00266 ...
#>  $ vcurv         : num  -0.00649 -0.12438 0.02357 0.01483 0.02896 ...
#>  $ carea         : num  1399 754 1752 567 276 ...
#>  $ cslope        : num  30.7 53.7 23.8 32.2 20.9 ...
#>  $ distroad      : num  300 300 159 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 0 300 300 300 195 0 300 300 ...
#>  $ distslidespast: num  21 100 5 100 31 11 2 100 2 39 ...
#>  $ log.carea     : num  3.15 2.88 3.24 2.75 2.44 ...
# the corresponding training sample - larger:
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	676 obs. of  13 variables:
#>  $ x             : num  712882 715392 715042 715382 712802 ...
#>  $ y             : num  9560002 9560172 9559312 9560142 9559952 ...
#>  $ dem           : num  1912 1989 2320 2021 1838 ...
#>  $ slope         : num  25.6 40.5 42.9 42 52.1 ...
#>  $ hcurv         : num  -0.00681 -0.01919 -0.01106 0.00958 0.00183 ...
#>  $ vcurv         : num  -0.00029 -0.04051 -0.04634 0.02642 -0.09203 ...
#>  $ carea         : num  5577 351155 501 671 634 ...
#>  $ cslope        : num  34.4 32.8 33.9 41.6 30.3 ...
#>  $ distroad      : num  300 300 300 300 300 300 300 30 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 9.15 ...
#>  $ distslidespast: num  9 40 100 21 2 100 41 20 100 100 ...
#>  $ log.carea     : num  3.75 5.55 2.7 2.83 2.8 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling_bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	276 obs. of  13 variables:
#>  $ x             : num  715042 715272 713512 714752 712742 ...
#>  $ y             : num  9559312 9557702 9559092 9561022 9560482 ...
#>  $ dem           : num  2320 2813 2166 1848 1928 ...
#>  $ slope         : num  42.9 31 56 33.4 34.7 ...
#>  $ hcurv         : num  -0.01106 -0.00123 0.02056 -0.00347 -0.00292 ...
#>  $ vcurv         : num  -0.04634 0.00393 -0.06976 0.02357 0.00712 ...
#>  $ carea         : num  501 2081 301 1752 3637 ...
#>  $ cslope        : num  33.9 37.6 49.4 23.8 29 ...
#>  $ distroad      : num  300 300 300 158.9 60.5 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 300 0 152 ...
#>  $ distslidespast: num  100 100 41 5 35 100 10 41 100 100 ...
#>  $ log.carea     : num  2.7 3.32 2.48 3.24 3.56 ...
# bootstrap training sample: same size as nrow(ecuador):
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	751 obs. of  13 variables:
#>  $ x             : num  715622 714552 715532 714412 715042 ...
#>  $ y             : num  9557952 9560382 9558792 9558562 9560342 ...
#>  $ dem           : num  2889 2053 2530 2382 2063 ...
#>  $ slope         : num  35 29.6 32.8 39.9 29.2 ...
#>  $ hcurv         : num  0.01635 -0.00255 -0.00585 -0.04601 0.04336 ...
#>  $ vcurv         : num  -0.00105 -0.00365 -0.00974 -0.08389 0.01403 ...
#>  $ carea         : num  869 1536 2137 47023 241 ...
#>  $ cslope        : num  43.7 23.2 31.6 37.4 32.3 ...
#>  $ distroad      : num  300 300 300 300 300 300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 1 2 2 2 1 1 2 2 1 2 ...
#>  $ distdeforest  : num  300 300 300 300 300 ...
#>  $ distslidespast: num  100 68 100 100 100 100 100 100 100 43 ...
#>  $ log.carea     : num  2.94 3.19 3.33 4.67 2.38 ...