Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or bootstrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the selection (see Examples).

Another example is bootstrap resampling. represampling_bootstrap with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling_list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

Examples

# Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor(as.character(floor(ecuador$dem / 200)))
smp <- as.resampling(parti)
summary(smp)
#>    n.train n.test
#> 10     600    151
#> 11     585    166
#> 12     660     91
#> 13     641    110
#> 14     727     24
#> 15     747      4
#> 8      730     21
#> 9      567    184
# Compare:
summary(parti)
#>  10  11  12  13  14  15   8   9 
#> 151 166  91 110  24   4  21 184 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition_cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	75 obs. of  13 variables:
#>  $ x             : num  715042 712742 712812 713962 714202 ...
#>  $ y             : num  9559312 9560482 9559942 9561282 9558402 ...
#>  $ dem           : num  2320 1928 1861 1796 2476 ...
#>  $ slope         : num  42.9 34.7 63.3 25.7 33.2 ...
#>  $ hcurv         : num  -0.01106 -0.00292 0.00514 0.00273 -0.01607 ...
#>  $ vcurv         : num  -0.04634 0.00712 -0.00644 -0.02403 0.00677 ...
#>  $ carea         : num  501 3637 1179 1787 1344 ...
#>  $ cslope        : num  33.91 29.02 36.52 4.87 29.55 ...
#>  $ distroad      : num  300 60.5 300 51 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 152.09 2.45 0 300 ...
#>  $ distslidespast: num  100 35 0 21 2 85 100 24 65 100 ...
#>  $ log.carea     : num  2.7 3.56 3.07 3.25 3.13 ...
# the corresponding training sample - larger:
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	676 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715382 712802 ...
#>  $ y             : num  9560002 9559582 9560172 9560142 9559952 ...
#>  $ dem           : num  1912 2199 1989 2021 1838 ...
#>  $ slope         : num  25.6 23.2 40.5 42 52.1 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 0.00958 0.00183 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 0.02642 -0.09203 ...
#>  $ carea         : num  5577 1399 351155 671 634 ...
#>  $ cslope        : num  34.4 30.7 32.8 41.6 30.3 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 9.15 ...
#>  $ distslidespast: num  9 21 40 21 2 100 100 41 5 20 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.83 2.8 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling_bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	276 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715382 714842 ...
#>  $ y             : num  9560002 9559582 9560172 9560142 9558892 ...
#>  $ dem           : num  1912 2199 1989 2021 2483 ...
#>  $ slope         : num  25.6 23.2 40.5 42 68.8 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 0.00958 -0.04921 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 0.02642 -0.12438 ...
#>  $ carea         : num  5577 1399 351155 671 754 ...
#>  $ cslope        : num  34.4 30.7 32.8 41.6 53.7 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 ...
#>  $ distslidespast: num  9 21 40 21 100 5 20 100 35 26 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.83 2.88 ...
# bootstrap training sample: same size as nrow(ecuador):
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	751 obs. of  13 variables:
#>  $ x             : num  714832 713962 712812 714032 715632 ...
#>  $ y             : num  9561052 9561082 9559892 9560252 9559102 ...
#>  $ dem           : num  1818 1839 1868 2160 2411 ...
#>  $ slope         : num  46.2 63.4 54.9 30.3 39.7 ...
#>  $ hcurv         : num  -0.01353 -0.04951 -0.01491 0.00537 0.0073 ...
#>  $ vcurv         : num  0.00063 -0.00529 0.01061 -0.00577 0.0177 ...
#>  $ carea         : num  5035 3282 5745 424 475 ...
#>  $ cslope        : num  19.2 29.1 32.3 27 33 ...
#>  $ distroad      : num  138 173 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 1 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  0 33.8 10 205.6 300 ...
#>  $ distslidespast: num  60 100 1 100 100 100 8 100 100 0 ...
#>  $ log.carea     : num  3.7 3.52 3.76 2.63 2.68 ...