Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or bootstrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the selection (see Examples).

Another example is bootstrap resampling. represampling_bootstrap with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling_list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

Examples

# Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor(as.character(floor(ecuador$dem / 200)))
smp <- as.resampling(parti)
summary(smp)
#>    n.train n.test
#> 10     600    151
#> 11     585    166
#> 12     660     91
#> 13     641    110
#> 14     727     24
#> 15     747      4
#> 8      730     21
#> 9      567    184
# Compare:
summary(parti)
#>  10  11  12  13  14  15   8   9 
#> 151 166  91 110  24   4  21 184 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition_cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	75 obs. of  13 variables:
#>  $ x             : num  715272 714022 713642 714202 714032 ...
#>  $ y             : num  9557702 9558862 9558712 9557412 9558502 ...
#>  $ dem           : num  2813 2331 2299 2544 2403 ...
#>  $ slope         : num  31 45.1 33.8 39.9 31.4 ...
#>  $ hcurv         : num  -0.00123 -0.00075 -0.00503 -0.02104 0.01924 ...
#>  $ vcurv         : num  0.00393 0.00475 0.02433 -0.02046 0.02606 ...
#>  $ carea         : num  2081 1001 1475 1024 565 ...
#>  $ cslope        : num  37.6 39.3 23.9 38.8 26.3 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 300 300 300 ...
#>  $ distslidespast: num  100 100 100 100 8 100 46 65 45 100 ...
#>  $ log.carea     : num  3.32 3 3.17 3.01 2.75 ...
# the corresponding training sample - larger:
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	676 obs. of  13 variables:
#>  $ x             : num  712882 715232 715392 715042 715382 ...
#>  $ y             : num  9560002 9559582 9560172 9559312 9560142 ...
#>  $ dem           : num  1912 2199 1989 2320 2021 ...
#>  $ slope         : num  25.6 23.2 40.5 42.9 42 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01919 -0.01106 0.00958 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04051 -0.04634 0.02642 ...
#>  $ carea         : num  5577 1399 351155 501 671 ...
#>  $ cslope        : num  34.4 30.7 32.8 33.9 41.6 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 ...
#>  $ distslidespast: num  9 21 40 100 21 2 100 41 5 20 ...
#>  $ log.carea     : num  3.75 3.15 5.55 2.7 2.83 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling_bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	279 obs. of  13 variables:
#>  $ x             : num  712882 715232 715272 714842 713512 ...
#>  $ y             : num  9560002 9559582 9557702 9558892 9559092 ...
#>  $ dem           : num  1912 2199 2813 2483 2166 ...
#>  $ slope         : num  25.6 23.2 31 68.8 56 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.00123 -0.04921 0.02056 ...
#>  $ vcurv         : num  -0.00029 -0.00649 0.00393 -0.12438 -0.06976 ...
#>  $ carea         : num  5577 1399 2081 754 301 ...
#>  $ cslope        : num  34.4 30.7 37.6 53.7 49.4 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 300 0 300 300 300 300 ...
#>  $ distslidespast: num  9 21 100 100 41 5 100 2 26 10 ...
#>  $ log.carea     : num  3.75 3.15 3.32 2.88 2.48 ...
# bootstrap training sample: same size as nrow(ecuador):
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	751 obs. of  13 variables:
#>  $ x             : num  715142 714932 715392 713702 713812 ...
#>  $ y             : num  9558482 9557982 9560162 9561122 9560592 ...
#>  $ dem           : num  2651 2650 1998 1781 2016 ...
#>  $ slope         : num  42.3 37.3 46.3 49.2 44.3 ...
#>  $ hcurv         : num  -0.00533 0.01633 0.00708 0.00607 -0.00719 ...
#>  $ vcurv         : num  0.01493 -0.01813 -0.02538 -0.04566 0.00079 ...
#>  $ carea         : num  2212 1131 808 12862 1648 ...
#>  $ cslope        : num  33.4 35.1 42.8 20.6 32.1 ...
#>  $ distroad      : num  300 300 300 117 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 1 2 2 1 1 2 1 1 1 2 ...
#>  $ distdeforest  : num  300 300 300 19.6 55.6 ...
#>  $ distslidespast: num  100 100 31 40 100 100 100 68 65 100 ...
#>  $ log.carea     : num  3.34 3.05 2.91 4.11 3.22 ...