Resampling objects such as partitionings or bootstrap samples

Create/coerce and print resampling objects, e.g., partitionings or bootstrap samples derived from a data set.

as.resampling(object, ...)

# S3 method for default
as.resampling(object, ...)

# S3 method for factor
as.resampling(object, ...)

# S3 method for list
as.resampling(object, ...)

validate.resampling(object)

is.resampling(x, ...)

# S3 method for resampling
print(x, ...)

Arguments

object

depending on the function/method, a list or a vector of type factor defining a partitioning of the dataset.

...

currently not used.

x

object of class resampling.

Value

as.resampling methods: An object of class resampling.

Details

A resampling object is a list of lists defining a set of training and test samples.

In the case of k-fold cross-validation partitioning, for example, the corresponding resampling object would be of length k, i.e. contain k lists. Each of these k lists defines a training set of size n(k-1)/k (where n is the overall sample size), and a test set of size n/k. The resampling object does, however, not contain the data itself, but only indices between 1 and n identifying the selection (see Examples).

Another example is bootstrap resampling. represampling_bootstrap with argument oob = TRUE generates represampling objects with indices of a bootstrap sample in the train component and indices of the out-of-bag sample in the test component (see Examples below).

as.resampling.factor: For each factor level of the input variable, as.resampling.factor determines the indices of samples in this level (= test samples) and outside this level (= training samples). Empty levels of object are dropped without warning.

as.resampling_list checks if the list in object has a valid resampling object structure (with components train and test etc.) and assigns the class attribute 'resampling' if successful.

See also

Examples

# Muenchow et al. (2012), see ?ecuador

# Partitioning by elevation classes in 200 m steps:
parti <- factor(as.character(floor(ecuador$dem / 200)))
smp <- as.resampling(parti)
summary(smp)
#>    n.train n.test
#> 10     600    151
#> 11     585    166
#> 12     660     91
#> 13     641    110
#> 14     727     24
#> 15     747      4
#> 8      730     21
#> 9      567    184
# Compare:
summary(parti)
#>  10  11  12  13  14  15   8   9 
#> 151 166  91 110  24   4  21 184 

# k-fold (non-spatial) cross-validation partitioning:
parti <- partition_cv(ecuador)
parti <- parti[[1]] # the first (and only) resampling object in parti
# data corresponding to the test sample of the first fold:
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	75 obs. of  13 variables:
#>  $ x             : num  715392 712992 714932 715302 715362 ...
#>  $ y             : num  9560172 9560672 9557982 9557472 9560102 ...
#>  $ dem           : num  1989 1926 2650 2857 2059 ...
#>  $ slope         : num  40.5 27.2 37.3 39.5 49.1 ...
#>  $ hcurv         : num  -0.01919 -0.00199 0.01633 -0.01021 0.02059 ...
#>  $ vcurv         : num  -0.04051 0.00659 -0.01813 -0.01579 -0.00628 ...
#>  $ carea         : num  351155 3554 1131 1319 556 ...
#>  $ cslope        : num  32.8 27.8 35.1 36.3 43.5 ...
#>  $ distroad      : num  300 30 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 183 300 300 300 ...
#>  $ distslidespast: num  40 20 100 100 26 31 100 100 100 100 ...
#>  $ log.carea     : num  5.55 3.55 3.05 3.12 2.75 ...
# the corresponding training sample - larger:
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	676 obs. of  13 variables:
#>  $ x             : num  712882 715232 715042 715382 712802 ...
#>  $ y             : num  9560002 9559582 9559312 9560142 9559952 ...
#>  $ dem           : num  1912 2199 2320 2021 1838 ...
#>  $ slope         : num  25.6 23.2 42.9 42 52.1 ...
#>  $ hcurv         : num  -0.00681 -0.00501 -0.01106 0.00958 0.00183 ...
#>  $ vcurv         : num  -0.00029 -0.00649 -0.04634 0.02642 -0.09203 ...
#>  $ carea         : num  5577 1399 501 671 634 ...
#>  $ cslope        : num  34.4 30.7 33.9 41.6 30.3 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  15 300 300 300 9.15 300 300 300 0 300 ...
#>  $ distslidespast: num  9 21 100 21 2 100 100 41 5 10 ...
#>  $ log.carea     : num  3.75 3.15 2.7 2.83 2.8 ...

# Bootstrap training sets, out-of-bag test sets:
parti <- represampling_bootstrap(ecuador, oob = TRUE)
parti <- parti[[1]] # the first (and only) resampling object in parti
# out-of-bag test sample: approx. one-third of nrow(ecuador):
str(ecuador[parti[[1]]$test, ])
#> 'data.frame':	279 obs. of  13 variables:
#>  $ x             : num  715232 715392 715042 715272 714932 ...
#>  $ y             : num  9559582 9560172 9559312 9557702 9557982 ...
#>  $ dem           : num  2199 1989 2320 2813 2650 ...
#>  $ slope         : num  23.2 40.5 42.9 31 37.3 ...
#>  $ hcurv         : num  -0.00501 -0.01919 -0.01106 -0.00123 0.01633 ...
#>  $ vcurv         : num  -0.00649 -0.04051 -0.04634 0.00393 -0.01813 ...
#>  $ carea         : num  1399 351155 501 2081 1131 ...
#>  $ cslope        : num  30.7 32.8 33.9 37.6 35.1 ...
#>  $ distroad      : num  300 300 300 300 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ distdeforest  : num  300 300 300 300 300 ...
#>  $ distslidespast: num  21 40 100 100 100 100 100 35 100 2 ...
#>  $ log.carea     : num  3.15 5.55 2.7 3.32 3.05 ...
# bootstrap training sample: same size as nrow(ecuador):
str(ecuador[parti[[1]]$train, ])
#> 'data.frame':	751 obs. of  13 variables:
#>  $ x             : num  715002 712812 714832 713962 714292 ...
#>  $ y             : num  9559922 9560032 9557912 9561282 9557832 ...
#>  $ dem           : num  2238 1847 2661 1796 2496 ...
#>  $ slope         : num  34.6 55 31.9 25.7 38.5 ...
#>  $ hcurv         : num  -0.01109 -0.00767 0.00191 0.00273 -0.00509 ...
#>  $ vcurv         : num  0.00298 -0.03713 -0.00261 -0.02403 0.00499 ...
#>  $ carea         : num  1534 6380 555 1787 885 ...
#>  $ cslope        : num  27.02 34.35 24.91 4.87 31.19 ...
#>  $ distroad      : num  300 300 300 51 300 ...
#>  $ slides        : Factor w/ 2 levels "FALSE","TRUE": 2 2 2 2 2 2 2 2 1 2 ...
#>  $ distdeforest  : num  300 0 300 0 300 300 7.49 300 300 300 ...
#>  $ distslidespast: num  29 0 0 21 75 91 100 89 100 73 ...
#>  $ log.carea     : num  3.19 3.8 2.74 3.25 2.95 ...