Partition the data for a (non-spatial) cross-validation

partition.cv creates a represampling() object for length(repetition)-repeated nfold-fold cross-validation.

partition.cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return.factor = FALSE)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition.cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.factor

if FALSE (default), return a represampling() object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return.factor = FALSE (the default), a represampling() object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling() objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return.factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

sperrorest(), represampling()

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition.cv(ecuador, nfold = 5, repetition = 1:2)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>             x       y  dem slope     hcurv     vcurv   carea  cslope distroad
#> 36710  715232 9559582 2199 23.25  -0.00501 -6.49e-03    1399 30.7157   300.00
#> 46178  714892 9559312 2380 32.77  -0.00266  2.90e-02     276 20.9313   300.00
#> 26531  714462 9560452 2105 19.52   0.00453  2.57e-03     357 15.9878   300.00
#> 43327  713962 9561082 1839 63.36  -0.04951 -5.29e-03    3282 29.0587   173.40
#> 43669  712392 9560162 2002 47.28  -0.00284  1.80e-02     629 28.4766    24.22
#> 49372  714862 9560982 1863 20.97  -0.00295  3.50e-04    2698 23.5405   213.54
#> 39629  714602 9558912 2475 40.26  -0.06175 -8.23e-02    2325 47.5486   300.00
#> 38287  714872 9561162 1719 48.24  -0.03021 -9.33e-02    2236 14.8969    41.43
#> 21703  713842 9559672 2283 24.99   0.01581  1.19e-02     288 22.2692   300.00
#> 8019   714672 9559272 2317 40.38  -0.01966 -2.23e-02     849 35.7801   300.00
#> 33792  714652 9558172 2598 59.46  -0.03525 -1.10e-01     395 51.2081   300.00
#> 2990   713542 9560372 1970 31.89   0.01903  1.45e-02     331 25.3213   300.00
#> 43964  714802 9558452 2684 37.98  -0.02123  1.16e-02     896 28.7573   300.00
#> 43007  713372 9559192 2120 46.01   0.03696  2.51e-02     298 38.9617   300.00
#> 40307  713302 9560702 1845 39.00   0.00129  2.51e-02    3343 29.5136    77.82
#> 38279  714192 9558612 2325 28.25  -0.04430 -6.07e-02  138869 37.6284   300.00
#> 23375  713162 9559632 2042 46.82  -0.00857  3.68e-02    1675 34.6399   300.00
#> 31576  713942 9558492 2387 46.14  -0.05634 -2.63e-02    4316 26.9353   300.00
#> 26215  714332 9560262 2138 25.72  -0.00491  5.81e-03    1855 23.8316   300.00
#> 4827   714962 9559882 2274 29.24   0.00051  8.89e-03     363 21.0740   300.00
#> 45575  715432 9559452 2305 24.23   0.00993 -4.13e-03     365 25.7281   300.00
#> 34280  713652 9560702 1896 41.27   0.00382  8.80e-04    2322 33.6212   300.00
#> 28230  715182 9558102 2792 39.07  -0.00669 -9.50e-03     950 37.9017   300.00
#> 15390  713132 9560622 1873 38.48   0.00252  1.12e-02    2360 29.3223    60.00
#> 13847  715052 9559232 2338 26.11  -0.01452 -1.44e-02    1155 20.5709   300.00
#> 46125  713732 9560062 2115 37.59  -0.04616  1.28e-02    9615 25.6571   300.00
#> 3466   713232 9560612 1813 51.57  -0.01804 -1.60e-02    5297 26.7949   101.32
#> 1635   713102 9559052 2173 31.48  -0.00064 -9.16e-03    2009 27.8268   300.00
#> 35142  714792 9558862 2575 25.66   0.06765  1.14e-03     150 27.1530   300.00
#> 28577  712812 9559962 1850 59.78   0.00646 -4.09e-02    6043 34.3912   300.00
#> 6557   713312 9559022 2182 27.63  -0.00669 -9.91e-03    1403 27.4195   300.00
#> 47011  713262 9559722 2004 37.19  -0.01846 -3.40e-04    4496 44.5343   300.00
#> 24100  715132 9558992 2420 29.21  -0.03229 -5.11e-03  361979 34.2978   300.00
#> 27084  713632 9558472 2380 23.14  -0.01752 -5.17e-03    1001 21.6727   300.00
#> 46201  714832 9561022 1846 34.57  -0.00856  1.80e-02    4730 22.1838   167.75
#> 31316  715022 9559332 2308 20.06   0.00940  6.79e-03     550 30.0098   300.00
#> 17423  715592 9558562 2655 33.22  -0.00034  2.94e-03     740 25.4605   300.00
#> 26567  714022 9558872 2327 46.67   0.00290  2.30e-03    1057 40.4371   300.00
#> 34707  713962 9561092 1819 65.69  -0.04749  4.78e-02    2972 27.7025   165.13
#> 16015  712582 9560312 1899 56.55   0.03480  2.53e-02     937 38.8328   152.09
#> 25244  713472 9559092 2137 18.35  -0.10372 -8.00e-05  825876 33.2189   300.00
#> 835    715052 9558212 2708 37.73  -0.01399 -7.31e-03    1385 36.6504   300.00
#> 4617   714292 9559312 2343 32.39   0.03504  5.83e-02     163 27.9517   300.00
#> 18892  712682 9560202 1839 40.92  -0.01364 -3.36e-02     473 27.0631   235.02
#> 38862  712952 9559892 1951 37.72  -0.00048 -3.32e-03    1419 34.9000   300.00
#> 7466   712962 9560392 1819 18.40  -0.00140 -2.33e-02    1457 28.7906   255.23
#> 30406  715312 9560152 2029 34.14   0.00953 -2.28e-02     590 43.1008   300.00
#> 2595   714912 9557662 2701 24.19  -0.02681 -1.32e-02    3458 30.4395   300.00
#> 13029  713542 9560242 2020 29.55   0.00576 -1.66e-03     651 34.1878   300.00
#> 6317   715042 9557702 2745 27.87   0.00090 -2.99e-03     688 27.3232   300.00
#> 7668   713372 9559062 2165 37.93  -0.01807 -1.88e-02     616 36.3204   300.00
#> 35083  714612 9558152 2596 60.60   0.04548  5.54e-02     352 43.5408   300.00
#> 20065  713772 9558852 2221 28.02  -0.00568 -2.55e-02    1750 30.3897   300.00
#> 31121  713462 9559132 2136 61.60  -0.03800 -6.02e-02    1564 38.1601   300.00
#> 24501  713302 9559142 2126 27.11   0.01784  5.96e-03     842 27.1525   300.00
#> 28046  714882 9558492 2712 28.31  -0.00131  2.01e-03     788 19.5911   300.00
#> 45500  712812 9560052 1852 56.77   0.00904 -2.01e-02    4230 34.9562   300.00
#> 17431  712802 9560022 1833 41.53  -0.00478 -7.13e-02   42971 25.3253   300.00
#> 40270  713312 9559012 2187 30.94  -0.00771 -5.99e-03    1379 27.5977   300.00
#> 7385   713752 9560122 2065 31.77  -0.02605 -3.59e-02   16588 28.7103   300.00
#> 47366  714892 9557612 2719 28.02   0.00538  1.12e-03     745 28.7275   300.00
#> 18897  712802 9559882 1852 50.23  -0.02627 -5.39e-02    8302 31.7883   300.00
#> 39722  715022 9557712 2736 30.26   0.00133  3.67e-03     806 26.5417   300.00
#> 6845   713412 9559172 2123 47.15  -0.01556 -4.39e-02     588 44.5962   300.00
#> 7041   713472 9559092 2137 18.35  -0.10372 -8.00e-05  825876 33.2189   300.00
#> 1973   715252 9557822 2861 15.57   0.01865  3.35e-03     195 19.3259   300.00
#> 32879  713912 9559672 2238 31.98  -0.00278 -4.12e-03    1350 29.7382   300.00
#> 10467  714642 9559652 2249 37.14  -0.00759 -4.00e-03     609 38.9961   300.00
#> 44767  714452 9560402 2092 12.46  -0.00195 -2.05e-03    1112 17.2953   300.00
#> 47718  712812 9559892 1868 54.94  -0.01491  1.06e-02    5745 32.3234   300.00
#> 7      714952 9559862 2284 16.11   0.00516  1.57e-02     183 13.2204   300.00
#> 18348  714042 9558882 2341 44.06  -0.00338  7.28e-03    1137 41.0106   300.00
#> 42201  714642 9557452 2592 13.23  -0.02084 -2.24e-02    9755 39.5587   300.00
#> 20797  713012 9560392 1842 36.65   0.01299  1.41e-03    4611 32.0570   285.07
#> 15432  715632 9558922 2510 35.83  -0.00444 -1.70e-04     760 28.7676   300.00
#> 29228  712782 9560422 1871 32.91  -0.00218 -3.72e-03    4155 30.3994   131.25
#>        slides distdeforest distslidespast log.carea
#> 36710    TRUE       300.00          21.00      3.15
#> 46178    TRUE       300.00          31.00      2.44
#> 26531    TRUE       300.00           2.00      2.55
#> 43327    TRUE        33.81         100.00      3.52
#> 43669    TRUE         0.00          89.00      2.80
#> 49372    TRUE        20.21          68.00      3.43
#> 39629    TRUE       300.00          85.00      3.37
#> 38287    TRUE         1.90         100.00      3.35
#> 21703    TRUE       300.00          34.00      2.46
#> 8019     TRUE       300.00          61.00      2.93
#> 33792    TRUE       300.00           5.00      2.60
#> 2990     TRUE       123.56           5.00      2.52
#> 43964    TRUE       300.00          46.00      2.95
#> 43007    TRUE       300.00          37.00      2.47
#> 40307    TRUE        76.41          54.00      3.52
#> 38279    TRUE       300.00          24.00      5.14
#> 23375    TRUE       195.00           2.00      3.22
#> 31576    TRUE       300.00           4.00      3.64
#> 26215    TRUE       300.00          10.00      3.27
#> 4827     TRUE       300.00           0.00      2.56
#> 45575    TRUE       300.00          59.00      2.56
#> 34280    TRUE         0.00         100.00      3.37
#> 28230    TRUE       300.00         100.00      2.98
#> 15390    TRUE       118.92           2.00      3.37
#> 13847    TRUE       300.00         100.00      3.06
#> 46125    TRUE       300.00         100.00      3.98
#> 3466     TRUE        95.02          19.00      3.72
#> 1635     TRUE       300.00         100.00      3.30
#> 35142    TRUE       300.00         100.00      2.18
#> 28577    TRUE         0.56           0.00      3.78
#> 6557     TRUE       300.00           5.00      3.15
#> 47011    TRUE       300.00          10.00      3.65
#> 24100    TRUE       300.00         100.00      5.56
#> 27084    TRUE       300.00          81.00      3.00
#> 46201    TRUE         0.00          55.00      3.67
#> 31316    TRUE       300.00          85.00      2.74
#> 17423    TRUE       300.00         100.00      2.87
#> 26567    TRUE       300.00         100.00      3.02
#> 34707    TRUE        39.56         100.00      3.47
#> 16015    TRUE        90.23          45.00      2.97
#> 25244    TRUE       300.00           6.00      5.92
#> 835      TRUE       300.00         100.00      3.14
#> 4617     TRUE       300.00           6.00      2.21
#> 18892    TRUE        27.24           0.00      2.68
#> 38862    TRUE       104.48          92.00      3.15
#> 7466     TRUE         0.00          60.00      3.16
#> 30406    TRUE       300.00          85.00      2.77
#> 2595     TRUE       300.00         100.00      3.54
#> 13029    TRUE        96.65           4.00      2.81
#> 6317     TRUE       300.00         100.00      2.84
#> 7668     TRUE       300.00           5.00      2.79
#> 35083    TRUE       300.00          23.00      2.55
#> 20065    TRUE       300.00           5.00      3.24
#> 31121    TRUE       300.00          45.00      3.19
#> 24501    TRUE       300.00           0.00      2.93
#> 28046    TRUE       300.00          15.00      2.90
#> 45500    TRUE         0.00           2.00      3.63
#> 17431    TRUE         1.11           5.00      4.63
#> 40270    TRUE       300.00           5.00      3.14
#> 7385     TRUE       300.00         100.00      4.22
#> 47366    TRUE       300.00         100.00      2.87
#> 18897    TRUE        20.00           5.00      3.92
#> 39722    TRUE       300.00         100.00      2.91
#> 6845     TRUE       300.00          29.00      2.77
#> 7041     TRUE       300.00           6.00      5.92
#> 1973     TRUE       300.00         100.00      2.29
#> 32879    TRUE       300.00           6.00      3.13
#> 10467    TRUE       300.00          90.00      2.78
#> 44767    TRUE       300.00          25.00      3.05
#> 47718    TRUE        10.00           1.00      3.76
#> 7        TRUE       300.00           0.00      2.26
#> 18348    TRUE       300.00         100.00      3.06
#> 42201    TRUE       300.00         100.00      3.99
#> 20797    TRUE         0.00          96.00      3.66
#> 15432    TRUE       300.00         100.00      2.88
#> 29228    TRUE        80.55           2.00      3.62
#>  [ reached getOption("max.print") -- omitted 324 rows ]