Partition the data for a (non-spatial) cross-validation

partition.cv creates a represampling() object for length(repetition)-repeated nfold-fold cross-validation.

partition.cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return.factor = FALSE)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition.cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.factor

if FALSE (default), return a represampling() object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return.factor=FALSE (the default), a represampling() object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling() objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return.factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

Partition example

Figure 1: 5-fold partitions created by partition.cv(). See examples.

See also

sperrorest(), represampling()

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition.cv(ecuador, nfold = 5, repetition = 1:2)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>               x       y     dem     slope      hcurv      vcurv        carea
#> 31358  712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029    5577.3916
#> 44302  714932.5 9557982 2649.56 37.300698    0.01633   -0.01813    1130.5791
#> 42632  714852.5 9557902 2674.80 30.652669    0.00221    0.00969     368.5606
#> 15762  714922.5 9558862 2535.97 54.910556    0.01305   -0.00555     430.4604
#> 40756  714022.5 9558862 2331.20 45.085476   -0.00075    0.00475    1001.0861
#> 37104  713962.5 9561282 1795.56 25.741848    0.00273   -0.02403    1787.3024
#> 26531  714462.5 9560452 2104.74 19.524110    0.00453    0.00257     357.1436
#> 14456  713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010    1527.1028
#> 32622  714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 49372  714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 20151  714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542     674.0147
#> 47255  713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 125    713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 22814  713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 23925  715202.5 9557652 2782.99 41.957126   -0.00542    0.01842     868.6137
#> 25690  715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 10899  713852.5 9559652 2282.82 28.814048    0.00952    0.01378     440.9104
#> 39061  712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 8019   714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 18843  715192.5 9559332 2265.75 41.153267   -0.07664   -0.05766  140591.4062
#> 45869  714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 38410  713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272    4955.5586
#> 31576  713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626    4316.0977
#> 2942   713232.5 9560652 1851.75 34.535031    0.00856    0.01114    4510.8833
#> 26215  714332.5 9560262 2138.08 25.721221   -0.00491    0.00581    1855.1630
#> 6676   715092.5 9559912 2217.44 41.707890   -0.00077   -0.00284    2010.4285
#> 45575  715432.5 9559452 2305.28 24.234396    0.00993   -0.00413     364.9788
#> 32821  715022.5 9559952 2208.33 43.416450   -0.00643    0.00314    2360.1501
#> 18391  715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629  360379.1875
#> 10320  714912.5 9558482 2724.21 25.282909    0.00356    0.00774     527.1068
#> 34280  713652.5 9560702 1896.37 41.273588    0.00382    0.00088    2322.1746
#> 28230  715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950     950.3138
#> 8073   712572.5 9560302 1917.62 52.009098    0.01937    0.00083     994.7876
#> 23031  714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 19730  715332.5 9558452 2640.35 47.642587   -0.00211    0.00642     613.6390
#> 28577  712812.5 9559962 1849.82 59.780125    0.00646   -0.04086    6042.5757
#> 15990  712852.5 9559982 1894.77 37.694320   -0.01545    0.02405    5681.2412
#> 29667  713892.5 9559712 2239.25 41.592725    0.00804    0.00346     543.0749
#> 47011  713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034    4495.6929
#> 2858   714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 19784  714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 20001  712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 36931  712852.5 9558892 2102.80 30.535786    0.00140   -0.00070    2562.7559
#> 27084  713632.5 9558472 2380.37 23.135463   -0.01752   -0.00517    1000.7119
#> 46201  714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 47884  715032.5 9557712 2738.31 28.289218    0.00006    0.00175     804.5679
#> 17423  715592.5 9558562 2654.69 33.219520   -0.00034    0.00294     739.5768
#> 18394  715022.5 9559942 2216.16 42.137608   -0.00613    0.00903    2360.1536
#> 2976   715122.5 9559142 2390.05 33.167954    0.01008    0.00102     483.4184
#> 5084   714912.5 9559302 2359.77 47.968027    0.00107    0.02373     527.3528
#> 49569  713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565    2232.6133
#> 18892  712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357     473.4709
#> 48808  715032.5 9557792 2707.02 25.437607   -0.02329   -0.01181   11655.4316
#> 30406  715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 13029  713542.5 9560242 2019.80 29.545142    0.00576   -0.00166     650.6246
#> 47100  712752.5 9560462 1910.00 41.911863   -0.00032    0.00542    3614.7334
#> 31462  714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 35083  714612.5 9558152 2596.12 60.598308    0.04548    0.05543     352.1558
#> 19966  715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 44825  714322.5 9557882 2534.80 40.688598    0.00114    0.03026     247.0894
#> 35014  713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 17728  713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 28046  714882.5 9558492 2711.77 28.305834   -0.00131    0.00201     788.0621
#> 7064   715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 2552   714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851    1012.5963
#> 42765  713552.5 9559632 2186.95 44.011180    0.00394   -0.00224     942.0322
#> 26010  712852.5 9560542 1928.25 30.978109    0.00165    0.00555    4439.0967
#> 41783  715572.5 9558652 2609.51 32.408212    0.00291   -0.00081     940.6002
#> 7385   713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 37714  713192.5 9560692 1885.05 18.034420    0.00188    0.00832    3022.3462
#> 18958  714192.5 9558372 2495.72 34.677125   -0.00178   -0.00822     472.8748
#> 7041   713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 1973   715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 8125   713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813  287033.4688
#> 10933  712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 47718  712812.5 9559892 1867.55 54.938631   -0.01491    0.01061    5744.7583
#>             cslope distroad slides distdeforest distslidespast log.carea
#> 31358  34.42788799   300.00   TRUE        15.00           9.00  3.746431
#> 44302  35.10111340   300.00   TRUE       300.00         100.00  3.053301
#> 42632  20.47808456   300.00   TRUE       300.00          10.00  2.566509
#> 15762  46.30186534   300.00   TRUE       300.00         100.00  2.633933
#> 40756  39.33527151   300.00   TRUE       300.00         100.00  3.000471
#> 37104   4.86613055    50.99   TRUE         0.00          21.00  3.252198
#> 26531  15.98781432   300.00   TRUE       300.00           2.00  2.552843
#> 14456  20.06899269   300.00   TRUE       247.02         100.00  3.183868
#> 32622  30.17138453   300.00   TRUE       300.00         100.00  3.265863
#> 49372  23.54054397   213.54   TRUE        20.21          68.00  3.430977
#> 20151  44.79555930   300.00   TRUE       300.00         100.00  2.828669
#> 47255  19.67766252   300.00   TRUE       260.03         100.00  3.112324
#> 125    29.59670787   207.04   TRUE        21.03         100.00  3.298701
#> 22814  34.91203733   300.00   TRUE       300.00           6.00  2.976574
#> 23925  37.53331924   300.00   TRUE       300.00         100.00  2.938827
#> 25690  21.74775903   300.00   TRUE       300.00         100.00  2.732242
#> 10899  24.77412210   300.00   TRUE       300.00          20.00  2.644350
#> 39061  35.24091511   300.00   TRUE        70.65          56.00  3.215230
#> 8019   35.78006839   300.00   TRUE       300.00          61.00  2.928936
#> 18843  28.52642270   300.00   TRUE       300.00         100.00  5.147959
#> 45869  23.45460030   300.00   TRUE       300.00          12.00  3.970546
#> 38410   8.76109765    69.52   TRUE        47.61          65.00  3.695093
#> 31576  26.93531891   300.00   TRUE       300.00           4.00  3.635091
#> 2942   29.82073436    68.13   TRUE       116.32          10.00  3.654262
#> 26215  23.83160653   300.00   TRUE       300.00          10.00  3.268382
#> 6676   34.62154773   300.00   TRUE       300.00         100.00  3.303289
#> 45575  25.72809683   300.00   TRUE       300.00          59.00  2.562268
#> 32821  29.48326222   300.00   TRUE       300.00          65.00  3.372940
#> 18391  34.50752913   300.00   TRUE       300.00         100.00  5.556760
#> 10320  16.45592083   300.00   TRUE       300.00          30.00  2.721899
#> 34280  33.62116342   300.00   TRUE         0.00         100.00  3.365895
#> 28230  37.90173111   300.00   TRUE       300.00         100.00  2.977867
#> 8073   36.07571461   138.71   TRUE        76.41          35.00  2.997730
#> 23031  26.74108621   300.00   TRUE       300.00           2.00  2.798938
#> 19730  38.47869961   300.00   TRUE       300.00         100.00  2.787913
#> 28577  34.39121869   300.00   TRUE         0.56           0.00  3.781222
#> 15990  34.14885755   300.00   TRUE         0.00           0.00  3.754443
#> 29667  34.25198995   300.00   TRUE       300.00          27.00  2.734860
#> 47011  44.53429054   300.00   TRUE       300.00          10.00  3.652797
#> 2858   35.31941032   300.00   TRUE       300.00           2.00  2.262887
#> 19784  27.95575674   300.00   TRUE       300.00         100.00  2.941780
#> 20001  34.67712463   264.46   TRUE        18.17          56.00  3.311398
#> 36931  27.20919273   300.00   TRUE       300.00         100.00  3.408707
#> 27084  21.67270156   300.00   TRUE       300.00          81.00  3.000309
#> 46201  22.18377991   167.75   TRUE         0.00          55.00  3.674828
#> 47884  27.05449413   300.00   TRUE       300.00         100.00  2.905563
#> 17423  25.46052554   300.00   TRUE       300.00         100.00  2.868983
#> 18394  28.61064750   300.00   TRUE       300.00          57.00  3.372940
#> 2976   26.88203383   300.00   TRUE       300.00         100.00  2.684323
#> 5084   31.59518465   300.00   TRUE       300.00          18.00  2.722101
#> 49569  41.84941031    25.00   TRUE        35.00         100.00  3.348814
#> 18892  27.06308850   235.02   TRUE        27.24           0.00  2.675293
#> 48808  33.77357019   300.00   TRUE       300.00         100.00  4.066528
#> 30406  43.10075014   300.00   TRUE       300.00          85.00  2.770965
#> 13029  34.18781868   300.00   TRUE        96.65           4.00  2.813330
#> 47100  29.64598224    82.80   TRUE       129.73          36.00  3.558076
#> 31462  33.75466258   300.00   TRUE       300.00           1.00  3.466909
#> 35083  43.54078173   300.00   TRUE       300.00          23.00  2.546735
#> 19966  29.22886896   300.00   TRUE       300.00          23.00  3.350329
#> 44825  29.02145824   300.00   TRUE       300.00         100.00  2.392854
#> 35014  32.53025178   300.00   TRUE       300.00          57.00  2.826223
#> 17728  28.55220580    30.14   TRUE       140.15           2.00  3.303483
#> 28046  19.59114589   300.00   TRUE       300.00          15.00  2.896560
#> 7064   29.13261205   300.00   TRUE       300.00         100.00  3.079714
#> 2552   32.21569795   300.00   TRUE       300.00         100.00  3.005436
#> 42765  42.35304022   300.00   TRUE       300.00          12.00  2.974066
#> 26010  25.22962355    69.52   TRUE       166.56           2.00  3.647295
#> 41783  28.18608577   300.00   TRUE       300.00         100.00  2.973405
#> 7385   28.71034216   300.00   TRUE       300.00         100.00  4.219794
#> 37714  30.73173726    13.57   TRUE       170.00           4.00  3.480344
#> 18958  29.98975691   300.00   TRUE       300.00           5.00  2.674746
#> 7041   33.21894705   300.00   TRUE       300.00           6.00  5.916915
#> 1973   19.32586643   300.00   TRUE       300.00         100.00  2.290100
#> 8125   33.18743437   300.00   TRUE       300.00          48.00  5.457933
#> 10933  34.67712463   264.46   TRUE        18.17          56.00  3.311398
#> 47718  32.32341401   300.00   TRUE        10.00           1.00  3.759272
#>  [ reached getOption("max.print") -- omitted 324 rows ]