Partition the data for a (non-spatial) cross-validation

partition.cv creates a represampling() object for length(repetition)-repeated nfold-fold cross-validation.

partition.cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return.factor = FALSE)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition.cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.factor

if FALSE (default), return a represampling() object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return.factor = FALSE (the default), a represampling() object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling() objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return.factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

Partition example

Fig. 1: Five-fold partitions created by partition.cv(). See examples.

See also

sperrorest(), represampling()

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition.cv(ecuador, nfold = 5, repetition = 1:2)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>               x       y     dem     slope      hcurv      vcurv       carea
#> 44302  714932.5 9557982 2649.56 37.300698    0.01633   -0.01813   1130.5791
#> 25090  715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 24051  715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339   1002.5877
#> 48187  712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767   3321.6113
#> 21939  713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 39885  715062.5 9561022 1840.94 38.435728    0.00250   -0.01340    604.1032
#> 42917  714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046   1024.1573
#> 38411  713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949   6217.8213
#> 27973  715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 8869   715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 14456  713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010   1527.1028
#> 1768   713912.5 9558552 2357.19 38.692413   -0.00645    0.00835   2425.3115
#> 20151  714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542    674.0147
#> 38827  713412.5 9560472 1869.10 16.871315   -0.00156    0.00406   2421.1145
#> 38993  714202.5 9558402 2475.75 33.159359   -0.01607    0.00677   1343.5006
#> 11463  713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 2990   713542.5 9560372 1970.07 31.887966    0.01903    0.01447    331.4833
#> 38410  713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272   4955.5586
#> 48498  714972.5 9557972 2681.66 38.346919    0.00528    0.01392   1255.0884
#> 18304  714582.5 9558932 2463.25 22.194666   -0.02131    0.00501   3889.7292
#> 30125  715162.5 9558992 2438.18 39.856663    0.00471   -0.01382    560.8309
#> 2942   713232.5 9560652 1851.75 34.535031    0.00856    0.01114   4510.8833
#> 6676   715092.5 9559912 2217.44 41.707890   -0.00077   -0.00284   2010.4285
#> 47210  714952.5 9557612 2744.35 40.884549   -0.00493    0.00733    598.7983
#> 45575  715432.5 9559452 2305.28 24.234396    0.00993   -0.00413    364.9788
#> 32821  715022.5 9559952 2208.33 43.416450   -0.00643    0.00314   2360.1501
#> 18391  715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629 360379.1875
#> 27708  715222.5 9557782 2844.65 29.542850    0.03206    0.02773    179.3077
#> 34280  713652.5 9560702 1896.37 41.273588    0.00382    0.00088   2322.1746
#> 34565  712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 27957  715012.5 9559332 2306.35 20.480376    0.00225    0.00325    616.8652
#> 3133   715042.5 9559252 2344.13 29.020885    0.00422    0.02357    275.7474
#> 3466   713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596   5296.9014
#> 28577  712812.5 9559962 1849.82 59.780125    0.00646   -0.04086   6042.5757
#> 34642  714962.5 9559862 2282.08 22.160862   -0.00028    0.01318    289.0063
#> 12868  712532.5 9560332 1913.92 61.937311   -0.03543   -0.05146   1347.6224
#> 11687  714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365   1536.2647
#> 47011  713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034   4495.6929
#> 17594  714832.5 9561052 1818.04 46.167220   -0.01353    0.00063   5035.0000
#> 46201  714832.5 9561022 1845.94 34.573992   -0.00856    0.01796   4729.6382
#> 4605   714852.5 9557892 2677.78 26.933600    0.00321    0.01169    350.1799
#> 43092  715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 47091  713382.5 9559182 2122.71 48.037927    0.04527    0.04613    289.6113
#> 13134  714652.5 9559632 2265.33 49.599810   -0.03552   -0.02058    431.4362
#> 34707  713962.5 9561092 1819.49 65.686173   -0.04749    0.04779   2972.0193
#> 9704   714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 21952  715352.5 9560172 2013.44 40.488063    0.02409    0.02771    674.3859
#> 35194  713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181   2406.1252
#> 421    714892.5 9561022 1849.01 26.132032    0.00002    0.00588    749.2502
#> 21053  712862.5 9559992 1901.01 30.950607   -0.00762    0.01252   5761.3486
#> 48808  715032.5 9557792 2707.02 25.437607   -0.02329   -0.01181  11655.4316
#> 7466   712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330   1456.9246
#> 10507  714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 33350  713132.5 9560672 1896.89 24.695054    0.00306   -0.00436   2602.7390
#> 6317   715042.5 9557702 2745.36 27.873824    0.00090   -0.00299    688.0706
#> 25700  715172.5 9559272 2324.49 35.534842   -0.00529    0.02009   1247.7061
#> 37416  713282.5 9560692 1850.61 30.259047    0.00736    0.01684   2556.8049
#> 32735  714922.5 9559002 2414.49 40.064647    0.00062   -0.00262   1760.2162
#> 31561  713302.5 9559192 2092.11 36.527778    0.00406   -0.00055    773.0359
#> 21013  712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 35014  713502.5 9559772 2190.48 44.235780    0.00927    0.01643    670.2289
#> 13146  713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 17728  713092.5 9560652 1900.14 30.279101    0.00645   -0.00395   2011.3270
#> 5103   714612.5 9559892 2220.36 49.007945   -0.00078    0.00219    420.7032
#> 45500  712812.5 9560052 1852.16 56.770377    0.00904   -0.02013   4229.5708
#> 42765  713552.5 9559632 2186.95 44.011180    0.00394   -0.00224    942.0322
#> 21680  712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399   3900.3445
#> 31834  713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228   2262.3796
#> 7385   713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594  16587.9980
#> 39166  712832.5 9560142 1922.46 39.660139    0.01469    0.02940    987.8085
#> 18958  714192.5 9558372 2495.72 34.677125   -0.00178   -0.00822    472.8748
#> 11612  715212.5 9557752 2820.51 39.162238    0.00017    0.01213    328.7922
#> 19892  714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058   1817.2371
#> 18348  714042.5 9558882 2341.41 44.063892   -0.00338    0.00728   1137.2450
#> 7086   712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 34615  713242.5 9560732 1879.60 27.593647    0.00384    0.00175   1026.6818
#>             cslope distroad slides distdeforest distslidespast log.carea
#> 44302  35.10111340   300.00   TRUE       300.00         100.00  3.053301
#> 25090  43.53161440   300.00   TRUE       300.00          26.00  2.745084
#> 24051  32.73880841   300.00   TRUE       300.00          11.00  3.001122
#> 48187  30.96664995   279.54   TRUE        21.35          46.00  3.521349
#> 21939  23.93244710   300.00   TRUE       300.00         100.00  3.168886
#> 39885   6.08481178   210.57   TRUE         0.00         100.00  2.781111
#> 42917  38.80356668   300.00   TRUE       300.00         100.00  3.010367
#> 38411  32.75828898   279.54   TRUE         0.00         100.00  3.793638
#> 27973  30.00981044   300.00   TRUE       300.00          85.00  2.740390
#> 8869   29.11026670   300.00   TRUE       300.00         100.00  3.036974
#> 14456  20.06899269   300.00   TRUE       247.02         100.00  3.183868
#> 1768   28.18551282   300.00   TRUE       300.00           7.00  3.384768
#> 20151  44.79555930   300.00   TRUE       300.00         100.00  2.828669
#> 38827  31.03597785   300.00   TRUE        52.52          39.00  3.384015
#> 38993  29.54857941   300.00   TRUE       300.00           2.00  3.128238
#> 11463  20.13373692   300.00   TRUE       300.00         100.00  2.768442
#> 2990   25.32129680   300.00   TRUE       123.56           5.00  2.520462
#> 38410   8.76109765    69.52   TRUE        47.61          65.00  3.695093
#> 48498  33.61829863   300.00   TRUE       300.00         100.00  3.098674
#> 18304  44.90843198   300.00   TRUE       300.00          57.00  3.589919
#> 30125  40.73558036   300.00   TRUE       300.00         100.00  2.748832
#> 2942   29.82073436    68.13   TRUE       116.32          10.00  3.654262
#> 6676   34.62154773   300.00   TRUE       300.00         100.00  3.303289
#> 47210  26.91526538   300.00   TRUE       300.00         100.00  2.777281
#> 45575  25.72809683   300.00   TRUE       300.00          59.00  2.562268
#> 32821  29.48326222   300.00   TRUE       300.00          65.00  3.372940
#> 18391  34.50752913   300.00   TRUE       300.00         100.00  5.556760
#> 27708  25.71090810   300.00   TRUE       300.00         100.00  2.253599
#> 34280  33.62116342   300.00   TRUE         0.00         100.00  3.365895
#> 34565  20.38010877   300.00   TRUE        68.13         100.00  2.749114
#> 27957  30.06023072   300.00   TRUE       300.00          75.00  2.790190
#> 3133   18.83426864   300.00   TRUE       300.00         100.00  2.440511
#> 3466   26.79494425   101.32   TRUE        95.02          19.00  3.724022
#> 28577  34.39121869   300.00   TRUE         0.56           0.00  3.781222
#> 34642  15.64862330   300.00   TRUE       300.00           2.00  2.460907
#> 12868  37.58832319   119.34   TRUE        79.16          61.00  3.129568
#> 11687  23.19390450   300.00   TRUE       300.00          68.00  3.186466
#> 47011  44.53429054   300.00   TRUE       300.00          10.00  3.652797
#> 17594  19.23190135   138.35   TRUE         0.00          60.00  3.701999
#> 46201  22.18377991   167.75   TRUE         0.00          55.00  3.674828
#> 4605   17.63564093   300.00   TRUE       300.00          10.00  2.544291
#> 43092  39.47851096   300.00   TRUE       300.00         100.00  2.903195
#> 47091  40.15173637   300.00   TRUE       300.00          25.00  2.461815
#> 13134  37.64390010   300.00   TRUE       300.00          67.00  2.634917
#> 34707  27.70250939   165.13   TRUE        39.56         100.00  3.473052
#> 9704   38.99608049   300.00   TRUE       300.00          90.00  2.784369
#> 21952  40.09157580   300.00   TRUE       300.00          64.00  2.828908
#> 35194  23.44944368   300.00   TRUE       300.00         100.00  3.381318
#> 421    10.71373781   182.02   TRUE         0.00          83.00  2.874627
#> 21053  34.25141699   300.00   TRUE         0.00           2.00  3.760524
#> 48808  33.77357019   300.00   TRUE       300.00         100.00  4.066528
#> 7466   28.79055625   255.23   TRUE         0.00          60.00  3.163437
#> 10507  27.16221019   300.00   TRUE       300.00          46.00  2.580865
#> 33350  28.95270330    10.00   TRUE       166.26           2.00  3.415431
#> 6317   27.32321133   300.00   TRUE       300.00         100.00  2.837633
#> 25700  32.54113797   300.00   TRUE       300.00         100.00  3.096112
#> 37416  28.77164864    72.23   TRUE        87.56          31.00  3.407698
#> 32735  39.35818982   300.00   TRUE       300.00         100.00  3.245566
#> 31561  31.44277788   300.00   TRUE       300.00           0.00  2.888200
#> 21013  24.96319818   300.00   TRUE        13.47          64.00  3.177528
#> 35014  32.53025178   300.00   TRUE       300.00          57.00  2.826223
#> 13146  33.57647271   300.00   TRUE       300.00           2.00  3.278424
#> 17728  28.55220580    30.14   TRUE       140.15           2.00  3.303483
#> 5103   34.91318325   300.00   TRUE       300.00          91.00  2.623976
#> 45500  34.95615508   300.00   TRUE         0.00           2.00  3.626296
#> 42765  42.35304022   300.00   TRUE       300.00          12.00  2.974066
#> 21680  28.34823283   104.12   TRUE       110.76           4.00  3.591103
#> 31834  24.64692547   215.68   TRUE         0.00         100.00  3.354565
#> 7385   28.71034216   300.00   TRUE       300.00         100.00  4.219794
#> 39166  29.81844253   300.00   TRUE         1.90          90.00  2.994673
#> 18958  29.98975691   300.00   TRUE       300.00           5.00  2.674746
#> 11612  35.16700355   300.00   TRUE       300.00         100.00  2.516921
#> 19892  27.02641920   300.00   TRUE       300.00          92.00  3.259412
#> 18348  41.01060010   300.00   TRUE       300.00         100.00  3.055854
#> 7086   34.20443445   300.00   TRUE       128.61          63.00  3.071308
#> 34615  34.18953755    17.57   TRUE       142.95          16.00  3.011436
#>  [ reached getOption("max.print") -- omitted 324 rows ]