Partition the data for a (non-spatial) cross-validation

partition.cv creates a represampling() object for length(repetition)-repeated nfold-fold cross-validation.

partition.cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return.factor = FALSE)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition.cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed() in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return.factor

if FALSE (default), return a represampling() object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return.factor = FALSE (the default), a represampling() object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling() objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return.factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

sperrorest(), represampling()

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition.cv(ecuador, nfold = 5, repetition = 1:2)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>               x       y     dem     slope      hcurv      vcurv        carea
#> 36710  715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649    1399.2329
#> 37912  712802.5 9559952 1838.40 52.101344    0.00183   -0.09203     634.3320
#> 31357  714752.5 9561022 1848.22 33.446411   -0.00347    0.02357    1752.0375
#> 44302  714932.5 9557982 2649.56 37.300698    0.01633   -0.01813    1130.5791
#> 46178  714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 8598   714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752    1321.5997
#> 37265  712812.5 9559942 1860.77 63.337047    0.00514   -0.00644    1179.0695
#> 48187  712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767    3321.6113
#> 21939  713642.5 9558712 2299.09 33.750652   -0.00503    0.02433    1475.3191
#> 15843  714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292    1233.9299
#> 9669   713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 39885  715062.5 9561022 1840.94 38.435728    0.00250   -0.01340     604.1032
#> 26447  715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 31023  714712.5 9561042 1838.66 44.120042   -0.00223    0.01783     377.8918
#> 38993  714202.5 9558402 2475.75 33.159359   -0.01607    0.00677    1343.5006
#> 24018  714292.5 9559282 2334.12 32.865432    0.00102    0.00388    1051.4222
#> 23925  715202.5 9557652 2782.99 41.957126   -0.00542    0.01842     868.6137
#> 25690  715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 30800  715182.5 9557582 2772.37 39.204064   -0.02475    0.00665   25318.3555
#> 16843  714882.5 9558512 2701.78 26.637954   -0.00270   -0.00769    1032.7177
#> 48055  712862.5 9559972 1902.48 29.016875   -0.00720    0.01250    3686.1240
#> 15109  713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 23375  713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 47022  715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 4827   714962.5 9559882 2274.42 29.244912    0.00051    0.00889     362.8363
#> 18391  715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629  360379.1875
#> 10320  714912.5 9558482 2724.21 25.282909    0.00356    0.00774     527.1068
#> 23031  714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 6763   714692.5 9561052 1830.60 48.394880   -0.01996    0.00626    1014.7159
#> 17004  713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 27013  713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 34642  714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 12460  712762.5 9559622 1906.41 24.261898    0.01047   -0.00887     366.8537
#> 11687  714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365    1536.2647
#> 24100  715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511  361979.2812
#> 20001  712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 34563  715022.5 9558162 2675.70 18.889846   -0.01993    0.01373   16031.7500
#> 6310   713802.5 9559752 2238.38 19.651306   -0.02379    0.01338   10787.6299
#> 4605   714852.5 9557892 2677.78 26.933600    0.00321    0.01169     350.1799
#> 6967   715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504    1020.2766
#> 47091  713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 13699  712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907    1943.4092
#> 17423  715592.5 9558562 2654.69 33.219520   -0.00034    0.00294     739.5768
#> 3944   715262.5 9558352 2654.20 33.555273   -0.04492    0.00072   54831.6289
#> 26567  714022.5 9558872 2327.42 46.665694    0.00290    0.00230    1056.5876
#> 2976   715122.5 9559142 2390.05 33.167954    0.01008    0.00102     483.4184
#> 23029  713502.5 9559102 2174.07 56.457542    0.03395    0.01495     266.5888
#> 34707  713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 4031   715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 35194  713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181    2406.1252
#> 835    715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731    1384.7959
#> 15766  715562.5 9558592 2646.16 29.020885    0.00194    0.00516     576.0335
#> 4617   714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 26698  713382.5 9559082 2155.56 34.264595   -0.00211    0.00241     585.6182
#> 19887  714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 32909  712812.5 9560452 1883.85 38.251235   -0.00221    0.00121    4159.6738
#> 6317   715042.5 9557702 2745.36 27.873824    0.00090   -0.00299     688.0706
#> 25700  715172.5 9559272 2324.49 35.534842   -0.00529    0.02009    1247.7061
#> 47100  712752.5 9560462 1910.00 41.911863   -0.00032    0.00542    3614.7334
#> 31462  714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 35083  714612.5 9558152 2596.12 60.598308    0.04548    0.05543     352.1558
#> 44825  714322.5 9557882 2534.80 40.688598    0.00114    0.03026     247.0894
#> 30886  714912.5 9558822 2563.86 49.753936   -0.00099    0.02169     495.8840
#> 31561  713302.5 9559192 2092.11 36.527778    0.00406   -0.00055     773.0359
#> 24501  713302.5 9559142 2125.66 27.108925    0.01784    0.00596     842.0994
#> 28046  714882.5 9558492 2711.77 28.305834   -0.00131    0.00201     788.0621
#> 45500  712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 2552   714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851    1012.5963
#> 5239   714832.5 9560992 1859.70 19.960131   -0.00320    0.00180    3904.5254
#> 12772  715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793     864.9472
#> 6845   713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394     587.8840
#> 43859  714892.5 9557602 2722.70 28.129363    0.00370   -0.00190     640.8842
#> 1973   715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 1424   714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524   10861.9482
#> 47718  712812.5 9559892 1867.55 54.938631   -0.01491    0.01061    5744.7583
#> 31117  712892.5 9558922 2129.84 30.550683    0.00521   -0.00870    1895.1399
#>              cslope distroad slides distdeforest distslidespast log.carea
#> 36710  30.715694439   300.00   TRUE       300.00          21.00  3.145890
#> 37912  30.294570460   300.00   TRUE         9.15           2.00  2.802317
#> 31357  23.817282586   158.92   TRUE         0.00           5.00  3.243543
#> 44302  35.101113403   300.00   TRUE       300.00         100.00  3.053301
#> 46178  20.931294172   300.00   TRUE       300.00          31.00  2.440958
#> 8598   39.598259137   300.00   TRUE       300.00         100.00  3.121100
#> 37265  36.517465073   300.00   TRUE         2.45           0.00  3.071539
#> 48187  30.966649953   279.54   TRUE        21.35          46.00  3.521349
#> 21939  23.932447103   300.00   TRUE       300.00         100.00  3.168886
#> 15843  42.221832881   300.00   TRUE       300.00         100.00  3.091290
#> 9669   31.664512548   300.00   TRUE       300.00           1.00  3.610558
#> 39885   6.084811784   210.57   TRUE         0.00         100.00  2.781111
#> 26447  33.036746467   300.00   TRUE       300.00         100.00  2.676242
#> 31023  16.677655501   135.00   TRUE         0.00           2.00  2.577367
#> 38993  29.548579410   300.00   TRUE       300.00           2.00  3.128238
#> 24018  29.305072348   300.00   TRUE       300.00           5.00  3.021777
#> 23925  37.533319243   300.00   TRUE       300.00         100.00  2.938827
#> 25690  21.747759030   300.00   TRUE       300.00         100.00  2.732242
#> 30800  33.248167894   300.00   TRUE       300.00         100.00  4.403435
#> 16843  21.398254775   300.00   TRUE       300.00          12.00  3.013982
#> 48055  33.825709351   300.00   TRUE         1.67           0.00  3.566570
#> 15109  22.635270654   300.00   TRUE       300.00          40.00  2.931675
#> 23375  34.639882378   300.00   TRUE       195.00           2.00  3.224136
#> 47022  43.531614401   300.00   TRUE       300.00          26.00  2.745084
#> 4827   21.073960663   300.00   TRUE       300.00           0.00  2.559711
#> 18391  34.507529127   300.00   TRUE       300.00         100.00  5.556760
#> 10320  16.455920834   300.00   TRUE       300.00          30.00  2.721899
#> 23031  26.741086214   300.00   TRUE       300.00           2.00  2.798938
#> 6763   16.218143349   125.47   TRUE         0.00           6.00  3.006344
#> 17004  27.189139210   300.00   TRUE       300.00         100.00  3.257531
#> 27013  30.918521499   300.00   TRUE       168.05          18.00  3.080938
#> 34642  15.648623301   300.00   TRUE       300.00           2.00  2.460907
#> 12460  36.256769276   300.00   TRUE         0.00         100.00  2.564493
#> 11687  23.193904505   300.00   TRUE       300.00          68.00  3.186466
#> 24100  34.297826574   300.00   TRUE       300.00         100.00  5.558684
#> 20001  34.677124635   264.46   TRUE        18.17          56.00  3.311398
#> 34563  38.156697325   300.00   TRUE       300.00         100.00  4.204981
#> 6310   23.826449911   300.00   TRUE       300.00         100.00  4.032926
#> 4605   17.635640934   300.00   TRUE       300.00          10.00  2.544291
#> 6967   41.119462083   300.00   TRUE       300.00          85.00  3.008718
#> 47091  40.151736367   300.00   TRUE       300.00          25.00  2.461815
#> 13699  34.789424363   300.00   TRUE        53.95          43.00  3.288564
#> 17423  25.460525542   300.00   TRUE       300.00         100.00  2.868983
#> 3944   39.360481652   300.00   TRUE       300.00         100.00  4.739031
#> 26567  40.437069349   300.00   TRUE       300.00         100.00  3.023906
#> 2976   26.882033832   300.00   TRUE       300.00         100.00  2.684323
#> 23029  46.601522267   300.00   TRUE       300.00          34.00  2.425842
#> 34707  27.702509395   165.13   TRUE        39.56         100.00  3.473052
#> 4031   42.761559124   300.00   TRUE       300.00          31.00  2.907317
#> 35194  23.449443681   300.00   TRUE       300.00         100.00  3.381318
#> 835    36.650391281   300.00   TRUE       300.00         100.00  3.141386
#> 15766  23.337143953   300.00   TRUE       300.00         100.00  2.760448
#> 4617   27.951746035   300.00   TRUE       300.00           6.00  2.211304
#> 26698  34.249698120   300.00   TRUE       300.00           7.00  2.767615
#> 19887  38.062159288   300.00   TRUE       300.00          63.00  2.760635
#> 32909  28.379172551   118.07   TRUE        98.15           0.00  3.619059
#> 6317   27.323211334   300.00   TRUE       300.00         100.00  2.837633
#> 25700  32.541137974   300.00   TRUE       300.00         100.00  3.096112
#> 47100  29.645982236    82.80   TRUE       129.73          36.00  3.558076
#> 31462  33.754662585   300.00   TRUE       300.00           1.00  3.466909
#> 35083  43.540781725   300.00   TRUE       300.00          23.00  2.546735
#> 44825  29.021458239   300.00   TRUE       300.00         100.00  2.392854
#> 30886  39.185156567   300.00   TRUE       300.00         100.00  2.695380
#> 31561  31.442777881   300.00   TRUE       300.00           0.00  2.888200
#> 24501  27.152469911   300.00   TRUE       300.00           0.00  2.925363
#> 28046  19.591145889   300.00   TRUE       300.00          15.00  2.896560
#> 45500  34.956155081   300.00   TRUE         0.00           2.00  3.626296
#> 2552   32.215697947   300.00   TRUE       300.00         100.00  3.005436
#> 5239   24.665260123   197.29   TRUE         4.67          57.00  3.591568
#> 12772  37.920065755   300.00   TRUE       300.00         100.00  2.936990
#> 6845   44.596169984   300.00   TRUE       300.00          29.00  2.769292
#> 43859  29.557746735   300.00   TRUE       300.00         100.00  2.806780
#> 1973   19.325866430   300.00   TRUE       300.00         100.00  2.290100
#> 1424   39.154216846   300.00   TRUE       300.00         100.00  4.035908
#> 47718  32.323414012   300.00   TRUE        10.00           1.00  3.759272
#> 31117  26.645975220   300.00   TRUE       291.23         100.00  3.277641
#>  [ reached getOption("max.print") -- omitted 324 rows ]