Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132  42970.7891
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484    495.0292
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273  72123.6172
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658