Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029   5577.3916
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056 -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665  25318.3555
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150   1274.3594
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346    543.0749
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178    800.1933
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632    620.0229
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075   8141.7549
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135   1293.4855
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603