Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660   -0.01919  -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.97661    0.00958   0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.95233   -0.00123   0.00393   2081.0056
#> 13748 715232.5 9559542 2212.12 26.12344    0.00023  -0.00043   1806.4552
#> 24051 715462.5 9559372 2319.46 32.32800   -0.00661  -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046   0.00324    861.3537
#> 27973 715022.5 9559332 2308.28 20.06326    0.00940   0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.52411    0.00453   0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010   1527.1028
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068   0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.97109    0.00183   0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730   0.01770    474.5059
#> 11463 713112.5 9558902 2244.08 21.89157    0.00160   0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.71360    0.00194   0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.81405    0.00952   0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.15553   -0.00769  -0.00281   1325.6138
#> 16843 714882.5 9558512 2701.78 26.63795   -0.00270  -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.18551    0.00602   0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.10784   -0.00017  -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.17312   -0.01029  -0.02312   2657.6936
#> 22566 712832.5 9559842 1897.49 33.06024    0.00672   0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.14316   -0.05634  -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440   0.00710    843.7961
#> 6676  715092.5 9559912 2217.44 41.70789   -0.00077  -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.88455   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713  -0.00173   8967.1924
#> 13847 715052.5 9559232 2337.87 26.11484   -0.01452  -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.39488   -0.01996   0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.64259   -0.00211   0.00642    613.6390
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543  -0.05146   1347.6224
#> 11687 714552.5 9560382 2052.66 29.57551   -0.00255  -0.00365   1536.2647
#> 6310  713802.5 9559752 2238.38 19.65131   -0.02379   0.01338  10787.6299
#> 27084 713632.5 9558472 2380.37 23.13546   -0.01752  -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.93360    0.00321   0.01169    350.1799
#> 31316 715022.5 9559332 2308.28 20.06326    0.00940   0.00679    550.0344
#> 29926 714702.5 9557622 2637.00 44.64258    0.01843   0.04407    344.8395
#> 4031  715392.5 9560162 1998.16 46.28582    0.00708  -0.02538    807.8237
#> 21952 715352.5 9560172 2013.44 40.48806    0.02409   0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.34897   -0.10372  -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.72526   -0.01399  -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.98633   -0.02917  -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089    0.00194   0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.39217    0.03504   0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762   0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.26460   -0.00211   0.00241    585.6182
#> 38862 712952.5 9559892 1950.86 37.72182   -0.00048  -0.00332   1418.6622
#> 34005 714922.5 9558842 2559.23 52.83129    0.03658   0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221   0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.85627   -0.00458  -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.54514    0.00576  -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.25905    0.00736   0.01684   2556.8049
#> 44650 715612.5 9558142 2764.88 39.36048   -0.00550  -0.00941   1290.7709
#> 21013 712742.5 9560032 1845.53 29.35263    0.00568  -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.23578    0.00927   0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.00795   -0.00078   0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.30583   -0.00131   0.00201    788.0621
#> 21680 712802.5 9560462 1893.75 38.73424   -0.00481  -0.00399   3900.3445
#> 41721 715182.5 9557562 2784.85 31.78999   -0.00588   0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.02394    0.00538   0.00112    745.3315
#> 12772 715172.5 9557642 2759.42 34.80547   -0.00567  -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.67712   -0.00178  -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.12936    0.00370  -0.00190    640.8842
#> 44767 714452.5 9560402 2092.27 12.46298   -0.00195  -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.57246    0.00025  -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.81517   -0.00332  -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.75123   -0.02556  -0.11694 102301.4609
#> 37720 715722.5 9557962 2820.42 45.07631   -0.01474   0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.73275   -0.02070  -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.59365    0.00384   0.00175   1026.6818
#> 5572  715062.5 9558262 2754.60 39.91167    0.00707   0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.98136    0.05379   0.00801    172.7491
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511  -0.01801   1071.6141
#> 11352 713002.5 9560562 1858.95 39.80395   -0.00442  -0.00118  19557.6621
#> 19756 715572.5 9558972 2485.95 38.81273    0.00546   0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.44603    0.01003   0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281   0.04759    158.9370
#> 22301 713722.5 9560712 1913.37 39.27110    0.00705  -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.42980    0.00774  -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.89750    0.00401  -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.30490   -0.00605   0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.95926   -0.00767  -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.21550   -0.01016  -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.87651    0.00010   0.00140   1066.3957
#> 49459 714692.5 9557342 2639.74 42.04708    0.00216  -0.02606    919.4241
#> 17227 714812.5 9558892 2545.43 54.07175    0.00458   0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.28730   -0.00144  -0.00856   1789.1596
#> 4479  713982.5 9557812 2399.24 41.23119    0.01895   0.04744    226.7911
#> 43834 713092.5 9560662 1905.55 31.32360    0.00758  -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186   0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.27259   -0.00166   0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.64751    0.00427   0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585  -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.75329    0.00166  -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.68017    0.00000   0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.05243   -0.04989  -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.02913   -0.01084  -0.02996   1599.3491
#> 32405 713022.5 9558982 2208.85 15.00462    0.00599   0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.62521    0.00159  -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.91793   -0.00849   0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.93724    0.00002  -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.32987    0.00325  -0.00505    351.4078
#> 15186 713322.5 9561212 1936.88 36.07973    0.00165   0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.02184   -0.00759   0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.18588   -0.00298  -0.00712   1010.5097
#> 11111 715142.5 9558482 2650.65 42.34330   -0.00533   0.01493   2212.0525
#> 30267 714442.5 9557512 2612.30 28.76592   -0.00458   0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.71742   -0.00494   0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.93224    0.03480   0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.55090   -0.00195  -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.49492    0.01306  -0.00626    450.7346
#> 42329 715852.5 9558702 2651.92 22.87076    0.00321   0.00548    532.8210
#> 16878 714232.5 9558172 2485.33 22.80830    0.00818   0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.59842    0.01263   0.03598    186.3527
#> 12974 714662.5 9557332 2667.61 41.94166   -0.00065   0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.99798    0.00990   0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.76043   -0.00457   0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.08505   -0.00196  -0.00175    725.1355
#> 14800 712952.5 9558682 2120.89 38.27358   -0.00022   0.00662   1269.2650
#> 22545 713402.5 9559562 2195.44 25.31557    0.04152   0.00098    164.0802
#> 3733  713762.5 9559632 2297.63 34.21475   -0.01338   0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.26401    0.02413   0.00528    355.3475
#> 45486 712762.5 9560172 1853.82 56.79731    0.01978   0.01282   2322.7944
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097   0.01487   1769.3392
#> 44715 714932.5 9560132 2195.78 30.39369    0.03479   0.02251    168.1733
#> 35826 712912.5 9560302 1891.22 50.86948   -0.00229   0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.93765   -0.00070  -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.28635    0.00165   0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.51535    0.00972   0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.28426    0.00804  -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.29457    0.00537  -0.00577    423.5598
#> 28286 713702.5 9559842 2245.07 27.96893    0.00511   0.01499    617.3855
#> 23499 714272.5 9558232 2508.35 28.96760   -0.00820  -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.45424   -0.00273  -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.03698    0.00343   0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.33640    0.00537  -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.44469   -0.00070  -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.03181    0.01600  -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.22166    0.00607  -0.04566  12862.4922
#> 48050 714022.5 9560572 2099.82 18.71395    0.01182  -0.00972    331.7462
#> 11394 713472.5 9561232 1889.48 34.31444    0.00108  -0.00948   1538.1741
#> 9526  714432.5 9560582 2076.13 31.08010   -0.00172  -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.88245    0.01014  -0.00884    672.2466
#> 32342 714862.5 9559622 2294.74 36.96724    0.01021   0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.96439   -0.00933  -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.16930   -0.00716  -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086   -0.00011  -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.57188   -0.00463   0.00253   1237.0536
#> 40852 715392.5 9557932 2939.71 18.29741    0.00505   0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983