Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659    3553.7166
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579    2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835    2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214   0.00466     724.0713
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189     287.5425
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392    1255.0884
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973     514.5488
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413     364.9788
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629  360379.1875
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416    1204.8629
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313     783.4548
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566  -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511  361979.2812
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058     431.4362
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974    2136.6939
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117    1018.2141
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588     749.2502
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236    2715.4138
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565    2232.6133
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408   76318.0156
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436    2602.7390
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596     842.0994
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132   42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180    3904.5254
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776    1021.8267
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133    2397.2681
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325     595.0426
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427     866.9398
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736     477.8591
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363     389.2691
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075    8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548     532.8210
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175     725.1355
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983     270.7949
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018     937.6911
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654    1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710     472.0236
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417    1520.6630
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236  -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047    2826.7700
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215     917.5296
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208     842.4825
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886   43249.3945
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884     672.2466
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803    1058.5630
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156     758.4840
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491