Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501   -0.00649    1399.2329
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958    0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393    2081.0056
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479    0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753    0.01487    1077.7585
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373   -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090    0.00890    2301.0454
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273   -0.02403    1787.3024
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940    0.00679     550.0344
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730    0.01770     474.5059
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175   -0.08226    2325.2053
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160    0.00311     586.7354
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191   -0.00261     554.9243
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832    0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151    0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151    0.00189    2317.1819
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902    0.01358     802.7622
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521   -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356    0.00774     527.1068
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664    0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382    0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669   -0.00950     950.3138
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996    0.00626    1014.7159
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064   -0.00916    2009.1686
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646   -0.04086    6042.5757
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440   -0.02651     419.0927
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566   -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255   -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606   -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904   -0.00866   29195.6016
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316    0.00714     475.3120
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843    0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006    0.00175     804.5679
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286   -0.00964     571.1364
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855   -0.02175    1122.9120
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119   -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917   -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194    0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329   -0.01181   11655.4316
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048   -0.00332    1418.6622
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057    0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306   -0.00436    2602.7390
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032    0.00542    3614.7334
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104    0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645   -0.00395    2011.3270
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394   -0.00224     942.0322
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320    0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604    0.00354    1147.7384
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024   -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.95066  0.03743   -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188    0.00832    3022.3462
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376   -0.01644 1057765.6250
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521   -0.00870    1895.1399
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038   -0.04212   46310.0195
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299    0.00141    4611.0718
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070   -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171   -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666   -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007   -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379    0.00801     172.7491
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124   -0.00456    1531.4274
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546    0.01484     495.0292
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051    0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281    0.04759     158.9370
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292   -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605    0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153   -0.02417   34908.6562
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106   -0.04634     500.5027
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267    0.00483     343.1384
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397    0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143    0.00477    2945.2100
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181   -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560    0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622    0.00152    1530.5657
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400   -0.01210     585.8992
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454    0.00144  317260.6875
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186    0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319    0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683   -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996    0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783   -0.00367    1966.2112
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251   -0.00389   14516.2002
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316   -0.00427     866.9398
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159   -0.01559     866.5775
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465   -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759    0.00158     837.3730
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007   -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533    0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112    0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242    0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458    0.00479     604.0975
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846    0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937   -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818    0.00502    1428.8546
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527   -0.02077    8759.9961
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338   -0.00399    1601.4852
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236    0.00483    1097.3082
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706   -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423   -0.00983     270.7949
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301   -0.01099    1146.2140
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785    0.00645     234.0114
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305   -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066    0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418    0.00513     161.9878
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479    0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858    0.00312     254.0656
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804   -0.00165     805.1403
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820   -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.63451  0.00082    0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952   -0.00438    1276.5448
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264    0.00624    1870.8307
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598    0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182   -0.00972     331.7462
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770    0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203    0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163   -0.00377    1727.6055
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305    0.00585    5714.3267
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992   -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732    0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983    0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986   -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021    0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298    0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962   -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241    0.00020     673.2861
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983