Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193    1809.3838
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517    1000.7119
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560     977.4719
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602     345.8933
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240    1201.4724
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010    3467.2690
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325     595.0426
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856    1789.1596
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945     259.7421
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431     315.9084
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624    1870.8307
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368    1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377    1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119     792.3544
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607