Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983