Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261     323.8441
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373   16031.7500
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495     266.5888
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351    1442.9247
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739     789.7573
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010    3467.2690
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923     811.5580
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955     713.4463
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388     180.4414
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491