Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.7898858  -0.04921   -0.12438    753.6304
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896    276.0310
#> 8598  714002.5 9558872 2310.02 39.5839352  -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.9525357  -0.01073    0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625  -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257    357.1436
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387   1989.3012
#> 39629 714602.5 9558912 2474.89 40.2583065  -0.06175   -0.08226   2325.2053
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.2040642  -0.02475    0.00665  25318.3555
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272   4955.5586
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382    560.8309
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.0602377   0.00672    0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628    556.0121
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450    873.0448
#> 34565 712972.5 9560172 2007.72 27.9471624   0.00598    0.01212    561.1948
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114    150.1498
#> 12868 712532.5 9560332 1913.92 61.9373106  -0.03543   -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991   1403.0193
#> 2858  714292.5 9559332 2335.04 42.3559050   0.04547    0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373  16031.7500
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169    350.1799
#> 3944  715262.5 9558352 2654.20 33.5552733  -0.04492    0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.4575423   0.03395    0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.6561279  -0.00855   -0.02175   1122.9120
#> 38862 712952.5 9559892 1950.86 37.7218224  -0.00048   -0.00332   1418.6622
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009   1247.7061
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328    592.3326
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102  35685.0000
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512  13178.3223
#> 17431 712802.5 9560022 1832.71 41.5251162  -0.00478   -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.4082118   0.00291   -0.00081    940.6002
#> 5239  714832.5 9560992 1859.70 19.9601307  -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354   1147.7384
#> 18897 712802.5 9559882 1851.71 50.2346476  -0.02627   -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.3688733  -0.00093    0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.9506633   0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386   0.01469    0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.2403562  -0.01169   -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.6771246  -0.00178   -0.00822    472.8748
#> 10467 714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400    608.6523
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742   0.03890   -0.00911    173.2806
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.9813582   0.05379    0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522   1355.7436
#> 11352 713002.5 9560562 1858.95 39.8039510  -0.00442   -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.3582039   0.00540    0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426   1315.5334
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044    288.5468
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 46792 713862.5 9559672 2272.03 35.4649416   0.00560    0.00610    454.8103
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.2311889   0.01895    0.04744    226.7911
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253    528.1924
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603   1562.0503
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908   0.00534    0.01476   1275.9446
#> 748   715102.5 9559692 2301.39 27.0218355  -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831  -0.00298   -0.00712   1010.5097
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363    389.2691
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 13286 713662.5 9559702 2268.14 41.9966605  -0.01341    0.04431    315.9084
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.5669070  -0.00978    0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734    335.0773
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.6842028   0.02218    0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.8963648   0.00358    0.02422    967.9738
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048   1550.6704
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196    576.1450
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756   0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750   1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.7668833  -0.00375    0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.2216583   0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.7737308   0.00452    0.00368   1071.8734
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172 111973.8906
#> 40653 714262.5 9561262 1791.94 12.2223357   0.00161    0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523    888.2811
#> 39883 712732.5 9560842 2082.69 25.0909678   0.00732    0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.2771580  -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458    485.7418
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525    162.0592
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057    722.3952
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072   0.00505    0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491