Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035    2697.5974
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651     419.0927
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270    4057.5637
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722    1373.4147
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739     789.7573
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380     601.0843
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477    2945.2100
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684     209.1778
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328    1066.7134
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445     355.5078
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945     259.7421
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853    1354.6001
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298     436.1459
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841     756.0055
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935     903.6397
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177     184.1042
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235