Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029    5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579    2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392    1255.0884
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518     854.4269
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917     116.1986
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593     939.0466
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118    2359.5059
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642     613.6390
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146    1347.6224
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072   54831.6289
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181   11655.4316
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542    3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224     942.0322
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970   0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171    4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325     595.0426
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476   11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528     355.3475
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714  480535.5000
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577     423.5598
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904     608.3561
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270   35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884     672.2466
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234    1707.3130
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603