Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333    258.6537
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102   0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769   1032.7177
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710    843.7961
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809   1102.9956
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961   1086.6963
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970   0.00200   1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152   1530.5657
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895   0.04744    226.7911
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158    837.3730
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923    811.5580
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035   1471.9783
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152   0.00098    164.0802
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624   1870.8307
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491    261.7371
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012    305.3624
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603