Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807 41.600174
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320 30.294570
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861 39.335272
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250   1413.3312 26.977145
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259 30.998736
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725 37.666818
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639 29.110267
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115 28.185513
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153 28.476575
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974 23.540544
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431 18.946568
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587 35.240915
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842 38.961703
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518 29.513629
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177 21.398255
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240 33.825709
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269 22.635271
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309 40.735580
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961 29.199648
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833 29.820734
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630 23.831607
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363 21.073961
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448 25.599754
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473 36.480223
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257  0.00593    939.0466 32.075896
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983 26.915265
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159 16.218143
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924 25.657050
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686 27.826841
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204 34.561960
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757 34.391219
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657 34.595765
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224 37.588323
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193 27.419468
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389 27.416030
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016 39.287716
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408 27.955757
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500 38.156697
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107 30.108359
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119 21.672702
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116 31.451945
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159 16.218143
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072  54831.6289 39.360482
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230   1056.5876 40.437069
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859 40.091576
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406 31.563099
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688 27.951746
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840 35.894087
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791 33.754663
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558 43.540782
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894 29.021458
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956 27.367902
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963 34.342517
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891 25.325308
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081    940.6002 28.186086
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315 28.727531
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100 27.588491
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094 31.788271
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659 26.364080
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242 30.971807
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708 41.828211
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743 27.569583
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.597847
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147 12.843422
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925 27.017825
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680 39.490543
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854 25.336194
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690 26.592690
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100 33.798207
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378 33.615434
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363 27.600523
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027 33.905923
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008 34.424450
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658 40.458269
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100 30.733456
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241 41.856859
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252   3631.7017 30.225243
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010 30.482501
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924 20.820713
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652 31.665086
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189    556.5211 29.094797
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585 -0.02665   1904.4769 36.185150
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487 27.251592
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002 32.263253
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491 41.558348
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644 27.225809
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590 20.891760
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089 35.857991
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432 36.366204
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105 30.892738
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113 50.344083
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715    520.7355 27.681883
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795 23.527366
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306 29.921575
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425 18.669830
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833    300.0125 25.894828
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527 21.442373
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273 34.261157
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773 19.295500
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290 36.084309
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953 32.227157
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236 34.783695
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762 27.000063
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802 23.684356
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554 24.875536
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179 22.100128
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450 21.877820
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547 32.236897
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560 44.227758
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248 25.587149
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179 21.777553
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630 37.882251
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978 26.147502
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451 32.145797
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964    543.0025 38.267851
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645 34.657644
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296 41.510792
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211  7.987605
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562 29.166417
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811 26.856251
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867 18.139271
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672 29.121153
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267 33.563295
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630 33.358176
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128 42.407471
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914 27.306023
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418 16.607182
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300 29.050106
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220 36.130719
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130 45.213246
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.978458
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820 40.158612
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579 31.663367
#>       distroad slides distdeforest distslidespast log.carea
#> 23178   300.00   TRUE       300.00             21  2.826839
#> 37912   300.00   TRUE         9.15              2  2.802317
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 8059    300.00   TRUE       300.00              4  3.150244
#> 11088   300.00   TRUE       300.00            100  3.212274
#> 24072   300.00   TRUE       300.00            100  3.351559
#> 9669    300.00   TRUE       300.00              1  3.610558
#> 8869    300.00   TRUE       300.00            100  3.036974
#> 1768    300.00   TRUE       300.00              7  3.384768
#> 43669    24.22   TRUE         0.00             89  2.798316
#> 49372   213.54   TRUE        20.21             68  3.430977
#> 37270   162.65   TRUE         0.00             77  3.703905
#> 39061   300.00   TRUE        70.65             56  3.215230
#> 43007   300.00   TRUE       300.00             37  2.473902
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 48055   300.00   TRUE         1.67              0  3.566570
#> 15109   300.00   TRUE       300.00             40  2.931675
#> 30125   300.00   TRUE       300.00            100  2.748832
#> 36030   300.00   TRUE         0.00            100  2.926238
#> 2942     68.13   TRUE       116.32             10  3.654262
#> 26215   300.00   TRUE       300.00             10  3.268382
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 40828   300.00   TRUE       300.00            100  2.941037
#> 47220   300.00   TRUE       300.00            100  3.327041
#> 48364   300.00   TRUE       300.00            100  2.972687
#> 38077   300.00   TRUE       300.00            100  2.777281
#> 6763    125.47   TRUE         0.00              6  3.006344
#> 46125   300.00   TRUE       300.00            100  3.982935
#> 1635    300.00   TRUE       300.00            100  3.303016
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 28577   300.00   TRUE         0.56              0  3.781222
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 34642   300.00   TRUE       300.00              2  2.460907
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 6557    300.00   TRUE       300.00              5  3.147064
#> 34359   300.00   TRUE       294.31            100  3.287475
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 19784   300.00   TRUE       300.00            100  2.941780
#> 34563   300.00   TRUE       300.00            100  4.204981
#> 46235   300.00   TRUE        41.37              0  2.752133
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 17423   300.00   TRUE       300.00            100  2.868983
#> 37341   300.00   TRUE       300.00              1  2.823221
#> 39278   125.47   TRUE         0.00              6  3.006344
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 26567   300.00   TRUE       300.00            100  3.023906
#> 21952   300.00   TRUE       300.00             64  2.828908
#> 36120   300.00   TRUE       300.00             94  4.013201
#> 4617    300.00   TRUE       300.00              6  2.211304
#> 39520   300.00   TRUE       300.00            100  2.934945
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 664     300.00   TRUE       300.00             12  2.885027
#> 27695   300.00   TRUE       300.00             41  3.042574
#> 38197   105.52   TRUE        65.32             50  3.036108
#> 45592   300.00   TRUE       300.00              2  2.460907
#> 17431   300.00   TRUE         1.11              5  4.633173
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 12137   300.00   TRUE       300.00            100  3.369811
#> 18897   300.00   TRUE        20.00              5  3.919162
#> 40216   300.00   TRUE       300.00            100  3.257022
#> 6217    300.00   TRUE       300.00              1  3.243763
#> 43201   300.00   TRUE       300.00            100  3.530273
#> 33875   300.00   TRUE         4.67              0  3.279136
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 30933    99.29   TRUE         0.00              6  3.137802
#> 28009   300.00   TRUE       300.00             29  3.185767
#> 7485    300.00   TRUE       300.00            100  2.627640
#> 47298   300.00   TRUE       300.00            100  3.126386
#> 37439   300.00   TRUE       300.00            100  3.539988
#> 10617    17.89   TRUE       145.83              4  3.064162
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 22301   300.00   TRUE         0.00            100  3.308786
#> 23783   300.00   TRUE         0.00            100  3.055889
#> 34984   300.00   TRUE       300.00            100  2.699406
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 48574    96.21   TRUE       101.32             10  3.143827
#> 1223    300.00   TRUE       300.00              4  3.251167
#> 38103   300.00   TRUE       300.00             31  2.907317
#> 17825    73.66   TRUE       137.63             18  3.469116
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 49459   300.00   TRUE       300.00            100  2.963516
#> 13278    91.73   TRUE       120.78             34  3.560110
#> 17674   300.00   TRUE       300.00              1  3.812044
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 17322    75.77   TRUE       138.35              1  3.235115
#> 43536   300.00   TRUE       300.00              8  2.745482
#> 36765   300.00   TRUE       300.00            100  3.279776
#> 39466   300.00   TRUE       300.00            100  3.060187
#> 17851   300.00   TRUE       300.00            100  4.161853
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 23176   300.00   TRUE       300.00            100  3.203943
#> 7317    300.00  FALSE       300.00             61  3.448129
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 48601   300.00  FALSE       245.19            100  2.998830
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 34250   300.00  FALSE       300.00            100  3.348210
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 15558   300.00  FALSE       300.00             45  2.657927
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 8635    300.00  FALSE       300.00             23  2.716617
#> 8066    300.00  FALSE       300.00            100  2.834598
#> 1488    300.00  FALSE       300.00            100  2.930251
#> 441     294.31  FALSE       152.09            100  3.054287
#> 28162   130.00  FALSE        75.00            100  3.197540
#> 1085    300.00  FALSE       300.00            100  2.477139
#> 44396   300.00  FALSE       300.00            100  2.270336
#> 20547   300.00  FALSE        42.56            100  4.078911
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 36126   300.00  FALSE       300.00            100  2.525145
#> 38689   300.00  FALSE       300.00             18  2.953242
#> 42934   300.00  FALSE       300.00             22  4.499525
#> 33175   300.00  FALSE       300.00            100  2.799219
#> 1691    300.00  FALSE       300.00            100  2.855262
#> 22545   300.00  FALSE       300.00             96  2.215056
#> 48869   138.98  FALSE         0.00             63  2.748851
#> 3733    300.00  FALSE       300.00            100  3.055767
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 41612   300.00  FALSE       300.00            100  3.004762
#> 26724   300.00  FALSE       185.65            100  3.274624
#> 14671   123.04  FALSE        75.11              0  3.182033
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 484     300.00  FALSE        55.63            100  3.216864
#> 38362   300.00  FALSE       300.00            100  2.734802
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 26283   300.00  FALSE       300.00            100  2.962620
#> 18246    18.49  FALSE         0.00             29  3.220009
#> 10981   300.00  FALSE       300.00            100  5.155983
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 9320    300.00  FALSE       300.00             93  2.700519
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 46639   300.00  FALSE       300.00            100  3.756965
#> 49812   300.00  FALSE       300.00             65  3.024717
#> 6577    300.00  FALSE       300.00             70  2.789872
#> 36141   300.00  FALSE       300.00             85  4.412069
#> 11570   300.00  FALSE       300.00            100  2.686405
#> 24962   300.00  FALSE       300.00            100  2.996437
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 14645   300.00  FALSE       300.00            100  3.232313
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 30296   300.00  FALSE       300.00             68  3.258250
#> 47629   300.00  FALSE       300.00            100  3.627607