Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029    5577.3916
#> 36052 714842.5 9558892 2483.27 68.7898858  -0.04921   -0.12438     753.6304
#> 27864 712992.5 9560672 1926.38 27.2240896  -0.00199    0.00659    3553.7166
#> 16749 715072.5 9558242 2745.27 49.0537816   0.00097    0.01483     566.7674
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896     276.0310
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324     861.3537
#> 11088 715532.5 9558762 2549.28 38.0501272  -0.00373   -0.00057    1630.3259
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.0050732  -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257     357.1436
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837  -0.04951   -0.00529    3282.1186
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264    5057.1431
#> 27375 714292.5 9558982 2458.96 43.6009423   0.00214    0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526     539.8119
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.4586952  -0.03525   -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 48055 712862.5 9559972 1902.48 29.0168746  -0.00720    0.01250    3686.1240
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203   10263.2734
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314    2360.1501
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 27013 713062.5 9559782 1976.81 41.6322593   0.01414    0.03416    1204.8629
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887     366.8537
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096     520.5570
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504    1020.2766
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907    1943.4092
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175     804.5679
#> 651   714602.5 9560152 2111.23 53.8517302  -0.01149    0.01949    1182.1506
#> 2976  715122.5 9559142 2390.05 33.1679538   0.01008    0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423   0.03395    0.01495     266.5888
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373     527.3528
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310    2833.9001
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063     576.2819
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319    3457.7519
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427     380.9472
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473    2930.2791
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.7539361  -0.00099    0.02169     495.8840
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180  -0.00104    0.01664    1898.5588
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112     745.3315
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 11612 715212.5 9557752 2820.51 39.1622383   0.00017    0.01213     328.7922
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870    1895.1399
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195    1901.6743
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522    1355.7436
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860  -0.01109    0.00298    1533.7925
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484     495.0292
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190     264.6404
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713    6379.7974
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.8087514  -0.01881    0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044     288.5468
#> 24723 712942.5 9560202 1983.18 33.0716969   0.00285   -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473  -0.01397    0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119    6589.5796
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921     525.3245
#> 28588 715532.5 9559362 2299.84 11.1446021  -0.04360   -0.03000  162178.1562
#> 17227 714812.5 9558892 2545.43 54.0717460   0.00458    0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.2311889   0.01895    0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 41408 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.8077801  -0.00486    0.02306    1718.3652
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403     474.8613
#> 36765 715302.5 9560262 1952.73 27.2034632  -0.00585   -0.02665    1904.4769
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301     338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150     503.5590
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.0218355  -0.00759    0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736     477.8591
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.7659191  -0.00458    0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463     851.6306
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.3029762   0.00540    0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199     836.4196
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502    1428.8546
#> 34347 714742.5 9560142 2161.95 26.5669070  -0.00978    0.00367    3002.8828
#> 34616 714522.5 9558742 2568.77 34.9979810   0.00990    0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.4863297  -0.00069   -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.2478257   0.00737    0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135  -0.00844    0.00514     897.9290
#> 36245 713142.5 9560572 1827.82 43.6759998  -0.00410    0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.9465824   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645     234.0114
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 5656  713132.5 9560142 2042.24 33.5953803  -0.00035    0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119  -0.00950   -0.05130     107.5831
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554   0.01858    0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360    4288.7329
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.9689348   0.00511    0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370  -0.00273   -0.01047    2826.7700
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532   16771.3066
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 42990 712432.5 9560692 2102.11 32.4500377  -0.00087   -0.00573    1490.0312
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.7319259   0.07379   -0.00249     148.1857
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491