Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 27864 712992.5 9560672 1926.38 27.2240896  -0.00199    0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.6526691   0.00221    0.00969     368.5606
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324     861.3537
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767    3321.6113
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107  -0.00858   -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.9525357  -0.01073    0.00174    2310.4871
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303   0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293   0.00884    0.01345    1120.6929
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.2583065  -0.06175   -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.6120312  -0.02894   -0.03416     947.4878
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328    2235.5259
#> 3835  714142.5 9558552 2388.42 32.9003825   0.00979   -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104    1641.4587
#> 18843 715192.5 9559332 2265.75 41.1532666  -0.07664   -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182  -0.02123    0.01164     895.6606
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514     297.7842
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769    1032.7177
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677    1675.4679
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973     514.5488
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.1282451  -0.00033   -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694    2123.4473
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924    8023.8657
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 47011 713262.5 9559722 2003.72 37.1929823  -0.01846   -0.00034    4495.6929
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.6513065  -0.02379    0.01338   10787.6299
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178     800.1933
#> 18751 714882.5 9559182 2381.81 39.2991752   0.00066   -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407     344.8395
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302  -0.01149    0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.8465245  -0.00585   -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560     977.4719
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197     711.1091
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.2122812  -0.01160    0.01690    5121.6924
#> 39732 715362.5 9559582 2230.97 23.3468842  -0.00161   -0.00409    2312.8413
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.5983082   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625    1063.9150
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328     592.3326
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055     773.0359
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.1089251   0.01784    0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388  -0.00481   -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.9601307  -0.00320    0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354    1147.7384
#> 12137 712842.5 9558902 2097.88 30.0986189   0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733  -0.00093    0.00664    1807.2659
#> 10467 714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.7916671  -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.1612670   0.00129    0.00171     788.0004
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211   49347.5547
#> 30287 713992.5 9557822 2403.92 18.2154742   0.03890   -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922  -0.00338    0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322    4605.1411
#> 5572  715062.5 9558262 2754.60 39.9116671   0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.9813582   0.05379    0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.5275756  -0.00193    0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970    1443.9200
#> 38668 715582.5 9558242 2735.55 35.4460340   0.01003    0.01196    1052.5341
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.5663341  -0.00693    0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171    4425.5410
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450    2589.9910
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.4840379  -0.01125   -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875     597.2893
#> 46792 713862.5 9559672 2272.03 35.4649416   0.00560    0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.1803772   0.00124    0.01166    1813.7841
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867    2247.8748
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655     130.9468
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996    1599.3491
#> 16208 712882.5 9560332 1846.34 48.6252092   0.00159   -0.01559     866.5775
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563   0.04336    0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923     811.5580
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 49835 713272.5 9560372 1898.68 32.8258980   0.00846    0.00794    1363.5321
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973     215.5101
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507     504.3085
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 24009 714042.5 9558962 2357.58 34.3619978   0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.9465824   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552    2515.0752
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863    1098.3462
#> 34532 713812.5 9559502 2328.80 44.9376528  -0.00070   -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788   0.00119    0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360    4288.7329
#> 24819 714032.5 9560882 1981.86 33.2842642   0.00804   -0.00165     805.1403
#> 40612 713542.5 9561152 1856.15 33.5684513   0.00420    0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904     608.3561
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 1577  713322.5 9559742 2027.47 29.5674870  -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.0235474   0.00340    0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.2223357   0.00161    0.00709    1181.2555
#> 9526  714432.5 9560582 2076.13 31.0800956  -0.00172   -0.00119     792.3544
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064    1485.5201
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803    1058.5630
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458     485.7418
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234    1707.3130
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020     673.2861
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603