Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969    368.5606
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149  12632.7793
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501   3889.7292
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385    674.0721
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 18892 712682.5 9560202 1839.08 40.9235105 -0.01364 -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684   2556.8049
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485   2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656   3036.7383
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298   1533.7925
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396    704.7076
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133   2397.2681
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341   1363.4000
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132    737.5982
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548    532.8210
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714 480535.5000
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162    290.2031
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710    472.0236
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260   1484.6593
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347   1893.3645
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172    351.5812
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775