Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.1013441   0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393    2081.0056
#> 15762 714922.5 9558862 2535.97 54.9105562   0.01305   -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.1234377   0.00023   -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487    1077.7585
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107  -0.00858   -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403    1787.3024
#> 8869  715062.5 9559392 2278.56 41.4019303   0.00310    0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311     586.7354
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.9003825   0.00979   -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189     287.5425
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749   0.00129    0.02510    3342.5518
#> 22933 712712.5 9560432 1902.17 38.1234658  -0.00746    0.00096    3210.5781
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.4003790   0.00151    0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311   0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.5384969  -0.02716    0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889     362.8363
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.0717109  -0.00669   -0.00950     950.3138
#> 27013 713062.5 9559782 1976.81 41.6322593   0.01414    0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245   0.00646   -0.04086    6042.5757
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313     783.4548
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556     660.3690
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452     437.3965
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025   0.01661    0.01199     196.4080
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964     571.1364
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.7968588  -0.00449   -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.0208853   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071  -0.00762    0.01252    5761.3486
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357     473.4709
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319    3457.7519
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684    2556.8049
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.2748872  -0.00845    0.00454    3476.0818
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316     454.9674
#> 38197 712522.5 9560322 1942.04 62.7772667  -0.01611    0.02961    1086.6963
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.7703772   0.00904   -0.02013    4229.5708
#> 21680 712802.5 9560462 1893.75 38.7342388  -0.00481   -0.00399    3900.3445
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.7710827  -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.2346476  -0.02627   -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.2607659   0.00133    0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.6601386   0.01469    0.02940     987.8085
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.5707010   0.01865    0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383   0.00017    0.01213     328.7922
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205    1112.2626
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211   49347.5547
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.8763323  -0.00310   -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078  622888.2500
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372    4154.5786
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522    1355.7436
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.2390497   0.01011    0.00739     789.7573
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484     495.0292
#> 16071 714562.5 9560062 2166.69 28.6960182   0.02281    0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.8354356   0.00181   -0.00001    1184.7952
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140    1066.3957
#> 30363 715192.5 9557542 2796.31 25.3218698  -0.00157   -0.00152     850.4633
#> 30626 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 14779 712502.5 9560202 2004.61 27.5226006   0.00115    0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.5225656  -0.01454    0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.3812407  -0.00465   -0.01036    2505.8369
#> 9984  712742.5 9560092 1848.56 23.6127367   0.00614   -0.00093     118.1402
#> 17851 715362.5 9557482 2869.26 29.1257366  -0.03251   -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.9576081   0.00378    0.02382     100.0000
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419     761.5427
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118    2273.4758
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.7332955  -0.01174   -0.01466    2633.9666
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028    2394.7261
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734     335.0773
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267   31588.1953
#> 8895  714132.5 9557692 2487.19 37.3642967   0.03423   -0.00983     270.7949
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098     164.0802
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 17746 714752.5 9558322 2648.52 47.7365517  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.0747293  -0.00112   -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487    1769.3392
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204     618.2027
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479   0.00972    0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577     423.5598
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964   0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.1104343  -0.01502    0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148     481.8292
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177     184.1042
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064    1485.5201
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803    1058.5630
#> 32342 714862.5 9559622 2294.74 36.9672369   0.01021    0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281     492.4785
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253    1237.0536
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775