Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041    772.9746
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312   2657.6936
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167   1938.5389
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560    977.4719
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632    620.0229
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565   1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875   1165.3953
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937  -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759   1880.4403
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780    825.4451
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955    713.4463
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624   1870.8307
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238    787.0053
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013    616.4128
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983