Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393   2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507 159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041    772.9746
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579   2246.7725
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046   1024.1573
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804    628.5153
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677   1343.5006
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269    491.8017
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096   3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398    267.2659
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284   2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917    116.1986
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205    629.4165
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338  10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025   4425.5151
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925    665.6116
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197    711.1091
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602    345.8933
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883    616.1993
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099    0.02169    495.8840
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599   1378.5760
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822    472.8748
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205   1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298   1533.7925
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246    0.00396    704.7076
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484    495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759    158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957    495.4453
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483    343.1384
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367   1966.2112
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842    246.5161
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336   2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208    951.3401
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995   1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409   1575.9425
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794   1363.5321
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734    335.0773
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422    967.9738
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268    252.0072
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204    618.2027
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555   1882.0179
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247    282.1076
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836    250.5111
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875   2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886  43249.3945
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177    184.1042
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119    792.3544
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803   1058.5630
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235