Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043    1806.4552
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024     394.8751
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696    4143.4019
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313     783.4548
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169     350.1799
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230    1056.5876
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731    1384.7959
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310    2833.9001
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512   13178.3223
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594   16587.9980
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878 -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322    4605.1411
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522    1355.7436
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094    1392.6008
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152    1530.5657
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252    3631.7017
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230    1044.6443
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093     118.1402
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177     997.3089
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493    2212.0525
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345  665381.5625
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333     422.8840
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645     234.0114
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884     672.2466
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235