Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438     753.6304
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406    2421.1145
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392    1255.0884
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491  0.00581    1855.1630
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642     613.6390
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651     419.0927
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796    4729.6382
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310    2833.9001
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555    4439.0967
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245    6353.3179
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583    1898.0105
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722    1373.4147
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363     565.5207
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325     595.0426
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119    6589.5796
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279   39495.1797
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684     209.1778
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466    2633.9666
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995    1026.7981
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199     836.4196
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077    8759.9961
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759    1880.4403
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902     218.7467
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645     234.0114
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863    1098.3462
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502     520.5218
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119     792.3544
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403     609.8665
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235