Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958    0.02642    671.1807
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021   -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221    0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292    0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917    0.00333    258.6537
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118    0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073    0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340    604.1032
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310    0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884    0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730   -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295    0.00035   2697.5974
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894   -0.03416    947.4878
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979   -0.00269    491.8017
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129    0.02510   3342.5518
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634   -0.02626   4316.0977
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426   -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856    0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491    0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628    556.0121
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993   -0.00413    364.9788
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033   -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669   -0.00950    950.3138
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252    0.01118   2359.5059
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616    0.01276   9614.6924
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642    613.6390
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651    419.0927
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463    0.00167   1938.5389
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140   -0.00070   2562.7559
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752   -0.00517   1000.7119
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066   -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288    0.01452    437.3965
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735   -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626   1014.7159
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492    0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286   -0.00964    571.1364
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008 825875.6875
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175    0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917   -0.01993  10308.6406
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797   -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140   -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658    0.04602    345.8933
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032    0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304   1569.7313
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282    0.00109    767.4097
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104    0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972    0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131    0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394   -0.00224    942.0322
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604    0.00354   1147.7384
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370   -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865    0.00335    195.0293
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412   1350.3611
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509   -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890   -0.00911    173.2806
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797   7557.7866
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466   -0.01776   1021.8267
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321   1170.5904
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540    0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546    0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669   -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051    0.02190    264.6404
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181   -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705   -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605    0.00075   2170.2249
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267    0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915   -0.02705    413.5702
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610   1872.9276
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019    0.00119   6589.5796
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181   -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124    0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354    0.00245    629.5983
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234    849.0547
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486    0.02306   1718.3652
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614   -0.00093    118.1402
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002   -0.01232    977.7320
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402   -0.01328   1066.7134
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220   -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195   -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161    0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544   -0.01014    271.4219
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978    0.00367   3002.8828
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174   -0.03065    704.4843
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457    0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236    0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844    0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706   -0.03035   1471.9783
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358    0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687   -0.00223   1056.3701
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224   -0.02036    476.9504
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413    0.00528    355.3475
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229    0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480    0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119    0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980   -0.00710    472.0236
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944    0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044   -0.00964    543.0025
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343    0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552    0.00532  16771.3066
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182   -0.00972    331.7462
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770    0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233    968.7780
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196   -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340    0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709   1181.2555
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014   -0.00884    672.2466
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181   -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933   -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087   -0.00573   1490.0312
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366    0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023   -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208    0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769