Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491