Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261    323.8441
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041    772.9746
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090    0.00890   2301.0454
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949   6217.8213
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388   1051.4222
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123    0.01164    895.6606
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270   -0.00769   1032.7177
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191   -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385    674.0721
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356    0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452   -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193   1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642    613.6390
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346    543.0749
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167   1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509    0.00499    885.0064
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752   -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714    475.3120
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066   -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175    804.5679
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585   -0.00974   2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779   2972.0193
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107    0.02373    527.3528
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252   5761.3486
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708    0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221    0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736    0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548    0.05543    352.1558
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114    0.03026    247.0894
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395   2011.3270
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132  42970.7891
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370   -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246    0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524  10861.9482
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521   -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949    0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078 622888.2500
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466   -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439   -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669   -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380    601.0843
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970    0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774   -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864    0.04226    338.3474
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610    454.8103
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252   3631.7017
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367   1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614   -0.00093    118.1402
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079   -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336   2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109    605.1908
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085   2229.5105
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204    853.2902
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479    604.0975
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475    515.1661
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033   3446.5520
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902    218.7467
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301   -0.01099   1146.2140
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714 480535.5000
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950   -0.05130    107.5831
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923    0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165    805.1403
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849    764.8193
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770    0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233    968.7780
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377   1727.6055
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109    915.4989
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250