Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976     301.2347
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261     323.8441
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250    1413.3312
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005     769.3426
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102   0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269     491.8017
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694    2123.4473
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357     275.7474
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346     543.0749
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175    1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181   11655.4316
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883     616.1993
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485    2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555    4439.0967
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112     745.3315
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058    1817.2371
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520    3390.5708
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372    4154.5786
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970   0.00200    1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450    2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610    1872.9276
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063    2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389   14516.2002
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382     100.0000
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181    1286.3949
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923     811.5580
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476   11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267   31588.1953
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650     766.6254
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904     608.3561
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907    1281.5044
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448     941.8503
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607