Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958    0.02642    671.1807
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221    0.00969    368.5606
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659    0.01041    772.9746
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090    0.00890   2301.0454
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118    0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073    0.00174   2310.4871
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295    0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979   -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730    0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407    0.00547    519.4228
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894   -0.03416    947.4878
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979   -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581    0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414   -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664   -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129    0.02510   3342.5518
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440    0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613   -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033   -0.03917    116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694   2123.4473
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252    0.01118   2359.5059
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642    613.6390
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651    419.0927
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804    0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255   -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866  29195.6016
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993    0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752   -0.00517   1000.7119
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527    0.04613    289.6113
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025   4425.5151
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552   -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617    0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359   -0.00279    930.0871
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478    0.01092    412.5849
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658    0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160    0.01690   5121.6924
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032    0.00542   3614.7334
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062   -0.00262   1760.2162
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568   -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168   -0.04102  35685.0000
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927   -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104    0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131    0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904   -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629   -0.00469   1021.5407
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743   -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188    0.00832   3022.3462
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025   -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332   -0.00058   1817.2371
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890   -0.00911    173.2806
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949    0.00520   3390.5708
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439   -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321   1170.5904
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011    0.00739    789.7573
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401   -0.00391   1794.7964
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364   -0.00094   1392.6008
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864    0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019    0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143    0.00477   2945.2100
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212    0.02732   4237.6934
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675   -0.01905   6487.0010
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465   -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319    0.00291    487.3438
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996    0.04684    209.1778
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783   -0.00367   1966.2112
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519    0.00609   4275.3457
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842    246.5161
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336    0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208    951.3401
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197    0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345   -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306   -0.00626    450.7346
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502   -0.00028   2394.7261
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341    0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978    0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538   -0.00052   3508.5591
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457    0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737    0.01033   3446.5520
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674   -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410    0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292    0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152    0.00098    164.0802
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338    0.00518   1137.0179
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068   -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424    0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586    0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462    0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305   -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035    0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118    0.00162    290.2031
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479    0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923    0.05547    128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719    0.00079   1647.6451
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044   -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273   -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354    0.00907   1281.5044
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182   -0.00972    331.7462
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196   -0.01935    903.6397
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163   -0.00377   1727.6055
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992   -0.01052    451.7209
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298    0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811   -0.02281    492.4785
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241    0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751    0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181   -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607