Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769