Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339   1002.5877
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005    769.3426
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684   2556.8049
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334   4905.2002
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118  19557.6621
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033   3446.5520
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769