Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659   3553.7166
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043   1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261    323.8441
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890   2301.0454
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340    604.1032
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804    628.5153
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149  12632.7793
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104   1641.4587
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626   4316.0977
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083    994.7876
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924   8023.8657
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517   1000.7119
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178    800.1933
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279    930.0871
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408  76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180   3904.5254
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524  10861.9482
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911    173.2806
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143   -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484    495.0292
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200   1305.6691
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015    790.2421
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291    487.3438
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331    699.3024
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150    503.5590
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109    605.1908
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452    578.5847
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445    355.5078
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626    450.7346
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014    271.4219
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476  11992.5273
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734    335.0773
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099   1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048   1550.6704
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312    593.7773
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948   1222.5433
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257    549.3252
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361   1011.0248
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849    764.8193
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624   1870.8307
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886  43249.3945
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012    305.3624
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595    0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573   1490.0312
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769