Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393   2081.0056
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633   -0.01813   1130.5791
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479    0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041    0.01741    746.4155
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046    0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333   -0.01767   3321.6113
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073    0.00174   2310.4871
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453    0.00257    357.1436
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730   -0.01491   1844.4353
#> 125   713992.5 9561062 1878.28 49.97109  0.00183    0.03387   1989.3012
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607    0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102    0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831   -0.00149  12632.7793
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966   -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903    0.01447    331.4833
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528    0.01392   1255.0884
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905   -0.01385    674.0721
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634   -0.02626   4316.0977
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151    0.00899   1394.2274
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856    0.01114   4510.8833
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108   -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206    0.02773    179.3077
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225    0.00325    616.8652
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452   -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205    0.02205    629.4165
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275    0.00605    798.6775
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184    0.03924   8023.8657
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743    0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804    0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047   -0.00887    366.8537
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606   -0.00556    660.3690
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140   -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509    0.00499    885.0064
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856    0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666   -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288    0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093   -0.00907   1943.4092
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267   -0.05063  18733.8848
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996    0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613    0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617    0.00307 147470.5312
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008    0.00102    483.4184
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759   -0.00400    608.6523
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119   -0.01560    977.4719
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140   -0.02330   1456.9246
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681   -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160    0.01690   5121.6924
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458   -0.00401   1722.6481
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090   -0.00299    688.0706
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736    0.01684   2556.8049
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568   -0.01558   1504.9717
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972    0.03512  13178.3223
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629   -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481   -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320    0.00180   3904.5254
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133    0.00367    806.0966
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370   -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017    0.01213    328.7922
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332   -0.00058   1817.2371
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259   -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038   -0.04212  46310.0195
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474    0.01334   4905.2002
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263    0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666   -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007   -0.02132   1580.1621
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428   -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924    0.02606    564.6876
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253    0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070    0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.35820  0.00540    0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669   -0.01981   3975.3608
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495   -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970    0.00200   1305.6691
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110   -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285   -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730   -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864    0.04226    338.3474
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124    0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488   -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893   -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115    0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770   -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454    0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465   -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195   -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025   -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251   -0.00389  14516.2002
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586    0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332    0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599    0.00301    338.9676
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465   -0.00105   4432.8760
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124    0.00204    853.2902
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363   -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458    0.00479    604.0975
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195   -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724    0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846    0.00794   1363.5321
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001    0.00759   1880.4403
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490   -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452   -0.01298    436.1459
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301   -0.01099   1146.2140
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393    0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785    0.00645    234.0114
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004    0.00714 480535.5000
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950   -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479    0.02251    168.1733
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165    0.00555   1882.0179
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820   -0.00260   1484.6593
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343    0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537   -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105   -0.01576   1045.1074
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129    0.00491    261.7371
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909   -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108   -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203    0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163   -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161    0.00709   1181.2555
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305    0.00585   5714.3267
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021    0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811   -0.02281    492.4785
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288    0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379   -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181   -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491