Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677    1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526     539.8119
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907    1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558    1504.9717
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456    1531.4274
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196    1052.5341
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603    1562.0503
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780     825.4451
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650     766.6254
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491