Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324     861.3537
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340     604.1032
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529    3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164     895.6606
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898     588.3761
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071    0.00150    1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598    0.01212     561.1948
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924    8023.8657
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181    2406.1252
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357     473.4709
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409    2312.8413
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656    3036.7383
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664    1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132   42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017    0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849    1120.5961
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010    3467.2690
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121    1783.0658
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020    0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715     520.7355
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466    2633.9666
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995    1026.7981
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069   -0.00900     614.5898
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033    3446.5520
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170     253.5129
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785    0.00645     234.0114
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148     481.8292
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491