Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526     539.8119
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974    2136.6939
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118   19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253     528.1924
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306    1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923     811.5580
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759    1880.4403
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033    3446.5520
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400    4126.7812
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931    1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250