Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806