Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681  -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.4766034   -0.01919  -0.04051 351155.1250
#> 31357 714752.5 9561022 1848.22 33.4464113   -0.00347   0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.6679573   -0.00292   0.00712   3636.5303
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677   1675.4679
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273  -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104  -0.02046   1024.1573
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068   0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387   1989.3012
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102   0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214   0.00466    724.0713
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194   0.00526    539.8119
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475   0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664  -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.6234412   -0.02169  -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164    895.6606
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270  -0.00769   1032.7177
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430  -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203  10263.2734
#> 26215 714332.5 9560262 2138.08 25.7212213   -0.00491   0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173   8967.1924
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987   0.02107   1392.5880
#> 5749  712852.5 9560162 1934.78 23.3887102    0.00071   0.00150   1274.3594
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559    0.01205   0.02205    629.4165
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383   0.00193   1809.3838
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887    366.8537
#> 11687 714552.5 9560382 2052.66 29.5755084   -0.00255  -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.8121470    0.00463   0.00167   1938.5389
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376   2048.3198
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672    565.1107
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442  -0.00178    800.1933
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617   0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.1730683    0.00359  -0.00279    930.0871
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395   0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286  -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.9680266    0.00107   0.02373    527.3528
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478   0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175   0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260  -0.03310   2833.9001
#> 30406 715312.5 9560152 2029.10 34.1373984    0.00953  -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924    0.03658   0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306  -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299    688.0706
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550  -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099   0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783    0.00406  -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195  -0.00656   3036.7383
#> 26010 712852.5 9560542 1928.25 30.9781091    0.00165   0.00555   4439.0967
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024  -0.00474   2343.2100
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188   0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469   0.02940    987.8085
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010    0.01865   0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017   0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458  -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205   1112.2626
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574    183.3220
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084  -0.02236   9755.3027
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175   1026.6818
#> 5911  713952.5 9561282 1801.16 33.7896130    0.00466  -0.01776   1021.8267
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247   0.01123    364.2903
#> 7485  714362.5 9558042 2484.02 47.9588593    0.03169   0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213  -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442  -0.00118  19557.6621
#> 10363 713252.5 9559722 1999.55 41.8299297   -0.03669  -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196   1052.5341
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317   0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495  -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705  -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292  -0.00109   2047.9031
#> 1997  714752.5 9561012 1853.19 26.5663341   -0.00693   0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016  -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267   0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110  -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285  -0.01325    595.0426
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610    454.8103
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212   0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.9395253    0.00902   0.01358    802.7622
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893  -0.00867   2247.8748
#> 33142 715652.5 9558212 2740.94 14.3812407   -0.00465  -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186   0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166   0.01476   1404.4454
#> 43536 713782.5 9558732 2281.50 32.7491217    0.00840   0.00189    556.5211
#> 6470  713742.5 9558802 2260.79 30.6801711    0.00000   0.00320    911.5911
#> 29431 714882.5 9559892 2258.03 21.9179275   -0.00849   0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009   0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439   -0.01433  -0.02177    997.3089
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085   2229.5105
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204    853.2902
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295   0.02715    520.7355
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815  -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165   0.00875   1165.3953
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195  -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.2571887    0.00075  -0.01945    259.7421
#> 39193 715332.5 9560852 1872.90 30.2905598   -0.00161   0.00171    756.8818
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818   0.00502   1428.8546
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405   0.00476  11992.5273
#> 4007  714132.5 9558412 2464.44 43.7201175    0.00174  -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.2169974    0.00973  -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338  -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563   -0.00846   0.00216   1412.2295
#> 42934 713832.5 9558672 2270.74 27.6119821   -0.03193  -0.00267  31588.1953
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983    270.7949
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041  -0.00011    201.1272
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338   0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.0747293   -0.00112  -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714 480535.5000
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948   1222.5433
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667   0.00863   1098.3462
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154  -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229   0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528   -0.00070  -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497   -0.00930  -0.00360   4288.7329
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804  -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873  -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262  -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247    282.1076
#> 23499 714272.5 9558232 2508.35 28.9676002   -0.00820  -0.00260   1484.6593
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607  -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013   -0.00598   0.02208    842.4825
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523    888.2811
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284  -0.00886  43249.3945
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196  -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508  -0.00177    184.1042
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014  -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.6572738   -0.01305   0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732   0.01858    298.4490
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716  -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253   1237.0536
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336   0.01403    609.8665
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983