Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357    1752.0375
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712    3636.5303
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752    1321.5997
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292    1233.9299
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542     674.0147
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149   12632.7793
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189     287.5425
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164     895.6606
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203   10263.2734
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114     150.1498
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167    1938.5389
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338   10787.6299
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169     350.1799
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530     937.1306
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319    3457.7519
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401    1722.6481
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055     773.0359
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870    1895.1399
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878 -0.04351    1442.9247
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322    4605.1411
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190     264.6404
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477    2945.2100
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000  162178.1562
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252    3631.7017
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665    1904.4769
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642    1688.1782
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403     241.2991
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466    2633.9666
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500     538.8209
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333     422.8840
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978  0.00367    3002.8828
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033    3446.5520
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011     201.1272
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032     994.9224
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312     254.0656
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215     917.5296
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935     903.6397
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983