Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109   2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150    503.5590
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290   2481.5632
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367   3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035   1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462    716.5762
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491