Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250