Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.85782   -0.01106   -0.04634    500.5027
#> 37912 712802.5 9559952 1838.40 52.10134    0.00183   -0.09203    634.3320
#> 42733 715112.5 9559062 2399.70 36.05452   -0.00041    0.01741    746.4155
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514   -0.00644   1179.0695
#> 20799 714792.5 9561002 1857.29 23.12916    0.00090    0.00890   2301.0454
#> 38411 713552.5 9560652 1851.63 46.00507   -0.00891   -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.06326    0.00940    0.00679    550.0344
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730    0.01770    474.5059
#> 22814 713862.5 9558582 2327.48 48.61203   -0.02894   -0.03416    947.4878
#> 10974 715312.5 9557502 2844.23 26.66488   -0.00831   -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.23560   -0.03021   -0.09328   2235.5259
#> 44328 713222.5 9558932 2227.94 29.15553   -0.00769   -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.99185    0.01581    0.01189    287.5425
#> 43007 713372.5 9559192 2120.47 46.00507    0.03696    0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.00467    0.00129    0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.63795   -0.00270   -0.00769   1032.7177
#> 30125 715162.5 9558992 2438.18 39.85666    0.00471   -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.17312   -0.01029   -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.25255   -0.04430   -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.18133   -0.00382    0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.04492    0.00832    0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.53503    0.00856    0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.53697   -0.00043    0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.23440    0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.52897   -0.03613   -0.01898    588.3761
#> 10320 714912.5 9558482 2724.21 25.28291    0.00356    0.00774    527.1068
#> 5749  712852.5 9560162 1934.78 23.38871    0.00071    0.00150   1274.3594
#> 27957 715012.5 9559332 2306.35 20.48038    0.00225    0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804   -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.80394   -0.01097    0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.48403   -0.00064   -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.66450    0.06765    0.00114    150.1498
#> 34642 714962.5 9559862 2282.08 22.16086   -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.15647   -0.01743    0.01313    783.4548
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224    0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.81215    0.00463    0.00167   1938.5389
#> 4605  714852.5 9557892 2677.78 26.93360    0.00321    0.01169    350.1799
#> 18751 714882.5 9559182 2381.81 39.29918    0.00066   -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.19146    0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.64579    0.00060    0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.28922    0.00006    0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.59979   -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.93664   -0.02617    0.00307 147470.5312
#> 39599 714112.5 9558562 2374.25 29.16814    0.00283    0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.48806    0.02409    0.02771    674.3859
#> 22299 713532.5 9560362 1975.58 27.70996    0.00478    0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.99491    0.00175    0.00236   2715.4138
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762    0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.26460   -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.17944   -0.02797   -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.64888   -0.01282    0.00542    860.8840
#> 35083 714612.5 9558152 2596.12 60.59831    0.04548    0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.56253   -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.21952   -0.00034    0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.68860    0.00114    0.03026    247.0894
#> 664   714452.5 9559852 2174.83 44.86088    0.00282    0.00109    767.4097
#> 27695 714142.5 9558452 2449.28 20.49012   -0.00539    0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104    0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.62441   -0.01972    0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.00795   -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238   -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478   -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165    0.00555   4439.0967
#> 12137 712842.5 9558902 2097.88 30.09862    0.00024   -0.00474   2343.2100
#> 39722 715022.5 9557712 2735.53 30.26077    0.00133    0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.03442    0.00188    0.00832   3022.3462
#> 32879 713912.5 9559672 2238.00 31.97677   -0.00278   -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.22372   -0.03986   -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.49091   -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.75123   -0.02556   -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.05243   -0.04989   -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.55068    0.00521   -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.10642    0.00516    0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.09863   -0.13038   -0.04212  46310.0195
#> 4588  713852.5 9558612 2308.97 52.18041   -0.01059   -0.07431   4079.0366
#> 27915 715232.5 9559552 2208.93 25.48058    0.00171   -0.00261   1627.8989
#> 19832 715182.5 9557662 2761.27 38.43286   -0.00666   -0.03583   1898.0105
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511   -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.23905    0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.05925    0.02213   -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.80395   -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.79288   -0.00246    0.00396    704.7076
#> 10617 713232.5 9560722 1879.95 26.49701    0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.43344    0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.40342    0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281    0.04759    158.9370
#> 49378 713252.5 9560732 1875.41 27.83544    0.00181   -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.27110    0.00705   -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.32798   -0.00292   -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.04632   -0.08153   -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.65787   -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.97833    0.01364   -0.00094   1392.6008
#> 38103 715392.5 9560162 1998.16 46.28582    0.00708   -0.02538    807.8237
#> 28588 715532.5 9559362 2299.84 11.14460   -0.04360   -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.93953    0.00902    0.01358    802.7622
#> 17674 713882.5 9558562 2339.20 41.03638   -0.03675   -0.01905   6487.0010
#> 14140 714672.5 9559262 2326.36 45.79251   -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.95483   -0.00770   -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.98938   -0.00983    0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.38124   -0.00465   -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186    0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.46298   -0.00195   -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585   -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.68017    0.00000    0.00320    911.5911
#> 7317  713682.5 9558582 2347.46 28.34766   -0.02196    0.00336   2806.2644
#> 16464 715252.5 9559982 2143.24 23.18588   -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.81713   -0.00162   -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.15324    0.00156   -0.00116   1817.5222
#> 26023 715832.5 9557632 3097.56 41.71018    0.00133   -0.00452    578.5847
#> 37832 712552.5 9559192 1998.10 41.55090   -0.00195   -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.25719    0.00075   -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.87076    0.00321    0.00548    532.8210
#> 39796 714802.5 9560372 2046.48 40.15575    0.00176    0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.80830    0.00818    0.00502   1428.8546
#> 39685 715332.5 9558102 2797.04 47.78812    0.00167    0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.49622   -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.70407    0.00001    0.00759   1880.4403
#> 43403 712632.5 9559262 2072.53 34.48633   -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.27967    0.00084   -0.00083   1309.2347
#> 44668 712612.5 9560472 1972.24 27.50713    0.00601   -0.00011    712.5228
#> 37029 713022.5 9559082 2163.08 36.03217    0.00490   -0.00780    825.4451
#> 33175 715822.5 9558832 2563.13 36.62289   -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.68420    0.02218    0.00902    218.7467
#> 14148 714162.5 9560352 2149.57 37.13339    0.00424    0.00196    576.1450
#> 33666 713632.5 9561012 1775.88 21.17881   -0.00950   -0.05130    107.5831
#> 41855 713182.5 9560452 1851.53 45.17830    0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.98875   -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.23024   -0.00233    0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.64905   -0.00930   -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.39369   -0.02608    0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.71778    0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.29457    0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.44717    0.00360    0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.56845    0.00420    0.01960    884.2003
#> 26283 715932.5 9557592 3075.48 32.76688   -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.46513    0.00105   -0.01576   1045.1074
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909   -0.01151 143213.1562
#> 16443 714472.5 9559532 2222.55 16.74412    0.03263   -0.02012    305.3624
#> 9320  713872.5 9559262 2378.76 29.85855    0.00309    0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.65727   -0.01305    0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732    0.01858    298.4490
#> 14645 715072.5 9560262 2034.04 38.16930   -0.00716   -0.03234   1707.3130
#> 32618 715132.5 9559642 2266.36 34.53274   -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.62044   -0.00241    0.00020    673.2861
#> 6206  714122.5 9560992 1917.28 37.96991   -0.01023   -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.34975   -0.00771    0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491