Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919   -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958    0.02642    671.1807
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347    0.02357   1752.0375
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896    276.0310
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514   -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767   3321.6113
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035   2697.5974
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024    394.8751
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412    0.00898    458.6938
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017   -0.00032    994.6740
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173   8967.1924
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898    588.3761
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452   -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193   1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318    289.0063
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224    0.00177    291.5937
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140   -0.00070   2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338  10787.6299
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279    930.0871
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395    0.01495    266.5888
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092    412.5849
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357    473.4709
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681   -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221    0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739    365.7626
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099    0.02169    495.8840
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643    670.2289
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611    0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213    328.7922
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524  10861.9482
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556   -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474    0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384    0.00175   1026.6818
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522   1355.7436
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213   -0.01363    565.5207
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070    0.00010   3467.2690
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196   1052.5341
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970    0.00200   1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450   2589.9910
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216   -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252   3631.7017
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115    0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770   -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186    0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427    0.02403    474.8613
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614   -0.00093    118.1402
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067   1123.7318
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463   1050.7473
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132    737.5982
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195   -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199    836.4196
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833    300.0125
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367   3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399   1601.4852
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990    0.02790    209.6952
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432   -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483   1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687   -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036    476.9504
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714 480535.5000
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418    0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.060776    0.01142    0.00868    235.8511
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420    0.01960    884.2003
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948   1538.1741
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595    0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992   -0.01052    451.7209
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962   -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241    0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751    0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603