Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603