Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029    5577.3916
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712    3636.5303
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741     746.4155
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752    1321.5997
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046    1024.1573
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730 -0.01491    1844.4353
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406    2421.1145
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120     512.9740
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518     854.4269
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581    1855.1630
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284    2010.4285
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678    1891.6584
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917     116.1986
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212     561.1948
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924    8023.8657
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887     366.8537
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034    4495.6929
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167    1938.5389
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511  361979.2812
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373   16031.7500
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270    4057.5637
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.84652 -0.00585 -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530     937.1306
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373     527.3528
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565    2232.6133
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181   11655.4316
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401    1722.6481
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009    1247.7061
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262    1760.2162
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512   13178.3223
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112     745.3315
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.95066  0.03743 -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190     640.8842
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849    1120.5961
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161    3065.3694
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118   19557.6621
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190     264.6404
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133    2397.2681
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094    1392.6008
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875     597.2893
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856    1789.1596
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288    2144.8164
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306    1718.3652
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000  0.00320     911.5911
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230     714.7432
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116    1817.5222
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118    2273.4758
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464    4029.5947
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923     811.5580
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333     422.8840
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216    1412.2295
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697     412.1185
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175     725.1355
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376     629.8236
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528     355.3475
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948    1222.5433
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099     909.0560
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165     805.1403
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499     617.3855
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578    1258.8865
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368    1071.8734
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948    1538.1741
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119     792.3544
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585    5714.3267
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249     148.1857
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057     722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775