Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199    0.00659   3553.7166
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046    0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514   -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191   -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503    0.02433   1475.3191
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010   1527.1028
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284    0.01804    628.5153
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223    0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607    0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214    0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021   -0.09328   2235.5259
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903    0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475    0.00665  25318.3555
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126   -0.02044 129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832    0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151    0.00899   1394.2274
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059   -0.00628    556.0121
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108   -0.00678   1891.6584
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598    0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937    0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252    0.01118   2359.5059
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616    0.01276   9614.6924
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543   -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047   -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355  14919.6260
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862    874.5408
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527    0.04613    289.6113
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492    0.00072  54831.6289
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779   2972.0193
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400    608.6523
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399   -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175    0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504    0.05825    162.6688
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329   -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797   -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708    0.00408  76318.0156
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736    0.01684   2556.8049
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304   1569.7313
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282    0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406   -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927   -0.00316    454.9674
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238   -0.00851   1012.5963
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228   2262.3796
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188    0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370   -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458   -0.14813 287033.4688
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556   -0.11694 102301.4609
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310   -0.00860    695.2302
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878   -0.04351   1442.9247
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166    0.01476   1404.4454
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253    0.00753   1337.7854
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258    0.00422   1159.2100
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759    158.9370
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705   -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605    0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767   -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016   -0.00294    734.9908
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526    706.1463
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445    0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181   -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124    0.01166   1813.7841
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020    0.00279  39495.1797
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895    0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893   -0.00867   2247.8748
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486    0.02306   1718.3652
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025   -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783   -0.00367   1966.2112
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614   -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382    100.0000
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284    0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390   -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007   -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738    0.00118   2273.4758
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165    0.00875   1165.3953
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195   -0.00995   1026.7981
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161    0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937   -0.02273  72123.6172
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341    0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978    0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538   -0.00052   3508.5591
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990    0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457    0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601   -0.00011    712.5228
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490   -0.00780    825.4451
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218    0.00902    218.7467
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224   -0.02036    476.9504
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785    0.00645    234.0114
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586    0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035    0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310   -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154   -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858    0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009   -0.00599  12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980   -0.00710    472.0236
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873   -0.01417   1520.6630
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273   -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343    0.01177   1558.2231
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361   -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293    0.00186   2751.4626
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233    968.7780
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163   -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709   1181.2555
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305    0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983    0.00803   1058.5630
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458    485.7418
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463    0.00253   1237.0536
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250