Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501   -0.00649    1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896     276.0310
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579    2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174    2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471   -0.01382     560.8309
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838     238.6685
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678    1891.6584
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694    2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146    1347.6224
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294   -0.00096     520.5570
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332    1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.860876  0.00282    0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520    3390.5708
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759     158.9370
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224    1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294     734.9908
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185     953.8614
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158     837.3730
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208     951.3401
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697     412.1185
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105     868.7802
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750    1892.8978
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607