Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649   1399.2329
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382    560.8309
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373  16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058   1817.2371
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911    173.2806
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801   1071.6141
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957    495.4453
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230    714.7432
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759   1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900    614.5898
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033   3446.5520
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361   1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047   2826.7700
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208    842.4825
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573   1490.0312
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806