Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034   -0.01919  -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.9766069    0.00958   0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183  -0.09203    634.3320
#> 16749 715072.5 9558242 2745.27 49.0537816    0.00097   0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251    0.00479   0.01261    323.8441
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498  -0.00752   1321.5997
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090   0.00890   2301.0454
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645   0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922   0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783    377.8918
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894  -0.03416    947.4878
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021  -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194   0.00526    539.8119
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952   -0.03525  -0.11024    394.8751
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164    895.6606
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905  -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.0602377    0.00672   0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051   0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733    598.7983
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257   0.00593    939.0466
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083    994.7876
#> 1635  713102.5 9559052 2172.87 31.4840308   -0.00064  -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948    0.01275   0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089  -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593    0.01414   0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.0547529   -0.03184   0.03924   8023.8657
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274   -0.00669  -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.1929823   -0.01846  -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323    183.1836
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229  -0.00511 361979.2812
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379   0.01338  10787.6299
#> 46201 714832.5 9561022 1845.94 34.5739922   -0.00856   0.01796   4729.6382
#> 46171 712822.5 9558932 2089.30 29.8545389   -0.01174   0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288   0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.6425796    0.01843   0.04407    344.8395
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302   -0.01149   0.01949   1182.1506
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290   0.00230   1056.5876
#> 9704  714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400    608.6523
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478   0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175   0.00236   2715.4138
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260  -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.2645950   -0.00211   0.00241    585.6182
#> 2595  714912.5 9557662 2700.72 24.1948618   -0.02681  -0.01319   3457.7519
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306  -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.3468842   -0.00161  -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576  -0.00166    650.6246
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739    365.7626
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856  -0.01625   1063.9150
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550  -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168  -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195  -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.2357795    0.00927   0.01643    670.2289
#> 24501 713302.5 9559142 2125.66 27.1089251    0.01784   0.00596    842.0994
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394  -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091    0.00165   0.00555   4439.0967
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228   2262.3796
#> 1973  715252.5 9557822 2861.00 15.5707010    0.01865   0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246   0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986  -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376   2048.3198
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129   0.00171    788.0004
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781  29995.1543
#> 20797 713012.5 9560392 1842.25 36.6498183    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506   -0.02070  -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017    759.9937
#> 22907 713172.5 9559612 2056.81 22.2725884   -0.00166   0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542  -0.00722   1373.4147
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193   0.02812   1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298   1533.7925
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442  -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253   0.00753   1337.7854
#> 10617 713232.5 9560722 1879.95 26.4970062    0.00258   0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384   -0.01102  -0.01778   4594.4351
#> 49378 713252.5 9560732 1875.41 27.8354356    0.00181  -0.00001   1184.7952
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678   -0.01341   0.00171   4425.5410
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915  -0.02705    413.5702
#> 38103 715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157  -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.0470808    0.00216  -0.02606    919.4241
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144  -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004  -0.02384    947.2075
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234    849.0547
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166   0.01476   1404.4454
#> 3771  714842.5 9561152 1724.51 45.5398315    0.02405  -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025  -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256   -0.00783  -0.00367   1966.2112
#> 6470  713742.5 9558802 2260.79 30.6801711    0.00000   0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781  29995.1543
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534   0.01476   1275.9446
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298  -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204    853.2902
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133  -0.00452    578.5847
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409   1575.9425
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544  -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818   0.00502   1428.8546
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527  -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.1066333   -0.00007  -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001   0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338  -0.00399   1601.4852
#> 8831  714372.5 9559852 2200.77 48.7604272   -0.00457   0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737   0.01033   3446.5520
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180   0.01170    253.5129
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452  -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983    270.7949
#> 12044 713612.5 9559292 2247.76 23.0747293   -0.00112  -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978   0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948   1222.5433
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066   0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418   0.00513    161.9878
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.9676002   -0.00820  -0.00260   1484.6593
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082   0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952  -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833   -0.00375   0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.1104343   -0.01502   0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.7218013   -0.00598   0.02208    842.4825
#> 4269  714162.5 9560572 2087.45  7.3344327    0.00293   0.00186   2751.4626
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172  -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014  -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595   0.03064   1485.5201
#> 36141 714222.5 9559422 2316.64 16.2771580   -0.07627  -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282   0.00458    485.7418
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811  -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716  -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305  -0.01135   1293.4855
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181  -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983