Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487   1077.7585
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993  10308.6406
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212  46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052   3508.5591
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018    937.6911
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491