Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603