Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660   -0.01919  -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.97661    0.00958   0.02642    671.1807
#> 44302 714932.5 9557982 2649.56 37.30070    0.01633  -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.65267    0.00221   0.00969    368.5606
#> 25090 715362.5 9560102 2059.29 49.11967    0.02059  -0.00628    556.0121
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514  -0.00644   1179.0695
#> 21939 713642.5 9558712 2299.09 33.75065   -0.00503   0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.18041   -0.01059  -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.74185    0.00273  -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.95254   -0.01073   0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.43573    0.00250  -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.06326    0.00940   0.00679    550.0344
#> 38827 713412.5 9560472 1869.10 16.87132   -0.00156   0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068   0.00588   1295.1610
#> 1393  714832.5 9557902 2664.72 31.29553    0.00550  -0.00120    512.9740
#> 10974 715312.5 9557502 2844.23 26.66488   -0.00831  -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.23560   -0.03021  -0.09328   2235.5259
#> 30800 715182.5 9557582 2772.37 39.20406   -0.02475   0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.15327   -0.07664  -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.00507    0.03696   0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.31922   -0.02678  -0.04272   4955.5586
#> 22933 712712.5 9560432 1902.17 38.12347   -0.00746   0.00096   3210.5781
#> 31576 713942.5 9558492 2387.12 46.14316   -0.05634  -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440   0.00710    843.7961
#> 25081 713392.5 9558452 2242.24 46.53850   -0.02716   0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.24491    0.00051   0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.70789   -0.00077  -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.93953    0.00902   0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.23440    0.00993  -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.25769   -0.02651  -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.54285    0.03206   0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.02912   -0.00521  -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.28291    0.00356   0.00774    527.1068
#> 13847 715052.5 9559232 2337.87 26.11484   -0.01452  -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.49606    0.01205   0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804  -0.01596   5296.9014
#> 27013 713062.5 9559782 1976.81 41.63226    0.01414   0.03416   1204.8629
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543  -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.69432   -0.01545   0.02405   5681.2412
#> 47351 714222.5 9558792 2420.94 28.66852   -0.09566  -0.03355  14919.6260
#> 24100 715132.5 9558992 2420.18 29.21397   -0.03229  -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.48099   -0.00509   0.00499    885.0064
#> 18564 715552.5 9558592 2648.36 26.29131    0.00316   0.00714    475.3120
#> 46171 712822.5 9558932 2089.30 29.85454   -0.01174   0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.48766   -0.00093  -0.00907   1943.4092
#> 39278 714692.5 9561052 1830.60 48.39488   -0.01996   0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.85173   -0.01149   0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.55527   -0.04492   0.00072  54831.6289
#> 39599 714112.5 9558562 2374.25 29.16814    0.00283   0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.34897   -0.10372  -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.72526   -0.01399  -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.02089    0.00194   0.00516    576.0335
#> 462   713712.5 9561182 1785.08 47.71707   -0.03260  -0.03310   2833.9001
#> 48808 715032.5 9557792 2707.02 25.43761   -0.02329  -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.95442   -0.04708   0.00408  76318.0156
#> 38862 712952.5 9559892 1950.86 37.72182   -0.00048  -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.64888   -0.01282   0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.83129    0.03658   0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.21228   -0.01160   0.01690   5121.6924
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209   0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.91186   -0.00032   0.00542   3614.7334
#> 32280 712572.5 9560182 1956.86 50.67812   -0.02554   0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.62725   -0.01168  -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.27489   -0.00845   0.00454   3476.0818
#> 16355 712962.5 9560212 1992.88 39.52607    0.01927  -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104   0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.05914   -0.00410   0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.62441   -0.01972   0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.77038    0.00904  -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478  -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165   0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.40821    0.00291  -0.00081    940.6002
#> 41721 715182.5 9557562 2784.85 31.78999   -0.00588   0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.36887   -0.00093   0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.95066    0.03743  -0.05242    507.9263
#> 18958 714192.5 9558372 2495.72 34.67712   -0.00178  -0.00822    472.8748
#> 47320 713422.5 9559182 2143.03 56.32519    0.01246   0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.97677   -0.00278  -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.81517   -0.00332  -0.00058   1817.2371
#> 30287 713992.5 9557822 2403.92 18.21547    0.03890  -0.00911    173.2806
#> 41125 712772.5 9559412 2021.88 38.87633   -0.00310  -0.00860    695.2302
#> 20797 713012.5 9560392 1842.25 36.64982    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005   -0.02143  -0.05078 622888.2500
#> 28312 715712.5 9558032 2783.23 24.25273   -0.00007  -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.78961    0.00466  -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.91167    0.00707   0.01833    369.3771
#> 12629 714712.5 9558872 2533.15 37.02568    0.04250  -0.00970   1443.9200
#> 19756 715572.5 9558972 2485.95 38.81273    0.00546   0.01484    495.0292
#> 10692 712712.5 9561022 2005.85 41.43344    0.01410  -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.68934   -0.01102  -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281   0.04759    158.9370
#> 1916  715202.5 9559512 2213.91 26.32798   -0.00292  -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.30490   -0.00605   0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.95926   -0.00767  -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.21550   -0.01016  -0.00294    734.9908
#> 44612 714932.5 9558822 2578.80 38.25525    0.01257   0.00044    288.5468
#> 26058 713402.5 9560442 1880.12 31.37746    0.00730  -0.02610   1872.9276
#> 16065 715242.5 9560622 1871.61 34.30642   -0.00019   0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.39754    0.00143   0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.46494    0.00560   0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.44550    0.02451  -0.03921    525.3245
#> 13278 712752.5 9560452 1901.25 42.11354   -0.00118  -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.68593   -0.00488  -0.00632    620.0229
#> 14140 714672.5 9559262 2326.36 45.79251   -0.01400  -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.52257   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186   0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.70001    0.00683  -0.00603   1562.0503
#> 29431 714882.5 9559892 2258.03 21.91793   -0.00849   0.00150    503.5590
#> 34250 713822.5 9559142 2316.36 33.75982   -0.02284   0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.43175    0.00390  -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.25369   -0.00124   0.00204    853.2902
#> 7517  714002.5 9561142 1772.85 26.18933    0.00815  -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.76592   -0.00458   0.00479    604.0975
#> 9665  713362.5 9558812 2275.28 31.49492    0.01306  -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.44473    0.02309  -0.03409   1575.9425
#> 47005 714942.5 9559192 2333.66 29.22085   -0.04937  -0.02273  72123.6172
#> 45071 712652.5 9560032 1900.80 14.47349    0.00987  -0.00507    504.3085
#> 48183 713642.5 9558262 2314.44 50.02323    0.01338  -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.06214    0.00196   0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.94166   -0.00065   0.00475    515.1661
#> 8413  713242.5 9561232 1967.45 38.19451    0.00236   0.00483   1097.3082
#> 33370 714202.5 9558962 2472.93 32.64141   -0.00307   0.02997    230.8641
#> 29107 713172.5 9559272 2067.58 28.26401    0.02413   0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.04765    0.02785   0.00645    234.0114
#> 44267 712972.5 9559042 2191.01 22.26915    0.01462   0.00268    252.0072
#> 33666 713632.5 9561012 1775.88 21.17881   -0.00950  -0.05130    107.5831
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097   0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.18382    0.00270  -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.33104    0.01154  -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.16621    0.01923   0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 3724  714322.5 9557402 2603.46 34.59748    0.02523   0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.56845    0.00420   0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.69410    0.00944   0.00836    250.5111
#> 38362 714402.5 9558042 2499.31 56.90101    0.03044  -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.63451    0.00082   0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.03698    0.00343   0.01177   1558.2231
#> 18246 713902.5 9561342 1818.52 19.11043   -0.01502   0.00282   1659.6211
#> 10127 713722.5 9558712 2311.10 19.48114    0.01129   0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.88684    0.01572   0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.81138   -0.04770   0.00270  35950.5508
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732   0.01858    298.4490
#> 24962 715152.5 9560472 1967.37 28.84556    0.00298   0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.34593    0.00811  -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.45004   -0.00087  -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.16930   -0.00716  -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086   -0.00011  -0.00419    323.0790
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603