Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712    3636.5303
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804     628.5153
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120     512.9740
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024     394.8751
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769    1032.7177
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629  360379.1875
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146    1347.6224
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169     350.1799
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949    1182.1506
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656    3036.7383
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322    4605.1411
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426    1315.5334
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152    1530.5657
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252    3631.7017
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867    2247.8748
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517    1060.5494
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945     259.7421
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794    1363.5321
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475     515.1661
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175     725.1355
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955     713.4463
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251     168.1733
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502     520.5218
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050    1471.4346
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709    1181.2555
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157     766.0220
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491