Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603