Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333     258.6537
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324     861.3537
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005     769.3426
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328    2235.5259
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766  140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114    4510.8833
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642     613.6390
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714     475.3120
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907    1943.4092
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009    1247.7061
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543     352.1558
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555    4439.0967
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017     759.9937
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477    2945.2100
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732    4237.6934
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279   39495.1797
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063    2430.2278
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389   14516.2002
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181    1286.3949
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419     761.5427
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363 -0.00363     389.2691
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333     422.8840
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048    1550.6704
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238     787.0053
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603