Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59291   300.00   TRUE       300.00            100  3.318273
#> 42632 20.47808   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 24129 24.12897   300.00   TRUE       300.00             89  2.886120
#> 44952 34.08239   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80357   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781   300.00   TRUE       300.00              2  2.552843
#> 20151 44.79556   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03598   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94657   162.65   TRUE         0.00             77  3.703905
#> 125   29.59671   207.04   TRUE        21.03            100  3.298701
#> 1393  23.41220   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32283   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862   300.00   TRUE       300.00             85  3.366461
#> 2990  25.32130   300.00   TRUE       123.56              5  2.520462
#> 43007 38.96170   300.00   TRUE       300.00             37  2.473902
#> 48055 33.82571   300.00   TRUE         1.67              0  3.566570
#> 38279 37.62843   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97143   300.00   TRUE       300.00            100  5.111976
#> 27978 14.32853   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 47210 26.91527   300.00   TRUE       300.00            100  2.777281
#> 40828 25.59975   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50753   300.00   TRUE       300.00            100  5.556760
#> 13585 47.06734   300.00   TRUE       300.00            100  2.769655
#> 4477  32.16986   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592   300.00   TRUE       300.00             30  2.721899
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91527   300.00   TRUE       300.00            100  2.777281
#> 46125 25.65705   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 47351 34.87881   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09767   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941   300.00   TRUE       300.00              2  2.262887
#> 20001 34.67712   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919   300.00   TRUE       300.00            100  3.408707
#> 27084 21.67270   300.00   TRUE       300.00             81  3.000309
#> 47091 40.15174   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74185   300.00   TRUE       300.00            100  2.640875
#> 13134 37.64390   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61065   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43707   300.00   TRUE       300.00            100  3.023906
#> 21952 40.09158   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518   300.00   TRUE       300.00             18  2.722101
#> 22299 21.71911   300.00   TRUE       127.54              2  2.615513
#> 15766 23.33714   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95175   300.00   TRUE       300.00              6  2.211304
#> 30406 43.10075   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18782   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84251   300.00   TRUE       300.00             54  2.563199
#> 35083 43.54078   300.00   TRUE       300.00             23  2.546735
#> 20065 30.38968   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96320   300.00   TRUE        13.47             64  3.177528
#> 38197 34.34252   105.52   TRUE        65.32             50  3.036108
#> 28046 19.59115   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261   300.00   TRUE       300.00            100  3.079714
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 20601 37.94871   300.00   TRUE       300.00            100  3.059843
#> 18897 31.78827   300.00   TRUE        20.00              5  3.919162
#> 39722 26.54170   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73174    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59617   300.00   TRUE       300.00             29  2.769292
#> 32879 29.73823   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15422   300.00   TRUE       300.00            100  4.035908
#> 10073 25.76133   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34442   300.00   TRUE       166.26            100  3.803001
#> 7     13.22043   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840   300.00   TRUE       300.00             65  2.238750
#> 37720 42.77359   300.00   TRUE       300.00            100  3.690657
#> 19832 36.63378   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59785   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342    99.29   TRUE         0.00              6  3.137802
#> 28009 27.01782   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407   300.00   TRUE       300.00             37  2.897494
#> 47298 25.33619   300.00   TRUE       300.00            100  3.126386
#> 19756 30.11638   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79821    17.89   TRUE       145.83              4  3.064162
#> 16071 29.62135   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99416    22.68   TRUE       134.63             25  3.073643
#> 48650 27.05220   300.00   TRUE       300.00            100  3.027918
#> 39405 27.72027   300.00   TRUE       300.00            100  3.413298
#> 42617 31.23193   300.00   TRUE       300.00            100  2.848895
#> 10456 21.66869    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297   300.00   TRUE       300.00            100  2.929656
#> 7771  16.57911   300.00   TRUE        35.59            100  3.258586
#> 46038 29.73078   300.00   TRUE       243.86            100  3.184852
#> 24532 32.54629   300.00   TRUE         7.49            100  6.276647
#> 9864  20.82071   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34574   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09480   300.00   TRUE       300.00              8  2.745482
#> 31253 38.66835   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40477   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55835   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83487   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22581   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620   181.84  FALSE         0.00            100  2.854150
#> 2664  39.29115   300.00  FALSE       300.00            100  2.881694
#> 35852 36.36219   300.00  FALSE       300.00            100  3.356690
#> 26023 45.07287   300.00  FALSE       300.00            100  2.762367
#> 49363 32.95768   300.00  FALSE         0.00            100  3.394725
#> 19493 18.57243   300.00  FALSE       291.23            100  2.550849
#> 13512 35.93477   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002   300.00  FALSE       117.12            100  2.879028
#> 16878 25.38891   300.00  FALSE       300.00             18  3.154988
#> 32849 29.55202   300.00  FALSE       300.00             45  3.149905
#> 38689 36.08431   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22716   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07246   300.00  FALSE       300.00            100  3.167901
#> 8895  38.07763   300.00  FALSE       300.00             90  2.432641
#> 39440 32.70615   300.00  FALSE       300.00             92  2.369237
#> 7553  39.77129   300.00  FALSE       300.00            100  2.739830
#> 7983  38.50620   300.00  FALSE        57.15             22  2.884583
#> 40909 22.87820   234.82  FALSE         0.00             43  4.109707
#> 1281  35.39389   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225   123.04  FALSE        75.11              0  3.182033
#> 28286 17.04836   300.00  FALSE       300.00            100  2.790556
#> 48184 30.95634   300.00  FALSE       300.00             75  2.485742
#> 2999  45.58395   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02070   300.00  FALSE       300.00             55  4.224567
#> 15138 20.64596   116.69  FALSE        19.61             40  4.109325
#> 10981 29.16642   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81463   300.00  FALSE        46.28            100  3.237445
#> 14881 26.85625   300.00  FALSE       300.00            100  2.948550
#> 46148 38.91071   300.00  FALSE       300.00            100  3.015637
#> 32342 37.61984   300.00  FALSE       300.00             81  2.961658
#> 15880 48.89679   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39465   300.00  FALSE       300.00            100  3.092389
#> 11025 36.81999   300.00  FALSE       300.00             50  2.828200
#> 30296 40.15861   300.00  FALSE       300.00             68  3.258250