Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642     671.1807
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056 -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659    3553.7166
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250    1413.3312
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345    1120.6929
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547     519.4228
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677    1343.5006
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160    9344.2930
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328     592.3326
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672    1114.8801
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899    1394.2274
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973     514.5488
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678    1891.6584
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739     558.2286
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626    1014.7159
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924    8023.8657
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313     783.4548
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862     874.5408
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796    4729.6382
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058     431.4362
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279     930.0871
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588     749.2502
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408   76318.0156
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121    4159.6738
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542    3614.7334
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553    1750.4175
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316     454.9674
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961    1086.6963
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474    2343.2100
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574     183.3220
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860     695.2302
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078  622888.2500
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261    1627.8989
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833     369.3771
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123     364.2903
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010    3467.2690
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380     601.0843
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171    4425.5410
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526     706.1463
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685    2568.4631
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632     620.0229
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585 -0.02665    1904.4769
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996    1599.3491
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105    4432.8760
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181    1286.3949
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538     265.4677
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875    1165.3953
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445     355.5078
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409    1575.9425
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199     836.4196
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476   11992.5273
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083    1309.2347
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267   31588.1953
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902     218.7467
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038    2017.3547
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060     383.3547
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710     472.0236
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079    1647.6451
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875    2156.9358
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841     756.0055
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238     787.0053
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119     792.3544
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172     351.5812
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249     148.1857
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983