Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921   -0.12438    753.6304
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292    0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075    0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917    0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863   -0.00507 159065.0781
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240   -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677   1675.4679
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104   -0.02046   1024.1573
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951   -0.00529   3282.1186
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068    0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183    0.03387   1989.3012
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550   -0.00120    512.9740
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214    0.00466    724.0713
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542    0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769   -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581    0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664   -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528    0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328    592.3326
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131    0.00501   3889.7292
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672    0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440    0.00710    843.7961
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902    0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643    0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521   -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694   2123.4473
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064   -0.00916   2009.1686
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193   -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416   1204.8629
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545    0.02405   5681.2412
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255   -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606   -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463    0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204   -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316    0.00714    475.3120
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527    0.04613    289.6113
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025   4425.5151
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626   1014.7159
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283    0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008 825875.6875
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002    0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194    0.00516    576.0335
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329   -0.01181  11655.4316
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140   -0.02330   1456.9246
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090   -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529    0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209    0.02739    365.7626
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568   -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168   -0.04102  35685.0000
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800   -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104    0.01664   1898.5588
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081    940.6002
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588    0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940    987.8085
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129    0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332   -0.00058   1817.2371
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241   1752.9242
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953   -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949    0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143   -0.05078 622888.2500
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354    0.00245    629.5983
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379    0.00801    172.7491
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511   -0.01801   1071.6141
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213   -0.01363    565.5207
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540    0.00879   1200.4210
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200   1305.6691
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133   2397.2681
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881    0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257    0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526    706.1463
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560    0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622    0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354    0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675   -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144   -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115    0.01015    790.2421
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025   -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251   -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519    0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989   -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382    100.0000
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849    0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002   -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465   -0.00105   4432.8760
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738    0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402   -0.01328   1066.7134
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197    0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220   -0.03290   2481.5632
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540    0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464   4029.5947
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502   -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846    0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263    0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341    0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527   -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987   -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538   -0.00052   3508.5591
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785    0.00645    234.0114
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452   -0.00948   1222.5433
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066    0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418    0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097    0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980   -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804   -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236   -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273   -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952   -0.00438   1276.5448
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502    0.00282   1659.6211
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203    0.00238    787.0053
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595    0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305    0.00585   5714.3267
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811   -0.02281    492.4785
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241    0.00020    673.2861
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200