Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976    301.2347
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292    0.00712   3636.5303
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090    0.00890   2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550   -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281   1325.6138
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272   4955.5586
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626   4316.0977
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173   8967.1924
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773    179.3077
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596   5296.9014
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440   -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991   1403.0193
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034   4495.6929
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862    874.5408
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373  16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613    289.6113
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907   1943.4092
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279    930.0871
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560    977.4719
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401   1722.6481
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807   -0.01883    616.1993
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553   1750.4175
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539    0.00809   1102.9956
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851   1012.5963
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771   -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399   3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133    0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188    0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412   1350.3611
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263    0.01322   4605.1411
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354    0.00245    629.5983
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739    789.7573
#> 332   714852.5 9561102 1787.11 45.358204    0.00540    0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196   1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317    0.00077   3056.9475
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426   1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121   1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610   1872.9276
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384    947.2075
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770   -0.03400   1982.7782
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655    130.9468
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166   -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382    100.0000
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002   -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534    0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298   -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118   2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028   2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133    828.8604
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457    0.01697    412.1185
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601   -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098    164.0802
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312    593.7773
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978    0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948   1222.5433
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950   -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009   -0.00599  12873.8018
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836    250.5111
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849    764.8193
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607   -0.04566  12862.4922
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491    261.7371
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284   -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233    968.7780
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222    991.8300
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491