Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345   1120.6929
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804    628.5153
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770    474.5059
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044 129412.3984
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212    561.1948
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117   1018.2141
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092    412.5849
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063    576.2819
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262   1760.2162
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337    486.7551
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776   1021.8267
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133   2397.2681
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185    953.8614
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744    226.7911
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144 317260.6875
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064    683.2795
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973    215.5101
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152  0.00098    164.0802
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759   4131.5356
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502    520.5218
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960    884.2003
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532  16771.3066
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624   1870.8307
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491