Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324     861.3537
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890    2301.0454
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345    1120.6929
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842     868.6137
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146    1347.6224
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323     183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178     800.1933
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779    2972.0193
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310    2833.9001
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454    3476.0818
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620     379.0855
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389   47022.9688
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058    1817.2371
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123     364.2903
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426    1315.5334
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010  0.00140    1066.3957
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116    1817.5222
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538     265.4677
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464    4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065     704.4843
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750    1892.8978
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532   16771.3066
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282    1659.6211
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270   35950.5508
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709    1181.2555
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172     351.5812
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057     722.3952
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491