Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633   -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752    1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767    3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292    1233.9299
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340     604.1032
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345    1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842     868.6137
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189     287.5425
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471   -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898     588.3761
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416    1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924    8023.8657
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279     930.0871
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560     977.4719
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883     616.1993
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656    3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643     670.2289
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133    0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242     507.9263
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556   -0.11694  102301.4609
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911     173.2806
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334    4905.2002
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132    1580.1621
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144  317260.6875
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306    1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665    1904.4769
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232     977.7320
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208     951.3401
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734     335.0773
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033    3446.5520
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099    1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060     383.3547
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960     884.2003
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238     787.0053
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884     672.2466
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983