Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649    1399.2329
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250    1413.3312
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403    1787.3024
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257     357.1436
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588    1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269     491.8017
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665   25318.3555
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973     514.5488
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150    1274.3594
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613     289.6113
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236    2715.4138
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357     473.4709
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330    1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602     345.8933
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474    2343.2100
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245    6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141    4611.0718
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583    1898.0105
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957     495.4453
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325     595.0426
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181  -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831     454.9113
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463     851.6306
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548     532.8210
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674  -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650     766.6254
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547     128.2613
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360    4288.7329
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904     608.3561
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875    2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172     351.5812
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020     673.2861
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388     180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336