Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491