Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102    483.4184
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491