Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813    1130.5791
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487    1077.7585
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250    1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767    3321.6113
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761     928.6858
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345    1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804     628.5153
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160    9344.2930
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261     554.9243
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629  360379.1875
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605     798.6775
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566  -0.03355   14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939     703.7817
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672     565.1107
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036     516.4709
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949    1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779    2972.0193
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181    2406.1252
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542     860.8840
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299     688.0706
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961    1086.6963
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132   42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081     940.6002
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448    1258.8196
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520    3390.5708
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776    1021.8267
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522    1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713    6379.7974
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610    1872.9276
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331     699.3024
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626     450.7346
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833     300.0125
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983     270.7949
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268     252.0072
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948    1222.5433
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204     618.2027
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360    4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502     520.5218
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907    1281.5044
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532   16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491     261.7371
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012     305.3624
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064    1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157     766.0220
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227     841.2025
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901