Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312   2657.6936
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077   3559.4204
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685   2568.4631
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734    335.0773
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462    716.5762
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607