Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304 53.745160
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441 15.665239
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537 24.437796
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997 39.598259
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426 24.128972
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299 42.221833
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761    928.6858 28.902283
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639 29.110267
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929 36.606274
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153 28.476575
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610 19.677663
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228 35.252374
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006 29.548579
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053 47.548622
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713 38.770335
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354 20.133737
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137 37.533319
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425 22.269151
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884 33.618299
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177 21.398255
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292 44.908432
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936 18.648630
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096   3210.5781 33.063103
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801 28.398653
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630 23.831607
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363 21.073961
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924 28.565957
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501 29.483262
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773    179.3077 25.710908
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774    527.1068 16.455921
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473 36.480223
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212    561.1948 20.380109
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876 36.075715
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686 27.826841
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555 28.203847
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757 34.391219
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657 34.595765
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887    366.8537 36.256769
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929 44.534291
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836 35.319410
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690 18.763795
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389 27.416030
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817 29.935899
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559 27.209193
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500 38.156697
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107 30.108359
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504   1020.2766 41.119462
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965 25.741848
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092 34.789424
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063  18733.8848 34.653060
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116 31.451945
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523 38.996080
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875 33.218947
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502 10.713738
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709 27.063088
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413 29.168135
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009   1247.7061 32.541138
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626 29.842507
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894 29.021458
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000 35.453482
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454   3476.0818 26.093071
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956 27.367902
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289 32.530252
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724 29.132612
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322 42.353040
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407 33.874984
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445 28.348233
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254 24.665260
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659 26.364080
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688 37.425030
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840 44.596170
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551 24.941426
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626 17.295304
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198 34.677125
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242 30.971807
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420 34.204434
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866 38.668349
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.597847
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854 25.336194
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100 33.798207
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341 32.922728
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681 25.235926
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008 34.424450
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910 27.720271
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657 29.730780
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562 31.625551
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010 30.482501
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421 16.521238
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652 31.665086
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585 -0.02665   1904.4769 36.185150
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398 29.404767
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318 26.486120
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161 19.775638
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644 27.225809
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089 35.857991
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432 36.366204
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949 38.982902
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525 33.383959
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666 42.063124
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209 24.344977
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075   8141.7549 35.462650
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580 21.112349
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196 27.986124
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101 34.567690
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527 21.442373
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538 -0.00052   3508.5591 31.257140
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843 32.092512
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347 24.137566
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228 30.102630
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355 25.007316
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701 25.759610
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179 22.100128
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114 32.706150
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944 33.421774
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433 30.485938
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463 31.272609
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426 40.203876
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236 32.282161
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076 25.726951
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855 17.048359
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231 38.417393
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019 32.003131
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645 34.657644
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193 45.583949
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066 30.020697
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055 42.010411
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741 40.603800
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906 23.095929
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624 26.538259
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042 11.038605
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544 27.569583
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466  4.191186
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606 38.910710
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128 42.407471
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300 29.050106
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785 37.533892
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130 45.213246
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156    758.4840 31.128797
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.978458
#>       distroad slides distdeforest distslidespast log.carea
#> 36052   300.00   TRUE       300.00            100  2.877158
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 8598    300.00   TRUE       300.00            100  3.121100
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 24129   300.00   TRUE       300.00             89  2.886120
#> 15843   300.00   TRUE       300.00            100  3.091290
#> 28293   300.00   TRUE       300.00              4  2.967869
#> 8869    300.00   TRUE       300.00            100  3.036974
#> 25169   195.00   TRUE        47.05              0  3.049487
#> 43669    24.22   TRUE         0.00             89  2.798316
#> 47255   300.00   TRUE       260.03            100  3.112324
#> 16811   300.00   TRUE       300.00             59  2.715521
#> 38993   300.00   TRUE       300.00              2  3.128238
#> 39629   300.00   TRUE       300.00             85  3.366461
#> 27375   300.00   TRUE       300.00            100  2.859781
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 23925   300.00   TRUE       300.00            100  2.938827
#> 21703   300.00   TRUE       300.00             34  2.458702
#> 48498   300.00   TRUE       300.00            100  3.098674
#> 22283   256.35   TRUE         0.00            100  2.772566
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 18304   300.00   TRUE       300.00             57  3.589919
#> 29472   111.09   TRUE         0.00             25  3.424505
#> 22933    87.56   TRUE       115.40              2  3.506583
#> 14522   300.00   TRUE       300.00            100  3.047228
#> 26215   300.00   TRUE       300.00             10  3.268382
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 6808    300.00   TRUE       300.00             40  3.952656
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 27708   300.00   TRUE       300.00            100  2.253599
#> 10320   300.00   TRUE       300.00             30  2.721899
#> 47220   300.00   TRUE       300.00            100  3.327041
#> 34565   300.00   TRUE        68.13            100  2.749114
#> 8073    138.71   TRUE        76.41             35  2.997730
#> 1635    300.00   TRUE       300.00            100  3.303016
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 28577   300.00   TRUE         0.56              0  3.781222
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 12460   300.00   TRUE         0.00            100  2.564493
#> 47011   300.00   TRUE       300.00             10  3.652797
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 34359   300.00   TRUE       294.31            100  3.287475
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 36931   300.00   TRUE       300.00            100  3.408707
#> 34563   300.00   TRUE       300.00            100  4.204981
#> 46235   300.00   TRUE        41.37              0  2.752133
#> 6967    300.00   TRUE       300.00             85  3.008718
#> 5578    300.00   TRUE       300.00            100  2.640875
#> 13699   300.00   TRUE        53.95             43  3.288564
#> 28743   300.00   TRUE       300.00             71  4.272628
#> 37341   300.00   TRUE       300.00              1  2.823221
#> 9704    300.00   TRUE       300.00             90  2.784369
#> 25244   300.00   TRUE       300.00              6  5.916915
#> 421     182.02   TRUE         0.00             83  2.874627
#> 18892   235.02   TRUE        27.24              0  2.675293
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 28845   300.00   TRUE       300.00             54  2.563199
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 10364   179.84   TRUE         0.00              0  3.541090
#> 664     300.00   TRUE       300.00             12  2.885027
#> 27695   300.00   TRUE       300.00             41  3.042574
#> 35014   300.00   TRUE       300.00             57  2.826223
#> 7064    300.00   TRUE       300.00            100  3.079714
#> 42765   300.00   TRUE       300.00             12  2.974066
#> 34192   273.94   TRUE         4.48             85  3.009256
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 5239    197.29   TRUE         4.67             57  3.591568
#> 40216   300.00   TRUE       300.00            100  3.257022
#> 37769   300.00   TRUE       300.00            100  4.672310
#> 6845    300.00   TRUE       300.00             29  2.769292
#> 46461   288.47   TRUE         0.00            100  2.687311
#> 44767   300.00   TRUE       300.00             25  3.046207
#> 10933   264.46   TRUE        18.17             56  3.311398
#> 18440   300.00   TRUE       300.00            100  4.477051
#> 6217    300.00   TRUE       300.00              1  3.243763
#> 7086    300.00   TRUE       128.61             63  3.071308
#> 24135   300.00   TRUE       122.87            100  3.878395
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 47298   300.00   TRUE       300.00            100  3.126386
#> 10617    17.89   TRUE       145.83              4  3.064162
#> 38668   300.00   TRUE       300.00            100  3.022236
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 1997    168.72   TRUE         0.00              5  3.379717
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 48574    96.21   TRUE       101.32             10  3.143827
#> 39405   300.00   TRUE       300.00            100  3.413298
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 46038   300.00   TRUE       243.86            100  3.184852
#> 28588   300.00   TRUE       300.00             78  5.209992
#> 17674   300.00   TRUE       300.00              1  3.812044
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 17322    75.77   TRUE       138.35              1  3.235115
#> 36765   300.00   TRUE       300.00            100  3.279776
#> 36744   300.00   TRUE       300.00            100  2.937989
#> 4785    300.00  FALSE         0.00            100  3.050663
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 7317    300.00  FALSE       300.00             61  3.448129
#> 48601   300.00  FALSE       245.19            100  2.998830
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 25212   300.00  FALSE       300.00            100  3.109374
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 26963   300.00  FALSE       300.00            100  3.420610
#> 12716   300.00  FALSE       300.00             25  2.731444
#> 43426   300.00  FALSE        82.11            100  3.910718
#> 441     294.31  FALSE       152.09            100  3.054287
#> 16542   300.00  FALSE       300.00            100  2.909320
#> 43243   300.00  FALSE       300.00            100  2.922424
#> 39796   300.00  FALSE       300.00             71  2.333468
#> 44396   300.00  FALSE       300.00            100  2.270336
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 18627   300.00  FALSE       300.00             20  3.545129
#> 4007    300.00  FALSE       300.00             68  2.847871
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 44668     9.15  FALSE       163.30              6  2.852799
#> 41753   300.00  FALSE       300.00            100  2.860419
#> 47059   300.00  FALSE       300.00            100  3.023816
#> 3733    300.00  FALSE       300.00            100  3.055767
#> 39440   300.00  FALSE       300.00             92  2.369237
#> 45486   300.00  FALSE         0.00             68  3.366011
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 5656    300.00  FALSE        94.43            100  2.853361
#> 35038   300.00  FALSE       300.00            100  3.347416
#> 26768   300.00  FALSE       207.22            100  2.673964
#> 3724    300.00  FALSE       300.00            100  2.450415
#> 28286   300.00  FALSE       300.00            100  2.790556
#> 12300   124.35  FALSE         0.00            100  3.192630
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 18308   300.00  FALSE       300.00             55  4.224567
#> 19767   300.00  FALSE       300.00            100  2.878525
#> 11394    46.04  FALSE        55.00            100  3.187005
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 16443   300.00  FALSE       300.00             75  2.484816
#> 41873   300.00  FALSE       300.00            100  2.265064
#> 9526    300.00  FALSE       300.00            100  2.898919
#> 37254   131.59  FALSE        80.55            100  2.827529
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 6577    300.00  FALSE       300.00             70  2.789872
#> 24962   300.00  FALSE       300.00            100  2.996437
#> 23808   300.00  FALSE       300.00             72  2.692387
#> 14645   300.00  FALSE       300.00            100  3.232313
#> 41373   300.00  FALSE       120.90            100  2.879946
#> 13548   300.00  FALSE       300.00            100  2.256336