Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029    5577.3916
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023   -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250    1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767    3321.6113
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273   -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250   -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046    1024.1573
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311     586.7354
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151    0.00899    1394.2274
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077   -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173    8967.1924
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108   -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987    0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088    2322.1746
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937    0.00083     994.7876
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991    1403.0193
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843    0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063   18733.8848
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283    0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855   -0.02175    1122.9120
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252    5761.3486
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690    5121.6924
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529    0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876    0.00282    0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406   -0.00055     773.0359
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539    0.00809    1102.9956
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291   -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242     507.9263
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370   -0.00190     640.8842
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509   -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241    1752.9242
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078  622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384    0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321    1170.5904
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298    1533.7925
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118   19557.6621
#> 332   714852.5 9561102 1787.11 45.358204    0.00540    0.00879    1200.4210
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181   -0.00001    1184.7952
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267    0.00483     343.1384
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526     706.1463
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758   -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306    1718.3652
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367    1966.2112
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599    0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002   -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177     997.3089
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570    0.02230     714.7432
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204     853.2902
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328    1066.7134
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197    0.00463     851.6306
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409    1575.9425
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028    2394.7261
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174   -0.03065     704.4843
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175     725.1355
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687   -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400    4126.7812
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018     937.6911
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705     968.9738
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342    0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418    0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204     618.2027
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229    0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360    4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502     520.5218
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750    1892.8978
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074   -0.02904     608.3561
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757   -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282    1659.6211
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865   -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108   -0.00948    1538.1741
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340    0.01809    1158.0426
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595    0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305    0.00585    5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169     649.9975
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222     991.8300
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607