Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056 -0.06976     301.2347
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483     566.7674
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761     928.6858
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257     357.1436
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542     674.0147
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416     947.4878
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328    2235.5259
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164     895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769    1032.7177
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530    9000.4356
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314    2360.1501
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629  360379.1875
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150    1274.3594
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325     616.8652
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405    5681.2412
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991    1403.0193
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556     660.3690
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338   10787.6299
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230    1056.5876
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771     674.3859
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825     162.6688
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357     473.4709
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166     650.6246
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316     454.9674
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389   47022.9688
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376    2048.3198
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583    1898.0105
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123     364.2903
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970    1443.9200
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981    3975.3608
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422    1159.2100
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915    2036.0378
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417   34908.6562
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450    2589.9910
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610    1872.9276
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685    2568.4631
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152     850.4633
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606     919.4241
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732    4237.6934
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632     620.0229
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400    1982.7782
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565    1410.8372
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419     761.5427
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118    2273.4758
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945     259.7421
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923     811.5580
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476   11992.5273
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697     412.1185
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175     725.1355
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462     716.5762
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645     234.0114
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148   10771.1963
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499     617.3855
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050    1471.4346
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576    1045.1074
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011     501.7867
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585    5714.3267
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799    1036.6606
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931    1812.3820
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983