Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806