Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393   2081.0056
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917    0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753    0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659    0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661   -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857    0.03677   1675.4679
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295    0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.97109  0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922    0.01088    266.4355
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194    0.00526    539.8119
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525   -0.11024    394.8751
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448   -0.00328    592.3326
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131    0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191   -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430   -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151    0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856    0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043    0.00973    514.5488
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077   -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902    0.01358    802.7622
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108   -0.00678   1891.6584
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664    0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257    0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382    0.00088   2322.1746
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422    0.02357    275.7474
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616    0.01276   9614.6924
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414    0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765    0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646   -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440   -0.02651    419.0927
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294   -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606   -0.00556    660.3690
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229   -0.00511 361979.2812
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940    0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060    0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034    0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661    0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613    0.00903   2360.1536
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492    0.00072  54831.6289
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008    0.00102    483.4184
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855   -0.02175   1122.9120
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405   -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364   -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211    0.00241    585.6182
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681   -0.01319   3457.7519
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057    0.02427    380.9472
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032    0.00542   3614.7334
#> 664   714452.5 9559852 2174.83 44.86088  0.00282    0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406   -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104    0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410    0.03620    379.0855
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629   -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481   -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588    0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538    0.00112    745.3315
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133    0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567   -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178   -0.00822    472.8748
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246    0.00534    344.1447
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204   -0.00376   2048.3198
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.10642  0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890   -0.00911    173.2806
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079    0.00321   1170.5904
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109    0.00298   1533.7925
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253    0.00753   1337.7854
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669   -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102   -0.01778   4594.4351
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970    0.00200   1305.6691
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397    0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157   -0.00152    850.4633
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458    0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144   -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488   -0.00632    620.0229
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115    0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758   -0.00288   2144.8164
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159   -0.01559    866.5775
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009    0.00109    605.1908
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759    0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226    0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298   -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124    0.00204    853.2902
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133   -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815   -0.03345 665381.5625
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480    0.01500    538.8209
#> 441   714312.5 9560972 1916.59 31.30298  0.00540    0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724    0.01464   4029.5947
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846    0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176    0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405    0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007   -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174   -0.03065    704.4843
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990    0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846    0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432   -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236    0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490   -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193   -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423   -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292    0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301   -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224   -0.02036    476.9504
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978    0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452   -0.00948   1222.5433
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009   -0.00599  12873.8018
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980   -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972    0.00308    422.5022
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420    0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944    0.00836    250.5111
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343    0.01177   1558.2231
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598    0.02208    842.4825
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129    0.00491    261.7371
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278   -0.01172 111973.8906
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309    0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788   -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305    0.00585   5714.3267
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933   -0.00157    766.0220
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181   -0.01149   4242.3579
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59291   300.00   TRUE       300.00            100  3.318273
#> 30059 20.19218   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28349   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73881   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59826   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51747   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988   300.00   TRUE       195.00              2  3.224136
#> 49372 23.54054   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94657   162.65   TRUE         0.00             77  3.703905
#> 125   29.59671   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61140   300.00   TRUE       300.00              4  2.425592
#> 25690 21.74776   300.00   TRUE       300.00            100  2.732242
#> 33792 51.20810   300.00   TRUE       300.00              5  2.596460
#> 22283 19.94237   256.35   TRUE         0.00            100  2.772566
#> 29420 28.85244   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843   300.00   TRUE       300.00             57  3.589919
#> 18005 24.90590   300.00   TRUE       300.00              0  2.744234
#> 38279 37.62843   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93532   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073    68.13   TRUE       116.32             10  3.654262
#> 34275 36.82915   300.00   TRUE       300.00            100  2.711427
#> 6676  34.62155   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71778   300.00   TRUE       300.00             83  2.904587
#> 3453  40.08470   300.00   TRUE       300.00              2  3.276843
#> 20194 28.11733   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116   300.00   TRUE         0.00            100  3.365895
#> 3133  18.83427   300.00   TRUE       300.00            100  2.440511
#> 46125 25.65705   300.00   TRUE       300.00            100  3.982935
#> 27013 30.91852   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39122   300.00   TRUE         0.56              0  3.781222
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 47429 32.09767   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379   300.00   TRUE       300.00              2  2.819787
#> 24100 34.29783   300.00   TRUE       300.00            100  5.558684
#> 31316 30.00981   300.00   TRUE       300.00             85  2.740390
#> 46171 29.07131   300.00   TRUE       270.22            100  3.645964
#> 17036 24.92825    25.33   TRUE       208.61             15  3.608265
#> 17423 25.46053   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61065   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048   300.00   TRUE       300.00            100  4.739031
#> 2976  26.88203   300.00   TRUE       300.00            100  2.684323
#> 16015 38.83279   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559   300.00   TRUE       300.00              8  3.050346
#> 49569 41.84941    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06309   235.02   TRUE        27.24              0  2.675293
#> 26698 34.24970   300.00   TRUE       300.00              7  2.767615
#> 30406 43.10075   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43953   300.00   TRUE       300.00            100  3.538794
#> 10507 27.16221   300.00   TRUE       300.00             46  2.580865
#> 47100 29.64598    82.80   TRUE       129.73             36  3.558076
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44278   300.00   TRUE       300.00              0  2.888200
#> 13146 33.57647   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556   300.00   TRUE       300.00              1  2.578737
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498   273.94   TRUE         4.48             85  3.009256
#> 21680 28.34823   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526   197.29   TRUE         4.67             57  3.591568
#> 31834 24.64693   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753   300.00   TRUE       300.00            100  2.872349
#> 39722 26.54170   300.00   TRUE       300.00            100  2.906387
#> 12772 37.92007   300.00   TRUE       300.00            100  2.936990
#> 18958 29.98976   300.00   TRUE       300.00              5  2.674746
#> 47320 45.66416   300.00   TRUE       300.00             43  2.536741
#> 10933 34.67712   264.46   TRUE        18.17             56  3.311398
#> 27244 37.77511   300.00   TRUE       300.00             54  5.009882
#> 7     13.22043   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840   300.00   TRUE       300.00             65  2.238750
#> 26101 28.64617   300.00   TRUE       300.00            100  3.068405
#> 28009 27.01782   300.00   TRUE       300.00             29  3.185767
#> 47298 25.33619   300.00   TRUE       300.00            100  3.126386
#> 10363 45.09350   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79821    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765   300.00   TRUE       300.00            100  3.662232
#> 14192 40.58661   300.00   TRUE       300.00            100  3.115833
#> 42617 31.23193   300.00   TRUE       300.00            100  2.848895
#> 30363 29.80297   300.00   TRUE       300.00            100  2.929656
#> 17227 43.59750   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28558   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505   300.00   TRUE       300.00            100  2.792408
#> 14779 16.52124    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51210    20.21   TRUE       150.14              2  3.331390
#> 16208 38.37385   256.60  FALSE         0.00             67  2.937807
#> 27884 15.49736   300.00  FALSE       300.00            100  2.781892
#> 748   38.13263   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371   300.00  FALSE       300.00             95  2.679300
#> 16464 35.04611   300.00  FALSE       300.00            100  3.004540
#> 18286 24.35071   300.00  FALSE       300.00             42  2.931097
#> 26023 45.07287   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477   144.37  FALSE        60.16            100  5.823071
#> 12716 24.34498   300.00  FALSE       300.00             25  2.731444
#> 441   25.11331   294.31  FALSE       152.09            100  3.054287
#> 13512 35.93477   300.00  FALSE       300.00            100  3.605261
#> 49835 32.36123   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56769   300.00  FALSE       300.00             71  2.333468
#> 20547 34.26116   300.00  FALSE        42.56            100  4.078911
#> 45071 27.80507   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67921   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251   300.00  FALSE       300.00             68  2.847871
#> 34616 31.49950   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55202   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13757   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89312   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22716   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07246   300.00  FALSE       300.00            100  3.167901
#> 33175 34.78369   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659   300.00  FALSE       300.00            100  2.639632
#> 8895  38.07763   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006   300.00  FALSE       300.00            100  2.855262
#> 34105 20.07701   300.00  FALSE       300.00             26  3.059266
#> 13737 40.96018   300.00  FALSE       300.00            100  2.678473
#> 45486 33.42177   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48594   300.00  FALSE       120.00            100  3.087264
#> 46509 45.57879   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26533   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820   234.82  FALSE         0.00             43  4.109707
#> 26768 32.28216   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90782   300.00  FALSE       300.00             48  2.625829
#> 40612 21.92709    55.73  FALSE       125.03            100  2.946551
#> 28286 17.04836   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044   300.00  FALSE       300.00            100  2.398827
#> 14763 10.14479   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739   124.35  FALSE         0.00            100  3.192630
#> 24446 31.36314   300.00  FALSE       300.00            100  2.925561
#> 6732  26.85625   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61941   300.00  FALSE       300.00             49  2.417865
#> 18504 23.09593   144.96  FALSE         4.48             89  5.049117
#> 9320  18.13927   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115   300.00  FALSE       300.00            100  2.951467
#> 46639 33.56329   300.00  FALSE       300.00            100  3.756965
#> 30403 36.13072   300.00  FALSE       209.91            100  2.884241
#> 46567 16.94809   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66337   300.00  FALSE       300.00            100  3.627607