Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969     368.5606
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339    1002.5877
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388    1051.4222
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032     994.6740
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114    4510.8833
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898     588.3761
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150    1274.3594
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338   10787.6299
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964     571.1364
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262    1760.2162
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512   13178.3223
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213     328.7922
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970    1443.9200
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077    3056.9475
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294     734.9908
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252    3631.7017
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230    1044.6443
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403     474.8613
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093     118.1402
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452     578.5847
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548     532.8210
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119     792.3544
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172     351.5812
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491