Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029    5577.3916
#> 21788 715392.5 9560172 1988.71 40.4766034  -0.01919   -0.04051  351155.1250
#> 44302 714932.5 9557982 2649.56 37.3006984   0.01633   -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691   0.00221    0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.0537816   0.00097    0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644    1179.0695
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732  -0.00891   -0.00949    6217.8213
#> 14456 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.9691094  -0.00295    0.00035    2697.5974
#> 47255 713552.5 9559962 2191.77 33.5787645  -0.00068    0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387    1989.3012
#> 22814 713862.5 9558582 2327.48 48.6120312  -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423   0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.6648828  -0.00831   -0.00149   12632.7793
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526     539.8119
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189     287.5425
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.0168746  -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058   0.00602    0.01518     854.4269
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276    9614.6924
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642     613.6390
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924    8023.8657
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887     366.8537
#> 34359 712862.5 9558912 2111.37 32.8121470   0.00463    0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.6513065  -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517    1000.7119
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270    4057.5637
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.1730683   0.00359   -0.00279     930.0871
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964     571.1364
#> 20220 714052.5 9558522 2378.67 39.6561279  -0.00855   -0.02175    1122.9120
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993   10308.6406
#> 38862 712952.5 9559892 1950.86 37.7218224  -0.00048   -0.00332    1418.6622
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401    1722.6481
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.5983082   0.04548    0.05543     352.1558
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102   35685.0000
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180  -0.00104    0.01664    1898.5588
#> 24501 713302.5 9559142 2125.66 27.1089251   0.01784    0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162  -0.00478   -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389   47022.9688
#> 12772 715172.5 9557642 2759.42 34.8054672  -0.00567   -0.03793     864.9472
#> 1973  715252.5 9557822 2861.00 15.5707010   0.01865    0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813  287033.4688
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524   10861.9482
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 43201 715712.5 9557922 2852.82 31.4622584  -0.00949    0.00520    3390.5708
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522    1355.7436
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510  -0.00442   -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.6858936  -0.00124   -0.00456    1531.4274
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190     264.6404
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.2155023  -0.01016   -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 14950 713382.5 9559172 2111.91 48.1066824   0.00915   -0.02705     413.5702
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 47722 712732.5 9561022 1999.03 37.4926392   0.00004   -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 34718 714802.5 9558802 2599.22 44.6500280   0.01996    0.04684     209.1778
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797    7557.7866
#> 16208 712882.5 9560332 1846.34 48.6252092   0.00159   -0.01559     866.5775
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290  -0.00162   -0.00517    1060.5494
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.7101816   0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363     389.2691
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345  665381.5625
#> 19493 714252.5 9560372 2156.93 24.8795463   0.00825    0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.4949170   0.01306   -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.2571887   0.00075   -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.8707563   0.00321    0.00548     532.8210
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028    2394.7261
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333     422.8840
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175   0.00174   -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734     335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.2478257   0.00737    0.01033    3446.5520
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780     825.4451
#> 36245 713142.5 9560572 1827.82 43.6759998  -0.00410    0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978   0.05041   -0.00011     201.1272
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.4571579  -0.03224   -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518    1137.0179
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220   0.00270   -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.1662069   0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788   0.00119    0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.5153479   0.00972    0.00308     422.5022
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836   0.00262   -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577     423.5598
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964   0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875    2156.9358
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327   0.00293    0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233     968.7780
#> 46148 715082.5 9558042 2725.39 42.8360436  -0.00181   -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.9643862  -0.00933   -0.00157     766.0220
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.0432868   0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.9699131  -0.01023   -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057     722.3952
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491