Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569   300.00   TRUE       300.00             21  3.145890
#> 4965  33.90592   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457   300.00   TRUE         9.15              2  2.802317
#> 42632 20.47808   300.00   TRUE       300.00             10  2.566509
#> 14254 15.66524   300.00   TRUE       300.00             10  2.510336
#> 5999  27.14216   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062   300.00   TRUE       300.00              6  2.888165
#> 24129 24.12897   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93245   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239   300.00   TRUE       300.00            100  3.363704
#> 38411 32.75829   279.54   TRUE         0.00            100  3.793638
#> 1768  28.18551   300.00   TRUE       300.00              7  3.384768
#> 23817 21.61140   300.00   TRUE       300.00              4  2.425592
#> 39629 47.54862   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13374   300.00   TRUE       300.00            100  2.768442
#> 23925 37.53332   300.00   TRUE       300.00            100  2.938827
#> 8019  35.78007   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32130   300.00   TRUE       123.56              5  2.520462
#> 43964 28.75732   300.00   TRUE       300.00             46  2.952143
#> 16843 21.39825   300.00   TRUE       300.00             12  3.013982
#> 35285 42.46649   300.00   TRUE       300.00            100  2.997681
#> 14522 28.39865   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32853   300.00   TRUE       300.00             12  2.377795
#> 6676  34.62155   300.00   TRUE       300.00            100  3.303289
#> 45575 25.72810   300.00   TRUE       300.00             59  2.562268
#> 34280 33.62116   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38011   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023   300.00   TRUE       300.00             75  2.790190
#> 46125 25.65705   300.00   TRUE       300.00            100  3.982935
#> 28577 34.39122   300.00   TRUE         0.56              0  3.781222
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259   300.00   TRUE       300.00            100  2.894014
#> 12868 37.58832   119.34   TRUE        79.16             61  3.129568
#> 47351 34.87881   300.00   TRUE       300.00            100  4.173758
#> 48434 29.93590   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190   138.35   TRUE         0.00             60  3.701999
#> 6967  41.11946   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851   300.00   TRUE       300.00            100  2.903195
#> 31316 30.00981   300.00   TRUE       300.00             85  2.740390
#> 13699 34.78942   300.00   TRUE        53.95             43  3.288564
#> 37341 31.45195   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626   300.00   TRUE       300.00            100  2.968524
#> 34707 27.70251   165.13   TRUE        39.56            100  3.473052
#> 835   36.65039   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911   300.00   TRUE       127.54              2  2.615513
#> 49569 41.84941    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357   300.00   TRUE       300.00            100  4.066528
#> 39520 35.89409   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43953   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917   118.07   TRUE        98.15              0  3.619059
#> 6317  27.32321   300.00   TRUE       300.00            100  2.837633
#> 23213 25.46053   300.00   TRUE       300.00            100  2.868983
#> 44650 37.36258   300.00   TRUE       300.00            100  3.110849
#> 30886 39.18516   300.00   TRUE       300.00            100  2.695380
#> 46065 29.77089   300.00   TRUE       251.11            100  3.482407
#> 31121 38.16014   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647   300.00   TRUE       300.00              2  3.278424
#> 38197 34.34252   105.52   TRUE        65.32             50  3.036108
#> 5103  34.91318   300.00   TRUE       300.00             91  2.623976
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 30268 47.27246   300.00   TRUE       300.00             41  2.705801
#> 7041  33.21895   300.00   TRUE       300.00              6  5.916915
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15422   300.00   TRUE       300.00            100  4.035908
#> 27244 37.77511   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94470   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97181   300.00   TRUE       300.00              1  3.243763
#> 7086  34.20443   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66835   300.00   TRUE       122.87            100  3.878395
#> 21683 19.74527   130.93   TRUE         0.00             21  3.399421
#> 22907 33.81368   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342    99.29   TRUE         0.00              6  3.137802
#> 17058 26.32283   300.00   TRUE       300.00              8  2.751808
#> 47244 28.03884   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18075   300.00   TRUE       300.00            100  3.159543
#> 10363 45.09350   300.00   TRUE       297.43              6  3.599377
#> 22301 33.61543   300.00   TRUE         0.00            100  3.308786
#> 23783 27.60052   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27356   300.00   TRUE       300.00             21  3.311309
#> 48574 34.42445    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69145   300.00   TRUE       300.00            100  2.460216
#> 29024 42.57306   300.00   TRUE       300.00             32  2.529363
#> 46792 27.60110   300.00   TRUE       300.00             18  2.657830
#> 30626 30.00351   300.00   TRUE       300.00              0  2.799064
#> 24532 32.54629   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172   300.00   TRUE       300.00             55  2.355626
#> 46971 16.78423   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90610   300.00   TRUE       300.00            100  3.398953
#> 31491 27.55870   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 23832 25.48516   300.00   TRUE       300.00              8  2.676567
#> 36765 36.18515   300.00   TRUE       300.00            100  3.279776
#> 13072 34.94470   300.00   TRUE       300.00            100  4.477051
#> 15211 30.83487   300.00  FALSE       300.00            100  2.844665
#> 36561 25.80487   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81841   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674   300.00  FALSE       300.00            100  3.105832
#> 40459 36.18687   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29534   300.00  FALSE       300.00            100  2.382556
#> 49835 32.36123   300.00  FALSE         0.00            100  3.134665
#> 29125 23.94505   300.00  FALSE       300.00            100  2.626221
#> 1085  25.89483   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83868   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26116   300.00  FALSE        42.56            100  4.078911
#> 34347 31.72009   300.00  FALSE       300.00            100  3.477538
#> 31374 26.67921   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251   300.00  FALSE       300.00             68  2.847871
#> 48183 30.94144   300.00  FALSE       300.00            100  3.204523
#> 32849 29.55202   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13757   300.00  FALSE       300.00            100  3.117018
#> 21925 31.24855   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00732   300.00  FALSE       300.00            100  2.860419
#> 42934 32.22716   300.00  FALSE       300.00             22  4.499525
#> 6143  28.75732   300.00  FALSE       185.58            100  2.985864
#> 47059 25.75961   300.00  FALSE       300.00            100  3.023816
#> 17746 36.39141   300.00  FALSE       300.00             87  2.773624
#> 12519 30.48594   300.00  FALSE       120.00            100  3.087264
#> 20764 39.85036   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792   300.00  FALSE       300.00            100  3.247811
#> 330   13.57222   300.00  FALSE       275.17            100  2.372638
#> 43222 21.26533   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 3868  26.14750   300.00  FALSE       300.00            100  3.277127
#> 28286 17.04836   300.00  FALSE       300.00            100  2.790556
#> 23499 28.51496   300.00  FALSE       300.00             50  3.171627
#> 44434 36.31979   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06500   300.00  FALSE       300.00            100  3.167741
#> 4943  37.77167   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36314   300.00  FALSE       300.00            100  2.925561
#> 1577  38.92962   300.00  FALSE       300.00             63  4.635980
#> 36141 27.30602   300.00  FALSE       300.00             85  4.412069
#> 47245 34.97621   300.00  FALSE       150.03            100  2.509309
#> 13548 16.97846   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809   300.00  FALSE       300.00             78  2.785235
#> 15363 26.94219   300.00  FALSE       300.00            100  2.858775
#> 40852 21.51800   300.00  FALSE       300.00            100  2.360983