Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607