Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603