Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029   5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203    634.3320
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357   1752.0375
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475   1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005    769.3426
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949   6217.8213
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730    0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783    377.8918
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416    947.4878
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164    895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367    793.9665
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314   2360.1501
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357    275.7474
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651    419.0927
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606   -0.00556    660.3690
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517   1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907   1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925    665.6116
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102    483.4184
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373    527.3528
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304   1569.7313
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664   1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219    420.7032
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132  42970.7891
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822    472.8748
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849   1120.5961
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583   1898.0105
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801   1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077   3056.9475
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685   2568.4631
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020    0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905   6487.0010
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684    209.1778
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367   1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320    911.5911
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996   1599.3491
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118   2273.4758
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995   1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028   2394.7261
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833    300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133    828.8604
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216   1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997    230.8641
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282   2322.7944
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360   4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710    472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044   -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849    764.8193
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172 111973.8906
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253   1237.0536
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607