Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649    1399.2329
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579    1318.7560
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057    1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005     769.3426
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542     674.0147
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088     266.4355
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766  140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272    4955.5586
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114    4510.8833
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596    5296.9014
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642     613.6390
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346     543.0749
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517    1000.7119
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175     804.5679
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058     431.4362
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974    2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495     266.5888
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181   11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332    1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020    1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316     454.9674
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555    4439.0967
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335     195.0293
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211   49347.5547
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860     695.2302
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739     789.7573
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450    2589.9910
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252    3631.7017
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158     837.3730
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548     532.8210
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794    1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853    1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033    3446.5520
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662    1269.2650
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099    1146.2140
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654    1576.2065
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650     766.6254
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148   10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750    1892.8978
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964     543.0025
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050    1471.4346
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012     305.3624
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013     616.4128
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057     722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775