Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649    1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958   0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813    1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767    3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005     769.3426
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949    6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835    2425.3115
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588    1295.1610
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311     586.7354
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842     868.6137
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769    1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626    4316.0977
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437    1155.1469
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405    5681.2412
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365    1536.2647
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092     412.5849
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825     162.6688
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684    2556.8049
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026     247.0894
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102   35685.0000
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940     987.8085
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212   46310.0195
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141    4611.0718
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776    1021.8267
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812    1423.7578
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753    1337.7854
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133    2397.2681
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450    2589.9910
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000  162178.1562
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291     293.4231
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655     130.9468
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427     866.9398
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331     699.3024
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463    1050.7473
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500     538.8209
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923     811.5580
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759    1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216    1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780     825.4451
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152   0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099    1146.2140
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268     252.0072
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130     107.5831
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204     618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417    1520.6630
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576    1045.1074
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491     261.7371
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011     501.7867
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156     758.4840
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491