Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769