Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983