Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769