Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183  -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199   0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023  -0.00043    1806.4552
#> 42733 715112.5 9559062 2399.70 36.0545152   -0.00041   0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659   0.01041     772.9746
#> 24129 714602.5 9560542 2038.30 25.5848574    0.00065   0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519   -0.00503   0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090   0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858  -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431    4079.0366
#> 37270 714852.5 9561032 1838.96 35.1589821   -0.00854   0.01264    5057.1431
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894  -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.6648828   -0.00831  -0.00149   12632.7793
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952   0.01378     440.9104
#> 33792 714652.5 9558172 2598.36 59.4586952   -0.03525  -0.11024     394.8751
#> 43007 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678  -0.04272    4955.5586
#> 48055 712862.5 9559972 1902.48 29.0168746   -0.00720   0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382     560.8309
#> 18005 714832.5 9557912 2660.76 31.9120303    0.00191  -0.00261     554.9243
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440   0.00710     843.7961
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051   0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.2343959    0.00993  -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643   0.00314    2360.1501
#> 27708 715222.5 9557782 2844.65 29.5428498    0.03206   0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774     527.1068
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559    0.01205   0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097   0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089  -0.00051     779.0555
#> 22735 714492.5 9559772 2170.30 39.6887865   -0.00440  -0.02651     419.0927
#> 46756 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514     297.7842
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224   0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294  -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229  -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.8898455   -0.01993   0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752  -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922   -0.00856   0.01796    4729.6382
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066  -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.6425796    0.01843   0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199     196.4080
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731    1384.7959
#> 36120 714622.5 9559942 2166.51 32.9863262   -0.02917  -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575   -0.00405  -0.01565    2232.6133
#> 48808 715032.5 9557792 2707.02 25.4376072   -0.02329  -0.01181   11655.4316
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.5983082    0.04548   0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856  -0.01625    1063.9150
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340   -0.00131   0.00201     788.0621
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588   0.01448    1258.8196
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133   0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743  -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469   0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.2403562   -0.01169  -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010    0.01865   0.00335     195.0293
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.6704448    0.01489   0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797    7557.7866
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175   0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175    1026.6818
#> 28312 715712.5 9558032 2783.23 24.2527305   -0.00007  -0.02132    1580.1621
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253   0.00753    1337.7854
#> 10363 713252.5 9559722 1999.55 41.8299297   -0.03669  -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042   0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.4626708   -0.02970   0.00200    1305.6691
#> 39624 715222.5 9559552 2205.54 23.8974967    0.00401  -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.0463191   -0.08153  -0.02417   34908.6562
#> 16065 715242.5 9560622 1871.61 34.3064209   -0.00019   0.00119    6589.5796
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157  -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031    0.02451  -0.03921     525.3245
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118  -0.00252    3631.7017
#> 35458 715242.5 9558442 2623.01 38.9548275   -0.00770  -0.03400    1982.7782
#> 23832 714252.5 9560182 2188.12 30.6475125    0.00427   0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.6500280    0.01996   0.04684     209.1778
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316  -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084  -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203   0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.0046187    0.00599   0.00301     338.9676
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290   -0.00162  -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156  -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336   0.01403     241.2991
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480   0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122   -0.00345  -0.00075    8141.7549
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 45166 715262.5 9558722 2531.59 13.5533166   -0.00502  -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973     215.5101
#> 1085  715722.5 9557532 3097.22 28.7842537    0.02327   0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507     504.3085
#> 32849 714302.5 9558222 2520.01 32.1572563   -0.00846   0.00216    1412.2295
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844   0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.6119821   -0.03193  -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.2735807   -0.00022   0.00662    1269.2650
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.1750320   -0.00687  -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410   0.00400    4126.7812
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301  -0.01099    1146.2140
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379  -0.01018     937.6911
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462   0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.5953803   -0.00035   0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130     107.5831
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667   0.00863    1098.3462
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009  -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972   0.00308     422.5022
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635  -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608   0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537  -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.5684513    0.00420   0.01960     884.2003
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.7668833   -0.00375   0.00215     917.5296
#> 18246 713902.5 9561342 1818.52 19.1104343   -0.01502   0.00282    1659.6211
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182  -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390   -0.06909  -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253  -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501   -0.00278  -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186    0.03263  -0.02012     305.3624
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983   0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986  -0.05013     616.4128
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241   0.00020     673.2861
#> 39234 712902.5 9559652 1953.39 29.7319259    0.07379  -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336   0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227     841.2025
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983