Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357    1752.0375
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752    1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579    2246.7725
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491    1844.4353
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264    5057.1431
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226    2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311     586.7354
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294   -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846   -0.00034    4495.6929
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904   -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888    0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664    1898.5588
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245    6353.3179
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334    4905.2002
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263    0.01322    4605.1411
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123     364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522    1355.7436
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441     424.2680
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246    0.00396     704.7076
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077    3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001    1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019    0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067    1123.7318
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177     997.3089
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077    8759.9961
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267   31588.1953
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655     560.8554
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836     250.5111
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047    2826.7700
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148     481.8292
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935     903.6397
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595    0.03064    1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775