Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.10134    0.00183   -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.97397    0.02056   -0.06976     301.2347
#> 16749 715072.5 9558242 2745.27 49.05378    0.00097    0.01483     566.7674
#> 8598  714002.5 9558872 2310.02 39.58394   -0.00498   -0.00752    1321.5997
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514   -0.00644    1179.0695
#> 37104 713962.5 9561282 1795.56 25.74185    0.00273   -0.02403    1787.3024
#> 8869  715062.5 9559392 2278.56 41.40193    0.00310    0.01440    1088.8639
#> 16811 714902.5 9559262 2362.67 50.70333   -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.65743    0.00922    0.01088     266.4355
#> 10974 715312.5 9557502 2844.23 26.66488   -0.00831   -0.00149   12632.7793
#> 38287 714872.5 9561162 1719.41 48.23560   -0.03021   -0.09328    2235.5259
#> 44328 713222.5 9558932 2227.94 29.15553   -0.00769   -0.00281    1325.6138
#> 2107  714142.5 9558562 2382.67 35.55375    0.01412    0.00898     458.6938
#> 29472 714792.5 9561072 1786.08 36.17312   -0.01029   -0.02312    2657.6936
#> 23375 713162.5 9559632 2041.52 46.81524   -0.00857    0.03677    1675.4679
#> 25081 713392.5 9558452 2242.24 46.53850   -0.02716    0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.11967    0.02059   -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713   -0.00173    8967.1924
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.94716    0.00598    0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.39488   -0.01996    0.00626    1014.7159
#> 4329  712882.5 9559982 1911.46 26.23173   -0.00193   -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.66450    0.06765    0.00114     150.1498
#> 45104 715062.5 9559242 2345.02 24.15647   -0.01743    0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.59273    0.00804    0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.30757   -0.00154    0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224    0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.81215    0.00463    0.00167    1938.5389
#> 19784 714982.5 9557722 2716.98 35.43629    0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.21397   -0.03229   -0.00511  361979.2812
#> 46201 714832.5 9561022 1845.94 34.57399   -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.93360    0.00321    0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.88150   -0.00666   -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.03793    0.04527    0.04613     289.6113
#> 37341 714412.5 9559142 2377.24 40.59979   -0.00735   -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.39488   -0.01996    0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.93664   -0.02617    0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.55527   -0.04492    0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.17307    0.00359   -0.00279     930.0871
#> 20220 714052.5 9558522 2378.67 39.65613   -0.00855   -0.02175    1122.9120
#> 35194 713692.5 9559952 2185.68 28.79686   -0.00449   -0.00181    2406.1252
#> 462   713712.5 9561182 1785.08 47.71707   -0.03260   -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.92351   -0.01364   -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.95442   -0.04708    0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953   -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221    0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.85627   -0.00458   -0.00401    1722.6481
#> 31462 714082.5 9558542 2367.84 34.72697   -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.21952   -0.00034    0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.06465    0.00062   -0.00262    1760.2162
#> 10364 714762.5 9561002 1856.21 25.27489   -0.00845    0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.86088    0.00282    0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.59697   -0.03800   -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.35263    0.00568   -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.62441   -0.01972    0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.77727   -0.01611    0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478   -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.01118    0.00394   -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.40821    0.00291   -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.23709   -0.00472   -0.00228    2262.3796
#> 30268 714172.5 9558612 2328.84 54.95066    0.03743   -0.05242     507.9263
#> 6845  713412.5 9559172 2123.32 47.15385   -0.01556   -0.04394     587.8840
#> 32879 713912.5 9559672 2238.00 31.97677   -0.00278   -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.97790    0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.22372   -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.79167   -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.16127    0.00129    0.00171     788.0004
#> 18440 714862.5 9558932 2448.29 41.05243   -0.04989   -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.21547    0.03890   -0.00911     173.2806
#> 4588  713852.5 9558612 2308.97 52.18041   -0.01059   -0.07431    4079.0366
#> 34615 713242.5 9560732 1879.60 27.59365    0.00384    0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.90611   -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.91167    0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.81029    0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.98136    0.05379    0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.32957    0.02247    0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.52758   -0.00193    0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.60379   -0.01109    0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.04570   -0.00789   -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.59002   -0.00253    0.00753    1337.7854
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281    0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.83544    0.00181   -0.00001    1184.7952
#> 39624 715222.5 9559552 2205.54 23.89750    0.00401   -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.56633   -0.00693    0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.85782   -0.01106   -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.87651    0.00010    0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.80875   -0.01881    0.00121    1783.0658
#> 42617 714092.5 9558392 2469.21 42.79995   -0.01397    0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.30642   -0.00019    0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.39754    0.00143    0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.89177    0.00445    0.00875     597.2893
#> 7771  715232.5 9560922 1816.88 37.18038    0.00124    0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.87118   -0.00212    0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.03638   -0.03675   -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.28730   -0.00144   -0.00856    1789.1596
#> 14779 712502.5 9560202 2004.61 27.52260    0.00115    0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.32360    0.00758   -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.98938   -0.00983    0.01253     528.1924
#> 48461 714672.5 9559272 2316.84 40.38149   -0.01966   -0.02234     849.0547
#> 41408 713172.5 9559612 2056.81 22.27259   -0.00166    0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.74912    0.00840    0.00189     556.5211
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585   -0.02665    1904.4769
#> 31983 714302.5 9559912 2204.77 23.46491   -0.01332    0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.62521    0.00159   -0.01559     866.5775
#> 30809 712532.5 9560032 1928.24 40.89945    0.00570    0.02230     714.7432
#> 16464 715252.5 9559982 2143.24 23.18588   -0.00298   -0.00712    1010.5097
#> 15558 714732.5 9559502 2283.66 53.43175    0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.15324    0.00156   -0.00116    1817.5222
#> 37832 712552.5 9559192 1998.10 41.55090   -0.00195   -0.00995    1026.7981
#> 47005 714942.5 9559192 2333.66 29.22085   -0.04937   -0.02273   72123.6172
#> 34347 714742.5 9560142 2161.95 26.56691   -0.00978    0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.49622   -0.00538   -0.00052    3508.5591
#> 8413  713242.5 9561232 1967.45 38.19451    0.00236    0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.24783    0.00737    0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.08505   -0.00196   -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.03217    0.00490   -0.00780     825.4451
#> 36245 713142.5 9560572 1827.82 43.67600   -0.00410    0.00400    4126.7812
#> 34810 714112.5 9559572 2337.51 10.28803   -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.11809   -0.01322   -0.00759    4131.5356
#> 13908 715022.5 9559372 2290.81 27.24300   -0.03004    0.00714  480535.5000
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.98875   -0.00270 1018.51947     842.1514
#> 44823 714202.5 9557442 2548.13 43.23024   -0.00233    0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.93765   -0.00070   -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480    0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649  100.79199     107.2986
#> 26768 715092.5 9560582 1960.72 36.65612    0.00980   -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.28426    0.00804   -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.39369   -0.02608    0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.29457    0.00537   -0.00577     423.5598
#> 14763 713462.5 9560852 1830.40 47.73999    0.01354    0.00907    1281.5044
#> 11394 713472.5 9561232 1889.48 34.31444    0.00108   -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.00219    0.00203    0.00238     787.0053
#> 22628 714942.5 9557732 2696.19 26.21167    0.00788   -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.88245    0.01014   -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.19775   -0.00595    0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.80144    0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.66261   -0.00983    0.00803    1058.5630
#> 32342 714862.5 9559622 2294.74 36.96724    0.01021    0.00109     915.4989
#> 47245 715372.5 9560792 1900.65 34.47086   -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.04329    0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.57188   -0.00463    0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.12459   -0.00166    0.00156     758.4840
#> 30296 713322.5 9559782 2042.96 52.62102   -0.02110   -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.27262   -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.62138    0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456   -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.29741    0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983