Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041    772.9746
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214   0.00466    724.0713
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114   4510.8833
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629 360379.1875
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593    939.0466
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991   1403.0193
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228   2262.3796
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389  47022.9688
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190    640.8842
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241   1752.9242
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812   1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970   1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417  34908.6562
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181  -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152   1530.5657
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905   6487.0010
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161   0.00171    756.8818
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399   1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079   1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788  -0.01029    894.2672
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491