Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813   1130.5791
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345   1120.6929
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311    586.7354
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378    440.9104
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044 129412.3984
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710    843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171    788.0004
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396    704.7076
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152   1530.5657
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367   1966.2112
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565   1410.8372
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528    355.3475
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552   2515.0752
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308    422.5022
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806