Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487    1077.7585
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257     357.1436
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406    2421.1145
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269     491.8017
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766  140591.4062
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769    1032.7177
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501    3889.7292
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070  138869.3906
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367     793.9665
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917     116.1986
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950     950.3138
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205     629.4165
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605     798.6775
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177     291.5937
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556     660.3690
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495     266.5888
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175    1122.9120
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252    5761.3486
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121    4159.6738
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299     688.0706
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026     247.0894
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620     379.0855
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599    1378.5760
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393    8301.6094
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520    3390.5708
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195    1901.6743
#> 29228 712782.5 9560422 1870.56 32.90611 -0.00218 -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321    1170.5904
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801    1071.6141
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426    1315.5334
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121    1783.0658
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152     850.4633
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856    1789.1596
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870    2099.7930
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189     556.5211
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781   29995.1543
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067    1123.7318
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150     503.5590
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181    1286.3949
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538     265.4677
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995    1026.7981
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431     315.9084
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065     704.4843
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734     335.0773
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422     967.9738
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759    4131.5356
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247     282.1076
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047    2826.7700
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177    1558.2231
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576    1045.1074
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233     968.7780
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709    1181.2555
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156     758.4840
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491