Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769