Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983