Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681  -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.9766069    0.00958   0.02642    671.1807
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969    368.5606
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917   0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753   0.01487   1077.7585
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273  -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104  -0.02046   1024.1573
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922   0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102   0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214   0.00466    724.0713
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021  -0.09328   2235.5259
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414  -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475   0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664  -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501   3889.7292
#> 29472 714792.5 9561072 1786.08 36.1731174   -0.01029  -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.1431560   -0.05634  -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151   0.00899   1394.2274
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173   8967.1924
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108  -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915   -0.02651  -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613  -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987   0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669  -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225   0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118   2359.5059
#> 46756 713372.5 9559192 2120.47 46.0050732    0.03696   0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.6343274   -0.00669  -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.1929823   -0.01846  -0.00034   4495.6929
#> 19784 714982.5 9557722 2716.98 35.4362937    0.00888   0.00862    874.5408
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379   0.01338  10787.6299
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442  -0.00178    800.1933
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093  -0.00907   1943.4092
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302   -0.01149   0.01949   1182.1506
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175   1122.9120
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119  -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175   0.00236   2715.4138
#> 375   715622.5 9558922 2510.87 37.3665885   -0.01017   0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364  -0.03357    473.4709
#> 39520 715342.5 9557622 2889.48 41.6488751   -0.01282   0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160   0.01690   5121.6924
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576  -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299    688.0706
#> 7668  713372.5 9559062 2164.82 37.9349627   -0.01807  -0.01883    616.1993
#> 20065 713772.5 9558852 2221.21 28.0199280   -0.00568  -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550  -0.00941   1290.7709
#> 31561 713302.5 9559192 2092.11 36.5277783    0.00406  -0.00055    773.0359
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611   0.02961   1086.6963
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394  -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388   -0.00481  -0.00399   3900.3445
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567  -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017   0.01213    328.7922
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.5724620    0.00025  -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521  -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431   4079.0366
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218  -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245    629.5983
#> 10689 713542.5 9559132 2211.95 31.9813582    0.05379   0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.5098139   -0.00428  -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593    0.03169   0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213  -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253   0.00753   1337.7854
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196   1052.5341
#> 4667  712652.5 9560472 1957.89 36.3243783   -0.00296   0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767  -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.0463191   -0.08153  -0.02417  34908.6562
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730  -0.02610   1872.9276
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144  -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.5226006    0.00115   0.01015    790.2421
#> 9864  714952.5 9557592 2752.40 31.9893796   -0.00983   0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234    849.0547
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025  -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.1257366   -0.03251  -0.00389  14516.2002
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586   0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378   0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196   0.00336   2806.2644
#> 48601 713692.5 9560152 2058.14 59.5572439   -0.01433  -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.4890128   -0.00226   0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085   2229.5105
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.7659191   -0.00458   0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825   0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480   0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626    450.7346
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937  -0.02273  72123.6172
#> 1085  715722.5 9557532 3097.22 28.7842537    0.02327   0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263   0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341   0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405   0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527  -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.1066333   -0.00007  -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175    0.00174  -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001   0.00759   1880.4403
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734    335.0773
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900    614.5898
#> 47059 713822.5 9558272 2413.20 38.1750320   -0.00687  -0.00223   1056.3701
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.3155672    0.04152   0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.6414056   -0.00307   0.02997    230.8641
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782  -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413   0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924   -0.01322  -0.00759   4131.5356
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650    766.6254
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119   0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.2863534    0.00165   0.00555   1882.0179
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635  -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873  -0.01417   1520.6630
#> 32430 715832.5 9558112 2824.39 23.6940967    0.00944   0.00836    250.5111
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273  -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952  -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354   0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.7668833   -0.00375   0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303    0.00105  -0.01576   1045.1074
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.7737308    0.00452   0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148    481.8292
#> 40653 714262.5 9561262 1791.94 12.2223357    0.00161   0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172  -0.00119    792.3544
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983   0.00803   1058.5630
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247   0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.3497490   -0.00771   0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491