Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769