Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491