Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603