Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005    769.3426
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096    520.5570
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961   1086.6963
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118   2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367   3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018    937.6911
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573   1490.0312
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250