Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174   2310.4871
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974   2136.6939
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684   2556.8049
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849   1120.5961
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161   3065.3694
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422   1159.2100
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483    343.1384
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744    226.7911
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655    130.9468
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116   1817.5222
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064    683.2795
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978  0.00367   3002.8828
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336