Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491