Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659   3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529   3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387   1989.3012
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856   0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276   9614.6924
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512  13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380    601.0843
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171   4425.5410
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905   6487.0010
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502   1428.8546
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135   1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491