Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.24719   -0.00501   -0.00649   1399.2329
#> 36052 714842.5 9558892 2483.27 68.78989   -0.04921   -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.30070    0.01633   -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.91056    0.01305   -0.00555    430.4604
#> 1300  714042.5 9558482 2408.27 24.14845    0.00659    0.01041    772.9746
#> 34512 713162.5 9559632 2041.52 46.81524   -0.00857    0.03677   1675.4679
#> 37104 713962.5 9561282 1795.56 25.74185    0.00273   -0.02403   1787.3024
#> 26531 714462.5 9560452 2104.74 19.52411    0.00453    0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.24149   -0.00841   -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.55641   -0.00730   -0.01491   1844.4353
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068    0.00588   1295.1610
#> 10974 715312.5 9557502 2844.23 26.66488   -0.00831   -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.95713   -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.71360    0.00194    0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.90038    0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.31521    0.00414   -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.38149   -0.01966   -0.02234    849.0547
#> 22283 713592.5 9560762 1832.60 36.01040    0.00448   -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.00467    0.00129    0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.92693   -0.00070   -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.10784   -0.00017   -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.85666    0.00471   -0.01382    560.8309
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440    0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.72122   -0.00491    0.00581   1855.1630
#> 4827  714962.5 9559882 2274.42 29.24491    0.00051    0.00889    362.8363
#> 32821 715022.5 9559952 2208.33 43.41645   -0.00643    0.00314   2360.1501
#> 34280 713652.5 9560702 1896.37 41.27359    0.00382    0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.38871    0.00071    0.00150   1274.3594
#> 6763  714692.5 9561052 1830.60 48.39488   -0.01996    0.00626   1014.7159
#> 1635  713102.5 9559052 2172.87 31.48403   -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.52508   -0.00089   -0.00051    779.0555
#> 34642 714962.5 9559862 2282.08 22.16086   -0.00028    0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.69432   -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.26190    0.01047   -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.54269   -0.00294   -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.19298   -0.01846   -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.35591    0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.81215    0.00463    0.00167   1938.5389
#> 17423 715592.5 9558562 2654.69 33.21952   -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.39488   -0.01996    0.00626   1014.7159
#> 44550 714842.5 9557642 2696.38 15.17307    0.00359   -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.16795    0.01008    0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.68617   -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.54692    0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.16814    0.00283    0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.14313   -0.00759   -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.65613   -0.00855   -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.79686   -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996    0.00478    0.01092    412.5849
#> 15766 715562.5 9558592 2646.16 29.02089    0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.39217    0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.92351   -0.01364   -0.03357    473.4709
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209    0.02739    365.7626
#> 35083 714612.5 9558152 2596.12 60.59831    0.04548    0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.21952   -0.00034    0.00294    739.5768
#> 31561 713302.5 9559192 2092.11 36.52778    0.00406   -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.26455   -0.00195   -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.23578    0.00927    0.01643    670.2289
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238   -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165    0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.47468    0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.96013   -0.00320    0.00180   3904.5254
#> 19892 714482.5 9560562 2059.63 19.81517   -0.00332   -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.49091   -0.06509   -0.01211  49347.5547
#> 7086  712762.5 9560962 2022.00 44.91359   -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.75679   -0.06953   -0.01797   7557.7866
#> 19832 715182.5 9557662 2761.27 38.43286   -0.00666   -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.25273   -0.00007   -0.02132   1580.1621
#> 1036  713042.5 9558892 2205.38 31.50981   -0.00428   -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511   -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.95886    0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.04570   -0.00789   -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.05925    0.02213   -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.02568    0.04250   -0.00970   1443.9200
#> 19756 715572.5 9558972 2485.95 38.81273    0.00546    0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.49701    0.00258    0.00422   1159.2100
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281    0.04759    158.9370
#> 20426 714702.5 9559712 2208.61 43.04632   -0.08153   -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.21550   -0.01016   -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.85782   -0.01106   -0.04634    500.5027
#> 48574 712692.5 9560412 1901.97 41.97833    0.01364   -0.00094   1392.6008
#> 46627 714852.5 9558932 2459.24 53.44550    0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.18038    0.00124    0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.49759   -0.00622    0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.99779    0.01354    0.00245    629.5983
#> 12074 714562.5 9560372 2045.17 26.28730   -0.00144   -0.00856   1789.1596
#> 35458 715242.5 9558442 2623.01 38.95483   -0.00770   -0.03400   1982.7782
#> 43834 713092.5 9560662 1905.55 31.32360    0.00758   -0.00288   2144.8164
#> 30378 714852.5 9557752 2651.70 41.29822   -0.00010   -0.00870   2099.7930
#> 23832 714252.5 9560182 2188.12 30.64751    0.00427    0.02403    474.8613
#> 17374 715712.5 9557952 2832.22 37.87480   -0.01519    0.00609   4275.3457
#> 4785  713792.5 9560472 2055.34 33.05279   -0.00586    0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.46491   -0.01332    0.01842    246.5161
#> 29431 714882.5 9559892 2258.03 21.91793   -0.00849    0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.32987    0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.60972   -0.00465   -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.48901   -0.00226    0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.67140   -0.01909   -0.01181   1286.3949
#> 20458 713612.5 9559562 2288.02 33.44011    0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.73330   -0.01174   -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.76592   -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.71742   -0.00494    0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.93224    0.03480    0.01500    538.8209
#> 39193 715332.5 9560852 1872.90 30.29056   -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.55332   -0.00502   -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.66583    0.01253   -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.22085   -0.04937   -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.15575    0.00176    0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.75409    0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.10663   -0.00007   -0.00853   1354.6001
#> 8831  714372.5 9559852 2200.77 48.76043   -0.00457    0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.33999    0.01432   -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451    0.00236    0.00483   1097.3082
#> 33956 714082.5 9558712 2327.95 39.68420    0.02218    0.00902    218.7467
#> 36245 713142.5 9560572 1827.82 43.67600   -0.00410    0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.21686   -0.01292    0.00462    716.5762
#> 34345 715852.5 9557902 2874.08 43.65595    0.00379   -0.01018    937.6911
#> 44267 712972.5 9559042 2191.01 22.26915    0.01462    0.00268    252.0072
#> 5656  713132.5 9560142 2042.24 33.59538   -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.93366   -0.00342    0.02552   2515.0752
#> 41855 713182.5 9560452 1851.53 45.17830    0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.98875   -0.00270 1018.51947    842.1514
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649  100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.65612    0.00980   -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.51535    0.00972    0.00308    422.5022
#> 3724  714322.5 9557402 2603.46 34.59748    0.02523    0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.44717    0.00360    0.00750   1892.8978
#> 48184 714202.5 9558222 2489.17 36.11353    0.02236   -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.90101    0.03044   -0.00964    543.0025
#> 44216 715772.5 9558402 2710.44 30.44469   -0.00070   -0.00050   1471.4346
#> 4943  712452.5 9559172 1927.46 34.46513    0.00105   -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.22166    0.00607   -0.04566  12862.4922
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909   -0.01151 143213.1562
#> 18504 713372.5 9560672 1786.89  6.47385   -0.00278   -0.01172 111973.8906
#> 41873 714832.5 9559942 2253.59 10.59800    0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.02355    0.00340    0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426   -0.00163   -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.90553   -0.01137   -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.21167    0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.88245    0.01014   -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.65727   -0.01305    0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.66261   -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.89051   -0.06986   -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.96439   -0.00933   -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.16930   -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086   -0.00011   -0.00419    323.0790
#> 32618 715132.5 9559642 2266.36 34.53274   -0.00305   -0.01135   1293.4855
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762