Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv      carea    cslope
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642   671.1807 41.600174
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712  3636.5303 29.022604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261   323.8441 15.665239
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896   276.0310 20.931294
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041   772.9746 27.530622
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324   861.3537 24.437796
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579  2246.7725 37.666818
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433  1475.3191 23.932447
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340   604.1032  6.084812
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491  1844.4353 30.171385
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606   564.6876 26.322827
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378   440.9104 24.774122
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234   849.0547 35.780068
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665 25318.3555 33.248168
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164   895.6606 28.757325
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392  1255.0884 33.618299
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328   592.3326 19.942369
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096  3210.5781 33.063103
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398   267.2659 24.586765
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710   843.7961 29.199648
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899  1394.2274 27.975237
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628   556.0121 43.531614
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889   362.8363 21.073961
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284  2010.4285 34.621548
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358   802.7622 22.717777
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739   558.2286 32.169861
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107  1392.5880 28.117331
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437  1155.1469 20.570904
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193  1809.3838 27.189139
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605   798.6775 30.291706
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051   779.0555 28.203847
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416  1204.8629 30.918521
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318   289.0063 15.648623
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313   783.4548 17.992594
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355 14919.6260 34.878806
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034  4495.6929 44.534291
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323   183.1836 35.319410
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556   660.3690 18.763795
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167  1938.5389 27.416030
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866 29195.6016 39.287716
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070  2562.7559 27.209193
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373 16031.7500 38.156697
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270  4057.5637 24.928248
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949  1182.1506 45.524362
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181  2406.1252 23.449444
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092   412.5849 21.719111
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236  2715.4138 27.013814
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252  5761.3486 34.251417
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197   711.1091 27.974091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357   473.4709 27.063088
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181 11655.4316 33.773570
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332  1418.6622 34.900005
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409  2312.8413 29.168135
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299   688.0706 27.323211
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473  2930.2791 33.754663
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026   247.0894 29.021458
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941  1290.7709 37.362578
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304  1569.7313 27.831425
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656  3036.7383 29.770887
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961  1086.6963 34.342517
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851  1012.5963 32.215698
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555  4439.0967 25.229624
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112   745.3315 28.727531
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393  8301.6094 31.788271
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367   806.0966 26.541697
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832  3022.3462 30.731737
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793   864.9472 37.920066
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335   195.0293 19.325866
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412  1350.3611 29.738228
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171   788.0004 25.761328
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431  4079.0366 31.664513
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334  4905.2002 42.773591
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372  4154.5786 30.399422
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132  1580.1621 41.642573
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833   369.3771 26.597847
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801   172.7491 30.648085
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522  1355.7436 25.726951
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739   789.7573 39.314072
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363   565.5207 28.038836
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970  1443.9200 32.180748
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484   495.0292 30.116381
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981  3975.3608 45.093497
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778  4594.4351 28.987654
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224  1137.3363 27.600523
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109  2047.9031 29.273560
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140  1066.3957 27.052202
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044   288.5468 34.691449
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450  2589.9910 27.720271
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705   413.5702 41.874047
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119  6589.5796 30.320354
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606   919.4241 41.856859
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358   802.7622 22.717777
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279 39495.1797 32.187050
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744   226.7911 32.011725
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867  2247.8748 31.598622
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230  1044.6443 16.784226
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253   528.1924 20.820713
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345  2417.1389 33.245303
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389 14516.2002 32.263253
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320   911.5911 28.093267
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067  1123.7318 26.486120
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230   714.7432 36.366204
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419   761.5427 39.291154
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403   241.2991 32.295339
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463  1050.7473 43.511561
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118  2273.4758 36.362194
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466  2633.9666 42.063124
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995  1026.7981 36.446418
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199   836.4196 27.986124
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476 11992.5273 34.261157
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052  3508.5591 31.257140
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853  1354.6001 26.679207
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065   704.4843 32.092512
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483  1097.3082 39.889895
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033  3446.5520 31.248545
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780   825.4451 28.893116
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983   270.7949 38.077629
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223  1056.3701 25.759610
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759  4131.5356 28.853009
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018   937.6911 44.395062
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038  2017.3547 32.236897
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948  1222.5433 30.485938
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955   713.4463 31.272609
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650   766.6254 38.506202
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863  1098.3462 28.701748
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251   168.1733 28.450219
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599 12873.8018 22.878205
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060   383.3547 31.206719
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740   909.4791 18.958600
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710   472.0236 32.282161
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577   423.5598 26.992042
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438  1276.5448 32.451184
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347  1893.3645 34.657644
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215   917.5296 41.510792
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624  1870.8307 32.099960
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282  1659.6211  7.987605
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208   842.4825 31.363137
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368  1071.8734 34.049163
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148   481.8292 24.257314
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270 35950.5508 32.711879
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935   903.6397 39.692797
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523   888.2811 26.856251
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029   894.2672 29.121153
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119   792.3544 27.569583
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064  1485.5201 29.185897
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157   766.0220 36.130719
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234  1707.3130 45.213246
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135  1293.4855 39.382827
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020   673.2861 36.819987
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525   162.0592  4.439850
#>       distroad slides distdeforest distslidespast log.carea
#> 23178   300.00   TRUE       300.00             21  2.826839
#> 28223    60.46   TRUE       152.09             35  3.560687
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 46178   300.00   TRUE       300.00             31  2.440958
#> 1300    300.00   TRUE       300.00              6  2.888165
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 24072   300.00   TRUE       300.00            100  3.351559
#> 21939   300.00   TRUE       300.00            100  3.168886
#> 39885   210.57   TRUE         0.00            100  2.781111
#> 32622   300.00   TRUE       300.00            100  3.265863
#> 7349    300.00   TRUE       300.00              8  2.751808
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 8019    300.00   TRUE       300.00             61  2.928936
#> 30800   300.00   TRUE       300.00            100  4.403435
#> 43964   300.00   TRUE       300.00             46  2.952143
#> 48498   300.00   TRUE       300.00            100  3.098674
#> 22283   256.35   TRUE         0.00            100  2.772566
#> 22933    87.56   TRUE       115.40              2  3.506583
#> 22566   300.00   TRUE         7.49              0  2.426944
#> 36030   300.00   TRUE         0.00            100  2.926238
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 47022   300.00   TRUE       300.00             26  2.745084
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 6676    300.00   TRUE       300.00            100  3.303289
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 4477    300.00   TRUE       300.00             11  2.746812
#> 20194   300.00   TRUE       300.00             41  3.143823
#> 13847   300.00   TRUE       300.00            100  3.062637
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 23141   300.00   TRUE       300.00            100  2.902371
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 27013   300.00   TRUE       168.05             18  3.080938
#> 34642   300.00   TRUE       300.00              2  2.460907
#> 45104   300.00   TRUE       300.00            100  2.894014
#> 47351   300.00   TRUE       300.00            100  4.173758
#> 47011   300.00   TRUE       300.00             10  3.652797
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 34359   300.00   TRUE       294.31            100  3.287475
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 36931   300.00   TRUE       300.00            100  3.408707
#> 34563   300.00   TRUE       300.00            100  4.204981
#> 17036    25.33   TRUE       208.61             15  3.608265
#> 651     300.00   TRUE       300.00            100  3.072673
#> 35194   300.00   TRUE       300.00            100  3.381318
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 40733   300.00   TRUE       300.00            100  3.433836
#> 21053   300.00   TRUE         0.00              2  3.760524
#> 375     300.00   TRUE       300.00            100  2.851936
#> 18892   235.02   TRUE        27.24              0  2.675293
#> 48808   300.00   TRUE       300.00            100  4.066528
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 44650   300.00   TRUE       300.00            100  3.110849
#> 32280   126.66   TRUE         0.00              5  3.195825
#> 46065   300.00   TRUE       251.11            100  3.482407
#> 38197   105.52   TRUE        65.32             50  3.036108
#> 2552    300.00   TRUE       300.00            100  3.005436
#> 26010    69.52   TRUE       166.56              2  3.647295
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 18897   300.00   TRUE        20.00              5  3.919162
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 37714    13.57   TRUE       170.00              4  3.480344
#> 12772   300.00   TRUE       300.00            100  2.936990
#> 1973    300.00   TRUE       300.00            100  2.290100
#> 32879   300.00   TRUE       300.00              6  3.130450
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 4588    300.00   TRUE       300.00              1  3.610558
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 29228   131.25   TRUE        80.55              2  3.618527
#> 28312   300.00   TRUE       300.00            100  3.198702
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 10689   300.00   TRUE       300.00             61  2.237416
#> 1036    300.00   TRUE       300.00            100  3.132178
#> 14545   300.00   TRUE       300.00             37  2.897494
#> 47244   300.00   TRUE       300.00              2  2.752448
#> 12629   300.00   TRUE       300.00            100  3.159543
#> 19756   300.00   TRUE       300.00            100  2.694631
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 49072   300.00   TRUE       300.00            100  3.662232
#> 23783   300.00   TRUE         0.00            100  3.055889
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 44612   300.00   TRUE       300.00            100  2.460216
#> 39405   300.00   TRUE       300.00            100  3.413298
#> 14950   300.00   TRUE       300.00             15  2.616549
#> 16065   300.00   TRUE       162.92            100  3.818858
#> 49459   300.00   TRUE       300.00            100  2.963516
#> 18498   300.00   TRUE       300.00             83  2.904587
#> 37115   300.00   TRUE       300.00            100  4.596544
#> 4479    300.00   TRUE       300.00             55  2.355626
#> 10213   300.00   TRUE       300.00            100  3.351772
#> 46971   300.00   TRUE       262.51            100  3.018968
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 10785   300.00   TRUE       300.00            100  3.383302
#> 17851   300.00   TRUE       300.00            100  4.161853
#> 6470    300.00   TRUE       300.00              2  2.959800
#> 4785    300.00  FALSE         0.00            100  3.050663
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 2664    300.00  FALSE       300.00            100  2.881694
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 14694   300.00  FALSE       300.00            100  3.021498
#> 35852   300.00  FALSE       300.00            100  3.356690
#> 26963   300.00  FALSE       300.00            100  3.420610
#> 37832   300.00  FALSE       207.18              2  3.011485
#> 43243   300.00  FALSE       300.00            100  2.922424
#> 20547   300.00  FALSE        42.56            100  4.078911
#> 18627   300.00  FALSE       300.00             20  3.545129
#> 31374   300.00  FALSE       300.00            100  3.131811
#> 4007    300.00  FALSE       300.00             68  2.847871
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 21925   300.00  FALSE       300.00             95  3.537385
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 47059   300.00  FALSE       300.00            100  3.023816
#> 29835   300.00  FALSE       300.00             45  3.616112
#> 34345   300.00  FALSE       300.00            100  2.972060
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 5656    300.00  FALSE        94.43            100  2.853361
#> 7983    300.00  FALSE        57.15             22  2.884583
#> 41855   234.40  FALSE        47.05            100  3.040739
#> 44715   300.00  FALSE       300.00            100  2.225757
#> 40909   234.82  FALSE         0.00             43  4.109707
#> 34532   300.00  FALSE       300.00             92  2.583601
#> 48988   300.00  FALSE       144.19            100  2.958793
#> 26768   300.00  FALSE       207.22            100  2.673964
#> 17628   300.00  FALSE       205.59            100  2.626915
#> 34336   300.00  FALSE       300.00            100  3.106036
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 26283   300.00  FALSE       300.00            100  2.962620
#> 40750   300.00  FALSE       300.00            100  3.272034
#> 18246    18.49  FALSE         0.00             29  3.220009
#> 24446   300.00  FALSE       300.00            100  2.925561
#> 7132    300.00  FALSE       225.26             12  3.030143
#> 37783   300.00  FALSE       300.00             18  2.682893
#> 20950   300.00  FALSE       300.00             81  4.555706
#> 1434    300.00  FALSE       300.00             35  2.955995
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 9526    300.00  FALSE       300.00            100  2.898919
#> 33881   300.00  FALSE       300.00            100  3.171879
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 14645   300.00  FALSE       300.00            100  3.232313
#> 32618   300.00  FALSE       300.00             56  3.111762
#> 11025   300.00  FALSE       300.00             50  2.828200
#> 28978   300.00  FALSE        10.00            100  2.209674