Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069    0.00958   0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199   0.00659    3553.7166
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753   0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.7668833   -0.00266   0.02896     276.0310
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498  -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272   -0.00373  -0.00057    1630.3259
#> 21939 713642.5 9558712 2299.09 33.7506519   -0.00503   0.02433    1475.3191
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529    3282.1186
#> 49372 714862.5 9560982 1863.12 20.9691094   -0.00295   0.00035    2697.5974
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783     377.8918
#> 22283 713592.5 9560762 1832.60 36.0103974    0.00448  -0.00328     592.3326
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973     514.5488
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774     527.1068
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669  -0.00950     950.3138
#> 6763  714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804  -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.4840308   -0.00064  -0.00916    2009.1686
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646  -0.04086    6042.5757
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743   0.01313     783.4548
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.5755084   -0.00255  -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 38686 714292.5 9557832 2496.22 38.4809914   -0.00509   0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316   0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666  -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066  -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288   0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093  -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034   0.00294     739.5768
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749   0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283   0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409   0.02771     674.3859
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762   0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.4376072   -0.02329  -0.01181   11655.4316
#> 10507 714892.5 9559282 2374.45 42.8423462   -0.01057   0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.8562718   -0.00458  -0.00401    1722.6481
#> 35083 714612.5 9558152 2596.12 60.5983082    0.04548   0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856  -0.01625    1063.9150
#> 46783 713592.5 9560762 1832.60 36.0103974    0.00448  -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.6781170   -0.02554   0.03304    1569.7313
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168  -0.04102   35685.0000
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904  -0.02013    4229.5708
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538   0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093   0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601  -0.08389   47022.9688
#> 7041  713472.5 9559092 2137.03 18.3489734   -0.10372  -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458  -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278  -0.00412    1350.3611
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129   0.00171     788.0004
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.9135886   -0.01259  -0.00281    1178.4420
#> 37720 715722.5 9557962 2820.42 45.0763086   -0.01474   0.01334    4905.2002
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542  -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247   0.01123     364.2903
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253   0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246   0.00396     704.7076
#> 28173 714562.5 9560382 2048.26 27.6858936   -0.00124  -0.00456    1531.4274
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546   0.01484     495.0292
#> 10692 712712.5 9561022 2005.85 41.4334430    0.01410  -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.6960182    0.02281   0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317   0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705  -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.4298025    0.00774  -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292  -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.8974967    0.00401  -0.00391    1794.7964
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 16065 715242.5 9560622 1871.61 34.3064209   -0.00019   0.00119    6589.5796
#> 10456 712442.5 9560292 2010.91 26.8917741    0.00445   0.00875     597.2893
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622   0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212   0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118  -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.0717460    0.00458   0.04291     293.4231
#> 11057 714962.5 9557642 2737.48 42.6859287   -0.00488  -0.00632     620.0229
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319   0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166   0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205    1112.2626
#> 36765 715302.5 9560262 1952.73 27.2034632   -0.00585  -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.7532859    0.00166  -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366   -0.03251  -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021   -0.01519   0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586   0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.0046187    0.00599   0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196   0.00336    2806.2644
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156  -0.00116    1817.5222
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738   0.00118    2273.4758
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165   0.00875    1165.3953
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197   0.00463     851.6306
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195  -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540   0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 16542 713772.5 9559622 2304.85 30.8102325   -0.00712   0.00923     811.5580
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973     215.5101
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527  -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507     504.3085
#> 12974 714662.5 9557332 2667.61 41.9416565   -0.00065   0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.0850495   -0.00196  -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.6767263    0.00706  -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.6842028    0.02218   0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410   0.00400    4126.7812
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041  -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224  -0.02036     476.9504
#> 34810 714112.5 9559572 2337.51 10.2880302   -0.00393   0.00032     994.9224
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424   0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978   0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.0930780   -0.00305  -0.00705     968.9738
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130     107.5831
#> 330   714232.5 9560362 2165.00 18.0607756    0.01142   0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858   0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009  -0.00599   12873.8018
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497   -0.00930  -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972   0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608   0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537  -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.1135298    0.02236  -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074  -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082   0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.8303070    0.01852   0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964    0.00537  -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.7737308    0.00452   0.00368    1071.8734
#> 4269  714162.5 9560572 2087.45  7.3344327    0.00293   0.00186    2751.4626
#> 11394 713472.5 9561232 1889.48 34.3144424    0.00108  -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.4738501   -0.00278  -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203   0.00238     787.0053
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172  -0.00119     792.3544
#> 46639 715232.5 9560042 2107.15 39.6572738   -0.01305   0.00585    5714.3267
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986  -0.05013     616.4128
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.4708598   -0.00011  -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305  -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336   0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247   0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775