Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491