Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149  12632.7793
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189    287.5425
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261    554.9243
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924   8023.8657
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714    475.3120
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036    516.4709
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058    431.4362
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236   2715.4138
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310   2833.9001
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132  42970.7891
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061   5744.7583
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450   2589.9910
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632    620.0229
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132    737.5982
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273  72123.6172
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476  11992.5273
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645    234.0114
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360   4288.7329
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983