Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(data, coords = c("x", "y"), nfold = 10, repetition = 1,
  seed1 = NULL, return_factor = FALSE)

Arguments

data
data.frame containing at least the columns specified by coords

coords
(ignored by partition_cv)

nfold
number of partitions (folds) in nfold-fold cross-validation partitioning

repetition
numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1
seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor
if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition) resampling objects. Each of these resampling objects is a list of length nfold corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[['1']][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx , ]

#>              x       y     dem    slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043   1806.4552
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250   1413.3312
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.64905 -0.00930 -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964    543.0025
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607