Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.1013441   0.00183   -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.6526691   0.00221    0.00969     368.5606
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475    1001.0861
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644    1179.0695
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005     769.3426
#> 28293 713762.5 9558812 2250.98 33.2854102   0.00118    0.00761     928.6858
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.7033271  -0.01407    0.00547     519.4228
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312  -0.02894   -0.03416     947.4878
#> 42598 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.8140475   0.00952    0.01378     440.9104
#> 8019  714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 22566 712832.5 9559842 1897.49 33.0602377   0.00672    0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.1813275  -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733   0.00151    0.00899    1394.2274
#> 47210 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694    2123.4473
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.4803764   0.00225    0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.0208853   0.00422    0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559   0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276    9614.6924
#> 17004 713102.5 9559032 2186.79 36.5335079  -0.01383    0.00193    1809.3838
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077    3559.4204
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096     520.5570
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169     350.1799
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025    4425.5151
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175     804.5679
#> 17423 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294     739.5768
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771     674.3859
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.0208853   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071  -0.00762    0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181   11655.4316
#> 42941 713252.5 9559742 1990.10 28.9544222  -0.04708    0.00408   76318.0156
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.9349627  -0.01807   -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294     739.5768
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102   35685.0000
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219     420.7032
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733  -0.00093    0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389   47022.9688
#> 12772 715172.5 9557642 2759.42 34.8054672  -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 11612 715212.5 9557752 2820.51 39.1622383   0.00017    0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.3251890   0.01246    0.00534     344.1447
#> 10933 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.1612670   0.00129    0.00171     788.0004
#> 19892 714482.5 9560562 2059.63 19.8151724  -0.00332   -0.00058    1817.2371
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.6704448   0.01489    0.00241    1752.9242
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078  622888.2500
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261    1627.8989
#> 15432 715632.5 9558922 2509.67 35.8276239  -0.00444   -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.9116671   0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.9597815  -0.01439   -0.01161    3065.3694
#> 43074 713172.5 9559622 2052.40 31.5275756  -0.00193    0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510  -0.00442   -0.00118   19557.6621
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753    1337.7854
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190     264.6404
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224    1137.3363
#> 7409  714772.5 9559122 2404.90 30.2155023  -0.01016   -0.00294     734.9908
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.9783258   0.01364   -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483     343.1384
#> 9606  714392.5 9559172 2353.05 30.7833035  -0.00185   -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969   0.00285   -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.7999473  -0.01397    0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.6859287  -0.00488   -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.5226006   0.00115    0.01015     790.2421
#> 6212  714592.5 9560362 2034.54 12.5225656  -0.01454    0.00144  317260.6875
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 21490 714672.5 9560382 2021.86 32.8041256  -0.00783   -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.6801711   0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427     866.9398
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301     338.9676
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.7659191  -0.00458    0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463   0.00825    0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500     538.8209
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502    1428.8546
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399    1601.4852
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780     825.4451
#> 47059 713822.5 9558272 2413.20 38.1750320  -0.00687   -0.00223    1056.3701
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048    1550.6704
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 35038 715092.5 9558152 2715.69 40.1838220   0.00270   -0.00250    2225.4426
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750    1892.8978
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066     306.0146
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248    1432.4019
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.3079371   0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803    1058.5630
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602   0.00298    0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234    1707.3130
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020     673.2861
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607