Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026