Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607