Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642     671.1807
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712    3636.5303
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250    1413.3312
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579    2246.7725
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761     928.6858
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440    1088.8639
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835    2425.3115
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416     947.4878
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164     895.6606
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096    3210.5781
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973     514.5488
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629  360379.1875
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950     950.3138
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276    9614.6924
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077     300.7915
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517    1000.7119
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025    4425.5151
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626    1014.7159
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779    2972.0193
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560     977.4719
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542     860.8840
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883     616.1993
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102   35685.0000
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656    3036.7383
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558    1504.9717
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448    1258.8196
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911     173.2806
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520    3390.5708
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583    1898.0105
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722    1373.4147
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970    1443.9200
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224    1137.3363
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294     734.9908
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121    1783.0658
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732    4237.6934
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744     226.7911
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036    2505.8369
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367    1966.2112
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177     997.3089
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158     837.3730
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452     578.5847
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409    1575.9425
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464    4029.5947
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014     271.4219
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538 -0.00052    3508.5591
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697     412.1185
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662    1269.2650
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223    1056.3701
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038    2017.3547
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740     909.4791
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308     422.5022
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502     520.5218
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904     608.3561
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215     917.5296
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151  143213.1562
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585    5714.3267
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124   25826.6914
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313