Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775