Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114   4510.8833
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512  13178.3223
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599   1378.5760
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 41753 715722.5 9558192 2771.39 26.085050  -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248   1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775