Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983