Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770    474.5059
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769   1032.7177
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312   2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485   2240.4185
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961   1086.6963
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123    364.2903
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298   1533.7925
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118  19557.6621
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905   6487.0010
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655    130.9468
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367   3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083   1309.2347
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902    218.7467
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884    672.2466
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946