Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336