Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.9766069    0.00958   0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123   0.00393   2081.0056
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199   0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813   1130.5791
#> 16749 715072.5 9558242 2745.27 49.0537816    0.00097   0.01483    566.7674
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659   0.01041    772.9746
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858  -0.00292   1233.9299
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310   0.01440   1088.8639
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284   0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542    674.0147
#> 26447 715632.5 9559102 2410.77 39.6567008    0.00730   0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271   -0.01407   0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922   0.01088    266.4355
#> 10974 715312.5 9557502 2844.23 26.6648828   -0.00831  -0.00149  12632.7793
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952   -0.03525  -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.8879661    0.01903   0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664  -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678  -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270  -0.00769   1032.7177
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032    994.6740
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044 129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832   0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.5350311    0.00856   0.01114   4510.8833
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051   0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490  -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643   0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774    527.1068
#> 6763  714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626   1014.7159
#> 29667 713892.5 9559712 2239.25 41.5927252    0.00804   0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887    366.8537
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294  -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823   -0.01846  -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527   -0.00904  -0.00866  29195.6016
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666  -0.00504   1020.2766
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903   2360.1536
#> 44550 714842.5 9557642 2696.38 15.1730683    0.00359  -0.00279    930.0871
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283   0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409   0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372  -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731   1384.7959
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762   0.01252   5761.3486
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063    576.2819
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.7269720   -0.02027  -0.01473   2930.2791
#> 10364 714762.5 9561002 1856.21 25.2748872   -0.00845   0.00454   3476.0818
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927  -0.00316    454.9674
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611   0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219    420.7032
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394  -0.00224    942.0322
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588   0.01448   1258.8196
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601  -0.08389  47022.9688
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822    472.8748
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453   0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205   1112.2626
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911    173.2806
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949   0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084  -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.6498183    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506   -0.02070  -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263   0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017    759.9937
#> 26428 713042.5 9560332 1895.15 33.9597815   -0.01439  -0.01161   3065.3694
#> 47086 713692.5 9559972 2175.51 25.0457041   -0.00789  -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.8039510   -0.00442  -0.00118  19557.6621
#> 14192 715162.5 9558122 2772.18 44.4626708   -0.02970   0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705  -0.00915   2036.0378
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016  -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267   0.00483    343.1384
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285  -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730  -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.4323251    0.04864   0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143   0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124   0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622   0.00152   1530.5657
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004  -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.7925059   -0.01400  -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529   0.01230   1044.6443
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010  -0.00870   2099.7930
#> 17851 715362.5 9557482 2869.26 29.1257366   -0.03251  -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781  29995.1543
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332   0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.2968763   -0.00079  -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159  -0.01559    866.5775
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465  -0.00105   4432.8760
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085   2229.5105
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336   0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007  -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738   0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544   -0.00812  -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402  -0.01328   1066.7134
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409   1575.9425
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937  -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544  -0.01014    271.4219
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341   0.04431    315.9084
#> 31374 714652.5 9560592 1999.69 27.1066333   -0.00007  -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.9979810    0.00990   0.02790    209.6952
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844   0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.2735807   -0.00022   0.00662   1269.2650
#> 33370 714202.5 9558962 2472.93 32.6414056   -0.00307   0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224  -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782  -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.1180924   -0.01322  -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517   -0.00068  -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379  -0.01018    937.6911
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066   0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667   0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.3936922    0.03479   0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119   0.00361   1011.0248
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262  -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.9676002   -0.00820  -0.00260   1484.6593
#> 26283 715932.5 9557592 3075.48 32.7668833   -0.00375   0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303    0.00105  -0.01576   1045.1074
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865  -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253  -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203   0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196  -0.01935    903.6397
#> 40653 714262.5 9561262 1791.94 12.2223357    0.00161   0.00709   1181.2555
#> 35257 713822.5 9557982 2350.80 29.3079371    0.01549  -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732   0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.2771580   -0.07627  -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573   1490.0312
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241   0.00020    673.2861
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110  -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.8520794    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603