Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684   2556.8049
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304   1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512  13178.3223
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212  46310.0195
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118  19557.6621
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419    761.5427
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734    335.0773
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400   4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674