Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946