Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029    5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183   -0.09203     634.3320
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347    0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199    0.00659    3553.7166
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741     746.4155
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835    2425.3115
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550   -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677    1343.5006
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664   -0.05766  140591.4062
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272    4955.5586
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250    3686.1240
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672    1114.8801
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151    0.00899    1394.2274
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114    4510.8833
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051    0.00889     362.8363
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739     558.2286
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276    9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991    1403.0193
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504    1020.2766
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395    0.01495     266.5888
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283    0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542     860.8840
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542    3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845    0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406   -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664    1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242     507.9263
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245    6353.3179
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212   46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236    9755.3027
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384    0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722    1373.4147
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801    1071.6141
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181   -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426    1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094    1392.6008
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610     454.8103
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675   -0.01905    6487.0010
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893   -0.00867    2247.8748
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036    2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367    1966.2112
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559     866.5775
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109     605.1908
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626     450.7346
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253   -0.00333     422.8840
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943     371.5514
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033    3446.5520
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301   -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036     476.9504
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196     576.1450
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948    1222.5433
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537   -0.00577     423.5598
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044   -0.00964     543.0025
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343    0.01177    1558.2231
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624    1870.8307
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491     261.7371
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377    1727.6055
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281     492.4785
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387