Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183   -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921   -0.12438    753.6304
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221    0.00969    368.5606
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514   -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333   -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503    0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104   -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891   -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453    0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841   -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284    0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.97109  0.00183    0.03387   1989.3012
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924    0.02606    564.6876
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894   -0.03416    947.4878
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952    0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581    0.01189    287.5425
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123    0.01164    895.6606
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430   -0.06070 138869.3906
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613   -0.01898    588.3761
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664    0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071    0.00150   1274.3594
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252    0.01118   2359.5059
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996    0.00626   1014.7159
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064   -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765    0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184    0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440   -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743    0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543   -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154    0.00124   2114.2253
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353    0.00063   5035.0000
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856    0.01796   4729.6382
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940    0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060    0.00270   4057.5637
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735   -0.00925    665.6116
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395    0.01495    266.5888
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372   -0.00008 825875.6875
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504    0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017    0.00197    711.1091
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048   -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282    0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160    0.01690   5121.6924
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529    0.02009   1247.7061
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448   -0.00328    592.3326
#> 664   714452.5 9559852 2174.83 44.86088  0.00282    0.00109    767.4097
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410    0.03620    379.0855
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394   -0.00224    942.0322
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024   -0.00474   2343.2100
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169   -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372   -0.00008 825875.6875
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759   -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129    0.00171    788.0004
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890   -0.00911    173.2806
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059   -0.07431   4079.0366
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175    0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007   -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166    0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379    0.00801    172.7491
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511   -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789   -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253    0.00753   1337.7854
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669   -0.01981   3975.3608
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181   -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705   -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774   -0.00224   1137.3363
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010    0.00140   1066.3957
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881    0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257    0.00044    288.5468
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019    0.00119   6589.5796
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216   -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124    0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902    0.01358    802.7622
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488   -0.00632    620.0229
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465   -0.01036   2505.8369
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683   -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996    0.04684    209.1778
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519    0.00609   4275.3457
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534    0.01476   1275.9446
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295    0.02715    520.7355
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363   -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174   -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220   -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195   -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.30298  0.00540    0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075   -0.01945    259.7421
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253   -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176    0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263    0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341    0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167    0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007   -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196    0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065    0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846    0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069   -0.00900    614.5898
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236    0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737    0.01033   3446.5520
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674   -0.00376    629.8236
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423   -0.00983    270.7949
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152    0.00098    164.0802
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338    0.00518   1137.0179
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379   -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004    0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978    0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452   -0.00948   1222.5433
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950   -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118    0.00162    290.2031
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070   -0.00060    383.3547
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635   -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873   -0.01417   1520.6630
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944    0.00836    250.5111
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070   -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264    0.00624   1870.8307
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452    0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284   -0.00886  43249.3945
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196   -0.01935    903.6397
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137   -0.01523    888.2811
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595    0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983    0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181   -0.00799   1036.6606
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166    0.00156    758.4840
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983