Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.667957  -0.00292    0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542     674.0147
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382     560.8309
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385     674.0721
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088    2322.1746
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373     527.3528
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809    1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380     601.0843
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152    1530.5657
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015     790.2421
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697     412.1185
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251     168.1733
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674