Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660   -0.01919  -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.95233   -0.00123   0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.97397    0.02056  -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.44641   -0.00347   0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.22409   -0.00199   0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.30070    0.01633  -0.01813   1130.5791
#> 1300  714042.5 9558482 2408.27 24.14845    0.00659   0.01041    772.9746
#> 42917 714202.5 9557412 2544.08 39.87786   -0.02104  -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.06326    0.00940   0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.40193    0.00310   0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010   1527.1028
#> 125   713992.5 9561062 1878.28 49.97109    0.00183   0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730   0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.29553    0.00550  -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.86543    0.00102   0.00388   1051.4222
#> 38287 714872.5 9561162 1719.41 48.23560   -0.03021  -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.71360    0.00194   0.00526    539.8119
#> 39061 712922.5 9559912 1931.50 36.31521    0.00414  -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.15327   -0.07664  -0.05766 140591.4062
#> 45869 714342.5 9560282 2127.11 27.62344   -0.02169  -0.00160   9344.2930
#> 48055 712862.5 9559972 1902.48 29.01687   -0.00720   0.01250   3686.1240
#> 22933 712712.5 9560432 1902.17 38.12347   -0.00746   0.00096   3210.5781
#> 14522 715012.5 9557732 2721.34 35.18133   -0.00382   0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.62670   -0.01537  -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.04492    0.00832   0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.53503    0.00856   0.01114   4510.8833
#> 36230 713322.5 9560662 1798.99 40.12825   -0.00033  -0.03917    116.1986
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669  -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.48038    0.00225   0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.02089    0.00422   0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.48042    0.00252   0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804  -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.53351   -0.01383   0.00193   1809.3838
#> 28577 712812.5 9559962 1849.82 59.78012    0.00646  -0.04086   6042.5757
#> 29667 713892.5 9559712 2239.25 41.59273    0.00804   0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.26190    0.01047  -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224   0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.35591    0.04547   0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.44543    0.00606  -0.00556    660.3690
#> 38686 714292.5 9557832 2496.22 38.48099   -0.00509   0.00499    885.0064
#> 46201 714832.5 9561022 1845.94 34.57399   -0.00856   0.01796   4729.6382
#> 5578  714962.5 9557612 2751.95 39.19146    0.00288   0.01452    437.3965
#> 37141 714382.5 9559822 2219.36 10.10640    0.01661   0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.93664   -0.02617   0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.85173   -0.01149   0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.16795    0.01008   0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.45754    0.03395   0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.54692    0.03480   0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.48806    0.02409   0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.79686   -0.00449  -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996    0.00478   0.01092    412.5849
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762   0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.26460   -0.00211   0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.40283   -0.00140  -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.64888   -0.01282   0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221   0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.84235   -0.01057   0.02427    380.9472
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209   0.02739    365.7626
#> 32735 714922.5 9559002 2414.49 40.06465    0.00062  -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.68860    0.00114   0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.01993   -0.00568  -0.02553   1750.4175
#> 31561 713302.5 9559192 2092.11 36.52778    0.00406  -0.00055    773.0359
#> 16355 712962.5 9560212 1992.88 39.52607    0.01927  -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104   0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.27910    0.00645  -0.00395   2011.3270
#> 24501 713302.5 9559142 2125.66 27.10893    0.01784   0.00596    842.0994
#> 26010 712852.5 9560542 1928.25 30.97811    0.00165   0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.96013   -0.00320   0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.75542   -0.00604   0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.77108   -0.02605  -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.36887   -0.00093   0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.95066    0.03743  -0.05242    507.9263
#> 18440 714862.5 9558932 2448.29 41.05243   -0.04989  -0.02781  29995.1543
#> 47702 714692.5 9559712 2200.51 33.09863   -0.13038  -0.04212  46310.0195
#> 20797 713012.5 9560392 1842.25 36.64982    0.01299   0.00141   4611.0718
#> 15432 715632.5 9558922 2509.67 35.82762   -0.00444  -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.90611   -0.00218  -0.00372   4154.5786
#> 26428 713042.5 9560332 1895.15 33.95978   -0.01439  -0.01161   3065.3694
#> 28173 714562.5 9560382 2048.26 27.68589   -0.00124  -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.90646   -0.00070   0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.82993   -0.03669  -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.44603    0.01003   0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.40342    0.03051   0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.26762   -0.00317   0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.89043    0.00042   0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.27110    0.00705  -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.32798   -0.00292  -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.95926   -0.00767  -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.21550   -0.01016  -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.72045    0.01267   0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.25525    0.01257   0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.64619   -0.02110  -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.78330   -0.00185  -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.79995   -0.01397   0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.28582    0.00708  -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.32187   -0.00157  -0.00152    850.4633
#> 18498 714842.5 9557712 2679.04 28.93953    0.00902   0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.23119    0.01895   0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.72390   -0.00529   0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.32360    0.00758  -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186   0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.80778   -0.00486   0.02306   1718.3652
#> 17374 715712.5 9557952 2832.22 37.87480   -0.01519   0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.75679   -0.06953  -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.05243   -0.04989  -0.02781  29995.1543
#> 32405 713022.5 9558982 2208.85 15.00462    0.00599   0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.34766   -0.02196   0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.91793   -0.00849   0.00150    503.5590
#> 48601 713692.5 9560152 2058.14 59.55724   -0.01433  -0.02177    997.3089
#> 31071 713472.5 9558462 2298.33 29.15324    0.00156  -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369   -0.00124   0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.31579   -0.00738   0.00118   2273.4758
#> 20458 713612.5 9559562 2288.02 33.44011    0.00112   0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.18933    0.00815  -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.28383    0.00197   0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.93224    0.03480   0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.49492    0.01306  -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.38952   -0.00724   0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.29056   -0.00161   0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.82590    0.00846   0.00794   1363.5321
#> 18627 714292.5 9559252 2330.78 18.49622   -0.00538  -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.10663   -0.00007  -0.00853   1354.6001
#> 34616 714522.5 9558742 2568.77 34.99798    0.00990   0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.48633   -0.00069  -0.00900    614.5898
#> 31201 714672.5 9560022 2224.47 38.86373    0.01180   0.01170    253.5129
#> 47059 713822.5 9558272 2413.20 38.17503   -0.00687  -0.00223   1056.3701
#> 1691  714222.5 9559022 2448.69 38.21686   -0.01292   0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.31557    0.04152   0.00098    164.0802
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224  -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.21475   -0.01338   0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.07473   -0.00112  -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.79731    0.01978   0.01282   2322.7944
#> 330   714232.5 9560362 2165.00 18.06078    0.01142   0.00868    235.8511
#> 44823 714202.5 9557442 2548.13 43.23024   -0.00233   0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.75618    0.00119   0.00361   1011.0248
#> 28286 713702.5 9559842 2245.07 27.96893    0.00511   0.01499    617.3855
#> 31098 714352.5 9559042 2410.15 30.33640    0.00537  -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.46513    0.00105  -0.01576   1045.1074
#> 48050 714022.5 9560572 2099.82 18.71395    0.01182  -0.00972    331.7462
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909  -0.01151 143213.1562
#> 20950 714912.5 9557782 2667.71 30.81138   -0.04770   0.00270  35950.5508
#> 40653 714262.5 9561262 1791.94 12.22234    0.00161   0.00709   1181.2555
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732   0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.66261   -0.00983   0.00803   1058.5630
#> 32618 715132.5 9559642 2266.36 34.53274   -0.00305  -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.12459   -0.00166   0.00156    758.4840
#> 40852 715392.5 9557932 2939.71 18.29741    0.00505   0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983