Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347    0.02357   1752.0375
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339   1002.5877
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046   1024.1573
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951   -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542    674.0147
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510   3342.5518
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746    0.00096   3210.5781
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773    179.3077
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950    950.3138
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605    798.6775
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318    289.0063
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224    0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904   -0.00866  29195.6016
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140   -0.00070   2562.7559
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925    665.6116
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194    0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658    0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221    0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306   -0.00436   2602.7390
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625   1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114    0.03026    247.0894
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927   -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395   2011.3270
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013   4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291   -0.00081    940.6002
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469    0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335    195.0293
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521   -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241   1752.9242
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263    0.01322   4605.1411
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193    0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801   1071.6141
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546    0.01484    495.0292
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756    353.9086
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610   1872.9276
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732   4237.6934
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384    947.2075
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144 317260.6875
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665   1904.4769
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336   2806.2644
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759    0.00158    837.3730
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204    853.2902
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493   2212.0525
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538    265.4677
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197    0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945    259.7421
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176    0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341    0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133    828.8604
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367   3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853   1354.6001
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900    614.5898
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514    897.9290
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312    593.7773
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950   -0.05130    107.5831
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608    0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820   -0.00260   1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607   -0.04566  12862.4922
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233    968.7780
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340    0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377   1727.6055
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983    0.00803   1058.5630
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573   1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751    0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983