Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969    368.5606
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896    276.0310
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629 360379.1875
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083    994.7876
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543 -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473   2930.2791
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262   1760.2162
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102  35685.0000
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372   4154.5786
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.8765069  0.00010  0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655    130.9468
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517   1060.5494
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900    614.5898
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196    576.1450
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234   1707.3130
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491