Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896   -0.00501  -0.00649    1399.2329
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056  -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199   0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969     368.5606
#> 40756 714022.5 9558862 2331.20 45.0854759   -0.00075   0.00475    1001.0861
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240  -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272   -0.00373  -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333  -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835   -0.02191  -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090   0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858  -0.00292    1233.9299
#> 28293 713762.5 9558812 2250.98 33.2854102    0.00118   0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645   0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284   0.01804     628.5153
#> 16811 714902.5 9559262 2362.67 50.7033271   -0.01407   0.00547     519.4228
#> 10974 715312.5 9557502 2844.23 26.6648828   -0.00831  -0.00149   12632.7793
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952   0.01378     440.9104
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164     895.6606
#> 40307 713302.5 9560702 1845.07 39.0046749    0.00129   0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.5537501    0.01412   0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.1855058    0.00602   0.01518     854.4269
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032     994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832   0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440   0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973     514.5488
#> 45575 715432.5 9559452 2305.28 24.2343959    0.00993  -0.00413     364.9788
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987   0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257   0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.3887102    0.00071   0.00150    1274.3594
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225   0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118    2359.5059
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077    3559.4204
#> 19128 714852.5 9557872 2681.49 20.7576880    0.00453   0.01077     300.7915
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224   0.00177     291.5937
#> 37348 715632.5 9558182 2743.16 13.6203527   -0.00904  -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353   0.00063    5035.0000
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379   0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316   0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666  -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 17036 712832.5 9560582 1951.38 27.6457866    0.00060   0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.5997894   -0.00735  -0.00925     665.6116
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395   0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480   0.02530     937.1306
#> 421   714892.5 9561022 1849.01 26.1320321    0.00002   0.00588     749.2502
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364  -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140  -0.02330    1456.9246
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529   0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739     365.7626
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126  -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168  -0.04102   35685.0000
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645  -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125   -0.01972   0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.2221961    0.00099   0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093   0.00664    1807.2659
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278  -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986  -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.7916671   -0.04376  -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175   0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263   0.01322    4605.1411
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542  -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247   0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.9750697    0.00511  -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497    0.01011   0.00739     789.7573
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196    1052.5341
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317   0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042   0.00957     495.4453
#> 23783 713062.5 9560252 1983.09 36.4298025    0.00774  -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605   0.00075    2170.2249
#> 14494 715332.5 9558792 2557.22 20.7204457    0.01267   0.00483     343.1384
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915  -0.02705     413.5702
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730  -0.02610    1872.9276
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622   0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.1446021   -0.04360  -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.9395253    0.00902   0.01358     802.7622
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004  -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.7925059   -0.01400  -0.01210     585.8992
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319   0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.7000149    0.00683  -0.00603    1562.0503
#> 21490 714672.5 9560382 2021.86 32.8041256   -0.00783  -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.7532859    0.00166  -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366   -0.03251  -0.00389   14516.2002
#> 31253 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797    7557.7866
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203   0.00642    1688.1782
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505     351.4078
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909  -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.4317521    0.00390  -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007  -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738   0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999   -0.00533   0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165   0.00875    1165.3953
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253  -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973     215.5101
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133     828.8604
#> 36126 715052.5 9558292 2759.47 29.0621382    0.00196   0.01734     335.0773
#> 31884 715782.5 9558612 2632.60 30.2796736    0.00084  -0.00083    1309.2347
#> 33175 715822.5 9558832 2563.13 36.6228893   -0.00674  -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983     270.7949
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301  -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782  -0.00048    1550.6704
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424   0.00196     576.1450
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780   -0.00305  -0.00705     968.9738
#> 21070 712832.5 9560062 1884.42 50.9336562   -0.00342   0.02552    2515.0752
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418   0.00513     161.9878
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650     766.6254
#> 330   714232.5 9560362 2165.00 18.0607756    0.01142   0.00868     235.8511
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262  -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537  -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074  -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952  -0.00438    1276.5448
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600  -0.01849     764.8193
#> 4943  712452.5 9559172 1927.46 34.4651303    0.00105  -0.01576    1045.1074
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491     261.7371
#> 20950 714912.5 9557782 2667.71 30.8113784   -0.04770   0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.4738501   -0.00278  -0.01172  111973.8906
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309   0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172  -0.00119     792.3544
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253    1237.0536
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110  -0.02931    1812.3820
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491