Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659   3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555    430.4604
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403   1787.3024
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357    275.7474
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566  -0.03355  14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993  10308.6406
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260  -0.03310   2833.9001
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401   1722.6481
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240   1201.4724
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556  -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351   1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812   1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077   3056.9475
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150    503.5590
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565   1410.8372
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900    614.5898
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011    712.5228
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499    617.3855
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624   1870.8307
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013    616.4128
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607