Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958    0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183   -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056   -0.06976    301.2347
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021   -0.01579   1318.7560
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863   -0.00507 159065.0781
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659    0.01041    772.9746
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498   -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191   -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090    0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273   -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118    0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104   -0.02046   1024.1573
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407    0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607    0.00677   1343.5006
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831   -0.00149  12632.7793
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769   -0.00281   1325.6138
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414   -0.00104   1641.4587
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017   -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471   -0.01382    560.8309
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126   -0.02044 129412.3984
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537   -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151    0.00899   1394.2274
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716    0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993   -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206    0.02773    179.3077
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071    0.00150   1274.3594
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452   -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383    0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064   -0.00916   2009.1686
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804    0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294   -0.00096    520.5570
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547    0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229   -0.00511 361979.2812
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066   -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843    0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006    0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661    0.01199    196.4080
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395    0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480    0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008 825875.6875
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194    0.00516    576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762    0.01252   5761.3486
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658    0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221    0.00121   4159.6738
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209    0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027   -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856   -0.01625   1063.9150
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099    0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282    0.00109    767.4097
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927    0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410    0.03620    379.0855
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028    0.01318    289.0063
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588    0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940    987.8085
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332   -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491    0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989   -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849   1120.5961
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384    0.00175   1026.6818
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218   -0.00372   4154.5786
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542   -0.00722   1373.4147
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193    0.02812   1423.7578
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246    0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070    0.00010   3467.2690
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042    0.00957    495.4453
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200   1305.6691
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016   -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881    0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185   -0.01185    953.8614
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157   -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360   -0.03000 162178.1562
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458    0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488   -0.00632    620.0229
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895    0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400   -0.01210    585.8992
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585   -0.02665   1904.4769
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084   -0.02996   1599.3491
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872    0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336    0.01403    241.2991
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738    0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.6402737  0.00363   -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242    0.01132    737.5982
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306   -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309   -0.03409   1575.9425
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712    0.00923    811.5580
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527   -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978    0.00367   3002.8828
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174   -0.03065    704.4843
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674   -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410    0.00400   4126.7812
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068   -0.00312    593.7773
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305   -0.00705    968.9738
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066    0.05257    549.3252
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947    842.1514
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229    0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911    100.0000
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262   -0.00502    520.5218
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343    0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215    917.5296
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361   -0.01841    756.0055
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452    0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284   -0.00886  43249.3945
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340    0.01809   1158.0426
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595    0.03064   1485.5201
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021    0.00109    915.4989
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288    0.00448    941.8503
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023   -0.00227    841.2025
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983