Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032    994.6740
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304   1569.7313
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961   1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119   6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166   1813.7841
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208    951.3401
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223   1056.3701
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803   1058.5630
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901