Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357    1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403    1787.3024
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124    2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167    1938.5389
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684    2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102   35685.0000
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849    1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722    1373.4147
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196    1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298     436.1459
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603