Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542   0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104   1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937   0.00083    994.7876
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357    275.7474
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887    366.8537
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197    711.1091
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542   3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281   1178.4420
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321   1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801   1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363    565.5207
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294    734.9908
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185    953.8614
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548    532.8210
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948   1222.5433
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650    766.6254
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347   1893.3645
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491    261.7371
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013    616.4128
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769