Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659   3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712   3636.5303
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579   2246.7725
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189    287.5425
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903    0.01447    331.4833
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032    994.6740
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398    267.2659
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416   1204.8629
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991   1403.0193
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796   4729.6382
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060    0.00270   4057.5637
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560    977.4719
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565   2232.6133
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283    590.1536
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009   1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473   2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020   1563.5549
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664   1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389  47022.9688
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822    472.8748
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245   6353.3179
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849   1120.5961
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384    0.00175   1026.6818
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079    0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879   1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632    620.0229
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234    849.0547
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609   4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067   1123.7318
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345 665381.5625
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500    538.8209
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794   1363.5321
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759   1880.4403
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084   -0.00083   1309.2347
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983    270.7949
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032    994.9224
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282   2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260   1484.6593
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050   1471.4346
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491    261.7371
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012    305.3624
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119    792.3544
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057    722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775