Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029    5577.3916
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739     558.2286
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167    1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907    1943.4092
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588     749.2502
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870    1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661    2508.5374
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981    3975.3608
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171    4425.5410
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526     706.1463
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291     487.3438
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052    3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943     371.5514
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312     254.0656
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281     492.4785
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149    4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983