Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438    753.6304
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174   2310.4871
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279    930.0871
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771    674.3859
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316    454.9674
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851   1012.5963
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822    472.8748
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298   1533.7925
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177    997.3089
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064    683.2795
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548    532.8210
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273  72123.6172
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223   1056.3701
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836    250.5111
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983