Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea   cslope
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056 37.59291
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166 27.81538
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604 46.30187
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441 15.66524
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781 27.14216
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537 24.43780
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725 37.66682
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191 23.93245
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353 30.17138
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186 29.05870
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228 35.25237
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119 21.74776
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930 23.45460
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842 38.96170
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938 35.42369
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740 42.46649
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309 40.73558
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936 18.64863
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679 34.63988
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977 26.93532
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734 15.45668
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274 27.97524
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488 36.82915
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285 34.62155
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924 28.56596
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501 29.48326
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594 28.23307
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204 34.56196
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757 34.39122
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647 23.19390
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817 29.93590
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198 34.67712
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107 30.10836
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799 17.63564
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766 41.11946
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768 25.46053
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116 31.45195
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080 12.38849
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506 45.52436
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289 39.36048
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871 24.82626
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888 46.60152
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193 27.70251
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237 42.76156
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523 38.99608
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849 21.71911
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502 10.71374
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091 27.97409
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819 38.06216
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622 34.90001
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536 43.10075
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519 30.43953
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706 27.32321
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061 32.54114
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049 28.77165
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894 29.02146
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097 30.84232
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549 38.16014
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855 33.98556
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994 27.15247
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002 28.18609
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384 37.94871
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966 26.54170
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447 45.66416
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626 17.29530
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198 34.67712
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371 27.02642
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547 29.82131
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220 13.22043
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708 41.82821
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027 39.55873
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718 32.05699
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002 42.77359
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989 33.57475
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454 33.81368
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491 30.64809
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578 33.59481
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608 45.09350
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100 33.79821
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843 33.64637
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453 23.51934
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086 46.64965
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952 33.99416
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031 29.27356
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410 42.13016
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658 40.45827
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910 27.72027
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702 41.87405
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796 30.32035
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103 27.60110
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562 31.62555
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017 30.22524
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421 16.52124
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164 28.51210
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547 35.78007
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454 33.81368
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652 31.66509
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211 29.09480
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543 34.94470
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590 20.89176
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134 39.13646
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632 32.95768
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421 14.48380
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818 31.48002
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196 27.98612
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172 36.13530
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101 34.56769
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546 25.38891
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125 25.89483
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828 31.72009
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852 30.94144
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661 37.33851
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185 32.55546
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347 24.13757
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082 39.88989
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228 30.10263
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129 33.72258
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783 32.07246
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504 40.96018
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475 25.23077
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450 21.87782
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433 30.48594
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463 31.27261
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392 29.83792
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656 21.26533
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613 19.64500
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366 36.10723
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560 44.22776
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791 18.95860
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248 25.58715
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146 30.95634
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025 38.26785
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700 36.31979
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645 34.65764
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193 45.58395
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066 30.02070
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055 42.01041
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462 23.36121
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508 32.71188
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906 23.09593
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397 39.69280
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811 26.85625
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672 29.12115
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989 37.61984
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785 37.53389
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536 31.39465
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861 36.81999
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857 33.48595
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665 16.94809
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691 33.52663
#>       distroad slides distdeforest distslidespast log.carea
#> 15412   300.00   TRUE       300.00            100  3.318273
#> 27864    30.00   TRUE       183.39             20  3.550683
#> 15762   300.00   TRUE       300.00            100  2.633933
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 5999    300.00   TRUE       300.00            100  5.201575
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 24072   300.00   TRUE       300.00            100  3.351559
#> 21939   300.00   TRUE       300.00            100  3.168886
#> 32622   300.00   TRUE       300.00            100  3.265863
#> 43327   173.40   TRUE        33.81            100  3.516154
#> 16811   300.00   TRUE       300.00             59  2.715521
#> 25690   300.00   TRUE       300.00            100  2.732242
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 43007   300.00   TRUE       300.00             37  2.473902
#> 2107    300.00   TRUE       300.00             45  2.661523
#> 35285   300.00   TRUE       300.00            100  2.997681
#> 30125   300.00   TRUE       300.00            100  2.748832
#> 29472   111.09   TRUE         0.00             25  3.424505
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 31576   300.00   TRUE       300.00              4  3.635091
#> 20290    95.53   TRUE         0.00             49  4.011286
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 34275   300.00   TRUE       300.00            100  2.711427
#> 6676    300.00   TRUE       300.00            100  3.303289
#> 6808    300.00   TRUE       300.00             40  3.952656
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 5749    300.00   TRUE         0.00            100  3.105292
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 28577   300.00   TRUE         0.56              0  3.781222
#> 11687   300.00   TRUE       300.00             68  3.186466
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 20001   264.46   TRUE        18.17             56  3.311398
#> 46235   300.00   TRUE        41.37              0  2.752133
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 6967    300.00   TRUE       300.00             85  3.008718
#> 17423   300.00   TRUE       300.00            100  2.868983
#> 37341   300.00   TRUE       300.00              1  2.823221
#> 37141   300.00   TRUE       300.00              1  2.293159
#> 651     300.00   TRUE       300.00            100  3.072673
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 44550   300.00   TRUE       300.00            100  2.968524
#> 23029   300.00   TRUE       300.00             34  2.425842
#> 34707   165.13   TRUE        39.56            100  3.473052
#> 4031    300.00   TRUE       300.00             31  2.907317
#> 9704    300.00   TRUE       300.00             90  2.784369
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 421     182.02   TRUE         0.00             83  2.874627
#> 375     300.00   TRUE       300.00            100  2.851936
#> 19887   300.00   TRUE       300.00             63  2.760635
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 30406   300.00   TRUE       300.00             85  2.770965
#> 2595    300.00   TRUE       300.00            100  3.538794
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 37416    72.23   TRUE        87.56             31  3.407698
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 664     300.00   TRUE       300.00             12  2.885027
#> 31121   300.00   TRUE       300.00             45  3.194113
#> 11618   300.00   TRUE       300.00              1  2.578737
#> 24501   300.00   TRUE       300.00              0  2.925363
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 20601   300.00   TRUE       300.00            100  3.059843
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 47320   300.00   TRUE       300.00             43  2.536741
#> 44767   300.00   TRUE       300.00             25  3.046207
#> 10933   264.46   TRUE        18.17             56  3.311398
#> 19892   300.00   TRUE       300.00             92  3.259412
#> 17345   300.00   TRUE       300.00              2  4.693266
#> 7       300.00   TRUE       300.00              0  2.263215
#> 43201   300.00   TRUE       300.00            100  3.530273
#> 42201   300.00   TRUE       300.00            100  3.989241
#> 20797   285.07   TRUE         0.00             96  3.663802
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 27915   300.00   TRUE       300.00              6  3.211627
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 10689   300.00   TRUE       300.00             61  2.237416
#> 43074   300.00   TRUE       205.02              1  3.153436
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 10617    17.89   TRUE       145.83              4  3.064162
#> 10692   300.00   TRUE        60.19             11  2.778935
#> 49133   300.00   TRUE       300.00             38  2.694996
#> 22297   300.00   TRUE       300.00             11  2.548891
#> 49378    22.68   TRUE       134.63             25  3.073643
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 20171   300.00   TRUE       300.00            100  3.645966
#> 1223    300.00   TRUE       300.00              4  3.251167
#> 39405   300.00   TRUE       300.00            100  3.413298
#> 14950   300.00   TRUE       300.00             15  2.616549
#> 16065   300.00   TRUE       162.92            100  3.818858
#> 46792   300.00   TRUE       300.00             18  2.657830
#> 28588   300.00   TRUE       300.00             78  5.209992
#> 13278    91.73   TRUE       120.78             34  3.560110
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 43834    20.21   TRUE       150.14              2  3.331390
#> 48461   300.00   TRUE       300.00             61  2.928936
#> 41408   300.00   TRUE       205.28              1  3.147505
#> 17322    75.77   TRUE       138.35              1  3.235115
#> 43536   300.00   TRUE       300.00              8  2.745482
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 18746     4.48  FALSE       205.00             27  3.028048
#> 49363   300.00  FALSE         0.00            100  3.394725
#> 25140   104.12  FALSE        46.28             74  2.414542
#> 39193   300.00  FALSE       117.12            100  2.879028
#> 43243   300.00  FALSE       300.00            100  2.922424
#> 47005   300.00  FALSE       300.00            100  4.858077
#> 39796   300.00  FALSE       300.00             71  2.333468
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 1085    300.00  FALSE       300.00            100  2.477139
#> 34347   300.00  FALSE       300.00            100  3.477538
#> 48183   300.00  FALSE       300.00            100  3.204523
#> 12974   300.00  FALSE       300.00            100  2.711947
#> 8831    300.00  FALSE       300.00             24  2.615022
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 44668     9.15  FALSE       163.30              6  2.852799
#> 31201   300.00  FALSE       300.00            100  2.404000
#> 20319   300.00  FALSE       300.00            100  3.167901
#> 13737   300.00  FALSE       300.00            100  2.678473
#> 29107   300.00  FALSE       300.00            100  2.550653
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 5656    300.00  FALSE        94.43            100  2.853361
#> 49831   300.00  FALSE       300.00            100  3.247811
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 36889   300.00  FALSE       300.00             48  2.108096
#> 44823   300.00  FALSE       300.00            100  2.749146
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 48988   300.00  FALSE       144.19            100  2.958793
#> 41612   300.00  FALSE       300.00            100  3.004762
#> 48184   300.00  FALSE       300.00             75  2.485742
#> 38362   300.00  FALSE       300.00            100  2.734802
#> 44434   300.00  FALSE       300.00            100  3.451290
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 18308   300.00  FALSE       300.00             55  4.224567
#> 19767   300.00  FALSE       300.00            100  2.878525
#> 48050   300.00  FALSE        72.23            100  2.520806
#> 20950   300.00  FALSE       300.00             81  4.555706
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 1434    300.00  FALSE       300.00             35  2.955995
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 32342   300.00  FALSE       300.00             81  2.961658
#> 23808   300.00  FALSE       300.00             72  2.692387
#> 26635   300.00  FALSE       300.00            100  3.092389
#> 11025   300.00  FALSE       300.00             50  2.828200
#> 39234   300.00  FALSE         0.00            100  2.170806
#> 46567   300.00  FALSE       300.00             78  2.785235
#> 48157    57.15  FALSE       157.06             45  2.887769