Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958   0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393    2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579    1318.7560
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752    1321.5997
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257     357.1436
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102   0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149   12632.7793
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189     287.5425
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160    9344.2930
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107    1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937   0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437    1155.1469
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651     419.0927
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991    1403.0193
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034    4495.6929
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338   10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504    1020.2766
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175    1122.9120
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236    2715.4138
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542    3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055     773.0359
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220    8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335     195.0293
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245    6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911     173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281    1178.4420
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410  -0.00380     601.0843
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756     353.9086
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143   0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152    1530.5657
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036    2505.8369
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301     338.9676
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230     714.7432
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419     761.5427
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463     851.6306
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464    4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794    1363.5321
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099    1146.2140
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502     520.5218
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236  -0.00066     306.0146
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532   16771.3066
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208     842.4825
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109     915.4989
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156     758.4840
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491