Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056   -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113   -0.00347    0.02357   1752.0375
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633   -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221    0.00969    368.5606
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917    0.00333    258.6537
#> 37265 712812.5 9559942 1860.77 63.3370465    0.00514   -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.3940835   -0.02191   -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090    0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858   -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073    0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104   -0.02046   1024.1573
#> 25169 712642.5 9560232 1883.10 46.8450293    0.00884    0.01345   1120.6929
#> 38827 713412.5 9560472 1869.10 16.8713152   -0.00156    0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.1589821   -0.00854    0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922    0.01088    266.4355
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842    868.6137
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581    0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475    0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.6234412   -0.02169   -0.00160   9344.2930
#> 48498 714972.5 9557972 2681.66 38.3469193    0.00528    0.01392   1255.0884
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017   -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471   -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126   -0.02044 129412.3984
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490   -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.0291164   -0.00521   -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.1282451   -0.00033   -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.8845494   -0.00493    0.00733    598.7983
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154    0.00124   2114.2253
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288    0.01452    437.3965
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613    0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617    0.00307 147470.5312
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395    0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480    0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372   -0.00008 825875.6875
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260   -0.03310   2833.9001
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140   -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.1373984    0.00953   -0.02283    590.1536
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306   -0.00436   2602.7390
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032    0.00542   3614.7334
#> 44825 714322.5 9557882 2534.80 40.6885978    0.00114    0.03026    247.0894
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550   -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170   -0.02554    0.03304   1569.7313
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167   -0.00539    0.00809   1102.9956
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645   -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125   -0.01972    0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.3058340   -0.00131    0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.2221961    0.00099    0.00240   1201.4724
#> 40270 713312.5 9559012 2186.56 30.9357102   -0.00771   -0.00599   1378.5760
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588    0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940    987.8085
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246    0.00534    344.1447
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759   -0.00400    608.6523
#> 41125 712772.5 9559412 2021.88 38.8763323   -0.00310   -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143   -0.05078 622888.2500
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263    0.01322   4605.1411
#> 22907 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130    0.00466   -0.01776   1021.8267
#> 1036  713042.5 9558892 2205.38 31.5098139   -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109    0.00298   1533.7925
#> 10363 713252.5 9559722 1999.55 41.8299297   -0.03669   -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182    0.02281    0.04759    158.9370
#> 23783 713062.5 9560252 1983.09 36.4298025    0.00774   -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605    0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767   -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016   -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.6578678   -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881    0.00121   1783.0658
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181   -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.0470808    0.00216   -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.9395253    0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118   -0.00252   3631.7017
#> 14140 714672.5 9559262 2326.36 45.7925059   -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.9893796   -0.00983    0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186    0.00063   2430.2278
#> 23832 714252.5 9560182 2188.12 30.6475125    0.00427    0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315    0.02405   -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025   -0.05345   2417.1389
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316   -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.0524260   -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084   -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203    0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009    0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465   -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759    0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284    0.01085   2229.5105
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124    0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007   -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133   -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815   -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165    0.00875   1165.3953
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290   2481.5632
#> 45166 715262.5 9558722 2531.59 13.5533166   -0.00502   -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.8258980    0.00846    0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253   -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937   -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544   -0.01014    271.4219
#> 13286 713662.5 9559702 2268.14 41.9966605   -0.01341    0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167    0.01133    828.8604
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844    0.00514    897.9290
#> 33175 715822.5 9558832 2563.13 36.6228893   -0.00674   -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648    0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423   -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301   -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041   -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185    0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224   -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782   -0.00048   1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302   -0.00393    0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.7365517   -0.00068   -0.00312    593.7773
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978    0.01282   2322.7944
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418    0.00513    161.9878
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233    0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209    0.00980   -0.00710    472.0236
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608    0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523    0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.5684513    0.00420    0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719    0.00079   1647.6451
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343    0.01177   1558.2231
#> 31098 714352.5 9559042 2410.15 30.3363964    0.00537   -0.00347   1893.3645
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182   -0.00972    331.7462
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865   -0.00805   7489.7417
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474    0.00340    0.01809   1158.0426
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014   -0.00884    672.2466
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992   -0.01052    451.7209
#> 46148 715082.5 9558042 2725.39 42.8360436   -0.00181   -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986   -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.2771580   -0.07627   -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.8455602    0.00298    0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087   -0.00573   1490.0312
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241    0.00020    673.2861
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208    0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769