Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017   300.00   TRUE       300.00             21  2.826839
#> 16435 49.44397   300.00   TRUE       300.00             41  2.478905
#> 44302 35.10111   300.00   TRUE       300.00            100  3.053301
#> 29391 28.28349   300.00   TRUE         0.56            100  3.032521
#> 19963 24.43780   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59826   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96665   279.54   TRUE        21.35             46  3.521349
#> 20799 24.80793   180.67   TRUE         0.00             16  3.361925
#> 28293 28.90228   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80357   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75829   279.54   TRUE         0.00            100  3.793638
#> 8869  29.11027   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627   195.00   TRUE        47.05              0  3.049487
#> 43327 29.05870   173.40   TRUE        33.81            100  3.516154
#> 7349  26.32283   300.00   TRUE       300.00              8  2.751808
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 8019  35.78007   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732   300.00   TRUE       300.00             46  2.952143
#> 40307 29.51363    77.82   TRUE        76.41             54  3.524078
#> 15109 22.63527   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46649   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310    87.56   TRUE       115.40              2  3.506583
#> 22566 24.58676   300.00   TRUE         7.49              0  2.426944
#> 6808  28.56596   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71778   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72810   300.00   TRUE       300.00             59  2.562268
#> 3453  40.08470   300.00   TRUE       300.00              2  3.276843
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116   300.00   TRUE         0.00            100  3.365895
#> 7879  28.58085   300.00   TRUE       300.00            100  2.912011
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576   300.00   TRUE         0.00              0  3.904384
#> 34359 27.41603   300.00   TRUE       294.31            100  3.287475
#> 27084 21.67270   300.00   TRUE       300.00             81  3.000309
#> 46171 29.07131   300.00   TRUE       270.22            100  3.645964
#> 28743 34.65306   300.00   TRUE       300.00             71  4.272628
#> 22332 27.06882   300.00   TRUE       300.00            100  5.168705
#> 44550 24.82626   300.00   TRUE       300.00            100  2.968524
#> 2718  36.73290   300.00   TRUE       300.00            100  2.990104
#> 49569 41.84941    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06309   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06216   300.00   TRUE       300.00             63  2.760635
#> 27875 21.26705   157.93   TRUE         0.00             57  3.709413
#> 33350 28.95270    10.00   TRUE       166.26              2  3.415431
#> 6317  27.32321   300.00   TRUE       300.00            100  2.837633
#> 25700 32.54114   300.00   TRUE       300.00            100  3.096112
#> 31462 33.75466   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32037   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46053   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22887   300.00   TRUE       300.00             23  3.350329
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77089   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96320   300.00   TRUE        13.47             64  3.177528
#> 11618 33.98556   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55221    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318   300.00   TRUE       300.00             91  2.623976
#> 42765 42.35304   300.00   TRUE       300.00             12  2.974066
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59766   300.00   TRUE       300.00              5  3.139431
#> 37769 37.42503   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73174    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21895   300.00   TRUE       300.00              6  5.916915
#> 1973  19.32587   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743   300.00   TRUE       300.00             48  5.457933
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 6217  30.97181   300.00   TRUE       300.00              1  3.243763
#> 47702 33.66643   300.00   TRUE       300.00             86  4.665675
#> 27915 33.57475   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18954    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44582   300.00   TRUE       300.00            100  3.663243
#> 5572  26.59785   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64809   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32283   300.00   TRUE       300.00              8  2.751808
#> 38668 32.92273   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765   300.00   TRUE       300.00            100  3.662232
#> 2801  25.74643   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934   300.00   TRUE       300.00             38  2.694996
#> 49378 33.99416    22.68   TRUE       134.63             25  3.073643
#> 39624 33.29458   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77833   300.00   TRUE       300.00            100  3.336505
#> 34984 33.90592   300.00   TRUE       300.00            100  2.699406
#> 39405 27.72027   300.00   TRUE       300.00            100  3.413298
#> 38103 42.76156   300.00   TRUE       300.00             31  2.907317
#> 16065 30.32035   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83354   300.00   TRUE       300.00            100  3.409673
#> 22847 24.74777   300.00   TRUE       300.00             41  3.134623
#> 46038 29.73078   300.00   TRUE       243.86            100  3.184852
#> 42640 34.90975   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00351   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71778   300.00   TRUE       300.00             83  2.904587
#> 17227 43.59750   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629   300.00   TRUE         7.49            100  6.276647
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 48461 35.78007   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55870   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 23832 25.48516   300.00   TRUE       300.00              8  2.676567
#> 48601 35.85799   300.00  FALSE       245.19            100  2.998830
#> 34250 30.89274   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18687   300.00  FALSE       216.66             10  3.025531
#> 14694 43.51156   300.00  FALSE       300.00            100  3.021498
#> 32869 14.60011   300.00  FALSE       300.00             78  2.867820
#> 441   25.11331   294.31  FALSE       152.09            100  3.054287
#> 42329 19.22789   300.00  FALSE       300.00            100  2.726581
#> 47005 36.13530   300.00  FALSE       300.00            100  4.858077
#> 44396 21.44237   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26116   300.00  FALSE        42.56            100  4.078911
#> 34347 31.72009   300.00  FALSE       300.00            100  3.477538
#> 4007  32.09251   300.00  FALSE       300.00             68  2.847871
#> 34616 31.49950   300.00  FALSE       300.00            100  2.321589
#> 44668 30.10263     9.15  FALSE       163.30              6  2.852799
#> 36245 25.85357   110.19  FALSE        81.50             20  3.615611
#> 13737 40.96018   300.00  FALSE       300.00            100  2.678473
#> 13908 34.37747   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712   300.00  FALSE         0.00             10  3.400551
#> 36721 26.66832   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20388   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820   234.82  FALSE         0.00             43  4.109707
#> 3724  25.72695   300.00  FALSE       300.00            100  2.450415
#> 34336 32.45118   300.00  FALSE       300.00            100  3.106036
#> 26283 41.51079   300.00  FALSE       300.00            100  2.962620
#> 40750 32.09996   300.00  FALSE       300.00            100  3.272034
#> 4943  37.77167   300.00  FALSE       300.00            100  3.019161
#> 15138 20.64596   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36121   300.00  FALSE        72.23            100  2.520806
#> 11394 40.60380    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69280   300.00  FALSE       300.00             35  2.955995
#> 28526 31.81463   300.00  FALSE        46.28            100  3.237445
#> 31440 24.03501   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35818   300.00  FALSE       300.00             65  3.024717
#> 24962 29.05011   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13072   300.00  FALSE       209.91            100  2.884241
#> 14645 45.21325   300.00  FALSE       300.00            100  3.232313
#> 26635 31.39465   300.00  FALSE       300.00            100  3.092389
#> 28978  4.43985   300.00  FALSE        10.00            100  2.209674
#> 30296 40.15861   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219   300.00  FALSE       300.00            100  2.858775
#> 40852 21.51800   300.00  FALSE       300.00            100  2.360983