Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403    1787.3024
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835    2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677    1343.5006
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526     539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032     994.6740
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385     674.0721
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413     364.9788
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739     558.2286
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598   0.01212     561.1948
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916    2009.1686
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178     800.1933
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452     437.3965
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684    2556.8049
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132   42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211   49347.5547
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911     173.2806
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713    6379.7974
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171    4425.5410
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252    3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306    1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559     866.5775
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232     977.7320
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712    1010.5097
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133     828.8604
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216    1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033    3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018     937.6911
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360    4288.7329
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750    1892.8978
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849     764.8193
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215     917.5296
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566   12862.4922
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585    5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169     649.9975
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603