Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491