Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475   1001.0861
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896    276.0310
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593    939.0466
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993  10308.6406
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332   1418.6622
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542   3614.7334
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558   1504.9717
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351   1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334   4905.2002
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439  -0.01161   3065.3694
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363    565.5207
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216  -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609   4275.3457
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598    186.3527
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367   3002.8828
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011    201.1272
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036    476.9504
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360   4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236  -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.634509    0.00082   0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012    305.3624
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124  25826.6914
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253   1237.0536
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592   300.00   TRUE       300.00            100  2.699406
#> 44302 35.10111   300.00   TRUE       300.00            100  3.053301
#> 13748 34.11906   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 46178 20.93129   300.00   TRUE       300.00             31  2.440958
#> 48187 30.96665   279.54   TRUE        21.35             46  3.521349
#> 21939 23.93245   300.00   TRUE       300.00            100  3.168886
#> 25169 36.60627   195.00   TRUE        47.05              0  3.049487
#> 49372 23.54054   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94657   162.65   TRUE         0.00             77  3.703905
#> 1393  23.41220   300.00   TRUE       300.00              1  2.710095
#> 21703 22.26915   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78007   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32130   300.00   TRUE       123.56              5  2.520462
#> 43964 28.75732   300.00   TRUE       300.00             46  2.952143
#> 29420 28.85244   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46649   300.00   TRUE       300.00            100  2.997681
#> 34482 32.97143   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676   300.00   TRUE         7.49              0  2.426944
#> 25081 34.87537   300.00   TRUE       300.00            100  3.617357
#> 18391 34.50753   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71091   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592   300.00   TRUE       300.00             30  2.721899
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 3466  26.79494   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684   300.00   TRUE       300.00            100  3.303016
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832   119.34   TRUE        79.16             61  3.129568
#> 17594 19.23190   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29783   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10836   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020   300.00   TRUE       300.00            100  2.676979
#> 31316 30.00981   300.00   TRUE       300.00             85  2.740390
#> 13699 34.78942   300.00   TRUE        53.95             43  3.288564
#> 47884 27.05449   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306   300.00   TRUE       300.00             71  4.272628
#> 39599 33.13472   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559   300.00   TRUE       300.00              8  3.050346
#> 5084  31.59518   300.00   TRUE       300.00             18  2.722101
#> 2718  36.73290   300.00   TRUE       300.00            100  2.990104
#> 421   10.71374   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56310   300.00   TRUE       300.00             94  4.013201
#> 38862 34.90001   300.00   TRUE       104.48             92  3.151879
#> 10507 27.16221   300.00   TRUE       300.00             46  2.580865
#> 39732 29.16814   300.00   TRUE       300.00             25  3.364146
#> 47100 29.64598    82.80   TRUE       129.73             36  3.558076
#> 46783 19.94237   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 664   30.84232   300.00   TRUE       300.00             12  2.885027
#> 21013 24.96320   300.00   TRUE        13.47             64  3.177528
#> 28046 19.59115   300.00   TRUE       300.00             15  2.896560
#> 45592 15.64862   300.00   TRUE       300.00              2  2.460907
#> 26010 25.22962    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 12772 37.92007   300.00   TRUE       300.00            100  2.936990
#> 18958 29.98976   300.00   TRUE       300.00              5  2.674746
#> 11612 35.16700   300.00   TRUE       300.00            100  2.516921
#> 47718 32.32341   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64598   300.00   TRUE       291.23            100  3.277641
#> 7     13.22043   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41318   300.00   TRUE       300.00            100  3.049449
#> 18348 41.01060   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66643   300.00   TRUE       300.00             86  4.665675
#> 42201 39.55873   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93574   300.00   TRUE       300.00            100  3.159244
#> 37720 42.77359   300.00   TRUE       300.00            100  3.690657
#> 29228 30.39942   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81368   300.00   TRUE       205.28              1  3.147505
#> 26428 35.47525   300.00   TRUE         0.00            100  3.486483
#> 47244 28.03884   300.00   TRUE       300.00              2  2.752448
#> 19756 30.11638   300.00   TRUE       300.00            100  2.694631
#> 49378 33.99416    22.68   TRUE       134.63             25  3.073643
#> 1997  25.23593   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13016   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027   300.00   TRUE       300.00            100  3.413298
#> 16065 30.32035   300.00   TRUE       162.92            100  3.818858
#> 46627 53.68328   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85686   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90975   300.00   TRUE       300.00            100  3.627130
#> 18498 22.71778   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28558   300.00   TRUE       300.00             63  3.252649
#> 14779 16.52124    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51210    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90610   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939   300.00   TRUE       300.00             57  3.322176
#> 36765 36.18515   300.00   TRUE       300.00            100  3.279776
#> 17374 43.14544   300.00   TRUE       300.00            100  3.630971
#> 10145 27.41603   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83487   300.00  FALSE       300.00            100  2.844665
#> 36561 25.80487   300.00  FALSE        70.00            100  2.990220
#> 8635  27.68188   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38396   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85631   300.00  FALSE       300.00            100  2.978336
#> 30267 25.27030   300.00  FALSE       300.00            100  2.781107
#> 19493 18.57243   300.00  FALSE       291.23            100  2.550849
#> 43426 35.46265   300.00  FALSE        82.11            100  3.910718
#> 7538  17.26494   300.00  FALSE       300.00             87  2.433645
#> 39796 34.56769   300.00  FALSE       300.00             71  2.333468
#> 16878 25.38891   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237   300.00  FALSE       300.00            100  2.270336
#> 34347 31.72009   300.00  FALSE       300.00            100  3.477538
#> 8413  39.88989    80.73  FALSE         0.00            100  3.040329
#> 24009 38.49703   300.00  FALSE       300.00            100  2.303471
#> 13737 40.96018   300.00  FALSE       300.00            100  2.678473
#> 29107 25.23077   300.00  FALSE       300.00            100  2.550653
#> 17746 36.39141   300.00  FALSE       300.00             87  2.773624
#> 26389 36.12212   300.00  FALSE        76.38             26  3.197613
#> 33410 32.87976   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27261   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036   300.00  FALSE       300.00             91  2.462702
#> 7983  38.50620   300.00  FALSE        57.15             22  2.884583
#> 42424 22.60720   275.21  FALSE        93.50             70  2.791131
#> 44823 36.10723   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55909   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389   300.00  FALSE       300.00            100  3.632329
#> 24819 29.93704   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61256   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695   300.00  FALSE       300.00            100  2.450415
#> 48184 30.95634   300.00  FALSE       300.00             75  2.485742
#> 974   22.75846    80.00  FALSE        90.05            100  3.099987
#> 14763 10.14479   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739   124.35  FALSE         0.00            100  3.192630
#> 15138 20.64596   116.69  FALSE        19.61             40  4.109325
#> 24446 31.36314   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916   300.00  FALSE       225.26             12  3.030143
#> 4269  19.11674   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731   300.00  FALSE       300.00             18  2.682893
#> 1577  38.92962   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71188   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380    46.04  FALSE        55.00            100  3.187005
#> 16443 26.53826   300.00  FALSE       300.00             75  2.484816
#> 46639 33.56329   300.00  FALSE       300.00            100  3.756965
#> 46148 38.91071   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602   300.00  FALSE       300.00             85  4.412069
#> 23808 37.53389   300.00  FALSE       300.00             72  2.692387
#> 14645 45.21325   300.00  FALSE       300.00            100  3.232313
#> 26635 31.39465   300.00  FALSE       300.00            100  3.092389
#> 13548 16.97846   300.00  FALSE       300.00            100  2.256336
#> 6206  32.46322   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861   300.00  FALSE       300.00             68  3.258250
#> 48157 33.52663    57.15  FALSE       157.06             45  2.887769