Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029    5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761     928.6858
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396     704.7076
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015     790.2421
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431     315.9084
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368    1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064    1485.5201
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458     485.7418
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491