Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607