Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257    357.1436
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547    519.4228
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773    179.3077
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150   1274.3594
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276   9614.6924
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086   6042.5757
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504   1020.2766
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092    412.5849
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690   5121.6924
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625   1063.9150
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304   1569.7313
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656   3036.7383
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463   1050.7473
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759   4131.5356
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705    968.9738
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165    805.1403
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502    520.5218
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875   2156.9358
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983