Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393   2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021   -0.01579   1318.7560
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339   1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752   1321.5997
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292   1233.9299
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264   5057.1431
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783    377.8918
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466    724.0713
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104   1641.4587
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328    592.3326
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710    843.7961
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314   2360.1501
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773    179.3077
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937    0.00083    994.7876
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077   3559.4204
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184    0.03924   8023.8657
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991   1403.0193
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504   1020.2766
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552   -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617    0.00307 147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359   -0.00279    930.0871
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473   2930.2791
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102  35685.0000
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800   -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643    670.2289
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469    0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412   1350.3611
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556   -0.11694 102301.4609
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372   4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722   1373.4147
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739    789.7573
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380    601.0843
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109   2047.9031
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094   1392.6008
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325    595.0426
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019    0.00119   6589.5796
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632    620.0229
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115    0.01015    790.2421
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840    0.00189    556.5211
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655    130.9468
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519    0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642   1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599    0.00301    338.9676
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181   1286.3949
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452    578.5847
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064    683.2795
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945    259.7421
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502   1428.8546
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399   1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196    0.01734    335.0773
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601   -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267  31588.1953
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782   -0.00048   1550.6704
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785    0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312    593.7773
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310   -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204    618.2027
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820   -0.00260   1484.6593
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375    0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208    842.4825
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305    0.00585   5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013    616.4128
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962   -0.00172    351.5812
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983