Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491