Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235