Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043   1806.4552 34.119064
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861 39.335272
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537 20.192179
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877 32.738808
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299 42.221833
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032  6.084812
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529   3282.1186 29.058700
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918 16.677656
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053 47.548622
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587 35.240915
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586  8.761098
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938 35.423689
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096   3210.5781 33.063103
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398    267.2659 24.586765
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685 14.328529
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367    793.9665 34.233082
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630 23.831607
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019 34.875368
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363 21.073961
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924 28.565957
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746 33.621163
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138 37.901731
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924 25.657050
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014 26.794944
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838 27.189139
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204 34.561960
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915 14.659698
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193 27.419468
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355  14919.6260 34.878806
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937 32.097669
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836 35.319410
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064 31.188958
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120 21.700204
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799 17.635641
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072  54831.6289 39.360482
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184 26.882034
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335 23.337144
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688 27.951746
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156 39.063117
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933 41.733673
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706 27.323211
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626 29.842507
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993 36.320368
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558 43.540782
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150 33.241865
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313 27.831425
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840 39.185157
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588 33.576473
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963 34.342517
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081    940.6002 28.186086
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384 37.948714
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196 29.979444
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626 17.295304
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583 32.323414
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399 26.645975
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242 30.971807
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220 13.220428
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450 41.010600
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983 30.003508
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454 33.813677
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776   1021.8267  4.138474
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904 28.646171
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436 25.726951
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621 27.998156
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484    495.0292 30.116381
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608 45.093497
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404 25.746431
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691 40.586611
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031 29.273560
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964 33.294577
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702 41.874047
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474 42.573056
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657 29.730780
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231 43.597505
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744    226.7911 32.011725
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748 31.598622
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875 31.931511
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547 35.780068
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454 33.813677
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002 32.263253
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782 27.416030
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590 20.891760
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232    977.7320 25.804873
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078 14.818407
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565   1410.8372 38.548600
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949 38.982902
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525 33.383959
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401 50.856307
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666 42.063124
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975 25.270304
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078 18.572427
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818 31.480020
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261 23.447725
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840 23.945052
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219 17.264937
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961 28.075505
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773 19.295500
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347 24.137566
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459 36.726595
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949 38.077629
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011    201.1272 38.497034
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450 21.877820
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547 32.236897
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065 36.122124
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433 30.485938
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463 31.272609
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031 39.850361
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752 34.677125
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392 29.837923
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733 28.450219
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027 22.607196
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282    561.2366 36.107227
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236 32.282161
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802 43.676573
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076 25.726951
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111 22.860443
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354  0.00907   1281.5044 10.144791
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019 32.003131
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645 34.657644
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066 30.020697
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875   2156.9358 29.323407
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825 31.363137
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462 23.361208
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055 31.814627
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672 29.121153
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466  4.191186
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975 30.881852
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209 24.035007
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630 33.358176
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606 38.910710
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128 42.407471
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536 31.394649
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820 40.158612
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579 31.663367
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691 33.526625
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544 36.254477
#>       distroad slides distdeforest distslidespast log.carea
#> 13748   300.00   TRUE       300.00              2  3.256827
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 30059   300.00   TRUE       300.00             41  2.412719
#> 24051   300.00   TRUE       300.00             11  3.001122
#> 15843   300.00   TRUE       300.00            100  3.091290
#> 9669    300.00   TRUE       300.00              1  3.610558
#> 39885   210.57   TRUE         0.00            100  2.781111
#> 43327   173.40   TRUE        33.81            100  3.516154
#> 31023   135.00   TRUE         0.00              2  2.577367
#> 39629   300.00   TRUE       300.00             85  3.366461
#> 39061   300.00   TRUE        70.65             56  3.215230
#> 38410    69.52   TRUE        47.61             65  3.695093
#> 2107    300.00   TRUE       300.00             45  2.661523
#> 22933    87.56   TRUE       115.40              2  3.506583
#> 22566   300.00   TRUE         7.49              0  2.426944
#> 27978   300.00   TRUE       300.00             12  2.377795
#> 38070   300.00   TRUE       300.00            100  2.899802
#> 26215   300.00   TRUE       300.00             10  3.268382
#> 25081   300.00   TRUE       300.00            100  3.617357
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 6808    300.00   TRUE       300.00             40  3.952656
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 28230   300.00   TRUE       300.00            100  2.977867
#> 46125   300.00   TRUE       300.00            100  3.982935
#> 3466    101.32   TRUE        95.02             19  3.724022
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 34642   300.00   TRUE       300.00              2  2.460907
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 6557    300.00   TRUE       300.00              5  3.147064
#> 47351   300.00   TRUE       300.00            100  4.173758
#> 47429   300.00   TRUE       300.00            100  2.464778
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 38686   300.00   TRUE       300.00             75  2.946946
#> 18564   300.00   TRUE       300.00            100  2.676979
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 17423   300.00   TRUE       300.00            100  2.868983
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 2976    300.00   TRUE       300.00            100  2.684323
#> 15766   300.00   TRUE       300.00            100  2.760448
#> 4617    300.00   TRUE       300.00              6  2.211304
#> 42941   300.00   TRUE       300.00             17  4.882627
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 28845   300.00   TRUE       300.00             54  2.563199
#> 7668    300.00   TRUE       300.00              5  2.789721
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 32280   126.66   TRUE         0.00              5  3.195825
#> 30886   300.00   TRUE       300.00            100  2.695380
#> 13146   300.00   TRUE       300.00              2  3.278424
#> 38197   105.52   TRUE        65.32             50  3.036108
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 20601   300.00   TRUE       300.00            100  3.059843
#> 41721   300.00   TRUE       300.00            100  3.099963
#> 44767   300.00   TRUE       300.00             25  3.046207
#> 47718   300.00   TRUE        10.00              1  3.759272
#> 31117   300.00   TRUE       291.23            100  3.277641
#> 6217    300.00   TRUE       300.00              1  3.243763
#> 7       300.00   TRUE       300.00              0  2.263215
#> 18348   300.00   TRUE       300.00            100  3.055854
#> 15981   300.00   TRUE       300.00              0  2.799064
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 5911     42.88   TRUE         0.00             13  3.009377
#> 26101   300.00   TRUE       300.00            100  3.068405
#> 1036    300.00   TRUE       300.00            100  3.132178
#> 11352   135.89   TRUE        82.11              4  4.291317
#> 19756   300.00   TRUE       300.00            100  2.694631
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 2801    300.00   TRUE       300.00             26  2.422656
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 14192   300.00   TRUE       300.00            100  3.115833
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 39624   300.00   TRUE       300.00              2  3.254015
#> 14950   300.00   TRUE       300.00             15  2.616549
#> 38103   300.00   TRUE       300.00             31  2.907317
#> 29024   300.00   TRUE       300.00             32  2.529363
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 46038   300.00   TRUE       243.86            100  3.184852
#> 18498   300.00   TRUE       300.00             83  2.904587
#> 17227   300.00   TRUE       300.00            100  2.467494
#> 4479    300.00   TRUE       300.00             55  2.355626
#> 10213   300.00   TRUE       300.00            100  3.351772
#> 6212    300.00   TRUE       300.00             47  5.501416
#> 48461   300.00   TRUE       300.00             61  2.928936
#> 41408   300.00   TRUE       205.28              1  3.147505
#> 17851   300.00   TRUE       300.00            100  4.161853
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 10145   276.16  FALSE        91.33            100  3.227418
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 36561   300.00  FALSE        70.00            100  2.990220
#> 46165   300.00  FALSE       300.00            100  2.545811
#> 15186    60.00  FALSE         0.00            100  3.149477
#> 25212   300.00  FALSE       300.00            100  3.109374
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 14717   300.00  FALSE       300.00            100  2.978336
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 26963   300.00  FALSE       300.00            100  3.420610
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 19493   300.00  FALSE       291.23            100  2.550849
#> 441     294.31  FALSE       152.09            100  3.054287
#> 39193   300.00  FALSE       117.12            100  2.879028
#> 45166   300.00  FALSE       300.00              1  3.379256
#> 29125   300.00  FALSE       300.00            100  2.626221
#> 7538    300.00  FALSE       300.00             87  2.433645
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 36126   300.00  FALSE       300.00            100  2.525145
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 29527   300.00  FALSE       300.00            100  2.639632
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 24009   300.00  FALSE       300.00            100  2.303471
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 26389   300.00  FALSE        76.38             26  3.197613
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 5656    300.00  FALSE        94.43            100  2.853361
#> 20764   300.00  FALSE       300.00             91  2.462702
#> 21070   300.00  FALSE         0.00             10  3.400551
#> 49831   300.00  FALSE       300.00            100  3.247811
#> 44715   300.00  FALSE       300.00            100  2.225757
#> 42424   275.21  FALSE        93.50             70  2.791131
#> 44823   300.00  FALSE       300.00            100  2.749146
#> 26768   300.00  FALSE       207.22            100  2.673964
#> 1479    300.00  FALSE       300.00            100  2.938910
#> 3724    300.00  FALSE       300.00            100  2.450415
#> 32430   300.00  FALSE       300.00            100  2.398827
#> 14763   110.04  FALSE        60.35            100  3.107720
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 18308   300.00  FALSE       300.00             55  4.224567
#> 280      60.55  FALSE       127.54             29  3.333837
#> 24446   300.00  FALSE       300.00            100  2.925561
#> 48050   300.00  FALSE        72.23            100  2.520806
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 37254   131.59  FALSE        80.55            100  2.827529
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 31440   300.00  FALSE       300.00             96  2.654870
#> 49812   300.00  FALSE       300.00             65  3.024717
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 6577    300.00  FALSE       300.00             70  2.789872
#> 26635   300.00  FALSE       300.00            100  3.092389
#> 30296   300.00  FALSE       300.00             68  3.258250
#> 47629   300.00  FALSE       300.00            100  3.627607
#> 48157    57.15  FALSE       157.06             45  2.887769
#> 24516   300.00  FALSE       122.13            100  3.751491