Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806