Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069    0.00958   0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123   0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056  -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.6679573   -0.00292   0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759   -0.00075   0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251    0.00479   0.01261    323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917   0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659   0.01041    772.9746
#> 24072 715282.5 9557602 2837.46 34.3940835   -0.02191  -0.00579   2246.7725
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453   0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841  -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730  -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529   3282.1186
#> 49372 714862.5 9560982 1863.12 20.9691094   -0.00295   0.00035   2697.5974
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550  -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175  -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160   0.00311    586.7354
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664  -0.05766 140591.4062
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.0168746   -0.00720   0.01250   3686.1240
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905  -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382   0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628    556.0121
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490  -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.2576915   -0.02651  -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613  -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987   0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197   -0.00664   0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257   0.00593    939.0466
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669  -0.00950    950.3138
#> 13847 715052.5 9559232 2337.87 26.1148433   -0.01452  -0.01437   1155.1469
#> 19730 715332.5 9558452 2640.35 47.6425866   -0.00211   0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089  -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.6322593    0.01414   0.03416   1204.8629
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355  14919.6260
#> 34359 712862.5 9558912 2111.37 32.8121470    0.00463   0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353   0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.5357857    0.00140  -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455   -0.01993   0.01373  16031.7500
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442  -0.00178    800.1933
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063  18733.8848
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286  -0.00964    571.1364
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409   0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478   0.01092    412.5849
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762   0.01252   5761.3486
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364  -0.03357    473.4709
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160   0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221   0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576  -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032   0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.9349627   -0.01807  -0.01883    616.1993
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126  -0.00485   2240.4185
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195  -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.5969737   -0.03800  -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568  -0.01558   1504.9717
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251    0.01784   0.00596    842.0994
#> 2552  714942.5 9557652 2719.32 34.8736492   -0.01238  -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478  -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604   0.00354   1147.7384
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133   0.00367    806.0966
#> 1973  715252.5 9557822 2861.00 15.5707010    0.01865   0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458  -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246   0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278  -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453   0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986  -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.8151724   -0.00332  -0.00058   1817.2371
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521  -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890  -0.00911    173.2806
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084  -0.02236   9755.3027
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666  -0.03583   1898.0105
#> 5572  715062.5 9558262 2754.60 39.9116671    0.00707   0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542  -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.9813582    0.05379   0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193   0.02812   1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593    0.03169   0.01441    424.2680
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051   0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317   0.00077   3056.9475
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705  -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605   0.00075   2170.2249
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257   0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185  -0.01185    953.8614
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181  -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610    454.8103
#> 28588 715532.5 9559362 2299.84 11.1446021   -0.04360  -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.7276774   -0.01020   0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144  -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004  -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.2311889    0.01895   0.04744    226.7911
#> 6212  714592.5 9560362 2034.54 12.5225656   -0.01454   0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.3812407   -0.00465  -0.01036   2505.8369
#> 6470  713742.5 9558802 2260.79 30.6801711    0.00000   0.00320    911.5911
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203   0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002  -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009   0.00109    605.1908
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156  -0.00116   1817.5222
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815  -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197   0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825   0.00445    355.5078
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626    450.7346
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527  -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507    504.3085
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338  -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.9416565   -0.00065   0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.7604272   -0.00457   0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.0850495   -0.00196  -0.00175    725.1355
#> 20319 715282.5 9560482 1901.16 27.6767263    0.00706  -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.2735807   -0.00022   0.00662   1269.2650
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.6414056   -0.00307   0.02997    230.8641
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413   0.00528    355.3475
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066   0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.2949337   -0.00097   0.01487   1769.3392
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858   0.00312    254.0656
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119   0.00361   1011.0248
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262  -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719   0.00079   1647.6451
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354   0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070    0.01852   0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050   1471.4346
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607  -0.04566  12862.4922
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390   -0.06909  -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284  -0.00886  43249.3945
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508  -0.00177    184.1042
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309   0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788  -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014  -0.00884    672.2466
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983   0.00803   1058.5630
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.8520794    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603