Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775