Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200