Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454   3476.0818
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870   2099.7930
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.87076  0.00321  0.00548    532.8210
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740    909.4791
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983