Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579   2246.7725
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770    474.5059
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665  25318.3555
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096   3210.5781
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672   1114.8801
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696   4143.4019
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236   2715.4138
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739    365.7626
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883    616.1993
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809   1102.9956
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596    842.0994
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213    328.7922
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776   1021.8267
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123    364.2903
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422   1159.2100
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744    226.7911
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109    605.1908
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945    259.7421
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133    828.8604
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858    298.4490
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983