Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579    2246.7725
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175    1122.9120
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166     650.6246
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542    3614.7334
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961    1086.6963
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242     507.9263
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722    1373.4147
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261    3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456    1531.4274
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500     538.8209
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923     811.5580
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098     164.0802
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388     180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336