Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 27864 712992.5 9560672 1926.38 27.2240896  -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.5335149  -0.01021   -0.01579    1318.7560
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.1484522   0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.5839352  -0.00498   -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835  -0.02191   -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005     769.3426
#> 25169 712642.5 9560232 1883.10 46.8450293   0.00884    0.01345    1120.6929
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387    1989.3012
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311     586.7354
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526     539.8119
#> 30800 715182.5 9557582 2772.37 39.2040642  -0.02475    0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272    4955.5586
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.1078353  -0.00017   -0.00032     994.6740
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.1813275  -0.00382    0.00672    1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203   10263.2734
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358     802.7622
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694    2123.4473
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150    1274.3594
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276    9614.6924
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077    3559.4204
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.7576880   0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313     783.4548
#> 15990 712852.5 9559982 1894.77 37.6943204  -0.01545    0.02405    5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556     660.3690
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939     703.7817
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504    1020.2766
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294     739.5768
#> 3944  715262.5 9558352 2654.20 33.5552733  -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245  -0.00585   -0.00974    2136.6939
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373     527.3528
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993   10308.6406
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241     585.6182
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009    1247.7061
#> 7668  713372.5 9559062 2164.82 37.9349627  -0.01807   -0.01883     616.1993
#> 44650 715612.5 9558142 2764.88 39.3604817  -0.00550   -0.00941    1290.7709
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219     420.7032
#> 5239  714832.5 9560992 1859.70 19.9601307  -0.00320    0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228    2262.3796
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671  -0.04376   -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694  102301.4609
#> 16009 715122.5 9559402 2264.81 50.0186426  -0.00589    0.00849    1120.5961
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321    1170.5904
#> 28009 715002.5 9559922 2237.68 34.6037860  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753    1337.7854
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484     495.0292
#> 49072 715472.5 9558902 2482.92 30.6893384  -0.01102   -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.8354356   0.00181   -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426    1315.5334
#> 20426 714702.5 9559712 2208.61 43.0463191  -0.08153   -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.8087514  -0.01881    0.00121    1783.0658
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119    6589.5796
#> 30363 715192.5 9557542 2796.31 25.3218698  -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.4649416   0.00560    0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732    4237.6934
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856    1789.1596
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291     487.3438
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.0527893  -0.00586    0.00067    1123.7318
#> 46165 714392.5 9557912 2542.56 16.3298701   0.00325   -0.00505     351.4078
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177     997.3089
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230     714.7432
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.1858831  -0.00298   -0.00712    1010.5097
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118    2273.4758
#> 26963 715702.5 9557882 2880.32 43.7332955  -0.01174   -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875    1165.3953
#> 9665  713362.5 9558812 2275.28 31.4949170   0.01306   -0.00626     450.7346
#> 42329 715852.5 9558702 2651.92 22.8707563   0.00321    0.00548     532.8210
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.2208476  -0.04937   -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502    1428.8546
#> 36833 713322.5 9560572 1794.25 33.7540896   0.00527   -0.02077    8759.9961
#> 43403 712632.5 9559262 2072.53 34.4863297  -0.00069   -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.2478257   0.00737    0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175     725.1355
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.2880302  -0.00393    0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.2640080   0.02413    0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196     576.1450
#> 44267 712972.5 9559042 2191.01 22.2691506   0.01462    0.00268     252.0072
#> 12519 713982.5 9560912 1956.98 28.8249337  -0.00452   -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780  -0.00305   -0.00705     968.9738
#> 33666 713632.5 9561012 1775.88 21.1788119  -0.00950   -0.05130     107.5831
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487    1769.3392
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209   0.00980   -0.00710     472.0236
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417    1520.6630
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578    1258.8865
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327   0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148     481.8292
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935     903.6397
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 6577  714142.5 9559902 2244.30 57.8905097  -0.06986   -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281     492.4785
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253    1237.0536
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403     609.8665
#> 40852 715392.5 9557932 2939.71 18.2974072   0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491