Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393    2081.0056
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475    1001.0861
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896     276.0310
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767    3321.6113
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.0977107  -0.00858   -0.00292    1233.9299
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303   0.00310    0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.2839150  -0.00284    0.01804     628.5153
#> 26447 715632.5 9559102 2410.77 39.6567008   0.00730    0.01770     474.5059
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.6120312  -0.02894   -0.03416     947.4878
#> 8019  714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392    1255.0884
#> 18005 714832.5 9557912 2660.76 31.9120303   0.00191   -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070  138869.3906
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367     793.9665
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 48364 715582.5 9558222 2746.23 25.0720602   0.01257    0.00593     939.0466
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083     994.7876
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114     150.1498
#> 15990 712852.5 9559982 1894.77 37.6943204  -0.01545    0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616   0.02224    0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096     520.5570
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862     874.5408
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373   16031.7500
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796    4729.6382
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.0208853   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071  -0.00762    0.01252    5761.3486
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.1373984   0.00953   -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.2512354  -0.00221    0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.6950539   0.00306   -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262    1760.2162
#> 46065 713072.5 9559162 2109.39 16.2645529  -0.00195   -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316     454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167  -0.00539    0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.0591409  -0.00410    0.03620     379.0855
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240    1201.4724
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.4746819   0.00629   -0.00469    1021.5407
#> 39722 715022.5 9557712 2735.53 30.2607659   0.00133    0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.0344196   0.00188    0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.5707010   0.01865    0.00335     195.0293
#> 10467 714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205    1112.2626
#> 10933 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376    2048.3198
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 18348 714042.5 9558882 2341.41 44.0638922  -0.00338    0.00728    1137.2450
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322    4605.1411
#> 28312 715712.5 9558032 2783.23 24.2527305  -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363     565.5207
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.4460340   0.01003    0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384  -0.01102   -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190     264.6404
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.3049047  -0.00605    0.00075    2170.2249
#> 7409  714772.5 9559122 2404.90 30.2155023  -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140    1066.3957
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.3218698  -0.00157   -0.00152     850.4633
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808   0.00216   -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.1446021  -0.04360   -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 35458 715242.5 9558442 2623.01 38.9548275  -0.00770   -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291     487.3438
#> 43536 713782.5 9558732 2281.50 32.7491217   0.00840    0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.2034632  -0.00585   -0.02665    1904.4769
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331     699.3024
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232     977.7320
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565    1410.8372
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875    1165.3953
#> 25140 713332.5 9560692 1809.95 52.2571887   0.00075   -0.01945     259.7421
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.2208476  -0.04937   -0.02273   72123.6172
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133     828.8604
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943     371.5514
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 8831  714372.5 9559852 2200.77 48.7604272  -0.00457    0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.8637272   0.01180    0.01170     253.5129
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462     716.5762
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 34345 715852.5 9557902 2874.08 43.6559462   0.00379   -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972  -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 20764 714542.5 9558152 2602.18 52.0789988   0.02118    0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513     161.9878
#> 7983  713512.5 9560432 1907.38 50.0192155   0.03310   -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487    1769.3392
#> 44715 714932.5 9560132 2195.78 30.3936922   0.03479    0.02251     168.1733
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282     561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360    4288.7329
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105     868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835   0.02523    0.02247     282.1076
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904     608.3561
#> 26283 715932.5 9557592 3075.48 32.7668833  -0.00375    0.00215     917.5296
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148     481.8292
#> 20950 714912.5 9557782 2667.71 30.8113784  -0.04770    0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012     305.3624
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.0432868   0.00962   -0.00172     351.5812
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491