Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056   -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633   -0.01813   1130.5791
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507 159065.0781
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588   1295.1610
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044 129412.3984
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889    362.8363
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898    588.3761
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593    939.0466
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950    950.3138
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276   9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077   3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365   1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939    703.7817
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452    437.3965
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530    937.1306
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092    412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002    0.00588    749.2502
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408  76318.0156
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409   2312.8413
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102  35685.0000
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664   1898.5588
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354   1147.7384
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335    195.0293
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833    369.3771
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801   1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422   1159.2100
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759    158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185    953.8614
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019    0.00119   6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000 162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015    790.2421
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144 317260.6875
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093    118.1402
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109    605.1908
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452    578.5847
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479    604.0975
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500    538.8209
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431    315.9084
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065    704.4843
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069   -0.00900    614.5898
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223   1056.3701
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528    355.3475
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038   2017.3547
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599  12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361   1011.0248
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247    282.1076
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044   -0.00964    543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532  16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805   7489.7417
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585   5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458    485.7418
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491