Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976    301.2347
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005    769.3426
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835   2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160   9344.2930
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365   1536.2647
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373  16031.7500
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363    565.5207
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937  -0.02273  72123.6172
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399   1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900    614.5898
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674  -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148  10771.1963
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079   1647.6451
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172  -0.00119    792.3544
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169    649.9975
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172    351.5812
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181  -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607