Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345    1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281    1325.6138
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898     458.6938
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385     674.0721
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588     749.2502
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454    3476.0818
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809    1102.9956
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010    3467.2690
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685    2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000  162178.1562
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452     578.5847
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199     836.4196
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011     201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048    1550.6704
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251     168.1733
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099     909.0560
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011     501.7867
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458     485.7418
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419     323.0790
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309