Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491