Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607