Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 4965  715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634    500.5027
#> 37912 712802.5 9559952 1838.40 52.1013441   0.00183   -0.09203    634.3320
#> 16749 715072.5 9558242 2745.27 49.0537816   0.00097    0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741    746.4155
#> 11088 715532.5 9558762 2549.28 38.0501272  -0.00373   -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005    769.3426
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403   1787.3024
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387   1989.3012
#> 1393  714832.5 9557902 2664.72 31.2955277   0.00550   -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783    377.8918
#> 10974 715312.5 9557502 2844.23 26.6648828  -0.00831   -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842    868.6137
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.8879661   0.01903    0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514    297.7842
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382    560.8309
#> 22933 712712.5 9560432 1902.17 38.1234658  -0.00746    0.00096   3210.5781
#> 14522 715012.5 9557732 2721.34 35.1813275  -0.00382    0.00672   1114.8801
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.2343959   0.00993   -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.0717109  -0.00669   -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.4803764   0.00225    0.00325    616.8652
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.5250811  -0.00089   -0.00051    779.0555
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499    885.0064
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613    289.6113
#> 41037 715532.5 9558792 2529.58 32.8465245  -0.00585   -0.00974   2136.6939
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560    977.4719
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063    576.2819
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319   3457.7519
#> 39732 715362.5 9559582 2230.97 23.3468842  -0.00161   -0.00409   2312.8413
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739    365.7626
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316    454.9674
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340  -0.00131    0.00201    788.0621
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.4746819   0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.9357102  -0.00771   -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448   1258.8196
#> 18897 712802.5 9559882 1851.71 50.2346476  -0.02627   -0.05393   8301.6094
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 46461 712682.5 9560032 1892.30 22.9778994   0.01453    0.01337    486.7551
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319  -0.02556   -0.11694 102301.4609
#> 18348 714042.5 9558882 2341.41 44.0638922  -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195   1901.6743
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322   4605.1411
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522   1355.7436
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970   1443.9200
#> 22297 714642.5 9558172 2606.18 59.4690084  -0.00495   -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.2711002   0.00705   -0.00915   2036.0378
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391   1794.7964
#> 7409  714772.5 9559122 2404.90 30.2155023  -0.01016   -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483    343.1384
#> 13278 712752.5 9560452 1901.25 42.1135439  -0.00118   -0.00252   3631.7017
#> 41408 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655    130.9468
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797   7557.7866
#> 28253 714102.5 9559892 2288.88  2.9576081   0.00378    0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301    338.9676
#> 27884 714072.5 9559202 2446.27 19.7567307  -0.00009    0.00109    605.1908
#> 25212 714272.5 9559012 2433.45 37.6714021  -0.01909   -0.01181   1286.3949
#> 40459 712542.5 9559192 1992.60 39.8171290  -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521   0.00390   -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948  -0.00124    0.00204    853.2902
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118   2273.4758
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875   1165.3953
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.5508993  -0.00195   -0.00995   1026.7981
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.8258980   0.00846    0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502   1428.8546
#> 34347 714742.5 9560142 2161.95 26.5669070  -0.00978    0.00367   3002.8828
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943    371.5514
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.9563087  -0.00301   -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759   4131.5356
#> 45486 712762.5 9560172 1853.82 56.7973062   0.01978    0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.8249337  -0.00452   -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780  -0.00305   -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.5953803  -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552   2515.0752
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756   0.01142    0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554   0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599  12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534   0.00165    0.00555   1882.0179
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835   0.02523    0.02247    282.1076
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836    250.5111
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.0318136   0.01600   -0.01849    764.8193
#> 40750 712962.5 9558702 2126.06 34.7040537  -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972    331.7462
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935    903.6397
#> 9320  713872.5 9559262 2378.76 29.8585496   0.00309    0.01011    501.7867
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436  -0.00181   -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.8455602   0.00298    0.00222    991.8300
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388    180.4414
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.2974072   0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983