Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603