Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027 33.905923
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123  0.00393   2081.0056 37.592907
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166 27.815382
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441 15.665239
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537 20.192179
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113 30.966650
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403   1787.3024  4.866131
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761    928.6858 28.902283
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610 19.677663
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876 26.322827
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918 16.677656
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006 29.548579
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416    947.4878 34.912037
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119 21.747759
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104 24.774122
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017 35.138929
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751 51.208103
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766 140591.4062 28.526423
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884 33.618299
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356 28.852436
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721 35.008294
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801 28.398653
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734 15.456682
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685 14.328529
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833 29.820734
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019 34.875368
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773    179.3077 25.710908
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138 37.901731
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686 27.826841
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555 28.203847
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757 34.391219
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657 34.595765
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927 36.516892
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224 37.588323
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412 34.148858
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346    543.0749 34.251990
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154  0.00124   2114.2253 40.408421
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836 35.319410
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064 31.188958
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299 23.826450
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120 21.700204
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159 16.218143
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536 28.610647
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072  54831.6289 39.360482
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364 29.354920
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875 33.218947
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528 31.595185
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688 27.951746
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252   5761.3486 34.251417
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001  9.812475
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622 34.900005
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933 41.733673
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246 34.187819
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150 33.241865
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162 39.358190
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674 28.246819
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270 28.552206
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081    940.6002 28.186086
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254 24.665260
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688 37.425030
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085 29.818443
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220   8936.2666 25.429013
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611 29.738228
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551 24.941426
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547 29.821307
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609 37.775107
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220 13.220428
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718 32.056989
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002 42.773591
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818 34.189538
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436 25.726951
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854 25.336194
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608 45.093497
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334 31.517262
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031 29.273560
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562 33.408596
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908 25.370571
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410 42.130160
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685   2568.4631 31.833535
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893 21.668691
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000 24.747766
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632    620.0229 31.225054
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164 28.512099
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454 33.813677
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626 17.295304
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367   1966.2112 34.060049
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398 29.404767
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318 26.486120
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161 19.775638
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782 27.416030
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676 15.762069
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427 39.291154
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525 33.383959
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691 29.823599
#> 42329 715852.5 9558702 2651.92 22.87076  0.00321  0.00548    532.8210 19.227891
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196 27.986124
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321 32.361229
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172 36.135302
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546 25.388906
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604 47.873489
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403 33.143317
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773 19.295500
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898 31.061761
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347 24.137566
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228 30.102630
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738 28.757325
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949 38.077629
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812 25.853575
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504 40.960180
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179 22.100128
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528    355.3475 25.230769
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356 28.853009
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000 34.377468
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944 33.421774
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738 32.879756
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252 39.771292
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511 13.572224
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179 21.777553
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978 26.147502
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003 21.927095
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111 22.860443
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354  0.00907   1281.5044 10.144791
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346 26.064996
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193 45.583949
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066 30.020697
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074 37.771670
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562 29.166417
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886  43249.3945 38.929617
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055 31.814627
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672 29.121153
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606 38.910710
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128 42.407471
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220 36.130719
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785 37.533892
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812 48.896791
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861 36.819987
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.978458
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820 40.158612
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288 23.996618
#>       distroad slides distdeforest distslidespast log.carea
#> 4965    300.00   TRUE       300.00            100  2.699406
#> 15412   300.00   TRUE       300.00            100  3.318273
#> 27864    30.00   TRUE       183.39             20  3.550683
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 30059   300.00   TRUE       300.00             41  2.412719
#> 48187   279.54   TRUE        21.35             46  3.521349
#> 9669    300.00   TRUE       300.00              1  3.610558
#> 37104    50.99   TRUE         0.00             21  3.252198
#> 28293   300.00   TRUE       300.00              4  2.967869
#> 47255   300.00   TRUE       260.03            100  3.112324
#> 7349    300.00   TRUE       300.00              8  2.751808
#> 31023   135.00   TRUE         0.00              2  2.577367
#> 38993   300.00   TRUE       300.00              2  3.128238
#> 22814   300.00   TRUE       300.00              6  2.976574
#> 25690   300.00   TRUE       300.00            100  2.732242
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 3835    300.00   TRUE       300.00             39  2.691790
#> 33792   300.00   TRUE       300.00              5  2.596460
#> 18843   300.00   TRUE       300.00            100  5.147959
#> 48498   300.00   TRUE       300.00            100  3.098674
#> 22283   256.35   TRUE         0.00            100  2.772566
#> 29420   253.02   TRUE         0.00            100  3.954264
#> 38032   300.00   TRUE       300.00             60  2.828706
#> 14522   300.00   TRUE       300.00            100  3.047228
#> 20290    95.53   TRUE         0.00             49  4.011286
#> 27978   300.00   TRUE       300.00             12  2.377795
#> 2942     68.13   TRUE       116.32             10  3.654262
#> 25081   300.00   TRUE       300.00            100  3.617357
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 27708   300.00   TRUE       300.00            100  2.253599
#> 28230   300.00   TRUE       300.00            100  2.977867
#> 1635    300.00   TRUE       300.00            100  3.303016
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 28577   300.00   TRUE         0.56              0  3.781222
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 22735   300.00   TRUE       300.00             90  2.622310
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 15990   300.00   TRUE         0.00              0  3.754443
#> 29667   300.00   TRUE       300.00             27  2.734860
#> 5728    300.00   TRUE       300.00              2  3.325151
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 38686   300.00   TRUE       300.00             75  2.946946
#> 6310    300.00   TRUE       300.00            100  4.032926
#> 18564   300.00   TRUE       300.00            100  2.676979
#> 39278   125.47   TRUE         0.00              6  3.006344
#> 18394   300.00   TRUE       300.00             57  3.372940
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 33240   300.00   TRUE       300.00            100  2.756740
#> 25244   300.00   TRUE       300.00              6  5.916915
#> 5084    300.00   TRUE       300.00             18  2.722101
#> 4617    300.00   TRUE       300.00              6  2.211304
#> 21053   300.00   TRUE         0.00              2  3.760524
#> 462      61.17   TRUE        57.16             74  3.452385
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 13029   300.00   TRUE        96.65              4  2.813330
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 23213   300.00   TRUE       300.00            100  2.868983
#> 32735   300.00   TRUE       300.00            100  3.245566
#> 664     300.00   TRUE       300.00             12  2.885027
#> 16355   300.00   TRUE        58.52            100  2.657980
#> 17728    30.14   TRUE       140.15              2  3.303483
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 5239    197.29   TRUE         4.67             57  3.591568
#> 37769   300.00   TRUE       300.00            100  4.672310
#> 39166   300.00   TRUE         1.90             90  2.994673
#> 8136    115.18   TRUE        84.59             30  3.951156
#> 32879   300.00   TRUE       300.00              6  3.130450
#> 46461   288.47   TRUE         0.00            100  2.687311
#> 17345   300.00   TRUE       300.00              2  4.693266
#> 27244   300.00   TRUE       300.00             54  5.009882
#> 7       300.00   TRUE       300.00              0  2.263215
#> 4588    300.00   TRUE       300.00              1  3.610558
#> 20797   285.07   TRUE         0.00             96  3.663802
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 1036    300.00   TRUE       300.00            100  3.132178
#> 47298   300.00   TRUE       300.00            100  3.126386
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 4667     24.78   TRUE       172.24             38  3.119102
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 20426   300.00   TRUE       300.00             76  4.542933
#> 7409    300.00   TRUE       300.00            100  2.866282
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 20171   300.00   TRUE       300.00            100  3.645966
#> 28260   300.00   TRUE       300.00            100  3.409673
#> 10456    24.34   TRUE         0.00             18  2.776185
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 22847   300.00   TRUE       300.00             41  3.134623
#> 11057   300.00   TRUE       300.00            100  2.792408
#> 43834    20.21   TRUE       150.14              2  3.331390
#> 41408   300.00   TRUE       205.28              1  3.147505
#> 41310   300.00   TRUE       300.00             25  3.046207
#> 21490   300.00   TRUE       300.00              1  3.293630
#> 36744   300.00   TRUE       300.00            100  2.937989
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 4785    300.00  FALSE         0.00            100  3.050663
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 10145   276.16  FALSE        91.33            100  3.227418
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 2664    300.00  FALSE       300.00            100  2.881694
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 20067   300.00  FALSE       268.03            100  2.590250
#> 42329   300.00  FALSE       300.00            100  2.726581
#> 43243   300.00  FALSE       300.00            100  2.922424
#> 49835   300.00  FALSE         0.00            100  3.134665
#> 47005   300.00  FALSE       300.00            100  4.858077
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 39685   300.00  FALSE       300.00            100  2.918481
#> 40139   300.00  FALSE         1.11             21  3.274260
#> 36126   300.00  FALSE       300.00            100  2.525145
#> 43403   300.00  FALSE       111.88             86  2.788585
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 44668     9.15  FALSE       163.30              6  2.852799
#> 6143    300.00  FALSE       185.58            100  2.985864
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 36245   110.19  FALSE        81.50             20  3.615611
#> 13737   300.00  FALSE       300.00            100  2.678473
#> 3733    300.00  FALSE       300.00            100  3.055767
#> 29107   300.00  FALSE       300.00            100  2.550653
#> 29835   300.00  FALSE       300.00             45  3.616112
#> 13908   300.00  FALSE       300.00             96  5.681725
#> 45486   300.00  FALSE         0.00             68  3.366011
#> 33410   300.00  FALSE       300.00            100  2.986312
#> 7553    300.00  FALSE       300.00            100  2.739830
#> 330     300.00  FALSE       275.17            100  2.372638
#> 26724   300.00  FALSE       185.65            100  3.274624
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 40612    55.73  FALSE       125.03            100  2.946551
#> 32430   300.00  FALSE       300.00            100  2.398827
#> 14763   110.04  FALSE        60.35            100  3.107720
#> 44216   300.00  FALSE       300.00            100  3.167741
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 18308   300.00  FALSE       300.00             55  4.224567
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 10981   300.00  FALSE       300.00            100  5.155983
#> 1577    300.00  FALSE       300.00             63  4.635980
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 22628   300.00  FALSE       300.00            100  2.951467
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 6577    300.00  FALSE       300.00             70  2.789872
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 23808   300.00  FALSE       300.00             72  2.692387
#> 15880   300.00  FALSE       300.00            100  2.546026
#> 11025   300.00  FALSE       300.00             50  2.828200
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 30296   300.00  FALSE       300.00             68  3.258250
#> 40043   194.79  FALSE        24.89             54  2.973603