Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982