Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292    1233.9299
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193    1809.3838
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063   18733.8848
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242     507.9263
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133    2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325     595.0426
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000  162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409    1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947     842.1514
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948    1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81444   300.00   TRUE       300.00             40  5.545499
#> 27864 27.81538    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906   300.00   TRUE       300.00            100  3.120164
#> 42733 31.28292   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73881   300.00   TRUE       300.00             11  3.001122
#> 37265 36.51747   300.00   TRUE         2.45              0  3.071539
#> 15843 42.22183   300.00   TRUE       300.00            100  3.091290
#> 42917 38.80357   300.00   TRUE       300.00            100  3.010367
#> 43327 29.05870   173.40   TRUE        33.81            100  3.516154
#> 7349  26.32283   300.00   TRUE       300.00              8  2.751808
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13374   300.00   TRUE       300.00            100  2.768442
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53332   300.00   TRUE       300.00            100  2.938827
#> 10899 24.77412   300.00   TRUE       300.00             20  2.644350
#> 30800 33.24817   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642   300.00   TRUE       300.00            100  5.147959
#> 22283 19.94237   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51363    77.82   TRUE        76.41             54  3.524078
#> 23375 34.63988   300.00   TRUE       195.00              2  3.224136
#> 27978 14.32853   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97524   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 6676  34.62155   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56596   300.00   TRUE       300.00             40  3.952656
#> 10320 16.45592   300.00   TRUE       300.00             30  2.721899
#> 38077 26.91527   300.00   TRUE       300.00            100  2.777281
#> 15390 29.32226    60.00   TRUE       118.92              2  3.372821
#> 7879  28.58085   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 19128 14.65970   300.00   TRUE       300.00             15  2.478266
#> 29667 34.25199   300.00   TRUE       300.00             27  2.734860
#> 17594 19.23190   138.35   TRUE         0.00             60  3.701999
#> 38686 31.18896   300.00   TRUE       300.00             75  2.946946
#> 28743 34.65306   300.00   TRUE       300.00             71  4.272628
#> 41037 31.57055   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43707   300.00   TRUE       300.00            100  3.023906
#> 34707 27.70251   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83279   152.09   TRUE        90.23             45  2.971800
#> 35194 23.44944   300.00   TRUE       300.00            100  3.381318
#> 40733 27.01381   300.00   TRUE       300.00            100  3.433836
#> 15766 23.33714   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409   300.00   TRUE       300.00            100  2.851936
#> 19887 38.06216   300.00   TRUE       300.00             63  2.760635
#> 39732 29.16814   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54114   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84251   300.00   TRUE       300.00             54  2.563199
#> 32735 39.35819   300.00   TRUE       300.00            100  3.245566
#> 44650 37.36258   300.00   TRUE       300.00            100  3.110849
#> 26010 25.22962    69.52   TRUE       166.56              2  3.647295
#> 30268 47.27246   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92007   300.00   TRUE       300.00            100  2.936990
#> 6845  44.59617   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700   300.00   TRUE       300.00            100  2.516921
#> 46461 24.94143   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15422   300.00   TRUE       300.00            100  4.035908
#> 20368 30.34442   300.00   TRUE       166.26            100  3.803001
#> 18440 34.94470   300.00   TRUE       300.00            100  4.477051
#> 47702 33.66643   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66835   300.00   TRUE       122.87            100  3.878395
#> 42201 39.55873   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451   300.00   TRUE       300.00              1  3.610558
#> 43245 24.44582   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76764   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942   131.25   TRUE        80.55              2  3.618527
#> 15981 30.00351   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64809   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525   300.00   TRUE         0.00            100  3.486483
#> 2801  25.74643   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62135   300.00   TRUE       300.00            100  2.201225
#> 22301 33.61543   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77833   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23593   168.72   TRUE         0.00              5  3.379717
#> 48650 27.05220   300.00   TRUE       300.00            100  3.027918
#> 14950 41.87405   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348   300.00   TRUE        36.63            100  2.774548
#> 10456 21.66869    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297   300.00   TRUE       300.00            100  2.929656
#> 28588 31.62555   300.00   TRUE       300.00             78  5.209992
#> 30626 30.00351   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71778   300.00   TRUE       300.00             83  2.904587
#> 24532 32.54629   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78423   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90610   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939   300.00   TRUE       300.00             57  3.322176
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 10785 33.24530   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18515   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325   300.00   TRUE       300.00            100  4.161853
#> 17374 43.14544   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66835   300.00   TRUE       122.87            100  3.878395
#> 31983 19.77564   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37385   256.60  FALSE         0.00             67  2.937807
#> 48601 35.85799   300.00  FALSE       245.19            100  2.998830
#> 15186 38.54860    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04611   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35071   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219   300.00  FALSE       300.00            100  3.356690
#> 30267 25.27030   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52737   300.00  FALSE       300.00            100  2.834598
#> 37832 36.44642   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331   294.31  FALSE       152.09            100  3.054287
#> 28162 18.66983   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93477   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505   300.00  FALSE       300.00            100  2.626221
#> 44396 21.44237   300.00  FALSE       300.00            100  2.270336
#> 34616 31.49950   300.00  FALSE       300.00            100  2.321589
#> 29944 18.62800   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989    80.73  FALSE         0.00            100  3.040329
#> 33175 34.78369   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07763   300.00  FALSE       300.00             90  2.432641
#> 34228 10.74869   300.00  FALSE       300.00            100  2.427333
#> 3733  22.10013   300.00  FALSE       300.00            100  3.055767
#> 17746 36.39141   300.00  FALSE       300.00             87  2.773624
#> 13908 34.37747   300.00  FALSE       300.00             96  5.681725
#> 20764 39.85036   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70175   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879   300.00  FALSE       162.53            100  2.925390
#> 36889 19.64500   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820   234.82  FALSE         0.00             43  4.109707
#> 1479  43.67657   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95634   300.00  FALSE       300.00             75  2.485742
#> 14763 10.14479   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06500   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42137    86.41  FALSE       172.46            100  6.734528
#> 40750 32.09996   300.00  FALSE       300.00            100  3.272034
#> 280   29.32341    60.55  FALSE       127.54             29  3.333837
#> 48050 23.36121   300.00  FALSE        72.23            100  2.520806
#> 11394 40.60380    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69280   300.00  FALSE       300.00             35  2.955995
#> 35257 30.88185   300.00  FALSE       300.00             79  2.812912
#> 49812 33.35818   300.00  FALSE       300.00             65  3.024717
#> 32342 37.61984   300.00  FALSE       300.00             81  2.961658
#> 30403 36.13072   300.00  FALSE       209.91            100  2.884241
#> 23808 37.53389   300.00  FALSE       300.00             72  2.692387
#> 14645 45.21325   300.00  FALSE       300.00            100  3.232313
#> 46567 16.94809   300.00  FALSE       300.00             78  2.785235
#> 24516 36.25448   300.00  FALSE       122.13            100  3.751491