Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv      carea
#> 4965  715042.5 9559312 2320.49 42.85782   -0.01106  -0.04634   500.5027
#> 47223 715302.5 9557472 2857.45 39.53351   -0.01021  -0.01579  1318.7560
#> 16749 715072.5 9558242 2745.27 49.05378    0.00097   0.01483   566.7674
#> 25090 715362.5 9560102 2059.29 49.11967    0.02059  -0.00628   556.0121
#> 42733 715112.5 9559062 2399.70 36.05452   -0.00041   0.01741   746.4155
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046   0.00324   861.3537
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514  -0.00644  1179.0695
#> 48187 712722.5 9560152 1835.29 29.32799   -0.00333  -0.01767  3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486    0.00065   0.00005   769.3426
#> 20799 714792.5 9561002 1857.29 23.12916    0.00090   0.00890  2301.0454
#> 25169 712642.5 9560232 1883.10 46.84503    0.00884   0.01345  1120.6929
#> 14456 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010  1527.1028
#> 43327 713962.5 9561082 1839.11 63.36168   -0.04951  -0.00529  3282.1186
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068   0.00588  1295.1610
#> 37270 714852.5 9561032 1838.96 35.15898   -0.00854   0.01264  5057.1431
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730   0.01770   474.5059
#> 44328 713222.5 9558932 2227.94 29.15553   -0.00769  -0.00281  1325.6138
#> 39061 712922.5 9559912 1931.50 36.31521    0.00414  -0.00104  1641.4587
#> 33792 714652.5 9558172 2598.36 59.45870   -0.03525  -0.11024   394.8751
#> 30800 715182.5 9557582 2772.37 39.20406   -0.02475   0.00665 25318.3555
#> 2107  714142.5 9558562 2382.67 35.55375    0.01412   0.00898   458.6938
#> 18005 714832.5 9557912 2660.76 31.91203    0.00191  -0.00261   554.9243
#> 29472 714792.5 9561072 1786.08 36.17312   -0.01029  -0.02312  2657.6936
#> 2942  713232.5 9560652 1851.75 34.53503    0.00856   0.01114  4510.8833
#> 47210 714952.5 9557612 2744.35 40.88455   -0.00493   0.00733   598.7983
#> 20194 714132.5 9558452 2447.22 24.02240   -0.00987   0.02107  1392.5880
#> 5749  712852.5 9560162 1934.78 23.38871    0.00071   0.00150  1274.3594
#> 23031 714042.5 9558492 2404.94 28.49606    0.01205   0.02205   629.4165
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804  -0.01596  5296.9014
#> 27013 713062.5 9559782 1976.81 41.63226    0.01414   0.03416  1204.8629
#> 12084 712832.5 9559992 1872.46 56.05475   -0.03184   0.03924  8023.8657
#> 46756 713372.5 9559192 2120.47 46.00507    0.03696   0.02514   297.7842
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543  -0.05146  1347.6224
#> 29667 713892.5 9559712 2239.25 41.59273    0.00804   0.00346   543.0749
#> 6557  713312.5 9559022 2181.69 27.63433   -0.00669  -0.00991  1403.0193
#> 34563 715022.5 9558162 2675.70 18.88985   -0.01993   0.01373 16031.7500
#> 6310  713802.5 9559752 2238.38 19.65131   -0.02379   0.01338 10787.6299
#> 43092 715532.5 9558192 2770.31 36.83316   -0.00442  -0.00178   800.1933
#> 47091 713382.5 9559182 2122.71 48.03793    0.04527   0.04613   289.6113
#> 37341 714412.5 9559142 2377.24 40.59979   -0.00735  -0.00925   665.6116
#> 39278 714692.5 9561052 1830.60 48.39488   -0.01996   0.00626  1014.7159
#> 18394 715022.5 9559942 2216.16 42.13761   -0.00613   0.00903  2360.1536
#> 26567 714022.5 9558872 2327.42 46.66569    0.00290   0.00230  1056.5876
#> 2976  715122.5 9559142 2390.05 33.16795    0.01008   0.00102   483.4184
#> 16015 712582.5 9560312 1899.18 56.54692    0.03480   0.02530   937.1306
#> 39599 714112.5 9558562 2374.25 29.16814    0.00283   0.00117  1018.2141
#> 35194 713692.5 9559952 2185.68 28.79686   -0.00449  -0.00181  2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996    0.00478   0.01092   412.5849
#> 15766 715562.5 9558592 2646.16 29.02089    0.00194   0.00516   576.0335
#> 48808 715032.5 9557792 2707.02 25.43761   -0.02329  -0.01181 11655.4316
#> 42941 713252.5 9559742 1990.10 28.95442   -0.04708   0.00408 76318.0156
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221   0.00121  4159.6738
#> 13029 713542.5 9560242 2019.80 29.54514    0.00576  -0.00166   650.6246
#> 6317  715042.5 9557702 2745.36 27.87382    0.00090  -0.00299   688.0706
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209   0.02739   365.7626
#> 19966 715362.5 9559572 2235.06 26.04953    0.00126  -0.00485  2240.4185
#> 27695 714142.5 9558452 2449.28 20.49012   -0.00539   0.00809  1102.9956
#> 5103  714612.5 9559892 2220.36 49.00795   -0.00078   0.00219   420.7032
#> 28046 714882.5 9558492 2711.77 28.30583   -0.00131   0.00201   788.0621
#> 45592 714962.5 9559862 2282.08 22.16086   -0.00028   0.01318   289.0063
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478  -0.07132 42970.7891
#> 42765 713552.5 9559632 2186.95 44.01118    0.00394  -0.00224   942.0322
#> 41783 715572.5 9558652 2609.51 32.40821    0.00291  -0.00081   940.6002
#> 34192 712702.5 9560102 1867.01 35.47468    0.00629  -0.00469  1021.5407
#> 40270 713312.5 9559012 2186.56 30.93571   -0.00771  -0.00599  1378.5760
#> 18897 712802.5 9559882 1851.71 50.23465   -0.02627  -0.05393  8301.6094
#> 40216 714602.5 9560592 2015.55 34.36887   -0.00093   0.00664  1807.2659
#> 37714 713192.5 9560692 1885.05 18.03442    0.00188   0.00832  3022.3462
#> 43859 714892.5 9557602 2722.70 28.12936    0.00370  -0.00190   640.8842
#> 10073 715572.5 9558612 2632.78 32.16127    0.00129   0.00171   788.0004
#> 17345 713922.5 9559682 2227.44 34.49091   -0.06509  -0.01211 49347.5547
#> 16009 715122.5 9559402 2264.81 50.01864   -0.00589   0.00849  1120.5961
#> 42201 714642.5 9557452 2592.29 13.22558   -0.02084  -0.02236  9755.3027
#> 34887 714662.5 9557392 2609.02 51.02075   -0.01878  -0.04351  1442.9247
#> 20797 713012.5 9560392 1842.25 36.64982    0.01299   0.00141  4611.0718
#> 27915 715232.5 9559552 2208.93 25.48058    0.00171  -0.00261  1627.8989
#> 15981 713302.5 9559172 2106.91 35.99779    0.01354   0.00245   629.5983
#> 5572  715062.5 9558262 2754.60 39.91167    0.00707   0.01833   369.3771
#> 26101 715562.5 9558682 2594.87 32.41108    0.00079   0.00321  1170.5904
#> 43074 713172.5 9559622 2052.40 31.52758   -0.00193   0.02812  1423.7578
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511  -0.01801  1071.6141
#> 37439 714602.5 9560672 1982.99 18.90646   -0.00070   0.00010  3467.2690
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281   0.04759   158.9370
#> 23626 713742.5 9560032 2137.92 39.30490   -0.00605   0.00075  2170.2249
#> 20426 714702.5 9559712 2208.61 43.04632   -0.08153  -0.02417 34908.6562
#> 14950 713382.5 9559172 2111.91 48.10668    0.00915  -0.02705   413.5702
#> 28260 715312.5 9558302 2681.75 34.48404   -0.01125  -0.02685  2568.4631
#> 22847 714582.5 9560382 2040.49 25.33619    0.00181  -0.00341  1363.4000
#> 18498 714842.5 9557712 2679.04 28.93953    0.00902   0.01358   802.7622
#> 10213 715542.5 9558782 2532.10 35.41338   -0.00893  -0.00867  2247.8748
#> 14779 712502.5 9560202 2004.61 27.52260    0.00115   0.01015   790.2421
#> 48461 714672.5 9559272 2316.84 40.38149   -0.01966  -0.02234   849.0547
#> 41408 713172.5 9559612 2056.81 22.27259   -0.00166   0.01476  1404.4454
#> 17322 712612.5 9560402 1924.48 43.80778   -0.00486   0.02306  1718.3652
#> 3771  714842.5 9561152 1724.51 45.53983    0.02405  -0.05655   130.9468
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585  -0.02665  1904.4769
#> 6470  713742.5 9558802 2260.79 30.68017    0.00000   0.00320   911.5911
#> 13072 714862.5 9558932 2448.29 41.05243   -0.04989  -0.02781 29995.1543
#> 29431 714882.5 9559892 2258.03 21.91793   -0.00849   0.00150   503.5590
#> 11288 712822.5 9560442 1873.42 39.60972   -0.00465  -0.00105  4432.8760
#> 16464 715252.5 9559982 2143.24 23.18588   -0.00298  -0.00712  1010.5097
#> 18286 714792.5 9557882 2650.57 28.25369   -0.00124   0.00204   853.2902
#> 35852 715112.5 9559962 2173.61 39.31579   -0.00738   0.00118  2273.4758
#> 32869 715332.5 9558892 2533.17 20.63507   -0.00242   0.01132   737.5982
#> 1488  714562.5 9560002 2150.93 11.28383    0.00197   0.00463   851.6306
#> 37832 712552.5 9559192 1998.10 41.55090   -0.00195  -0.00995  1026.7981
#> 441   714312.5 9560972 1916.59 31.30298    0.00540   0.00650  1133.1501
#> 13512 715152.5 9557992 2778.81 43.38952   -0.00724   0.01464  4029.5947
#> 42329 715852.5 9558702 2651.92 22.87076    0.00321   0.00548   532.8210
#> 45166 715262.5 9558722 2531.59 13.55332   -0.00502  -0.00028  2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629   -0.00031  -0.00199   836.4196
#> 49835 713272.5 9560372 1898.68 32.82590    0.00846   0.00794  1363.5321
#> 39685 715332.5 9558102 2797.04 47.78812    0.00167   0.01133   828.8604
#> 34616 714522.5 9558742 2568.77 34.99798    0.00990   0.02790   209.6952
#> 8413  713242.5 9561232 1967.45 38.19451    0.00236   0.00483  1097.3082
#> 33175 715822.5 9558832 2563.13 36.62289   -0.00674  -0.00376   629.8236
#> 22545 713402.5 9559562 2195.44 25.31557    0.04152   0.00098   164.0802
#> 33370 714202.5 9558962 2472.93 32.64141   -0.00307   0.02997   230.8641
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224  -0.02036   476.9504
#> 39440 714222.5 9557872 2508.52 39.04765    0.02785   0.00645   234.0114
#> 44267 712972.5 9559042 2191.01 22.26915    0.01462   0.00268   252.0072
#> 5656  713132.5 9560142 2042.24 33.59538   -0.00035   0.00955   713.4463
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097   0.01487  1769.3392
#> 41855 713182.5 9560452 1851.53 45.17830    0.00667   0.00863  1098.3462
#> 44715 714932.5 9560132 2195.78 30.39369    0.03479   0.02251   168.1733
#> 35826 712912.5 9560302 1891.22 50.86948   -0.00229   0.00099   909.0560
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740   909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199   107.2986
#> 1281  715102.5 9560032 2120.64 34.64905   -0.00930  -0.00360  4288.7329
#> 26724 714252.5 9560862 1975.16 28.28635    0.00165   0.00555  1882.0179
#> 15008 714492.5 9559012 2449.93 29.51535    0.00972   0.00308   422.5022
#> 24819 714032.5 9560882 1981.86 33.28426    0.00804  -0.00165   805.1403
#> 3868  714542.5 9560622 2024.15 29.44717    0.00360   0.00750  1892.8978
#> 28286 713702.5 9559842 2245.07 27.96893    0.00511   0.01499   617.3855
#> 44434 713022.5 9558762 2161.07 31.45424   -0.00273  -0.01047  2826.7700
#> 31098 714352.5 9559042 2410.15 30.33640    0.00537  -0.00347  1893.3645
#> 26283 715932.5 9557592 3075.48 32.76688   -0.00375   0.00215   917.5296
#> 40750 712962.5 9558702 2126.06 34.70405   -0.00264   0.00624  1870.8307
#> 24446 714742.5 9560082 2200.60 28.72180   -0.00598   0.02208   842.4825
#> 19767 714272.5 9558052 2431.67 55.72473    0.00361  -0.01841   756.0055
#> 10127 713722.5 9558712 2311.10 19.48114    0.01129   0.00491   261.7371
#> 7132  713192.5 9559612 2060.27 20.77373    0.00452   0.00368  1071.8734
#> 40653 714262.5 9561262 1791.94 12.22234    0.00161   0.00709  1181.2555
#> 22628 714942.5 9557732 2696.19 26.21167    0.00788  -0.01029   894.2672
#> 31440 714842.5 9558392 2704.41 25.80144    0.00992  -0.01052   451.7209
#> 49812 713392.5 9559222 2133.99 35.66261   -0.00983   0.00803  1058.5630
#> 36141 714222.5 9559422 2316.64 16.27716   -0.07627  -0.00124 25826.6914
#> 42990 712432.5 9560692 2102.11 32.45004   -0.00087  -0.00573  1490.0312
#> 11025 714402.5 9559252 2325.64 49.62044   -0.00241   0.00020   673.2861
#> 15363 714742.5 9559212 2373.13 33.58335   -0.00247   0.01057   722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775