Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324     861.3537
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057    1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835    2425.3115
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226    2325.2053
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281    1325.6138
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447     331.4833
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261     554.9243
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739     558.2286
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598   0.01212     561.1948
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642     613.6390
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338   10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175    0.00066  -0.00036     516.4709
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903    2360.1536
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102     483.4184
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117    1018.2141
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409    2312.8413
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941    1290.7709
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132   42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448    1258.8196
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793     864.9472
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334    4905.2002
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017     759.9937
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812    1423.7578
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753    1337.7854
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075    2170.2249
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044     288.5468
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166    1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895   0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109     605.1908
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064     683.2795
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409    1575.9425
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014     271.4219
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943     371.5514
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362     283.7090
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462     716.5762
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655     560.8554
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268     252.0072
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165     805.1403
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960     884.2003
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050    1471.4346
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270   35950.5508
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109     915.4989
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448     941.8503
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603