Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355  14919.6260
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775