Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766 140591.4062
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241    585.6182
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776   1021.8267
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632    620.0229
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840  0.00189    556.5211
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626    450.7346
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.64905 -0.00930 -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983