Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665   25318.3555
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973     514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107    1392.5880
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313     783.4548
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063    5035.0000
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373   16031.7500
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178     800.1933
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270    4057.5637
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175    1122.9120
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241     585.6182
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332    1418.6622
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389   47022.9688
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870    1895.1399
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661    2508.5374
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161    3065.3694
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422    1159.2100
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075    2170.2249
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294     734.9908
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606     919.4241
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252    3631.7017
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655     130.9468
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150     503.5590
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109     605.1908
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712    1010.5097
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204     853.2902
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445     355.5078
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409    1575.9425
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298     436.1459
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462     716.5762
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048    1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249     148.1857
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491