Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804     628.5153
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.5384969 -0.02716  0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357     275.7474
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006  0.00175     804.5679
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373     527.3528
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454    3476.0818
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851    1012.5963
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367     806.0966
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141    4611.0718
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484     495.0292
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380     601.0843
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705     413.5702
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119    6589.5796
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632     620.0229
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336    2806.2644
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181    1286.3949
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452     578.5847
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945     259.7421
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794    1363.5321
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973     215.5101
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179  0.00167  0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077    8759.9961
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759    1880.4403
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513     161.9878
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502     520.5218
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972     331.7462
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263 -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585    5714.3267
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603