Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603