Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681   -0.00029    5577.3916
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056   -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357    1752.0375
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305   -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292    0.00712    3636.5303
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863   -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498   -0.00752    1321.5997
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191   -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065    0.00005     769.3426
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431    4079.0366
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310    0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951   -0.00529    3282.1186
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407    0.00547     519.4228
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214    0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160    0.00311     586.7354
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952    0.01378     440.9104
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581    0.01189     287.5425
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412    0.00898     458.6938
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430   -0.06070  138869.3906
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382    0.00672    1114.8801
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993   -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108   -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206    0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356    0.00774     527.1068
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694    2123.4473
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646   -0.04086    6042.5757
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545    0.02405    5681.2412
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224    0.00177     291.5937
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458    0.00672     565.1107
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174    0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060    0.00270    4057.5637
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735   -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290    0.00230    1056.5876
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855   -0.02175    1122.9120
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449   -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399   -0.00731    1384.7959
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405   -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211    0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797   -0.00063     576.2819
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458   -0.00401    1722.6481
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576   -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529    0.02009    1247.7061
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548    0.05543     352.1558
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304    1569.7313
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568   -0.01558    1504.9717
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320    0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605   -0.03594   16587.9980
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133    0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556   -0.04394     587.8840
#> 10467 714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491    0.01061    5744.7583
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084   -0.02236    9755.3027
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143   -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175    0.04195    1901.6743
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263    0.01322    4605.1411
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079    0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428   -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169    0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789   -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250   -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124   -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669   -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410   -0.00380     601.0843
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296    0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767   -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016   -0.00294     734.9908
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881    0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257    0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915   -0.02705     413.5702
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143    0.00477    2945.2100
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451   -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212    0.02732    4237.6934
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004   -0.02384     947.2075
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115    0.01015     790.2421
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186    0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486    0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996    0.04684     209.1778
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084   -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203    0.00642    1688.1782
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159   -0.01559     866.5775
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849    0.00150     503.5590
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872    0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298   -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156   -0.00116    1817.5222
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007   -0.00463    1050.7473
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402   -0.01328    1066.7134
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494    0.01064     683.2795
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345   -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464    4029.5947
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405    0.00476   11992.5273
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174   -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338   -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065    0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457    0.01697     412.1185
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185    0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224   -0.02036     476.9504
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586    0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452   -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035    0.00955     713.4463
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418    0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154   -0.00204     618.2027
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119    0.00361    1011.0248
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745    0.00875    2156.9358
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361   -0.01841     756.0055
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284   -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309    0.01011     501.7867
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549   -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627   -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933   -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087   -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962   -0.00172     351.5812
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166    0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751    0.00388     180.4414
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336