Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203    634.3320
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896    276.0310
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257    357.1436
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.2583065 -0.06175 -0.08226   2325.2053
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269    491.8017
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665  25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385    674.0721
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629 360379.1875
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036    516.4709
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495    266.5888
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181   2406.1252
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473   2930.2791
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485   2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583   1898.0105
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123    364.2903
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011  0.00739    789.7573
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484    495.0292
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177   1558.2231
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935    903.6397
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250