Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393    2081.0056
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969     368.5606
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324     861.3537
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761     928.6858
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466     724.0713
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605     798.6775
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373     527.3528
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408   76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809    1102.9956
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961    1086.6963
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456    1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044     288.5468
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265    1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445     355.5078
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734     335.0773
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298     436.1459
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011     201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947     842.1514
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499     617.3855
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492     772.2691
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983