Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223   0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416    947.4878
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104   1641.4587
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325    616.8652
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887    366.8537
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903   2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072  54831.6289
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529   0.02009   1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558   1504.9717
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175   1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298   1533.7925
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915   2036.0378
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044    288.5468
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298  -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945    259.7421
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069  -0.00900    614.5898
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513    161.9878
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247    282.1076
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234   1707.3130
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020    673.2861
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603