Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593     939.0466
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542    3614.7334
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961    1086.6963
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118   19557.6621
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133    2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685    2568.4631
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015     790.2421
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306    1718.3652
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559     866.5775
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452     578.5847
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409    1575.9425
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853    1354.6001
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257     549.3252
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060     383.3547
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532   16771.3066
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491