Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983