Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.7898858   -0.04921  -0.12438     753.6304
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199   0.00659    3553.7166
#> 28223 712742.5 9560482 1927.69 34.6679573   -0.00292   0.00712    3636.5303
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023  -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753   0.01487    1077.7585
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498  -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333  -0.01767    3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677    1675.4679
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431    4079.0366
#> 28293 713762.5 9558812 2250.98 33.2854102    0.00118   0.00761     928.6858
#> 25169 712642.5 9560232 1883.10 46.8450293    0.00884   0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645   0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284   0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922   0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223   0.01783     377.8918
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979  -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414  -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.4586952   -0.03525  -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.8879661    0.01903   0.01447     331.4833
#> 15109 713872.5 9559602 2288.64 25.1855058    0.00602   0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.8566631    0.00471  -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.1731174   -0.01029  -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382   0.00672    1114.8801
#> 38070 715212.5 9559222 2342.41 55.6771737    0.00426  -0.00367     793.9665
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108  -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915   -0.02651  -0.03629  360379.1875
#> 34280 713652.5 9560702 1896.37 41.2735877    0.00382   0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669  -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646  -0.04086    6042.5757
#> 12868 712532.5 9560332 1913.92 61.9373106   -0.03543  -0.05146    1347.6224
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355   14919.6260
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353   0.00063    5035.0000
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752  -0.00517    1000.7119
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316   0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.5739922   -0.00856   0.01796    4729.6382
#> 31316 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661   0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617   0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 2976  715122.5 9559142 2390.05 33.1679538    0.01008   0.00102     483.4184
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283   0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175    1122.9120
#> 40733 712862.5 9558892 2108.60 29.9949135    0.00175   0.00236    2715.4138
#> 38862 712952.5 9559892 1950.86 37.7218224   -0.00048  -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.6488751   -0.01282   0.00542     860.8840
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299     688.0706
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126  -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550  -0.00941    1290.7709
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568  -0.01558    1504.9717
#> 13146 713792.5 9558552 2341.40 29.8253180   -0.00104   0.01664    1898.5588
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394  -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.4082118    0.00291  -0.00081     940.6002
#> 5239  714832.5 9560992 1859.70 19.9601307   -0.00320   0.00180    3904.5254
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024  -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601  -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188   0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567  -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246   0.00534     344.1447
#> 44767 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129   0.00171     788.0004
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319   -0.02556  -0.11694  102301.4609
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084  -0.02236    9755.3027
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175   0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384   0.00175    1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263   0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666  -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218  -0.00372    4154.5786
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540   0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196    1052.5341
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051   0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495  -0.03756     353.9086
#> 39624 715222.5 9559552 2205.54 23.8974967    0.00401  -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767  -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.5663341   -0.00693   0.01133    2397.2681
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110  -0.02450    2589.9910
#> 28260 715312.5 9558302 2681.75 34.4840379   -0.01125  -0.02685    2568.4631
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124   0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118  -0.00252    3631.7017
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144  -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514   -0.02312  -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893  -0.00867    2247.8748
#> 33142 715652.5 9558212 2740.94 14.3812407   -0.00465  -0.01036    2505.8369
#> 21490 714672.5 9560382 2021.86 32.8041256   -0.00783  -0.00367    1966.2112
#> 17374 715712.5 9557952 2832.22 37.8748021   -0.01519   0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586   0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378   0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332   0.01842     246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159  -0.01559     866.5775
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465  -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.0797253    0.00165   0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.4890128   -0.00226   0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.9754961    0.00872   0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298  -0.00712    1010.5097
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336   0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133  -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815  -0.03345  665381.5625
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480   0.01500     538.8209
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544  -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263   0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405   0.00476   11992.5273
#> 48183 713642.5 9558262 2314.44 50.0232262    0.01338  -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.9416565   -0.00065   0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900     614.5898
#> 44668 712612.5 9560472 1972.24 27.5071308    0.00601  -0.00011     712.5228
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301  -0.01099    1146.2140
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586   0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462   0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.5953803   -0.00035   0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650     766.6254
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009  -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233   0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229   0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534    0.00165   0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804  -0.00165     805.1403
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 4943  712452.5 9559172 1927.46 34.4651303    0.00105  -0.01576    1045.1074
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148     481.8292
#> 11394 713472.5 9561232 1889.48 34.3144424    0.00108  -0.00948    1538.1741
#> 18504 713372.5 9560672 1786.89  6.4738501   -0.00278  -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203   0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508  -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 46639 715232.5 9560042 2107.15 39.6572738   -0.01305   0.00585    5714.3267
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282   0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.8455602    0.00298   0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.9643862   -0.00933  -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166   0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.5082486    0.02751   0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336   0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227     841.2025
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491