Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123   0.00393    2081.0056
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633  -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.5335149   -0.01021  -0.01579    1318.7560
#> 40756 714022.5 9558862 2331.20 45.0854759   -0.00075   0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.9841669   -0.01863  -0.00507  159065.0781
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068   0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550  -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.1593594   -0.01607   0.00677    1343.5006
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021  -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194   0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952   0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414  -0.00104    1641.4587
#> 45869 714342.5 9560282 2127.11 27.6234412   -0.02169  -0.00160    9344.2930
#> 22283 713592.5 9560762 1832.60 36.0103974    0.00448  -0.00328     592.3326
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032     994.6740
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746   0.00096    3210.5781
#> 23375 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382   0.00672    1114.8801
#> 2942  713232.5 9560652 1851.75 34.5350311    0.00856   0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.7212213   -0.00491   0.00581    1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973     514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733     598.7983
#> 27708 715222.5 9557782 2844.65 29.5428498    0.03206   0.02773     179.3077
#> 34280 713652.5 9560702 1896.37 41.2735877    0.00382   0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733     598.7983
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225   0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804  -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097   0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383   0.00193    1809.3838
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646  -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.0547529   -0.03184   0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353   0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229  -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316   0.00714     475.3120
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093  -0.00907    1943.4092
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286  -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749   0.04779    2972.0193
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175    1122.9120
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731    1384.7959
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364  -0.03357     473.4709
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160   0.01690    5121.6924
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306  -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529   0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736   0.01684    2556.8049
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034   0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.0103974    0.00448  -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280   -0.00568  -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170   -0.02554   0.03304    1569.7313
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168  -0.04102   35685.0000
#> 35014 713502.5 9559772 2190.48 44.2357795    0.00927   0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180   -0.00104   0.01664    1898.5588
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604   0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588   0.01448    1258.8196
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743  -0.05242     507.9263
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017   0.01213     328.7922
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986  -0.02524   10861.9482
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.7916671   -0.04376  -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129   0.00171     788.0004
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259   -0.13038  -0.04212   46310.0195
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017     759.9937
#> 26101 715562.5 9558682 2594.87 32.4110766    0.00079   0.00321    1170.5904
#> 12629 714712.5 9558872 2533.15 37.0256786    0.04250  -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.8299297   -0.03669  -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.4334430    0.01410  -0.00380     601.0843
#> 48650 715592.5 9558612 2626.74 31.8765069    0.00010   0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881   0.00121    1783.0658
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143   0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379   -0.01125  -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181  -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610     454.8103
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212   0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245     629.5983
#> 37115 715612.5 9559202 2358.15  9.7276774   -0.01020   0.00279   39495.1797
#> 35458 715242.5 9558442 2623.01 38.9548275   -0.00770  -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.3236027    0.00758  -0.00288    2144.8164
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205    1112.2626
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316  -0.00427     866.9398
#> 15211 715772.5 9558082 2801.07 36.2968763   -0.00079  -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187    0.00599   0.00301     338.9676
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009   0.00109     605.1908
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156  -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204     853.2902
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295   0.02715     520.7355
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.6350750   -0.00242   0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.7659191   -0.00458   0.00479     604.0975
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480   0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195  -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540   0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.2905598   -0.00161   0.00171     756.8818
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253  -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176   0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818   0.00502    1428.8546
#> 45071 712652.5 9560032 1900.80 14.4734869    0.00987  -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538  -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.7201175    0.00174  -0.03065     704.4843
#> 8413  713242.5 9561232 1967.45 38.1945125    0.00236   0.00483    1097.3082
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180   0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.6228893   -0.00674  -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648    0.00358   0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983     270.7949
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782  -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302   -0.00393   0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645     234.0114
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004   0.00714  480535.5000
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310  -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667   0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270  -0.00250    2225.4426
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804  -0.00165     805.1403
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873  -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.2945705    0.00537  -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.5684513    0.00420   0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511   0.01499     617.3855
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082   0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.7040537   -0.00264   0.00624    1870.8307
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182  -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 39883 712732.5 9560842 2082.69 25.0909678    0.00732   0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.8360436   -0.00181  -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986  -0.05013     616.4128
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811  -0.02281     492.4785
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253    1237.0536
#> 46567 714472.5 9560352 2080.13 21.3816390    0.01336   0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181  -0.01149    4242.3579
#> 40043 713322.5 9560562 1802.09 39.8520794    0.04456  -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491