Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438     753.6304
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884    0.01345    1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284    0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831   -0.00149   12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378     440.9104
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126   -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973     514.5488
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629  360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898     588.3761
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694    2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950     950.3138
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440   -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077     300.7915
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096     520.5570
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771   -0.01939     703.7817
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373   16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552   -0.02058     431.4362
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771     674.3859
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588     749.2502
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260   -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427     380.9472
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529    0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807   -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262    1760.2162
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304    1569.7313
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165    0.00555    4439.0967
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389   47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394     587.8840
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509   -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241    1752.9242
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801     172.7491
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801    1071.6141
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441     424.2680
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669   -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778    4594.4351
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181   -0.00001    1184.7952
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713    6379.7974
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171    4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325     595.0426
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000  162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744     226.7911
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758   -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614   -0.00093     118.1402
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465   -0.00105    4432.8760
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872    0.05419     761.5427
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715     520.7355
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500     538.8209
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014     271.4219
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174   -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475     515.1661
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432   -0.00362     283.7090
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997     230.8641
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528     355.3475
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196     576.1450
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948    1222.5433
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165     805.1403
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502     520.5218
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836     250.5111
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719    0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273   -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578    1258.8865
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532   16771.3066
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148     481.8292
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884     672.2466
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181   -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109     915.4989
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208    0.00492     772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769