Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483     566.7674
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339    1002.5877
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120     512.9740
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024     394.8751
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385     674.0721
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596    5296.9014
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167    1938.5389
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338   10787.6299
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613     289.6113
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036     516.4709
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779    2972.0193
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588     749.2502
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409    2312.8413
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171     788.0004
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911     173.2806
#> 24135 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666 -0.03583    1898.0105
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833     369.3771
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801    1071.6141
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200    1305.6691
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044     288.5468
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291     293.4231
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093     118.1402
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831     454.9113
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064     683.2795
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.8795463  0.00825  0.00445     355.5078
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.5669070 -0.00978  0.00367    3002.8828
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645     234.0114
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863    1098.3462
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532   16771.3066
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935     903.6397
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448     941.8503
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227     841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901