Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649    1399.2329
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043    1806.4552
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057    1630.3259
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010    1527.1028
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804     628.5153
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606     564.6876
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842     868.6137
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510    3342.5518
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261     554.9243
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973     514.5488
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898     588.3761
#> 10320 714912.5 9558482 2724.21 25.28291  0.00356  0.00774     527.1068
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437    1155.1469
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346     543.0749
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096     520.5570
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866   29195.6016
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178     800.1933
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452     437.3965
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588     749.2502
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516     576.0335
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009    1247.7061
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262    1760.2162
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512   13178.3223
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961    1086.6963
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474    2343.2100
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376    2048.3198
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078  622888.2500
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321    1170.5904
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739     789.7573
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010    3467.2690
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778    4594.4351
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915    2036.0378
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152     850.4633
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252    3631.7017
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144  317260.6875
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684     209.1778
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389   14516.2002
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232     977.7320
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105    4432.8760
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158     837.3730
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290    2481.5632
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409    1575.9425
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923     811.5580
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273   72123.6172
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399    1601.4852
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697     412.1185
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170     253.5129
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099    1146.2140
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251     168.1733
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308     422.5022
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247     282.1076
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438    1276.5448
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624    1870.8307
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841     756.0055
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119     792.3544
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419     323.0790
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607