Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029   5577.3916 34.427888
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649   1399.2329 30.715694
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166 27.815382
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861 39.335272
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441 15.665239
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537 24.437796
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997 39.598259
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191 23.932447
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454 24.807927
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299 42.221833
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403   1787.3024  4.866131
#> 26531 714462.5 9560452 2104.74 19.52411  0.00453  0.00257    357.1436 15.987814
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147 44.795559
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610 19.677663
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713 38.770335
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354 20.133737
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138 25.504070
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587 35.240915
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547 35.780068
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751 51.208103
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606 28.757325
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240 33.825709
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269 22.635271
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961 29.199648
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488 36.829154
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363 21.073961
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285 34.621548
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924 28.565957
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788 25.728097
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448 25.599754
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501 29.483262
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880 28.117331
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257  0.00593    939.0466 32.075896
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059 29.322261
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204 34.561960
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498 27.153043
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915 14.659698
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647 23.193905
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817 29.935899
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559 27.209193
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064 31.188958
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120 21.700204
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382 22.183780
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613    289.6113 40.151736
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768 25.460526
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193 27.702509
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306 38.832788
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181   2406.1252 23.449444
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.650391
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849 21.719111
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091 27.974091
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622 34.900005
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840 35.894087
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738 28.379173
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009   1247.7061 32.541138
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993 36.320368
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150 33.241865
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313 27.831425
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359 31.442778
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855 33.985565
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270 28.552206
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891 25.325308
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322 42.353040
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407 33.874984
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254 24.665260
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196 29.979444
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594  16587.9980 28.710342
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315 28.727531
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094 31.788271
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659 26.364080
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462 30.731737
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293 19.325866
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004 25.761328
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708 41.828211
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500 32.694691
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002 42.773591
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411 24.445817
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694 35.475255
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903 39.713424
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436 25.726951
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925 27.017825
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207 28.038836
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621 27.998156
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690 26.592690
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210  0.556915
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100 33.798207
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341 32.922728
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475 23.160673
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713   6379.7974 34.347101
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410 42.130160
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702 41.874047
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474 42.573056
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796 30.320354
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870   2099.7930 28.569395
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278 27.558697
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389 33.245303
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487 27.251592
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457 43.145441
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866 38.668349
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024 30.834870
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590 20.891760
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446 27.146740
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565   1410.8372 38.548600
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113 50.344083
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991 32.295339
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134 39.136455
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345 665381.5625 22.984775
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975 25.270304
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795 23.527366
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632 32.957678
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075   8141.7549 35.462650
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626    450.7346 34.081249
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464   4029.5947 35.934767
#> 42329 715852.5 9558702 2651.92 22.87076  0.00321  0.00548    532.8210 19.227891
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818 31.480020
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367   3002.8828 31.720089
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347 24.137566
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520 31.248545
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783 32.072458
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504 40.960180
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511 13.572224
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613 19.645004
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018 22.878205
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560 44.227758
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740    909.4791 18.958600
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577    423.5598 26.992042
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978 26.147502
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451 32.145797
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561 30.964358
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019 32.003131
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906 23.095929
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201 29.185897
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209 24.035007
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220 36.130719
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790 34.976209
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812 48.896791
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536 31.394649
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.978458
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665 16.948092
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952 26.942194
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691 33.526625
#>       distroad slides distdeforest distslidespast log.carea
#> 31358   300.00   TRUE        15.00              9  3.746431
#> 36710   300.00   TRUE       300.00             21  3.145890
#> 27864    30.00   TRUE       183.39             20  3.550683
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 8598    300.00   TRUE       300.00            100  3.121100
#> 21939   300.00   TRUE       300.00            100  3.168886
#> 20799   180.67   TRUE         0.00             16  3.361925
#> 15843   300.00   TRUE       300.00            100  3.091290
#> 37104    50.99   TRUE         0.00             21  3.252198
#> 26531   300.00   TRUE       300.00              2  2.552843
#> 20151   300.00   TRUE       300.00            100  2.828669
#> 47255   300.00   TRUE       260.03            100  3.112324
#> 27375   300.00   TRUE       300.00            100  2.859781
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 44328   300.00   TRUE       300.00            100  3.122417
#> 39061   300.00   TRUE        70.65             56  3.215230
#> 8019    300.00   TRUE       300.00             61  2.928936
#> 33792   300.00   TRUE       300.00              5  2.596460
#> 43964   300.00   TRUE       300.00             46  2.952143
#> 48055   300.00   TRUE         1.67              0  3.566570
#> 15109   300.00   TRUE       300.00             40  2.931675
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 36030   300.00   TRUE         0.00            100  2.926238
#> 34275   300.00   TRUE       300.00            100  2.711427
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 6676    300.00   TRUE       300.00            100  3.303289
#> 6808    300.00   TRUE       300.00             40  3.952656
#> 45575   300.00   TRUE       300.00             59  2.562268
#> 40828   300.00   TRUE       300.00            100  2.941037
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 20194   300.00   TRUE       300.00             41  3.143823
#> 48364   300.00   TRUE       300.00            100  2.972687
#> 15390    60.00   TRUE       118.92              2  3.372821
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 35142   300.00   TRUE       300.00            100  2.176525
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 11687   300.00   TRUE       300.00             68  3.186466
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 36931   300.00   TRUE       300.00            100  3.408707
#> 38686   300.00   TRUE       300.00             75  2.946946
#> 18564   300.00   TRUE       300.00            100  2.676979
#> 46201   167.75   TRUE         0.00             55  3.674828
#> 47091   300.00   TRUE       300.00             25  2.461815
#> 17423   300.00   TRUE       300.00            100  2.868983
#> 34707   165.13   TRUE        39.56            100  3.473052
#> 16015   152.09   TRUE        90.23             45  2.971800
#> 35194   300.00   TRUE       300.00            100  3.381318
#> 835     300.00   TRUE       300.00            100  3.141386
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 375     300.00   TRUE       300.00            100  2.851936
#> 38862   300.00   TRUE       104.48             92  3.151879
#> 39520   300.00   TRUE       300.00            100  2.934945
#> 32909   118.07   TRUE        98.15              0  3.619059
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 7668    300.00   TRUE       300.00              5  2.789721
#> 29801   300.00   TRUE       300.00              2  3.026907
#> 32280   126.66   TRUE         0.00              5  3.195825
#> 31561   300.00   TRUE       300.00              0  2.888200
#> 11618   300.00   TRUE       300.00              1  2.578737
#> 17728    30.14   TRUE       140.15              2  3.303483
#> 17431   300.00   TRUE         1.11              5  4.633173
#> 42765   300.00   TRUE       300.00             12  2.974066
#> 34192   273.94   TRUE         4.48             85  3.009256
#> 5239    197.29   TRUE         4.67             57  3.591568
#> 41721   300.00   TRUE       300.00            100  3.099963
#> 7385    300.00   TRUE       300.00            100  4.219794
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 18897   300.00   TRUE        20.00              5  3.919162
#> 40216   300.00   TRUE       300.00            100  3.257022
#> 37714    13.57   TRUE       170.00              4  3.480344
#> 1973    300.00   TRUE       300.00            100  2.290100
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 43201   300.00   TRUE       300.00            100  3.530273
#> 4588    300.00   TRUE       300.00              1  3.610558
#> 46138   300.00   TRUE       300.00              8  5.794410
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 43245   300.00   TRUE       300.00            100  3.663243
#> 26428   300.00   TRUE         0.00            100  3.486483
#> 34164   300.00   TRUE       300.00              1  2.561448
#> 1036    300.00   TRUE       300.00            100  3.132178
#> 28009   300.00   TRUE       300.00             29  3.185767
#> 47244   300.00   TRUE       300.00              2  2.752448
#> 11352   135.89   TRUE        82.11              4  4.291317
#> 37439   300.00   TRUE       300.00            100  3.539988
#> 332      94.82   TRUE         0.00             90  3.079334
#> 10617    17.89   TRUE       145.83              4  3.064162
#> 38668   300.00   TRUE       300.00            100  3.022236
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 750     300.00   TRUE         0.00              0  3.804807
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 20171   300.00   TRUE       300.00            100  3.645966
#> 14950   300.00   TRUE       300.00             15  2.616549
#> 29024   300.00   TRUE       300.00             32  2.529363
#> 16065   300.00   TRUE       162.92            100  3.818858
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 30378   300.00   TRUE       300.00             57  3.322176
#> 31491   300.00   TRUE       300.00            100  3.385647
#> 10785   300.00   TRUE       300.00            100  3.383302
#> 39466   300.00   TRUE       300.00            100  3.060187
#> 17374   300.00   TRUE       300.00            100  3.630971
#> 31253   300.00   TRUE       122.87            100  3.878395
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 25584   300.00  FALSE       300.00            100  3.105832
#> 15186    60.00  FALSE         0.00            100  3.149477
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 15558   300.00  FALSE       300.00             45  2.657927
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 18746     4.48  FALSE       205.00             27  3.028048
#> 7517    144.37  FALSE        60.16            100  5.823071
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 8066    300.00  FALSE       300.00            100  2.834598
#> 49363   300.00  FALSE         0.00            100  3.394725
#> 43426   300.00  FALSE        82.11            100  3.910718
#> 9665    300.00  FALSE       300.00             64  2.653921
#> 13512   300.00  FALSE       300.00            100  3.605261
#> 42329   300.00  FALSE       300.00            100  2.726581
#> 39193   300.00  FALSE       117.12            100  2.879028
#> 34347   300.00  FALSE       300.00            100  3.477538
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 21925   300.00  FALSE       300.00             95  3.537385
#> 20319   300.00  FALSE       300.00            100  3.167901
#> 13737   300.00  FALSE       300.00            100  2.678473
#> 330     300.00  FALSE       275.17            100  2.372638
#> 36889   300.00  FALSE       300.00             48  2.108096
#> 40909   234.82  FALSE         0.00             43  4.109707
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 48988   300.00  FALSE       144.19            100  2.958793
#> 17628   300.00  FALSE       205.59            100  2.626915
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 484     300.00  FALSE        55.63            100  3.216864
#> 9007    300.00  FALSE        14.98             35  2.784158
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 33881   300.00  FALSE       300.00            100  3.171879
#> 31440   300.00  FALSE       300.00             96  2.654870
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 47245   300.00  FALSE       150.03            100  2.509309
#> 15880   300.00  FALSE       300.00            100  2.546026
#> 26635   300.00  FALSE       300.00            100  3.092389
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 46567   300.00  FALSE       300.00             78  2.785235
#> 15363   300.00  FALSE       300.00            100  2.858775
#> 48157    57.15  FALSE       157.06             45  2.887769