Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340    604.1032
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345   1120.6929
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588   1295.1610
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104   1641.4587
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272   4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032    994.6740
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856    0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107   1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083    994.7876
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276   9614.6924
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114    150.1498
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338  10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178    800.1933
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117   1018.2141
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401   1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739    365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883    616.1993
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240   1201.4724
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354   1147.7384
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393   8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793    864.9472
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534    344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776   1021.8267
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441    424.2680
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224   1137.3363
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610   1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685   2568.4631
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384    947.2075
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036   2505.8369
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367   1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505    351.4078
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204    853.2902
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132    737.5982
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548    532.8210
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273  72123.6172
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759   1880.4403
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734    335.0773
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032    994.9224
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305   -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513    161.9878
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060    383.3547
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066    306.0146
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177   1558.2231
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177    184.1042
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803   1058.5630
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419    323.0790
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156    758.4840
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946