Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261     323.8441
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661   -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761     928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949    6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392    1255.0884
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672    1114.8801
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774     527.1068
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146    1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063    5035.0000
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972    0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132   42970.7891
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822     472.8748
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058    1817.2371
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334    4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200    1305.6691
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294     734.9908
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632     620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158     837.3730
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463     851.6306
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171     756.8818
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399    1601.4852
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298     436.1459
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948    1222.5433
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251     168.1733
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099     909.0560
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233     968.7780
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935     903.6397
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135    1293.4855
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249     148.1857
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769