Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212    561.1948
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178    800.1933
#> 31316 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722   1373.4147
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380    601.0843
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 28260 715312.5 9558302 2681.75 34.48404 -0.01125 -0.02685   2568.4631
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218  0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260   1484.6593
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 26635 715092.5 9560412 2010.59 30.57188 -0.00463  0.00253   1237.0536
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983