Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491