Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579    2246.7725
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761     928.6858
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378     440.9104
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189     287.5425
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070  138869.3906
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694    2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846   -0.00034    4495.6929
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252    5761.3486
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690    5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262    1760.2162
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099    0.02169     495.8840
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558    1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822     472.8748
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870    1895.1399
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010    3467.2690
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140    1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110   -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856    1789.1596
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609    4275.3457
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363     389.2691
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431     315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216    1412.2295
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175     725.1355
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035    1471.9783
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165     805.1403
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247     282.1076
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119     792.3544
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775