Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613     289.6113
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241    1752.9242
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526     706.1463
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856    1789.1596
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288    2144.8164
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265    1148.6487
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011     501.7867
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419     323.0790
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983