Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969    368.5606
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005    769.3426
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501   3889.7292
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270   4057.5637
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656   3036.7383
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627 -0.05393   8301.6094
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213 -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121   1783.0658
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000 162178.1562
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400   1982.7782
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684    209.1778
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565   1410.8372
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463   1050.7473
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422    967.9738
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152  0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262 -0.00502    520.5218
#> 484   713812.5 9560592 2016.41 44.3010967 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603