Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357    1752.0375
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250   -0.01340     604.1032
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607    0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226    2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903    0.01447     331.4833
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164     895.6606
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471   -0.01382     560.8309
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284    2010.4285
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314    2360.1501
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325     616.8652
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077     300.7915
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175    0.00236    2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241     585.6182
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436    2602.7390
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558    1504.9717
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399    3900.3445
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822     472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335     195.0293
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911     173.2806
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212   46310.0195
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363     565.5207
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001    1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417   34908.6562
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067    1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105    4432.8760
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118    2273.4758
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466    2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945     259.7421
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734     335.0773
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038    2017.3547
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247     282.1076
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576    1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799    1036.6606
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607