Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203    634.3320
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542    674.0147
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226   2325.2053
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665  25318.3555
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530   9000.4356
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312   2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189   2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453    0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146   1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887    366.8537
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888    0.00862    874.5408
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373  16031.7500
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307 147470.5312
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102    483.4184
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357    473.4709
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009   1247.7061
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026    247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927    0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219    420.7032
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393   8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940    987.8085
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412   1350.3611
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263    0.01322   4605.1411
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109    0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246    0.00396    704.7076
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981   3975.3608
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778   4594.4351
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224   1137.3363
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044    288.5468
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015    790.2421
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063   2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345   2417.1389
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301    338.9676
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230    714.7432
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204    853.2902
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466   2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445    355.5078
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598    186.3527
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662   1269.2650
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997    230.8641
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196    576.1450
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305   -0.00705    968.9738
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650    766.6254
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740    909.4791
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836    250.5111
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347   1893.3645
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233    968.7780
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377   1727.6055
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799   1036.6606
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235