Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769