Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555    430.4604
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065   0.00005    769.3426
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588   1295.1610
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311    586.7354
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281   1325.6138
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665  25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392   1255.0884
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367    793.9665
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086   6042.5757
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887    366.8537
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495    266.5888
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241   1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259  -0.00281   1178.4420
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522   1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456   1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171   4425.5410
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143   0.00477   2945.2100
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166   1813.7841
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909  -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538    265.4677
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937  -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833    300.0125
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399   1601.4852
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196    576.1450
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948   1222.5433
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203   0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281    492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419    323.0790
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309