Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769