Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588    1295.1610
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367     793.9665
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694    2123.4473
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124    2114.2253
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063   18733.8848
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530     937.1306
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512   13178.3223
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211   49347.5547
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380     601.0843
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417   34908.6562
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450    2589.9910
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166    1813.7841
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744     226.7911
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291     487.3438
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538     265.4677
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400    4126.7812
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099    1146.2140
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955     713.4463
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282     561.2366
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740     909.4791
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710     472.0236
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907    1281.5044
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948    1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064    1485.5201
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419     323.0790
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491