Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104    1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769    1032.7177
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261     554.9243
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203   10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593     939.0466
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598   0.01212     561.1948
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437    1155.1469
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338   10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063   18733.8848
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230    1056.5876
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771     674.3859
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408   76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436    2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542    3614.7334
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026     247.0894
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020    1563.5549
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211   49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195    1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522    1355.7436
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010    3467.2690
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196    1052.5341
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970   0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389   14516.2002
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331     699.3024
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736     477.8591
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204     853.2902
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759    4131.5356
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948    1222.5433
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863    1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233   0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308     422.5022
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904     608.3561
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972     331.7462
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064    1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235