Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393   2081.0056
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305   -0.00555    430.4604
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075    0.00475   1001.0861
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659    0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661   -0.01339   1002.5877
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065    0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090    0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104   -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.00507 -0.00891   -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884    0.01345   1120.6929
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295    0.00035   2697.5974
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068    0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407    0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550   -0.00120    512.9740
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214    0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160    0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841   -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194    0.00526    539.8119
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123    0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528    0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448   -0.00328    592.3326
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131    0.00501   3889.7292
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029   -0.02312   2657.6936
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672    0.04398    267.2659
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151    0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426   -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043    0.00973    514.5488
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902    0.01358    802.7622
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613   -0.01898    588.3761
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071    0.00150   1274.3594
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452   -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383    0.00193   1809.3838
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453    0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545    0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154    0.00124   2114.2253
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669   -0.00991   1403.0193
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463    0.00167   1938.5389
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353    0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140   -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509    0.00499    885.0064
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174    0.00025   4425.5151
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093   -0.00907   1943.4092
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060    0.00270   4057.5637
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149    0.01949   1182.1506
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480    0.02530    937.1306
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478    0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175    0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364   -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797   -0.00063    576.2819
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221    0.00121   4159.6738
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736    0.01684   2556.8049
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062   -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550   -0.00941   1290.7709
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406   -0.00055    773.0359
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238   -0.00851   1012.5963
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771   -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472   -0.00228   2262.3796
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521   -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038   -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953   -0.01797   7557.7866
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175    0.04195   1901.6743
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354    0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166    0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379    0.00801    172.7491
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109    0.00298   1533.7925
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250   -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.35820  0.00540    0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546    0.01484    495.0292
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281    0.04759    158.9370
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181   -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970    0.00200   1305.6691
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010    0.00140   1066.3957
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708   -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019    0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143    0.00477   2945.2100
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181   -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451   -0.03921    525.3245
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360   -0.03000 162178.1562
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118   -0.00252   3631.7017
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186    0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195   -0.00205   1112.2626
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683   -0.00603   1562.0503
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316   -0.00427    866.9398
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156   -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533    0.01493   2212.0525
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112    0.03538    265.4677
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825    0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345   -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.30298  0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306   -0.00626    450.7346
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161    0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937   -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544   -0.01014    271.4219
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341    0.04431    315.9084
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065    0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236    0.00483   1097.3082
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193   -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218    0.00902    218.7467
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152    0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185    0.01655    560.8554
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393    0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785    0.00645    234.0114
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004    0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035    0.00955    713.4463
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.06078  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119    0.00361   1011.0248
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523    0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511    0.01499    617.3855
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820   -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273   -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074   -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952   -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070   -0.00050   1471.4346
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502    0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129    0.00491    261.7371
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203    0.00238    787.0053
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161    0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309    0.01011    501.7867
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172   -0.00119    792.3544
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298    0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962   -0.00172    351.5812
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026