Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958    0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123    0.00393    2081.0056
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097    0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075    0.00475    1001.0861
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661   -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046    0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514   -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191   -0.00579    2246.7725
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645    0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979   -0.01542     674.0147
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922    0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223    0.01783     377.8918
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542    0.01842     868.6137
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952    0.01378     440.9104
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475    0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514     297.7842
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131    0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017   -0.00032     994.6740
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029   -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126   -0.02044  129412.3984
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151    0.00189    2317.1819
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651   -0.03629  360379.1875
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937    0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252    0.01118    2359.5059
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416    1204.8629
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743    0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.9373106 -0.03543   -0.05146    1347.6224
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463    0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904   -0.00866   29195.6016
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442   -0.00178     800.1933
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661    0.01199     196.4080
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260   -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.9235105 -0.01364   -0.03357     473.4709
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221    0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576   -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209    0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856   -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550   -0.00941    1290.7709
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168   -0.04102   35685.0000
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568   -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539    0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927    0.01643     670.2289
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784    0.00596     842.0994
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588    0.01448    1258.8196
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627   -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093    0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133    0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556   -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008  825875.6875
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986   -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129    0.00171     788.0004
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890   -0.00911     173.2806
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384    0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444   -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166    0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466   -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542   -0.00722    1373.4147
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924    0.02606     564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011    0.00739     789.7573
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442   -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124   -0.00456    1531.4274
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051    0.02190     264.6404
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970    0.00200    1305.6691
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257    0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526     706.1463
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125   -0.02685    2568.4631
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020    0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675   -0.01905    6487.0010
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115    0.01015     790.2421
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983    0.01253     528.1924
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683   -0.00603    1562.0503
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378    0.02382     100.0000
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165    0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759    0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909   -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298   -0.00712    1010.5097
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161    0.00171     756.8818
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253   -0.00333     422.8840
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405    0.00476   11992.5273
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527   -0.02077    8759.9961
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001    0.00759    1880.4403
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990    0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069   -0.00900     614.5898
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180    0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193   -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706   -0.03035    1471.9783
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423   -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687   -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292    0.00462     716.5762
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978    0.01282    2322.7944
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066    0.05257     549.3252
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310   -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009   -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233    0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070   -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635   -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608    0.00148   10771.1963
#> 29716 715272.5 9559992 2144.94 25.7177836  0.00262   -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537   -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750    1892.8978
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944    0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044   -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273   -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354    0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852    0.00248    1432.4019
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523     888.2811
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253   -0.01233     968.7780
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709    1181.2555
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732    0.01858     298.4490
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962   -0.00172     351.5812
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379   -0.00249     148.1857
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456   -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491