Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340    604.1032
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276   9614.6924
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166   1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367   3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400   4126.7812
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603