Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408  76318.0156
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983