Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438     753.6304
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479    0.01261     323.8441
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373   -0.00057    1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005     769.3426
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440    1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010    1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269     491.8017
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475    0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392    1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173    8967.1924
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356    0.00774     527.1068
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252    0.01118    2359.5059
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916    2009.1686
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346     543.0749
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771   -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140   -0.00070    2562.7559
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169     350.1799
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666   -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025    4425.5151
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372   -0.00008  825875.6875
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329   -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241     585.6182
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690    5121.6924
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529    0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099    0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845    0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656    3036.7383
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611    0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078    0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389   47022.9688
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469    0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058    1817.2371
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728    1137.2450
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078  622888.2500
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739     789.7573
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317    0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756     353.9086
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915   -0.02705     413.5702
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019    0.00119    6589.5796
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610     454.8103
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458    0.04291     293.4231
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744     226.7911
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893   -0.00867    2247.8748
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230    1044.6443
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306    1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840    0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655     130.9468
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665    1904.4769
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002   -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465   -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298   -0.00712    1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162   -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463    1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538     265.4677
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176    0.02973     215.5101
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987   -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052    3508.5591
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035    1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422     967.9738
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098     164.0802
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586    0.00654    1576.2065
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118    0.00162     290.2031
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487    1769.3392
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980   -0.00710     472.0236
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247     282.1076
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375    0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576    1045.1074
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972     331.7462
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278   -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011     501.7867
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181   -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419     323.0790
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775