Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491