Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949   6217.8213
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838    238.6685
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118   2359.5059
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199    196.4080
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245   6353.3179
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334   4905.2002
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685   2568.4631
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048   1550.6704
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960    884.2003
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491