Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057    1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005     769.3426
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696    4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355   14919.6260
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504    1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731    1384.7959
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357     473.4709
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332    1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055     773.0359
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102   -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759     158.9370
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200    1305.6691
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140    1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110   -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488   -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331     699.3024
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463    1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548     532.8210
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399    1601.4852
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196   -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514     897.9290
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785    0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268     252.0072
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836     250.5111
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624    1870.8307
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119     792.3544
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227     841.2025
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505    0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983