Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458    485.7418
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336