Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976     301.2347
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250    1413.3312
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090   0.00890    2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046    1024.1573
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214   0.00466     724.0713
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526     539.8119
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164     895.6606
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518     854.4269
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398     267.2659
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937   0.00083     994.7876
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346     543.0749
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338   10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407     344.8395
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917  -0.01993   10308.6406
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357     473.4709
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883     616.1993
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026     247.0894
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809    1102.9956
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224     942.0322
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870    1895.1399
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797    7557.7866
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141    4611.0718
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428  -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169   0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003   0.01196    1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756     353.9086
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044     288.5468
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358     802.7622
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382     100.0000
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505     351.4078
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158     837.3730
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419     761.5427
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479     604.0975
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445     355.5078
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734     335.0773
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697     412.1185
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035    1471.9783
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018     937.6911
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418   0.00513     161.9878
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417    1520.6630
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499     617.3855
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788  -0.01029     894.2672
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463   0.00253    1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249     148.1857
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806