Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041    772.9746
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491