Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 25090 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732  -0.00891   -0.00949    6217.8213
#> 20151 714602.5 9559922 2184.66 48.3415951  -0.00979   -0.01542     674.0147
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311     586.7354
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501    3889.7292
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 22566 712832.5 9559842 1897.49 33.0602377   0.00672    0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790   0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.5350311   0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314    2360.1501
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764   0.00225    0.00325     616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437    1155.1469
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 19128 714852.5 9557872 2681.49 20.7576880   0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313     783.4548
#> 12868 712532.5 9560332 1913.92 61.9373106  -0.03543   -0.05146    1347.6224
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096     520.5570
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178     800.1933
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452     437.3965
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 2976  715122.5 9559142 2390.05 33.1679538   0.01008    0.00102     483.4184
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373     527.3528
#> 7466  712962.5 9560392 1819.03 18.4028314  -0.00140   -0.02330    1456.9246
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166     650.6246
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684    2556.8049
#> 35083 714612.5 9558152 2596.12 60.5983082   0.04548    0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262    1760.2162
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 13146 713792.5 9558552 2341.40 29.8253180  -0.00104    0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219     420.7032
#> 17431 712802.5 9560022 1832.71 41.5251162  -0.00478   -0.07132   42970.7891
#> 21680 712802.5 9560462 1893.75 38.7342388  -0.00481   -0.00399    3900.3445
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189   0.00024   -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389   47022.9688
#> 12772 715172.5 9557642 2759.42 34.8054672  -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394     587.8840
#> 47320 713422.5 9559182 2143.03 56.3251890   0.01246    0.00534     344.1447
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239  -0.00444   -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245     629.5983
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.5275756  -0.00193    0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613  -0.00070    0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 49378 713252.5 9560732 1875.41 27.8354356   0.00181   -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.4626708  -0.02970    0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 42617 714092.5 9558392 2469.21 42.7999473  -0.01397    0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 18498 714842.5 9557712 2679.04 28.9395253   0.00902    0.01358     802.7622
#> 17227 714812.5 9558892 2545.43 54.0717460   0.00458    0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.7491217   0.00840    0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.6500280   0.01996    0.04684     209.1778
#> 9984  712742.5 9560092 1848.56 23.6127367   0.00614   -0.00093     118.1402
#> 6470  713742.5 9558802 2260.79 30.6801711   0.00000    0.00320     911.5911
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.0527893  -0.00586    0.00067    1123.7318
#> 15211 715772.5 9558082 2801.07 36.2968763  -0.00079   -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.3476599  -0.02196    0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150     503.5590
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232     977.7320
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230     714.7432
#> 25212 714272.5 9559012 2433.45 37.6714021  -0.01909   -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085    2229.5105
#> 26023 715832.5 9557632 3097.56 41.7101816   0.00133   -0.00452     578.5847
#> 42329 715852.5 9558702 2651.92 22.8707563   0.00321    0.00548     532.8210
#> 44396 715512.5 9558102 2845.31 26.5984197   0.01263    0.03598     186.3527
#> 34347 714742.5 9560142 2161.95 26.5669070  -0.00978    0.00367    3002.8828
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943     371.5514
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.6481135  -0.00844    0.00514     897.9290
#> 20319 715282.5 9560482 1901.16 27.6767263   0.00706   -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.8963648   0.00358    0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.6759998  -0.00410    0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.9465824   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645     234.0114
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257     549.3252
#> 35826 712912.5 9560302 1891.22 50.8694849  -0.00229    0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479   0.00972    0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705   0.00537   -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.5684513   0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079    1647.6451
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578    1258.8865
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.0318136   0.01600   -0.01849     764.8193
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.7737308   0.00452    0.00368    1071.8734
#> 16845 712772.5 9560022 1831.76 21.3999736  -0.00865   -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012     305.3624
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 9320  713872.5 9559262 2378.76 29.8585496   0.00309    0.01011     501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029     894.2672
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436  -0.00181   -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458     485.7418
#> 15880 715892.5 9557572 3113.36 53.0432868   0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253    1237.0536
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389