Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309