Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901