Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806