Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.44641   -0.00347   0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.12344    0.00023  -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.08548   -0.00075   0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.27793    0.01917   0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.09556    0.00753   0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.33705    0.00514  -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.81524   -0.00857   0.03677   1675.4679
#> 15843 714042.5 9558902 2332.30 50.09771   -0.00858  -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.28541    0.00118   0.00761    928.6858
#> 38411 713552.5 9560652 1851.63 46.00507   -0.00891  -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.55641   -0.00730  -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.87132   -0.00156   0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.15898   -0.00854   0.01264   5057.1431
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.15936   -0.01607   0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.25831   -0.06175  -0.08226   2325.2053
#> 42598 713542.5 9559972 2184.39 35.24149   -0.00841  -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.95713   -0.00542   0.01842    868.6137
#> 8019  714672.5 9559272 2316.84 40.38149   -0.01966  -0.02234    849.0547
#> 18843 715192.5 9559332 2265.75 41.15327   -0.07664  -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.98252   -0.02123   0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.34692    0.00528   0.01392   1255.0884
#> 22933 712712.5 9560432 1902.17 38.12347   -0.00746   0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.25255   -0.04430  -0.06070 138869.3906
#> 47022 715362.5 9560102 2059.29 49.11967    0.02059  -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.24491    0.00051   0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.70789   -0.00077  -0.00284   2010.4285
#> 40828 714842.5 9557632 2698.65 17.53652    0.00490  -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.25769   -0.02651  -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.54285    0.03206   0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.02912   -0.00521  -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.28291    0.00356   0.00774    527.1068
#> 34565 712972.5 9560172 2007.72 27.94716    0.00598   0.01212    561.1948
#> 1635  713102.5 9559052 2172.87 31.48403   -0.00064  -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.37689    0.01275   0.00605    798.6775
#> 19128 714852.5 9557872 2681.49 20.75769    0.00453   0.01077    300.7915
#> 12460 712762.5 9559622 1906.41 24.26190    0.01047  -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224   0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.54269   -0.00294  -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.19298   -0.01846  -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.62035   -0.00904  -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.70215   -0.01771  -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.16722   -0.01353   0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.21397   -0.03229  -0.00511 361979.2812
#> 4605  714852.5 9557892 2677.78 26.93360    0.00321   0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.83316   -0.00442  -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.29918    0.00066  -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.85454   -0.01174   0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.64258    0.01843   0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.59981   -0.03552  -0.02058    431.4362
#> 23029 713502.5 9559102 2174.07 56.45754    0.03395   0.01495    266.5888
#> 4031  715392.5 9560162 1998.16 46.28582    0.00708  -0.02538    807.8237
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762   0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.36659   -0.01017   0.00197    711.1091
#> 6317  715042.5 9557702 2745.36 27.87382    0.00090  -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209   0.02739    365.7626
#> 37416 713282.5 9560692 1850.61 30.25905    0.00736   0.01684   2556.8049
#> 23213 715592.5 9558562 2654.69 33.21952   -0.00034   0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.01040    0.00448  -0.00328    592.3326
#> 10364 714762.5 9561002 1856.21 25.27489   -0.00845   0.00454   3476.0818
#> 21013 712742.5 9560032 1845.53 29.35263    0.00568  -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.52607    0.01927  -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104   0.01664   1898.5588
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238  -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.40821    0.00291  -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.47468    0.00629  -0.00469   1021.5407
#> 18897 712802.5 9559882 1851.71 50.23465   -0.02627  -0.05393   8301.6094
#> 12772 715172.5 9557642 2759.42 34.80547   -0.00567  -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.15385   -0.01556  -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.34897   -0.10372  -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.12936    0.00370  -0.00190    640.8842
#> 17345 713922.5 9559682 2227.44 34.49091   -0.06509  -0.01211  49347.5547
#> 47702 714692.5 9559712 2200.51 33.09863   -0.13038  -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.22558   -0.02084  -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.86005   -0.02143  -0.05078 622888.2500
#> 43245 713722.5 9560012 2153.38 34.19126   -0.01263   0.01322   4605.1411
#> 26428 713042.5 9560332 1895.15 33.95978   -0.01439  -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511  -0.01801   1071.6141
#> 37439 714602.5 9560672 1982.99 18.90646   -0.00070   0.00010   3467.2690
#> 19756 715572.5 9558972 2485.95 38.81273    0.00546   0.01484    495.0292
#> 22297 714642.5 9558172 2606.18 59.46901   -0.00495  -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.32438   -0.00296   0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.32798   -0.00292  -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.89750    0.00401  -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.56633   -0.00693   0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.87651    0.00010   0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.65787   -0.01341   0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.72045    0.01267   0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.80875   -0.01881   0.00121   1783.0658
#> 28260 715312.5 9558302 2681.75 34.48404   -0.01125  -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.04708    0.00216  -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.49759   -0.00622   0.00152   1530.5657
#> 13278 712752.5 9560452 1901.25 42.11354   -0.00118  -0.00252   3631.7017
#> 46971 713532.5 9559952 2198.03 29.72390   -0.00529   0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.98938   -0.00983   0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.38149   -0.01966  -0.02234    849.0547
#> 10785 714052.5 9557852 2377.76 20.98401   -0.00025  -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.61274    0.00614  -0.00093    118.1402
#> 31983 714302.5 9559912 2204.77 23.46491   -0.01332   0.01842    246.5161
#> 748   715102.5 9559692 2301.39 27.02184   -0.00759   0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.75982   -0.02284   0.01085   2229.5105
#> 14694 715032.5 9559482 2307.44 44.17333   -0.00007  -0.00463   1050.7473
#> 26963 715702.5 9557882 2880.32 43.73330   -0.01174  -0.01466   2633.9666
#> 43426 713782.5 9560822 1876.35 23.82301   -0.00345  -0.00075   8141.7549
#> 13512 715152.5 9557992 2778.81 43.38952   -0.00724   0.01464   4029.5947
#> 39796 714802.5 9560372 2046.48 40.15575    0.00176   0.02973    215.5101
#> 16878 714232.5 9558172 2485.33 22.80830    0.00818   0.00502   1428.8546
#> 45071 712652.5 9560032 1900.80 14.47349    0.00987  -0.00507    504.3085
#> 40139 712852.5 9560072 1907.89 43.70407    0.00001   0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.21700    0.00973  -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.02323    0.01338  -0.00399   1601.4852
#> 34616 714522.5 9558742 2568.77 34.99798    0.00990   0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.48633   -0.00069  -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.27967    0.00084  -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.33999    0.01432  -0.00362    283.7090
#> 37029 713022.5 9559082 2163.08 36.03217    0.00490  -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.61198   -0.03193  -0.00267  31588.1953
#> 48869 712562.5 9560122 1972.89 24.94658    0.00185   0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224  -0.02036    476.9504
#> 29107 713172.5 9559272 2067.58 28.26401    0.02413   0.00528    355.3475
#> 5656  713132.5 9560142 2042.24 33.59538   -0.00035   0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.93366   -0.00342   0.02552   2515.0752
#> 44823 714202.5 9557442 2548.13 43.23024   -0.00233   0.01282    561.2366
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 29716 715272.5 9559992 2144.94 25.71778    0.00262  -0.00502    520.5218
#> 40612 713542.5 9561152 1856.15 33.56845    0.00420   0.01960    884.2003
#> 48184 714202.5 9558222 2489.17 36.11353    0.02236  -0.00066    306.0146
#> 7986  712752.5 9560242 1857.41 39.83031    0.01852   0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.33640    0.00537  -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.01744   -0.02552   0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.76688   -0.00375   0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.00410    0.00745   0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.72180   -0.00598   0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.48114    0.01129   0.00491    261.7371
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909  -0.01151 143213.1562
#> 20950 714912.5 9557782 2667.71 30.81138   -0.04770   0.00270  35950.5508
#> 18504 713372.5 9560672 1786.89  6.47385   -0.00278  -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.00219    0.00203   0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.53426   -0.00163  -0.00377   1727.6055
#> 46148 715082.5 9558042 2725.39 42.83604   -0.00181  -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.89051   -0.06986  -0.05013    616.4128
#> 32342 714862.5 9559622 2294.74 36.96724    0.01021   0.00109    915.4989
#> 23808 715172.5 9558702 2533.50 22.34593    0.00811  -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.04329    0.00962  -0.00172    351.5812
#> 6206  714122.5 9560992 1917.28 37.96991   -0.01023  -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.58335   -0.00247   0.01057    722.3952
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456  -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.29741    0.00505   0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983