Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976     301.2347
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767    3321.6113
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046    1024.1573
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 10974 715312.5 9557502 2844.23 26.6648828 -0.00831 -0.00149   12632.7793
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314    2360.1501
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150    1274.3594
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118    2359.5059
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 34642 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887     366.8537
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373     527.3528
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588     749.2502
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121    4159.6738
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081     940.6002
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474    2343.2100
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801    1071.6141
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301     338.9676
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122 -0.00345 -0.00075    8141.7549
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548     532.8210
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014     271.4219
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223    1056.3701
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011     201.1272
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282    2322.7944
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.9332227  0.03418  0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312     254.0656
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964     543.0025
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849     764.8193
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368    1071.8734
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525     162.0592
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607