Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491