Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957  -0.00292    0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312   2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.950663   0.03743   -0.05242    507.9263
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583   1898.0105
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261   3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867   2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075   8141.7549
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033   3446.5520
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714 480535.5000
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906   300.00   TRUE       300.00            100  3.120164
#> 28223 29.02260    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161   300.00   TRUE       300.00             26  2.745084
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 48187 30.96665   279.54   TRUE        21.35             46  3.521349
#> 15843 42.22183   300.00   TRUE       300.00            100  3.091290
#> 44952 34.08239   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11027   300.00   TRUE       300.00            100  3.036974
#> 43669 28.47658    24.22   TRUE         0.00             89  2.798316
#> 38827 31.03598   300.00   TRUE        52.52             39  3.384015
#> 23817 21.61140   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220   300.00   TRUE       300.00              1  2.710095
#> 39629 47.54862   300.00   TRUE       300.00             85  3.366461
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 44328 25.50407   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915   300.00   TRUE       300.00             34  2.458702
#> 30800 33.24817   300.00   TRUE       300.00            100  4.403435
#> 43964 28.75732   300.00   TRUE       300.00             46  2.952143
#> 2107  35.42369   300.00   TRUE       300.00             45  2.661523
#> 30125 40.73558   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863   111.09   TRUE         0.00             25  3.424505
#> 22566 24.58676   300.00   TRUE         7.49              0  2.426944
#> 36030 29.19965   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 32821 29.48326   300.00   TRUE       300.00             65  3.372940
#> 18391 34.50753   300.00   TRUE       300.00            100  5.556760
#> 13585 47.06734   300.00   TRUE       300.00            100  2.769655
#> 38077 26.91527   300.00   TRUE       300.00            100  2.777281
#> 27957 30.06023   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83427   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090   300.00   TRUE       300.00            100  3.062637
#> 3466  26.79494   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 2950  28.20385   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39122   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862   300.00   TRUE       300.00              2  2.460907
#> 12868 37.58832   119.34   TRUE        79.16             61  3.129568
#> 47429 32.09767   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603   300.00   TRUE       294.31            100  3.287475
#> 20001 34.67712   264.46   TRUE        18.17             56  3.311398
#> 34563 38.15670   300.00   TRUE       300.00            100  4.204981
#> 6310  23.82645   300.00   TRUE       300.00            100  4.032926
#> 43092 39.47851   300.00   TRUE       300.00            100  2.903195
#> 5578  25.74185   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942   300.00   TRUE        53.95             43  3.288564
#> 47884 27.05449   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306   300.00   TRUE       300.00             71  4.272628
#> 651   45.52436   300.00   TRUE       300.00            100  3.072673
#> 26567 40.43707   300.00   TRUE       300.00            100  3.023906
#> 2976  26.88203   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70251   165.13   TRUE        39.56            100  3.473052
#> 4031  42.76156   300.00   TRUE       300.00             31  2.907317
#> 5084  31.59518   300.00   TRUE       300.00             18  2.722101
#> 39520 35.89409   300.00   TRUE       300.00            100  2.934945
#> 13029 34.18782   300.00   TRUE        96.65              4  2.813330
#> 30886 39.18516   300.00   TRUE       300.00            100  2.695380
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 46065 29.77089   300.00   TRUE       251.11            100  3.482407
#> 35014 32.53025   300.00   TRUE       300.00             57  2.826223
#> 34192 33.87498   273.94   TRUE         4.48             85  3.009256
#> 40270 27.59766   300.00   TRUE       300.00              5  3.139431
#> 21680 28.34823   104.12   TRUE       110.76              4  3.591103
#> 47366 28.72753   300.00   TRUE       300.00            100  2.872349
#> 39722 26.54170   300.00   TRUE       300.00            100  2.906387
#> 30268 47.27246   300.00   TRUE       300.00             41  2.705801
#> 16009 33.41318   300.00   TRUE       300.00            100  3.049449
#> 24135 38.66835   300.00   TRUE       122.87            100  3.878395
#> 4588  31.66451   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527   130.93   TRUE         0.00             21  3.399421
#> 19832 36.63378   300.00   TRUE       300.00            100  3.278299
#> 5572  26.59785   300.00   TRUE       300.00            100  2.567470
#> 26428 35.47525   300.00   TRUE         0.00            100  3.486483
#> 14545 39.31407   300.00   TRUE       300.00             37  2.897494
#> 47086 23.38699   300.00   TRUE       300.00            100  3.525504
#> 37439 26.59269   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09350   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79821    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765   300.00   TRUE       300.00            100  3.662232
#> 48650 27.05220   300.00   TRUE       300.00            100  3.027918
#> 10456 21.66869    24.34   TRUE         0.00             18  2.776185
#> 28588 31.62555   300.00   TRUE       300.00             78  5.209992
#> 11057 31.22505   300.00   TRUE       300.00            100  2.792408
#> 10213 31.59862   300.00   TRUE       300.00            100  3.351772
#> 43834 28.51210    20.21   TRUE       150.14              2  3.331390
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 36765 36.18515   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09327   300.00   TRUE       300.00              2  2.959800
#> 23176 41.55835   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77564   300.00  FALSE       300.00             65  2.391845
#> 11288 28.53731   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860    60.00  FALSE         0.00            100  3.149477
#> 15558 50.34408   300.00  FALSE       300.00             45  2.657927
#> 11111 33.38396   300.00  FALSE       300.00            100  3.344795
#> 20067 29.82360   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030   300.00  FALSE       300.00            100  2.781107
#> 43426 35.46265   300.00  FALSE        82.11            100  3.910718
#> 28162 18.66983   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93477   300.00  FALSE       300.00            100  3.605261
#> 49835 32.36123   300.00  FALSE         0.00            100  3.134665
#> 40139 33.14332   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27049   300.00  FALSE       154.49            100  2.570019
#> 12974 37.33851   300.00  FALSE       300.00            100  2.711947
#> 21925 31.24855   300.00  FALSE       300.00             95  3.537385
#> 31201 33.72258   300.00  FALSE       300.00            100  2.404000
#> 29527 36.72659   300.00  FALSE       300.00            100  2.639632
#> 48869 24.87554   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74869   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96018   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10013   300.00  FALSE       300.00            100  3.055767
#> 14148 21.87782   300.00  FALSE       214.75            100  2.760532
#> 13908 34.37747   300.00  FALSE       300.00             96  5.681725
#> 20764 39.85036   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20388   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60720   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 1281  35.39389   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61256   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204   300.00  FALSE       205.59            100  2.626915
#> 44434 36.31979   300.00  FALSE       300.00            100  3.451290
#> 974   22.75846    80.00  FALSE        90.05            100  3.099987
#> 44216 26.06500   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58395   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02070   300.00  FALSE       300.00             55  4.224567
#> 4943  37.77167   300.00  FALSE       300.00            100  3.019161
#> 15138 20.64596   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36121   300.00  FALSE        72.23            100  2.520806
#> 16845 22.88279   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92962   300.00  FALSE       300.00             63  4.635980
#> 18504 23.09593   144.96  FALSE         4.48             89  5.049117
#> 22628 29.12115   300.00  FALSE       300.00            100  2.951467
#> 32342 37.61984   300.00  FALSE       300.00             81  2.961658
#> 15880 48.89679   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39465   300.00  FALSE       300.00            100  3.092389
#> 32618 39.38283   300.00  FALSE       300.00             56  3.111762
#> 15363 26.94219   300.00  FALSE       300.00            100  2.858775
#> 24516 36.25448   300.00  FALSE       122.13            100  3.751491