Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579   1318.7560
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426 -0.00367    793.9665
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178    800.1933
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279    930.0871
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236   2715.4138
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.95066  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261   1627.8989
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257  0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158    837.3730
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 8066  715842.5 9558732 2636.38 33.71742 -0.00494  0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950 -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066    306.0146
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156    758.4840
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227    841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901