Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203    634.3320
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579   1318.7560
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665  25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991   1403.0193
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613    289.6113
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407    344.8395
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903   2360.1536
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690   5121.6924
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 3418  715372.5 9558492 2601.00 22.6272492 -0.01168 -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656   3036.7383
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809   1102.9956
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512  13178.3223
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238 -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422   1159.2100
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109   2047.9031
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921    525.3245
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251 -0.00389  14516.2002
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273  72123.6172
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714 480535.5000
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532  16771.3066
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875   2156.9358
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172 111973.8906
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884    672.2466
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172    351.5812
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983