Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491