Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903   2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491