Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970    1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769