Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603