Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034  -0.01919   -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.9766069   0.00958    0.02642    671.1807
#> 36052 714842.5 9558892 2483.27 68.7898858  -0.04921   -0.12438    753.6304
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251   0.02240   -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352  -0.00498   -0.00752   1321.5997
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005    769.3426
#> 42917 714202.5 9557412 2544.08 39.8778625  -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303   0.00310    0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.3616837  -0.04951   -0.00529   3282.1186
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.5787645  -0.00068    0.00588   1295.1610
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.2583065  -0.06175   -0.08226   2325.2053
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514    297.7842
#> 48055 712862.5 9559972 1902.48 29.0168746  -0.00720    0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.1855058   0.00602    0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.9120303   0.00191   -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.0602377   0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626   4316.0977
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889    362.8363
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314   2360.1501
#> 48364 715582.5 9558222 2746.23 25.0720602   0.01257    0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.9471624   0.00598    0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764   0.00225    0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185   0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.5662016  -0.01804   -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642    613.6390
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.6943204  -0.01545    0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.1929823  -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939    703.7817
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.6513065  -0.02379    0.01338  10787.6299
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 651   714602.5 9560152 2111.23 53.8517302  -0.01149    0.01949   1182.1506
#> 41037 715532.5 9558792 2529.58 32.8465245  -0.00585   -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935   0.00290    0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.6561279  -0.00855   -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.9680266   0.00107    0.02373    527.3528
#> 7466  712962.5 9560392 1819.03 18.4028314  -0.00140   -0.02330   1456.9246
#> 32909 712812.5 9560452 1883.85 38.2512354  -0.00221    0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166    650.6246
#> 7668  713372.5 9559062 2164.82 37.9349627  -0.01807   -0.01883    616.1993
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262   1760.2162
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.2645529  -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167  -0.00539    0.00809   1102.9956
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240   1201.4724
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388  -0.00481   -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189   0.00024   -0.00474   2343.2100
#> 8136  713242.5 9560602 1798.18 41.2403562  -0.01169   -0.03220   8936.2666
#> 20368 715252.5 9560622 1865.83 33.5724620   0.00025   -0.01245   6353.3179
#> 31117 712892.5 9558922 2129.84 30.5506826   0.00521   -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.1064166   0.00516    0.01574    183.3220
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236   9755.3027
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175   1026.6818
#> 1036  713042.5 9558892 2205.38 31.5098139  -0.00428   -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.3582039   0.00540    0.00879   1200.4210
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224   1137.3363
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140   1066.3957
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 49459 714692.5 9557342 2639.74 42.0470808   0.00216   -0.02606    919.4241
#> 18498 714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.1135439  -0.00118   -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.0717460   0.00458    0.04291    293.4231
#> 41310 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.2034632  -0.00585   -0.02665   1904.4769
#> 27884 714072.5 9559202 2446.27 19.7567307  -0.00009    0.00109    605.1908
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419    761.5427
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493   2212.0525
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955  -0.01174   -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.6350750  -0.00242    0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171    756.8818
#> 20547 713732.5 9560822 1851.67 36.9683829  -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734    335.0773
#> 8831  714372.5 9559852 2200.77 48.7604272  -0.00457    0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362    283.7090
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267  31588.1953
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967   0.03423   -0.00983    270.7949
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.2880302  -0.00393    0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645    234.0114
#> 12519 713982.5 9560912 1956.98 28.8249337  -0.00452   -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.5953803  -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552   2515.0752
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513    161.9878
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209   0.00980   -0.00710    472.0236
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750   1892.8978
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 15138 713702.5 9561122 1780.68 49.2216583   0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523    888.2811
#> 1577  713322.5 9559742 2027.47 29.5674870  -0.03284   -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784  -0.04770    0.00270  35950.5508
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.0800956  -0.00172   -0.00119    792.3544
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064   1485.5201
#> 32342 714862.5 9559622 2294.74 36.9672369   0.01021    0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.4500377  -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234   1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253   1237.0536
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.6210169  -0.02110   -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769