Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438     753.6304
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661   -0.01339    1002.5877
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529    3282.1186
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284    2010.4285
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598    0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118    2359.5059
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193   -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651     419.0927
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124    2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294   -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511  361979.2812
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181   11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332    1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099    0.02169     495.8840
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238   -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220    8936.2666
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211   49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870    1895.1399
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121    1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610     454.8103
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252    3631.7017
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464    4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216    1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844    0.00514     897.9290
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400    4126.7812
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011     201.1272
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165     805.1403
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368    1071.8734
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029     894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884     672.2466
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250