Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292    1233.9299
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408   76318.0156
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542    3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512   13178.3223
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538     265.4677
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298     436.1459
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705     968.9738
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740     909.4791
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368    1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491