Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958  0.02642    671.1807
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741    746.4155
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292   1233.9299
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730 -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264   5057.1431
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150   1274.3594
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154  0.00124   2114.2253
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270   4057.5637
#> 37141 714382.5 9559822 2219.36 10.10640  0.01661  0.01199    196.4080
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181   2406.1252
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643    670.2289
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245    629.5983
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776   1021.8267
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970   1443.9200
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713   6379.7974
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177    997.3089
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.44473  0.02309 -0.03409   1575.9425
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945    259.7421
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.80830  0.00818  0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367   3002.8828
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099   1146.2140
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.26401  0.02413  0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578   1258.8865
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595  0.03064   1485.5201
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135   1293.4855
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762