Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542     674.0147
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466     724.0713
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720    0.01250    3686.1240
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696    4143.4019
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773     179.3077
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257    0.00593     939.0466
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405    5681.2412
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846   -0.00034    4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749    0.04779    2972.0193
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558    1504.9717
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793     864.9472
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524   10861.9482
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245    6353.3179
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661    2508.5374
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522    1355.7436
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739     789.7573
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016   -0.00294     734.9908
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152     850.4633
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177     997.3089
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328    1066.7134
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500     538.8209
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065    0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069   -0.00900     614.5898
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997     230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948    1222.5433
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262   -0.00502     520.5218
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044   -0.00964     543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578    1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585    5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858     298.4490
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775