Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393    2081.0056
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023   -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487    1077.7585
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339    1002.5877
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090    0.00890    2301.0454
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440    1088.8639
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407    0.00547     519.4228
#> 31023 714712.5 9561042 1838.66 44.120042   -0.00223    0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447     331.4833
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514     297.7842
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392    1255.0884
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501    3889.7292
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191   -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430   -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126   -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203   10263.2734
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077   -0.00284    2010.4285
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108   -0.00678    1891.6584
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694    2123.4473
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051     779.0555
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606   -0.00556     660.3690
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752   -0.00517    1000.7119
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008    0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260   -0.03310    2833.9001
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708    0.00408   76318.0156
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427     380.9472
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.860876    0.00282    0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620     379.0855
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078    0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132   42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291   -0.00081     940.6002
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601   -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133    0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169   -0.03220    8936.2666
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246    0.00534     344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337     486.7551
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281    1178.4420
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661    2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017     759.9937
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476    1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193    0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739     789.7573
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380     601.0843
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778    4594.4351
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401   -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325     595.0426
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000  162178.1562
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186    0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306    1718.3652
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655     130.9468
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519    0.00609    4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849    0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534    0.01476    1275.9446
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831     454.9113
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812   -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825    0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195   -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650    1133.1501
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548     532.8210
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077    8759.9961
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052    3508.5591
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399    1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900     614.5898
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514     897.9290
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035    1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268     252.0072
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118    0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418    0.00513     161.9878
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204     618.2027
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105     868.7802
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273   -0.01047    2826.7700
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907    1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347    1893.3645
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148     481.8292
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012     305.3624
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595    0.03064    1485.5201
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109     915.4989
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419     323.0790
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607