Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659    3553.7166
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483     566.7674
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433    1475.3191
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340     604.1032
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679     550.0344
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542     674.0147
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783     377.8918
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010    1527.1028
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766  140591.4062
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769    1032.7177
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.7212213 -0.00491  0.00581    1855.1630
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866   29195.6016
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066 -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230    1056.5876
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332    1418.6622
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401    1722.6481
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009    1247.7061
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328     592.3326
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.4901167 -0.00539  0.00809    1102.9956
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201     788.0621
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.7916671 -0.04376 -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870    1895.1399
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833     369.3771
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713    6379.7974
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000  162178.1562
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144  317260.6875
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559     866.5775
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505     351.4078
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463     851.6306
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540  0.00650    1133.1501
#> 39796 714802.5 9560372 2046.48 40.1557471  0.00176  0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502    1428.8546
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033    3446.5520
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400    4126.7812
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413  0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705     968.9738
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417    1520.6630
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875    2156.9358
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566   12862.4922
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886   43249.3945
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119     792.3544
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491