Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474   2343.2100
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010   3467.2690
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607