Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034    4495.6929
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215     917.5296
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491