Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487   1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035   2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838    238.6685
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879   1200.4210
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573   1490.0312
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135   1293.4855
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603