Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183   -0.09203     634.3320
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555     430.4604
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292    1233.9299
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884    0.01345    1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550   -0.00120     512.9740
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952    0.01378     440.9104
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664   -0.05766  140591.4062
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312    2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973     514.5488
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993   -0.00413     364.9788
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917     116.1986
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088    2322.1746
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077     300.7915
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346     543.0749
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511  361979.2812
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178     800.1933
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613    0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585   -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283    0.00117    1018.2141
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560     977.4719
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330    1456.9246
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306   -0.00436    2602.7390
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542    3614.7334
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554    0.03304    1569.7313
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845    0.00454    3476.0818
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132   42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771   -0.00599    1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228    2262.3796
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213     328.7922
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521   -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911     173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212   46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236    9755.3027
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583    1898.0105
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924    0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181   -0.00001    1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133    2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171    4425.5410
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185     953.8614
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477    2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152    1530.5657
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279   39495.1797
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870    2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427    0.02403     474.8613
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002   -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534    0.01476    1275.9446
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181    1286.3949
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452     578.5847
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197    0.00463     851.6306
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409    1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937   -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476   11992.5273
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457    0.01697     412.1185
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483    1097.3082
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298     436.1459
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423   -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687   -0.00223    1056.3701
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782   -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032     994.9224
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487    1769.3392
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251     168.1733
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555    1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105     868.7802
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420    0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066     306.0146
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757   -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532   16771.3066
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624    1870.8307
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208     842.4825
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148     481.8292
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196   -0.01935     903.6397
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340    0.01809    1158.0426
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983