Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177    291.5937
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212  46310.0195
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334   4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200   1305.6691
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477   2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367   3002.8828
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836    250.5111
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066    306.0146
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250