Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061   5744.7583
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250