Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642    671.1807
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475   1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041    772.9746
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057   1630.3259
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431   4079.0366
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035   2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406   2421.1145
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730    0.01770    474.5059
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010   1527.1028
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665  25318.3555
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720    0.01250   3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261    554.9243
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581   1855.1630
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898    588.3761
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276   9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416   1204.8629
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294   -0.00096    520.5570
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167   1938.5389
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679    550.0344
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907   1943.4092
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102    483.4184
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436   2602.7390
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809   1102.9956
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335    195.0293
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849   1120.5961
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263    0.01322   4605.1411
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422   1159.2100
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110   -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325    595.0426
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610    454.8103
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020    0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856   1789.1596
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036   2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486    0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427    0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345   2417.1389
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150    503.5590
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226    0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538    265.4677
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445    355.5078
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923    811.5580
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431    315.9084
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367   3002.8828
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216   1412.2295
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410    0.00400   4126.7812
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282   2322.7944
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251    168.1733
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360   4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577    423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066    306.0146
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576   1045.1074
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709   1181.2555
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052    451.7209
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288    0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769