Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175    725.1355
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491