Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041    772.9746
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345   1120.6929
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547    519.4228
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214   0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250   3686.1240
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367    793.9665
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276   9614.6924
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294  -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319   3457.7519
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454   3476.0818
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228   2262.3796
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753   1337.7854
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391   1794.7964
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216  -0.02606    919.4241
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306   1718.3652
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427    866.9398
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433  -0.02177    997.3089
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565   1410.8372
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028   2394.7261
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322  -0.00759   4131.5356
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268    252.0072
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487   1769.3392
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599  12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836    250.5111
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566  12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972    331.7462
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769