Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.600942   0.00214    0.00466    724.0713
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367    793.9665
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118   2359.5059
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276   9614.6924
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009   1247.7061
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822    472.8748
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484    495.0292
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 14779 712502.5 9560202 2004.61 27.522601   0.00115    0.01015    790.2421
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 21925 714542.5 9559232 2314.33 50.247826   0.00737    0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035   1471.9783
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603