Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501   -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051  351155.1250
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056   -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347    0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221    0.00969     368.5606
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041    0.01741     746.4155
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857    0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273   -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118    0.00761     928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104   -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310    0.01440    1088.8639
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979   -0.01542     674.0147
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223    0.01783     377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102    0.00388    1051.4222
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678   -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528    0.01392    1255.0884
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412    0.00898     458.6938
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070   -0.00530    9000.4356
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746    0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382    0.00672    1114.8801
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108   -0.00678    1891.6584
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033   -0.03917     116.1986
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422    0.02357     275.7474
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097    0.00156     816.6036
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765    0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440   -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545    0.02405    5681.2412
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047   -0.00887     366.8537
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255   -0.00365    1536.2647
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140   -0.00070    2562.7559
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993    0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379    0.01338   10787.6299
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527    0.04613     289.6113
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552   -0.02058     431.4362
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290    0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480    0.02530     937.1306
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119   -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175    0.00236    2715.4138
#> 18892 712682.5 9560202 1839.08 40.9235105 -0.01364   -0.03357     473.4709
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211    0.00241     585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282    0.00542     860.8840
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458   -0.00401    1722.6481
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306   -0.00436    2602.7390
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161   -0.00409    2312.8413
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.1089251  0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131    0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081     940.6002
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604    0.00354    1147.7384
#> 18897 712802.5 9559882 1851.71 50.2346476 -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743   -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940     987.8085
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178   -0.00822     472.8748
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453    0.01337     486.7551
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195   -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129    0.00171     788.0004
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521   -0.00870    1895.1399
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849    1120.5961
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310   -0.00860     695.2302
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175    0.04195    1901.6743
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166    0.01476    1404.4454
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109    0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497  0.01011    0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.0592454  0.02213   -0.01363     565.5207
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250   -0.00970    1443.9200
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070    0.00010    3467.2690
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003    0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102   -0.01778    4594.4351
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281    0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495   -0.03756     353.9086
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774   -0.00224    1137.3363
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881    0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915   -0.02705     413.5702
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285   -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143    0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125   -0.02685    2568.4631
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252    3631.7017
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488   -0.00632     620.0229
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234     849.0547
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405   -0.05655     130.9468
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783   -0.00367    1966.2112
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251   -0.00389   14516.2002
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084   -0.02996    1599.3491
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599    0.00301     338.9676
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534    0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009    0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433   -0.02177     997.3089
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226    0.00736     477.8591
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390   -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271 -0.00007   -0.00463    1050.7473
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112    0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174   -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165    0.00875    1165.3953
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945     259.7421
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846    0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937   -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327    0.01833     300.0125
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987   -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174   -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747  0.00001    0.00759    1880.4403
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990    0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084   -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196   -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180    0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674   -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358    0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410    0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292    0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.3155672  0.04152    0.00098     164.0802
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782   -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322   -0.00759    4131.5356
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004    0.00714  480535.5000
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667    0.00863    1098.3462
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009   -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608    0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873   -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523    0.02247     282.1076
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944    0.00836     250.5111
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852    0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757   -0.00003 5426606.0000
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182   -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909   -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865   -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770    0.00270   35950.5508
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137   -0.01523     888.2811
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595    0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992   -0.01052     451.7209
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298    0.00222     991.8300
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166    0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379   -0.00249     148.1857
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806