Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345    1120.6929
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769