Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367   3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674