Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057   1630.3259
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903   2360.1536
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.695054   0.00306   -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409   2312.8413
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 38197 712522.5 9560322 1942.04 62.777267  -0.01611    0.02961   1086.6963
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109   2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363    389.2691
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014    271.4219
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81444   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111   300.00   TRUE       300.00            100  3.053301
#> 16749 32.18304   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 46178 20.93129   300.00   TRUE       300.00             31  2.440958
#> 11088 30.99874   300.00   TRUE       300.00            100  3.212274
#> 44952 34.08239   300.00   TRUE       300.00            100  3.363704
#> 27973 30.00981   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551   300.00   TRUE       300.00              7  3.384768
#> 32622 30.17138   300.00   TRUE       300.00            100  3.265863
#> 125   29.59671   207.04   TRUE        21.03            100  3.298701
#> 39629 47.54862   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 21703 22.26915   300.00   TRUE       300.00             34  2.458702
#> 23375 34.63988   300.00   TRUE       195.00              2  3.224136
#> 14522 28.39865   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668    95.53   TRUE         0.00             49  4.011286
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 26215 23.83161   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396   300.00   TRUE       300.00              0  2.559711
#> 6808  28.56596   300.00   TRUE       300.00             40  3.952656
#> 18391 34.50753   300.00   TRUE       300.00            100  5.556760
#> 47220 36.48022   300.00   TRUE       300.00            100  3.327041
#> 34280 33.62116   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38011   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83427   300.00   TRUE       300.00            100  2.440511
#> 23031 26.74109   300.00   TRUE       300.00              2  2.798938
#> 3466  26.79494   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18914   300.00   TRUE       300.00            100  3.257531
#> 19730 38.47870   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20385   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259   300.00   TRUE       300.00            100  2.894014
#> 17594 19.23190   138.35   TRUE         0.00             60  3.701999
#> 46201 22.18378   167.75   TRUE         0.00             55  3.674828
#> 31316 30.00981   300.00   TRUE       300.00             85  2.740390
#> 47884 27.05449   300.00   TRUE       300.00            100  2.905563
#> 18394 28.61065   300.00   TRUE       300.00             57  3.372940
#> 34707 27.70251   165.13   TRUE        39.56            100  3.473052
#> 21952 40.09158   300.00   TRUE       300.00             64  2.828908
#> 35194 23.44944   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039   300.00   TRUE       300.00            100  3.141386
#> 421   10.71374   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56310   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95175   300.00   TRUE       300.00              6  2.211304
#> 375   27.97409   300.00   TRUE       300.00            100  2.851936
#> 39520 35.89409   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367   300.00   TRUE       300.00            100  2.538942
#> 33350 28.95270    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16814   300.00   TRUE       300.00             25  3.364146
#> 46783 19.94237   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307   179.84   TRUE         0.00              0  3.541090
#> 38197 34.34252   105.52   TRUE        65.32             50  3.036108
#> 41783 28.18609   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526   197.29   TRUE         4.67             57  3.591568
#> 31834 24.64693   215.68   TRUE         0.00            100  3.354565
#> 37769 37.42503   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73174    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21895   300.00   TRUE       300.00              6  5.916915
#> 43859 29.55775   300.00   TRUE       300.00            100  2.806780
#> 1424  39.15422   300.00   TRUE       300.00            100  4.035908
#> 10933 34.67712   264.46   TRUE        18.17             56  3.311398
#> 17345 29.82131   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94470   300.00   TRUE       300.00            100  4.477051
#> 43201 41.82821   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475   300.00   TRUE       300.00              6  3.211627
#> 26428 35.47525   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32283   300.00   TRUE       300.00              8  2.751808
#> 14545 39.31407   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03884   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852   300.00   TRUE       300.00             60  2.848009
#> 12629 32.18075   300.00   TRUE       300.00            100  3.159543
#> 1916  29.27356   300.00   TRUE       300.00             21  3.311309
#> 7409  25.37057   300.00   TRUE       300.00            100  2.866282
#> 14494 21.18511   300.00   TRUE       300.00             28  2.535469
#> 1223  40.45827   300.00   TRUE       300.00              4  3.251167
#> 22847 24.74777   300.00   TRUE       300.00             41  3.134623
#> 49459 41.85686   300.00   TRUE       300.00            100  2.963516
#> 46038 29.73078   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555   300.00   TRUE       300.00             78  5.209992
#> 17674 30.48250   300.00   TRUE       300.00              1  3.812044
#> 11057 31.22505   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581   300.00   TRUE        65.02              1  2.976445
#> 35458 37.47316   300.00   TRUE       300.00            100  3.297274
#> 30378 28.56939   300.00   TRUE       300.00             57  3.322176
#> 41310 17.29530   300.00   TRUE       300.00             25  3.046207
#> 39466 27.25159   300.00   TRUE       300.00            100  3.060187
#> 17374 43.14544   300.00   TRUE       300.00            100  3.630971
#> 36744 29.40477   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55835   300.00   TRUE       300.00            100  3.203943
#> 16208 38.37385   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81841   300.00  FALSE       300.00            100  2.545811
#> 15186 38.54860    60.00  FALSE         0.00            100  3.149477
#> 25212 38.98290   300.00  FALSE       300.00            100  3.109374
#> 40459 36.18687   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29534   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188   300.00  FALSE       300.00             23  2.716617
#> 20067 29.82360   300.00  FALSE       268.03            100  2.590250
#> 8066  23.52737   300.00  FALSE       300.00            100  2.834598
#> 441   25.11331   294.31  FALSE       152.09            100  3.054287
#> 42329 19.22789   300.00  FALSE       300.00            100  2.726581
#> 16542 21.11235   300.00  FALSE       300.00            100  2.909320
#> 7538  17.26494   300.00  FALSE       300.00             87  2.433645
#> 16878 25.38891   300.00  FALSE       300.00             18  3.154988
#> 4007  32.09251   300.00  FALSE       300.00             68  2.847871
#> 8413  39.88989    80.73  FALSE         0.00            100  3.040329
#> 29527 36.72659   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732   300.00  FALSE       185.58            100  2.985864
#> 24009 38.49703   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74869   300.00  FALSE       300.00            100  2.427333
#> 39440 32.70615   300.00  FALSE       300.00             92  2.369237
#> 45486 33.42177   300.00  FALSE         0.00             68  3.366011
#> 7983  38.50620   300.00  FALSE        57.15             22  2.884583
#> 44715 28.45022   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60720   275.21  FALSE        93.50             70  2.791131
#> 34532 31.20672   300.00  FALSE       300.00             92  2.583601
#> 1281  35.39389   300.00  FALSE       300.00            100  3.632329
#> 5861  27.61256   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709    55.73  FALSE       125.03            100  2.946551
#> 484   32.14580   300.00  FALSE        55.63            100  3.216864
#> 6732  26.85625   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61941   300.00  FALSE       300.00             49  2.417865
#> 34334 25.08982   300.00  FALSE       300.00            100  2.986224
#> 41873 11.03860   300.00  FALSE       300.00            100  2.265064
#> 9526  27.56958   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185   300.00  FALSE       300.00             79  2.812912
#> 11570 16.60718   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61984   300.00  FALSE       300.00             81  2.961658
#> 26635 31.39465   300.00  FALSE       300.00            100  3.092389
#> 41373 31.12880   300.00  FALSE       120.90            100  2.879946
#> 30296 40.15861   300.00  FALSE       300.00             68  3.258250