Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712    3636.5303
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661   -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440    1088.8639
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922    0.01088     266.4355
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542    0.01842     868.6137
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581    0.01189     287.5425
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392    1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070   -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071    0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118    2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437    1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156     816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184    0.03924    8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543   -0.05146    1347.6224
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771   -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495     266.5888
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364   -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558    1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555    4439.0967
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017    0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246    0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986   -0.02524   10861.9482
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870    1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141    4611.0718
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970    0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397    0.00526     706.1463
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905    6487.0010
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465   -0.01036    2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105    4432.8760
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284    0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464    4029.5947
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152    0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099    1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909   -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305    0.00585    5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603