Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438    753.6304
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712   3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804    628.5153
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581   0.01189    287.5425
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475   0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070 138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025   4425.5151
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949   1182.1506
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017   0.00197    711.1091
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681  -0.01319   3457.7519
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485   2240.4185
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399   3900.3445
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267   0.00483    343.1384
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685   2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216  -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004  -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486   0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079  -0.00331    699.3024
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409   1575.9425
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367   3002.8828
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483   1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298    436.1459
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223   1056.3701
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196    576.1450
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650    766.6254
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360   4288.7329
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079   1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044  -0.00964    543.0025
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852   0.00248   1432.4019
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208    842.4825
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148    481.8292
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169    649.9975
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087  -0.00573   1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172    351.5812
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388    180.4414
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983