Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.2471896   -0.00501   -0.00649   1399.2329
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059   -0.00628    556.0121
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753    0.01487   1077.7585
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104   -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951   -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150   -0.00284    0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.8713152   -0.00156    0.00406   2421.1145
#> 26447 715632.5 9559102 2410.77 39.6567008    0.00730    0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550   -0.00120    512.9740
#> 42598 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194    0.00526    539.8119
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414   -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.8879661    0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664   -0.05766 140591.4062
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017   -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.9120303    0.00191   -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.1731174   -0.01029   -0.02312   2657.6936
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.4003790    0.00151    0.00189   2317.1819
#> 40828 714842.5 9557632 2698.65 17.5365192    0.00490   -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108   -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.1282451   -0.00033   -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257    0.00593    939.0466
#> 6763  714692.5 9561052 1830.60 48.3948802   -0.01996    0.00626   1014.7159
#> 35142 714792.5 9558862 2575.17 25.6644985    0.06765    0.00114    150.1498
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743    0.01313    783.4548
#> 47011 713262.5 9559722 2003.72 37.1929823   -0.01846   -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470    0.00463    0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.4362937    0.00888    0.00862    874.5408
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204   -0.00376   2048.3198
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379    0.01338  10787.6299
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666   -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442   -0.00178    800.1933
#> 47884 715032.5 9557712 2738.31 28.2892182    0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267   -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552   -0.02058    431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613    0.00903   2360.1536
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286   -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480    0.02530    937.1306
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855   -0.02175   1122.9120
#> 15766 715562.5 9558592 2646.16 29.0208853    0.00194    0.00516    576.0335
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762    0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.3121575   -0.00405   -0.01565   2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260   -0.03310   2833.9001
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529    0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209    0.02739    365.7626
#> 20065 713772.5 9558852 2221.21 28.0199280   -0.00568   -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169    495.8840
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195   -0.00656   3036.7383
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611    0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904   -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478   -0.07132  42970.7891
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605   -0.03594  16587.9980
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093    0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940    987.8085
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370   -0.00190    640.8842
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453    0.01337    486.7551
#> 10467 714642.5 9559652 2249.12 37.1431350   -0.00759   -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.4629780   -0.00195   -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.5724620    0.00025   -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491    0.01061   5744.7583
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236   9755.3027
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384    0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444   -0.00017    759.9937
#> 15981 713302.5 9559172 2106.91 35.9977924    0.01354    0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.8102886    0.00542   -0.00722   1373.4147
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193    0.02812   1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109    0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.0457041   -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253    0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070    0.00010   3467.2690
#> 49072 715472.5 9558902 2482.92 30.6893384   -0.01102   -0.01778   4594.4351
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705   -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292   -0.00109   2047.9031
#> 1997  714752.5 9561012 1853.19 26.5663341   -0.00693    0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016   -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106   -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678   -0.01341    0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110   -0.02450   2589.9910
#> 42617 714092.5 9558392 2469.21 42.7999473   -0.01397    0.00526    706.1463
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157   -0.00152    850.4633
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622    0.00152   1530.5657
#> 17227 714812.5 9558892 2545.43 54.0717460    0.00458    0.04291    293.4231
#> 4479  713982.5 9557812 2399.24 41.2311889    0.01895    0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893   -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.5226006    0.00115    0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.3236027    0.00758   -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186    0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476   1404.4454
#> 9984  712742.5 9560092 1848.56 23.6127367    0.00614   -0.00093    118.1402
#> 6470  713742.5 9558802 2260.79 30.6801711    0.00000    0.00320    911.5911
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084   -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332    0.01842    246.5161
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325   -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009    0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465   -0.00105   4432.8760
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909   -0.01181   1286.3949
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124    0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738    0.00118   2273.4758
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.6402737    0.00363   -0.00363    389.2691
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.9322444    0.03480    0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195   -0.00995   1026.7981
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306   -0.00626    450.7346
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544   -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263    0.03598    186.3527
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527   -0.02077   8759.9961
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844    0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452   -0.01298    436.1459
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.5351636  197.75496 -197.75366    267.5057
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379   -0.01018    937.6911
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.9336562   -0.00342    0.02552   2515.0752
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270   -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229    0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209    0.00980   -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.9676002   -0.00820   -0.00260   1484.6593
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074   -0.02904    608.3561
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343    0.01177   1558.2231
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745    0.00875   2156.9358
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607   -0.04566  12862.4922
#> 11394 713472.5 9561232 1889.48 34.3144424    0.00108   -0.00948   1538.1741
#> 16443 714472.5 9559532 2222.55 16.7441186    0.03263   -0.02012    305.3624
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788   -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595    0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.3079371    0.01549   -0.00169    649.9975
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021    0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716   -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.1106510   -0.00288    0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.2293858    0.00366    0.00525    162.0592
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674