Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284  0.01804    628.5153
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.1532666 -0.07664 -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.8121470  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939    703.7817
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107  0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330   1456.9246
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542   3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473   2930.2791
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316    454.9674
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395   2011.3270
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240   1201.4724
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399   3900.3445
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601 -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 27244 714162.5 9558622 2316.00 36.7512319 -0.02556 -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878 -0.04351   1442.9247
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 26101 715562.5 9558682 2594.87 32.4110766  0.00079  0.00321   1170.5904
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.8299297 -0.03669 -0.01981   3975.3608
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 20171 715732.5 9557962 2815.53 48.6578678 -0.01341  0.00171   4425.5410
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166   1813.7841
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208    951.3401
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399   1601.4852
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216   1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900    614.5898
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983    270.7949
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487   1769.3392
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972  0.00308    422.5022
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499    617.3855
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260   1484.6593
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 18308 714842.5 9558742 2568.20 26.0174405 -0.02552  0.00532  16771.3066
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129  0.00491    261.7371
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9526  714432.5 9560582 2076.13 31.0800956 -0.00172 -0.00119    792.3544
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169    649.9975
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573   1490.0312
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020    673.2861
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901