Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438    753.6304
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712   3636.5303
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896    276.0310
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677   1675.4679
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008  0.00730  0.01770    474.5059
#> 31023 714712.5 9561042 1838.66 44.1200421 -0.00223  0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447    331.4833
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385    674.0721
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950    950.3138
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051    779.0555
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178    800.1933
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175   1122.9120
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588    749.2502
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602    345.8933
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009   1247.7061
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294    739.5768
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972  0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.3058340 -0.00131  0.00201    788.0621
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594  16587.9980
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058   1817.2371
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241   1752.9242
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728   1137.2450
#> 43245 713722.5 9560012 2153.38 34.1912564 -0.01263  0.01322   4605.1411
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.0256786  0.04250 -0.00970   1443.9200
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212  0.02732   4237.6934
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744    226.7911
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320    911.5911
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332  0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565   1410.8372
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230    714.7432
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875   1165.3953
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544 -0.01014    271.4219
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790    209.6952
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.6759998 -0.00410  0.00400   4126.7812
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099   1146.2140
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868    235.8511
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 38362 714402.5 9558042 2499.31 56.9010116  0.03044 -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177   1558.2231
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186   2751.4626
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803   1058.5630
#> 32342 714862.5 9559622 2294.74 36.9672369  0.01021  0.00109    915.4989
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172    351.5812
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.2293858  0.00366  0.00525    162.0592
#> 39234 712902.5 9559652 1953.39 29.7319259  0.07379 -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491