Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123  0.00393   2081.0056 37.592907
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606 20.478085
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604 46.301865
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075  0.00475   1001.0861 39.335272
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917  0.00333    258.6537 20.192179
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113 30.966650
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191 23.932447
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454 24.807927
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340    604.1032  6.084812
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046   1024.1573 38.803567
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028 20.068993
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730 -0.01491   1844.4353 30.171385
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974 23.540544
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012 29.596708
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740 23.412201
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028 20.068993
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104 24.774122
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138 25.504070
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833 25.321297
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930 23.454600
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177 21.398255
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938 35.423689
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243 24.905902
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019 34.875368
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363 21.073961
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584 40.084700
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473 36.480223
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746 33.621163
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983 26.915265
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474 18.834269
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059 29.322261
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224 37.588323
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193 27.419468
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937 32.097669
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255 -0.00365   1536.2647 23.193905
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836 35.319410
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690 18.763795
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000 19.231901
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408 27.955757
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064 31.188958
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799 17.635641
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092 34.789424
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536 28.610647
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184 26.882034
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237 42.761559
#> 49569 713232.5 9561122 1868.73 37.31216 -0.00405 -0.01565   2232.6133 41.849410
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091 27.974091
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330   1456.9246 28.790556
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602    345.8933 41.733673
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519 30.439529
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481 31.445643
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706 27.323211
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558 43.540782
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185 29.228869
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326 19.942369
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941   1290.7709 37.362578
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855 33.985565
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708 34.956155
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063 15.648623
#> 5239  714832.5 9560992 1859.70 19.96013 -0.00320  0.00180   3904.5254 24.665260
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315 28.727531
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659 26.364080
#> 43859 714892.5 9557602 2722.70 28.12936  0.00370 -0.00190    640.8842 29.557747
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293 19.325866
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004 25.761328
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583 32.323414
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609 37.775107
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027 39.558725
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718 32.056989
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002 42.773591
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743 27.569583
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818 34.189538
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.597847
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903 39.713424
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621 27.998156
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274 23.883746
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608 45.093497
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370 29.621345
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363 27.600523
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562 33.408596
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908 25.370571
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241 41.856859
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934 34.909745
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384    947.2075 33.115815
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421 16.521238
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443 16.784226
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870   2099.7930 28.569395
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278 27.558697
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438 23.345738
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503 25.681687
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585 -0.02665   1904.4769 36.185150
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487 27.251592
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457 43.145441
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161 19.775638
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301    338.9676 15.762069
#> 30809 712532.5 9560032 1928.24 40.89945  0.00570  0.02230    714.7432 36.366204
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949 38.982902
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097 35.046109
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758 36.362194
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401 50.856307
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328   1066.7134 39.136455
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538    265.4677 21.429767
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975 25.270304
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075   8141.7549 35.462650
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 9665  713362.5 9558812 2275.28 31.49492  0.01306 -0.00626    450.7346 34.081249
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321 32.361229
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840 23.945052
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101 34.567690
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604 47.873489
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961 28.075505
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852 30.941440
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082 39.889895
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520 31.248545
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451 28.893116
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953 32.227157
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783 32.072458
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459 36.726595
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554 24.875536
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450 21.877820
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911 44.395062
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065 36.122124
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031 39.850361
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511 13.572224
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179 21.777553
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022 26.907817
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417   1520.6630 37.882251
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218 32.693545
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578   1258.8865 22.758457
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645 34.657644
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624   1870.8307 32.099960
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074 37.771670
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491    261.7371 18.619409
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417 22.882788
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811 26.856251
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975 30.881852
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209 24.035007
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490 17.044348
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606 38.910710
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989 37.619836
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312 22.237065
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130 45.213246
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414 16.978458
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492    772.2691 33.526625
#>       distroad slides distdeforest distslidespast log.carea
#> 15412   300.00   TRUE       300.00            100  3.318273
#> 42632   300.00   TRUE       300.00             10  2.566509
#> 15762   300.00   TRUE       300.00            100  2.633933
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 30059   300.00   TRUE       300.00             41  2.412719
#> 48187   279.54   TRUE        21.35             46  3.521349
#> 34512   300.00   TRUE       195.00              2  3.224136
#> 21939   300.00   TRUE       300.00            100  3.168886
#> 20799   180.67   TRUE         0.00             16  3.361925
#> 39885   210.57   TRUE         0.00            100  2.781111
#> 42917   300.00   TRUE       300.00            100  3.010367
#> 14456   300.00   TRUE       247.02            100  3.183868
#> 32622   300.00   TRUE       300.00            100  3.265863
#> 49372   213.54   TRUE        20.21             68  3.430977
#> 125     207.04   TRUE        21.03            100  3.298701
#> 1393    300.00   TRUE       300.00              1  2.710095
#> 42598   300.00   TRUE       247.02            100  3.183868
#> 10899   300.00   TRUE       300.00             20  2.644350
#> 44328   300.00   TRUE       300.00            100  3.122417
#> 2990    300.00   TRUE       123.56              5  2.520462
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 2107    300.00   TRUE       300.00             45  2.661523
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 25081   300.00   TRUE       300.00            100  3.617357
#> 4827    300.00   TRUE       300.00              0  2.559711
#> 3453    300.00   TRUE       300.00              2  3.276843
#> 47220   300.00   TRUE       300.00            100  3.327041
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 38077   300.00   TRUE       300.00            100  2.777281
#> 3133    300.00   TRUE       300.00            100  2.440511
#> 15390    60.00   TRUE       118.92              2  3.372821
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 6557    300.00   TRUE       300.00              5  3.147064
#> 47429   300.00   TRUE       300.00            100  2.464778
#> 11687   300.00   TRUE       300.00             68  3.186466
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 17594   138.35   TRUE         0.00             60  3.701999
#> 19784   300.00   TRUE       300.00            100  2.941780
#> 38686   300.00   TRUE       300.00             75  2.946946
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 13699   300.00   TRUE        53.95             43  3.288564
#> 18394   300.00   TRUE       300.00             57  3.372940
#> 2976    300.00   TRUE       300.00            100  2.684323
#> 4031    300.00   TRUE       300.00             31  2.907317
#> 49569    25.00   TRUE        35.00            100  3.348814
#> 375     300.00   TRUE       300.00            100  2.851936
#> 7466    255.23   TRUE         0.00             60  3.163437
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 2595    300.00   TRUE       300.00            100  3.538794
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 19966   300.00   TRUE       300.00             23  3.350329
#> 46783   256.35   TRUE         0.00            100  2.772566
#> 44650   300.00   TRUE       300.00            100  3.110849
#> 664     300.00   TRUE       300.00             12  2.885027
#> 11618   300.00   TRUE       300.00              1  2.578737
#> 45500   300.00   TRUE         0.00              2  3.626296
#> 45592   300.00   TRUE       300.00              2  2.460907
#> 5239    197.29   TRUE         4.67             57  3.591568
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 40216   300.00   TRUE       300.00            100  3.257022
#> 43859   300.00   TRUE       300.00            100  2.806780
#> 1973    300.00   TRUE       300.00            100  2.290100
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 47718   300.00   TRUE        10.00              1  3.759272
#> 27244   300.00   TRUE       300.00             54  5.009882
#> 42201   300.00   TRUE       300.00            100  3.989241
#> 20797   285.07   TRUE         0.00             96  3.663802
#> 37720   300.00   TRUE       300.00            100  3.690657
#> 33875   300.00   TRUE         4.67              0  3.279136
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 34164   300.00   TRUE       300.00              1  2.561448
#> 11352   135.89   TRUE        82.11              4  4.291317
#> 28173   300.00   TRUE       300.00             59  3.185096
#> 10363   300.00   TRUE       297.43              6  3.599377
#> 16071   300.00   TRUE       300.00            100  2.201225
#> 23783   300.00   TRUE         0.00            100  3.055889
#> 20426   300.00   TRUE       300.00             76  4.542933
#> 7409    300.00   TRUE       300.00            100  2.866282
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 49459   300.00   TRUE       300.00            100  2.963516
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 42640   300.00   TRUE       300.00            100  3.627130
#> 47722   300.00   TRUE        65.02              1  2.976445
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 46971   300.00   TRUE       262.51            100  3.018968
#> 30378   300.00   TRUE       300.00             57  3.322176
#> 31491   300.00   TRUE       300.00            100  3.385647
#> 49346   300.00   TRUE       300.00             29  2.687835
#> 11332   300.00   TRUE        27.04             68  3.193695
#> 36765   300.00   TRUE       300.00            100  3.279776
#> 39466   300.00   TRUE       300.00            100  3.060187
#> 17374   300.00   TRUE       300.00            100  3.630971
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 25212   300.00  FALSE       300.00            100  3.109374
#> 16464   300.00  FALSE       300.00            100  3.004540
#> 35852   300.00  FALSE       300.00            100  3.356690
#> 14717   300.00  FALSE       300.00            100  2.978336
#> 18746     4.48  FALSE       205.00             27  3.028048
#> 20458   300.00  FALSE       300.00             28  2.424012
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 43426   300.00  FALSE        82.11            100  3.910718
#> 441     294.31  FALSE       152.09            100  3.054287
#> 9665    300.00  FALSE       300.00             64  2.653921
#> 49835   300.00  FALSE         0.00            100  3.134665
#> 29125   300.00  FALSE       300.00            100  2.626221
#> 39796   300.00  FALSE       300.00             71  2.333468
#> 39685   300.00  FALSE       300.00            100  2.918481
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 48183   300.00  FALSE       300.00            100  3.204523
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 21925   300.00  FALSE       300.00             95  3.537385
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 42934   300.00  FALSE       300.00             22  4.499525
#> 20319   300.00  FALSE       300.00            100  3.167901
#> 29527   300.00  FALSE       300.00            100  2.639632
#> 48869   138.98  FALSE         0.00             63  2.748851
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 34345   300.00  FALSE       300.00            100  2.972060
#> 26389   300.00  FALSE        76.38             26  3.197613
#> 20764   300.00  FALSE       300.00             91  2.462702
#> 330     300.00  FALSE       275.17            100  2.372638
#> 26724   300.00  FALSE       185.65            100  3.274624
#> 15008   300.00  FALSE       300.00             48  2.625829
#> 14671   123.04  FALSE        75.11              0  3.182033
#> 29716   300.00  FALSE       300.00            100  2.716439
#> 974      80.00  FALSE        90.05            100  3.099987
#> 31098   300.00  FALSE       300.00             45  3.277234
#> 40750   300.00  FALSE       300.00            100  3.272034
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 10127   300.00  FALSE       300.00             49  2.417865
#> 16845   300.00  FALSE        25.55             33  3.874467
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 31440   300.00  FALSE       300.00             96  2.654870
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 46148   300.00  FALSE       300.00            100  3.015637
#> 32342   300.00  FALSE       300.00             81  2.961658
#> 42990   205.00  FALSE         0.00            100  3.173195
#> 14645   300.00  FALSE       300.00            100  3.232313
#> 13548   300.00  FALSE       300.00            100  2.256336
#> 48157    57.15  FALSE       157.06             45  2.887769