Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.7898858   -0.04921   -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633   -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.9105562    0.01305   -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023   -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917    0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659    0.01041    772.9746
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333   -0.01767   3321.6113
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059   -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273   -0.02403   1787.3024
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940    0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645    0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979   -0.01542    674.0147
#> 37270 714852.5 9561032 1838.96 35.1589821   -0.00854    0.01264   5057.1431
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102    0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160    0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979   -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414   -0.00104   1641.4587
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664   -0.05766 140591.4062
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270   -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.5537501    0.01412    0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.1855058    0.00602    0.01518    854.4269
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746    0.00096   3210.5781
#> 31576 713942.5 9558492 2387.12 46.1431560   -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537   -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151    0.00899   1394.2274
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051    0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493    0.00733    598.7983
#> 38077 714952.5 9557612 2744.35 40.8845494   -0.00493    0.00733    598.7983
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937    0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252    0.01118   2359.5059
#> 46125 713732.5 9560062 2115.18 37.5934798   -0.04616    0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097    0.00156    816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948    0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866   -0.00211    0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.6322593    0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646   -0.04086   6042.5757
#> 22735 714492.5 9559772 2170.30 39.6887865   -0.00440   -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545    0.02405   5681.2412
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154    0.00124   2114.2253
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353    0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721   -0.03229   -0.00511 361979.2812
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321    0.01169    350.1799
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093   -0.00907   1943.4092
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661    0.01199    196.4080
#> 33240 714972.5 9557652 2741.04 39.4343932   -0.00286   -0.00964    571.1364
#> 9704  714642.5 9559652 2249.12 37.1431350   -0.00759   -0.00400    608.6523
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478    0.01092    412.5849
#> 48808 715032.5 9557792 2707.02 25.4376072   -0.02329   -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.7218224   -0.00048   -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140   -0.02330   1456.9246
#> 18265 715522.5 9558782 2539.23 35.8562718   -0.00458   -0.00401   1722.6481
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576   -0.00166    650.6246
#> 7668  713372.5 9559062 2164.82 37.9349627   -0.01807   -0.01883    616.1993
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927   -0.00316    454.9674
#> 7064  715592.5 9558662 2597.69 34.2221961    0.00099    0.00240   1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478   -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.0111801    0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.7342388   -0.00481   -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.9601307   -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.7554242   -0.00604    0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605   -0.03594  16587.9980
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024   -0.00474   2343.2100
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133    0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743   -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567   -0.03793    864.9472
#> 47320 713422.5 9559182 2143.03 56.3251890    0.01246    0.00534    344.1447
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890   -0.00911    173.2806
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949    0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.8600517   -0.02143   -0.05078 622888.2500
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444   -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666   -0.03583   1898.0105
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218   -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.2527305   -0.00007   -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.9116671    0.00707    0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.9597815   -0.01439   -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247    0.01123    364.2903
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253    0.00753   1337.7854
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546    0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062    0.00258    0.00422   1159.2100
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042    0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.8354356    0.00181   -0.00001   1184.7952
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016   -0.00294    734.9908
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730   -0.02610   1872.9276
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143    0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379   -0.01125   -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157   -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124    0.01166   1813.7841
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622    0.00152   1530.5657
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010   -0.00870   2099.7930
#> 43536 713782.5 9558732 2281.50 32.7491217    0.00840    0.00189    556.5211
#> 21490 714672.5 9560382 2021.86 32.8041256   -0.00783   -0.00367   1966.2112
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316   -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.4649135   -0.01332    0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159   -0.01559    866.5775
#> 11288 712822.5 9560442 1873.42 39.6097183   -0.00465   -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759    0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298   -0.00712   1010.5097
#> 35852 715112.5 9559962 2173.61 39.3157909   -0.00738    0.00118   2273.4758
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112    0.03538    265.4677
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165    0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197    0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290   2481.5632
#> 43426 713782.5 9560822 1876.35 23.8230122   -0.00345   -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540    0.00650   1133.1501
#> 42329 715852.5 9558702 2651.92 22.8707563    0.00321    0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.8258980    0.00846    0.00794   1363.5321
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937   -0.02273  72123.6172
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167    0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.7540896    0.00527   -0.02077   8759.9961
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974    0.00973   -0.00943    371.5514
#> 8831  714372.5 9559852 2200.77 48.7604272   -0.00457    0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.2796736    0.00084   -0.00083   1309.2347
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844    0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180    0.01170    253.5129
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410    0.00400   4126.7812
#> 20135 714772.5 9560232 2107.52 37.8507379   -0.00782   -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413    0.00528    355.3475
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462    0.00268    252.0072
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.0192155    0.03310   -0.07650    766.6254
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204    618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229    0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497   -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534    0.00165    0.00555   1882.0179
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719    0.00079   1647.6451
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082    0.02578   1258.8865
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343    0.01177   1558.2231
#> 19767 714272.5 9558052 2431.67 55.7247292    0.00361   -0.01841    756.0055
#> 10981 713832.5 9560022 2125.33 34.8484390   -0.06909   -0.01151 143213.1562
#> 20950 714912.5 9557782 2667.71 30.8113784   -0.04770    0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196   -0.01935    903.6397
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309    0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788   -0.01029    894.2672
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811   -0.02281    492.4785
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305   -0.01135   1293.4855
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208    0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769