Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203     634.3320
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438     753.6304
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.2779266  0.01917  0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174    2310.4871
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310  0.01440    1088.8639
#> 32622 714972.5 9557762 2687.38 30.5564122 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529    3282.1186
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588    1295.1610
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466     724.0713
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250    3686.1240
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440  0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605     798.6775
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 15990 712852.5 9559982 1894.77 37.6943204 -0.01545  0.02405    5681.2412
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669 -0.00991    1403.0193
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373   16031.7500
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230    1056.5876
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917 -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.7170711 -0.03260 -0.03310    2833.9001
#> 7466  712962.5 9560392 1819.03 18.4028314 -0.00140 -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684    2556.8049
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883     616.1993
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171     788.0004
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849    1120.5961
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281    1178.4420
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431    4079.0366
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753    1337.7854
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001    1184.7952
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224    1137.3363
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226     338.3474
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152     850.4633
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124  0.01166    1813.7841
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744     226.7911
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603    1562.0503
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665    1904.4769
#> 13072 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150     503.5590
#> 11288 712822.5 9560442 1873.42 39.6097183 -0.00465 -0.00105    4432.8760
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419     761.5427
#> 31071 713472.5 9558462 2298.33 29.1532385  0.00156 -0.00116    1817.5222
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112  0.03538     265.4677
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075 -0.01945     259.7421
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028    2394.7261
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077    8759.9961
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.8963648  0.00358  0.02422     967.9738
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301 -0.01099    1146.2140
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518    1137.0179
#> 17746 714752.5 9558322 2648.52 47.7365517 -0.00068 -0.00312     593.7773
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268     252.0072
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552    2515.0752
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251     168.1733
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820 -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.6345090  0.00082  0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020     673.2861
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200