Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203     634.3320
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644    1179.0695
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174    2310.4871
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269     491.8017
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514     297.7842
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773     179.3077
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071   0.00150    1274.3594
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937   0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596    5296.9014
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124    2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323     183.1836
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796    4729.6382
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492   0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530     937.1306
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252    5761.3486
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602     345.8933
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883     616.1993
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195  -0.00656    3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927  -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961    1086.6963
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588   0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112     745.3315
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171     788.0004
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860     695.2302
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583    1898.0105
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 28173 714562.5 9560382 2048.26 27.685894   -0.00124  -0.00456    1531.4274
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915    2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075    2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121    1783.0658
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312  -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609    4275.3457
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446    0.00570   0.02230     714.7432
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345  665381.5625
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995    1026.7981
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199     836.4196
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833     300.0125
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216    1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099    1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462   0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.060776    0.01142   0.00868     235.8511
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547     128.2613
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361    1011.0248
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148   10771.1963
#> 974   713402.5 9561062 1854.56 23.634509    0.00082   0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215     917.5296
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177     184.1042
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585    5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803    1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627  -0.00124   25826.6914
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020     673.2861
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491