Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203     634.3320
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884    0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542     674.0147
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149   12632.7793
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905   -0.01385     674.0721
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398     267.2659
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856    0.01114    4510.8833
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358     802.7622
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357     275.7474
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714     475.3120
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449   -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399   -0.00731    1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175    0.00236    2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504    0.05825     162.6688
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690    5121.6924
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306   -0.00436    2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263    0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224    1137.3363
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684     209.1778
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079   -0.00331     699.3024
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105    4432.8760
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517    1060.5494
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493    2212.0525
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490   -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267   31588.1953
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462     716.5762
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785    0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654    1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480    0.00740     909.4791
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803    1058.5630
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135    1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491