Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813   1130.5791
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 43327 713962.5 9561082 1839.11 63.3616837 -0.04951 -0.00529   3282.1186
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581  0.01189    287.5425
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514    297.7842
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270 -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898    458.6938
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382    560.8309
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694   2123.4473
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097  0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086   6042.5757
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 2858  714292.5 9559332 2335.04 42.3559050  0.04547  0.03323    183.1836
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 6310  713802.5 9559752 2238.38 19.6513065 -0.02379  0.01338  10787.6299
#> 18564 715552.5 9558592 2648.36 26.2913143  0.00316  0.00714    475.3120
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452    437.3965
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 20220 714052.5 9558522 2378.67 39.6561279 -0.00855 -0.02175   1122.9120
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236   2715.4138
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252   5761.3486
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.8423462 -0.01057  0.02427    380.9472
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548  0.05543    352.1558
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278  0.00568 -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318    289.0063
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.6771246 -0.00178 -0.00822    472.8748
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.2237194 -0.03986 -0.02524  10861.9482
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211  49347.5547
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334   4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 36292 712862.5 9559602 1915.77 37.9750697  0.00511 -0.01801   1071.6141
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118  19557.6621
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.8127340  0.00546  0.01484    495.0292
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915   2036.0378
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685   2568.4631
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291    293.4231
#> 10213 715542.5 9558782 2532.10 35.4133754 -0.00893 -0.00867   2247.8748
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996   1599.3491
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298 -0.00712   1010.5097
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466   2633.9666
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011    201.1272
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035  0.00955    713.4463
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250   2225.4426
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208    842.4825
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948   1538.1741
#> 43972 713642.5 9558702 2304.16 24.0235474  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029    894.2672
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135   1293.4855
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769