Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029   5577.3916
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339   1002.5877
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273   -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174   2310.4871
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440   1088.8639
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831   -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328   2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842    868.6137
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024    394.8751
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392   1255.0884
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017   -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385    674.0721
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151    0.00899   1394.2274
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973    514.5488
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450    873.0448
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212    561.1948
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276   9614.6924
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566   -0.03355  14919.6260
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338  10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796   4729.6382
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175    804.5679
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585   -0.00974   2136.6939
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357    473.4709
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048   -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542    860.8840
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553   1750.4175
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238   -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538    0.00112    745.3315
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376   2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491    0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521   -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241   1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444   -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261   3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190    264.6404
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075   2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019    0.00119   6589.5796
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279  39495.1797
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144 317260.6875
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683   -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251   -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067   1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842    246.5161
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642   1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599    0.00301    338.9676
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715    520.7355
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345 665381.5625
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409   1575.9425
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167    0.01133    828.8604
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174   -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196    0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475    515.1661
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422    967.9738
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978    0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950   -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257    549.3252
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923    0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247    282.1076
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836    250.5111
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909   -0.01151 143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203    0.00238    787.0053
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161    0.00709   1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788   -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172   -0.00119    792.3544
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858    298.4490
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458    485.7418
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933   -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281    492.4785
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751    0.00388    180.4414
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607