Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183  -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056  -0.06976     301.2347
#> 47223 715302.5 9557472 2857.45 39.5335149   -0.01021  -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.9105562    0.01305  -0.00555     430.4604
#> 25090 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 1300  714042.5 9558482 2408.27 24.1484522    0.00659   0.01041     772.9746
#> 11088 715532.5 9558762 2549.28 38.0501272   -0.00373  -0.00057    1630.3259
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431    4079.0366
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 25169 712642.5 9560232 1883.10 46.8450293    0.00884   0.01345    1120.6929
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894  -0.03416     947.4878
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979  -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110    0.00414  -0.00104    1641.4587
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070  -0.00530    9000.4356
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430  -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905  -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382   0.00672    1114.8801
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440   0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151   0.00899    1394.2274
#> 47022 715362.5 9560102 2059.29 49.1196718    0.02059  -0.00628     556.0121
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108  -0.00678    1891.6584
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613  -0.01898     588.3761
#> 36230 713322.5 9560662 1798.99 40.1282451   -0.00033  -0.03917     116.1986
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212     561.1948
#> 23141 715632.5 9558942 2495.99 34.3768948    0.01275   0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089  -0.00051     779.0555
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405    5681.2412
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224   0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294  -0.00096     520.5570
#> 37348 715632.5 9558182 2743.16 13.6203527   -0.00904  -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.7021461   -0.01771  -0.01939     703.7817
#> 34563 715022.5 9558162 2675.70 18.8898455   -0.01993   0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752  -0.00517    1000.7119
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442  -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.0379275    0.04527   0.04613     289.6113
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974    2136.6939
#> 44550 714842.5 9557642 2696.38 15.1730683    0.00359  -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.1679538    0.01008   0.00102     483.4184
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749   0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350   -0.00759  -0.00400     608.6523
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372  -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.7968588   -0.00449  -0.00181    2406.1252
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762   0.01252    5761.3486
#> 375   715622.5 9558922 2510.87 37.3665885   -0.01017   0.00197     711.1091
#> 38862 712952.5 9559892 1950.86 37.7218224   -0.00048  -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140  -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.1948618   -0.02681  -0.01319    3457.7519
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032   0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736   0.01684    2556.8049
#> 32280 712572.5 9560182 1956.86 50.6781170   -0.02554   0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282   0.00109     767.4097
#> 11618 713852.5 9558402 2432.14 54.0591409   -0.00410   0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645  -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667   -0.01611   0.02961    1086.6963
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904  -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.8736492   -0.01238  -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 40270 713312.5 9559012 2186.56 30.9357102   -0.00771  -0.00599    1378.5760
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188   0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458  -0.14813  287033.4688
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521  -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 4588  713852.5 9558612 2308.97 52.1804123   -0.01059  -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.6498183    0.01299   0.00141    4611.0718
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175   0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.8276239   -0.00444  -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666  -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.2527305   -0.00007  -0.02132    1580.1621
#> 26428 713042.5 9560332 1895.15 33.9597815   -0.01439  -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924   0.02606     564.6876
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546   0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062    0.00258   0.00422    1159.2100
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042   0.00957     495.4453
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292  -0.00109    2047.9031
#> 48650 715592.5 9558612 2626.74 31.8765069    0.00010   0.00140    1066.3957
#> 48574 712692.5 9560412 1901.97 41.9783258    0.01364  -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185  -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285  -0.01325     595.0426
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143   0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379   -0.01125  -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741    0.00445   0.00875     597.2893
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610     454.8103
#> 49459 714692.5 9557342 2639.74 42.0470808    0.00216  -0.02606     919.4241
#> 37115 715612.5 9559202 2358.15  9.7276774   -0.01020   0.00279   39495.1797
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893  -0.00867    2247.8748
#> 14779 712502.5 9560202 2004.61 27.5226006    0.00115   0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.3236027    0.00758  -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.5225656   -0.01454   0.00144  317260.6875
#> 36765 715302.5 9560262 1952.73 27.2034632   -0.00585  -0.02665    1904.4769
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316  -0.00427     866.9398
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002  -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534   0.01476    1275.9446
#> 30809 712532.5 9560032 1928.24 40.8994463    0.00570   0.02230     714.7432
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298  -0.00712    1010.5097
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336   0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007  -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295   0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.3432999   -0.00533   0.01493    2212.0525
#> 43426 713782.5 9560822 1876.35 23.8230122   -0.00345  -0.00075    8141.7549
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 39193 715332.5 9560852 1872.90 30.2905598   -0.00161   0.00171     756.8818
#> 1085  715722.5 9557532 3097.22 28.7842537    0.02327   0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263   0.03598     186.3527
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844   0.00514     897.9290
#> 33956 714082.5 9558712 2327.95 39.6842028    0.02218   0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.8963648    0.00358   0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423  -0.00983     270.7949
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410   0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301  -0.01099    1146.2140
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224  -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338   0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.1180924   -0.01322  -0.00759    4131.5356
#> 44267 712972.5 9559042 2191.01 22.2691506    0.01462   0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978   0.01282    2322.7944
#> 21070 712832.5 9560062 1884.42 50.9336562   -0.00342   0.02552    2515.0752
#> 44823 714202.5 9557442 2548.13 43.2302386   -0.00233   0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497   -0.00930  -0.00360    4288.7329
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.5684513    0.00420   0.01960     884.2003
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719   0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273  -0.01047    2826.7700
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343   0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.8303070    0.01852   0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964    0.00537  -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865  -0.00805    7489.7417
#> 36016 715262.5 9557682 2814.30 37.0021874    0.00203   0.00238     787.0053
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788  -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.6572738   -0.01305   0.00585    5714.3267
#> 46148 715082.5 9558042 2725.39 42.8360436   -0.00181  -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.9643862   -0.00933  -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 28978 713982.5 9560462 2133.10  5.2293858    0.00366   0.00525     162.0592
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181  -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208   0.00492     772.2691
#> 40852 715392.5 9557932 2939.71 18.2974072    0.00505   0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983