Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832    3022.3462
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769