Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483    566.7674
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560    977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775