Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029    5577.3916
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203     634.3320
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483     566.7674
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896     276.0310
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324     861.3537
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579    2246.7725
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433    1475.3191
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.43573  0.00250 -0.01340     604.1032
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730 -0.01491    1844.4353
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189     287.5425
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164     895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626    4316.0977
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838     238.6685
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114    4510.8833
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413     364.9788
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107    1392.5880
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212     561.1948
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077    3559.4204
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077     300.7915
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154  0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355   14919.6260
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323     183.1836
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.64579  0.00060  0.00270    4057.5637
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117    1018.2141
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241     585.6182
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319    3457.7519
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436    2602.7390
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883     616.1993
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596     842.0994
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240    1201.4724
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081     940.6002
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474    2343.2100
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793     864.9472
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534     344.1447
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171     788.0004
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781   29995.1543
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 21683 714792.5 9561052 1807.84 55.73275 -0.02070 -0.02661    2508.5374
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321    1170.5904
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.02568  0.04250 -0.00970    1443.9200
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778    4594.4351
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483     343.1384
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285 -0.01325     595.0426
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610     454.8103
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291     293.4231
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867    2247.8748
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253     528.1924
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306    1718.3652
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301     338.9676
#> 36561 715292.5 9560802 1844.64 31.93724  0.00002 -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109     605.1908
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736     477.8591
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493    2212.0525
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463     851.6306
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273   72123.6172
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077    8759.9961
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367    3002.8828
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759    1880.4403
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790     209.6952
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033    3446.5520
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298     436.1459
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099    1146.2140
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705     968.9738
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599   12873.8018
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060     383.3547
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165     805.1403
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.11353  0.02236 -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.90101  0.03044 -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578    1258.8865
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368    1071.8734
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270   35950.5508
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011     501.7867
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884     672.2466
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.81402  0.00282  0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222     991.8300
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020     673.2861
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149    4242.3579
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603