Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250