Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633 -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475    1001.0861
#> 29391 712842.5 9560152 1930.16 28.0955584  0.00753  0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266  0.02896     276.0310
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240 -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767    3321.6113
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 25169 712642.5 9560232 1883.10 46.8450293  0.00884  0.01345    1120.6929
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088     266.4355
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311     586.7354
#> 3835  714142.5 9558552 2388.42 32.9003825  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.3814924 -0.01966 -0.02234     849.0547
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510    3342.5518
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683 -0.00905 -0.01385     674.0721
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189    2317.1819
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173    8967.1924
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206  0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774     527.1068
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150    1274.3594
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605     798.6775
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924    8023.8657
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058     431.4362
#> 26567 714022.5 9558872 2327.42 46.6656935  0.00290  0.00230    1056.5876
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495     266.5888
#> 21053 712862.5 9559992 1901.01 30.9506071 -0.00762  0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.3121575 -0.00405 -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319    3457.7519
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034  0.00294     739.5768
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550 -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282  0.00109     767.4097
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927 -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395    2011.3270
#> 7064  715592.5 9558662 2597.69 34.2221961  0.00099  0.00240    1201.4724
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469    1021.5407
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832    3022.3462
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453  0.01337     486.7551
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.0638922 -0.00338  0.00728    1137.2450
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833     369.3771
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 49072 715472.5 9558902 2482.92 30.6893384 -0.01102 -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110 -0.02450    2589.9910
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226     338.3474
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216 -0.02606     919.4241
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210     585.8992
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.7567938 -0.06953 -0.01797    7557.7866
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208     951.3401
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165  0.00875    1165.3953
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548     532.8210
#> 47005 714942.5 9559192 2333.66 29.2208476 -0.04937 -0.02273   72123.6172
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507     504.3085
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262  0.01338 -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.7604272 -0.00457  0.01697     412.1185
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033    3446.5520
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780     825.4451
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902     218.7467
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036     476.9504
#> 39440 714222.5 9557872 2508.52 39.0476467  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714  480535.5000
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730 -0.00066  0.05257     549.3252
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177    1558.2231
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368    1071.8734
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238     787.0053
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157     766.0220
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020     673.2861
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247  0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769