Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv      carea
#> 16435 713512.5 9559092 2166.13 55.9739659   0.02056   -0.06976   301.2347
#> 44302 714932.5 9557982 2649.56 37.3006984   0.01633   -0.01813  1130.5791
#> 13748 715232.5 9559542 2212.12 26.1234377   0.00023   -0.00043  1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475  1001.0861
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261   323.8441
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741   746.4155
#> 34512 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677  1675.4679
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433  1475.3191
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403  1787.3024
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491  1844.4353
#> 43669 712392.5 9560162 2001.90 47.2839150  -0.00284    0.01804   628.5153
#> 16811 714902.5 9559262 2362.67 50.7033271  -0.01407    0.00547   519.4228
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388  1051.4222
#> 42598 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010  1527.1028
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189   287.5425
#> 33792 714652.5 9558172 2598.36 59.4586952  -0.03525   -0.11024   394.8751
#> 2990  713542.5 9560372 1970.07 31.8879661   0.01903    0.01447   331.4833
#> 43007 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514   297.7842
#> 40307 713302.5 9560702 1845.07 39.0046749   0.00129    0.02510  3342.5518
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769  1032.7177
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385   674.0721
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677  1675.4679
#> 14522 715012.5 9557732 2721.34 35.1813275  -0.00382    0.00672  1114.8801
#> 36030 712852.5 9559572 1938.79 28.6788295  -0.01440    0.00710   843.7961
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581  1855.1630
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694  2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877   0.00382    0.00088  2322.1746
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083   994.7876
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437  1155.1469
#> 19730 715332.5 9558452 2640.35 47.6425866  -0.00211    0.00642   613.6390
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114   150.1498
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651   419.0927
#> 2858  714292.5 9559332 2335.04 42.3559050   0.04547    0.03323   183.1836
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939   703.7817
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063  5035.0000
#> 20001 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376  2048.3198
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517  1000.7119
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178   800.1933
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175   804.5679
#> 13134 714652.5 9559632 2265.33 49.5998104  -0.03552   -0.02058   431.4362
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925   665.6116
#> 37141 714382.5 9559822 2219.36 10.1064025   0.01661    0.01199   196.4080
#> 23029 713502.5 9559102 2174.07 56.4575423   0.03395    0.01495   266.5888
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538   807.8237
#> 16015 712582.5 9560312 1899.18 56.5469237   0.03480    0.02530   937.1306
#> 20220 714052.5 9558522 2378.67 39.6561279  -0.00855   -0.02175  1122.9120
#> 421   714892.5 9561022 1849.01 26.1320321   0.00002    0.00588   749.2502
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236  2715.4138
#> 21053 712862.5 9559992 1901.01 30.9506071  -0.00762    0.01252  5761.3486
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357   473.4709
#> 7466  712962.5 9560392 1819.03 18.4028314  -0.00140   -0.02330  1456.9246
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542   860.8840
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166   650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009  1247.7061
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625  1063.9150
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553  1750.4175
#> 27695 714142.5 9558452 2449.28 20.4901167  -0.00539    0.00809  1102.9956
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240  1201.4724
#> 45500 712812.5 9560052 1852.16 56.7703772   0.00904   -0.02013  4229.5708
#> 2552  714942.5 9557652 2719.32 34.8736492  -0.01238   -0.00851  1012.5963
#> 41783 715572.5 9558652 2609.51 32.4082118   0.00291   -0.00081   940.6002
#> 43859 714892.5 9557602 2722.70 28.1293630   0.00370   -0.00190   640.8842
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524 10861.9482
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205  1112.2626
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061  5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781 29995.1543
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281  1178.4420
#> 43201 715712.5 9557922 2852.82 31.4622584  -0.00949    0.00520  3390.5708
#> 34887 714662.5 9557392 2609.02 51.0207457  -0.01878   -0.04351  1442.9247
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261  1627.8989
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322  4605.1411
#> 22907 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476  1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776  1021.8267
#> 26428 713042.5 9560332 1895.15 33.9597815  -0.01439   -0.01161  3065.3694
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321  1170.5904
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970  1443.9200
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190   264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182   0.02281    0.04759   158.9370
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077  3056.9475
#> 22297 714642.5 9558172 2606.18 59.4690084  -0.00495   -0.03756   353.9086
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044   288.5468
#> 24723 712942.5 9560202 1983.18 33.0716969   0.00285   -0.01325   595.0426
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538   807.8237
#> 30363 715192.5 9557542 2796.31 25.3218698  -0.00157   -0.00152   850.4633
#> 47722 712732.5 9561022 1999.03 37.4926392   0.00004   -0.02384   947.2075
#> 4479  713982.5 9557812 2399.24 41.2311889   0.01895    0.04744   226.7911
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210   585.8992
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291   487.3438
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655   130.9468
#> 10785 714052.5 9557852 2377.76 20.9840063  -0.00025   -0.05345  2417.1389
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609  4275.3457
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842   246.5161
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230   714.7432
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419   761.5427
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116  1817.5222
#> 18286 714792.5 9557882 2650.57 28.2536948  -0.00124    0.00204   853.2902
#> 27965 715042.5 9560342 2063.39 29.1973563   0.04336    0.01403   241.2991
#> 14717 715382.5 9558062 2799.05 50.5904544  -0.00812   -0.02208   951.3401
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538   265.4677
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500   538.8209
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028  2394.7261
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333   422.8840
#> 1085  715722.5 9557532 3097.22 28.7842537   0.02327    0.01833   300.0125
#> 13286 713662.5 9559702 2268.14 41.9966605  -0.01341    0.04431   315.9084
#> 20547 713732.5 9560822 1851.67 36.9683829  -0.01405    0.00476 11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179   0.00167    0.01133   828.8604
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853  1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759  1880.4403
#> 43403 712632.5 9559262 2072.53 34.4863297  -0.00069   -0.00900   614.5898
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780   825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267 31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263   0.00706   -0.03035  1471.9783
#> 8895  714132.5 9557692 2487.19 37.3642967   0.03423   -0.00983   270.7949
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098   164.0802
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366   267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048  1550.6704
#> 34810 714112.5 9559572 2337.51 10.2880302  -0.00393    0.00032   994.9224
#> 12044 713612.5 9559292 2247.76 23.0747293  -0.00112   -0.00038  2017.3547
#> 33666 713632.5 9561012 1775.88 21.1788119  -0.00950   -0.05130   107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988   0.02118    0.00162   290.2031
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552  2515.0752
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257   549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513   161.9878
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487  1769.3392
#> 35038 715092.5 9558152 2715.69 40.1838220   0.00270   -0.00250  2225.4426
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204   618.2027
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740   909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911   100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788   0.00119    0.00361  1011.0248
#> 1281  715102.5 9560032 2120.64 34.6490497  -0.00930   -0.00360  4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534   0.00165    0.00555  1882.0179
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148 10771.1963
#> 3724  714322.5 9557402 2603.46 34.5974835   0.02523    0.02247   282.1076
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836   250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260  1484.6593
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438  1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248  1432.4019
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532 16771.3066
#> 26283 715932.5 9557592 3075.48 32.7668833  -0.00375    0.00215   917.5296
#> 15138 713702.5 9561122 1780.68 49.2216583   0.00607   -0.04566 12862.4922
#> 6732  715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523   888.2811
#> 1577  713322.5 9559742 2027.47 29.5674870  -0.03284   -0.00886 43249.3945
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233   968.7780
#> 28526 715152.5 9560732 1871.53 29.5342555  -0.00163   -0.00377  1727.6055
#> 46639 715232.5 9560042 2107.15 39.6572738  -0.01305    0.00585  5714.3267
#> 35257 713822.5 9557982 2350.80 29.3079371   0.01549   -0.00169   649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354   0.00992   -0.01052   451.7209
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458   485.7418
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156   758.4840
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525   162.0592
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149  4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607