Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681  -0.00029   5577.3916
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056  -0.06976    301.2347
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969    368.5606
#> 30059 713832.5 9559662 2288.02 18.2779266    0.01917   0.00333    258.6537
#> 19963 714612.5 9560552 2031.70 26.8482293    0.00046   0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240  -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498  -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907   -0.00333  -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.1291603    0.00090   0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273  -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174   2310.4871
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730  -0.01491   1844.4353
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068   0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.1589821   -0.00854   0.01264   5057.1431
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894  -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214   0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160   0.00311    586.7354
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952   0.01378    440.9104
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678  -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193    0.00528   0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749    0.00129   0.02510   3342.5518
#> 35285 714602.5 9560112 2142.75 51.1078353   -0.00017  -0.00032    994.6740
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832   0.00838    238.6685
#> 25081 713392.5 9558452 2242.24 46.5384969   -0.02716   0.03696   4143.4019
#> 18391 715112.5 9559432 2235.02 22.2576915   -0.02651  -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164   -0.00521  -0.00739    558.2286
#> 28230 715182.5 9558102 2791.83 39.0717109   -0.00669  -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225   0.00325    616.8652
#> 46125 713732.5 9560062 2115.18 37.5934798   -0.04616   0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097   0.00156    816.6036
#> 2950  715042.5 9557712 2740.72 26.5250811   -0.00089  -0.00051    779.0555
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743   0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.5927252    0.00804   0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.5755084   -0.00255  -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.4454260    0.00606  -0.00556    660.3690
#> 48434 713352.5 9559042 2176.63 35.7021461   -0.01771  -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204  -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.5357857    0.00140  -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.4809914   -0.00509   0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321   0.01169    350.1799
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058    431.4362
#> 22332 715482.5 9558922 2475.28 19.9366394   -0.02617   0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974   2136.6939
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749   0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480   0.02530    937.1306
#> 35194 713692.5 9559952 2185.68 28.7968588   -0.00449  -0.00181   2406.1252
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260  -0.03310   2833.9001
#> 48808 715032.5 9557792 2707.02 25.4376072   -0.02329  -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984    0.00953  -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221   0.00121   4159.6738
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032   0.00542   3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720   -0.02027  -0.01473   2930.2791
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099   0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168  -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.2645529   -0.00195  -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568  -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927  -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167   -0.00539   0.00809   1102.9956
#> 13146 713792.5 9558552 2341.40 29.8253180   -0.00104   0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219    420.7032
#> 41783 715572.5 9558652 2609.51 32.4082118    0.00291  -0.00081    940.6002
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605  -0.03594  16587.9980
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024  -0.00474   2343.2100
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567  -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562   -0.01169  -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734   -0.10372  -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278  -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453   0.01337    486.7551
#> 44767 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129   0.00171    788.0004
#> 20368 715252.5 9560622 1865.83 33.5724620    0.00025  -0.01245   6353.3179
#> 18440 714862.5 9558932 2448.29 41.0524260   -0.04989  -0.02781  29995.1543
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.8763323   -0.00310  -0.00860    695.2302
#> 34887 714662.5 9557392 2609.02 51.0207457   -0.01878  -0.04351   1442.9247
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218  -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.2527305   -0.00007  -0.02132   1580.1621
#> 43074 713172.5 9559622 2052.40 31.5275756   -0.00193   0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924   0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213  -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.0256786    0.04250  -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936   -0.00124  -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070   0.00010   3467.2690
#> 49072 715472.5 9558902 2482.92 30.6893384   -0.01102  -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051   0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.3243783   -0.00296   0.00426   1315.5334
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915  -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185  -0.01185    953.8614
#> 26058 713402.5 9560442 1880.12 31.3774607    0.00730  -0.02610   1872.9276
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181  -0.00341   1363.4000
#> 13278 712752.5 9560452 1901.25 42.1135439   -0.00118  -0.00252   3631.7017
#> 11057 714962.5 9557642 2737.48 42.6859287   -0.00488  -0.00632    620.0229
#> 14140 714672.5 9559262 2326.36 45.7925059   -0.01400  -0.01210    585.8992
#> 6212  714592.5 9560362 2034.54 12.5225656   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186   0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.7000149    0.00683  -0.00603   1562.0503
#> 31253 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797   7557.7866
#> 32405 713022.5 9558982 2208.85 15.0046187    0.00599   0.00301    338.9676
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009   0.00109    605.1908
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759   0.00158    837.3730
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909  -0.01181   1286.3949
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204    853.2902
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409   1575.9425
#> 49835 713272.5 9560372 1898.68 32.8258980    0.00846   0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.9946898    0.00544  -0.01014    271.4219
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.1066333   -0.00007  -0.00853   1354.6001
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737   0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495   -0.00196  -0.00175    725.1355
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410   0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041  -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185   0.01655    560.8554
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785   0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424   0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462    0.00379  -0.01018    937.6911
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586   0.00654   1576.2065
#> 5656  713132.5 9560142 2042.24 33.5953803   -0.00035   0.00955    713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066   0.05257    549.3252
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911    100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534    0.00165   0.00555   1882.0179
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635  -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247    282.1076
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511   0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719   0.00079   1647.6451
#> 12300 712372.5 9560062 1903.64 50.0369772    0.00343   0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050   1471.4346
#> 19767 714272.5 9558052 2431.67 55.7247292    0.00361  -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523    888.2811
#> 11394 713472.5 9561232 1889.48 34.3144424    0.00108  -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253  -0.01233    968.7780
#> 40653 714262.5 9561262 1791.94 12.2223357    0.00161   0.00709   1181.2555
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014  -0.00884    672.2466
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986  -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282   0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.8455602    0.00298   0.00222    991.8300
#> 47245 715372.5 9560792 1900.65 34.4708598   -0.00011  -0.00419    323.0790
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110  -0.02931   1812.3820
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208   0.00492    772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769