Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114     150.1498
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121    4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607