Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309   -0.00681  -0.00029    5577.3916
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056  -0.06976     301.2347
#> 19963 714612.5 9560552 2031.70 26.8482293    0.00046   0.00324     861.3537
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273  -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278    0.00250  -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104  -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453   0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.8450293    0.00884   0.01345    1120.6929
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979  -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068   0.00588    1295.1610
#> 16811 714902.5 9559262 2362.67 50.7033271   -0.01407   0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550  -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.1593594   -0.01607   0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321    0.00102   0.00388    1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423    0.00214   0.00466     724.0713
#> 44328 713222.5 9558932 2227.94 29.1555304   -0.00769  -0.00281    1325.6138
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664  -0.05766  140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217   -0.02678  -0.04272    4955.5586
#> 18304 714582.5 9558932 2463.25 22.1946661   -0.02131   0.00501    3889.7292
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126  -0.02044  129412.3984
#> 22566 712832.5 9559842 1897.49 33.0602377    0.00672   0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382   0.00672    1114.8801
#> 2942  713232.5 9560652 1851.75 34.5350311    0.00856   0.01114    4510.8833
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356   0.00774     527.1068
#> 48364 715582.5 9558222 2746.23 25.0720602    0.01257   0.00593     939.0466
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804  -0.01596    5296.9014
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.6644985    0.06765   0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646  -0.04086    6042.5757
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545   0.02405    5681.2412
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154   0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887     366.8537
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.4454260    0.00606  -0.00556     660.3690
#> 37348 715632.5 9558182 2743.16 13.6203527   -0.00904  -0.00866   29195.6016
#> 6310  713802.5 9559752 2238.38 19.6513065   -0.02379   0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752  -0.00517    1000.7119
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442  -0.00178     800.1933
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066  -0.00036     516.4709
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613   0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175    1122.9120
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119  -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.9863262   -0.02917  -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.7170711   -0.03260  -0.03310    2833.9001
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.9544222   -0.04708   0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140  -0.02330    1456.9246
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576  -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529   0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070   -0.00209   0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.7269720   -0.02027  -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856  -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034   0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.0646468    0.00062  -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126  -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817   -0.00550  -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.2748872   -0.00845   0.00454    3476.0818
#> 35014 713502.5 9559772 2190.48 44.2357795    0.00927   0.01643     670.2289
#> 16370 713852.5 9558632 2296.97 34.6244125   -0.01972   0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078   0.00219     420.7032
#> 41783 715572.5 9558652 2609.51 32.4082118    0.00291  -0.00081     940.6002
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133   0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178  -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017   0.01213     328.7922
#> 33910 715202.5 9559572 2195.66 13.7916671   -0.04376  -0.01644 1057765.6250
#> 6217  714442.5 9559932 2151.30 36.6704448    0.01489   0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 24135 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797    7557.7866
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175   0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.7896130    0.00466  -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.9116671    0.00707   0.01833     369.3771
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.2390497    0.01011   0.00739     789.7573
#> 12629 714712.5 9558872 2533.15 37.0256786    0.04250  -0.00970    1443.9200
#> 28173 714562.5 9560382 2048.26 27.6858936   -0.00124  -0.00456    1531.4274
#> 23783 713062.5 9560252 1983.09 36.4298025    0.00774  -0.00224    1137.3363
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605   0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.0463191   -0.08153  -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.6578678   -0.01341   0.00171    4425.5410
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157  -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.4649416    0.00560   0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622   0.00152    1530.5657
#> 17227 714812.5 9558892 2545.43 54.0717460    0.00458   0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.0363832   -0.03675  -0.01905    6487.0010
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893  -0.00867    2247.8748
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010  -0.00870    2099.7930
#> 41310 714452.5 9560402 2092.27 12.4629780   -0.00195  -0.00205    1112.2626
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025  -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367    0.00614  -0.00093     118.1402
#> 31253 712612.5 9559492 1903.67 46.7567938   -0.06953  -0.01797    7557.7866
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378   0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203   0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763   -0.00079  -0.00331     699.3024
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505     351.4078
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284   0.01085    2229.5105
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815  -0.03345  665381.5625
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197   0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825   0.00445     355.5078
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626     450.7346
#> 43243 715682.5 9558462 2693.23 27.4062902   -0.00031  -0.00199     836.4196
#> 20547 713732.5 9560822 1851.67 36.9683829   -0.01405   0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.7881179    0.00167   0.01133     828.8604
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069  -0.00900     614.5898
#> 8413  713242.5 9561232 1967.45 38.1945125    0.00236   0.00483    1097.3082
#> 42934 713832.5 9558672 2270.74 27.6119821   -0.03193  -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452  -0.01298     436.1459
#> 36245 713142.5 9560572 1827.82 43.6759998   -0.00410   0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185   0.01655     560.8554
#> 13737 714272.5 9559622 2257.05 55.4571579   -0.03224  -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413   0.00528     355.3475
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130     107.5831
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270  -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009  -0.00599   12873.8018
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635  -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608   0.00148   10771.1963
#> 3868  714542.5 9560622 2024.15 29.4471659    0.00360   0.00750    1892.8978
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511   0.01499     617.3855
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719   0.00079    1647.6451
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082   0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952  -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354   0.00907    1281.5044
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745   0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.1104343   -0.01502   0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.3344327    0.00293   0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572   0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865  -0.00805    7489.7417
#> 18504 713372.5 9560672 1786.89  6.4738501   -0.00278  -0.01172  111973.8906
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052     451.7209
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716  -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305  -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241   0.00020     673.2861
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247   0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.6213801    0.00208   0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.3497490   -0.00771   0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491