Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939    703.7817
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102    483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512  13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458    485.7418
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769