Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043    1806.4552
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324     861.3537
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403    1787.3024
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761     928.6858
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046    1024.1573
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529    3282.1186
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035    2697.5974
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842     868.6137
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269     491.8017
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024     394.8751
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160    9344.2930
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769    1032.7177
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710     843.7961
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.25769 -0.02651 -0.03629  360379.1875
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257  0.00593     939.0466
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150    1274.3594
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118    2359.5059
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276    9614.6924
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605     798.6775
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651     419.0927
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866   29195.6016
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862     874.5408
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714     475.3120
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613     289.6113
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407     344.8395
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495     266.5888
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964     571.1364
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117    1018.2141
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236    2715.4138
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197     711.1091
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319    3457.7519
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625    1063.9150
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262    1760.2162
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020    1563.5549
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.23578  0.00927  0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.27910  0.00645 -0.00395    2011.3270
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512   13178.3223
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201     788.0621
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555    4439.0967
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594   16587.9980
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.15385 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813  287033.4688
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524   10861.9482
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078  622888.2500
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776    1021.8267
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321    1170.5904
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363     565.5207
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077    3056.9475
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109    2047.9031
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075    2170.2249
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230    1044.6443
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367    1966.2112
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559     866.5775
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109     605.1908
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565    1410.8372
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.44011  0.00112  0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466    2633.9666
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479     604.0975
#> 25140 713332.5 9560692 1809.95 52.25719  0.00075 -0.01945     259.7421
#> 13512 715152.5 9557992 2778.81 43.38952 -0.00724  0.01464    4029.5947
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973     215.5101
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507     504.3085
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175     725.1355
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983     270.7949
#> 24009 714042.5 9558962 2357.58 34.36200  0.05041 -0.00011     201.1272
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038    2017.3547
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948    1222.5433
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547     128.2613
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740     909.4791
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.96760 -0.00820 -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.63451  0.00082  0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532   16771.3066
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841     756.0055
#> 10127 713722.5 9558712 2311.10 19.48114  0.01129  0.00491     261.7371
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233     968.7780
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377    1727.6055
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124   25826.6914
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448     941.8503
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931    1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250