Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337    486.7551
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715    520.7355
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769