Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332   1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995   1026.7981
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585   5714.3267
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250