Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.7898858 -0.04921 -0.12438     753.6304
#> 47223 715302.5 9557472 2857.45 39.5335149 -0.01021 -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475    1001.0861
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 37265 712812.5 9559942 1860.77 63.3370465  0.00514 -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579    2246.7725
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073  0.00174    2310.4871
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.0050732 -0.00891 -0.00949    6217.8213
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035    2697.5974
#> 125   713992.5 9561062 1878.28 49.9710871  0.00183  0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120     512.9740
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328    2235.5259
#> 2990  713542.5 9560372 1970.07 31.8879661  0.01903  0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.6234412 -0.02169 -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 47022 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902  0.01358     802.7622
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118    2359.5059
#> 23031 714042.5 9558492 2404.94 28.4960559  0.01205  0.02205     629.4165
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077     300.7915
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034    4495.6929
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771 -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672     565.1107
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 29926 714702.5 9557622 2637.00 44.6425796  0.01843  0.04407     344.8395
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 40733 712862.5 9558892 2108.60 29.9949135  0.00175  0.00236    2715.4138
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009    1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032  0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554  0.03304    1569.7313
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410  0.03620     379.0855
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028  0.01318     289.0063
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664    1807.2659
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017  0.01213     328.7922
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205    1112.2626
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204 -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245    6353.3179
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245     629.5983
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801     172.7491
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123     364.2903
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193  0.02812    1423.7578
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.0457041 -0.00789 -0.00261    3353.5422
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693  0.01133    2397.2681
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 17674 713882.5 9558562 2339.20 41.0363832 -0.03675 -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306    1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189     556.5211
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.6801711  0.00000  0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 36561 715292.5 9560802 1844.64 31.9372405  0.00002 -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534  0.01476    1275.9446
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204     853.2902
#> 8635  713832.5 9557902 2397.71 47.1194761  0.02295  0.02715     520.7355
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208     951.3401
#> 26963 715702.5 9557882 2880.32 43.7332955 -0.01174 -0.01466    2633.9666
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242  0.01132     737.5982
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.8707563  0.00321  0.00548     532.8210
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431     315.9084
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083    1309.2347
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376     629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218  0.00902     218.7467
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.0372984  0.00586  0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948    1222.5433
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650     766.6254
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220  0.00270 -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142  0.00868     235.8511
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523  0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576    1045.1074
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948    1538.1741
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523     888.2811
#> 42990 712432.5 9560692 2102.11 32.4500377 -0.00087 -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419     323.0790
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931    1812.3820
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491