Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393    2081.0056
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741     746.4155
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514   -0.00644    1179.0695
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440    1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550   -0.00120     512.9740
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466     724.0713
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447     331.4833
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664   -0.05766  140591.4062
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382     560.8309
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151    0.00899    1394.2274
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993   -0.00413     364.9788
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917     116.1986
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156     816.6036
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743    0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991    1403.0193
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255   -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606   -0.00556     660.3690
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771   -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140   -0.00070    2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169     350.1799
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843    0.04407     344.8395
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855   -0.02175    1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181    2406.1252
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588     749.2502
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364   -0.03357     473.4709
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542     860.8840
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690    5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427     380.9472
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553    1750.4175
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656    3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131    0.00201     788.0621
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771   -0.00599    1378.5760
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813  287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911     173.2806
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878   -0.04351    1442.9247
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171   -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384    0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372    4154.5786
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193    0.02812    1423.7578
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213   -0.01363     565.5207
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396     704.7076
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422    1159.2100
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042    0.00957     495.4453
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774   -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285   -0.01325     595.0426
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610     454.8103
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732    4237.6934
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758   -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036    2505.8369
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306    1718.3652
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345    2417.1389
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427     866.9398
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842     246.5161
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162   -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493    2212.0525
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598     186.3527
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432   -0.00362     283.7090
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997     230.8641
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759    4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312     593.7773
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268     252.0072
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667    0.00863    1098.3462
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312     254.0656
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608    0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502     520.5218
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750    1892.8978
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907    1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347    1893.3645
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624    1870.8307
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576    1045.1074
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129    0.00491     261.7371
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172   -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169     649.9975
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241    0.00020     673.2861
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249     148.1857
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247    0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603