Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530    9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064 -0.00916    2009.1686
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312     593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769