Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324     861.3537
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333  -0.01767    3321.6113
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588    1295.1610
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922   0.01088     266.4355
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378     440.9104
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096    3210.5781
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398     267.2659
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426  -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581    1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173    8967.1924
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593     939.0466
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193    1809.3838
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991    1403.0193
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070    2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714     475.3120
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063   18733.8848
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495     266.5888
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119  -0.01560     977.4719
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565    2232.6133
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427     380.9472
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539   0.00809    1102.9956
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784   0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220    8936.2666
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516   0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849    1120.5961
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474   0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195    1901.6743
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606     564.6876
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442  -0.00118   19557.6621
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778    4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957     495.4453
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970   0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325     595.0426
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119    6589.5796
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216  -0.02606     919.4241
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291     293.4231
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234     849.0547
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150     503.5590
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736     477.8591
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402  -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112   0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064     683.2795
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309  -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075  -0.01945     259.7421
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507     504.3085
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033    3446.5520
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218   0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097   0.01487    1769.3392
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499     617.3855
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841     756.0055
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865  -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172  111973.8906
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788  -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803    1058.5630
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020     673.2861
#> 13548 715812.5 9558122 2821.65 15.508249    0.02751   0.00388     180.4414
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379  -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 11352 27.998156   135.89   TRUE        82.11              4  4.291317
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983