Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659   0.02056   -0.06976    301.2347
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741    746.4155
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.5839352  -0.00498   -0.00752   1321.5997
#> 24129 714602.5 9560542 2038.30 25.5848574   0.00065    0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.0977107  -0.00858   -0.00292   1233.9299
#> 27973 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679    550.0344
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406   2421.1145
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783    377.8918
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311    586.7354
#> 3835  714142.5 9558552 2388.42 32.9003825   0.00979   -0.00269    491.8017
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.9825182  -0.02123    0.01164    895.6606
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501   3889.7292
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070 138869.3906
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.5384969  -0.02716    0.03696   4143.4019
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494  -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173   8967.1924
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.2829086   0.00356    0.00774    527.1068
#> 3133  715042.5 9559252 2344.13 29.0208853   0.00422    0.02357    275.7474
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626   1014.7159
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114    150.1498
#> 47429 714962.5 9557832 2705.12 32.0575616   0.02224    0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.5755084  -0.00255   -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.8121470   0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866  29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169    350.1799
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200  -0.00034    0.00294    739.5768
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964    571.1364
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771    674.3859
#> 421   714892.5 9561022 1849.01 26.1320321   0.00002    0.00588    749.2502
#> 4617  714292.5 9559312 2342.95 32.3921689   0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357    473.4709
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.4028314  -0.00140   -0.02330   1456.9246
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468   0.00062   -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.6885978   0.00114    0.03026    247.0894
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354   1147.7384
#> 6845  713412.5 9559172 2123.32 47.1538536  -0.01556   -0.04394    587.8840
#> 16009 715122.5 9559402 2264.81 50.0186426  -0.00589    0.00849   1120.5961
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322   4605.1411
#> 19832 715182.5 9557662 2761.27 38.4328630  -0.00666   -0.03583   1898.0105
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245    629.5983
#> 26428 713042.5 9560332 1895.15 33.9597815  -0.01439   -0.01161   3065.3694
#> 34164 714212.5 9558492 2421.45 51.3295700   0.02247    0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 19756 715572.5 9558972 2485.95 38.8127340   0.00546    0.01484    495.0292
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981   3975.3608
#> 10617 713232.5 9560722 1879.95 26.4970062   0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340   0.01003    0.01196   1052.5341
#> 22301 713722.5 9560712 1913.37 39.2711002   0.00705   -0.00915   2036.0378
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140   1066.3957
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610   1872.9276
#> 16065 715242.5 9560622 1871.61 34.3064209  -0.00019    0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.3218698  -0.00157   -0.00152    850.4633
#> 37115 715612.5 9559202 2358.15  9.7276774  -0.01020    0.00279  39495.1797
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856   1789.1596
#> 14779 712502.5 9560202 2004.61 27.5226006   0.00115    0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.9548275  -0.00770   -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.5225656  -0.01454    0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801  -0.00486    0.02306   1718.3652
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781  29995.1543
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150    503.5590
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.6714021  -0.01909   -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.1858831  -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.8171290  -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.4317521   0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.2536948  -0.00124    0.00204    853.2902
#> 35852 715112.5 9559962 2173.61 39.3157909  -0.00738    0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210   0.00165    0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 37832 712552.5 9559192 1998.10 41.5508993  -0.00195   -0.00995   1026.7981
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333    422.8840
#> 44396 715512.5 9558102 2845.31 26.5984197   0.01263    0.03598    186.3527
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399   1601.4852
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483   1097.3082
#> 14800 712952.5 9558682 2120.89 38.2735807  -0.00022    0.00662   1269.2650
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759   4131.5356
#> 33666 713632.5 9561012 1775.88 21.1788119  -0.00950   -0.05130    107.5831
#> 7983  713512.5 9560432 1907.38 50.0192155   0.03310   -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599  12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260   1484.6593
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964    543.0025
#> 9007  714952.5 9560902 1863.75 32.7703211   0.00074   -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.0318136   0.01600   -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.7668833  -0.00375    0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.4651303   0.00105   -0.01576   1045.1074
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.7737308   0.00452    0.00368   1071.8734
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233    968.7780
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.5342555  -0.00163   -0.00377   1727.6055
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884    672.2466
#> 32342 714862.5 9559622 2294.74 36.9672369   0.01021    0.00109    915.4989
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.7319259   0.07379   -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769