Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032    994.6740
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276   9614.6924
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077    300.7915
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964    571.1364
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684   2556.8049
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534    344.1447
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336   2806.2644
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109    605.1908
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419    761.5427
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268    252.0072
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836    250.5111
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491    261.7371
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491