Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 24072 715282.5 9557602 2837.46 34.394083  -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804    628.5153
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677   1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378    440.9104
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993  10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.366589  -0.01017    0.00197    711.1091
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 28845 714892.5 9559272 2374.42 45.750107  -0.00209    0.02739    365.7626
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395   2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245   6353.3179
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325    595.0426
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230    714.7432
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736    477.8591
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376    629.8236
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162    290.2031
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809   1158.0426
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858    298.4490
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983