Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041     772.9746
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804     628.5153
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730    0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164     895.6606
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367     793.9665
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898     588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598    0.01212     561.1948
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437    1155.1469
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405    5681.2412
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991    1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294   -0.00096     520.5570
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373   16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175     804.5679
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964     571.1364
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026     247.0894
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316     454.9674
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220    8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058    1817.2371
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310   -0.00860     695.2302
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141    4611.0718
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292   -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767   -0.03713    6379.7974
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171    4425.5410
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319    0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586    0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203    0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181    1286.3949
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116    1817.5222
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363     389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995    1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945     259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069   -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084   -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223    1056.3701
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785    0.00645     234.0114
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750    1892.8978
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119     792.3544
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803    1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172     351.5812
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946