Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203    634.3320
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023 -0.00043   1806.4552
#> 42733 715112.5 9559062 2399.70 36.0545152 -0.00041  0.01741    746.4155
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333 -0.01767   3321.6113
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257    357.1436
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.1589821 -0.00854  0.01264   5057.1431
#> 1393  714832.5 9557902 2664.72 31.2955277  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281   1325.6138
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 38279 714192.5 9558612 2325.14 28.2525489 -0.04430 -0.06070 138869.3906
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 20194 714132.5 9558452 2447.22 24.0224015 -0.00987  0.02107   1392.5880
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669 -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937  0.00083    994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185  0.00252  0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437   1155.1469
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616  0.01276   9614.6924
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 19128 714852.5 9557872 2681.49 20.7576880  0.00453  0.01077    300.7915
#> 29667 713892.5 9559712 2239.25 41.5927252  0.00804  0.00346    543.0749
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511 361979.2812
#> 46201 714832.5 9561022 1845.94 34.5739922 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504   1020.2766
#> 17036 712832.5 9560582 1951.38 27.6457866  0.00060  0.00270   4057.5637
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058    431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081 -0.00613  0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 16015 712582.5 9560312 1899.18 56.5469237  0.03480  0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953 -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.2512354 -0.00221  0.00121   4159.6738
#> 7668  713372.5 9559062 2164.82 37.9349627 -0.01807 -0.01883    616.1993
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 34192 712702.5 9560102 1867.01 35.4746819  0.00629 -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180   3904.5254
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865  0.00335    195.0293
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.2154742  0.03890 -0.00911    173.2806
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517 -0.02143 -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.7327506 -0.02070 -0.02661   2508.5374
#> 15981 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542 -0.00722   1373.4147
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879   1200.4210
#> 10692 712712.5 9561022 2005.85 41.4334430  0.01410 -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047 -0.00605  0.00075   2170.2249
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121   1783.0658
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 37115 715612.5 9559202 2358.15  9.7276774 -0.01020  0.00279  39495.1797
#> 14779 712502.5 9560202 2004.61 27.5226006  0.00115  0.01015    790.2421
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306   1718.3652
#> 43536 713782.5 9558732 2281.50 32.7491217  0.00840  0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.6500280  0.01996  0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632 -0.00585 -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505    351.4078
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736    477.8591
#> 18286 714792.5 9557882 2650.57 28.2536948 -0.00124  0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.1973563  0.04336  0.01403    241.2991
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 30267 714442.5 9557512 2612.30 28.7659191 -0.00458  0.00479    604.0975
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794   1363.5321
#> 45071 712652.5 9560032 1900.80 14.4734869  0.00987 -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035   1471.9783
#> 14800 712952.5 9558682 2120.89 38.2735807 -0.00022  0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292  0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.6414056 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759   4131.5356
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950 -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342  0.02552   2515.0752
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547    128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417   1520.6630
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 18246 713902.5 9561342 1818.52 19.1104343 -0.01502  0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.7139475  0.01182 -0.00972    331.7462
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284 -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.4521619 -0.01196 -0.01935    903.6397
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.6572738 -0.01305  0.00585   5714.3267
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.6210169 -0.02110 -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801  0.00208  0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769