Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.4464113 -0.00347  0.02357   1752.0375
#> 24072 715282.5 9557602 2837.46 34.3940835 -0.02191 -0.00579   2246.7725
#> 21939 713642.5 9558712 2299.09 33.7506519 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 42917 714202.5 9557412 2544.08 39.8778625 -0.02104 -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098  0.00453  0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588   1295.1610
#> 16811 714902.5 9559262 2362.67 50.7033271 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088    266.4355
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.6120312 -0.02894 -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.9571264 -0.00542  0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.0168746 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.1078353 -0.00017 -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029 -0.02312   2657.6936
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257  0.00593    939.0466
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325    616.8652
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064 -0.00916   2009.1686
#> 28577 712812.5 9559962 1849.82 59.7801245  0.00646 -0.04086   6042.5757
#> 47011 713262.5 9559722 2003.72 37.1929823 -0.01846 -0.00034   4495.6929
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509  0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.8898455 -0.01993  0.01373  16031.7500
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752 -0.00517   1000.7119
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307 147470.5312
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.1730683  0.00359 -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495    266.5888
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.5451417  0.00576 -0.00166    650.6246
#> 44825 714322.5 9557882 2534.80 40.6885978  0.00114  0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 38197 712522.5 9560322 1942.04 62.7772667 -0.01611  0.02961   1086.6963
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013   4229.5708
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.3688733 -0.00093  0.00664   1807.2659
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 7041  713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278 -0.00412   1350.3611
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171    788.0004
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574    183.3220
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299  0.00141   4611.0718
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.2527305 -0.00007 -0.02132   1580.1621
#> 26428 713042.5 9560332 1895.15 33.9597815 -0.01439 -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298   1533.7925
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 13259 714862.5 9560992 1859.28 21.2676204 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957    495.4453
#> 14192 715162.5 9558122 2772.18 44.4626708 -0.02970  0.00200   1305.6691
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296  0.00426   1315.5334
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 7409  714772.5 9559122 2404.90 30.2155023 -0.01016 -0.00294    734.9908
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.8087514 -0.01881  0.00121   1783.0658
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610    454.8103
#> 18498 714842.5 9557712 2679.04 28.9395253  0.00902  0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118 -0.00252   3631.7017
#> 4479  713982.5 9557812 2399.24 41.2311889  0.01895  0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210    585.8992
#> 17322 712612.5 9560402 1924.48 43.8077801 -0.00486  0.02306   1718.3652
#> 41310 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.9840063 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427    866.9398
#> 32405 713022.5 9558982 2208.85 15.0046187  0.00599  0.00301    338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433 -0.02177    997.3089
#> 15186 713322.5 9561212 1936.88 36.0797253  0.00165  0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 15558 714732.5 9559502 2283.66 53.4317521  0.00390 -0.02831    454.9113
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815 -0.03345 665381.5625
#> 12716 713962.5 9557762 2402.60 18.9322444  0.03480  0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199    836.4196
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170    253.5129
#> 33175 715822.5 9558832 2563.13 36.6228893 -0.00674 -0.00376    629.8236
#> 47059 713822.5 9558272 2413.20 38.1750320 -0.00687 -0.00223   1056.3701
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 12519 713982.5 9560912 1956.98 28.8249337 -0.00452 -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.0930780 -0.00305 -0.00705    968.9738
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 44823 714202.5 9557442 2548.13 43.2302386 -0.00233  0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.1971887 -0.00873 -0.01417   1520.6630
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511  0.01499    617.3855
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 26283 715932.5 9557592 3075.48 32.7668833 -0.00375  0.00215    917.5296
#> 6732  715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.9643862 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716 -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962 -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.5718820 -0.00463  0.00253   1237.0536
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389