Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345    1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783     377.8918
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842     868.6137
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501    3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921   0.00832    0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939     703.7817
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199     196.4080
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728    1137.2450
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252    3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291     293.4231
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230    1044.6443
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067    1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476    1275.9446
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077    8759.9961
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308     422.5022
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750    1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438    1276.5448
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624    1870.8307
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585    5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968   0.00732    0.01858     298.4490
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249     148.1857
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806