Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183   -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056   -0.06976    301.2347
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890   2301.0454
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679    550.0344
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416    947.4878
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024    394.8751
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766 140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328    592.3326
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626   4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491    0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051    0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613   -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114    150.1498
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669   -0.00991   1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355  14919.6260
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796   4729.6382
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025   4425.5151
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181  11655.4316
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166    650.6246
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485   2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845    0.00454   3476.0818
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596    842.0994
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448   1258.8196
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589    0.00849   1120.5961
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132   1580.1621
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522   1355.7436
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001   1184.7952
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075   2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185    953.8614
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538    807.8237
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341   1363.4000
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291    293.4231
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400   1982.7782
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840    0.00189    556.5211
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519    0.00609   4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181   1286.3949
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118   2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452    578.5847
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445    355.5078
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548    532.8210
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273  72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502   1428.8546
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367   3002.8828
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065    0.00475    515.1661
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298    436.1459
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011    201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032    994.9224
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162    290.2031
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282    561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360   4288.7329
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417   1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499    617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260   1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177   1558.2231
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805   7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238    787.0053
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014   -0.00884    672.2466
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732    0.01858    298.4490
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931   1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250