Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029    5577.3916
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056   -0.06976     301.2347
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097    0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514   -0.00644    1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433    1475.3191
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679     550.0344
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491    1844.4353
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942    0.00214    0.00466     724.0713
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831   -0.00149   12632.7793
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398     267.2659
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189    2317.1819
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356    0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088    2322.1746
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205    0.02205     629.4165
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097    0.00156     816.6036
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086    6042.5757
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566   -0.03355   14919.6260
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063    5035.0000
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452     437.3965
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175     804.5679
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613    0.00903    2360.1536
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395    0.01495     266.5888
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107    0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560     977.4719
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236    2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197     711.1091
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427     380.9472
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161   -0.00409    2312.8413
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739     365.7626
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927   -0.00316     454.9674
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611    0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925    0.01784    0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771   -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242     507.9263
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169   -0.03220    8936.2666
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417    0.00516    0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589    0.00849    1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281    1178.4420
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212   46310.0195
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722    1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522    1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011    0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213   -0.01363     565.5207
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396     704.7076
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778    4594.4351
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181   -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774   -0.00224    1137.3363
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140    1066.3957
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864    0.04226     338.3474
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152    1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000  162178.1562
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770   -0.03400    1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186    0.00063    2430.2278
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405   -0.05655     130.9468
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320     911.5911
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378    0.02382     100.0000
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534    0.01476    1275.9446
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181    1286.3949
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466    2633.9666
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500     538.8209
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476   11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527   -0.02077    8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943     371.5514
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601   -0.00011     712.5228
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902     218.7467
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068   -0.00312     593.7773
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251     168.1733
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165     805.1403
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360    0.00750    1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066     306.0146
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578    1258.8865
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070   -0.00050    1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375    0.00215     917.5296
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278   -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983    0.00803    1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222     991.8300
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933   -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811   -0.02281     492.4785
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491