Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102    483.4184
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492    772.2691
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769