Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919   -0.04051 351155.1250
#> 4965  715042.5 9559312 2320.49 42.85782 -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.97661  0.00958    0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.95233 -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.97397  0.02056   -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633   -0.01813   1130.5791
#> 47223 715302.5 9557472 2857.45 39.53351 -0.01021   -0.01579   1318.7560
#> 40756 714022.5 9558862 2331.20 45.08548 -0.00075    0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.27793  0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041    0.01741    746.4155
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373   -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191   -0.00579   2246.7725
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118    0.00761    928.6858
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310    0.01440   1088.8639
#> 32622 714972.5 9557762 2687.38 30.55641 -0.00730   -0.01491   1844.4353
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068    0.00588   1295.1610
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223    0.01783    377.8918
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664   -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528    0.01392   1255.0884
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191   -0.00261    554.9243
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905   -0.01385    674.0721
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.67717  0.00426   -0.00367    793.9665
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856    0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716    0.03696   4143.4019
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071    0.00150   1274.3594
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616    0.01276   9614.6924
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383    0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414    0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765    0.00114    150.1498
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743    0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545    0.02405   5681.2412
#> 11687 714552.5 9560382 2052.66 29.57551 -0.00255   -0.00365   1536.2647
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463    0.00167   1938.5389
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752   -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316    0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066   -0.00036    516.4709
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149    0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008    0.00102    483.4184
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286   -0.00964    571.1364
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708   -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759   -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409    0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449   -0.00181   2406.1252
#> 421   714892.5 9561022 1849.01 26.13203  0.00002    0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917   -0.01993  10308.6406
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260   -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017    0.00197    711.1091
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211    0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140   -0.02330   1456.9246
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681   -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160    0.01690   5121.6924
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306   -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161   -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529    0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032    0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736    0.01684   2556.8049
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099    0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168   -0.04102  35685.0000
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195   -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568   -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104    0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410    0.03620    379.0855
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784    0.00596    842.0994
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394   -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291   -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771   -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604    0.00354   1147.7384
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605   -0.03594  16587.9980
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567   -0.03793    864.9472
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865    0.00335    195.0293
#> 10467 714642.5 9559652 2249.12 37.14313 -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489    0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.10642  0.00516    0.01574    183.3220
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878   -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299    0.00141   4611.0718
#> 29228 712782.5 9560422 1870.56 32.90611 -0.00218   -0.00372   4154.5786
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466   -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542   -0.00722   1373.4147
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247    0.01123    364.2903
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109    0.00298   1533.7925
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011    0.00739    789.7573
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253    0.00753   1337.7854
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070    0.00010   3467.2690
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003    0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051    0.02190    264.6404
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042    0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495   -0.03756    353.9086
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774   -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767   -0.03713   6379.7974
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341    0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110   -0.02450   2589.9910
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157   -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216   -0.02606    919.4241
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488   -0.00632    620.0229
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770   -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454    0.00144 317260.6875
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983    0.01253    528.1924
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010   -0.00870   2099.7930
#> 43536 713782.5 9558732 2281.50 32.74912  0.00840    0.00189    556.5211
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599    0.00301    338.9676
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433   -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226    0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909   -0.01181   1286.3949
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533    0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402   -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363   -0.00363    389.2691
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815   -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197    0.00463    851.6306
#> 441   714312.5 9560972 1916.59 31.30298  0.00540    0.00650   1133.1501
#> 42329 715852.5 9558702 2651.92 22.87076  0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712    0.00923    811.5580
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263    0.03598    186.3527
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341    0.04431    315.9084
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007   -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174   -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973   -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457    0.01697    412.1185
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601   -0.00011    712.5228
#> 33956 714082.5 9558712 2327.95 39.68420  0.02218    0.00902    218.7467
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112   -0.00038   2017.3547
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586    0.00654   1576.2065
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418    0.00513    161.9878
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.98875 -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858    0.00312    254.0656
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804   -0.00165    805.1403
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262   -0.00502    520.5218
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074   -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.73999  0.01354    0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343    0.01177   1558.2231
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909   -0.01151 143213.1562
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788   -0.01029    894.2672
#> 33881 712992.5 9558822 2173.26 33.19775 -0.00595    0.03064   1485.5201
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751    0.00388    180.4414
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379   -0.00249    148.1857
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505    0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 38668 32.922728   300.00   TRUE       300.00            100  3.022236
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983