Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 13748 715232.5 9559542 2212.12 26.12344  0.00023 -0.00043   1806.4552
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588   1295.1610
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032    994.6740
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838
#> 23141 715632.5 9558942 2495.99 34.37689  0.01275  0.00605    798.6775
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504   1020.2766
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025   4425.5151
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903   2360.1536
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469   1021.5407
#> 40216 714602.5 9560592 2015.55 34.36887 -0.00093  0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220   8936.2666
#> 1973  715252.5 9557822 2861.00 15.57070  0.01865  0.00335    195.0293
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860    695.2302
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621
#> 10689 713542.5 9559132 2211.95 31.98136  0.05379  0.00801    172.7491
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298   1533.7925
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257  0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.37746  0.00730 -0.02610   1872.9276
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.87118 -0.00212  0.02732   4237.6934
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036   2505.8369
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603   1562.0503
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.65003  0.01996  0.04684    209.1778
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.67673  0.00706 -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099   1146.2140
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.17881 -0.00950 -0.05130    107.5831
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165    805.1403
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972    331.7462
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607