Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896   -0.00501   -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260   -0.00123    0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659    0.02056   -0.06976     301.2347
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023   -0.00043    1806.4552
#> 24051 715462.5 9559372 2319.46 32.3279977   -0.00661   -0.01339    1002.5877
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010    1527.1028
#> 26447 715632.5 9559102 2410.77 39.6567008    0.00730    0.01770     474.5059
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175   -0.08226    2325.2053
#> 42598 713542.5 9559972 2184.39 35.2414881   -0.00841   -0.00010    1527.1028
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842     868.6137
#> 40307 713302.5 9560702 1845.07 39.0046749    0.00129    0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270   -0.00769    1032.7177
#> 29420 713592.5 9560772 1826.82 34.9269342   -0.00070   -0.00530    9000.4356
#> 18005 714832.5 9557912 2660.76 31.9120303    0.00191   -0.00261     554.9243
#> 22933 712712.5 9560432 1902.17 38.1234658   -0.00746    0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430   -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.1698683   -0.00905   -0.01385     674.0721
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151    0.00899    1394.2274
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713   -0.00173    8967.1924
#> 36230 713322.5 9560662 1798.99 40.1282451   -0.00033   -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.2829086    0.00356    0.00774     527.1068
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987    0.02107    1392.5880
#> 1635  713102.5 9559052 2172.87 31.4840308   -0.00064   -0.00916    2009.1686
#> 19128 714852.5 9557872 2681.49 20.7576880    0.00453    0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.0050732    0.03696    0.02514     297.7842
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047   -0.00887     366.8537
#> 20001 712702.5 9560132 1854.55 36.2911467   -0.00204   -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458    0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752   -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922   -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321    0.01169     350.1799
#> 46171 712822.5 9558932 2089.30 29.8545389   -0.01174    0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288    0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552   -0.02058     431.4362
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492    0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585   -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290    0.00230    1056.5876
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395    0.01495     266.5888
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283    0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626    0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372   -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399   -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119   -0.01560     977.4719
#> 2595  714912.5 9557662 2700.72 24.1948618   -0.02681   -0.01319    3457.7519
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221    0.00121    4159.6738
#> 7668  713372.5 9559062 2164.82 37.9349627   -0.01807   -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126   -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.0199280   -0.00568   -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168   -0.04102   35685.0000
#> 17728 713092.5 9560652 1900.14 30.2791006    0.00645   -0.00395    2011.3270
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478   -0.07132   42970.7891
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538    0.00112     745.3315
#> 39722 715022.5 9557712 2735.53 30.2607659    0.00133    0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556   -0.04394     587.8840
#> 46461 712682.5 9560032 1892.30 22.9778994    0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194   -0.03986   -0.02524   10861.9482
#> 20368 715252.5 9560622 1865.83 33.5724620    0.00025   -0.01245    6353.3179
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521   -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890   -0.00911     173.2806
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.1804123   -0.01059   -0.07431    4079.0366
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263    0.01322    4605.1411
#> 19832 715182.5 9557662 2761.27 38.4328630   -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218   -0.00372    4154.5786
#> 1036  713042.5 9558892 2205.38 31.5098139   -0.00428   -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924    0.02606     564.6876
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246    0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070    0.00010    3467.2690
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317    0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.8904255    0.00042    0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.4626708   -0.02970    0.00200    1305.6691
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605    0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.5663341   -0.00693    0.01133    2397.2681
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881    0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110   -0.02450    2589.9910
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915   -0.02705     413.5702
#> 7771  715232.5 9560922 1816.88 37.1803772    0.00124    0.01166    1813.7841
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144   -0.00856    1789.1596
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529    0.01230    1044.6443
#> 9864  714952.5 9557592 2752.40 31.9893796   -0.00983    0.01253     528.1924
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010   -0.00870    2099.7930
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319    0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.7000149    0.00683   -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.6127367    0.00614   -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.2034632   -0.00585   -0.02665    1904.4769
#> 17374 715712.5 9557952 2832.22 37.8748021   -0.01519    0.00609    4275.3457
#> 29431 714882.5 9559892 2258.03 21.9179275   -0.00849    0.00150     503.5590
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909   -0.01181    1286.3949
#> 31071 713472.5 9558462 2298.33 29.1532385    0.00156   -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.1973563    0.04336    0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295    0.02715     520.7355
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402   -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.6402737    0.00363   -0.00363     389.2691
#> 30267 714442.5 9557512 2612.30 28.7659191   -0.00458    0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.0369281   -0.03220   -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195   -0.00995    1026.7981
#> 43243 715682.5 9558462 2693.23 27.4062902   -0.00031   -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.2208476   -0.04937   -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818    0.00502    1428.8546
#> 32849 714302.5 9558222 2520.01 32.1572563   -0.00846    0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069   -0.00900     614.5898
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737    0.01033    3446.5520
#> 37029 713022.5 9559082 2163.08 36.0321698    0.00490   -0.00780     825.4451
#> 47059 713822.5 9558272 2413.20 38.1750320   -0.00687   -0.00223    1056.3701
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636  197.75496 -197.75366     267.5057
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.1180924   -0.01322   -0.00759    4131.5356
#> 17746 714752.5 9558322 2648.52 47.7365517   -0.00068   -0.00312     593.7773
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978    0.01282    2322.7944
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667    0.00863    1098.3462
#> 42424 714692.5 9560902 1921.34 25.3310371    0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009   -0.00599   12873.8018
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788    0.00119    0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209    0.00980   -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635   -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851   -0.02608    0.00148   10771.1963
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502     520.5218
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511    0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967    0.00944    0.00836     250.5111
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074   -0.02904     608.3561
#> 14763 713462.5 9560852 1830.40 47.7399894    0.01354    0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745    0.00875    2156.9358
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129    0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412    0.01572    0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865   -0.00805    7489.7417
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196   -0.01935     903.6397
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309    0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.0800956   -0.00172   -0.00119     792.3544
#> 37254 713482.5 9560892 1823.73 39.8824462    0.01014   -0.00884     672.2466
#> 6577  714142.5 9559902 2244.30 57.8905097   -0.06986   -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.3459270    0.00811   -0.02281     492.4785
#> 15880 715892.5 9557572 3113.36 53.0432868    0.00962   -0.00172     351.5812
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166    0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181   -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607