Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096   3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212    561.1948
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083    994.7876
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294    734.9908
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650    766.6254
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 23925 37.533319   300.00   TRUE       300.00            100  2.938827
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 8066  23.527366   300.00  FALSE       300.00            100  2.834598
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607