Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752   1321.5997
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177    291.5937
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055    773.0359
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911    173.2806
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298   1533.7925
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456   1531.4274
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036   2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566  12862.4922
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234   1707.3130
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525    162.0592
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491