Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599    1378.5760
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393    8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336