Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498 -0.00752    1321.5997
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948    1538.1741
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607