Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029    5577.3916
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752    1321.5997
#> 34512 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174    2310.4871
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345    1120.6929
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010    1527.1028
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035    2697.5974
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120     512.9740
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388    1051.4222
#> 21703 713842.5 9559672 2282.58 24.99185  0.01581  0.01189     287.5425
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530    9000.4356
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261     554.9243
#> 22566 712832.5 9559842 1897.49 33.06024  0.00672  0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838     238.6685
#> 34275 715332.5 9558442 2650.77 45.53697 -0.00043  0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696    4143.4019
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917     116.1986
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150    1274.3594
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205     629.4165
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077    3559.4204
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318     289.0063
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146    1347.6224
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887     366.8537
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991    1403.0193
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499     885.0064
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504    1020.2766
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407     344.8395
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925     665.6116
#> 41037 715532.5 9558792 2529.58 32.84652 -0.00585 -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102     483.4184
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495     266.5888
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400     608.6523
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092     412.5849
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197     711.1091
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357     473.4709
#> 7466  712962.5 9560392 1819.03 18.40283 -0.00140 -0.02330    1456.9246
#> 34005 714922.5 9558842 2559.23 52.83129  0.03658  0.04602     345.8933
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121    4159.6738
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009    1247.7061
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304    1569.7313
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558    1504.9717
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224     942.0322
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599    1378.5760
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474    2343.2100
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367     806.0966
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008  825875.6875
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171     788.0004
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911     173.2806
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261    1627.8989
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322    4605.1411
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441     424.2680
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396     704.7076
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756     353.9086
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140    1066.3957
#> 44612 714932.5 9558822 2578.80 38.25525  0.01257  0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185     953.8614
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152    1530.5657
#> 18498 714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291     293.4231
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632     620.0229
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.38124 -0.00465 -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063    2430.2278
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603    1562.0503
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093     118.1402
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996    1599.3491
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505     351.4078
#> 48601 713692.5 9560152 2058.14 59.55724 -0.01433 -0.02177     997.3089
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517    1060.5494
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116    1817.5222
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715     520.7355
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345  665381.5625
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.82301 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995    1026.7981
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171     756.8818
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973     215.5101
#> 1085  715722.5 9557532 3097.22 28.78425  0.02327  0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476   11992.5273
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077    8759.9961
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507     504.3085
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223    1056.3701
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997     230.8641
#> 20135 714772.5 9560232 2107.52 37.85074 -0.00782 -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032     994.9224
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654    1576.2065
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705     968.9738
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417    1520.6630
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577     423.5598
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836     250.5111
#> 34336 713792.5 9559072 2288.95 36.93286 -0.00952 -0.00438    1276.5448
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849     764.8193
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208     842.4825
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809    1158.0426
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858     298.4490
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222     991.8300
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172     351.5812
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227     841.2025
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769