Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203    634.3320
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410    0.00118   0.00761    928.6858
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387   1989.3012
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 10974 715312.5 9557502 2844.23 26.664883   -0.00831  -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720   0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905  -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059  -0.00628    556.0121
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450    873.0448
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118   2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276   9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365   1536.2647
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499    885.0064
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517   1000.7119
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613    289.6113
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063  18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199    196.4080
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068    0.00359  -0.00279    930.0871
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329  -0.01181  11655.4316
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048  -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690   5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299    688.0706
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126  -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117   -0.02554   0.03304   1569.7313
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219    420.7032
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291  -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627  -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832   3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556  -0.04394    587.8840
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813 287033.4688
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129   0.00171    788.0004
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556  -0.11694 102301.4609
#> 16009 715122.5 9559402 2264.81 50.018643   -0.00589   0.00849   1120.5961
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338   0.00728   1137.2450
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175   1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132   1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075   2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705    413.5702
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609   4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109    605.1908
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226   0.00736    477.8591
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493   2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363  -0.00363    389.2691
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875   1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290   2481.5632
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161   0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028   2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 47005 714942.5 9559192 2333.66 29.220848   -0.04937  -0.02273  72123.6172
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431    315.9084
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007  -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338  -0.00399   1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175    725.1355
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267  31588.1953
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282   2322.7944
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417   1520.6630
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074  -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952  -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050   1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576   1045.1074
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884    672.2466
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931   1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250