Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976     301.2347
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250    1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677    1343.5006
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392    1255.0884
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314    2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205     629.4165
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779    2972.0193
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332    1418.6622
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473    2930.2791
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664    1898.5588
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594   16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664    1807.2659
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322    4605.1411
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.045704  -0.00789   -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363     565.5207
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196    1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077    3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044     288.5468
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 30626 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945     259.7421
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367    3002.8828
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400    4126.7812
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547     128.2613
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499     617.3855
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047    2826.7700
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576    1045.1074
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 18498 22.7177766   300.00   TRUE       300.00             83  2.904587
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769