Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633   -0.01813    1130.5791
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644    1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949    6217.8213
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175   -0.08226    2325.2053
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129    0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471   -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746    0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382    0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973     514.5488
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521   -0.00739     558.2286
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088    2322.1746
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596    5296.9014
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416    1204.8629
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124    2114.2253
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613     289.6113
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072   54831.6289
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119   -0.01560     977.4719
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057    0.02427     380.9472
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543     352.1558
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114    0.03026     247.0894
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927   -0.00316     454.9674
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771   -0.00599    1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865    0.00335     195.0293
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171     788.0004
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728    1137.2450
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774   -0.00224    1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110   -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157   -0.00152     850.4633
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020    0.00279   39495.1797
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603    1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367    1966.2112
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759    0.00158     837.3730
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298   -0.00712    1010.5097
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738    0.00118    2273.4758
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759    1880.4403
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737    0.01033    3446.5520
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301   -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997     230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268     252.0072
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487    1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251     168.1733
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099     909.0560
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044   -0.00964     543.0025
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537   -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247    0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775