Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483    566.7674
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916   2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672    565.1107
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025   4425.5151
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199    196.4080
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964    571.1364
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684   2556.8049
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102  35685.0000
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354   1147.7384
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822    472.8748
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911    173.2806
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981   3975.3608
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196   1052.5341
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 13278 712752.5 9560452 1901.25 42.113544  -0.00118   -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290   2481.5632
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109    915.4989
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573   1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983