Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383  0.00193   1809.3838
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405   5681.2412
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862    874.5408
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530    937.1306
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594  16587.9980
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394    587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208    951.3401
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904    608.3561
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119    792.3544
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157    766.0220
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901