Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>         cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42789   300.00   TRUE        15.00              9  3.746431
#> 36052 53.74516   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527   300.00   TRUE       300.00            100  3.000471
#> 21939 23.93245   300.00   TRUE       300.00            100  3.168886
#> 32622 30.17138   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870   173.40   TRUE        33.81            100  3.516154
#> 31023 16.67766   135.00   TRUE         0.00              2  2.577367
#> 22814 34.91204   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13374   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46551   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690    41.43   TRUE         1.90            100  3.349380
#> 23925 37.53332   300.00   TRUE       300.00            100  2.938827
#> 33792 51.20810   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32130   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460   300.00   TRUE       300.00             12  3.970546
#> 18304 44.90843   300.00   TRUE       300.00             57  3.589919
#> 34482 32.97143   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93532   300.00   TRUE       300.00              4  3.635091
#> 35942 27.16450   300.00   TRUE       300.00            100  3.364960
#> 38070 34.23308   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83161   300.00   TRUE       300.00             10  3.268382
#> 45575 25.72810   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470   300.00   TRUE       300.00              2  3.276843
#> 48364 32.07590   300.00   TRUE       300.00            100  2.972687
#> 13847 20.57090   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814   125.47   TRUE         0.00              6  3.006344
#> 23141 30.29171   300.00   TRUE       300.00            100  2.902371
#> 12084 34.59576   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14886   300.00   TRUE         0.00              0  3.754443
#> 47429 32.09767   300.00   TRUE       300.00            100  2.464778
#> 18564 21.70020   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564   300.00   TRUE       300.00             10  2.544291
#> 29926 32.09480   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92825    25.33   TRUE       208.61             15  3.608265
#> 37141 12.38849   300.00   TRUE       300.00              1  2.293159
#> 26567 40.43707   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35492   300.00   TRUE       300.00            100  2.756740
#> 20220 32.17559   300.00   TRUE       300.00              8  3.050346
#> 21053 34.25142   300.00   TRUE         0.00              2  3.760524
#> 26698 34.24970   300.00   TRUE       300.00              7  2.767615
#> 2595  30.43953   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564   300.00   TRUE       300.00            100  3.236197
#> 39732 29.16814   300.00   TRUE       300.00             25  3.364146
#> 47100 29.64598    82.80   TRUE       129.73             36  3.558076
#> 30886 39.18516   300.00   TRUE       300.00            100  2.695380
#> 21013 24.96320   300.00   TRUE        13.47             64  3.177528
#> 24501 27.15247   300.00   TRUE       300.00              0  2.925363
#> 42765 42.35304   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64693   215.68   TRUE         0.00            100  3.354565
#> 7385  28.71034   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753   300.00   TRUE       300.00            100  2.872349
#> 12137 27.58849   300.00   TRUE       300.00            100  3.369811
#> 39166 29.81844   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59617   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98976   300.00   TRUE       300.00              5  2.674746
#> 11612 35.16700   300.00   TRUE       300.00            100  2.516921
#> 19892 27.02642   300.00   TRUE       300.00             92  3.259412
#> 7086  34.20443   300.00   TRUE       128.61             63  3.071308
#> 15432 28.76764   300.00   TRUE       300.00            100  2.880810
#> 10689 30.64809   300.00   TRUE       300.00             61  2.237416
#> 28009 27.01782   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699   300.00   TRUE       300.00            100  3.525504
#> 11352 27.99816   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619   300.00   TRUE       300.00            100  3.126386
#> 28173 23.88375   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79821    17.89   TRUE       145.83              4  3.064162
#> 38668 32.92273   300.00   TRUE       300.00            100  3.022236
#> 16071 29.62135   300.00   TRUE       300.00            100  2.201225
#> 49378 33.99416    22.68   TRUE       134.63             25  3.073643
#> 14192 40.58661   300.00   TRUE       300.00            100  3.115833
#> 4667  31.51726    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77833   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40860   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23593   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592   300.00   TRUE       300.00            100  2.699406
#> 9606  33.82399   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23193   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76156   300.00   TRUE       300.00             31  2.907317
#> 46627 53.68328   300.00   TRUE       300.00            100  2.720428
#> 28588 31.62555   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90975   300.00   TRUE       300.00            100  3.627130
#> 47722 33.11581   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172   300.00   TRUE       300.00             55  2.355626
#> 6212  31.93151   300.00   TRUE       300.00             47  5.501416
#> 30378 28.56939   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34574   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81368   300.00   TRUE       205.28              1  3.147505
#> 41310 17.29530   300.00   TRUE       300.00             25  3.046207
#> 17374 43.14544   300.00   TRUE       300.00            100  3.630971
#> 4785  26.48612   300.00  FALSE         0.00            100  3.050663
#> 36561 25.80487   300.00  FALSE        70.00            100  2.990220
#> 27884 15.49736   300.00  FALSE       300.00            100  2.781892
#> 15558 50.34408   300.00  FALSE       300.00             45  2.657927
#> 18286 24.35071   300.00  FALSE       300.00             42  2.931097
#> 18746 39.13646     4.48  FALSE       205.00             27  3.028048
#> 26963 42.06312   300.00  FALSE       300.00            100  3.420610
#> 8066  23.52737   300.00  FALSE       300.00            100  2.834598
#> 43426 35.46265   300.00  FALSE        82.11            100  3.910718
#> 37832 36.44642   300.00  FALSE       207.18              2  3.011485
#> 39193 31.48002   300.00  FALSE       117.12            100  2.879028
#> 49835 32.36123   300.00  FALSE         0.00            100  3.134665
#> 47005 36.13530   300.00  FALSE       300.00            100  4.858077
#> 18627 31.25714   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14332   300.00  FALSE         1.11             21  3.274260
#> 43403 31.06176   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546   300.00  FALSE       300.00             24  2.615022
#> 44668 30.10263     9.15  FALSE       163.30              6  2.852799
#> 33956 38.59558   300.00  FALSE       300.00            100  2.339942
#> 47059 25.75961   300.00  FALSE       300.00            100  3.023816
#> 1691  27.00006   300.00  FALSE       300.00            100  2.855262
#> 29107 25.23077   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70615   300.00  FALSE       300.00             92  2.369237
#> 34345 44.39506   300.00  FALSE       300.00            100  2.972060
#> 45486 33.42177   300.00  FALSE         0.00             68  3.366011
#> 5656  31.27261   300.00  FALSE        94.43            100  2.853361
#> 21070 34.67712   300.00  FALSE         0.00             10  3.400551
#> 42424 22.60720   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26533   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860   300.00  FALSE       144.19            100  2.958793
#> 3724  25.72695   300.00  FALSE       300.00            100  2.450415
#> 484   32.14580   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979   300.00  FALSE       300.00            100  3.451290
#> 31098 34.65764   300.00  FALSE       300.00             45  3.277234
#> 42638 23.42137    86.41  FALSE       172.46            100  6.734528
#> 18308 30.02070   300.00  FALSE       300.00             55  4.224567
#> 280   29.32341    60.55  FALSE       127.54             29  3.333837
#> 10981 29.16642   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88279   300.00  FALSE        25.55             33  3.874467
#> 1434  39.69280   300.00  FALSE       300.00             35  2.955995
#> 14881 26.85625   300.00  FALSE       300.00            100  2.948550
#> 31440 24.03501   300.00  FALSE       300.00             96  2.654870
#> 36141 27.30602   300.00  FALSE       300.00             85  4.412069
#> 30403 36.13072   300.00  FALSE       209.91            100  2.884241
#> 15880 48.89679   300.00  FALSE       300.00            100  2.546026
#> 46567 16.94809   300.00  FALSE       300.00             78  2.785235
#> 40043 23.99662   194.79  FALSE        24.89             54  2.973603