Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766  140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392    1255.0884
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032     994.6740
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774     527.1068
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 13699 712902.5 9559902 1917.21 30.487657  -0.00093   -0.00907    1943.4092
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694   0.00290    0.00230    1056.5876
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684    2556.8049
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102   35685.0000
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240    1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469    1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793     864.9472
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061    5744.7583
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731  -0.00007   -0.02132    1580.1621
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363     565.5207
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422    1159.2100
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133    2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094    1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038  -0.01125   -0.02685    2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341    1363.4000
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210     585.8992
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996    1599.3491
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446   0.00570    0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479     604.0975
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399    1601.4852
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983     270.7949
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011     201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196     576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552    2515.0752
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650     766.6254
#> 49831 715262.5 9557332 2884.30 22.294934  -0.00097    0.01487    1769.3392
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710     472.0236
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578    1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576    1045.1074
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124   25826.6914
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 37416 28.771649    72.23   TRUE        87.56             31  3.407698
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983