Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041    772.9746
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284   2010.4285
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089 -0.00051    779.0555
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416   1204.8629
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267 -0.05063  18733.8848
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925    665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224    942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870   1895.1399
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797   7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833    300.0125
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399   1601.4852
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662   1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148    481.8292
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149   4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607