Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043    1806.4552
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507  159065.0781
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728   0.00250   -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412  -0.00730   -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783     377.8918
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.155530  -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678    1891.6584
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407     344.8395
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 2976  715122.5 9559142 2390.05 33.167954   0.01008    0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393  -0.00286   -0.00964     571.1364
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731    1384.7959
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588     749.2502
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241     585.6182
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602     345.8933
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427     380.9472
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972  -0.02027   -0.01473    2930.2791
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397   0.00448   -0.00328     592.3326
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109     767.4097
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180    3904.5254
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474    2343.2100
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667  -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245     629.5983
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 38668 715582.5 9558242 2735.55 35.446034   0.01003    0.01196    1052.5341
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426    1315.5334
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185     953.8614
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477    2945.2100
#> 42640 714332.5 9558522 2403.10 38.871176  -0.00212    0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400    1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336    2806.2644
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712    1010.5097
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345  665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132     737.5982
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171     756.8818
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943     371.5514
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462     716.5762
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038    2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499     617.3855
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836     250.5111
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576    1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491     261.7371
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585    5714.3267
#> 32342 714862.5 9559622 2294.74 36.967237   0.01021    0.00109     915.4989
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603