Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034   -0.01919  -0.04051  351155.1250
#> 15762 714922.5 9558862 2535.97 54.9105562    0.01305  -0.00555     430.4604
#> 19963 714612.5 9560552 2031.70 26.8482293    0.00046   0.00324     861.3537
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498  -0.00752    1321.5997
#> 24129 714602.5 9560542 2038.30 25.5848574    0.00065   0.00005     769.3426
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273  -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 27973 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310   0.01440    1088.8639
#> 25169 712642.5 9560232 1883.10 46.8450293    0.00884   0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.3616837   -0.04951  -0.00529    3282.1186
#> 125   713992.5 9561062 1878.28 49.9710871    0.00183   0.03387    1989.3012
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021  -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542   0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194   0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979  -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.9918461    0.01581   0.01189     287.5425
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234     849.0547
#> 18005 714832.5 9557912 2660.76 31.9120303    0.00191  -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430  -0.06070  138869.3906
#> 23375 713162.5 9559632 2041.52 46.8152355   -0.00857   0.03677    1675.4679
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151   0.00899    1394.2274
#> 6808  715192.5 9559492 2219.03 21.8176599    0.00713  -0.00173    8967.1924
#> 34565 712972.5 9560172 2007.72 27.9471624    0.00598   0.01212     561.1948
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937   0.00083     994.7876
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252   0.01118    2359.5059
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047  -0.00887     366.8537
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566  -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294  -0.00096     520.5570
#> 34359 712862.5 9558912 2111.37 32.8121470    0.00463   0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.7021461   -0.01771  -0.01939     703.7817
#> 34563 715022.5 9558162 2675.70 18.8898455   -0.01993   0.01373   16031.7500
#> 31316 715022.5 9559332 2308.28 20.0632631    0.00940   0.00679     550.0344
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267  -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996   0.00626    1014.7159
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 41037 715532.5 9558792 2529.58 32.8465245   -0.00585  -0.00974    2136.6939
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290   0.00230    1056.5876
#> 4031  715392.5 9560162 1998.16 46.2858225    0.00708  -0.02538     807.8237
#> 22299 713532.5 9560362 1975.58 27.7099578    0.00478   0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.7531604   -0.00119  -0.01560     977.4719
#> 49569 713232.5 9561122 1868.73 37.3121575   -0.00405  -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.2645950   -0.00211   0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.4028314   -0.00140  -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.1948618   -0.02681  -0.01319    3457.7519
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090  -0.00299     688.0706
#> 35083 714612.5 9558152 2596.12 60.5983082    0.04548   0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856  -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126  -0.00485    2240.4185
#> 10364 714762.5 9561002 1856.21 25.2748872   -0.00845   0.00454    3476.0818
#> 35014 713502.5 9559772 2190.48 44.2357795    0.00927   0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.0591409   -0.00410   0.03620     379.0855
#> 40270 713312.5 9559012 2186.56 30.9357102   -0.00771  -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.7342388   -0.00481  -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228    2262.3796
#> 37769 714412.5 9558562 2382.08 39.8767166   -0.04601  -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188   0.00832    3022.3462
#> 12772 715172.5 9557642 2759.42 34.8054672   -0.00567  -0.03793     864.9472
#> 33910 715202.5 9559572 2195.66 13.7916671   -0.04376  -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491   0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521  -0.00870    1895.1399
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338   0.00728    1137.2450
#> 37720 715722.5 9557962 2820.42 45.0763086   -0.01474   0.01334    4905.2002
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263   0.01322    4605.1411
#> 14545 713562.5 9559602 2222.17 54.2390497    0.01011   0.00739     789.7573
#> 47086 713692.5 9559972 2175.51 25.0457041   -0.00789  -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253   0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.6858936   -0.00124  -0.00456    1531.4274
#> 49378 713252.5 9560732 1875.41 27.8354356    0.00181  -0.00001    1184.7952
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292  -0.00109    2047.9031
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257   0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.6461919   -0.02110  -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185  -0.01185     953.8614
#> 29024 713492.5 9559112 2173.47 53.4323251    0.04864   0.04226     338.3474
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157  -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.3361937    0.00181  -0.00341    1363.4000
#> 17227 714812.5 9558892 2545.43 54.0717460    0.00458   0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.4885514   -0.02312  -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.6859287   -0.00488  -0.00632     620.0229
#> 33142 715652.5 9558212 2740.94 14.3812407   -0.00465  -0.01036    2505.8369
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319   0.00291     487.3438
#> 23832 714252.5 9560182 2188.12 30.6475125    0.00427   0.02403     474.8613
#> 17851 715362.5 9557482 2869.26 29.1257366   -0.03251  -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.6801711    0.00000   0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.8748021   -0.01519   0.00609    4275.3457
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084  -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081    0.00378   0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763   -0.00079  -0.00331     699.3024
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196   0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002  -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908    0.00534   0.01476    1275.9446
#> 15186 713322.5 9561212 1936.88 36.0797253    0.00165   0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.4890128   -0.00226   0.00736     477.8591
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124   0.00204     853.2902
#> 11111 715142.5 9558482 2650.65 42.3432999   -0.00533   0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.5904544   -0.00812  -0.02208     951.3401
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112   0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466    2633.9666
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197   0.00463     851.6306
#> 9665  713362.5 9558812 2275.28 31.4949170    0.01306  -0.00626     450.7346
#> 28162 714062.5 9561152 1785.47 48.4447275    0.02309  -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.8707563    0.00321   0.00548     532.8210
#> 49835 713272.5 9560372 1898.68 32.8258980    0.00846   0.00794    1363.5321
#> 44396 715512.5 9558102 2845.31 26.5984197    0.01263   0.03598     186.3527
#> 34347 714742.5 9560142 2161.95 26.5669070   -0.00978   0.00367    3002.8828
#> 34616 714522.5 9558742 2568.77 34.9979810    0.00990   0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.2796736    0.00084  -0.00083    1309.2347
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844   0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272    0.01180   0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.6119821   -0.03193  -0.00267   31588.1953
#> 33956 714082.5 9558712 2327.95 39.6842028    0.02218   0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.8963648    0.00358   0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.2168579   -0.01292   0.00462     716.5762
#> 33370 714202.5 9558962 2472.93 32.6414056   -0.00307   0.02997     230.8641
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948    1222.5433
#> 41855 713182.5 9560452 1851.53 45.1782951    0.00667   0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.1838220    0.00270  -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.3936922    0.03479   0.02251     168.1733
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497   -0.00930  -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534    0.00165   0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.2842642    0.00804  -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632    0.01635  -0.00105     868.7802
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247     282.1076
#> 32430 715832.5 9558112 2824.39 23.6940967    0.00944   0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.9010116    0.03044  -0.00964     543.0025
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273  -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952  -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070    0.01852   0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.0174405   -0.02552   0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.0041019    0.00745   0.00875    2156.9358
#> 6732  715482.5 9559232 2376.03 29.9055321   -0.01137  -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129   0.00491     261.7371
#> 7132  713192.5 9559612 2060.27 20.7737308    0.00452   0.00368    1071.8734
#> 22628 714942.5 9557732 2696.19 26.2116732    0.00788  -0.01029     894.2672
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595   0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052     451.7209
#> 36141 714222.5 9559422 2316.64 16.2771580   -0.07627  -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.9672369    0.01021   0.00109     915.4989
#> 26635 715092.5 9560412 2010.59 30.5718820   -0.00463   0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305  -0.01135    1293.4855
#> 6206  714122.5 9560992 1917.28 37.9699131   -0.01023  -0.00227     841.2025
#> 47629 714472.5 9558482 2437.33 37.2726234   -0.01181  -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607