Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150   1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964    571.1364
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175   1122.9120
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169    495.8840
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395   2011.3270
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555   4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399   3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694 102301.4609
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118  19557.6621
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140   1066.3957
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419    761.5427
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463    851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333    422.8840
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077   8759.9961
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216   1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376    629.8236
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162    290.2031
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013    616.4128
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249    148.1857
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983