Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537 -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899    1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545  0.02405    5681.2412
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064     683.2795
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182 -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983