Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea   cslope
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029    5577.3916 34.42789
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051  351155.1250 32.81444
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203     634.3320 30.29457
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606 20.47808
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487    1077.7585 28.28349
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752    1321.5997 39.59826
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426 24.12897
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299 42.22183
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431    4079.0366 31.66451
#> 47255 713552.5 9559962 2191.77 33.57876 -0.00068  0.00588    1295.1610 19.67766
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088     266.4355 21.61140
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388    1051.4222 29.30507
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149   12632.7793 32.46551
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328    2235.5259 14.89690
#> 18843 715192.5 9559332 2265.75 41.15327 -0.07664 -0.05766  140591.4062 28.52642
#> 15109 713872.5 9559602 2288.64 25.18551  0.00602  0.01518     854.4269 22.63527
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032     994.6740 42.46649
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261     554.9243 24.90590
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677    1675.4679 34.63988
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710     843.7961 29.19965
#> 2942  713232.5 9560652 1851.75 34.53503  0.00856  0.01114    4510.8833 29.82073
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696    4143.4019 34.87537
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284    2010.4285 34.62155
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358     802.7622 22.71778
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314    2360.1501 29.48326
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678    1891.6584 40.08470
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694    2123.4473 36.48022
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357     275.7474 18.83427
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437    1155.1469 20.57090
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156     816.6036 28.58085
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193    1809.3838 27.18914
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916    2009.1686 27.82684
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416    1204.8629 30.91852
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924    8023.8657 34.59576
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146    1347.6224 37.58832
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034    4495.6929 44.53429
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866   29195.6016 39.28772
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939     703.7817 29.93590
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376    2048.3198 34.67712
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070    2562.7559 27.20919
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178     800.1933 39.47851
#> 46171 712822.5 9558932 2089.30 29.85454 -0.01174  0.00025    4425.5151 29.07131
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626    1014.7159 16.21814
#> 18394 715022.5 9559942 2216.16 42.13761 -0.00613  0.00903    2360.1536 28.61065
#> 41037 715532.5 9558792 2529.58 32.84652 -0.00585 -0.00974    2136.6939 31.57055
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279     930.0871 24.82626
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102     483.4184 26.88203
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779    2972.0193 27.70251
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117    1018.2141 33.13472
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400     608.6523 38.99608
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771     674.3859 40.09158
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092     412.5849 21.71911
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516     576.0335 23.33714
#> 10507 714892.5 9559282 2374.45 42.84235 -0.01057  0.02427     380.9472 27.16221
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401    1722.6481 31.44564
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009    1247.7061 32.54114
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739     365.7626 29.84251
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542    3614.7334 29.64598
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543     352.1558 43.54078
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169     495.8840 39.18516
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109     767.4097 30.84232
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020    1563.5549 38.16014
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224     942.0322 42.35304
#> 34192 712702.5 9560102 1867.01 35.47468  0.00629 -0.00469    1021.5407 33.87498
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354    1147.7384 37.94871
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112     745.3315 28.72753
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367     806.0966 26.54170
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822     472.8748 29.98976
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213     328.7922 35.16700
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524   10861.9482 39.15422
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250 32.67349
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171     788.0004 25.76133
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211   49347.5547 29.82131
#> 6217  714442.5 9559932 2151.30 36.67044  0.01489  0.00241    1752.9242 30.97181
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849    1120.5961 33.41318
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911     173.2806 19.84840
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450 41.01060
#> 41125 712772.5 9559412 2021.88 38.87633 -0.00310 -0.00860     695.2302 36.55585
#> 15981 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983 30.00351
#> 28009 715002.5 9559922 2237.68 34.60379 -0.01109  0.00298    1533.7925 27.01782
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456    1531.4274 23.88375
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077    3056.9475 23.16067
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957     495.4453 23.51934
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426    1315.5334 31.51726
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075    2170.2249 26.77833
#> 750   712812.5 9560032 1847.09 54.95926 -0.00767 -0.03713    6379.7974 34.34710
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000  162178.1562 31.62555
#> 30626 713302.5 9559172 2106.91 35.99779  0.01354  0.00245     629.5983 30.00351
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291     293.4231 43.59750
#> 47722 712732.5 9561022 1999.03 37.49264  0.00004 -0.02384     947.2075 33.11581
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210     585.8992 35.48729
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870    2099.7930 28.56939
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291     487.3438 23.34574
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205    1112.2626 17.29530
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403     474.8613 25.48516
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367    1966.2112 34.06005
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000  0.00320     911.5911 28.09327
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781   29995.1543 34.94470
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161 19.77564
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642    1688.1782 27.41603
#> 32405 713022.5 9558982 2208.85 15.00462  0.00599  0.00301     338.9676 15.76207
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150     503.5590 20.89176
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505     351.4078 14.81841
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204     853.2902 24.35071
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463    1050.7473 43.51156
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118    2273.4758 36.36219
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328    1066.7134 39.13646
#> 7517  714002.5 9561142 1772.85 26.18933  0.00815 -0.03345  665381.5625 22.98477
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132     737.5982 14.60011
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650    1133.1501 25.11331
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077    8759.9961 28.07550
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853    1354.6001 26.67921
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943     371.5514 21.27049
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399    1601.4852 30.94144
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475     515.1661 37.33851
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900     614.5898 31.06176
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697     412.1185 32.55546
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083    1309.2347 24.13757
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362     283.7090 18.62800
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170     253.5129 33.72258
#> 14800 712952.5 9558682 2120.89 38.27358 -0.00022  0.00662    1269.2650 31.52299
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298     436.1459 36.72659
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400    4126.7812 25.85357
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098     164.0802 23.68436
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032     994.9224 15.55466
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714  480535.5000 34.37747
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268     252.0072 19.73782
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948    1222.5433 30.48594
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955     713.4463 31.27261
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552    2515.0752 34.67712
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650     766.6254 38.50620
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868     235.8511 13.57222
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312     254.0656 21.26533
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599   12873.8018 22.87820
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282     561.2366 36.10723
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060     383.3547 31.20672
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165     805.1403 29.93704
#> 14671 712672.5 9560372 1872.21 39.19719 -0.00873 -0.01417    1520.6630 37.88225
#> 17628 714032.5 9560252 2160.03 30.29457  0.00537 -0.00577     423.5598 26.99204
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750    1892.8978 26.14750
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960     884.2003 21.92709
#> 48050 714022.5 9560572 2099.82 18.71395  0.01182 -0.00972     331.7462 23.36121
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523     888.2811 26.85625
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172  111973.8906 23.09593
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238     787.0053 39.52377
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011     501.7867 18.13927
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585    5714.3267 33.56329
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858     298.4490 17.04435
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492     772.2691 33.52663
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661    5642.7544 36.25448
#>       distroad slides distdeforest distslidespast log.carea
#> 31358   300.00   TRUE        15.00              9  3.746431
#> 21788   300.00   TRUE       300.00             40  5.545499
#> 37912   300.00   TRUE         9.15              2  2.802317
#> 42632   300.00   TRUE       300.00             10  2.566509
#> 29391   300.00   TRUE         0.56            100  3.032521
#> 8598    300.00   TRUE       300.00            100  3.121100
#> 24129   300.00   TRUE       300.00             89  2.886120
#> 15843   300.00   TRUE       300.00            100  3.091290
#> 9669    300.00   TRUE       300.00              1  3.610558
#> 47255   300.00   TRUE       260.03            100  3.112324
#> 23817   300.00   TRUE       300.00              4  2.425592
#> 24018   300.00   TRUE       300.00              5  3.021777
#> 10974   300.00   TRUE       300.00            100  4.101499
#> 38287    41.43   TRUE         1.90            100  3.349380
#> 18843   300.00   TRUE       300.00            100  5.147959
#> 15109   300.00   TRUE       300.00             40  2.931675
#> 35285   300.00   TRUE       300.00            100  2.997681
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 36030   300.00   TRUE         0.00            100  2.926238
#> 2942     68.13   TRUE       116.32             10  3.654262
#> 25081   300.00   TRUE       300.00            100  3.617357
#> 6676    300.00   TRUE       300.00            100  3.303289
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 3453    300.00   TRUE       300.00              2  3.276843
#> 47220   300.00   TRUE       300.00            100  3.327041
#> 3133    300.00   TRUE       300.00            100  2.440511
#> 13847   300.00   TRUE       300.00            100  3.062637
#> 7879    300.00   TRUE       300.00            100  2.912011
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 1635    300.00   TRUE       300.00            100  3.303016
#> 27013   300.00   TRUE       168.05             18  3.080938
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 47011   300.00   TRUE       300.00             10  3.652797
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 48434   300.00   TRUE       300.00              6  2.847438
#> 20001   264.46   TRUE        18.17             56  3.311398
#> 36931   300.00   TRUE       300.00            100  3.408707
#> 43092   300.00   TRUE       300.00            100  2.903195
#> 46171   300.00   TRUE       270.22            100  3.645964
#> 39278   125.47   TRUE         0.00              6  3.006344
#> 18394   300.00   TRUE       300.00             57  3.372940
#> 41037   300.00   TRUE       300.00            100  3.329742
#> 44550   300.00   TRUE       300.00            100  2.968524
#> 2976    300.00   TRUE       300.00            100  2.684323
#> 34707   165.13   TRUE        39.56            100  3.473052
#> 39599   300.00   TRUE       300.00             25  3.007839
#> 9704    300.00   TRUE       300.00             90  2.784369
#> 21952   300.00   TRUE       300.00             64  2.828908
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 15766   300.00   TRUE       300.00            100  2.760448
#> 10507   300.00   TRUE       300.00             46  2.580865
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 28845   300.00   TRUE       300.00             54  2.563199
#> 47100    82.80   TRUE       129.73             36  3.558076
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 30886   300.00   TRUE       300.00            100  2.695380
#> 664     300.00   TRUE       300.00             12  2.885027
#> 31121   300.00   TRUE       300.00             45  3.194113
#> 42765   300.00   TRUE       300.00             12  2.974066
#> 34192   273.94   TRUE         4.48             85  3.009256
#> 20601   300.00   TRUE       300.00            100  3.059843
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 11612   300.00   TRUE       300.00            100  2.516921
#> 1424    300.00   TRUE       300.00            100  4.035908
#> 33910   300.00   TRUE       300.00              6  6.024389
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 17345   300.00   TRUE       300.00              2  4.693266
#> 6217    300.00   TRUE       300.00              1  3.243763
#> 16009   300.00   TRUE       300.00            100  3.049449
#> 30287   300.00   TRUE       300.00             65  2.238750
#> 18348   300.00   TRUE       300.00            100  3.055854
#> 41125   300.00   TRUE         0.00            100  2.842129
#> 15981   300.00   TRUE       300.00              0  2.799064
#> 28009   300.00   TRUE       300.00             29  3.185767
#> 28173   300.00   TRUE       300.00             59  3.185096
#> 13259   203.85   TRUE        10.45             74  3.485288
#> 49133   300.00   TRUE       300.00             38  2.694996
#> 4667     24.78   TRUE       172.24             38  3.119102
#> 23626   300.00   TRUE       300.00            100  3.336505
#> 750     300.00   TRUE         0.00              0  3.804807
#> 28588   300.00   TRUE       300.00             78  5.209992
#> 30626   300.00   TRUE       300.00              0  2.799064
#> 17227   300.00   TRUE       300.00            100  2.467494
#> 47722   300.00   TRUE        65.02              1  2.976445
#> 14140   300.00   TRUE       300.00             65  2.767823
#> 30378   300.00   TRUE       300.00             57  3.322176
#> 49346   300.00   TRUE       300.00             29  2.687835
#> 41310   300.00   TRUE       300.00             25  3.046207
#> 23832   300.00   TRUE       300.00              8  2.676567
#> 21490   300.00   TRUE       300.00              1  3.293630
#> 6470    300.00   TRUE       300.00              2  2.959800
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 10145   276.16  FALSE        91.33            100  3.227418
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 29431   300.00  FALSE       300.00             42  2.702050
#> 46165   300.00  FALSE       300.00            100  2.545811
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 14694   300.00  FALSE       300.00            100  3.021498
#> 35852   300.00  FALSE       300.00            100  3.356690
#> 18746     4.48  FALSE       205.00             27  3.028048
#> 7517    144.37  FALSE        60.16            100  5.823071
#> 32869   300.00  FALSE       300.00             78  2.867820
#> 441     294.31  FALSE       152.09            100  3.054287
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 31374   300.00  FALSE       300.00            100  3.131811
#> 31497   300.00  FALSE       154.49            100  2.570019
#> 48183   300.00  FALSE       300.00            100  3.204523
#> 12974   300.00  FALSE       300.00            100  2.711947
#> 43403   300.00  FALSE       111.88             86  2.788585
#> 8831    300.00  FALSE       300.00             24  2.615022
#> 31884   300.00  FALSE       300.00            100  3.117018
#> 29944   300.00  FALSE       300.00             48  2.452873
#> 31201   300.00  FALSE       300.00            100  2.404000
#> 14800   300.00  FALSE       300.00            100  3.103552
#> 29527   300.00  FALSE       300.00            100  2.639632
#> 36245   110.19  FALSE        81.50             20  3.615611
#> 22545   300.00  FALSE       300.00             96  2.215056
#> 34810   300.00  FALSE       300.00            100  2.997789
#> 13908   300.00  FALSE       300.00             96  5.681725
#> 44267   300.00  FALSE       226.30            100  2.401413
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 5656    300.00  FALSE        94.43            100  2.853361
#> 21070   300.00  FALSE         0.00             10  3.400551
#> 7983    300.00  FALSE        57.15             22  2.884583
#> 330     300.00  FALSE       275.17            100  2.372638
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 40909   234.82  FALSE         0.00             43  4.109707
#> 44823   300.00  FALSE       300.00            100  2.749146
#> 34532   300.00  FALSE       300.00             92  2.583601
#> 24819   300.00  FALSE       150.00            100  2.905872
#> 14671   123.04  FALSE        75.11              0  3.182033
#> 17628   300.00  FALSE       205.59            100  2.626915
#> 3868    300.00  FALSE       300.00            100  3.277127
#> 40612    55.73  FALSE       125.03            100  2.946551
#> 48050   300.00  FALSE        72.23            100  2.520806
#> 6732    300.00  FALSE       300.00            100  2.948550
#> 18504   144.96  FALSE         4.48             89  5.049117
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 9320    300.00  FALSE       300.00             93  2.700519
#> 46639   300.00  FALSE       300.00            100  3.756965
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 48157    57.15  FALSE       157.06             45  2.887769
#> 24516   300.00  FALSE       122.13            100  3.751491