Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183  -0.09203     634.3320
#> 42632 714852.5 9557902 2674.80 30.6526691    0.00221   0.00969     368.5606
#> 28223 712742.5 9560482 1927.69 34.6679573   -0.00292   0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759   -0.00075   0.00475    1001.0861
#> 24051 715462.5 9559372 2319.46 32.3279977   -0.00661  -0.01339    1002.5877
#> 28293 713762.5 9558812 2250.98 33.2854102    0.00118   0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073   0.00174    2310.4871
#> 38411 713552.5 9560652 1851.63 46.0050732   -0.00891  -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453   0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310   0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645   0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.5564122   -0.00730  -0.01491    1844.4353
#> 38827 713412.5 9560472 1869.10 16.8713152   -0.00156   0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.5787645   -0.00068   0.00588    1295.1610
#> 1393  714832.5 9557902 2664.72 31.2955277    0.00550  -0.00120     512.9740
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194   0.00526     539.8119
#> 2990  713542.5 9560372 1970.07 31.8879661    0.01903   0.01447     331.4833
#> 43964 714802.5 9558452 2684.33 37.9825182   -0.02123   0.01164     895.6606
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430  -0.06070  138869.3906
#> 20290 714812.5 9561092 1775.73 34.6267043   -0.01537  -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.4003790    0.00151   0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.5369667   -0.00043   0.00973     514.5488
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493   0.00733     598.7983
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383   0.00193    1809.3838
#> 4329  712882.5 9559982 1911.46 26.2317267   -0.00193  -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.6644985    0.06765   0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.7801245    0.00646  -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028   0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865   -0.00440  -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.7576880    0.00453   0.01077     300.7915
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224   0.00177     291.5937
#> 40738 714972.5 9557642 2745.48 42.5426892   -0.00294  -0.00096     520.5570
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547   0.03323     183.1836
#> 34563 715022.5 9558162 2675.70 18.8898455   -0.01993   0.01373   16031.7500
#> 46235 712712.5 9561042 2023.11 41.2770255    0.01458   0.00672     565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000    0.00321   0.01169     350.1799
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288   0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572   -0.00093  -0.00907    1943.4092
#> 13134 714652.5 9559632 2265.33 49.5998104   -0.03552  -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.5997894   -0.00735  -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.8517302   -0.01149   0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492   0.00072   54831.6289
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480   0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855  -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.9680266    0.00107   0.02373     527.3528
#> 835   715052.5 9558212 2707.65 37.7252601   -0.01399  -0.00731    1384.7959
#> 4617  714292.5 9559312 2342.95 32.3921689    0.03504   0.05825     162.6688
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797  -0.00063     576.2819
#> 2595  714912.5 9557662 2700.72 24.1948618   -0.02681  -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160   0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.2512354   -0.00221   0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.3468842   -0.00161  -0.00409    2312.8413
#> 32735 714922.5 9559002 2414.49 40.0646468    0.00062  -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.6885978    0.00114   0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099   0.02169     495.8840
#> 10364 714762.5 9561002 1856.21 25.2748872   -0.00845   0.00454    3476.0818
#> 21013 712742.5 9560032 1845.53 29.3526278    0.00568  -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927  -0.00316     454.9674
#> 28046 714882.5 9558492 2711.77 28.3058340   -0.00131   0.00201     788.0621
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472  -0.00228    2262.3796
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093   0.00664    1807.2659
#> 30268 714172.5 9558612 2328.84 54.9506633    0.03743  -0.05242     507.9263
#> 6845  713412.5 9559172 2123.32 47.1538536   -0.01556  -0.04394     587.8840
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370  -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017   0.01213     328.7922
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509  -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.6704448    0.01489   0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516   0.01574     183.3220
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949   0.00520    3390.5708
#> 21683 714792.5 9561052 1807.84 55.7327506   -0.02070  -0.02661    2508.5374
#> 1036  713042.5 9558892 2205.38 31.5098139   -0.00428  -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.6037860   -0.01109   0.00298    1533.7925
#> 47244 714262.5 9559302 2350.47 17.0592454    0.02213  -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.8127340    0.00546   0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.4970062    0.00258   0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003   0.01196    1052.5341
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495  -0.03756     353.9086
#> 23626 713742.5 9560032 2137.92 39.3049047   -0.00605   0.00075    2170.2249
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767  -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106  -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.8765069    0.00010   0.00140    1066.3957
#> 44612 714932.5 9558822 2578.80 38.2552461    0.01257   0.00044     288.5468
#> 14950 713382.5 9559172 2111.91 48.1066824    0.00915  -0.02705     413.5702
#> 29024 713492.5 9559112 2173.47 53.4323251    0.04864   0.04226     338.3474
#> 30626 713302.5 9559172 2106.91 35.9977924    0.01354   0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.4885514   -0.02312  -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.7239045   -0.00529   0.01230    1044.6443
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966  -0.02234     849.0547
#> 30378 714852.5 9557752 2651.70 41.2982249   -0.00010  -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186   0.00063    2430.2278
#> 34718 714802.5 9558802 2599.22 44.6500280    0.01996   0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025  -0.05345    2417.1389
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084  -0.02996    1599.3491
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196   0.00336    2806.2644
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325  -0.00505     351.4078
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759   0.00158     837.3730
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298  -0.00712    1010.5097
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295   0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112   0.03538     265.4677
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174  -0.01466    2633.9666
#> 8066  715842.5 9558732 2636.38 33.7174203   -0.00494   0.01064     683.2795
#> 1488  714562.5 9560002 2150.93 11.2838308    0.00197   0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463    0.00825   0.00445     355.5078
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540   0.00650    1133.1501
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253  -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818   0.00502    1428.8546
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538  -0.00052    3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333   -0.00007  -0.00853    1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001   0.00759    1880.4403
#> 8413  713242.5 9561232 1967.45 38.1945125    0.00236   0.00483    1097.3082
#> 37029 713022.5 9559082 2163.08 36.0321698    0.00490  -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821   -0.03193  -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263    0.00706  -0.03035    1471.9783
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452  -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.6842028    0.02218   0.00902     218.7467
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413   0.00528     355.3475
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424   0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586   0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452  -0.00948    1222.5433
#> 33666 713632.5 9561012 1775.88 21.1788119   -0.00950  -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118   0.00162     290.2031
#> 7553  714642.5 9559132 2391.59 63.9678730   -0.00066   0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.9332227    0.03418   0.00513     161.9878
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649 100.79199     107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799   0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209    0.00980  -0.00710     472.0236
#> 3724  714322.5 9557402 2603.46 34.5974835    0.02523   0.02247     282.1076
#> 44434 713022.5 9558762 2161.07 31.4542370   -0.00273  -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.6345090    0.00082   0.02578    1258.8865
#> 44216 715772.5 9558402 2710.44 30.4446854   -0.00070  -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.0318136    0.01600  -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787   -0.03757  -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.7040537   -0.00264   0.00624    1870.8307
#> 4269  714162.5 9560572 2087.45  7.3344327    0.00293   0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.8484390   -0.06909  -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.3999736   -0.00865  -0.00805    7489.7417
#> 1577  713322.5 9559742 2027.47 29.5674870   -0.03284  -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784   -0.04770   0.00270   35950.5508
#> 43972 713642.5 9558702 2304.16 24.0235474    0.00340   0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.5342555   -0.00163  -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309   0.01011     501.7867
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595   0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992  -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983   0.00803    1058.5630
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087  -0.00573    1490.0312
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241   0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166   0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.7319259    0.07379  -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.6210169   -0.02110  -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247   0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775