Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739     365.7626
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070  0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981    3975.3608
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035    1471.9783
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308     422.5022
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074 -0.02904     608.3561
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 30268 47.272456   300.00   TRUE       300.00             41  2.705801
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 16065 30.320354   300.00   TRUE       162.92            100  3.818858
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200