Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921  -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090   -0.00199   0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043    1806.4552
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261     323.8441
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339    1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324     861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644    1179.0695
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403    1787.3024
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104  -0.02046    1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010    1527.1028
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677    1343.5006
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378     440.9104
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160    9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678  -0.04272    4955.5586
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382     560.8309
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746   0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430  -0.06070  138869.3906
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917     116.1986
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422   0.02357     275.7474
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605     798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114     150.1498
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924    8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743   0.01313     783.4548
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887     366.8537
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034    4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167    1938.5389
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993   0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463   -0.01752  -0.00517    1000.7119
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613     289.6113
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288   0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907    1943.4092
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925     665.6116
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949    1182.1506
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117    1018.2141
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588     749.2502
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330    1456.9246
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166     650.6246
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473    2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941    1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102   35685.0000
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611   0.02961    1086.6963
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555    4439.0967
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664    1807.2659
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822     472.8748
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058    1817.2371
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746   -0.01878  -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583    1898.0105
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007  -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833     369.3771
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261    3353.5422
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010    3467.2690
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279   39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.036383   -0.03675  -0.01905    6487.0010
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867    2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253     528.1924
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870    2099.7930
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996   0.04684     209.1778
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389   14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316  -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565    1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085    2229.5105
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204     853.2902
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533   0.01493    2212.0525
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712   0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028    2394.7261
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598     186.3527
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367    3002.8828
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790     209.6952
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267   31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400    4126.7812
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224  -0.02036     476.9504
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312     593.7773
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654    1576.2065
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 484   713812.5 9560592 2016.41 44.301097   -0.00719   0.00079    1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273  -0.01047    2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347    1893.3645
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105  -0.01576    1045.1074
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709    1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064    1485.5201
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281     492.4785
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135    1293.4855
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491