Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv      carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029  5577.3916
#> 16435 713512.5 9559092 2166.13 55.973966    0.02056  -0.06976   301.2347
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221   0.00969   368.5606
#> 14254 714852.5 9557882 2680.00 23.335425    0.00479   0.01261   323.8441
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324   861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644  1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677  1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431  4079.0366
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884   0.01345  1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010  1527.1028
#> 43669 712392.5 9560162 2001.90 47.283915   -0.00284   0.01804   628.5153
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068   0.00588  1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264  5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770   474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547   519.4228
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120   512.9740
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416   947.4878
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328  2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048    0.00952   0.01378   440.9104
#> 44328 713222.5 9558932 2227.94 29.155530   -0.00769  -0.00281  1325.6138
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164   895.6606
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898   458.6938
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471  -0.01382   560.8309
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261   554.9243
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838   238.6685
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284  2010.4285
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773   179.3077
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774   527.1068
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694  2123.4473
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257   0.00593   939.0466
#> 38077 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733   598.7983
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437  1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156   816.6036
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642   613.6390
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416  1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114   150.1498
#> 12084 712832.5 9559992 1872.46 56.054753   -0.03184   0.03924  8023.8657
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077   300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146  1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177   291.5937
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379   0.01338 10787.6299
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527   0.04613   289.6113
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174   0.00025  4425.5151
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779  2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117  1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400   608.6523
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771   674.3859
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731  1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236  2715.4138
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408 76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140  -0.02330  1456.9246
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602   345.8933
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121  4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401  1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436  2602.7390
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409  2312.8413
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684  2556.8049
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553  1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102 35685.0000
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195  -0.00656  3036.7383
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395  2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512 13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078   0.00219   420.7032
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132 42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224   942.0322
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472  -0.00228  2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112   745.3315
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474  2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664  1807.2659
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335   195.0293
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412  1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058  1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061  5744.7583
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521  -0.00870  1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489   0.00241  1752.9242
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212 46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236  9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661  2508.5374
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722  1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801   172.7491
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321  1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924   0.02606   564.6876
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261  3353.5422
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970  1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010  3467.2690
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190   264.6404
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915  2036.0378
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605   0.00075  2170.2249
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044   288.5468
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226   338.3474
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610   454.8103
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921   525.3245
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020   0.00279 39495.1797
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230  1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288  2144.8164
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036  2505.8369
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291   487.3438
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367  1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093   118.1402
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849   0.00150   503.5590
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476  1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009   0.00109   605.1908
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433  -0.02177   997.3089
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419   761.5427
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116  1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463  1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452   578.5847
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174  -0.01466  2633.9666
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463   851.6306
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650  1133.1501
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502  -0.00028  2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833    0.01253  -0.00333   422.8840
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818   0.00502  1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833   300.0125
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263   0.03598   186.3527
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507   504.3085
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943   371.5514
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734   335.0773
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475   515.1661
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175   725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780   825.4451
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022   0.00662  1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032   -0.00687  -0.00223  1056.3701
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518  1137.0179
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645   234.0114
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705   968.9738
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955   713.4463
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650   766.6254
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312   254.0656
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547   128.2613
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009  -0.00599 12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480   0.00740   909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199   107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361  1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710   472.0236
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308   422.5022
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750  1892.8978
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177  1558.2231
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972   331.7462
#> 10127 713722.5 9558712 2311.10 19.481138    0.01129   0.00491   261.7371
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233   968.7780
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709  1181.2555
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799  1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986  -0.05013   616.4128
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458   485.7418
#> 23808 715172.5 9558702 2533.50 22.345927    0.00811  -0.02281   492.4785
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011  -0.00419   323.0790
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156   758.4840
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227   841.2025
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931  1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250