Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203    634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393   2081.0056
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712   3636.5303
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385    674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889    362.8363
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678   1891.6584
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593    939.0466
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325    616.8652
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167   1938.5389
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825    162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409   2312.8413
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543    352.1558
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102  35685.0000
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171    788.0004
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236   9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226    338.3474
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306   1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603   1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427    866.9398
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565   1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181   1286.3949
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199    836.4196
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483   1097.3082
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487   1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251    168.1733
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555   1882.0179
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750   1892.8978
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172 111973.8906
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809   1158.0426
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011    501.7867
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931   1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 10213 31.598622   300.00   TRUE       300.00            100  3.351772
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 15211 30.834870   300.00  FALSE       300.00            100  2.844665
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 47059 25.759610   300.00  FALSE       300.00            100  3.023816
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3868  26.147502   300.00  FALSE       300.00            100  3.277127
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250