Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919 -0.04051  351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 16435 713512.5 9559092 2166.13 55.9739659  0.02056 -0.06976     301.2347
#> 42632 714852.5 9557902 2674.80 30.6526691  0.00221  0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555     430.4604
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483     566.7674
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498 -0.00752    1321.5997
#> 34512 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.5848574  0.00065  0.00005     769.3426
#> 47255 713552.5 9559962 2191.77 33.5787645 -0.00068  0.00588    1295.1610
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922  0.01088     266.4355
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414 -0.00104    1641.4587
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.5537501  0.01412  0.00898     458.6938
#> 22933 712712.5 9560432 1902.17 38.1234658 -0.00746  0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838     238.6685
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367     793.9665
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173    8967.1924
#> 18391 715112.5 9559432 2235.02 22.2576915 -0.02651 -0.03629  360379.1875
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694    2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.3887102  0.00071  0.00150    1274.3594
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212     561.1948
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.1148433 -0.01452 -0.01437    1155.1469
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804 -0.01596    5296.9014
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414  0.03416    1204.8629
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924    8023.8657
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887     366.8537
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866   29195.6016
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229 -0.00511  361979.2812
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442 -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.0379275  0.04527  0.04613     289.6113
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093 -0.00907    1943.4092
#> 22332 715482.5 9558922 2475.28 19.9366394 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974    2136.6939
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449 -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.7099578  0.00478  0.01092     412.5849
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 38862 712952.5 9559892 1950.86 37.7218224 -0.00048 -0.00332    1418.6622
#> 25700 715172.5 9559272 2324.49 35.5348425 -0.00529  0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.7269720 -0.02027 -0.01473    2930.2791
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856 -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262    1760.2162
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406 -0.00055     773.0359
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020    1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664    1898.5588
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219     420.7032
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478 -0.07132   42970.7891
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224     942.0322
#> 41721 715182.5 9557562 2784.85 31.7899903 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605 -0.03594   16587.9980
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793     864.9472
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813  287033.4688
#> 19892 714482.5 9560562 2059.63 19.8151724 -0.00332 -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.4909134 -0.06509 -0.01211   49347.5547
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516  0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589  0.00849    1120.5961
#> 47702 714692.5 9559712 2200.51 33.0986259 -0.13038 -0.04212   46310.0195
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949  0.00520    3390.5708
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474  0.01334    4905.2002
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 17058 714032.5 9558502 2402.95 31.3528235  0.01924  0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860 -0.01109  0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118   19557.6621
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456    1531.4274
#> 332   714852.5 9561102 1787.11 45.3582039  0.00540  0.00879    1200.4210
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003  0.01196    1052.5341
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267  0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044     288.5468
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538     807.8237
#> 28260 715312.5 9558302 2681.75 34.4840379 -0.01125 -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181 -0.00341    1363.4000
#> 46792 713862.5 9559672 2272.03 35.4649416  0.00560  0.00610     454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031  0.02451 -0.03921     525.3245
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312 -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.7239045 -0.00529  0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319  0.00291     487.3438
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316 -0.00427     866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479 -0.00203  0.00642    1688.1782
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 748   715102.5 9559692 2301.39 27.0218355 -0.00759  0.00158     837.3730
#> 34250 713822.5 9559142 2316.36 33.7598192 -0.02284  0.01085    2229.5105
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517    1060.5494
#> 14717 715382.5 9558062 2799.05 50.5904544 -0.00812 -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452     578.5847
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502    1428.8546
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052    3508.5591
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069 -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483    1097.3082
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.8637272  0.01180  0.01170     253.5129
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038    2017.3547
#> 44267 712972.5 9559042 2191.01 22.2691506  0.01462  0.00268     252.0072
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097  0.01487    1769.3392
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547     128.2613
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930 -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.2863534  0.00165  0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632  0.01635 -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.5684513  0.00420  0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.6940967  0.00944  0.00836     250.5111
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066     306.0146
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852  0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964  0.00537 -0.00347    1893.3645
#> 15138 713702.5 9561122 1780.68 49.2216583  0.00607 -0.04566   12862.4922
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841     756.0055
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233     968.7780
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309  0.01011     501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788 -0.01029     894.2672
#> 35257 713822.5 9557982 2350.80 29.3079371  0.01549 -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627 -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222     991.8300
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281     492.4785
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388     180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 22933 33.06310253    87.56   TRUE       115.40              2  3.506583
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336