Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315  -0.00156    0.00406   2421.1145
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096    520.5570
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485   2240.4185
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399   3900.3445
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911    173.2806
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400   1982.7782
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 21490 714672.5 9560382 2021.86 32.804126  -0.00783   -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665   1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 15211 715772.5 9558082 2801.07 36.296876  -0.00079   -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158    837.3730
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403    241.2991
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945    259.7421
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 41753 715722.5 9558192 2771.39 26.085050  -0.00196   -0.00175    725.1355
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048   1550.6704
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863   1098.3462
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057    722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775