Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642     671.1807
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393    2081.0056
#> 29391 712842.5 9560152 1930.16 28.095558   0.00753    0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767    3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292    1233.9299
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035    2697.5974
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770     474.5059
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526     539.8119
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160    9344.2930
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769    1032.7177
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070  138869.3906
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 8073  712572.5 9560302 1917.62 52.009098   0.01937    0.00083     994.7876
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051     779.0555
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177     291.5937
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 48434 713352.5 9559042 2176.63 35.702146  -0.01771   -0.01939     703.7817
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495     266.5888
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175    1122.9120
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542     860.8840
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842  -0.00529    0.02009    1247.7061
#> 3418  715372.5 9558492 2601.00 22.627249  -0.01168   -0.04102   35685.0000
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316     454.9674
#> 17728 713092.5 9560652 1900.14 30.279101   0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201     788.0621
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555    4439.0967
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 21680 712802.5 9560462 1893.75 38.734239  -0.00481   -0.00399    3900.3445
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335     195.0293
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205    1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849    1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801    1071.6141
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.829930  -0.03669   -0.01981    3975.3608
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384     947.2075
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234     849.0547
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063    2430.2278
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286   0.00166   -0.01265    1148.6487
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232     977.7320
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085    2229.5105
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199     836.4196
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973     215.5101
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889  -0.00674   -0.00376     629.8236
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751  -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204     618.2027
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 10127 713722.5 9558712 2311.10 19.481138   0.01129    0.00491     261.7371
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253    1237.0536
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156     758.4840
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249     148.1857
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 43669 28.476575    24.22   TRUE         0.00             89  2.798316
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 2950  28.203847   300.00   TRUE       300.00            100  2.891568
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 44550 24.826261   300.00   TRUE       300.00            100  2.968524
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 48461 35.780068   300.00   TRUE       300.00             61  2.928936
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983