Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.662025   0.02240   -0.00250   1413.3312
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 22933 712712.5 9560432 1902.17 38.123466  -0.00746    0.00096   3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114   4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629 360379.1875
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620  -0.00664    0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 5728  713502.5 9559662 2140.27 34.307567  -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365   1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034   4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556    660.3690
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499    885.0064
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169   0.03504    0.05825    162.6688
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166    650.6246
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809   1102.9956
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 6845  713412.5 9559172 2123.32 47.153854  -0.01556   -0.04394    587.8840
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892  -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212  46310.0195
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161   3065.3694
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739    789.7573
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001   1184.7952
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610    454.8103
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150    503.5590
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736    477.8591
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475    515.1661
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 6143  712762.5 9559022 2096.24 29.896365   0.00358    0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528    355.3475
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312    593.7773
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 7553  714642.5 9559132 2391.59 63.967873  -0.00066    0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250   2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 15008 714492.5 9559012 2449.93 29.515348   0.00972    0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215    917.5296
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 11394 713472.5 9561232 1889.48 34.314442   0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172 111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119    792.3544
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458    485.7418
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 20799 24.807927   180.67   TRUE         0.00             16  3.361925
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 47702 33.666427   300.00   TRUE       300.00             86  4.665675
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983