Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.3006984  0.01633   -0.01813    1130.5791
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097    0.01483     566.7674
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023   -0.00043    1806.4552
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266    0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251  0.02240   -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373   -0.00057    1630.3259
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073    0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340     604.1032
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 14456 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607    0.00677    1343.5006
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841   -0.00010    1527.1028
#> 21703 713842.5 9559672 2282.58 24.9918461  0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110  0.00414   -0.00104    1641.4587
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471   -0.01382     560.8309
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672    0.04398     267.2659
#> 31576 713942.5 9558492 2387.12 46.1431560 -0.05634   -0.02626    4316.0977
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440    0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973     514.5488
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033   -0.03917     116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664    0.00694    2123.4473
#> 48364 715582.5 9558222 2746.23 25.0720602  0.01257    0.00593     939.0466
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493    0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225    0.00325     616.8652
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.8039440 -0.01097    0.00156     816.6036
#> 27013 713062.5 9559782 1976.81 41.6322593  0.01414    0.03416    1204.8629
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743    0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514     297.7842
#> 6557  713312.5 9559022 2181.69 27.6343274 -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566   -0.03355   14919.6260
#> 48434 713352.5 9559042 2176.63 35.7021461 -0.01771   -0.01939     703.7817
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862     874.5408
#> 18751 714882.5 9559182 2381.81 39.2991752  0.00066   -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093   -0.00907    1943.4092
#> 47884 715032.5 9557712 2738.31 28.2892182  0.00006    0.00175     804.5679
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267   -0.05063   18733.8848
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735   -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149    0.01949    1182.1506
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372   -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.9680266  0.00107    0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.7968588 -0.00449   -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119   -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002    0.00588     749.2502
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504    0.05825     162.6688
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658    0.04602     345.8933
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458   -0.00401    1722.6481
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090   -0.00299     688.0706
#> 47100 712752.5 9560462 1910.00 41.9118627 -0.00032    0.00542    3614.7334
#> 35083 714612.5 9558152 2596.12 60.5983082  0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.5625261 -0.00856   -0.01625    1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550   -0.00941    1290.7709
#> 664   714452.5 9559852 2174.83 44.8608765  0.00282    0.00109     767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800   -0.06020    1563.5549
#> 16370 713852.5 9558632 2296.97 34.6244125 -0.01972    0.03512   13178.3223
#> 17431 712802.5 9560022 1832.71 41.5251162 -0.00478   -0.07132   42970.7891
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228    2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827 -0.02605   -0.03594   16587.9980
#> 37769 714412.5 9558562 2382.08 39.8767166 -0.04601   -0.08389   47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196  0.00188    0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469    0.02940     987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567   -0.03793     864.9472
#> 1973  715252.5 9557822 2861.00 15.5707010  0.01865    0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.1622383  0.00017    0.01213     328.7922
#> 32879 713912.5 9559672 2238.00 31.9767745 -0.00278   -0.00412    1350.3611
#> 10933 712702.5 9560132 1854.55 36.2911467 -0.00204   -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025   -0.01245    6353.3179
#> 16009 715122.5 9559402 2264.81 50.0186426 -0.00589    0.00849    1120.5961
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059   -0.07431    4079.0366
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474    0.01334    4905.2002
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166    0.01476    1404.4454
#> 43074 713172.5 9559622 2052.40 31.5275756 -0.00193    0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169    0.01441     424.2680
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258    0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051    0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181   -0.00001    1184.7952
#> 1997  714752.5 9561012 1853.19 26.5663341 -0.00693    0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.6461919 -0.02110   -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185   -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730   -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157   -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.3361937  0.00181   -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.1803772  0.00124    0.01166    1813.7841
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212    0.02732    4237.6934
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312   -0.01848 1890804.2500
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400   -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983    0.01253     528.1924
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405   -0.05655     130.9468
#> 36744 714922.5 9557722 2692.39 25.9807712  0.00316   -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586    0.00067    1123.7318
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196    0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.6885908  0.00534    0.01476    1275.9446
#> 16464 715252.5 9559982 2143.24 23.1858831 -0.00298   -0.00712    1010.5097
#> 18746 713292.5 9560952 1890.51 41.8855067 -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088  0.00112    0.03538     265.4677
#> 8066  715842.5 9558732 2636.38 33.7174203 -0.00494    0.01064     683.2795
#> 441   714312.5 9560972 1916.59 31.3029762  0.00540    0.00650    1133.1501
#> 16542 713772.5 9559622 2304.85 30.8102325 -0.00712    0.00923     811.5580
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253   -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.9946898  0.00544   -0.01014     271.4219
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196    0.01734     335.0773
#> 43403 712632.5 9559262 2072.53 34.4863297 -0.00069   -0.00900     614.5898
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601   -0.00011     712.5228
#> 33956 714082.5 9558712 2327.95 39.6842028  0.02218    0.00902     218.7467
#> 34105 715432.5 9558592 2621.73 18.9563087 -0.00301   -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185    0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782   -0.00048    1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080  0.02413    0.00528     355.3475
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379   -0.01018     937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112   -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004    0.00714  480535.5000
#> 21070 712832.5 9560062 1884.42 50.9336562 -0.00342    0.02552    2515.0752
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.3310371  0.01154   -0.00204     618.2027
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480    0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 1281  715102.5 9560032 2120.64 34.6490497 -0.00930   -0.00360    4288.7329
#> 15008 714492.5 9559012 2449.93 29.5153479  0.00972    0.00308     422.5022
#> 3724  714322.5 9557402 2603.46 34.5974835  0.02523    0.02247     282.1076
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360    0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.9676002 -0.00820   -0.00260    1484.6593
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074   -0.02904     608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952   -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070  0.01852    0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070   -0.00050    1471.4346
#> 10127 713722.5 9558712 2311.10 19.4811380  0.01129    0.00491     261.7371
#> 1577  713322.5 9559742 2027.47 29.5674870 -0.03284   -0.00886   43249.3945
#> 16443 714472.5 9559532 2222.55 16.7441186  0.03263   -0.02012     305.3624
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.8585496  0.00309    0.01011     501.7867
#> 22628 714942.5 9557732 2696.19 26.2116732  0.00788   -0.01029     894.2672
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992   -0.01052     451.7209
#> 14645 715072.5 9560262 2034.04 38.1693024 -0.00716   -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.4708598 -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.0432868  0.00962   -0.00172     351.5812
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336    0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.5833482 -0.00247    0.01057     722.3952
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456   -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491