Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv      carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634   500.5027
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958    0.02642   671.1807
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347    0.02357  1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021   -0.01579  1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969   368.5606
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555   430.4604
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917    0.00333   258.6537
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741   746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041   772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339  1002.5877
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752  1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767  3321.6113
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273   -0.02403  1787.3024
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835  2425.3115
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264  5057.1431
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606   564.6876
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842   868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526   539.8119
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269   491.8017
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234   849.0547
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525   -0.11024   394.8751
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272  4955.5586
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382   560.8309
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312  2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672  1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114  4510.8833
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696  4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733   598.7983
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108   -0.00678  1891.6584
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773   179.3077
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917   116.1986
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694  2123.4473
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382    0.00088  2322.1746
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276  9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596  5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916  2009.1686
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077  3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416  1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086  6042.5757
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346   543.0749
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124  2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991  1403.0193
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034  4495.6929
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323   183.1836
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167  1938.5389
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338 10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321    0.01169   350.1799
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907  1943.4092
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552   -0.02058   431.4362
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530   937.1306
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400   608.6523
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181  2406.1252
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399   -0.00731  1384.7959
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588   749.2502
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197   711.1091
#> 48808 715032.5 9557792 2707.02 25.437607   -0.02329   -0.01181 11655.4316
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681   -0.01319  3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160    0.01690  5121.6924
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161   -0.00409  2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090   -0.00299   688.0706
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736    0.01684  2556.8049
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473  2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294   739.5768
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062   -0.00262  1760.2162
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553  1750.4175
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406   -0.00055   773.0359
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611    0.02961  1086.6963
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013  4229.5708
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469  1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320    0.00180  3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354  1147.7384
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469    0.02940   987.8085
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246    0.00534   344.1447
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337   486.7551
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245  6353.3179
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728  1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212 46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860   695.2302
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949    0.00520  3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431  4079.0366
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321  1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441   424.2680
#> 332   714852.5 9561102 1787.11 45.358204    0.00540    0.00879  1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422  1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443    0.01410   -0.00380   601.0843
#> 14192 715162.5 9558122 2772.18 44.462671   -0.02970    0.00200  1305.6691
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140  1066.3957
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526   706.1463
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685  2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921   525.3245
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216   -0.02606   919.4241
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354    0.00245   629.5983
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279 39495.1797
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400   -0.01210   585.8992
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234   849.0547
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205  1112.2626
#> 34718 714802.5 9558802 2599.22 44.650028    0.01996    0.04684   209.1778
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367  1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614   -0.00093   118.1402
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797  7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067  1123.7318
#> 15211 715772.5 9558082 2801.07 36.296876   -0.00079   -0.00331   699.3024
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325   -0.00505   351.4078
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565  1410.8372
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452   578.5847
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466  2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132   737.5982
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479   604.0975
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199   836.4196
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794  1363.5321
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987   -0.00507   504.3085
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367  3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853  1354.6001
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065    0.00475   515.1661
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432   -0.00362   283.7090
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483  1097.3082
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490   -0.00780   825.4451
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170   253.5129
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267 31588.1953
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298   436.1459
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098   164.0802
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782   -0.00048  1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338    0.00518  1137.0179
#> 44267 712972.5 9559042 2191.01 22.269151    0.01462    0.00268   252.0072
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705   968.9738
#> 330   714232.5 9560362 2165.00 18.060776    0.01142    0.00868   235.8511
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947   842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204   618.2027
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923    0.05547   128.2613
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282   561.2366
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740   909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199   107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360  4288.7329
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417  1520.6630
#> 9007  714952.5 9560902 1863.75 32.770321    0.00074   -0.02904   608.3561
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907  1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343    0.01177  1558.2231
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875  2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598    0.02208   842.4825
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186  2751.4626
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161    0.00709  1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523   888.2811
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169   649.9975
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992   -0.01052   451.7209
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858   298.4490
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222   991.8300
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448   941.8503
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156   758.4840
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208    0.00492   772.2691
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586   941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603