Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv        carea
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969     368.5606
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767    3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005     769.3426
#> 15843 714042.5 9558902 2332.30 50.09771 -0.00858 -0.00292    1233.9299
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761     928.6858
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345    1120.6929
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.15898 -0.00854  0.01264    5057.1431
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.20406 -0.02475  0.00665   25318.3555
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032     994.6740
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312    2657.6936
#> 22933 712712.5 9560432 1902.17 38.12347 -0.00746  0.00096    3210.5781
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838     238.6685
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189    2317.1819
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898     588.3761
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325     616.8652
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642     613.6390
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114     150.1498
#> 34642 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077     300.7915
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146    1347.6224
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405    5681.2412
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346     543.0749
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.88150 -0.00666 -0.00504    1020.2766
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907    1943.4092
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495     266.5888
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092     412.5849
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993   10308.6406
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252    5761.3486
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241     585.6182
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332    1418.6622
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401    1722.6481
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739     365.7626
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883     616.1993
#> 23213 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055     773.0359
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013    4229.5708
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081     940.6002
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599    1378.5760
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.24036 -0.01169 -0.03220    8936.2666
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337     486.7551
#> 33910 715202.5 9559572 2195.66 13.79167 -0.04376 -0.01644 1057765.6250
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728    1137.2450
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.07631 -0.01474  0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.48058  0.00171 -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583    1898.0105
#> 5911  713952.5 9561282 1801.16 33.78961  0.00466 -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.81029  0.00542 -0.00722    1373.4147
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606     564.6876
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739     789.7573
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981    3975.3608
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422    1159.2100
#> 10692 712712.5 9561022 2005.85 41.43344  0.01410 -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759     158.9370
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200    1305.6691
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109    2047.9031
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171    4425.5410
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744     226.7911
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867    2247.8748
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230    1044.6443
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.70001  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655     130.9468
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265    1148.6487
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609    4275.3457
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842     246.5161
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476    1275.9446
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158     837.3730
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.17333 -0.00007 -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.88551 -0.00402 -0.01328    1066.7134
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995    1026.7981
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028    2394.7261
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333     422.8840
#> 34347 714742.5 9560142 2161.95 26.56691 -0.00978  0.00367    3002.8828
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514     897.9290
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098     164.0802
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759    4131.5356
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257     549.3252
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863    1098.3462
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361    1011.0248
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308     422.5022
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836     250.5111
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177    1558.2231
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368    1071.8734
#> 1577  713322.5 9559742 2027.47 29.56749 -0.03284 -0.00886   43249.3945
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012     305.3624
#> 43972 713642.5 9558702 2304.16 24.02355  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011     501.7867
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.65727 -0.01305  0.00585    5714.3267
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799    1036.6606
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419     323.0790
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.38164  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931    1812.3820
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057     722.3952
#> 48157 713142.5 9560992 1978.48 44.62138  0.00208  0.00492     772.2691
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 40270 27.5976581   300.00   TRUE       300.00              5  3.139431
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983