Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896    276.0310
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073   0.00174   2310.4871
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035   2697.5974
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528    0.00550  -0.00120    512.9740
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029  -0.02312   2657.6936
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537  -0.00203  10263.2734
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696   4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889    362.8363
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211   0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463   0.00167   1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600    0.00321   0.01169    350.1799
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270   4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006   0.00175    804.5679
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307 147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230   1056.5876
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373    527.3528
#> 421   714892.5 9561022 1849.01 26.132032    0.00002   0.00588    749.2502
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658   0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690   5121.6924
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807  -0.01883    616.1993
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664   1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972   0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904  -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224    942.0322
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599   1378.5760
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567  -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890  -0.00911    173.2806
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372   4154.5786
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321   1170.5904
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739    789.7573
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.906461   -0.00070   0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171   4425.5410
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110  -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705    413.5702
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610   1872.9276
#> 10456 712442.5 9560292 2010.91 26.891774    0.00445   0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152    850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921    525.3245
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186   0.00063   2430.2278
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189    556.5211
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599   0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 40459 712542.5 9559192 1992.60 39.817129   -0.00162  -0.00517   1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463   1050.7473
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825   0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650   1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161   0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544  -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327   0.01833    300.0125
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405   0.00476  11992.5273
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035   1471.9783
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038   2017.3547
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552   2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257    549.3252
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105    868.7802
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 28286 713702.5 9559842 2245.07 27.968935    0.00511   0.01499    617.3855
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552   0.00532  16771.3066
#> 24446 714742.5 9560082 2200.60 28.721801   -0.00598   0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948   1538.1741
#> 22628 714942.5 9557732 2696.19 26.211673    0.00788  -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884    672.2466
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732   0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282   0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157    766.0220
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305  -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855    229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 40270 27.597658   300.00   TRUE       300.00              5  3.139431
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 7553  39.771292   300.00  FALSE       300.00            100  2.739830
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 11394 40.603800    46.04  FALSE        55.00            100  3.187005
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983