Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.9766069   0.00958    0.02642     671.1807
#> 16435 713512.5 9559092 2166.13 55.9739659   0.02056   -0.06976     301.2347
#> 25090 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.6620251   0.02240   -0.00250    1413.3312
#> 28293 713762.5 9558812 2250.98 33.2854102   0.00118    0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.9525357  -0.01073    0.00174    2310.4871
#> 43327 713962.5 9561082 1839.11 63.3616837  -0.04951   -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.2839150  -0.00284    0.01804     628.5153
#> 49372 714862.5 9560982 1863.12 20.9691094  -0.00295    0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.9710871   0.00183    0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.7033271  -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088     266.4355
#> 1393  714832.5 9557902 2664.72 31.2955277   0.00550   -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783     377.8918
#> 39629 714602.5 9558912 2474.89 40.2583065  -0.06175   -0.08226    2325.2053
#> 27375 714292.5 9558982 2458.96 43.6009423   0.00214    0.00466     724.0713
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328    2235.5259
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104    1641.4587
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 43964 714802.5 9558452 2684.33 37.9825182  -0.02123    0.01164     895.6606
#> 15109 713872.5 9559602 2288.64 25.1855058   0.00602    0.01518     854.4269
#> 34482 715152.5 9557642 2753.52  4.9956190  -0.01126   -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203   10263.2734
#> 15340 715452.5 9558852 2523.95 34.1654733   0.00151    0.00899    1394.2274
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889     362.8363
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.2343959   0.00993   -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678    1891.6584
#> 13585 714552.5 9560092 2123.69 57.5289733  -0.03613   -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.2829086   0.00356    0.00774     527.1068
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150    1274.3594
#> 3133  715042.5 9559252 2344.13 29.0208853   0.00422    0.02357     275.7474
#> 17004 713102.5 9559032 2186.79 36.5335079  -0.01383    0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 27013 713062.5 9559782 1976.81 41.6322593   0.01414    0.03416    1204.8629
#> 12868 712532.5 9560332 1913.92 61.9373106  -0.03543   -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616   0.02224    0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.1929823  -0.01846   -0.00034    4495.6929
#> 2858  714292.5 9559332 2335.04 42.3559050   0.04547    0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.8121470   0.00463    0.00167    1938.5389
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511  361979.2812
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.8898455  -0.01993    0.01373   16031.7500
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175     804.5679
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925     665.6116
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.5469237   0.03480    0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 35194 713692.5 9559952 2185.68 28.7968588  -0.00449   -0.00181    2406.1252
#> 22299 713532.5 9560362 1975.58 27.7099578   0.00478    0.01092     412.5849
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310    2833.9001
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.1794410  -0.02797   -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.8312924   0.03658    0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.2122812  -0.01160    0.01690    5121.6924
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558    1504.9717
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395    2011.3270
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.7710827  -0.02605   -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112     745.3315
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389   47022.9688
#> 12772 715172.5 9557642 2759.42 34.8054672  -0.00567   -0.03793     864.9472
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412    1350.3611
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524   10861.9482
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212   46310.0195
#> 4588  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431    4079.0366
#> 37720 715722.5 9557962 2820.42 45.0763086  -0.01474    0.01334    4905.2002
#> 15432 715632.5 9558922 2509.67 35.8276239  -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.4328630  -0.00666   -0.03583    1898.0105
#> 26428 713042.5 9560332 1895.15 33.9597815  -0.01439   -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860  -0.01109    0.00298    1533.7925
#> 28173 714562.5 9560382 2048.26 27.6858936  -0.00124   -0.00456    1531.4274
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077    3056.9475
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426    1315.5334
#> 1997  714752.5 9561012 1853.19 26.5663341  -0.00693    0.01133    2397.2681
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140    1066.3957
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450    2589.9910
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.6859287  -0.00488   -0.00632     620.0229
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867    2247.8748
#> 6212  714592.5 9560362 2034.54 12.5225656  -0.01454    0.00144  317260.6875
#> 33142 715652.5 9558212 2740.94 14.3812407  -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249  -0.00010   -0.00870    2099.7930
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403     474.8613
#> 21490 714672.5 9560382 2021.86 32.8041256  -0.00783   -0.00367    1966.2112
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.5955479  -0.00203    0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.6252092   0.00159   -0.01559     866.5775
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177     997.3089
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419     761.5427
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463    1050.7473
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.6350750  -0.00242    0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463     851.6306
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.7842537   0.02327    0.01833     300.0125
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052    3508.5591
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175     725.1355
#> 31201 714672.5 9560022 2224.47 38.8637272   0.01180    0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376     629.8236
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462     716.5762
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 20764 714542.5 9558152 2602.18 52.0789988   0.02118    0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.9336562  -0.00342    0.02552    2515.0752
#> 35826 712912.5 9560302 1891.22 50.8694849  -0.00229    0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.9376528  -0.00070   -0.00060     383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 26724 714252.5 9560862 1975.16 28.2863534   0.00165    0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148   10771.1963
#> 29716 715272.5 9559992 2144.94 25.7177836   0.00262   -0.00502     520.5218
#> 44434 713022.5 9558762 2161.07 31.4542370  -0.00273   -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438    1276.5448
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875    2156.9358
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151  143213.1562
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935     903.6397
#> 35257 713822.5 9557982 2350.80 29.3079371   0.01549   -0.00169     649.9975
#> 39883 712732.5 9560842 2082.69 25.0909678   0.00732    0.01858     298.4490
#> 32618 715132.5 9559642 2266.36 34.5327393  -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020     673.2861
#> 6206  714122.5 9560992 1917.28 37.9699131  -0.01023   -0.00227     841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 47011 44.5342905   300.00   TRUE       300.00             10  3.652797
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 12772 37.9200658   300.00   TRUE       300.00            100  2.936990
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901