Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183   -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.7898858   -0.04921   -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.2240896   -0.00199    0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.3006984    0.01633   -0.01813   1130.5791
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753    0.01487   1077.7585
#> 5999  715552.5 9558812 2518.03  7.9841669   -0.01863   -0.00507 159065.0781
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240   -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498   -0.00752   1321.5997
#> 15843 714042.5 9558902 2332.30 50.0977107   -0.00858   -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123   -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.8778625   -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.5241098    0.00453    0.00257    357.1436
#> 20151 714602.5 9559922 2184.66 48.3415951   -0.00979   -0.01542    674.0147
#> 38993 714202.5 9558402 2475.75 33.1593594   -0.01607    0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175   -0.08226   2325.2053
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.1555304   -0.00769   -0.00281   1325.6138
#> 3835  714142.5 9558552 2388.42 32.9003825    0.00979   -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.2040642   -0.02475    0.00665  25318.3555
#> 18843 715192.5 9559332 2265.75 41.1532666   -0.07664   -0.05766 140591.4062
#> 16843 714882.5 9558512 2701.78 26.6379538   -0.00270   -0.00769   1032.7177
#> 34482 715152.5 9557642 2753.52  4.9956190   -0.01126   -0.02044 129412.3984
#> 15340 715452.5 9558852 2523.95 34.1654733    0.00151    0.00899   1394.2274
#> 45575 715432.5 9559452 2305.28 24.2343959    0.00993   -0.00413    364.9788
#> 5749  712852.5 9560162 1934.78 23.3887102    0.00071    0.00150   1274.3594
#> 46125 713732.5 9560062 2115.18 37.5934798   -0.04616    0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.8039440   -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383    0.00193   1809.3838
#> 46756 713372.5 9559192 2120.47 46.0050732    0.03696    0.02514    297.7842
#> 11687 714552.5 9560382 2052.66 29.5755084   -0.00255   -0.00365   1536.2647
#> 37348 715632.5 9558182 2743.16 13.6203527   -0.00904   -0.00866  29195.6016
#> 43092 715532.5 9558192 2770.31 36.8331648   -0.00442   -0.00178    800.1933
#> 47091 713382.5 9559182 2122.71 48.0379275    0.04527    0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.0632631    0.00940    0.00679    550.0344
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.3948802   -0.01996    0.00626   1014.7159
#> 3944  715262.5 9558352 2654.20 33.5552733   -0.04492    0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290    0.00230   1056.5876
#> 16015 712582.5 9560312 1899.18 56.5469237    0.03480    0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.3489734   -0.10372   -0.00008 825875.6875
#> 36120 714622.5 9559942 2166.51 32.9863262   -0.02917   -0.01993  10308.6406
#> 375   715622.5 9558922 2510.87 37.3665885   -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.9235105   -0.01364   -0.03357    473.4709
#> 19887 714662.5 9559632 2266.47 45.1794410   -0.02797   -0.00063    576.2819
#> 30406 715312.5 9560152 2029.10 34.1373984    0.00953   -0.02283    590.1536
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160    0.01690   5121.6924
#> 33350 713132.5 9560672 1896.89 24.6950539    0.00306   -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529    0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.2590471    0.00736    0.01684   2556.8049
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294    739.5768
#> 13146 713792.5 9558552 2341.40 29.8253180   -0.00104    0.01664   1898.5588
#> 24501 713302.5 9559142 2125.66 27.1089251    0.01784    0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.7703772    0.00904   -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.5251162   -0.00478   -0.07132  42970.7891
#> 26010 712852.5 9560542 1928.25 30.9781091    0.00165    0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.4746819    0.00629   -0.00469   1021.5407
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605   -0.03594  16587.9980
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.2403562   -0.01169   -0.03220   8936.2666
#> 1973  715252.5 9557822 2861.00 15.5707010    0.01865    0.00335    195.0293
#> 8125  713452.5 9559132 2120.44 42.1404729   -0.05458   -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.9767745   -0.00278   -0.00412   1350.3611
#> 20368 715252.5 9560622 1865.83 33.5724620    0.00025   -0.01245   6353.3179
#> 7     714952.5 9559862 2284.19 16.1064166    0.00516    0.01574    183.3220
#> 34615 713242.5 9560732 1879.60 27.5936474    0.00384    0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263    0.01322   4605.1411
#> 34164 714212.5 9558492 2421.45 51.3295700    0.02247    0.01123    364.2903
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924    0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.9750697    0.00511   -0.01801   1071.6141
#> 39325 714142.5 9558592 2357.72 40.7928761   -0.00246    0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.6858936   -0.00124   -0.00456   1531.4274
#> 10692 712712.5 9561022 2005.85 41.4334430    0.01410   -0.00380    601.0843
#> 14192 715162.5 9558122 2772.18 44.4626708   -0.02970    0.00200   1305.6691
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016   -0.00294    734.9908
#> 1223  713632.5 9559632 2226.22 50.8087514   -0.01881    0.00121   1783.0658
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285   -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.4323251    0.04864    0.04226    338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353    0.00143    0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.8917741    0.00445    0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.3218698   -0.00157   -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.4455031    0.02451   -0.03921    525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808    0.00216   -0.02606    919.4241
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212    0.02732   4237.6934
#> 17674 713882.5 9558562 2339.20 41.0363832   -0.03675   -0.01905   6487.0010
#> 12074 714562.5 9560372 2045.17 26.2873036   -0.00144   -0.00856   1789.1596
#> 47722 712732.5 9561022 1999.03 37.4926392    0.00004   -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893   -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.5226006    0.00115    0.01015    790.2421
#> 48461 714672.5 9559272 2316.84 40.3814924   -0.01966   -0.02234    849.0547
#> 31491 712832.5 9558892 2091.39 29.7004132    0.00186    0.00063   2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476   1404.4454
#> 17374 715712.5 9557952 2832.22 37.8748021   -0.01519    0.00609   4275.3457
#> 31253 712612.5 9559492 1903.67 46.7567938   -0.06953   -0.01797   7557.7866
#> 13072 714862.5 9558932 2448.29 41.0524260   -0.04989   -0.02781  29995.1543
#> 32405 713022.5 9558982 2208.85 15.0046187    0.00599    0.00301    338.9676
#> 36561 715292.5 9560802 1844.64 31.9372405    0.00002   -0.01232    977.7320
#> 18286 714792.5 9557882 2650.57 28.2536948   -0.00124    0.00204    853.2902
#> 14694 715032.5 9559482 2307.44 44.1733271   -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.1194761    0.02295    0.02715    520.7355
#> 14717 715382.5 9558062 2799.05 50.5904544   -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067   -0.00402   -0.01328   1066.7134
#> 26023 715832.5 9557632 3097.56 41.7101816    0.00133   -0.00452    578.5847
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112    0.03538    265.4677
#> 20067 714242.5 9560442 2125.23 38.6402737    0.00363   -0.00363    389.2691
#> 37832 712552.5 9559192 1998.10 41.5508993   -0.00195   -0.00995   1026.7981
#> 39796 714802.5 9560372 2046.48 40.1557471    0.00176    0.02973    215.5101
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759   1880.4403
#> 31497 714032.5 9560672 2060.77 16.2169974    0.00973   -0.00943    371.5514
#> 34616 714522.5 9558742 2568.77 34.9979810    0.00990    0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069   -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.7604272   -0.00457    0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737    0.01033   3446.5520
#> 37029 713022.5 9559082 2163.08 36.0321698    0.00490   -0.00780    825.4451
#> 14800 712952.5 9558682 2120.89 38.2735807   -0.00022    0.00662   1269.2650
#> 47059 713822.5 9558272 2413.20 38.1750320   -0.00687   -0.00223   1056.3701
#> 34105 715432.5 9558592 2621.73 18.9563087   -0.00301   -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041   -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185    0.01655    560.8554
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413    0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.1333947    0.00424    0.00196    576.1450
#> 12519 713982.5 9560912 1956.98 28.8249337   -0.00452   -0.00948   1222.5433
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229    0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511    0.01499    617.3855
#> 484   713812.5 9560592 2016.41 44.3010967   -0.00719    0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298    0.02236   -0.00066    306.0146
#> 34336 713792.5 9559072 2288.95 36.9328595   -0.00952   -0.00438   1276.5448
#> 31098 714352.5 9559042 2410.15 30.3363964    0.00537   -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.0174405   -0.02552    0.00532  16771.3066
#> 15138 713702.5 9561122 1780.68 49.2216583    0.00607   -0.04566  12862.4922
#> 1434  714672.5 9559902 2194.92 42.4521619   -0.01196   -0.01935    903.6397
#> 16443 714472.5 9559532 2222.55 16.7441186    0.03263   -0.02012    305.3624
#> 40653 714262.5 9561262 1791.94 12.2223357    0.00161    0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.8585496    0.00309    0.01011    501.7867
#> 36141 714222.5 9559422 2316.64 16.2771580   -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195    0.00282    0.00458    485.7418
#> 42990 712432.5 9560692 2102.11 32.4500377   -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.1693024   -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.4708598   -0.00011   -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.1106510   -0.00288    0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.5327393   -0.00305   -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.1245906   -0.00166    0.00156    758.4840
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247    0.01057    722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775