Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.24719 -0.00501 -0.00649   1399.2329
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441
#> 42733 715112.5 9559062 2399.70 36.05452 -0.00041  0.01741    746.4155
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250   1413.3312
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.89157  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.66488 -0.00831 -0.00149  12632.7793
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328   2235.5259
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.52897 -0.03613 -0.01898    588.3761
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150   1274.3594
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.48099 -0.00509  0.00499    885.0064
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709
#> 17423 715592.5 9558562 2654.69 33.21952 -0.00034  0.00294    739.5768
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 19887 714662.5 9559632 2266.47 45.17944 -0.02797 -0.00063    576.2819
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 27875 714832.5 9561032 1838.22 41.21228 -0.01160  0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738
#> 13029 713542.5 9560242 2019.80 29.54514  0.00576 -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 30886 714912.5 9558822 2563.86 49.75394 -0.00099  0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.52778  0.00406 -0.00055    773.0359
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.16086 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081    940.6002
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 12772 715172.5 9557642 2759.42 34.80547 -0.00567 -0.03793    864.9472
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534    344.1447
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 22907 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812   1423.7578
#> 19756 715572.5 9558972 2485.95 38.81273  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285 -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 12074 714562.5 9560372 2045.17 26.28730 -0.00144 -0.00856   1789.1596
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487
#> 36744 714922.5 9557722 2692.39 25.98077  0.00316 -0.00427    866.9398
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 42329 715852.5 9558702 2651.92 22.87076  0.00321  0.00548    532.8210
#> 29125 714992.5 9558792 2584.41 24.66583  0.01253 -0.00333    422.8840
#> 47005 714942.5 9559192 2333.66 29.22085 -0.04937 -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961
#> 18627 714292.5 9559252 2330.78 18.49622 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853   1354.6001
#> 32849 714302.5 9558222 2520.01 32.15726 -0.00846  0.00216   1412.2295
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.89636  0.00358  0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 48869 712562.5 9560122 1972.89 24.94658  0.00185  0.01655    560.8554
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 330   714232.5 9560362 2165.00 18.06078  0.01142  0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282    561.2366
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165    805.1403
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.00410  0.00745  0.00875   2156.9358
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 11394 713472.5 9561232 1889.48 34.31444  0.00108 -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128
#> 36141 714222.5 9559422 2316.64 16.27716 -0.07627 -0.00124  25826.6914
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.04329  0.00962 -0.00172    351.5812
#> 126   715042.5 9561262 1733.49 30.11065 -0.00288  0.00448    941.8503
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603