Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555    430.4604
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339   1002.5877
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257    357.1436
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264   5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783    377.8918
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.926934 -0.00070 -0.00530   9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086   6042.5757
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373  16031.7500
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613    289.6113
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779   2972.0193
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304   1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008 825875.6875
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141   4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321   1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422   1159.2100
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391   1794.7964
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189    556.5211
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665   1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736    477.8591
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650   1133.1501
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083   1309.2347
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018    937.6911
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599  12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148  10771.1963
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960    884.2003
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047   2826.7700
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233    968.7780
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029    894.2672
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057    722.3952
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775