Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 37912 712802.5 9559952 1838.40 52.1013441   0.00183   -0.09203    634.3320
#> 27864 712992.5 9560672 1926.38 27.2240896  -0.00199    0.00659   3553.7166
#> 47223 715302.5 9557472 2857.45 39.5335149  -0.01021   -0.01579   1318.7560
#> 16749 715072.5 9558242 2745.27 49.0537816   0.00097    0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.1234377   0.00023   -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475   1001.0861
#> 30059 713832.5 9559662 2288.02 18.2779266   0.01917    0.00333    258.6537
#> 21939 713642.5 9558712 2299.09 33.7506519  -0.00503    0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890   2301.0454
#> 37104 713962.5 9561282 1795.56 25.7418478   0.00273   -0.02403   1787.3024
#> 42917 714202.5 9557412 2544.08 39.8778625  -0.02104   -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406   2421.1145
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677   1343.5006
#> 27375 714292.5 9558982 2458.96 43.6009423   0.00214    0.00466    724.0713
#> 10974 715312.5 9557502 2844.23 26.6648828  -0.00831   -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.2355979  -0.03021   -0.09328   2235.5259
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.4586952  -0.03525   -0.11024    394.8751
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898    458.6938
#> 34482 715152.5 9557642 2753.52  4.9956190  -0.01126   -0.02044 129412.3984
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733   0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367    793.9665
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.2735877   0.00382    0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.3887102   0.00071    0.00150   1274.3594
#> 46125 713732.5 9560062 2115.18 37.5934798  -0.04616    0.01276   9614.6924
#> 2950  715042.5 9557712 2740.72 26.5250811  -0.00089   -0.00051    779.0555
#> 12460 712762.5 9559622 1906.41 24.2618978   0.01047   -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616   0.02224    0.00177    291.5937
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511 361979.2812
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796   4729.6382
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613    289.6113
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.6570711  -0.01267   -0.05063  18733.8848
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.9366394  -0.02617    0.00307 147470.5312
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731   1384.7959
#> 462   713712.5 9561182 1785.08 47.7170711  -0.03260   -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197    711.1091
#> 38862 712952.5 9559892 1950.86 37.7218224  -0.00048   -0.00332   1418.6622
#> 32909 712812.5 9560452 1883.85 38.2512354  -0.00221    0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401   1722.6481
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299    688.0706
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.3604817  -0.00550   -0.00941   1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304   1569.7313
#> 3418  715372.5 9558492 2601.00 22.6272492  -0.01168   -0.04102  35685.0000
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.2357795   0.00927    0.01643    670.2289
#> 11618 713852.5 9558402 2432.14 54.0591409  -0.00410    0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512  13178.3223
#> 7064  715592.5 9558662 2597.69 34.2221961   0.00099    0.00240   1201.4724
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555   4439.0967
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448   1258.8196
#> 37714 713192.5 9560692 1885.05 18.0344196   0.00188    0.00832   3022.3462
#> 18958 714192.5 9558372 2495.72 34.6771246  -0.00178   -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 44767 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205   1112.2626
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.6704448   0.01489    0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.1064166   0.00516    0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.0186426  -0.00589    0.00849   1120.5961
#> 41125 712772.5 9559412 2021.88 38.8763323  -0.00310   -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175   1026.6818
#> 19832 715182.5 9557662 2761.27 38.4328630  -0.00666   -0.03583   1898.0105
#> 43074 713172.5 9559622 2052.40 31.5275756  -0.00193    0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 14545 713562.5 9559602 2222.17 54.2390497   0.01011    0.00739    789.7573
#> 28173 714562.5 9560382 2048.26 27.6858936  -0.00124   -0.00456   1531.4274
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957    495.4453
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.3049047  -0.00605    0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450   2589.9910
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610   1872.9276
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 22847 714582.5 9560382 2040.49 25.3361937   0.00181   -0.00341   1363.4000
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732   4237.6934
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230   1044.6443
#> 43536 713782.5 9558732 2281.50 32.7491217   0.00840    0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315   0.02405   -0.05655    130.9468
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842    246.5161
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177    997.3089
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736    477.8591
#> 15558 714732.5 9559502 2283.66 53.4317521   0.00390   -0.02831    454.9113
#> 11111 715142.5 9558482 2650.65 42.3432999  -0.00533    0.01493   2212.0525
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.6350750  -0.00242    0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.0369281  -0.03220   -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.5508993  -0.00195   -0.00995   1026.7981
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.9966605  -0.01341    0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.7540896   0.00527   -0.02077   8759.9961
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175   0.00174   -0.03065    704.4843
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.0321698   0.00490   -0.00780    825.4451
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263   0.00706   -0.03035   1471.9783
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376    629.8236
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.7365517  -0.00068   -0.00312    593.7773
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282    561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105    868.7802
#> 28286 713702.5 9559842 2245.07 27.9689348   0.00511    0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836    250.5111
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578   1258.8865
#> 44216 715772.5 9558402 2710.44 30.4446854  -0.00070   -0.00050   1471.4346
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491    261.7371
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233    968.7780
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.8905097  -0.06986   -0.05013    616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.9643862  -0.00933   -0.00157    766.0220
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419    323.0790
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131  -0.01023   -0.00227    841.2025
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 6206  32.4632157   294.31  FALSE        35.90            100  2.924901