Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659  0.01041     772.9746
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664 -0.05766  140591.4062
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626    4316.0977
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 48364 715582.5 9558222 2746.23 25.072060  0.01257  0.00593     939.0466
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032  0.00542    3614.7334
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118  0.00162     290.2031
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079    1647.6451
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983