Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501   -0.00649   1399.2329
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555    430.4604
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858   -0.00292   1233.9299
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310    0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730   -0.01491   1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529   3282.1186
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281   1325.6138
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769   1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312   2657.6936
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634   -0.02626   4316.0977
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426   -0.00367    793.9665
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902    0.01358    802.7622
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413    364.9788
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205    629.4165
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193   1809.3838
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804    0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154    0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365   1536.2647
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229   -0.00511 361979.2812
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714    475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169    350.1799
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527    0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036    516.4709
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175    804.5679
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149    0.01949   1182.1506
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495    266.5888
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286   -0.00964    571.1364
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409    0.02771    674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008 825875.6875
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092    412.5849
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993  10308.6406
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408  76318.0156
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602    345.8933
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941   1290.7709
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081    940.6002
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320    0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354   1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228   2262.3796
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389  47022.9688
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178   -0.00822    472.8748
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129    0.00171    788.0004
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141   4611.0718
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474    0.01334   4905.2002
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079    0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247    0.01123    364.2903
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246    0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051    0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181   -0.00001   1184.7952
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267    0.00483    343.1384
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121   1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226    338.3474
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341   1363.4000
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610    454.8103
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606    919.4241
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252   3631.7017
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015    790.2421
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870   2099.7930
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603   1562.0503
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665   1904.4769
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953   -0.01797   7557.7866
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109    605.1908
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517   1060.5494
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715    520.7355
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875   1165.3953
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075   8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995   1026.7981
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548    532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794   1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527   -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007   -0.00853   1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338   -0.00399   1601.4852
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457    0.01697    412.1185
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292    0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041   -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518   1137.0179
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112   -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714 480535.5000
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130    107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310   -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667    0.00863   1098.3462
#> 330   714232.5 9560362 2165.00 18.060776  0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547    128.2613
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148  10771.1963
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944    0.00836    250.5111
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904    608.3561
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572    0.03148    481.8292
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270  35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935    903.6397
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523    888.2811
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124  25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379   -0.00249    148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403    609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 25090 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 22299 21.7191111   300.00   TRUE       127.54              2  2.615513
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235