Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813    1130.5791
#> 28223 712742.5 9560482 1927.69 34.667957  -0.00292    0.00712    3636.5303
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.129160   0.00090    0.00890    2301.0454
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345    1120.6929
#> 1768  713912.5 9558552 2357.19 38.692413  -0.00645    0.00835    2425.3115
#> 43669 712392.5 9560162 2001.90 47.283915  -0.00284    0.01804     628.5153
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 3835  714142.5 9558552 2388.42 32.900383   0.00979   -0.00269     491.8017
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164     895.6606
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677    1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189    2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973     514.5488
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691  -0.02651   -0.03629  360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773     179.3077
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593     939.0466
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150    1274.3594
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688   0.00453    0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991  -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 46201 714832.5 9561022 1845.94 34.573992  -0.00856    0.01796    4729.6382
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403   0.01661    0.01199     196.4080
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181   11655.4316
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 27695 714142.5 9558452 2449.28 20.490117  -0.00539    0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925   0.01784    0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013    4229.5708
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448    1258.8196
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112     745.3315
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213     328.7922
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171     788.0004
#> 27244 714162.5 9558622 2316.00 36.751232  -0.02556   -0.11694  102301.4609
#> 47702 714692.5 9559712 2200.51 33.098626  -0.13038   -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860     695.2302
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050   0.01011    0.00739     789.7573
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396     704.7076
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484     495.0292
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075    2170.2249
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823   0.00708   -0.02538     807.8237
#> 46792 713862.5 9559672 2272.03 35.464942   0.00560    0.00610     454.8103
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166    1813.7841
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744     226.7911
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867    2247.8748
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093     118.1402
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609    4275.3457
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 748   715102.5 9559692 2301.39 27.021836  -0.00759    0.00158     837.3730
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493    2212.0525
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328    1066.7134
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445     355.5078
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028    2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690   0.00544   -0.01014     271.4219
#> 12974 714662.5 9557332 2667.61 41.941657  -0.00065    0.00475     515.1661
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697     412.1185
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955     713.4463
#> 20764 714542.5 9558152 2602.18 52.078999   0.02118    0.00162     290.2031
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740     909.4791
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555    1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347    1893.3645
#> 26283 715932.5 9557592 3075.48 32.766883  -0.00375    0.00215     917.5296
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368    1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547   0.00340    0.01809    1158.0426
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884     672.2466
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013     616.4128
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172     351.5812
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 31576 26.9353189   300.00   TRUE       300.00              4  3.635091
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 23626 26.7783285   300.00   TRUE       300.00            100  3.336505
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674