Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642     671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056   -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043    1806.4552
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046    0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057    1630.3259
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761     928.6858
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949    6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804     628.5153
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979   -0.01542     674.0147
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903    0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838     238.6685
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629  360379.1875
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 2950  715042.5 9557712 2740.72 26.525081 -0.00089   -0.00051     779.0555
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086    6042.5757
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255   -0.00365    1536.2647
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888    0.00862     874.5408
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993    0.01373   16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307  147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230    1056.5876
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917   -0.01993   10308.6406
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241     585.6182
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078    0.00219     420.7032
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399    3900.3445
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664    1807.2659
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133    0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832    3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469    0.02940     987.8085
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245    6353.3179
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911     173.2806
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522    1355.7436
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246    0.00396     704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401   -0.00391    1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693    0.01133    2397.2681
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226     338.3474
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019    0.00119    6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732    4237.6934
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893   -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454    0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870    2099.7930
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320     911.5911
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378    0.02382     100.0000
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599    0.00301     338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493    2212.0525
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945     259.7421
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170     253.5129
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782   -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393    0.00032     994.9224
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066    0.05257     549.3252
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233    0.01282     561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229    0.00099     909.0560
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608    0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047    2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865   -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238     787.0053
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419     323.0790
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 18391 34.5075291   300.00   TRUE       300.00            100  5.556760
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 44767 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250