Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976    301.2347
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677   1675.4679
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024    394.8751
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312   2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033 -0.03917    116.1986
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514    297.7842
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547  0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063   5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666 -0.00504   1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974   2136.6939
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310   2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283    590.1536
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684   2556.8049
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620    379.0855
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180   3904.5254
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213    328.7922
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739    789.7573
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200   1305.6691
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075   2170.2249
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713   6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483    343.1384
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119   6589.5796
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744    226.7911
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204    853.2902
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945    259.7421
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548    532.8210
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273  72123.6172
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038   2017.3547
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955    713.4463
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513    161.9878
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577    423.5598
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719  0.00079   1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066    306.0146
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532  16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884    672.2466
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573   1490.0312
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388    180.4414
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336