Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 16435 713512.5 9559092 2166.13 55.973966  0.02056 -0.06976     301.2347
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328    2235.5259
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144 -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316 -0.00427     866.9398
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282  0.00458     485.7418
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253    1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 8073  36.075715   138.71   TRUE        76.41             35  2.997730
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 25244 33.218947   300.00   TRUE       300.00              6  5.916915
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 47320 45.664163   300.00   TRUE       300.00             43  2.536741
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 36126 19.295500   300.00  FALSE       300.00            100  2.525145
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 20135 32.332581   300.00  FALSE       300.00            100  3.190519
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 10127 18.619409   300.00  FALSE       300.00             49  2.417865
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 16845 22.882788   300.00  FALSE        25.55             33  3.874467
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 11570 16.607182   300.00  FALSE       300.00            100  2.686405
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250
#> 48157 33.526625    57.15  FALSE       157.06             45  2.887769
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983