Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579   2246.7725
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046   1024.1573
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547    519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120    512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226   2325.2053
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010   1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526    539.8119
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234    849.0547
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160   9344.2930
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272   4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 31576 713942.5 9558492 2387.12 46.143156 -0.05634 -0.02626   4316.0977
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437   1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156    816.6036
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796   4729.6382
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199    196.4080
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538    807.8237
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771    674.3859
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236   2715.4138
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516    576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109    767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596    842.0994
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354   1147.7384
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627 -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813 287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412   1350.3611
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058   1817.2371
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574    183.3220
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860    695.2302
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078 622888.2500
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583   1898.0105
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833    369.3771
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753   1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484    495.0292
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380    601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077   3056.9475
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001   1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224   1137.3363
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526    706.1463
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477   2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124  0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632    620.0229
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210    585.8992
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109    605.1908
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345 665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923    811.5580
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598    186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759   1880.4403
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518   1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645    234.0114
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282    561.2366
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165    805.1403
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578   1258.8865
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186   2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177    184.1042
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 27375 38.770335   300.00   TRUE       300.00            100  2.859781
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 5728  40.408421   300.00   TRUE       300.00              2  3.325151
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 24501 27.152470   300.00   TRUE       300.00              0  2.925363
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 40216 26.364080   300.00   TRUE       300.00            100  3.257022
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 49133 23.519345   300.00   TRUE       300.00             38  2.694996
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 13072 34.944696   300.00   TRUE       300.00            100  4.477051
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946