Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029    5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393    2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555     430.4604
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324     861.3537
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761     928.6858
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677    1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388    1051.4222
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010    1527.1028
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769    1032.7177
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312    2657.6936
#> 14522 715012.5 9557732 2721.34 35.181327  -0.00382    0.00672    1114.8801
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107    1392.5880
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212     561.1948
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118    2359.5059
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642     613.6390
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077    3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114     150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 47351 714222.5 9558792 2420.94 28.668516  -0.09566   -0.03355   14919.6260
#> 11687 714552.5 9560382 2052.66 29.575508  -0.00255   -0.00365    1536.2647
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063    5035.0000
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338   10787.6299
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517    1000.7119
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169     350.1799
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539  -0.01174    0.00025    4425.5151
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 13134 714652.5 9559632 2265.33 49.599810  -0.03552   -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588     749.2502
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883     616.1993
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262    1760.2162
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055     773.0359
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851    1012.5963
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393    8301.6094
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 8125  713452.5 9559132 2120.44 42.140473  -0.05458   -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412    1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245    6353.3179
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484     495.0292
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190     264.6404
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436   0.00181   -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224    1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713    6379.7974
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450    2589.9910
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586  -0.00622    0.00152    1530.5657
#> 35458 715242.5 9558442 2623.01 38.954828  -0.00770   -0.03400    1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306    1718.3652
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842     246.5161
#> 27884 714072.5 9559202 2446.27 19.756731  -0.00009    0.00109     605.1908
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715     520.7355
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875    1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.257189   0.00075   -0.01945     259.7421
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548     532.8210
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273   72123.6172
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312     593.7773
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151   0.01462    0.00268     252.0072
#> 41855 713182.5 9560452 1851.53 45.178295   0.00667    0.00863    1098.3462
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361    1011.0248
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079    1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964     543.0025
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282    1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585    5714.3267
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057     722.3952
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 10899 24.774122   300.00   TRUE       300.00             20  2.644350
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 16843 21.398255   300.00   TRUE       300.00             12  3.013982
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 38070 34.233082   300.00   TRUE       300.00            100  2.899802
#> 26215 23.831607   300.00   TRUE       300.00             10  3.268382
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 37341 31.451945   300.00   TRUE       300.00              1  2.823221
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 20368 30.344418   300.00   TRUE       166.26            100  3.803001
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 39440 32.706150   300.00  FALSE       300.00             92  2.369237
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 44267 19.737823   300.00  FALSE       226.30            100  2.401413
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 23499 28.514964   300.00  FALSE       300.00             50  3.171627
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 15363 26.942194   300.00  FALSE       300.00            100  2.858775