Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 25090 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.33543  0.00479  0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.87786 -0.02104 -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 125   713992.5 9561062 1878.28 49.97109  0.00183  0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.04492  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.16547  0.00151  0.00899   1394.2274
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.05756  0.02224  0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016
#> 48434 713352.5 9559042 2176.63 35.70215 -0.01771 -0.01939    703.7817
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 29926 714702.5 9557622 2637.00 44.64258  0.01843  0.04407    344.8395
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.16814  0.00283  0.00117   1018.2141
#> 35194 713692.5 9559952 2185.68 28.79686 -0.00449 -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.36659 -0.01017  0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.92351 -0.01364 -0.03357    473.4709
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738
#> 47100 712752.5 9560462 1910.00 41.91186 -0.00032  0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791
#> 29801 714392.5 9559162 2356.18 31.56253 -0.00856 -0.01625   1063.9150
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454   3476.0818
#> 16355 712962.5 9560212 1992.88 39.52607  0.01927 -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.05914 -0.00410  0.03620    379.0855
#> 7064  715592.5 9558662 2597.69 34.22220  0.00099  0.00240   1201.4724
#> 20601 715542.5 9558182 2773.29 32.75542 -0.00604  0.00354   1147.7384
#> 41721 715182.5 9557562 2784.85 31.78999 -0.00588  0.01448   1258.8196
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966
#> 11612 715212.5 9557752 2820.51 39.16224  0.00017  0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 16009 715122.5 9559402 2264.81 50.01864 -0.00589  0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.21547  0.03890 -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937
#> 28312 715712.5 9558032 2783.23 24.25273 -0.00007 -0.02132   1580.1621
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 47298 714602.5 9560572 2025.50 28.59002 -0.00253  0.00753   1337.7854
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341
#> 49072 715472.5 9558902 2482.92 30.68934 -0.01102 -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 7409  714772.5 9559122 2404.90 30.21550 -0.01016 -0.00294    734.9908
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 46792 713862.5 9559672 2272.03 35.46494  0.00560  0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 17674 713882.5 9558562 2339.20 41.03638 -0.03675 -0.01905   6487.0010
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 48461 714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 17322 712612.5 9560402 1924.48 43.80778 -0.00486  0.02306   1718.3652
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402
#> 4785  713792.5 9560472 2055.34 33.05279 -0.00586  0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.59555 -0.00203  0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 11288 712822.5 9560442 1873.42 39.60972 -0.00465 -0.00105   4432.8760
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.19736  0.04336  0.01403    241.2991
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604
#> 36126 715052.5 9558292 2759.47 29.06214  0.00196  0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.33999  0.01432 -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.11809 -0.01322 -0.00759   4131.5356
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 12519 713982.5 9560912 1956.98 28.82493 -0.00452 -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.59538 -0.00035  0.00955    713.4463
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 41855 713182.5 9560452 1851.53 45.17830  0.00667  0.00863   1098.3462
#> 35038 715092.5 9558152 2715.69 40.18382  0.00270 -0.00250   2225.4426
#> 26768 715092.5 9560582 1960.72 36.65612  0.00980 -0.00710    472.0236
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165    805.1403
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624   1870.8307
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562
#> 18504 713372.5 9560672 1786.89  6.47385 -0.00278 -0.01172 111973.8906
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 24962 715152.5 9560472 1967.37 28.84556  0.00298  0.00222    991.8300
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387