Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333    258.6537
#> 42733 715112.5 9559062 2399.70 36.054515  -0.00041    0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.148452   0.00659    0.01041    772.9746
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 37104 713962.5 9561282 1795.56 25.741848   0.00273   -0.02403   1787.3024
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 23925 715202.5 9557652 2782.99 41.957126  -0.00542    0.01842    868.6137
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769   1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 29472 714792.5 9561072 1786.08 36.173117  -0.01029   -0.02312   2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581   1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696   4143.4019
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593    939.0466
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051    779.0555
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 27084 713632.5 9558472 2380.37 23.135463  -0.01752   -0.00517   1000.7119
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974   2136.6939
#> 23029 713502.5 9559102 2174.07 56.457542   0.03395    0.01495    266.5888
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299    688.0706
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684   2556.8049
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599   1378.5760
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524  10861.9482
#> 47718 712812.5 9559892 1867.55 54.938631  -0.01491    0.01061   5744.7583
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241   1752.9242
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094   1392.6008
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291    293.4231
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.792506  -0.01400   -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208   0.00319    0.00291    487.3438
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403    474.8613
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345   2417.1389
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109   0.00112    0.03538    265.4677
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923    811.5580
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399   1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900    614.5898
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599  12873.8018
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 23499 714272.5 9558232 2508.35 28.967600  -0.00820   -0.00260   1484.6593
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162  -0.01196   -0.01935    903.6397
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052    451.7209
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 27084 21.67270156   300.00   TRUE       300.00             81  3.000309
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983