Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051  351155.1250
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438     753.6304
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976     301.2347
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659    3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957  -0.00292    0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896     276.0310
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339    1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644    1179.0695
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311     586.7354
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149   12632.7793
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024     394.8751
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514     297.7842
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530    9000.4356
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 25081 713392.5 9558452 2242.24 46.538497  -0.02716    0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450     873.0448
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596    5296.9014
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651     419.0927
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991    1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905   0.04547    0.03323     183.1836
#> 37828 713482.5 9558652 2330.58 20.445426   0.00606   -0.00556     660.3690
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167    1938.5389
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338   10787.6299
#> 46235 712712.5 9561042 2023.11 41.277025   0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036     516.4709
#> 5578  714962.5 9557612 2751.95 39.191459   0.00288    0.01452     437.3965
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307  147470.5312
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279     930.0871
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.020885   0.00194    0.00516     576.0335
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543     352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625    1063.9150
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169     495.8840
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020    1563.5549
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132   42970.7891
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599    1378.5760
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220    8936.2666
#> 1424  714012.5 9558902 2305.97 32.223719  -0.03986   -0.02524   10861.9482
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870    1895.1399
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351    1442.9247
#> 37720 715722.5 9557962 2820.42 45.076309  -0.01474    0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661    2508.5374
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175    1026.6818
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.411077   0.00079    0.00321    1170.5904
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812    1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441     424.2680
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957     495.4453
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200    1305.6691
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417   34908.6562
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634     500.5027
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705     413.5702
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226     338.3474
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152     850.4633
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606     919.4241
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551  -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384     947.2075
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.647512   0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.539831   0.02405   -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642    1688.1782
#> 48601 713692.5 9560152 2058.14 59.557244  -0.01433   -0.02177     997.3089
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181    1286.3949
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463    1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363     389.2691
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463     851.6306
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464    4029.5947
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333     422.8840
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502    1428.8546
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507     504.3085
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790     209.6952
#> 43403 712632.5 9559262 2072.53 34.486330  -0.00069   -0.00900     614.5898
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759    4131.5356
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282    2322.7944
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250    2225.4426
#> 40909 712722.5 9560272 1827.15 29.293040   0.00009   -0.00599   12873.8018
#> 26768 715092.5 9560582 1960.72 36.656121   0.00980   -0.00710     472.0236
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750    1892.8978
#> 48184 714202.5 9558222 2489.17 36.113530   0.02236   -0.00066     306.0146
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439  -0.06909   -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487  -0.03284   -0.00886   43249.3945
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238     787.0053
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052     451.7209
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157     766.0220
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419     323.0790
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525     162.0592
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227     841.2025
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931    1812.3820
#> 48157 713142.5 9560992 1978.48 44.621380   0.00208    0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769