Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.4766034   -0.01919   -0.04051 351155.1250
#> 37912 712802.5 9559952 1838.40 52.1013441    0.00183   -0.09203    634.3320
#> 29391 712842.5 9560152 1930.16 28.0955584    0.00753    0.01487   1077.7585
#> 42733 715112.5 9559062 2399.70 36.0545152   -0.00041    0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.3279977   -0.00661   -0.01339   1002.5877
#> 11088 715532.5 9558762 2549.28 38.0501272   -0.00373   -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.8152355   -0.00857    0.03677   1675.4679
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073    0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310    0.01440   1088.8639
#> 1768  713912.5 9558552 2357.19 38.6924129   -0.00645    0.00835   2425.3115
#> 23817 714842.5 9557832 2677.58 26.6574344    0.00922    0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.1200421   -0.00223    0.01783    377.8918
#> 38993 714202.5 9558402 2475.75 33.1593594   -0.01607    0.00677   1343.5006
#> 22814 713862.5 9558582 2327.48 48.6120312   -0.02894   -0.03416    947.4878
#> 38287 714872.5 9561162 1719.41 48.2355979   -0.03021   -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983    0.00194    0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.8140475    0.00952    0.01378    440.9104
#> 2990  713542.5 9560372 1970.07 31.8879661    0.01903    0.01447    331.4833
#> 40307 713302.5 9560702 1845.07 39.0046749    0.00129    0.02510   3342.5518
#> 38279 714192.5 9558612 2325.14 28.2525489   -0.04430   -0.06070 138869.3906
#> 14522 715012.5 9557732 2721.34 35.1813275   -0.00382    0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.0449214    0.00832    0.00838    238.6685
#> 4827  714962.5 9559882 2274.42 29.2449118    0.00051    0.00889    362.8363
#> 32821 715022.5 9559952 2208.33 43.4164499   -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.5289733   -0.03613   -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164   -0.00521   -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987    0.02107   1392.5880
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937    0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764    0.00225    0.00325    616.8652
#> 15390 713132.5 9560622 1873.26 38.4804185    0.00252    0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.5662016   -0.01804   -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.5335079   -0.01383    0.00193   1809.3838
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743    0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.3075669   -0.00154    0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.2618978    0.01047   -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.6685162   -0.09566   -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.0575616    0.02224    0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.3559050    0.04547    0.03323    183.1836
#> 48434 713352.5 9559042 2176.63 35.7021461   -0.01771   -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.1672203   -0.01353    0.00063   5035.0000
#> 18564 715552.5 9558592 2648.36 26.2913143    0.00316    0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.8815030   -0.00666   -0.00504   1020.2766
#> 46171 712822.5 9558932 2089.30 29.8545389   -0.01174    0.00025   4425.5151
#> 17036 712832.5 9560582 1951.38 27.6457866    0.00060    0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661    0.01199    196.4080
#> 44550 714842.5 9557642 2696.38 15.1730683    0.00359   -0.00279    930.0871
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779   2972.0193
#> 421   714892.5 9561022 1849.01 26.1320321    0.00002    0.00588    749.2502
#> 42941 713252.5 9559742 1990.10 28.9544222   -0.04708    0.00408  76318.0156
#> 27875 714832.5 9561032 1838.22 41.2122812   -0.01160    0.01690   5121.6924
#> 39732 715362.5 9559582 2230.97 23.3468842   -0.00161   -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.8738238    0.00090   -0.00299    688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529    0.02009   1247.7061
#> 23213 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294    739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262    0.00126   -0.00485   2240.4185
#> 3418  715372.5 9558492 2601.00 22.6272492   -0.01168   -0.04102  35685.0000
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927   -0.00316    454.9674
#> 27695 714142.5 9558452 2449.28 20.4901167   -0.00539    0.00809   1102.9956
#> 45592 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 21680 712802.5 9560462 1893.75 38.7342388   -0.00481   -0.00399   3900.3445
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903   -0.00588    0.01448   1258.8196
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024   -0.00474   2343.2100
#> 40216 714602.5 9560592 2015.55 34.3688733   -0.00093    0.00664   1807.2659
#> 39166 712832.5 9560142 1922.46 39.6601386    0.01469    0.02940    987.8085
#> 43859 714892.5 9557602 2722.70 28.1293630    0.00370   -0.00190    640.8842
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491    0.01061   5744.7583
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521   -0.00870   1895.1399
#> 18348 714042.5 9558882 2341.41 44.0638922   -0.00338    0.00728   1137.2450
#> 41125 712772.5 9559412 2021.88 38.8763323   -0.00310   -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.4622584   -0.00949    0.00520   3390.5708
#> 34887 714662.5 9557392 2609.02 51.0207457   -0.01878   -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.6498183    0.01299    0.00141   4611.0718
#> 43245 713722.5 9560012 2153.38 34.1912564   -0.01263    0.01322   4605.1411
#> 29228 712782.5 9560422 1870.56 32.9061121   -0.00218   -0.00372   4154.5786
#> 22907 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.7896130    0.00466   -0.01776   1021.8267
#> 47086 713692.5 9559972 2175.51 25.0457041   -0.00789   -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.5900210   -0.00253    0.00753   1337.7854
#> 38668 715582.5 9558242 2735.55 35.4460340    0.01003    0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051    0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204   -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.4690084   -0.00495   -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.2711002    0.00705   -0.00915   2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783   -0.00296    0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.3279836   -0.00292   -0.00109   2047.9031
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767   -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.8578160   -0.01106   -0.04634    500.5027
#> 9606  714392.5 9559172 2353.05 30.7833035   -0.00185   -0.01185    953.8614
#> 46038 713042.5 9559102 2146.38 29.4975862   -0.00622    0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.8711757   -0.00212    0.02732   4237.6934
#> 10213 715542.5 9558782 2532.10 35.4133754   -0.00893   -0.00867   2247.8748
#> 9864  714952.5 9557592 2752.40 31.9893796   -0.00983    0.01253    528.1924
#> 49346 715322.5 9558782 2553.74 25.3952084    0.00319    0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.2725884   -0.00166    0.01476   1404.4454
#> 36765 715302.5 9560262 1952.73 27.2034632   -0.00585   -0.02665   1904.4769
#> 4785  713792.5 9560472 2055.34 33.0527893   -0.00586    0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203    0.00642   1688.1782
#> 7317  713682.5 9558582 2347.46 28.3476599   -0.02196    0.00336   2806.2644
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325   -0.00505    351.4078
#> 27884 714072.5 9559202 2446.27 19.7567307   -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439   -0.01433   -0.02177    997.3089
#> 748   715102.5 9559692 2301.39 27.0218355   -0.00759    0.00158    837.3730
#> 24836 714202.5 9558272 2507.87 38.4890128   -0.00226    0.00736    477.8591
#> 15558 714732.5 9559502 2283.66 53.4317521    0.00390   -0.02831    454.9113
#> 11111 715142.5 9558482 2650.65 42.3432999   -0.00533    0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.7332955   -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.5098210    0.00165    0.00875   1165.3953
#> 8066  715842.5 9558732 2636.38 33.7174203   -0.00494    0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.3029762    0.00540    0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.3895209   -0.00724    0.01464   4029.5947
#> 4007  714132.5 9558412 2464.44 43.7201175    0.00174   -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759   1880.4403
#> 43403 712632.5 9559262 2072.53 34.4863297   -0.00069   -0.00900    614.5898
#> 41753 715722.5 9558192 2771.39 26.0850495   -0.00196   -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.0321698    0.00490   -0.00780    825.4451
#> 20319 715282.5 9560482 1901.16 27.6767263    0.00706   -0.03035   1471.9783
#> 24009 714042.5 9558962 2357.58 34.3619978    0.05041   -0.00011    201.1272
#> 3733  713762.5 9559632 2297.63 34.2147477   -0.01338    0.00518   1137.0179
#> 17746 714752.5 9558322 2648.52 47.7365517   -0.00068   -0.00312    593.7773
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654   1576.2065
#> 20764 714542.5 9558152 2602.18 52.0789988    0.02118    0.00162    290.2031
#> 46509 712512.5 9559532 1860.09 89.9887513   -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.3936922    0.03479    0.02251    168.1733
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.8694849   -0.00229    0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972    0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 9007  714952.5 9560902 1863.75 32.7703211    0.00074   -0.02904    608.3561
#> 7986  712752.5 9560242 1857.41 39.8303070    0.01852    0.00248   1432.4019
#> 48050 714022.5 9560572 2099.82 18.7139475    0.01182   -0.00972    331.7462
#> 10127 713722.5 9558712 2311.10 19.4811380    0.01129    0.00491    261.7371
#> 10981 713832.5 9560022 2125.33 34.8484390   -0.06909   -0.01151 143213.1562
#> 16443 714472.5 9559532 2222.55 16.7441186    0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003    0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474    0.00340    0.01809   1158.0426
#> 14881 715482.5 9559232 2376.03 29.9055321   -0.01137   -0.01523    888.2811
#> 33881 712992.5 9558822 2173.26 33.1977476   -0.00595    0.03064   1485.5201
#> 35257 713822.5 9557982 2350.80 29.3079371    0.01549   -0.00169    649.9975
#> 31440 714842.5 9558392 2704.41 25.8014354    0.00992   -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.6626120   -0.00983    0.00803   1058.5630
#> 24962 715152.5 9560472 1967.37 28.8455602    0.00298    0.00222    991.8300
#> 11025 714402.5 9559252 2325.64 49.6204369   -0.00241    0.00020    673.2861
#> 15363 714742.5 9559212 2373.13 33.5833482   -0.00247    0.01057    722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 48434 29.9358989   300.00   TRUE       300.00              6  2.847438
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775