Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841  -0.00010   1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491   1844.4353
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979  -0.01542    674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406   2421.1145
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175  -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160   0.00311    586.7354
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169  -0.00160   9344.2930
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899   1394.2274
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043   0.00973    514.5488
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664   0.00694   2123.4473
#> 8073  712572.5 9560302 1917.62 52.009098    0.01937   0.00083    994.7876
#> 27957 715012.5 9559332 2306.35 20.480376    0.00225   0.00325    616.8652
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205    629.4165
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414   0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556    660.3690
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856   0.01796   4729.6382
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907   1943.4092
#> 13134 714652.5 9559632 2265.33 49.599810   -0.03552  -0.02058    431.4362
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405  -0.01565   2232.6133
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797  -0.00063    576.2819
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161  -0.00409   2312.8413
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027  -0.01473   2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647    0.00062  -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114   0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550  -0.00941   1290.7709
#> 28046 714882.5 9558492 2711.77 28.305834   -0.00131   0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478  -0.07132  42970.7891
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832   3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370  -0.00190    640.8842
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017   0.01213    328.7922
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 44767 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205   1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299   0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263   0.01322   4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 15981 713302.5 9559172 2106.91 35.997792    0.01354   0.00245    629.5983
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109   0.00298   1533.7925
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.358204    0.00540   0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767  -0.03713   6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341   0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226    338.3474
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115   0.01015    790.2421
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655    130.9468
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166  -0.01265   1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737   -0.03251  -0.00389  14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781  29995.1543
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586   0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158    837.3730
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124   0.00204    853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403    241.2991
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295   0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118   2273.4758
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195  -0.00995   1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724   0.01464   4029.5947
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457   0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033   3446.5520
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514    897.9290
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413   0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 17746 714752.5 9558322 2648.52 47.736552   -0.00068  -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379  -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305  -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.178812   -0.00950  -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066   0.05257    549.3252
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312    254.0656
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547    128.2613
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119   0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710    472.0236
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165    805.1403
#> 3868  714542.5 9560622 2024.15 29.447166    0.00360   0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960    884.2003
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354   0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537  -0.00347   1893.3645
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284  -0.00886  43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595   0.03064   1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435    0.00992  -0.01052    451.7209
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983   0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181  -0.00799   1036.6606
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771   0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491