Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813   1130.5791
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 30059 713832.5 9559662 2288.02 18.277927    0.01917   0.00333    258.6537
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507 159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046   0.00324    861.3537
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752   1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431   4079.0366
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854   0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730   0.01770    474.5059
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267   -0.07664  -0.05766 140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672   1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899   1394.2274
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713  -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902   0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651  -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206   0.02773    179.3077
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107   1392.5880
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804  -0.01596   5296.9014
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051    779.0555
#> 35142 714792.5 9558862 2575.17 25.664499    0.06765   0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.005073    0.03696   0.02514    297.7842
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545   0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047  -0.00887    366.8537
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140  -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458   0.00672    565.1107
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442  -0.00178    800.1933
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294    739.5768
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735  -0.00925    665.6116
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613   0.00903   2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708  -0.02538    807.8237
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478   0.01092    412.5849
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436   2602.7390
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736   0.01684   2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543    352.1558
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553   1750.4175
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168  -0.04102  35685.0000
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195  -0.00656   3036.7383
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645  -0.00395   2011.3270
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394  -0.00224    942.0322
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024  -0.00474   2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242    507.9263
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865   0.00335    195.0293
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038  -0.04212  46310.0195
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310  -0.00860    695.2302
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542  -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193   0.02812   1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789  -0.00261   3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253   0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246   0.00396    704.7076
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422   1159.2100
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281   0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077   3056.9475
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610   1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526    706.1463
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685   2568.4631
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252   3631.7017
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291    293.4231
#> 43834 713092.5 9560662 1905.55 31.323603    0.00758  -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 28253 714102.5 9559892 2288.88  2.957608    0.00378   0.02382    100.0000
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336   2806.2644
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165   0.00565   1410.8372
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496    0.00872   0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458   0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290   2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548    532.8210
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031  -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846   0.00794   1363.5321
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978   0.00367   3002.8828
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052   3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943    371.5514
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410   0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292   0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.956309   -0.00301  -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307   0.02997    230.8641
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645    234.0114
#> 41855 713182.5 9560452 1851.53 45.178295    0.00667   0.00863   1098.3462
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154  -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312    254.0656
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165   0.00555   1882.0179
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972   0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836    250.5111
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849    764.8193
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502   0.00282   1659.6211
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253  -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172 111973.8906
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021   0.00109    915.4989
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241   0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403    609.8665
#> 48157 713142.5 9560992 1978.48 44.621380    0.00208   0.00492    772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769