Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029   5577.3916
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043   1806.4552
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628    556.0121
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896    276.0310
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858
#> 25169 712642.5 9560232 1883.10 46.845029  0.00884  0.01345   1120.6929
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189    287.5425
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898    458.6938
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382    560.8309
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581   1855.1630
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973    514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696   4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173   8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898    588.3761
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694   2123.4473
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357    275.7474
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991   1403.0193
#> 47011 713262.5 9559722 2003.72 37.192982 -0.01846 -0.00034   4495.6929
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499    885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672    565.1107
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279    930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230   1056.5876
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400    608.6523
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993  10308.6406
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181  11655.4316
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330   1456.9246
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221  0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009   1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625   1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328    592.3326
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643    670.2289
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240   1201.4724
#> 40270 713312.5 9559012 2186.56 30.935710 -0.00771 -0.00599   1378.5760
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940    987.8085
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376   2048.3198
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728   1137.2450
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175   1026.6818
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606    564.6876
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441    424.2680
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778   4594.4351
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133   2397.2681
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044    288.5468
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450   2589.9910
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732   4237.6934
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063   2430.2278
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265   1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320    911.5911
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712   1010.5097
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517   1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463   1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794   1363.5321
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099   1146.2140
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863   1098.3462
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154 -0.00204    618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547    128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930 -0.00360   4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105    868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417   1520.6630
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836    250.5111
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875   2156.9358
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841    756.0055
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151 143213.1562
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172    351.5812
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15981 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946