Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305 -0.00555     430.4604
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261     323.8441
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333 -0.01767    3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422  0.02357     275.7474
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554  0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800 -0.06020    1563.5549
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201     788.0621
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567 -0.03793     864.9472
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341  0.00171    4425.5410
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185     953.8614
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212  0.02732    4237.6934
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 24009 714042.5 9558962 2357.58 34.361998  0.05041 -0.00011     201.1272
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624    1870.8307
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 14254 15.665239   300.00   TRUE       300.00             10  2.510336
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 32622 30.171385   300.00   TRUE       300.00            100  3.265863
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 25081 34.875368   300.00   TRUE       300.00            100  3.617357
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 36931 27.209193   300.00   TRUE       300.00            100  3.408707
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 37141 12.388493   300.00   TRUE       300.00              1  2.293159
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 5103  34.913183   300.00   TRUE       300.00             91  2.623976
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 38103 42.761559   300.00   TRUE       300.00             31  2.907317
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 17374 43.145441   300.00   TRUE       300.00            100  3.630971
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 40459 36.186868   300.00  FALSE       216.66             10  3.025531
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 39193 31.480020   300.00  FALSE       117.12            100  2.879028
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 10981 29.166417   300.00  FALSE       300.00            100  5.155983
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607