Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633   -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969     368.5606
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097    0.01483     566.7674
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507  159065.0781
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373   -0.00057    1630.3259
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.129160    0.00090    0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250   -0.01340     604.1032
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310    0.01440    1088.8639
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088     266.4355
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894   -0.03416     947.4878
#> 11463 713112.5 9558902 2244.08 21.891571    0.00160    0.00311     586.7354
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414   -0.00104    1641.4587
#> 30800 715182.5 9557582 2772.37 39.204064   -0.02475    0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123    0.01164     895.6606
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272    4955.5586
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250    3686.1240
#> 30125 715162.5 9558992 2438.18 39.856663    0.00471   -0.01382     560.8309
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672    1114.8801
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838     238.6685
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581    1855.1630
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696    4143.4019
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450     873.0448
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917     116.1986
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212     561.1948
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124    2114.2253
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991    1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167    1938.5389
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771   -0.01939     703.7817
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063    5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338   10787.6299
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452     437.3965
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270    4057.5637
#> 37341 714412.5 9559142 2377.24 40.599789   -0.00735   -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661    0.01199     196.4080
#> 18394 715022.5 9559942 2216.16 42.137608   -0.00613    0.00903    2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072   54831.6289
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504    0.05825     162.6688
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252    5761.3486
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241     585.6182
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681   -0.01319    3457.7519
#> 37416 713282.5 9560692 1850.61 30.259047    0.00736    0.01684    2556.8049
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598    0.00114    0.03026     247.0894
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133    0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394     587.8840
#> 11612 715212.5 9557752 2820.51 39.162238    0.00017    0.01213     328.7922
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781   29995.1543
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911     173.2806
#> 47702 714692.5 9559712 2200.51 33.098626   -0.13038   -0.04212   46310.0195
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666   -0.03583    1898.0105
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321    1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522    1355.7436
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118   19557.6621
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 38668 715582.5 9558242 2735.55 35.446034    0.01003    0.01196    1052.5341
#> 16071 714562.5 9560062 2166.69 28.696018    0.02281    0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042    0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705   -0.00915    2036.0378
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450    2589.9910
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185   -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526     706.1463
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166    1813.7841
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744     226.7911
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870    2099.7930
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186    0.00063    2430.2278
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166   -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427     866.9398
#> 15186 713322.5 9561212 1936.88 36.079725    0.00165    0.00565    1410.8372
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759    0.00158     837.3730
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007   -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118    2273.4758
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.440109    0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345  665381.5625
#> 30267 714442.5 9557512 2612.30 28.765919   -0.00458    0.00479     604.0975
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161    0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199     836.4196
#> 1085  715722.5 9557532 3097.22 28.784254    0.02327    0.01833     300.0125
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174   -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399    1601.4852
#> 43403 712632.5 9559262 2072.53 34.486330   -0.00069   -0.00900     614.5898
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514     897.9290
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267   31588.1953
#> 22545 713402.5 9559562 2195.44 25.315567    0.04152    0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424    0.00196     576.1450
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018     937.6911
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586    0.00654    1576.2065
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342    0.02552    2515.0752
#> 7553  714642.5 9559132 2391.59 63.967873   -0.00066    0.05257     549.3252
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523    0.02247     282.1076
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820   -0.00260    1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044   -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438    1276.5448
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264    0.00624    1870.8307
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576    1045.1074
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186    2751.4626
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011     501.7867
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884     672.2466
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595    0.03064    1485.5201
#> 6577  714142.5 9559902 2244.30 57.890510   -0.06986   -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419     323.0790
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962   -0.00172     351.5812
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023   -0.00227     841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 11612 35.16700355   300.00   TRUE       300.00            100  2.516921
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 43403 31.06176095   300.00  FALSE       111.88             86  2.788585
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901