Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761     928.6858
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466     724.0713
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525 -0.11024     394.8751
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664  0.00694    2123.4473
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495     266.5888
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168 -0.04102   35685.0000
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568 -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458 -0.14813  287033.4688
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299  0.00141    4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010  0.00140    1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560  0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113 -0.00844  0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097  0.00944  0.00836     250.5111
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607 -0.04566   12862.4922
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 38032 35.0082942   300.00   TRUE       300.00             60  2.828706
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 26101 28.6461709   300.00   TRUE       300.00            100  3.068405
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 15138 20.6459612   116.69  FALSE        19.61             40  4.109325
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607