Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051  351155.1250
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475    1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261     323.8441
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661   -0.01339    1002.5877
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191   -0.00579    2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503    0.02433    1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073    0.00174    2310.4871
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149   12632.7793
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126   -0.02044  129412.3984
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993   -0.00413     364.9788
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108   -0.00678    1891.6584
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211    0.00642     613.6390
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440   -0.02651     419.0927
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323     183.1836
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379    0.01338   10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714     475.3120
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060    0.00270    4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102     483.4184
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181   11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211    0.00241     585.6182
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953   -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542    3614.7334
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736    0.01684    2556.8049
#> 664   714452.5 9559852 2174.83 44.860876  0.00282    0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104    0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395    2011.3270
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201     788.0621
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165    0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604    0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024   -0.00474    2343.2100
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890   -0.00911     173.2806
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466   -0.01776    1021.8267
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522    1355.7436
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169    0.01441     424.2680
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442   -0.00118   19557.6621
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010    3467.2690
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258    0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417   34908.6562
#> 20171 715732.5 9557962 2815.53 48.657868 -0.01341    0.00171    4425.5410
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864    0.04226     338.3474
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458    0.04291     293.4231
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783   -0.00367    1966.2112
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000    0.00320     911.5911
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165    0.00565    1410.8372
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162   -0.00517    1060.5494
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336    0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295    0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533    0.01493    2212.0525
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195   -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724    0.01464    4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544   -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818    0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263    0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759    1880.4403
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170     253.5129
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983     270.7949
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004    0.00714  480535.5000
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009   -0.00599   12873.8018
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523    0.02247     282.1076
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502    0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238     787.0053
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788   -0.01029     894.2672
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172   -0.00119     792.3544
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458     485.7418
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157     766.0220
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011   -0.00419     323.0790
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 23031 26.7410862   300.00   TRUE       300.00              2  2.798938
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 20171 42.1301596   300.00   TRUE       300.00            100  3.645966
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309