Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919   -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123    0.00393    2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199    0.00659    3553.7166
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221    0.00969     368.5606
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628     556.0121
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917    0.00333     258.6537
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041     772.9746
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767    3321.6113
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273   -0.02403    1787.3024
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046    1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854    0.01264    5057.1431
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024     394.8751
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678   -0.04272    4955.5586
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029   -0.02312    2657.6936
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430   -0.06070  138869.3906
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643    0.00314    2360.1501
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088    2322.1746
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225    0.00325     616.8652
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452   -0.01437    1155.1469
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224    0.00177     291.5937
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606   -0.00556     660.3690
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442   -0.00178     800.1933
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395    0.01495     266.5888
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329   -0.01181   11655.4316
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708    0.00408   76318.0156
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140   -0.02330    1456.9246
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458   -0.00401    1722.6481
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117 -0.02554    0.03304    1569.7313
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611    0.02961    1086.6963
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291   -0.00081     940.6002
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112     745.3315
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133    0.00367     806.0966
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241    1752.9242
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520    3390.5708
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218   -0.00372    4154.5786
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789   -0.00261    3353.5422
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250   -0.00970    1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540    0.00879    1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605    0.00075    2170.2249
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094    1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121    1783.0658
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285   -0.01325     595.0426
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000  162178.1562
#> 12074 714562.5 9560372 2045.17 26.287304 -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895    0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230    1044.6443
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683   -0.00603    1562.0503
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996    1599.3491
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009    0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177     997.3089
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872    0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204     853.2902
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133   -0.00452     578.5847
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112    0.03538     265.4677
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197    0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445     355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712    0.00923     811.5580
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327    0.01833     300.0125
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341    0.04431     315.9084
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978    0.00367    3002.8828
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084   -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218    0.00902     218.7467
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983     270.7949
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413    0.00528     355.3475
#> 17746 714752.5 9558322 2648.52 47.736552 -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452   -0.00948    1222.5433
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162     290.2031
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552    2515.0752
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923    0.05547     128.2613
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980   -0.00710     472.0236
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165     805.1403
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 9007  714952.5 9560902 1863.75 32.770321  0.00074   -0.02904     608.3561
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 280   713182.5 9560632 1862.83 39.004102  0.00745    0.00875    2156.9358
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566   12862.4922
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233     968.7780
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595    0.03064    1485.5201
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181   -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013     616.4128
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811   -0.02281     492.4785
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023   -0.00227     841.2025
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771    0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 47223 36.259061   300.00   TRUE       300.00            100  3.120164
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 42733 31.282923   300.00   TRUE       300.00            100  2.872981
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 47351 34.878806   300.00   TRUE       300.00            100  4.173758
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 37828 18.763795   300.00   TRUE       300.00              2  2.819787
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 13134 37.643900   300.00   TRUE       300.00             67  2.634917
#> 23029 46.601522   300.00   TRUE       300.00             34  2.425842
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 32280 27.831425   126.66   TRUE         0.00              5  3.195825
#> 38197 34.342517   105.52   TRUE        65.32             50  3.036108
#> 41783 28.186086   300.00   TRUE       300.00            100  2.973405
#> 7385  28.710342   300.00   TRUE       300.00            100  4.219794
#> 47366 28.727531   300.00   TRUE       300.00            100  2.872349
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 29228 30.399422   131.25   TRUE        80.55              2  3.618527
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10363 45.093497   300.00   TRUE       297.43              6  3.599377
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 24723 28.173481   300.00   TRUE        36.63            100  2.774548
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 35458 37.473159   300.00   TRUE       300.00            100  3.297274
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 27884 15.497362   300.00  FALSE       300.00            100  2.781892
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 18286 24.350706   300.00  FALSE       300.00             42  2.931097
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20458 21.429767   300.00  FALSE       300.00             28  2.424012
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 15008 26.907817   300.00  FALSE       300.00             48  2.625829
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 15138 20.645961   116.69  FALSE        19.61             40  4.109325
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491