Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057   1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.584857   0.00065    0.00005    769.3426
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311    586.7354
#> 10974 715312.5 9557502 2844.23 26.664883  -0.00831   -0.00149  12632.7793
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 18843 715192.5 9559332 2265.75 41.153267  -0.07664   -0.05766 140591.4062
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 22566 712832.5 9559842 1897.49 33.060238   0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401  -0.00987    0.02107   1392.5880
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 17004 713102.5 9559032 2186.79 36.533508  -0.01383    0.00193   1809.3838
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146   1347.6224
#> 24100 715132.5 9558992 2420.18 29.213972  -0.03229   -0.00511 361979.2812
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169    350.1799
#> 6967  715342.5 9560192 1995.36 44.881503  -0.00666   -0.00504   1020.2766
#> 47091 713382.5 9559182 2122.71 48.037927   0.04527    0.04613    289.6113
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903   2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639  -0.02617    0.00307 147470.5312
#> 4031  715392.5 9560162 1998.16 46.285823   0.00708   -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128  -0.00855   -0.02175   1122.9120
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236   2715.4138
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941   1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936  -0.00099    0.02169    495.8840
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.352628   0.00568   -0.01558   1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 17431 712802.5 9560022 1832.71 41.525116  -0.00478   -0.07132  42970.7891
#> 39722 715022.5 9557712 2735.53 30.260766   0.00133    0.00367    806.0966
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337    486.7551
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911    173.2806
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 41125 712772.5 9559412 2021.88 38.876332  -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484    495.0292
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 34984 715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 26058 713402.5 9560442 1880.12 31.377461   0.00730   -0.02610   1872.9276
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230   1044.6443
#> 9864  714952.5 9557592 2752.40 31.989380  -0.00983    0.01253    528.1924
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445    355.5078
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794   1363.5321
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.968383  -0.01405    0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133    828.8604
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 34616 714522.5 9558742 2568.77 34.997981   0.00990    0.02790    209.6952
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 29527 714332.5 9557502 2567.02 48.904813  -0.00452   -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400   4126.7812
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960    884.2003
#> 32430 715832.5 9558112 2824.39 23.694097   0.00944    0.00836    250.5111
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 41873 714832.5 9559942 2253.59 10.598000   0.01508   -0.00177    184.1042
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169    649.9975
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 31357 23.817283   158.92   TRUE         0.00              5  3.243543
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 24129 24.128972   300.00   TRUE       300.00             89  2.886120
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 10974 32.465508   300.00   TRUE       300.00            100  4.101499
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 18843 28.526423   300.00   TRUE       300.00            100  5.147959
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 4827  21.073961   300.00   TRUE       300.00              0  2.559711
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 12868 37.588323   119.34   TRUE        79.16             61  3.129568
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 4605  17.635641   300.00   TRUE       300.00             10  2.544291
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 2595  30.439529   300.00   TRUE       300.00            100  3.538794
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 10364 26.093071   179.84   TRUE         0.00              0  3.541090
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 46065 29.770887   300.00   TRUE       251.11            100  3.482407
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 46461 24.941426   288.47   TRUE         0.00            100  2.687311
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 41125 36.555853   300.00   TRUE         0.00            100  2.842129
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 29024 42.573056   300.00   TRUE       300.00             32  2.529363
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 31071 31.737851   300.00  FALSE       300.00             76  3.259480
#> 14694 43.511561   300.00  FALSE       300.00            100  3.021498
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 30267 25.270304   300.00  FALSE       300.00            100  2.781107
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 49835 32.361229   300.00  FALSE         0.00            100  3.134665
#> 1085  25.894828   300.00  FALSE       300.00            100  2.477139
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 20547 34.261157   300.00  FALSE        42.56            100  4.078911
#> 39685 47.873489   300.00  FALSE       300.00            100  2.918481
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 34616 31.499501   300.00  FALSE       300.00            100  2.321589
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 29527 36.726595   300.00  FALSE       300.00            100  2.639632
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 32430 22.860443   300.00  FALSE       300.00            100  2.398827
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 34334 25.089822   300.00  FALSE       300.00            100  2.986224
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 13548 16.978458   300.00  FALSE       300.00            100  2.256336
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235