Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.05378  0.00097  0.01483    566.7674
#> 46178 714892.5 9559312 2379.54 32.76688 -0.00266  0.02896    276.0310
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 44952 715352.5 9557472 2873.35 38.95254 -0.01073  0.00174   2310.4871
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730  0.01770    474.5059
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 38287 714872.5 9561162 1719.41 48.23560 -0.03021 -0.09328   2235.5259
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.90038  0.00979 -0.00269    491.8017
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 22283 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 16843 714882.5 9558512 2701.78 26.63795 -0.00270 -0.00769   1032.7177
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 3466  713232.5 9560612 1812.72 51.56620 -0.01804 -0.01596   5296.9014
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555
#> 28577 712812.5 9559962 1849.82 59.78012  0.00646 -0.04086   6042.5757
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 37341 714412.5 9559142 2377.24 40.59979 -0.00735 -0.00925    665.6116
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230   1056.5876
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959
#> 421   714892.5 9561022 1849.01 26.13203  0.00002  0.00588    749.2502
#> 36120 714622.5 9559942 2166.51 32.98633 -0.02917 -0.01993  10308.6406
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688
#> 48808 715032.5 9557792 2707.02 25.43761 -0.02329 -0.01181  11655.4316
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594  16587.9980
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.03442  0.00188  0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534    344.1447
#> 10933 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210
#> 49133 714162.5 9558462 2451.13 21.89043  0.00042  0.00957    495.4453
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086
#> 22301 713722.5 9560712 1913.37 39.27110  0.00705 -0.00915   2036.0378
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.32798 -0.00292 -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 49459 714692.5 9557342 2639.74 42.04708  0.00216 -0.02606    919.4241
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 46971 713532.5 9559952 2198.03 29.72390 -0.00529  0.01230   1044.6443
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.27259 -0.00166  0.01476   1404.4454
#> 10785 714052.5 9557852 2377.76 20.98401 -0.00025 -0.05345   2417.1389
#> 21490 714672.5 9560382 2021.86 32.80413 -0.00783 -0.00367   1966.2112
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 15186 713322.5 9561212 1936.88 36.07973  0.00165  0.00565   1410.8372
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.73330 -0.01174 -0.01466   2633.9666
#> 49363 712772.5 9559552 1935.91 29.03693 -0.03220 -0.03290   2481.5632
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.55332 -0.00502 -0.00028   2394.7261
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 12974 714662.5 9557332 2667.61 41.94166 -0.00065  0.00475    515.1661
#> 8831  714372.5 9559852 2200.77 48.76043 -0.00457  0.01697    412.1185
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.08505 -0.00196 -0.00175    725.1355
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 39440 714222.5 9557872 2508.52 39.04765  0.02785  0.00645    234.0114
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.79731  0.01978  0.01282   2322.7944
#> 7553  714642.5 9559132 2391.59 63.96787 -0.00066  0.05257    549.3252
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392
#> 48988 714682.5 9560852 1942.97 20.29073  0.00480  0.00740    909.4791
#> 26724 714252.5 9560862 1975.16 28.28635  0.00165  0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.39369 -0.02608  0.00148  10771.1963
#> 3724  714322.5 9557402 2603.46 34.59748  0.02523  0.02247    282.1076
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978
#> 44216 715772.5 9558402 2710.44 30.44469 -0.00070 -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.70405 -0.00264  0.00624   1870.8307
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397
#> 9320  713872.5 9559262 2378.76 29.85855  0.00309  0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.21167  0.00788 -0.01029    894.2672
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220
#> 23808 715172.5 9558702 2533.50 22.34593  0.00811 -0.02281    492.4785
#> 13548 715812.5 9558122 2821.65 15.50825  0.02751  0.00388    180.4414
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 38077 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 332    0.55691498    94.82   TRUE         0.00             90  3.079334
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 7553  39.77129239   300.00  FALSE       300.00            100  2.739830
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983