Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681   -0.00029    5577.3916
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633   -0.01813    1130.5791
#> 15762 714922.5 9558862 2535.97 54.910556  0.01305   -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475    1001.0861
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.583935 -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373   -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.327991 -0.00333   -0.01767    3321.6113
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453    0.00257     357.1436
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529    3282.1186
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730    0.01770     474.5059
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606     564.6876
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223    0.01783     377.8918
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160    0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769   -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269     491.8017
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475    0.00665   25318.3555
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160    9344.2930
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720    0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191   -0.00261     554.9243
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440    0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151    0.00899    1394.2274
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173    8967.1924
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694    2123.4473
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071    0.00150    1274.3594
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416    1204.8629
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047   -0.00887     366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752   -0.00517    1000.7119
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856    0.01796    4729.6382
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288    0.01452     437.3965
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008    0.00102     483.4184
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855   -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373     527.3528
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175    0.00236    2715.4138
#> 32909 712812.5 9560452 1883.85 38.251235 -0.00221    0.00121    4159.6738
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161   -0.00409    2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090   -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856   -0.01625    1063.9150
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 3418  715372.5 9558492 2601.00 22.627249 -0.01168   -0.04102   35685.0000
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020    1563.5549
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539    0.00809    1102.9956
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201     788.0621
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224     942.0322
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629   -0.00469    1021.5407
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190     640.8842
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453    0.01337     486.7551
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211   49347.5547
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728    1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259   -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949    0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236    9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661    2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444   -0.00017     759.9937
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439   -0.01161    3065.3694
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003    0.01196    1052.5341
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705   -0.00915    2036.0378
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705     413.5702
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143    0.00477    2945.2100
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610     454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354    0.00245     629.5983
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312   -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384     947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405   -0.05655     130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332    0.01842     246.5161
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849    0.00150     503.5590
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476    1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909   -0.01181    1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445     355.5078
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306   -0.00626     450.7346
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075   -0.01945     259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253   -0.00333     422.8840
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538   -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170     253.5129
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422     967.9738
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655     560.8554
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307    0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224   -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785    0.00645     234.0114
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379   -0.01018     937.6911
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305   -0.00705     968.9738
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035    0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950   -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.078999  0.02118    0.00162     290.2031
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312     254.0656
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070   -0.00060     383.3547
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972    0.00308     422.5022
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873   -0.01417    1520.6630
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 484   713812.5 9560592 2016.41 44.301097 -0.00719    0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273   -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438    1276.5448
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343    0.01177    1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852    0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552    0.00532   16771.3066
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377    1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161    0.00709    1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137   -0.01523     888.2811
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549   -0.00169     649.9975
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803    1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627   -0.00124   25826.6914
#> 11570 715572.5 9558482 2685.72 16.814019  0.00282    0.00458     485.7418
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021    0.00109     915.4989
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751    0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366    0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 47220 36.4802228   300.00   TRUE       300.00            100  3.327041
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 18897 31.7882714   300.00   TRUE        20.00              5  3.919162
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 484   32.1457971   300.00  FALSE        55.63            100  3.216864
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235