Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183  -0.09203     634.3320
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633  -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483     566.7674
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075   0.00475    1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863  -0.00507  159065.0781
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896     276.0310
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057    1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 37104 713962.5 9561282 1795.56 25.741848    0.00273  -0.02403    1787.3024
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440    1088.8639
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295   0.00035    2697.5974
#> 22814 713862.5 9558582 2327.48 48.612031   -0.02894  -0.03416     947.4878
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269     491.8017
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903   0.01447     331.4833
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672   0.04398     267.2659
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440   0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899    1394.2274
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716   0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051   0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.707890   -0.00077  -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733     598.7983
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613  -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987   0.02107    1392.5880
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088    2322.1746
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118    2359.5059
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156     816.6036
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089  -0.00051     779.0555
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077     300.7915
#> 37348 715632.5 9558182 2743.16 13.620353   -0.00904  -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229  -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.480991   -0.00509   0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714     475.3120
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267  -0.05063   18733.8848
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661   0.01199     196.4080
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230    1056.5876
#> 15766 715562.5 9558592 2646.16 29.020885    0.00194   0.00516     576.0335
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282   0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121    4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306  -0.00436    2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299     688.0706
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568  -0.02553    1750.4175
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109     767.4097
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927   0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410   0.03620     379.0855
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099   0.00240    1201.4724
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481  -0.00399    3900.3445
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354    1147.7384
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367     806.0966
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188   0.00832    3022.3462
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220    8936.2666
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376  -0.01644 1057765.6250
#> 34615 713242.5 9560732 1879.60 27.593647    0.00384   0.00175    1026.6818
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776    1021.8267
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511  -0.01801    1071.6141
#> 14545 713562.5 9559602 2222.17 54.239050    0.01011   0.00739     789.7573
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970    1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258   0.00422    1159.2100
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042   0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705  -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016  -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140    1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094    1392.6008
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185     953.8614
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397   0.00526     706.1463
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157  -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181  -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354   0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458   0.04291     293.4231
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230    1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036    2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831    0.02405  -0.05655     130.9468
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320     911.5911
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196   0.00336    2806.2644
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476    1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105    4432.8760
#> 34250 713822.5 9559142 2316.36 33.759819   -0.02284   0.01085    2229.5105
#> 14694 715032.5 9559482 2307.44 44.173327   -0.00007  -0.00463    1050.7473
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738   0.00118    2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133  -0.00452     578.5847
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165   0.00875    1165.3953
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290    2481.5632
#> 39193 715332.5 9560852 1872.90 30.290560   -0.00161   0.00171     756.8818
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 13286 713662.5 9559702 2268.14 41.996660   -0.01341   0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118    0.00167   0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077    8759.9961
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001   0.00759    1880.4403
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236   0.00483    1097.3082
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737   0.01033    3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131    0.00601  -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175     725.1355
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844   0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193  -0.00267   31588.1953
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655     560.8554
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785   0.00645     234.0114
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112  -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586   0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978   0.01282    2322.7944
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162     290.2031
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251     168.1733
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060     383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930  -0.00360    4288.7329
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804  -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635  -0.00105     868.7802
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836     250.5111
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875    2156.9358
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186    2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770   0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278  -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263  -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177     184.1042
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549  -0.00169     649.9975
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166   0.00156     758.4840
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110  -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 40756 39.3352715   300.00   TRUE       300.00            100  3.000471
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 22814 34.9120373   300.00   TRUE       300.00              6  2.976574
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 32909 28.3791726   118.07   TRUE        98.15              0  3.619059
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 46971 16.7842257   300.00   TRUE       262.51            100  3.018968
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 1281  35.3938948   300.00  FALSE       300.00            100  3.632329
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603