Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344   0.00183   -0.09203     634.3320
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579    1318.7560
#> 16749 715072.5 9558242 2745.27 49.053782   0.00097    0.01483     566.7674
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261     323.8441
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752    1321.5997
#> 11088 715532.5 9558762 2549.28 38.050127  -0.00373   -0.00057    1630.3259
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433    1475.3191
#> 27973 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588    1295.1610
#> 7349  714032.5 9558502 2402.95 31.352824   0.01924    0.02606     564.6876
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526     539.8119
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447     331.4833
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665   25318.3555
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250    3686.1240
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203   10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.677174   0.00426   -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114    4510.8833
#> 26215 714332.5 9560262 2138.08 25.721221  -0.00491    0.00581    1855.1630
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628     556.0121
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413     364.9788
#> 36230 713322.5 9560662 1798.99 40.128245  -0.00033   -0.03917     116.1986
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950     950.3138
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357     275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437    1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156     816.6036
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416    1204.8629
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924    8023.8657
#> 12868 712532.5 9560332 1913.92 61.937311  -0.03543   -0.05146    1347.6224
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096     520.5570
#> 47011 713262.5 9559722 2003.72 37.192982  -0.01846   -0.00034    4495.6929
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376    2048.3198
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070    2562.7559
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714     475.3120
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679     550.0344
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270    4057.5637
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925     665.6116
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072   54831.6289
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181    2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993   10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252    5761.3486
#> 49569 713232.5 9561122 1868.73 37.312158  -0.00405   -0.01565    2232.6133
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398   0.00953   -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602     345.8933
#> 39732 715362.5 9559582 2230.97 23.346884  -0.00161   -0.00409    2312.8413
#> 23213 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482  -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304    1569.7313
#> 46065 713072.5 9559162 2109.39 16.264553  -0.00195   -0.00656    3036.7383
#> 20601 715542.5 9558182 2773.29 32.755424  -0.00604    0.00354    1147.7384
#> 18958 714192.5 9558372 2495.72 34.677125  -0.00178   -0.00822     472.8748
#> 47320 713422.5 9559182 2143.03 56.325189   0.01246    0.00534     344.1447
#> 46461 712682.5 9560032 1892.30 22.977899   0.01453    0.01337     486.7551
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400     608.6523
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.215474   0.03890   -0.00911     173.2806
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141    4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195    1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863  -0.00666   -0.03583    1898.0105
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781  -0.01439   -0.01161    3065.3694
#> 34164 714212.5 9558492 2421.45 51.329570   0.02247    0.01123     364.2903
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298    1533.7925
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970    1443.9200
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759     158.9370
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294     734.9908
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044     288.5468
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119    6589.5796
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279   39495.1797
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144  317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225  -0.00010   -0.00870    2099.7930
#> 10785 714052.5 9557852 2377.76 20.984006  -0.00025   -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320     911.5911
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426  -0.04989   -0.02781   29995.1543
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382     100.0000
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105    4432.8760
#> 24836 714202.5 9558272 2507.87 38.489013  -0.00226    0.00736     477.8591
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517    1060.5494
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356   0.04336    0.01403     241.2991
#> 35852 715112.5 9559962 2173.61 39.315791  -0.00738    0.00118    2273.4758
#> 49363 712772.5 9559552 1935.91 29.036928  -0.03220   -0.03290    2481.5632
#> 39193 715332.5 9560852 1872.90 30.290560  -0.00161    0.00171     756.8818
#> 1085  715722.5 9557532 3097.22 28.784254   0.02327    0.01833     300.0125
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598     186.3527
#> 39685 715332.5 9558102 2797.04 47.788118   0.00167    0.01133     828.8604
#> 34347 714742.5 9560142 2161.95 26.566907  -0.00978    0.00367    3002.8828
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943     371.5514
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674   0.00084   -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011     712.5228
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902     218.7467
#> 47059 713822.5 9558272 2413.20 38.175032  -0.00687   -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406  -0.00307    0.02997     230.8641
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518    1137.0179
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528     355.3475
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714  480535.5000
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513     161.9878
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740     909.4791
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360    4288.7329
#> 1479  715622.5 9557952 2888.67 34.953863   0.01635   -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.717784   0.00262   -0.00502     520.5218
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960     884.2003
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050    1471.4346
#> 42638 714322.5 9561182 1754.22  4.172279  -0.03757   -0.00003 5426606.0000
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282    1659.6211
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186    2751.4626
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148     481.8292
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805    7489.7417
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270   35950.5508
#> 18504 713372.5 9560672 1786.89  6.473850  -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012     305.3624
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523     888.2811
#> 35257 713822.5 9557982 2350.80 29.307937   0.01549   -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044  -0.00181   -0.00799    1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019   0.00282    0.00458     485.7418
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038  -0.00087   -0.00573    1490.0312
#> 32618 715132.5 9559642 2266.36 34.532739  -0.00305   -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149    4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 7466  28.79055625   255.23   TRUE         0.00             60  3.163437
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26428 35.47525484   300.00   TRUE         0.00            100  3.486483
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 29107 25.23076947   300.00  FALSE       300.00            100  2.550653
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607