Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357    1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104 -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952  0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314    2360.1501
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937  0.00083     994.7876
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 34563 715022.5 9558162 2675.70 18.889846 -0.01993  0.01373   16031.7500
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954  0.01008  0.00102     483.4184
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422 -0.04708  0.00408   76318.0156
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 20601 715542.5 9558182 2773.29 32.755424 -0.00604  0.00354    1147.7384
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 41125 712772.5 9559412 2021.88 38.876332 -0.00310 -0.00860     695.2302
#> 46138 715202.5 9559582 2194.40 18.860052 -0.02143 -0.05078  622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257  0.00044     288.5468
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606     919.4241
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 10213 715542.5 9558782 2532.10 35.413375 -0.00893 -0.00867    2247.8748
#> 33142 715652.5 9558212 2740.94 14.381241 -0.00465 -0.01036    2505.8369
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870    2099.7930
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 16464 715252.5 9559982 2143.24 23.185883 -0.00298 -0.00712    1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118    2273.4758
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 20458 713612.5 9559562 2288.02 33.440109  0.00112  0.03538     265.4677
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 18627 714292.5 9559252 2330.78 18.496224 -0.00538 -0.00052    3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 41753 715722.5 9558192 2771.39 26.085050 -0.00196 -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267   31588.1953
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185  0.01655     560.8554
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032     994.9224
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 7983  713512.5 9560432 1907.38 50.019216  0.03310 -0.07650     766.6254
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309  0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096 -0.00172 -0.00119     792.3544
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860 -0.00011 -0.00419     323.0790
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 15363 714742.5 9559212 2373.13 33.583348 -0.00247  0.01057     722.3952
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 42917 38.8035667   300.00   TRUE       300.00            100  3.010367
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 34563 38.1566973   300.00   TRUE       300.00            100  4.204981
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 37029 28.8931157   300.00  FALSE       237.21            100  2.916688
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 6732  26.8562507   300.00  FALSE       300.00            100  2.948550
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775