Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034  -0.01919   -0.04051 351155.1250
#> 31357 714752.5 9561022 1848.22 33.4464113  -0.00347    0.02357   1752.0375
#> 42632 714852.5 9557902 2674.80 30.6526691   0.00221    0.00969    368.5606
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487   1077.7585
#> 8598  714002.5 9558872 2310.02 39.5839352  -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767   3321.6113
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431   4079.0366
#> 26531 714462.5 9560452 2104.74 19.5241098   0.00453    0.00257    357.1436
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835   2425.3115
#> 20151 714602.5 9559922 2184.66 48.3415951  -0.00979   -0.01542    674.0147
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.2955277   0.00550   -0.00120    512.9740
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842    868.6137
#> 3835  714142.5 9558552 2388.42 32.9003825   0.00979   -0.00269    491.8017
#> 15109 713872.5 9559602 2288.64 25.1855058   0.00602    0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501   3889.7292
#> 18005 714832.5 9557912 2660.76 31.9120303   0.00191   -0.00261    554.9243
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838    238.6685
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628    556.0121
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253   0.00902    0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.2576915  -0.02651   -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.0291164  -0.00521   -0.00739    558.2286
#> 28230 715182.5 9558102 2791.83 39.0717109  -0.00669   -0.00950    950.3138
#> 34565 712972.5 9560172 2007.72 27.9471624   0.00598    0.01212    561.1948
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437   1155.1469
#> 1635  713102.5 9559052 2172.87 31.4840308  -0.00064   -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.5250811  -0.00089   -0.00051    779.0555
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651    419.0927
#> 46756 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514    297.7842
#> 40738 714972.5 9557642 2745.48 42.5426892  -0.00294   -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.1929823  -0.01846   -0.00034   4495.6929
#> 37828 713482.5 9558652 2330.58 20.4454260   0.00606   -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.2139721  -0.03229   -0.00511 361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914  -0.00509    0.00499    885.0064
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714    475.3120
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613    289.6113
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 17036 712832.5 9560582 1951.38 27.6457866   0.00060    0.00270   4057.5637
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903   2360.1536
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 16015 712582.5 9560312 1899.18 56.5469237   0.03480    0.02530    937.1306
#> 21952 715352.5 9560172 2013.44 40.4880626   0.02409    0.02771    674.3859
#> 40733 712862.5 9558892 2108.60 29.9949135   0.00175    0.00236   2715.4138
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542    860.8840
#> 34005 714922.5 9558842 2559.23 52.8312924   0.03658    0.04602    345.8933
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427    380.9472
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009   1247.7061
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684   2556.8049
#> 664   714452.5 9559852 2174.83 44.8608765   0.00282    0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020   1563.5549
#> 13146 713792.5 9558552 2341.40 29.8253180  -0.00104    0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.0591409  -0.00410    0.03620    379.0855
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512  13178.3223
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.5251162  -0.00478   -0.07132  42970.7891
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.3688733  -0.00093    0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.0344196   0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.6601386   0.01469    0.02940    987.8085
#> 1973  715252.5 9557822 2861.00 15.5707010   0.01865    0.00335    195.0293
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412   1350.3611
#> 10933 712702.5 9560132 1854.55 36.2911467  -0.00204   -0.00376   2048.3198
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781  29995.1543
#> 47702 714692.5 9559712 2200.51 33.0986259  -0.13038   -0.04212  46310.0195
#> 43201 715712.5 9557922 2852.82 31.4622584  -0.00949    0.00520   3390.5708
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141   4611.0718
#> 43245 713722.5 9560012 2153.38 34.1912564  -0.01263    0.01322   4605.1411
#> 34164 714212.5 9558492 2421.45 51.3295700   0.02247    0.01123    364.2903
#> 47244 714262.5 9559302 2350.47 17.0592454   0.02213   -0.01363    565.5207
#> 22297 714642.5 9558172 2606.18 59.4690084  -0.00495   -0.03756    353.9086
#> 4667  712652.5 9560472 1957.89 36.3243783  -0.00296    0.00426   1315.5334
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109   2047.9031
#> 20426 714702.5 9559712 2208.61 43.0463191  -0.08153   -0.02417  34908.6562
#> 1997  714752.5 9561012 1853.19 26.5663341  -0.00693    0.01133   2397.2681
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226    338.3474
#> 28260 715312.5 9558302 2681.75 34.4840379  -0.01125   -0.02685   2568.4631
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.4649416   0.00560    0.00610    454.8103
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921    525.3245
#> 13278 712752.5 9560452 1901.25 42.1135439  -0.00118   -0.00252   3631.7017
#> 43834 713092.5 9560662 1905.55 31.3236027   0.00758   -0.00288   2144.8164
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291    487.3438
#> 39466 714942.5 9557772 2680.35 16.7532859   0.00166   -0.01265   1148.6487
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.5955479  -0.00203    0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.9372405   0.00002   -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.6885908   0.00534    0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.6097183  -0.00465   -0.00105   4432.8760
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736    477.8591
#> 34250 713822.5 9559142 2316.36 33.7598192  -0.02284    0.01085   2229.5105
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463   1050.7473
#> 14717 715382.5 9558062 2799.05 50.5904544  -0.00812   -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363    389.2691
#> 7517  714002.5 9561142 1772.85 26.1893279   0.00815   -0.03345 665381.5625
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 9665  713362.5 9558812 2275.28 31.4949170   0.01306   -0.00626    450.7346
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.5533166  -0.00502   -0.00028   2394.7261
#> 13286 713662.5 9559702 2268.14 41.9966605  -0.01341    0.04431    315.9084
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 12974 714662.5 9557332 2667.61 41.9416565  -0.00065    0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.2796736   0.00084   -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.2478257   0.00737    0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.6481135  -0.00844    0.00514    897.9290
#> 33956 714082.5 9558712 2327.95 39.6842028   0.02218    0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462    716.5762
#> 24009 714042.5 9558962 2357.58 34.3619978   0.05041   -0.00011    201.1272
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196    576.1450
#> 45486 712762.5 9560172 1853.82 56.7973062   0.01978    0.01282   2322.7944
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 36889 714302.5 9559802 2243.60 21.1662069   0.01923    0.05547    128.2613
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 29716 715272.5 9559992 2144.94 25.7177836   0.00262   -0.00502    520.5218
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750   1892.8978
#> 40612 713542.5 9561152 1856.15 33.5684513   0.00420    0.01960    884.2003
#> 12300 712372.5 9560062 1903.64 50.0369772   0.00343    0.01177   1558.2231
#> 26283 715932.5 9557592 3075.48 32.7668833  -0.00375    0.00215    917.5296
#> 280   713182.5 9560632 1862.83 39.0041019   0.00745    0.00875   2156.9358
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491    261.7371
#> 20950 714912.5 9557782 2667.71 30.8113784  -0.04770    0.00270  35950.5508
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 37254 713482.5 9560892 1823.73 39.8824462   0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476  -0.00595    0.03064   1485.5201
#> 49812 713392.5 9559222 2133.99 35.6626120  -0.00983    0.00803   1058.5630
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448    941.8503
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491