Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 23178 715382.5 9560142 2021.07 41.9766069  0.00958  0.02642    671.1807
#> 27864 712992.5 9560672 1926.38 27.2240896 -0.00199  0.00659   3553.7166
#> 40756 714022.5 9558862 2331.20 45.0854759 -0.00075  0.00475   1001.0861
#> 24051 715462.5 9559372 2319.46 32.3279977 -0.00661 -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.8482293  0.00046  0.00324    861.3537
#> 11088 715532.5 9558762 2549.28 38.0501272 -0.00373 -0.00057   1630.3259
#> 28293 713762.5 9558812 2250.98 33.2854102  0.00118  0.00761    928.6858
#> 27973 715022.5 9559332 2308.28 20.0632631  0.00940  0.00679    550.0344
#> 42598 713542.5 9559972 2184.39 35.2414881 -0.00841 -0.00010   1527.1028
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378    440.9104
#> 43964 714802.5 9558452 2684.33 37.9825182 -0.02123  0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.3469193  0.00528  0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448 -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.0046749  0.00129  0.02510   3342.5518
#> 18005 714832.5 9557912 2660.76 31.9120303  0.00191 -0.00261    554.9243
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.4003790  0.00151  0.00189   2317.1819
#> 2942  713232.5 9560652 1851.75 34.5350311  0.00856  0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643  0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.5289733 -0.03613 -0.01898    588.3761
#> 4477  715432.5 9559392 2325.25 37.0291164 -0.00521 -0.00739    558.2286
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917    116.1986
#> 47220 715052.5 9557752 2726.93 32.7216197 -0.00664  0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.2735877  0.00382  0.00088   2322.1746
#> 34565 712972.5 9560172 2007.72 27.9471624  0.00598  0.01212    561.1948
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211  0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077   3559.4204
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 47351 714222.5 9558792 2420.94 28.6685162 -0.09566 -0.03355  14919.6260
#> 37828 713482.5 9558652 2330.58 20.4454260  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866  29195.6016
#> 17594 714832.5 9561052 1818.04 46.1672203 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888  0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.5357857  0.00140 -0.00070   2562.7559
#> 46235 712712.5 9561042 2023.11 41.2770255  0.01458  0.00672    565.1107
#> 4605  714852.5 9557892 2677.78 26.9336000  0.00321  0.01169    350.1799
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058    431.4362
#> 37141 714382.5 9559822 2219.36 10.1064025  0.01661  0.01199    196.4080
#> 41037 715532.5 9558792 2529.58 32.8465245 -0.00585 -0.00974   2136.6939
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008  0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423  0.03395  0.01495    266.5888
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759 -0.00400    608.6523
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560    977.4719
#> 15766 715562.5 9558592 2646.16 29.0208853  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.3921689  0.03504  0.05825    162.6688
#> 2595  714912.5 9557662 2700.72 24.1948618 -0.02681 -0.01319   3457.7519
#> 18265 715522.5 9558782 2539.23 35.8562718 -0.00458 -0.00401   1722.6481
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553   1750.4175
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 40270 713312.5 9559012 2186.56 30.9357102 -0.00771 -0.00599   1378.5760
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604  0.00354   1147.7384
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112    745.3315
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 39166 712832.5 9560142 1922.46 39.6601386  0.01469  0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.8054672 -0.00567 -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220   8936.2666
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 10073 715572.5 9558612 2632.78 32.1612670  0.00129  0.00171    788.0004
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261   1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.8276239 -0.00444 -0.00017    759.9937
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476   1404.4454
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707  0.01833    369.3771
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247  0.01123    364.2903
#> 1036  713042.5 9558892 2205.38 31.5098139 -0.00428 -0.00522   1355.7436
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441    424.2680
#> 11352 713002.5 9560562 1858.95 39.8039510 -0.00442 -0.00118  19557.6621
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753   1337.7854
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258  0.00422   1159.2100
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.8354356  0.00181 -0.00001   1184.7952
#> 23783 713062.5 9560252 1983.09 36.4298025  0.00774 -0.00224   1137.3363
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767 -0.03713   6379.7974
#> 48574 712692.5 9560412 1901.97 41.9783258  0.01364 -0.00094   1392.6008
#> 44612 714932.5 9558822 2578.80 38.2552461  0.01257  0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185    953.8614
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708 -0.02538    807.8237
#> 29024 713492.5 9559112 2173.47 53.4323251  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875    597.2893
#> 30626 713302.5 9559172 2106.91 35.9977924  0.01354  0.00245    629.5983
#> 17227 714812.5 9558892 2545.43 54.0717460  0.00458  0.04291    293.4231
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384    947.2075
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400 -0.01210    585.8992
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 11332 712742.5 9560072 1850.34 22.7000149  0.00683 -0.00603   1562.0503
#> 9984  712742.5 9560092 1848.56 23.6127367  0.00614 -0.00093    118.1402
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150    503.5590
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533  0.01493   2212.0525
#> 28162 714062.5 9561152 1785.47 48.4447275  0.02309 -0.03409   1575.9425
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 49835 713272.5 9560372 1898.68 32.8258980  0.00846  0.00794   1363.5321
#> 29125 714992.5 9558792 2584.41 24.6658331  0.01253 -0.00333    422.8840
#> 16878 714232.5 9558172 2485.33 22.8083039  0.00818  0.00502   1428.8546
#> 1085  715722.5 9557532 3097.22 28.7842537  0.02327  0.01833    300.0125
#> 13286 713662.5 9559702 2268.14 41.9966605 -0.01341  0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.7540896  0.00527 -0.02077   8759.9961
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.9416565 -0.00065  0.00475    515.1661
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 8413  713242.5 9561232 1967.45 38.1945125  0.00236  0.00483   1097.3082
#> 21925 714542.5 9559232 2314.33 50.2478257  0.00737  0.01033   3446.5520
#> 41753 715722.5 9558192 2771.39 26.0850495 -0.00196 -0.00175    725.1355
#> 42934 713832.5 9558672 2270.74 27.6119821 -0.03193 -0.00267  31588.1953
#> 48869 712562.5 9560122 1972.89 24.9465824  0.00185  0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.8507379 -0.00782 -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338  0.00518   1137.0179
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118  0.00162    290.2031
#> 7983  713512.5 9560432 1907.38 50.0192155  0.03310 -0.07650    766.6254
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858  0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.8694849 -0.00229  0.00099    909.0560
#> 48988 714682.5 9560852 1942.97 20.2907274  0.00480  0.00740    909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 3868  714542.5 9560622 2024.15 29.4471659  0.00360  0.00750   1892.8978
#> 48184 714202.5 9558222 2489.17 36.1135298  0.02236 -0.00066    306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370 -0.00273 -0.01047   2826.7700
#> 14763 713462.5 9560852 1830.40 47.7399894  0.01354  0.00907   1281.5044
#> 12300 712372.5 9560062 1903.64 50.0369772  0.00343  0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.4446854 -0.00070 -0.00050   1471.4346
#> 280   713182.5 9560632 1862.83 39.0041019  0.00745  0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576   1045.1074
#> 19767 714272.5 9558052 2431.67 55.7247292  0.00361 -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 14881 715482.5 9559232 2376.03 29.9055321 -0.01137 -0.01523    888.2811
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799   1036.6606
#> 24962 715152.5 9560472 1967.37 28.8455602  0.00298  0.00222    991.8300
#> 126   715042.5 9561262 1733.49 30.1106510 -0.00288  0.00448    941.8503
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135   1293.4855
#> 13548 715812.5 9558122 2821.65 15.5082486  0.02751  0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.3816390  0.01336  0.01403    609.8665
#> 6206  714122.5 9560992 1917.28 37.9699131 -0.01023 -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.2726234 -0.01181 -0.01149   4242.3579
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 11088 30.99873559   300.00   TRUE       300.00            100  3.212274
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 37348 39.28771601   300.00   TRUE       300.00            100  4.465317
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30626 30.00350790   300.00   TRUE       300.00              0  2.799064
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491