Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 36052 714842.5 9558892 2483.27 68.789886   -0.04921   -0.12438    753.6304
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753    0.01487   1077.7585
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005    769.3426
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250   -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730   -0.01491   1844.4353
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387   1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088    266.4355
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226   2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432    0.00102    0.00388   1051.4222
#> 23925 715202.5 9557652 2782.99 41.957126   -0.00542    0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189    287.5425
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160   9344.2930
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123    0.01164    895.6606
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 23375 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398    267.2659
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838    238.6685
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450    873.0448
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917    116.1986
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593    939.0466
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064   -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642    613.6390
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051    779.0555
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646   -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318    289.0063
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545    0.02405   5681.2412
#> 12460 712762.5 9559622 1906.41 24.261898    0.01047   -0.00887    366.8537
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566   -0.03355  14919.6260
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.213972   -0.03229   -0.00511 361979.2812
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672    565.1107
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714    475.3120
#> 46171 712822.5 9558932 2089.30 29.854539   -0.01174    0.00025   4425.5151
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843    0.04407    344.8395
#> 37141 714382.5 9559822 2219.36 10.106403    0.01661    0.01199    196.4080
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779   2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771    674.3859
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107    0.02373    527.3528
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560    977.4719
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565   2232.6133
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241    585.6182
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 33350 713132.5 9560672 1896.89 24.695054    0.00306   -0.00436   2602.7390
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 7668  713372.5 9559062 2164.82 37.934963   -0.01807   -0.01883    616.1993
#> 46065 713072.5 9559162 2109.39 16.264553   -0.00195   -0.00656   3036.7383
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620    379.0855
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240   1201.4724
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594  16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024   -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743   -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467   -0.00567   -0.03793    864.9472
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453    0.01337    486.7551
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911    173.2806
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728   1137.2450
#> 7086  712762.5 9560962 2022.00 44.913589   -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797   7557.7866
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132   1580.1621
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833    369.3771
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298   1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441    424.2680
#> 49133 714162.5 9558462 2451.13 21.890426    0.00042    0.00957    495.4453
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171   4425.5410
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267    0.00483    343.1384
#> 42617 714092.5 9558392 2469.21 42.799947   -0.01397    0.00526    706.1463
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477   2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152    850.4633
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451   -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358    802.7622
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118   -0.00252   3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279  39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746    0.00458    0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639    0.00004   -0.02384    947.2075
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744    226.7911
#> 31491 712832.5 9558892 2091.39 29.700413    0.00186    0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780   -0.00486    0.02306   1718.3652
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996   1599.3491
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559    866.5775
#> 27884 714072.5 9559202 2446.27 19.756731   -0.00009    0.00109    605.1908
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 25212 714272.5 9559012 2433.45 37.671402   -0.01909   -0.01181   1286.3949
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831    454.9113
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812   -0.02208    951.3401
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466   2633.9666
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794   1363.5321
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014    271.4219
#> 44396 715512.5 9558102 2845.31 26.598420    0.01263    0.03598    186.3527
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759   1880.4403
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175    725.1355
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170    253.5129
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889   -0.00674   -0.00376    629.8236
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902    218.7467
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400   4126.7812
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647    0.02785    0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.074729   -0.00112   -0.00038   2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948   1222.5433
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.515348    0.00972    0.00308    422.5022
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537   -0.00577    423.5598
#> 38362 714402.5 9558042 2499.31 56.901012    0.03044   -0.00964    543.0025
#> 44434 713022.5 9558762 2161.07 31.454237   -0.00273   -0.01047   2826.7700
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343    0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852    0.00248   1432.4019
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361   -0.01841    756.0055
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182   -0.00972    331.7462
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452    0.00368   1071.8734
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284   -0.00886  43249.3945
#> 16443 714472.5 9559532 2222.55 16.744119    0.03263   -0.02012    305.3624
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172   -0.00119    792.3544
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181   -0.00799   1036.6606
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458    485.7418
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419    323.0790
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 30296 713322.5 9559782 2042.96 52.621017   -0.02110   -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247    0.01057    722.3952
#> 47629 714472.5 9558482 2437.33 37.272623   -0.01181   -0.01149   4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 29472 18.64863032   111.09   TRUE         0.00             25  3.424505
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 17322 31.66508551    75.77   TRUE       138.35              1  3.235115
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 33956 38.59558300   300.00  FALSE       300.00            100  2.339942
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 38362 38.26785114   300.00  FALSE       300.00            100  2.734802
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607