Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813   1130.5791
#> 28223 712742.5 9560482 1927.69 34.66796 -0.00292  0.00712   3636.5303
#> 8598  714002.5 9558872 2310.02 39.58394 -0.00498 -0.00752   1321.5997
#> 28293 713762.5 9558812 2250.98 33.28541  0.00118  0.00761    928.6858
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.36168 -0.04951 -0.00529   3282.1186
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 38827 713412.5 9560472 1869.10 16.87132 -0.00156  0.00406   2421.1145
#> 7349  714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 30125 715162.5 9558992 2438.18 39.85666  0.00471 -0.01382    560.8309
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 6808  715192.5 9559492 2219.03 21.81766  0.00713 -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.02912 -0.00521 -0.00739    558.2286
#> 38077 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 45104 715062.5 9559242 2345.02 24.15647 -0.01743  0.01313    783.4548
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.16722 -0.01353  0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 24100 715132.5 9558992 2420.18 29.21397 -0.03229 -0.00511 361979.2812
#> 20001 712702.5 9560132 1854.55 36.29115 -0.00204 -0.00376   2048.3198
#> 46201 714832.5 9561022 1845.94 34.57399 -0.00856  0.01796   4729.6382
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.28922  0.00006  0.00175    804.5679
#> 26567 714022.5 9558872 2327.42 46.66569  0.00290  0.00230   1056.5876
#> 4031  715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 25244 713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252   5761.3486
#> 38862 712952.5 9559892 1950.86 37.72182 -0.00048 -0.00332   1418.6622
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283    590.1536
#> 2595  714912.5 9557662 2700.72 24.19486 -0.02681 -0.01319   3457.7519
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 19966 715362.5 9559572 2235.06 26.04953  0.00126 -0.00485   2240.4185
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.27489 -0.00845  0.00454   3476.0818
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.00795 -0.00078  0.00219    420.7032
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 2552  714942.5 9557652 2719.32 34.87365 -0.01238 -0.00851   1012.5963
#> 42765 713552.5 9559632 2186.95 44.01118  0.00394 -0.00224    942.0322
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100
#> 7041  713472.5 9559092 2137.03 18.34897 -0.10372 -0.00008 825875.6875
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 46461 712682.5 9560032 1892.30 22.97790  0.01453  0.01337    486.7551
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 19892 714482.5 9560562 2059.63 19.81517 -0.00332 -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141
#> 14545 713562.5 9559602 2222.17 54.23905  0.01011  0.00739    789.7573
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370
#> 23783 713062.5 9560252 1983.09 36.42980  0.00774 -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.04632 -0.08153 -0.02417  34908.6562
#> 34984 715042.5 9559312 2320.49 42.85782 -0.01106 -0.04634    500.5027
#> 39405 713122.5 9559052 2169.30 37.64619 -0.02110 -0.02450   2589.9910
#> 14950 713382.5 9559172 2111.91 48.10668  0.00915 -0.02705    413.5702
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 46627 714852.5 9558932 2459.24 53.44550  0.02451 -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.49759 -0.00622  0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421
#> 31491 712832.5 9558892 2091.39 29.70041  0.00186  0.00063   2430.2278
#> 3771  714842.5 9561152 1724.51 45.53983  0.02405 -0.05655    130.9468
#> 6470  713742.5 9558802 2260.79 30.68017  0.00000  0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559    866.5775
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736    477.8591
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902
#> 14717 715382.5 9558062 2799.05 50.59045 -0.00812 -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 32869 715332.5 9558892 2533.17 20.63507 -0.00242  0.01132    737.5982
#> 12716 713962.5 9557762 2402.60 18.93224  0.03480  0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 39193 715332.5 9560852 1872.90 30.29056 -0.00161  0.00171    756.8818
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 21925 714542.5 9559232 2314.33 50.24783  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.50713  0.00601 -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.86373  0.01180  0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 13737 714272.5 9559622 2257.05 55.45716 -0.03224 -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.21475 -0.01338  0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032    994.9224
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027
#> 3868  714542.5 9560622 2024.15 29.44717  0.00360  0.00750   1892.8978
#> 44434 713022.5 9558762 2161.07 31.45424 -0.00273 -0.01047   2826.7700
#> 9007  714952.5 9560902 1863.75 32.77032  0.00074 -0.02904    608.3561
#> 31098 714352.5 9559042 2410.15 30.33640  0.00537 -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544
#> 31440 714842.5 9558392 2704.41 25.80144  0.00992 -0.01052    451.7209
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857
#> 40043 713322.5 9560562 1802.09 39.85208  0.04456 -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.29741  0.00505  0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 27973 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 25169 36.6062735   195.00   TRUE        47.05              0  3.049487
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 2858  35.3194103   300.00   TRUE       300.00              2  2.262887
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 28046 19.5911459   300.00   TRUE       300.00             15  2.896560
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 44434 36.3197946   300.00  FALSE       300.00            100  3.451290
#> 9007  30.9643581   300.00  FALSE        14.98             35  2.784158
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 47245 34.9762086   300.00  FALSE       150.03            100  2.509309
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491