Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.9523260  -0.00123    0.00393   2081.0056
#> 15762 714922.5 9558862 2535.97 54.9105562   0.01305   -0.00555    430.4604
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741    746.4155
#> 46178 714892.5 9559312 2379.54 32.7668833  -0.00266    0.02896    276.0310
#> 1300  714042.5 9558482 2408.27 24.1484522   0.00659    0.01041    772.9746
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339   1002.5877
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 20799 714792.5 9561002 1857.29 23.1291603   0.00090    0.00890   2301.0454
#> 9669  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431   4079.0366
#> 38827 713412.5 9560472 1869.10 16.8713152  -0.00156    0.00406   2421.1145
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264   5057.1431
#> 10974 715312.5 9557502 2844.23 26.6648828  -0.00831   -0.00149  12632.7793
#> 42598 713542.5 9559972 2184.39 35.2414881  -0.00841   -0.00010   1527.1028
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189    287.5425
#> 18843 715192.5 9559332 2265.75 41.1532666  -0.07664   -0.05766 140591.4062
#> 38410 713712.5 9561172 1776.08 36.3192217  -0.02678   -0.04272   4955.5586
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769   1032.7177
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.1946661  -0.02131    0.00501   3889.7292
#> 34275 715332.5 9558442 2650.77 45.5369667  -0.00043    0.00973    514.5488
#> 6808  715192.5 9559492 2219.03 21.8176599   0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314   2360.1501
#> 8073  712572.5 9560302 1917.62 52.0090979   0.01937    0.00083    994.7876
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514    297.7842
#> 47351 714222.5 9558792 2420.94 28.6685162  -0.09566   -0.03355  14919.6260
#> 11687 714552.5 9560382 2052.66 29.5755084  -0.00255   -0.00365   1536.2647
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866  29195.6016
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517   1000.7119
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907   1943.4092
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175    804.5679
#> 37341 714412.5 9559142 2377.24 40.5997894  -0.00735   -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.5552733  -0.04492    0.00072  54831.6289
#> 26567 714022.5 9558872 2327.42 46.6656935   0.00290    0.00230   1056.5876
#> 2976  715122.5 9559142 2390.05 33.1679538   0.01008    0.00102    483.4184
#> 23029 713502.5 9559102 2174.07 56.4575423   0.03395    0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354   0.00283    0.00117   1018.2141
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731   1384.7959
#> 36120 714622.5 9559942 2166.51 32.9863262  -0.02917   -0.01993  10308.6406
#> 4617  714292.5 9559312 2342.95 32.3921689   0.03504    0.05825    162.6688
#> 49569 713232.5 9561122 1868.73 37.3121575  -0.00405   -0.01565   2232.6133
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197    711.1091
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357    473.4709
#> 48808 715032.5 9557792 2707.02 25.4376072  -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241    585.6182
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542    860.8840
#> 2595  714912.5 9557662 2700.72 24.1948618  -0.02681   -0.01319   3457.7519
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739    365.7626
#> 29801 714392.5 9559162 2356.18 31.5625261  -0.00856   -0.01625   1063.9150
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553   1750.4175
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558   1504.9717
#> 16370 713852.5 9558632 2296.97 34.6244125  -0.01972    0.03512  13178.3223
#> 24501 713302.5 9559142 2125.66 27.1089251   0.01784    0.00596    842.0994
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318    289.0063
#> 41783 715572.5 9558652 2609.51 32.4082118   0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.4746819   0.00629   -0.00469   1021.5407
#> 37769 714412.5 9558562 2382.08 39.8767166  -0.04601   -0.08389  47022.9688
#> 30268 714172.5 9558612 2328.84 54.9506633   0.03743   -0.05242    507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562  -0.01169   -0.03220   8936.2666
#> 43859 714892.5 9557602 2722.70 28.1293630   0.00370   -0.00190    640.8842
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281   1178.4420
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.8600517  -0.02143   -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195   1901.6743
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372   4154.5786
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245    629.5983
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.9813582   0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 47086 713692.5 9559972 2175.51 25.0457041  -0.00789   -0.00261   3353.5422
#> 11352 713002.5 9560562 1858.95 39.8039510  -0.00442   -0.00118  19557.6621
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396    704.7076
#> 332   714852.5 9561102 1787.11 45.3582039   0.00540    0.00879   1200.4210
#> 10617 713232.5 9560722 1879.95 26.4970062   0.00258    0.00422   1159.2100
#> 14192 715162.5 9558122 2772.18 44.4626708  -0.02970    0.00200   1305.6691
#> 23783 713062.5 9560252 1983.09 36.4298025   0.00774   -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109   2047.9031
#> 48574 712692.5 9560412 1901.97 41.9783258   0.01364   -0.00094   1392.6008
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483    343.1384
#> 24723 712942.5 9560202 1983.18 33.0716969   0.00285   -0.01325    595.0426
#> 38103 715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538    807.8237
#> 46792 713862.5 9559672 2272.03 35.4649416   0.00560    0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.4975862  -0.00622    0.00152   1530.5657
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.1135439  -0.00118   -0.00252   3631.7017
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856   1789.1596
#> 4479  713982.5 9557812 2399.24 41.2311889   0.01895    0.04744    226.7911
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867   2247.8748
#> 35458 715242.5 9558442 2623.01 38.9548275  -0.00770   -0.03400   1982.7782
#> 30378 714852.5 9557752 2651.70 41.2982249  -0.00010   -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063   2430.2278
#> 49346 715322.5 9558782 2553.74 25.3952084   0.00319    0.00291    487.3438
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403    474.8613
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609   4275.3457
#> 36744 714922.5 9557722 2692.39 25.9807712   0.00316   -0.00427    866.9398
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177    997.3089
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 31071 713472.5 9558462 2298.33 29.1532385   0.00156   -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715    520.7355
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463    851.6306
#> 49363 712772.5 9559552 1935.91 29.0369281  -0.03220   -0.03290   2481.5632
#> 9665  713362.5 9558812 2275.28 31.4949170   0.01306   -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171    756.8818
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199    836.4196
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399   1601.4852
#> 8413  713242.5 9561232 1967.45 38.1945125   0.00236    0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011    712.5228
#> 47059 713822.5 9558272 2413.20 38.1750320  -0.00687   -0.00223   1056.3701
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518   1137.0179
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196    576.1450
#> 33666 713632.5 9561012 1775.88 21.1788119  -0.00950   -0.05130    107.5831
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.1838220   0.00270   -0.00250   2225.4426
#> 46509 712512.5 9559532 1860.09 89.9887513  -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.1906554   0.01858    0.00312    254.0656
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851  -0.02608    0.00148  10771.1963
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417   1520.6630
#> 3868  714542.5 9560622 2024.15 29.4471659   0.00360    0.00750   1892.8978
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964    543.0025
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578   1258.8865
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.3363964   0.00537   -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532  16771.3066
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491    261.7371
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148    481.8292
#> 10981 713832.5 9560022 2125.33 34.8484390  -0.06909   -0.01151 143213.1562
#> 1577  713322.5 9559742 2027.47 29.5674870  -0.03284   -0.00886  43249.3945
#> 20950 714912.5 9557782 2667.71 30.8113784  -0.04770    0.00270  35950.5508
#> 36016 715262.5 9557682 2814.30 37.0021874   0.00203    0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474   0.00340    0.01809   1158.0426
#> 14881 715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523    888.2811
#> 9526  714432.5 9560582 2076.13 31.0800956  -0.00172   -0.00119    792.3544
#> 39883 712732.5 9560842 2082.69 25.0909678   0.00732    0.01858    298.4490
#> 46148 715082.5 9558042 2725.39 42.8360436  -0.00181   -0.00799   1036.6606
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281    492.4785
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072   0.00505    0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661   5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 42598 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 18843 28.5264227   300.00   TRUE       300.00            100  5.147959
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 49569 41.8494103    25.00   TRUE        35.00            100  3.348814
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 48808 33.7735702   300.00   TRUE       300.00            100  4.066528
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 37769 37.4250302   300.00   TRUE       300.00            100  4.672310
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 46138 32.6946907   300.00   TRUE       300.00              8  5.794410
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 11352 27.9981556   135.89   TRUE        82.11              4  4.291317
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 14192 40.5866113   300.00   TRUE       300.00            100  3.115833
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 46792 27.6010959   300.00   TRUE       300.00             18  2.657830
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 31098 34.6576441   300.00  FALSE       300.00             45  3.277234
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491