Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123   0.00393   2081.0056
#> 16749 715072.5 9558242 2745.27 49.053782    0.00097   0.01483    566.7674
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266   0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240  -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.050127   -0.00373  -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857   0.03677   1675.4679
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250  -0.01340    604.1032
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645   0.00835   2425.3115
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021  -0.09328   2235.5259
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979  -0.00269    491.8017
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528   0.01392   1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070  -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602   0.01518    854.4269
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501   3889.7292
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126  -0.02044 129412.3984
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832   0.00838    238.6685
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189   2317.1819
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493   0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396    0.00993  -0.00413    364.9788
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669  -0.00950    950.3138
#> 15390 713132.5 9560622 1873.26 38.480418    0.00252   0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452  -0.01437   1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616   0.01276   9614.6924
#> 7879  714722.5 9559232 2356.19 36.803944   -0.01097   0.00156    816.6036
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916   2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275   0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077   3559.4204
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804   0.00346    543.0749
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547   0.03323    183.1836
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888   0.00862    874.5408
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714    475.3120
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504   1020.2766
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 29926 714702.5 9557622 2637.00 44.642580    0.01843   0.04407    344.8395
#> 41037 715532.5 9558792 2529.58 32.846524   -0.00585  -0.00974   2136.6939
#> 2976  715122.5 9559142 2390.05 33.167954    0.01008   0.00102    483.4184
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286  -0.00964    571.1364
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530    937.1306
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855  -0.02175   1122.9120
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825    162.6688
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357    473.4709
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221   0.00121   4159.6738
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739    365.7626
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099   0.02169    495.8840
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020   1563.5549
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851   1012.5963
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.960131   -0.00320   0.00180   3904.5254
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604   0.00354   1147.7384
#> 39722 715022.5 9557712 2735.53 30.260766    0.00133   0.00367    806.0966
#> 8136  713242.5 9560602 1798.18 41.240356   -0.01169  -0.03220   8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178  -0.00822    472.8748
#> 7041  713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008 825875.6875
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986  -0.02524  10861.9482
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631   -0.01491   0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509  -0.01211  49347.5547
#> 24135 712612.5 9559492 1903.67 46.756794   -0.06953  -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949   0.00520   3390.5708
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070  -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175   0.04195   1901.6743
#> 19832 715182.5 9557662 2761.27 38.432863   -0.00666  -0.03583   1898.0105
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466  -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707   0.01833    369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379   0.00801    172.7491
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250  -0.00970   1443.9200
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484    495.0292
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102  -0.01778   4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495  -0.03756    353.9086
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001   1184.7952
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109   2047.9031
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693   0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181  -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921    525.3245
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622   0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000 162178.1562
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632    620.0229
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400  -0.01210    585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529   0.01230   1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454   0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465  -0.01036   2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737    0.00614  -0.00093    118.1402
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000   0.00320    911.5911
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084  -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203   0.00642   1688.1782
#> 25584 715062.5 9558532 2675.93 37.688591    0.00534   0.01476   1275.9446
#> 11288 712822.5 9560442 1873.42 39.609718   -0.00465  -0.00105   4432.8760
#> 748   715102.5 9559692 2301.39 27.021836   -0.00759   0.00158    837.3730
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831    454.9113
#> 14717 715382.5 9558062 2799.05 50.590454   -0.00812  -0.02208    951.3401
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242   0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220  -0.03290   2481.5632
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626    450.7346
#> 45071 712652.5 9560032 1900.80 14.473487    0.00987  -0.00507    504.3085
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973  -0.00943    371.5514
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362    283.7090
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196  -0.00175    725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490  -0.00780    825.4451
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011    201.1272
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004   0.00714 480535.5000
#> 20764 714542.5 9558152 2602.18 52.078999    0.02118   0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.933656   -0.00342   0.02552   2515.0752
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270  -0.00250   2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479   0.02251    168.1733
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873  -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 17628 714032.5 9560252 2160.03 30.294570    0.00537  -0.00577    423.5598
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236  -0.00066    306.0146
#> 26283 715932.5 9557592 3075.48 32.766883   -0.00375   0.00215    917.5296
#> 19767 714272.5 9558052 2431.67 55.724729    0.00361  -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293   0.00186   2751.4626
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163  -0.00377   1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309   0.01011    501.7867
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172  -0.00119    792.3544
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274   -0.01305   0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386   -0.00933  -0.00157    766.0220
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456  -0.01586    941.0288
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 16749 32.183039   300.00   TRUE       300.00            100  2.753405
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 11088 30.998736   300.00   TRUE       300.00            100  3.212274
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 27973 30.009810   300.00   TRUE       300.00             85  2.740390
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 38993 29.548579   300.00   TRUE       300.00              2  3.128238
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 45575 25.728097   300.00   TRUE       300.00             59  2.562268
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 15390 29.322261    60.00   TRUE       118.92              2  3.372821
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 1635  27.826841   300.00   TRUE       300.00            100  3.303016
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 6967  41.119462   300.00   TRUE       300.00             85  3.008718
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 29926 32.094804   300.00   TRUE       300.00            100  2.537617
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 2976  26.882034   300.00   TRUE       300.00            100  2.684323
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 9704  38.996080   300.00   TRUE       300.00             90  2.784369
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 31121 38.160135   300.00   TRUE       300.00             45  3.194113
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 39722 26.541697   300.00   TRUE       300.00            100  2.906387
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 7041  33.218947   300.00   TRUE       300.00              6  5.916915
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 19832 36.633776   300.00   TRUE       300.00            100  3.278299
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 23783 27.600523   300.00   TRUE         0.00            100  3.055889
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 22847 24.747766   300.00   TRUE       300.00             41  3.134623
#> 46627 53.683281   300.00   TRUE       300.00            100  2.720428
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 28588 31.625551   300.00   TRUE       300.00             78  5.209992
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 10785 33.245303   300.00   TRUE       300.00            100  3.383302
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 6470  28.093267   300.00   TRUE       300.00              2  2.959800
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 748   38.132633   300.00  FALSE       300.00             12  2.922919
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 41753 25.007316   300.00  FALSE       300.00            100  2.860419
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 24009 38.497034   300.00  FALSE       300.00            100  2.303471
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 20764 39.850361   300.00  FALSE       300.00             91  2.462702
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 35038 40.203876   300.00  FALSE       300.00            100  3.347416
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 26283 41.510792   300.00  FALSE       300.00            100  2.962620
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 9526  27.569583   300.00  FALSE       300.00            100  2.898919
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 40043 23.996618   194.79  FALSE        24.89             54  2.973603