Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 42632 714852.5 9557902 2674.80 30.65267  0.00221  0.00969    368.5606
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 21939 713642.5 9558712 2299.09 33.75065 -0.00503  0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.12916  0.00090  0.00890   2301.0454
#> 8869  715062.5 9559392 2278.56 41.40193  0.00310  0.01440   1088.8639
#> 25169 712642.5 9560232 1883.10 46.84503  0.00884  0.01345   1120.6929
#> 43669 712392.5 9560162 2001.90 47.28392 -0.00284  0.01804    628.5153
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974
#> 23817 714842.5 9557832 2677.58 26.65743  0.00922  0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.29553  0.00550 -0.00120    512.9740
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.81405  0.00952  0.01378    440.9104
#> 44328 713222.5 9558932 2227.94 29.15553 -0.00769 -0.00281   1325.6138
#> 8019  714672.5 9559272 2316.84 40.38149 -0.01966 -0.02234    849.0547
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692  0.00528  0.01392   1255.0884
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240
#> 18005 714832.5 9557912 2660.76 31.91203  0.00191 -0.00261    554.9243
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961
#> 35942 712862.5 9558902 2109.97 31.40038  0.00151  0.00189   2317.1819
#> 25081 713392.5 9558452 2242.24 46.53850 -0.02716  0.03696   4143.4019
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121
#> 4827  714962.5 9559882 2274.42 29.24491  0.00051  0.00889    362.8363
#> 6676  715092.5 9559912 2217.44 41.70789 -0.00077 -0.00284   2010.4285
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.02089  0.00422  0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.11484 -0.01452 -0.01437   1155.1469
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390
#> 27013 713062.5 9559782 1976.81 41.63226  0.01414  0.03416   1204.8629
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498
#> 46756 713372.5 9559192 2120.47 46.00507  0.03696  0.02514    297.7842
#> 29667 713892.5 9559712 2239.25 41.59273  0.00804  0.00346    543.0749
#> 47351 714222.5 9558792 2420.94 28.66852 -0.09566 -0.03355  14919.6260
#> 19784 714982.5 9557722 2716.98 35.43629  0.00888  0.00862    874.5408
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 651   714602.5 9560152 2111.23 53.85173 -0.01149  0.01949   1182.1506
#> 44550 714842.5 9557642 2696.38 15.17307  0.00359 -0.00279    930.0871
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335
#> 4617  714292.5 9559312 2342.95 32.39217  0.03504  0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.71707 -0.03260 -0.03310   2833.9001
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 20065 713772.5 9558852 2221.21 28.01993 -0.00568 -0.02553   1750.4175
#> 32280 712572.5 9560182 1956.86 50.67812 -0.02554  0.03304   1569.7313
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020   1563.5549
#> 38197 712522.5 9560322 1942.04 62.77727 -0.01611  0.02961   1086.6963
#> 24501 713302.5 9559142 2125.66 27.10893  0.01784  0.00596    842.0994
#> 28046 714882.5 9558492 2711.77 28.30583 -0.00131  0.00201    788.0621
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094
#> 37769 714412.5 9558562 2382.08 39.87672 -0.04601 -0.08389  47022.9688
#> 39166 712832.5 9560142 1922.46 39.66014  0.01469  0.02940    987.8085
#> 32879 713912.5 9559672 2238.00 31.97677 -0.00278 -0.00412   1350.3611
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 17345 713922.5 9559682 2227.44 34.49091 -0.06509 -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.75123 -0.02556 -0.11694 102301.4609
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 34164 714212.5 9558492 2421.45 51.32957  0.02247  0.01123    364.2903
#> 47086 713692.5 9559972 2175.51 25.04570 -0.00789 -0.00261   3353.5422
#> 10363 713252.5 9559722 1999.55 41.82993 -0.03669 -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.40342  0.03051  0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.26762 -0.00317  0.00077   3056.9475
#> 49378 713252.5 9560732 1875.41 27.83544  0.00181 -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 29024 713492.5 9559112 2173.47 53.43233  0.04864  0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.33619  0.00181 -0.00341   1363.4000
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231
#> 4479  713982.5 9557812 2399.24 41.23119  0.01895  0.04744    226.7911
#> 14140 714672.5 9559262 2326.36 45.79251 -0.01400 -0.01210    585.8992
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 9984  712742.5 9560092 1848.56 23.61274  0.00614 -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024
#> 25212 714272.5 9559012 2433.45 37.67140 -0.01909 -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.97550  0.00872  0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.43175  0.00390 -0.02831    454.9113
#> 31071 713472.5 9558462 2298.33 29.15324  0.00156 -0.00116   1817.5222
#> 8635  713832.5 9557902 2397.71 47.11948  0.02295  0.02715    520.7355
#> 35852 715112.5 9559962 2173.61 39.31579 -0.00738  0.00118   2273.4758
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525
#> 20067 714242.5 9560442 2125.23 38.64027  0.00363 -0.00363    389.2691
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078
#> 37832 712552.5 9559192 1998.10 41.55090 -0.00195 -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196
#> 49835 713272.5 9560372 1898.68 32.82590  0.00846  0.00794   1363.5321
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843
#> 40139 712852.5 9560072 1907.89 43.70407  0.00001  0.00759   1880.4403
#> 48183 713642.5 9558262 2314.44 50.02323  0.01338 -0.00399   1601.4852
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 38689 714622.5 9559342 2264.90 42.64811 -0.00844  0.00514    897.9290
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 44267 712972.5 9559042 2191.01 22.26915  0.01462  0.00268    252.0072
#> 33410 715342.5 9560452 1942.88 29.09308 -0.00305 -0.00705    968.9738
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 42424 714692.5 9560902 1921.34 25.33104  0.01154 -0.00204    618.2027
#> 24819 714032.5 9560882 1981.86 33.28426  0.00804 -0.00165    805.1403
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922
#> 24446 714742.5 9560082 2200.60 28.72180 -0.00598  0.02208    842.4825
#> 7132  713192.5 9559612 2060.27 20.77373  0.00452  0.00368   1071.8734
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 37254 713482.5 9560892 1823.73 39.88245  0.01014 -0.00884    672.2466
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790
#> 32618 715132.5 9559642 2266.36 34.53274 -0.00305 -0.01135   1293.4855
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857
#> 15363 714742.5 9559212 2373.13 33.58335 -0.00247  0.01057    722.3952
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 25169 36.60627353   195.00   TRUE        47.05              0  3.049487
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 44328 25.50407033   300.00   TRUE       300.00            100  3.122417
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 2990  25.32129680   300.00   TRUE       123.56              5  2.520462
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47086 23.38699128   300.00   TRUE       300.00            100  3.525504
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 4479  32.01172497   300.00   TRUE       300.00             55  2.355626
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491