Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642    671.1807
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347  0.02357   1752.0375
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813   1130.5791
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250   1413.3312
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005    769.3426
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431   4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949   6217.8213
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804    628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406   2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588   1295.1610
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088    266.4355
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149  12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842    868.6137
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281   1325.6138
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104   1641.4587
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250   3686.1240
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114   4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773    179.3077
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950    950.3138
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616  0.01276   9614.6924
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077   3559.4204
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651    419.0927
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365   1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556    660.3690
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866  29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511 361979.2812
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070   2562.7559
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338  10787.6299
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714    475.3120
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626   1014.7159
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289
#> 23029 713502.5 9559102 2174.07 56.457542  0.03395  0.01495    266.5888
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117   1018.2141
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731   1384.7959
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357    473.4709
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166    650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262   1760.2162
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941   1290.7709
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316    454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469   1021.5407
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093  0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241   1752.9242
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281   1178.4420
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334   4905.2002
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123    364.2903
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396    704.7076
#> 12629 714712.5 9558872 2533.15 37.025679  0.04250 -0.00970   1443.9200
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879   1200.4210
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669 -0.01981   3975.3608
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281  0.04759    158.9370
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426   1315.5334
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634    500.5027
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705    413.5702
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685   2568.4631
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341   1363.4000
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905   6487.0010
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015    790.2421
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144 317260.6875
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010 -0.00870   2099.7930
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403    474.8613
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655    130.9468
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345   2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093    118.1402
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609   4275.3457
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150    503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505    351.4078
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158    837.3730
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222
#> 35852 715112.5 9559962 2173.61 39.315791 -0.00738  0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452    578.5847
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466   2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494  0.01064    683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445    355.5078
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075   8141.7549
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626    450.7346
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171    756.8818
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507    504.3085
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900    614.5898
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726  0.00706 -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298    436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902    218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422    967.9738
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997    230.8641
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048   1550.6704
#> 34810 714112.5 9559572 2337.51 10.288030 -0.00393  0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759   4131.5356
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714 480535.5000
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705    968.9738
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257    549.3252
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250   2225.4426
#> 15008 714492.5 9559012 2449.93 29.515348  0.00972  0.00308    422.5022
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964    543.0025
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375  0.00215    917.5296
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264  0.00624   1870.8307
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368   1071.8734
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805   7489.7417
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886  43249.3945
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012    305.3624
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523    888.2811
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858    298.4490
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803   1058.5630
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799   1036.6606
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109    915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222    991.8300
#> 26635 715092.5 9560412 2010.59 30.571882 -0.00463  0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156    758.4840
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 36052 53.7451601   300.00   TRUE       300.00            100  2.877158
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 22735 36.5168921   300.00   TRUE       300.00             90  2.622310
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 12629 32.1807475   300.00   TRUE       300.00            100  3.159543
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 30626 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 14763 10.1447907   110.04  FALSE        60.35            100  3.107720
#> 26283 41.5107923   300.00  FALSE       300.00            100  2.962620
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 14881 26.8562507   300.00  FALSE       300.00            100  2.948550
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946