Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423 -0.00681 -0.00029   5577.3916
#> 21788 715392.5 9560172 1988.71 40.47660 -0.01919 -0.04051 351155.1250
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304
#> 27864 712992.5 9560672 1926.38 27.22409 -0.00199  0.00659   3553.7166
#> 44302 714932.5 9557982 2649.56 37.30070  0.01633 -0.01813   1130.5791
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585
#> 1300  714042.5 9558482 2408.27 24.14845  0.00659  0.01041    772.9746
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.32799 -0.00333 -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426
#> 37104 713962.5 9561282 1795.56 25.74185  0.00273 -0.02403   1787.3024
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147
#> 26447 715632.5 9559102 2410.77 39.65670  0.00730  0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228
#> 38993 714202.5 9558402 2475.75 33.15936 -0.01607  0.00677   1343.5006
#> 23925 715202.5 9557652 2782.99 41.95713 -0.00542  0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119
#> 39061 712922.5 9559912 1931.50 36.31521  0.00414 -0.00104   1641.4587
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938
#> 29420 713592.5 9560772 1826.82 34.92693 -0.00070 -0.00530   9000.4356
#> 18304 714582.5 9558932 2463.25 22.19467 -0.02131  0.00501   3889.7292
#> 38279 714192.5 9558612 2325.14 28.25255 -0.04430 -0.06070 138869.3906
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721
#> 31576 713942.5 9558492 2387.12 46.14316 -0.05634 -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.62670 -0.01537 -0.00203  10263.2734
#> 26215 714332.5 9560262 2138.08 25.72122 -0.00491  0.00581   1855.1630
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983
#> 36230 713322.5 9560662 1798.99 40.12825 -0.00033 -0.03917    116.1986
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.72162 -0.00664  0.00694   2123.4473
#> 34280 713652.5 9560702 1896.37 41.27359  0.00382  0.00088   2322.1746
#> 5749  712852.5 9560162 1934.78 23.38871  0.00071  0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.07171 -0.00669 -0.00950    950.3138
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412
#> 40738 714972.5 9557642 2745.48 42.54269 -0.00294 -0.00096    520.5570
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929
#> 34359 712862.5 9558912 2111.37 32.81215  0.00463  0.00167   1938.5389
#> 34563 715022.5 9558162 2675.70 18.88985 -0.01993  0.01373  16031.7500
#> 46235 712712.5 9561042 2023.11 41.27703  0.01458  0.00672    565.1107
#> 5578  714962.5 9557612 2751.95 39.19146  0.00288  0.01452    437.3965
#> 13699 712902.5 9559902 1917.21 30.48766 -0.00093 -0.00907   1943.4092
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063  18733.8848
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362
#> 39278 714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159
#> 3944  715262.5 9558352 2654.20 33.55527 -0.04492  0.00072  54831.6289
#> 2976  715122.5 9559142 2390.05 33.16795  0.01008  0.00102    483.4184
#> 34707 713962.5 9561092 1819.49 65.68617 -0.04749  0.04779   2972.0193
#> 5084  714912.5 9559302 2359.77 47.96803  0.00107  0.02373    527.3528
#> 2718  715062.5 9558222 2719.56 44.75316 -0.00119 -0.01560    977.4719
#> 21053 712862.5 9559992 1901.01 30.95061 -0.00762  0.01252   5761.3486
#> 42941 713252.5 9559742 1990.10 28.95442 -0.04708  0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.64888 -0.01282  0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.13740  0.00953 -0.02283    590.1536
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738
#> 6317  715042.5 9557702 2745.36 27.87382  0.00090 -0.00299    688.0706
#> 28845 714892.5 9559272 2374.42 45.75011 -0.00209  0.02739    365.7626
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162
#> 46783 713592.5 9560762 1832.60 36.01040  0.00448 -0.00328    592.3326
#> 44650 715612.5 9558142 2764.88 39.36048 -0.00550 -0.00941   1290.7709
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588
#> 7385  713752.5 9560122 2064.61 31.77108 -0.02605 -0.03594  16587.9980
#> 12137 712842.5 9558902 2097.88 30.09862  0.00024 -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.95066  0.03743 -0.05242    507.9263
#> 8125  713452.5 9559132 2120.44 42.14047 -0.05458 -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 47718 712812.5 9559892 1867.55 54.93863 -0.01491  0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543
#> 7     714952.5 9559862 2284.19 16.10642  0.00516  0.01574    183.3220
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.64982  0.01299  0.00141   4611.0718
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818
#> 43245 713722.5 9560012 2153.38 34.19126 -0.01263  0.01322   4605.1411
#> 26101 715562.5 9558682 2594.87 32.41108  0.00079  0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.50981 -0.00428 -0.00522   1355.7436
#> 17058 714032.5 9558502 2402.95 31.35282  0.01924  0.02606    564.6876
#> 39325 714142.5 9558592 2357.72 40.79288 -0.00246  0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.68589 -0.00124 -0.00456   1531.4274
#> 10617 713232.5 9560722 1879.95 26.49701  0.00258  0.00422   1159.2100
#> 39624 715222.5 9559552 2205.54 23.89750  0.00401 -0.00391   1794.7964
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249
#> 1997  714752.5 9561012 1853.19 26.56633 -0.00693  0.01133   2397.2681
#> 48574 712692.5 9560412 1901.97 41.97833  0.01364 -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.80875 -0.01881  0.00121   1783.0658
#> 9606  714392.5 9559172 2353.05 30.78330 -0.00185 -0.01185    953.8614
#> 24723 712942.5 9560202 1983.18 33.07170  0.00285 -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.79995 -0.01397  0.00526    706.1463
#> 38103 715392.5 9560162 1998.16 46.28582  0.00708 -0.02538    807.8237
#> 17825 712732.5 9560462 1915.51 40.39754  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.89177  0.00445  0.00875    597.2893
#> 10213 715542.5 9558782 2532.10 35.41338 -0.00893 -0.00867   2247.8748
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438
#> 41310 714452.5 9560402 2092.27 12.46298 -0.00195 -0.00205   1112.2626
#> 36765 715302.5 9560262 1952.73 27.20346 -0.00585 -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002
#> 31253 712612.5 9559492 1903.67 46.75679 -0.06953 -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.02913 -0.01084 -0.02996   1599.3491
#> 7317  713682.5 9558582 2347.46 28.34766 -0.02196  0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.91793 -0.00849  0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.68859  0.00534  0.01476   1275.9446
#> 748   715102.5 9559692 2301.39 27.02184 -0.00759  0.00158    837.3730
#> 34250 713822.5 9559142 2316.36 33.75982 -0.02284  0.01085   2229.5105
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847
#> 19493 714252.5 9560372 2156.93 24.87955  0.00825  0.00445    355.5078
#> 16542 713772.5 9559622 2304.85 30.81023 -0.00712  0.00923    811.5580
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.15575  0.00176  0.02973    215.5101
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.10663 -0.00007 -0.00853   1354.6001
#> 31884 715782.5 9558612 2632.60 30.27967  0.00084 -0.00083   1309.2347
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451
#> 33175 715822.5 9558832 2563.13 36.62289 -0.00674 -0.00376    629.8236
#> 36245 713142.5 9560572 1827.82 43.67600 -0.00410  0.00400   4126.7812
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762
#> 22545 713402.5 9559562 2195.44 25.31557  0.04152  0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.64141 -0.00307  0.02997    230.8641
#> 17746 714752.5 9558322 2648.52 47.73655 -0.00068 -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.65595  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.03730  0.00586  0.00654   1576.2065
#> 20764 714542.5 9558152 2602.18 52.07900  0.02118  0.00162    290.2031
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.01922  0.03310 -0.07650    766.6254
#> 36889 714302.5 9559802 2243.60 21.16621  0.01923  0.05547    128.2613
#> 40909 712722.5 9560272 1827.15 29.29304  0.00009 -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.23024 -0.00233  0.01282    561.2366
#> 1281  715102.5 9560032 2120.64 34.64905 -0.00930 -0.00360   4288.7329
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193
#> 26283 715932.5 9557592 3075.48 32.76688 -0.00375  0.00215    917.5296
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074
#> 37783 713832.5 9557892 2405.28 40.88684  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417
#> 20950 714912.5 9557782 2667.71 30.81138 -0.04770  0.00270  35950.5508
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042
#> 40653 714262.5 9561262 1791.94 12.22234  0.00161  0.00709   1181.2555
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975
#> 46148 715082.5 9558042 2725.39 42.83604 -0.00181 -0.00799   1036.6606
#> 6577  714142.5 9559902 2244.30 57.89051 -0.06986 -0.05013    616.4128
#> 14645 715072.5 9560262 2034.04 38.16930 -0.00716 -0.03234   1707.3130
#> 6206  714122.5 9560992 1917.28 37.96991 -0.01023 -0.00227    841.2025
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 5749  28.23306831   300.00   TRUE         0.00            100  3.105292
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 2976  26.88203383   300.00   TRUE       300.00            100  2.684323
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 20764 39.85036057   300.00  FALSE       300.00             91  2.462702
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 15008 26.90781693   300.00  FALSE       300.00             48  2.625829
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 46148 38.91070978   300.00  FALSE       300.00            100  3.015637
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250