Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065  0.00005     769.3426
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770     474.5059
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665   25318.3555
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528  0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 22566 712832.5 9559842 1897.49 33.060238  0.00672  0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 15390 713132.5 9560622 1873.26 38.480418  0.00252  0.01118    2359.5059
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 12868 712532.5 9560332 1913.92 61.937311 -0.03543 -0.05146    1347.6224
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566 -0.03355   14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 37348 715632.5 9558182 2743.16 13.620353 -0.00904 -0.00866   29195.6016
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735 -0.00925     665.6116
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 44650 715612.5 9558142 2764.88 39.360482 -0.00550 -0.00941    1290.7709
#> 10364 714762.5 9561002 1856.21 25.274887 -0.00845  0.00454    3476.0818
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 38197 712522.5 9560322 1942.04 62.777267 -0.01611  0.02961    1086.6963
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013    4229.5708
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478 -0.07132   42970.7891
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511 -0.01801    1071.6141
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 1997  714752.5 9561012 1853.19 26.566334 -0.00693  0.01133    2397.2681
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915 -0.02705     413.5702
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152    1530.5657
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 17322 712612.5 9560402 1924.48 43.807780 -0.00486  0.02306    1718.3652
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242  0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 25140 713332.5 9560692 1809.95 52.257189  0.00075 -0.01945     259.7421
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001  0.00759    1880.4403
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380 -0.00035  0.00955     713.4463
#> 7553  714642.5 9559132 2391.59 63.967873 -0.00066  0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248    1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293  0.00186    2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 19963 24.4377959   300.00   TRUE       300.00             89  2.935182
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 24129 24.1289716   300.00   TRUE       300.00             89  2.886120
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 1393  23.4122014   300.00   TRUE       300.00              1  2.710095
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 18005 24.9059024   300.00   TRUE       300.00              0  2.744234
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 47022 43.5316144   300.00   TRUE       300.00             26  2.745084
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 15390 29.3222611    60.00   TRUE       118.92              2  3.372821
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 12868 37.5883232   119.34   TRUE        79.16             61  3.129568
#> 47351 34.8788058   300.00   TRUE       300.00            100  4.173758
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 41037 31.5705475   300.00   TRUE       300.00            100  3.329742
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 38197 34.3425173   105.52   TRUE        65.32             50  3.036108
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 17431 25.3253075   300.00   TRUE         1.11              5  4.633173
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 28312 41.6425726   300.00   TRUE       300.00            100  3.198702
#> 10689 30.6480854   300.00   TRUE       300.00             61  2.237416
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 36292 31.5155435   300.00   TRUE        16.03            100  3.030038
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 1997  25.2359261   168.72   TRUE         0.00              5  3.379717
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17322 31.6650855    75.77   TRUE       138.35              1  3.235115
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25584 27.1467403   300.00  FALSE       300.00            100  3.105832
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 44216 26.0649960   300.00  FALSE       300.00            100  3.167741
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 22628 29.1211529   300.00  FALSE       300.00            100  2.951467
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603