Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 21788 715392.5 9560172 1988.71 40.476603 -0.01919 -0.04051  351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240 -0.00250    1413.3312
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 44952 715352.5 9557472 2873.35 38.952536 -0.01073  0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407  0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550 -0.00120     512.9740
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311     586.7354
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412  0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832  0.00838     238.6685
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077 -0.00284    2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651 -0.03629  360379.1875
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987  0.02107    1392.5880
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.054753 -0.03184  0.03924    8023.8657
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 40738 714972.5 9557642 2745.48 42.542689 -0.00294 -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.575508 -0.00255 -0.00365    1536.2647
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517    1000.7119
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326 -0.02917 -0.01993   10308.6406
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 23213 715592.5 9558562 2654.69 33.219520 -0.00034  0.00294     739.5768
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481 -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.960131 -0.00320  0.00180    3904.5254
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524   10861.9482
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491  0.01061    5744.7583
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132    1580.1621
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186  0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122  0.00840  0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534  0.01476    1275.9446
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433 -0.02177     997.3089
#> 15186 713322.5 9561212 1936.88 36.079725  0.00165  0.00565    1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065  0.00475     515.1661
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 21925 714542.5 9559232 2314.33 50.247826  0.00737  0.01033    3446.5520
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674 -0.00376     629.8236
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 34105 715432.5 9558592 2621.73 18.956309 -0.00301 -0.01099    1146.2140
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196     576.1450
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060     383.3547
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 26768 715092.5 9560582 1960.72 36.656121  0.00980 -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770  0.00270   35950.5508
#> 23808 715172.5 9558702 2533.50 22.345927  0.00811 -0.02281     492.4785
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 32618 715132.5 9559642 2266.36 34.532739 -0.00305 -0.01135    1293.4855
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 21788 32.814439   300.00   TRUE       300.00             40  5.545499
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37104  4.866131    50.99   TRUE         0.00             21  3.252198
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 8869  29.110267   300.00   TRUE       300.00            100  3.036974
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 11463 20.133737   300.00   TRUE       300.00            100  2.768442
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 2107  35.423689   300.00   TRUE       300.00             45  2.661523
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 28577 34.391219   300.00   TRUE         0.56              0  3.781222
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 11687 23.193905   300.00   TRUE       300.00             68  3.186466
#> 24100 34.297827   300.00   TRUE       300.00            100  5.558684
#> 27084 21.672702   300.00   TRUE       300.00             81  3.000309
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 34192 33.874984   273.94   TRUE         4.48             85  3.009256
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 39166 29.818443   300.00   TRUE         1.90             90  2.994673
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 18958 29.989757   300.00   TRUE       300.00              5  2.674746
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 1424  39.154217   300.00   TRUE       300.00            100  4.035908
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 47718 32.323414   300.00   TRUE        10.00              1  3.759272
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 14545 39.314072   300.00   TRUE       300.00             37  2.897494
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 26058 34.498362   300.00   TRUE        76.02             39  3.272521
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 13278 30.225243    91.73   TRUE       120.78             34  3.560110
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 47722 33.115815   300.00   TRUE        65.02              1  2.976445
#> 14140 35.487287   300.00   TRUE       300.00             65  2.767823
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 23832 25.485163   300.00   TRUE       300.00              8  2.676567
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 36561 25.804873   300.00  FALSE        70.00            100  2.990220
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 44396 21.442373   300.00  FALSE       300.00            100  2.270336
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 12974 37.338514   300.00  FALSE       300.00            100  2.711947
#> 43403 31.061761   300.00  FALSE       111.88             86  2.788585
#> 21925 31.248545   300.00  FALSE       300.00             95  3.537385
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33175 34.783695   300.00  FALSE       300.00            100  2.799219
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 41855 28.701748   234.40  FALSE        47.05            100  3.040739
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 26768 32.282161   300.00  FALSE       207.22            100  2.673964
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 12300 38.417393   124.35  FALSE         0.00            100  3.192630
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250