Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 31357 714752.5 9561022 1848.22 33.44641 -0.00347  0.02357   1752.0375 23.817283
#> 29391 712842.5 9560152 1930.16 28.09556  0.00753  0.01487   1077.7585 28.283489
#> 24051 715462.5 9559372 2319.46 32.32800 -0.00661 -0.01339   1002.5877 32.738808
#> 11088 715532.5 9558762 2549.28 38.05013 -0.00373 -0.00057   1630.3259 30.998736
#> 24129 714602.5 9560542 2038.30 25.58486  0.00065  0.00005    769.3426 24.128972
#> 27973 715022.5 9559332 2308.28 20.06326  0.00940  0.00679    550.0344 30.009810
#> 1768  713912.5 9558552 2357.19 38.69241 -0.00645  0.00835   2425.3115 28.185513
#> 49372 714862.5 9560982 1863.12 20.96911 -0.00295  0.00035   2697.5974 23.540544
#> 20151 714602.5 9559922 2184.66 48.34160 -0.00979 -0.01542    674.0147 44.795559
#> 24018 714292.5 9559282 2334.12 32.86543  0.00102  0.00388   1051.4222 29.305072
#> 27375 714292.5 9558982 2458.96 43.60094  0.00214  0.00466    724.0713 38.770335
#> 33792 714652.5 9558172 2598.36 59.45870 -0.03525 -0.11024    394.8751 51.208103
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930 23.454600
#> 43964 714802.5 9558452 2684.33 37.98252 -0.02123  0.01164    895.6606 28.757325
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586  8.761098
#> 2107  714142.5 9558562 2382.67 35.55375  0.01412  0.00898    458.6938 35.423689
#> 35285 714602.5 9560112 2142.75 51.10784 -0.00017 -0.00032    994.6740 42.466486
#> 29472 714792.5 9561072 1786.08 36.17312 -0.01029 -0.02312   2657.6936 18.648630
#> 38032 714762.5 9560442 1988.33 35.16987 -0.00905 -0.01385    674.0721 35.008294
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801 28.398653
#> 8288  714842.5 9557712 2679.04 28.93953  0.00902  0.01358    802.7622 22.717777
#> 27708 715222.5 9557782 2844.65 29.54285  0.03206  0.02773    179.3077 25.710908
#> 48364 715582.5 9558222 2746.23 25.07206  0.01257  0.00593    939.0466 32.075896
#> 34565 712972.5 9560172 2007.72 27.94716  0.00598  0.01212    561.1948 20.380109
#> 8073  712572.5 9560302 1917.62 52.00910  0.01937  0.00083    994.7876 36.075715
#> 27957 715012.5 9559332 2306.35 20.48038  0.00225  0.00325    616.8652 30.060231
#> 23031 714042.5 9558492 2404.94 28.49606  0.01205  0.02205    629.4165 26.741086
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036 28.580854
#> 17004 713102.5 9559032 2186.79 36.53351 -0.01383  0.00193   1809.3838 27.189139
#> 1635  713102.5 9559052 2172.87 31.48403 -0.00064 -0.00916   2009.1686 27.826841
#> 19730 715332.5 9558452 2640.35 47.64259 -0.00211  0.00642    613.6390 38.478700
#> 2950  715042.5 9557712 2740.72 26.52508 -0.00089 -0.00051    779.0555 28.203847
#> 4329  712882.5 9559982 1911.46 26.23173 -0.00193 -0.00077   3559.4204 34.561960
#> 35142 714792.5 9558862 2575.17 25.66450  0.06765  0.00114    150.1498 27.153043
#> 12084 712832.5 9559992 1872.46 56.05475 -0.03184  0.03924   8023.8657 34.595765
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927 36.516892
#> 19128 714852.5 9557872 2681.49 20.75769  0.00453  0.01077    300.7915 14.659698
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412 34.148858
#> 6557  713312.5 9559022 2181.69 27.63433 -0.00669 -0.00991   1403.0193 27.419468
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929 44.534291
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690 18.763795
#> 37348 715632.5 9558182 2743.16 13.62035 -0.00904 -0.00866  29195.6016 39.287716
#> 36931 712852.5 9558892 2102.80 30.53579  0.00140 -0.00070   2562.7559 27.209193
#> 6310  713802.5 9559752 2238.38 19.65131 -0.02379  0.01338  10787.6299 23.826450
#> 18564 715552.5 9558592 2648.36 26.29131  0.00316  0.00714    475.3120 21.700204
#> 4605  714852.5 9557892 2677.78 26.93360  0.00321  0.01169    350.1799 17.635641
#> 43092 715532.5 9558192 2770.31 36.83316 -0.00442 -0.00178    800.1933 39.478511
#> 18751 714882.5 9559182 2381.81 39.29918  0.00066 -0.00036    516.4709 38.075910
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312 27.068818
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888 46.601522
#> 33240 714972.5 9557652 2741.04 39.43439 -0.00286 -0.00964    571.1364 29.354920
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306 38.832788
#> 20220 714052.5 9558522 2378.67 39.65613 -0.00855 -0.02175   1122.9120 32.175591
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.650391
#> 26698 713382.5 9559082 2155.56 34.26460 -0.00211  0.00241    585.6182 34.249698
#> 18265 715522.5 9558782 2539.23 35.85627 -0.00458 -0.00401   1722.6481 31.445643
#> 33350 713132.5 9560672 1896.89 24.69505  0.00306 -0.00436   2602.7390 28.952703
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993 36.320368
#> 35083 714612.5 9558152 2596.12 60.59831  0.04548  0.05543    352.1558 43.540782
#> 44825 714322.5 9557882 2534.80 40.68860  0.00114  0.03026    247.0894 29.021458
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000 35.453482
#> 21013 712742.5 9560032 1845.53 29.35263  0.00568 -0.01558   1504.9717 24.963198
#> 17431 712802.5 9560022 1832.71 41.52512 -0.00478 -0.07132  42970.7891 25.325308
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760 27.597658
#> 47366 714892.5 9557612 2718.78 28.02394  0.00538  0.00112    745.3315 28.727531
#> 18897 712802.5 9559882 1851.71 50.23465 -0.02627 -0.05393   8301.6094 31.788271
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966 26.541697
#> 1424  714012.5 9558902 2305.97 32.22372 -0.03986 -0.02524  10861.9482 39.154217
#> 10073 715572.5 9558612 2632.78 32.16127  0.00129  0.00171    788.0004 25.761328
#> 20368 715252.5 9560622 1865.83 33.57246  0.00025 -0.01245   6353.3179 30.344418
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027 39.558725
#> 4588  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247 42.935738
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743 27.569583
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818 34.189538
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694 35.475255
#> 43074 713172.5 9559622 2052.40 31.52758 -0.00193  0.02812   1423.7578 33.594807
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680 39.490543
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207 28.038836
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086 46.649651
#> 14192 715162.5 9558122 2772.18 44.46267 -0.02970  0.00200   1305.6691 40.586611
#> 23626 713742.5 9560032 2137.92 39.30490 -0.00605  0.00075   2170.2249 26.778328
#> 20171 715732.5 9557962 2815.53 48.65787 -0.01341  0.00171   4425.5410 42.130160
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796 30.320354
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 7771  715232.5 9560922 1816.88 37.18038  0.00124  0.01166   1813.7841 16.579107
#> 17227 714812.5 9558892 2545.43 54.07175  0.00458  0.04291    293.4231 43.597505
#> 11057 714962.5 9557642 2737.48 42.68593 -0.00488 -0.00632    620.0229 31.225054
#> 35458 715242.5 9558442 2623.01 38.95483 -0.00770 -0.03400   1982.7782 37.473159
#> 6212  714592.5 9560362 2034.54 12.52257 -0.01454  0.00144 317260.6875 31.931511
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924 20.820713
#> 30378 714852.5 9557752 2651.70 41.29822 -0.00010 -0.00870   2099.7930 28.569395
#> 49346 715322.5 9558782 2553.74 25.39521  0.00319  0.00291    487.3438 23.345738
#> 17851 715362.5 9557482 2869.26 29.12574 -0.03251 -0.00389  14516.2002 32.263253
#> 17374 715712.5 9557952 2832.22 37.87480 -0.01519  0.00609   4275.3457 43.145441
#> 31983 714302.5 9559912 2204.77 23.46491 -0.01332  0.01842    246.5161 19.775638
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024 30.834870
#> 16208 712882.5 9560332 1846.34 48.62521  0.00159 -0.01559    866.5775 38.373848
#> 46165 714392.5 9557912 2542.56 16.32987  0.00325 -0.00505    351.4078 14.818407
#> 27884 714072.5 9559202 2446.27 19.75673 -0.00009  0.00109    605.1908 15.497362
#> 40459 712542.5 9559192 1992.60 39.81713 -0.00162 -0.00517   1060.5494 36.186868
#> 18286 714792.5 9557882 2650.57 28.25369 -0.00124  0.00204    853.2902 24.350706
#> 11111 715142.5 9558482 2650.65 42.34330 -0.00533  0.01493   2212.0525 33.383959
#> 26023 715832.5 9557632 3097.56 41.71018  0.00133 -0.00452    578.5847 45.072871
#> 46930 713642.5 9560602 1945.03 34.50982  0.00165  0.00875   1165.3953 31.920625
#> 441   714312.5 9560972 1916.59 31.30298  0.00540  0.00650   1133.1501 25.113313
#> 44396 715512.5 9558102 2845.31 26.59842  0.01263  0.03598    186.3527 21.442373
#> 20547 713732.5 9560822 1851.67 36.96838 -0.01405  0.00476  11992.5273 34.261157
#> 39685 715332.5 9558102 2797.04 47.78812  0.00167  0.01133    828.8604 47.873489
#> 36833 713322.5 9560572 1794.25 33.75409  0.00527 -0.02077   8759.9961 28.075505
#> 45071 712652.5 9560032 1900.80 14.47349  0.00987 -0.00507    504.3085 27.805069
#> 31497 714032.5 9560672 2060.77 16.21700  0.00973 -0.00943    371.5514 21.270485
#> 8413  713242.5 9561232 1967.45 38.19451  0.00236  0.00483   1097.3082 39.889895
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451 28.893116
#> 42934 713832.5 9558672 2270.74 27.61198 -0.03193 -0.00267  31588.1953 32.227157
#> 29527 714332.5 9557502 2567.02 48.90481 -0.00452 -0.01298    436.1459 36.726595
#> 1691  714222.5 9559022 2448.69 38.21686 -0.01292  0.00462    716.5762 27.000063
#> 34810 714112.5 9559572 2337.51 10.28803 -0.00393  0.00032    994.9224 15.554658
#> 13908 715022.5 9559372 2290.81 27.24300 -0.03004  0.00714 480535.5000 34.377468
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752 34.677125
#> 36721 714092.5 9559752 2301.81 28.93322  0.03418  0.00513    161.9878 26.668321
#> 49831 715262.5 9557332 2884.30 22.29493 -0.00097  0.01487   1769.3392 29.837923
#> 44715 714932.5 9560132 2195.78 30.39369  0.03479  0.02251    168.1733 28.450219
#> 35826 712912.5 9560302 1891.22 50.86948 -0.00229  0.00099    909.0560 44.227758
#> 41612 713952.5 9560002 2198.00 28.75618  0.00119  0.00361   1011.0248 25.587149
#> 1479  715622.5 9557952 2888.67 34.95386  0.01635 -0.00105    868.7802 43.676573
#> 29716 715272.5 9559992 2144.94 25.71778  0.00262 -0.00502    520.5218 32.693545
#> 40612 713542.5 9561152 1856.15 33.56845  0.00420  0.01960    884.2003 21.927095
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111 22.860443
#> 484   713812.5 9560592 2016.41 44.30110 -0.00719  0.00079   1647.6451 32.145797
#> 12300 712372.5 9560062 1903.64 50.03698  0.00343  0.01177   1558.2231 38.417393
#> 15138 713702.5 9561122 1780.68 49.22166  0.00607 -0.04566  12862.4922 20.645961
#> 19767 714272.5 9558052 2431.67 55.72473  0.00361 -0.01841    756.0055 42.010411
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562 29.166417
#> 16845 712772.5 9560022 1831.76 21.39997 -0.00865 -0.00805   7489.7417 22.882788
#> 34334 714642.5 9557912 2582.25 30.99129  0.00253 -0.01233    968.7780 25.089822
#> 36016 715262.5 9557682 2814.30 37.00219  0.00203  0.00238    787.0053 39.523775
#> 16443 714472.5 9559532 2222.55 16.74412  0.03263 -0.02012    305.3624 26.538259
#> 28526 715152.5 9560732 1871.53 29.53426 -0.00163 -0.00377   1727.6055 31.814627
#> 35257 713822.5 9557982 2350.80 29.30794  0.01549 -0.00169    649.9975 30.881852
#> 49812 713392.5 9559222 2133.99 35.66261 -0.00983  0.00803   1058.5630 33.358176
#> 32342 714862.5 9559622 2294.74 36.96724  0.01021  0.00109    915.4989 37.619836
#> 30403 713952.5 9560752 2003.57 42.96439 -0.00933 -0.00157    766.0220 36.130719
#> 42990 712432.5 9560692 2102.11 32.45004 -0.00087 -0.00573   1490.0312 22.237065
#> 11025 714402.5 9559252 2325.64 49.62044 -0.00241  0.00020    673.2861 36.819987
#> 41373 715332.5 9560862 1873.10 29.12459 -0.00166  0.00156    758.4840 31.128797
#> 39234 712902.5 9559652 1953.39 29.73193  0.07379 -0.00249    148.1857 33.485945
#> 30296 713322.5 9559782 2042.96 52.62102 -0.02110 -0.02931   1812.3820 40.158612
#> 24516 713802.5 9560862 1873.21 23.34975 -0.00771  0.00661   5642.7544 36.254477
#>       distroad slides distdeforest distslidespast log.carea
#> 31357   158.92   TRUE         0.00              5  3.243543
#> 29391   300.00   TRUE         0.56            100  3.032521
#> 24051   300.00   TRUE       300.00             11  3.001122
#> 11088   300.00   TRUE       300.00            100  3.212274
#> 24129   300.00   TRUE       300.00             89  2.886120
#> 27973   300.00   TRUE       300.00             85  2.740390
#> 1768    300.00   TRUE       300.00              7  3.384768
#> 49372   213.54   TRUE        20.21             68  3.430977
#> 20151   300.00   TRUE       300.00            100  2.828669
#> 24018   300.00   TRUE       300.00              5  3.021777
#> 27375   300.00   TRUE       300.00            100  2.859781
#> 33792   300.00   TRUE       300.00              5  2.596460
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 43964   300.00   TRUE       300.00             46  2.952143
#> 38410    69.52   TRUE        47.61             65  3.695093
#> 2107    300.00   TRUE       300.00             45  2.661523
#> 35285   300.00   TRUE       300.00            100  2.997681
#> 29472   111.09   TRUE         0.00             25  3.424505
#> 38032   300.00   TRUE       300.00             60  2.828706
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 14522   300.00   TRUE       300.00            100  3.047228
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 27708   300.00   TRUE       300.00            100  2.253599
#> 48364   300.00   TRUE       300.00            100  2.972687
#> 34565   300.00   TRUE        68.13            100  2.749114
#> 8073    138.71   TRUE        76.41             35  2.997730
#> 27957   300.00   TRUE       300.00             75  2.790190
#> 23031   300.00   TRUE       300.00              2  2.798938
#> 7879    300.00   TRUE       300.00            100  2.912011
#> 17004   300.00   TRUE       300.00            100  3.257531
#> 1635    300.00   TRUE       300.00            100  3.303016
#> 19730   300.00   TRUE       300.00            100  2.787913
#> 2950    300.00   TRUE       300.00            100  2.891568
#> 4329    300.00   TRUE        16.35              2  3.551379
#> 35142   300.00   TRUE       300.00            100  2.176525
#> 12084   300.00   TRUE         0.00              0  3.904384
#> 22735   300.00   TRUE       300.00             90  2.622310
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 15990   300.00   TRUE         0.00              0  3.754443
#> 6557    300.00   TRUE       300.00              5  3.147064
#> 47011   300.00   TRUE       300.00             10  3.652797
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 37348   300.00   TRUE       300.00            100  4.465317
#> 36931   300.00   TRUE       300.00            100  3.408707
#> 6310    300.00   TRUE       300.00            100  4.032926
#> 18564   300.00   TRUE       300.00            100  2.676979
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 43092   300.00   TRUE       300.00            100  2.903195
#> 18751   300.00   TRUE       300.00            100  2.713046
#> 22332   300.00   TRUE       300.00            100  5.168705
#> 23029   300.00   TRUE       300.00             34  2.425842
#> 33240   300.00   TRUE       300.00            100  2.756740
#> 16015   152.09   TRUE        90.23             45  2.971800
#> 20220   300.00   TRUE       300.00              8  3.050346
#> 835     300.00   TRUE       300.00            100  3.141386
#> 26698   300.00   TRUE       300.00              7  2.767615
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 33350    10.00   TRUE       166.26              2  3.415431
#> 7668    300.00   TRUE       300.00              5  2.789721
#> 35083   300.00   TRUE       300.00             23  2.546735
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 21013   300.00   TRUE        13.47             64  3.177528
#> 17431   300.00   TRUE         1.11              5  4.633173
#> 40270   300.00   TRUE       300.00              5  3.139431
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 18897   300.00   TRUE        20.00              5  3.919162
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 1424    300.00   TRUE       300.00            100  4.035908
#> 10073   300.00   TRUE       300.00            100  2.896526
#> 20368   300.00   TRUE       166.26            100  3.803001
#> 42201   300.00   TRUE       300.00            100  3.989241
#> 4588    300.00   TRUE       300.00              1  3.610558
#> 34887   300.00   TRUE       300.00            100  3.159244
#> 33875   300.00   TRUE         4.67              0  3.279136
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 26428   300.00   TRUE         0.00            100  3.486483
#> 43074   300.00   TRUE       205.02              1  3.153436
#> 7485    300.00   TRUE       300.00            100  2.627640
#> 47244   300.00   TRUE       300.00              2  2.752448
#> 22297   300.00   TRUE       300.00             11  2.548891
#> 14192   300.00   TRUE       300.00            100  3.115833
#> 23626   300.00   TRUE       300.00            100  3.336505
#> 20171   300.00   TRUE       300.00            100  3.645966
#> 16065   300.00   TRUE       162.92            100  3.818858
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 7771    300.00   TRUE        35.59            100  3.258586
#> 17227   300.00   TRUE       300.00            100  2.467494
#> 11057   300.00   TRUE       300.00            100  2.792408
#> 35458   300.00   TRUE       300.00            100  3.297274
#> 6212    300.00   TRUE       300.00             47  5.501416
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 30378   300.00   TRUE       300.00             57  3.322176
#> 49346   300.00   TRUE       300.00             29  2.687835
#> 17851   300.00   TRUE       300.00            100  4.161853
#> 17374   300.00   TRUE       300.00            100  3.630971
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 16208   256.60  FALSE         0.00             67  2.937807
#> 46165   300.00  FALSE       300.00            100  2.545811
#> 27884   300.00  FALSE       300.00            100  2.781892
#> 40459   300.00  FALSE       216.66             10  3.025531
#> 18286   300.00  FALSE       300.00             42  2.931097
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 26023   300.00  FALSE       300.00            100  2.762367
#> 46930   300.00  FALSE         0.00            100  3.066473
#> 441     294.31  FALSE       152.09            100  3.054287
#> 44396   300.00  FALSE       300.00            100  2.270336
#> 20547   300.00  FALSE        42.56            100  4.078911
#> 39685   300.00  FALSE       300.00            100  2.918481
#> 36833   187.46  FALSE        18.49             45  3.942504
#> 45071   264.04  FALSE         0.00            100  2.702696
#> 31497   300.00  FALSE       154.49            100  2.570019
#> 8413     80.73  FALSE         0.00            100  3.040329
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 42934   300.00  FALSE       300.00             22  4.499525
#> 29527   300.00  FALSE       300.00            100  2.639632
#> 1691    300.00  FALSE       300.00            100  2.855262
#> 34810   300.00  FALSE       300.00            100  2.997789
#> 13908   300.00  FALSE       300.00             96  5.681725
#> 21070   300.00  FALSE         0.00             10  3.400551
#> 36721   300.00  FALSE       300.00            100  2.209482
#> 49831   300.00  FALSE       300.00            100  3.247811
#> 44715   300.00  FALSE       300.00            100  2.225757
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 41612   300.00  FALSE       300.00            100  3.004762
#> 1479    300.00  FALSE       300.00            100  2.938910
#> 29716   300.00  FALSE       300.00            100  2.716439
#> 40612    55.73  FALSE       125.03            100  2.946551
#> 32430   300.00  FALSE       300.00            100  2.398827
#> 484     300.00  FALSE        55.63            100  3.216864
#> 12300   124.35  FALSE         0.00            100  3.192630
#> 15138   116.69  FALSE        19.61             40  4.109325
#> 19767   300.00  FALSE       300.00            100  2.878525
#> 10981   300.00  FALSE       300.00            100  5.155983
#> 16845   300.00  FALSE        25.55             33  3.874467
#> 34334   300.00  FALSE       300.00            100  2.986224
#> 36016   300.00  FALSE       300.00            100  2.895978
#> 16443   300.00  FALSE       300.00             75  2.484816
#> 28526   300.00  FALSE        46.28            100  3.237445
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 49812   300.00  FALSE       300.00             65  3.024717
#> 32342   300.00  FALSE       300.00             81  2.961658
#> 30403   300.00  FALSE       209.91            100  2.884241
#> 42990   205.00  FALSE         0.00            100  3.173195
#> 11025   300.00  FALSE       300.00             50  2.828200
#> 41373   300.00  FALSE       120.90            100  2.879946
#> 39234   300.00  FALSE         0.00            100  2.170806
#> 30296   300.00  FALSE       300.00             68  3.258250
#> 24516   300.00  FALSE       122.13            100  3.751491