Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633   -0.01813    1130.5791
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097    0.01483     566.7674
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753    0.01487    1077.7585
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863   -0.00507  159065.0781
#> 24129 714602.5 9560542 2038.30 25.584857  0.00065    0.00005     769.3426
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090    0.00890    2301.0454
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940    0.00679     550.0344
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295    0.00035    2697.5974
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068    0.00588    1295.1610
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102    0.00388    1051.4222
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831   -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841   -0.00010    1527.1028
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021   -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194    0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 18843 715192.5 9559332 2265.75 41.153267 -0.07664   -0.05766  140591.4062
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392    1255.0884
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898     458.6938
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017   -0.00032     994.6740
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203   10263.2734
#> 6676  715092.5 9559912 2217.44 41.707890 -0.00077   -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490   -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206    0.02773     179.3077
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 3133  715042.5 9559252 2344.13 29.020885  0.00422    0.02357     275.7474
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205     629.4165
#> 46125 713732.5 9560062 2115.18 37.593480 -0.04616    0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804   -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.484031 -0.00064   -0.00916    2009.1686
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416    1204.8629
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765    0.00114     150.1498
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028    0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743    0.01313     783.4548
#> 47351 714222.5 9558792 2420.94 28.668516 -0.09566   -0.03355   14919.6260
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458    0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.881503 -0.00666   -0.00504    1020.2766
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552   -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.599789 -0.00735   -0.00925     665.6116
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974    2136.6939
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480    0.02530     937.1306
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372   -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478    0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194    0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405   -0.01565    2232.6133
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260   -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017    0.00197     711.1091
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160    0.01690    5121.6924
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062   -0.00262    1760.2162
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397  0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568   -0.02553    1750.4175
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406   -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656    3036.7383
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645   -0.00395    2011.3270
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784    0.00596     842.0994
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131    0.00201     788.0621
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472   -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588    0.01448    1258.8196
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601   -0.08389   47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 8125  713452.5 9559132 2120.44 42.140473 -0.05458   -0.14813  287033.4688
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278   -0.00412    1350.3611
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195   -0.00205    1112.2626
#> 10933 712702.5 9560132 1854.55 36.291147 -0.00204   -0.00376    2048.3198
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332   -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521   -0.00870    1895.1399
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818  0.01299    0.00141    4611.0718
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171   -0.00261    1627.8989
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666   -0.03583    1898.0105
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193    0.02812    1423.7578
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253    0.00753    1337.7854
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 16071 714562.5 9560062 2166.69 28.696018  0.02281    0.04759     158.9370
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495   -0.03756     353.9086
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705     413.5702
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185   -0.01185     953.8614
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125   -0.02685    2568.4631
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181   -0.00341    1363.4000
#> 7771  715232.5 9560922 1816.88 37.180377  0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152    1530.5657
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115    0.01015     790.2421
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 31491 712832.5 9558892 2091.39 29.700413  0.00186    0.00063    2430.2278
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025   -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614   -0.00093     118.1402
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159   -0.01559     866.5775
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124    0.00204     853.2902
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165    0.00875    1165.3953
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458    0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290    2481.5632
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.302976  0.00540    0.00650    1133.1501
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321    0.00548     532.8210
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973     215.5101
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973   -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196    0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990    0.02790     209.6952
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338    0.00518    1137.0179
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424    0.00196     576.1450
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097    0.01487    1769.3392
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479    0.02251     168.1733
#> 42424 714692.5 9560902 1921.34 25.331037  0.01154   -0.00204     618.2027
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119    0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165    0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635   -0.00105     868.7802
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236   -0.00066     306.0146
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578    1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354    0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886   43249.3945
#> 20950 714912.5 9557782 2667.71 30.811378 -0.04770    0.00270   35950.5508
#> 9320  713872.5 9559262 2378.76 29.858550  0.00309    0.01011     501.7867
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992   -0.01052     451.7209
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983    0.00803    1058.5630
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573    1490.0312
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288    0.00448     941.8503
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336    0.01403     609.8665
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110   -0.02931    1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 18843 28.52642270   300.00   TRUE       300.00            100  5.147959
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 28046 19.59114589   300.00   TRUE       300.00             15  2.896560
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 19832 36.63377551   300.00   TRUE       300.00            100  3.278299
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 47298 25.33619370   300.00   TRUE       300.00            100  3.126386
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603