Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816  -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669   0.00221    0.00969    368.5606
#> 28223 712742.5 9560482 1927.69 34.667957  -0.00292    0.00712   3636.5303
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475   1001.0861
#> 5999  715552.5 9558812 2518.03  7.984167  -0.01863   -0.00507 159065.0781
#> 8598  714002.5 9558872 2310.02 39.583935  -0.00498   -0.00752   1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949   6217.8213
#> 25169 712642.5 9560232 1883.10 46.845029   0.00884    0.01345   1120.6929
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.657434   0.00922    0.01088    266.4355
#> 1393  714832.5 9557902 2664.72 31.295528   0.00550   -0.00120    512.9740
#> 31023 714712.5 9561042 1838.66 44.120042  -0.00223    0.01783    377.8918
#> 22814 713862.5 9558582 2327.48 48.612031  -0.02894   -0.03416    947.4878
#> 42598 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 2990  713542.5 9560372 1970.07 31.887966   0.01903    0.01447    331.4833
#> 45869 714342.5 9560282 2127.11 27.623441  -0.02169   -0.00160   9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919   0.00528    0.01392   1255.0884
#> 48055 712862.5 9559972 1902.48 29.016875  -0.00720    0.01250   3686.1240
#> 18005 714832.5 9557912 2660.76 31.912030   0.00191   -0.00261    554.9243
#> 38279 714192.5 9558612 2325.14 28.252549  -0.04430   -0.06070 138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619  -0.01126   -0.02044 129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868  -0.00905   -0.01385    674.0721
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 36030 712852.5 9559572 1938.79 28.678829  -0.01440    0.00710    843.7961
#> 6808  715192.5 9559492 2219.03 21.817660   0.00713   -0.00173   8967.1924
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 13585 714552.5 9560092 2123.69 57.528973  -0.03613   -0.01898    588.3761
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 10320 714912.5 9558482 2724.21 25.282909   0.00356    0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096    520.5570
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 18751 714882.5 9559182 2381.81 39.299175   0.00066   -0.00036    516.4709
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949   1182.1506
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.725260  -0.01399   -0.00731   1384.7959
#> 18892 712682.5 9560202 1839.08 40.923511  -0.01364   -0.03357    473.4709
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427    380.9472
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 16370 713852.5 9558632 2296.97 34.624413  -0.01972    0.03512  13178.3223
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 8136  713242.5 9560602 1798.18 41.240356  -0.01169   -0.03220   8936.2666
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 10933 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746  -0.01878   -0.04351   1442.9247
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 15981 713302.5 9559172 2106.91 35.997792   0.01354    0.00245    629.5983
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461  -0.00070    0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.358204   0.00540    0.00879   1200.4210
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 39624 715222.5 9559552 2205.54 23.897497   0.00401   -0.00391   1794.7964
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751  -0.01881    0.00121   1783.0658
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 17227 714812.5 9558892 2545.43 54.071746   0.00458    0.04291    293.4231
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144 317260.6875
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 17322 712612.5 9560402 1924.48 43.807780  -0.00486    0.02306   1718.3652
#> 36744 714922.5 9557722 2692.39 25.980771   0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996   1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.957608   0.00378    0.02382    100.0000
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 15186 713322.5 9561212 1936.88 36.079725   0.00165    0.00565   1410.8372
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419    761.5427
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 45071 712652.5 9560032 1900.80 14.473487   0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.720118   0.00174   -0.03065    704.4843
#> 31497 714032.5 9560672 2060.77 16.216997   0.00973   -0.00943    371.5514
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 36245 713142.5 9560572 1827.82 43.676000  -0.00410    0.00400   4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 48869 712562.5 9560122 1972.89 24.946582   0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 5656  713132.5 9560142 2042.24 33.595380  -0.00035    0.00955    713.4463
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 28286 713702.5 9559842 2245.07 27.968935   0.00511    0.01499    617.3855
#> 38362 714402.5 9558042 2499.31 56.901012   0.03044   -0.00964    543.0025
#> 34336 713792.5 9559072 2288.95 36.932859  -0.00952   -0.00438   1276.5448
#> 40750 712962.5 9558702 2126.06 34.704054  -0.00264    0.00624   1870.8307
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523    888.2811
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.197748  -0.00595    0.03064   1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157    766.0220
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913  -0.01023   -0.00227    841.2025
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 27864 27.815382    30.00   TRUE       183.39             20  3.550683
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 28223 29.022604    60.46   TRUE       152.09             35  3.560687
#> 13748 34.119064   300.00   TRUE       300.00              2  3.256827
#> 40756 39.335272   300.00   TRUE       300.00            100  3.000471
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 1393  23.412201   300.00   TRUE       300.00              1  2.710095
#> 31023 16.677656   135.00   TRUE         0.00              2  2.577367
#> 22814 34.912037   300.00   TRUE       300.00              6  2.976574
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 48055 33.825709   300.00   TRUE         1.67              0  3.566570
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 835   36.650391   300.00   TRUE       300.00            100  3.141386
#> 18892 27.063088   235.02   TRUE        27.24              0  2.675293
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 8136  25.429013   115.18   TRUE        84.59             30  3.951156
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 10933 34.677125   264.46   TRUE        18.17             56  3.311398
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 28173 23.883746   300.00   TRUE       300.00             59  3.185096
#> 37439 26.592690   300.00   TRUE       300.00            100  3.539988
#> 332    0.556915    94.82   TRUE         0.00             90  3.079334
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 17227 43.597505   300.00   TRUE       300.00            100  2.467494
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 17322 31.665086    75.77   TRUE       138.35              1  3.235115
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 23176 41.558348   300.00   TRUE       300.00            100  3.203943
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 28253  2.957608   300.00  FALSE       300.00            100  2.000000
#> 10145 27.416030   276.16  FALSE        91.33            100  3.227418
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 15186 38.548600    60.00  FALSE         0.00            100  3.149477
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 2664  39.291154   300.00  FALSE       300.00            100  2.881694
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 45071 27.805069   264.04  FALSE         0.00            100  2.702696
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 29944 18.628004   300.00  FALSE       300.00             48  2.452873
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 22545 23.684356   300.00  FALSE       300.00             96  2.215056
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 48869 24.875536   138.98  FALSE         0.00             63  2.748851
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 5656  31.272609   300.00  FALSE        94.43            100  2.853361
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 38362 38.267851   300.00  FALSE       300.00            100  2.734802
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 40750 32.099960   300.00  FALSE       300.00            100  3.272034
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 14881 26.856251   300.00  FALSE       300.00            100  2.948550
#> 9320  18.139271   300.00  FALSE       300.00             93  2.700519
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 33881 29.185897   300.00  FALSE       300.00            100  3.171879
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 30403 36.130719   300.00  FALSE       209.91            100  2.884241
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607