Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958  0.02642     671.1807
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 24051 715462.5 9559372 2319.46 32.327998 -0.00661 -0.01339    1002.5877
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110  0.00453  0.00257     357.1436
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261     554.9243
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 14522 715012.5 9557732 2721.34 35.181327 -0.00382  0.00672    1114.8801
#> 36030 712852.5 9559572 1938.79 28.678829 -0.01440  0.00710     843.7961
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151  0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856  0.01114    4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205     629.4165
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 22735 714492.5 9559772 2170.30 39.688786 -0.00440 -0.02651     419.0927
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 19784 714982.5 9557722 2716.98 35.436294  0.00888  0.00862     874.5408
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509  0.00499     885.0064
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 47091 713382.5 9559182 2122.71 48.037927  0.04527  0.04613     289.6113
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174  0.00025    4425.5151
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585 -0.00974    2136.6939
#> 33240 714972.5 9557652 2741.04 39.434393 -0.00286 -0.00964     571.1364
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092     412.5849
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.392169  0.03504  0.05825     162.6688
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 33350 713132.5 9560672 1896.89 24.695054  0.00306 -0.00436    2602.7390
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556 -0.04394     587.8840
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 32879 713912.5 9559672 2238.00 31.976775 -0.00278 -0.00412    1350.3611
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 15981 713302.5 9559172 2106.91 35.997792  0.01354  0.00245     629.5983
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 7485  714362.5 9558042 2484.02 47.958859  0.03169  0.01441     424.2680
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 13259 714862.5 9560992 1859.28 21.267620 -0.00317  0.00077    3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 14192 715162.5 9558122 2772.18 44.462671 -0.02970  0.00200    1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 17674 713882.5 9558562 2339.20 41.036383 -0.03675 -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529  0.01230    1044.6443
#> 49346 715322.5 9558782 2553.74 25.395208  0.00319  0.00291     487.3438
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 39466 714942.5 9557772 2680.35 16.753286  0.00166 -0.01265    1148.6487
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196  0.00336    2806.2644
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.489013 -0.00226  0.00736     477.8591
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.119476  0.02295  0.02715     520.7355
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 1488  714562.5 9560002 2150.93 11.283831  0.00197  0.00463     851.6306
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 13908 715022.5 9559372 2290.81 27.242997 -0.03004  0.00714  480535.5000
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099     909.0560
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952 -0.00438    1276.5448
#> 31098 714352.5 9559042 2410.15 30.336396  0.00537 -0.00347    1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361 -0.01841     756.0055
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 43972 713642.5 9558702 2304.16 24.023547  0.00340  0.01809    1158.0426
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163 -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.882446  0.01014 -0.00884     672.2466
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166  0.00156     758.4840
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.427888   300.00   TRUE        15.00              9  3.746431
#> 23178 41.600174   300.00   TRUE       300.00             21  2.826839
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 30059 20.192179   300.00   TRUE       300.00             41  2.412719
#> 29391 28.283489   300.00   TRUE         0.56            100  3.032521
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 44328 25.504070   300.00   TRUE       300.00            100  3.122417
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 18005 24.905902   300.00   TRUE       300.00              0  2.744234
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 14522 28.398653   300.00   TRUE       300.00            100  3.047228
#> 36030 29.199648   300.00   TRUE         0.00            100  2.926238
#> 35942 27.164502   300.00   TRUE       300.00            100  3.364960
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 6808  28.565957   300.00   TRUE       300.00             40  3.952656
#> 10320 16.455921   300.00   TRUE       300.00             30  2.721899
#> 34280 33.621163   300.00   TRUE         0.00            100  3.365895
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 27957 30.060231   300.00   TRUE       300.00             75  2.790190
#> 23031 26.741086   300.00   TRUE       300.00              2  2.798938
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 17594 19.231901   138.35   TRUE         0.00             60  3.701999
#> 19784 27.955757   300.00   TRUE       300.00            100  2.941780
#> 38686 31.188958   300.00   TRUE       300.00             75  2.946946
#> 46235 30.108359   300.00   TRUE        41.37              0  2.752133
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 18751 38.075910   300.00   TRUE       300.00            100  2.713046
#> 46171 29.071306   300.00   TRUE       270.22            100  3.645964
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 651   45.524362   300.00   TRUE       300.00            100  3.072673
#> 41037 31.570547   300.00   TRUE       300.00            100  3.329742
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 20220 32.175591   300.00   TRUE       300.00              8  3.050346
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 15766 23.337144   300.00   TRUE       300.00            100  2.760448
#> 4617  27.951746   300.00   TRUE       300.00              6  2.211304
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 33350 28.952703    10.00   TRUE       166.26              2  3.415431
#> 6317  27.323211   300.00   TRUE       300.00            100  2.837633
#> 25700 32.541138   300.00   TRUE       300.00            100  3.096112
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 6845  44.596170   300.00   TRUE       300.00             29  2.769292
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10467 38.996080   300.00   TRUE       300.00             90  2.784369
#> 18440 34.944696   300.00   TRUE       300.00            100  4.477051
#> 6217  30.971807   300.00   TRUE       300.00              1  3.243763
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 34887 42.935738   300.00   TRUE       300.00            100  3.159244
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15981 30.003508   300.00   TRUE       300.00              0  2.799064
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 34164 39.713424   300.00   TRUE       300.00              1  2.561448
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 49378 33.994159    22.68   TRUE       134.63             25  3.073643
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 39624 33.294577   300.00   TRUE       300.00              2  3.254015
#> 750   34.347101   300.00   TRUE         0.00              0  3.804807
#> 34984 33.905923   300.00   TRUE       300.00            100  2.699406
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 46971 16.784226   300.00   TRUE       262.51            100  3.018968
#> 49346 23.345738   300.00   TRUE       300.00             29  2.687835
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 36765 36.185150   300.00   TRUE       300.00            100  3.279776
#> 39466 27.251592   300.00   TRUE       300.00            100  3.060187
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 30809 36.366204   181.84  FALSE         0.00            100  2.854150
#> 24836 25.583711   300.00  FALSE       300.00             95  2.679300
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 49363 32.957678   300.00  FALSE         0.00            100  3.394725
#> 9665  34.081249   300.00  FALSE       300.00             64  2.653921
#> 29125 23.945052   300.00  FALSE       300.00            100  2.626221
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 37029 28.893116   300.00  FALSE       237.21            100  2.916688
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 28286 17.048359   300.00  FALSE       300.00            100  2.790556
#> 44434 36.319795   300.00  FALSE       300.00            100  3.451290
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 24446 31.363137   300.00  FALSE       300.00            100  2.925561
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 43972 22.852422   300.00  FALSE       300.00            100  3.063725
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 32342 37.619836   300.00  FALSE       300.00             81  2.961658
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 15880 48.896791   300.00  FALSE       300.00            100  2.546026
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 30296 40.158612   300.00  FALSE       300.00             68  3.258250