Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976    301.2347
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410   0.00118    0.00761    928.6858
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 47255 713552.5 9559962 2191.77 33.578765  -0.00068    0.00588   1295.1610
#> 125   713992.5 9561062 1878.28 49.971087   0.00183    0.03387   1989.3012
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104   1641.4587
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 43964 714802.5 9558452 2684.33 37.982518  -0.02123    0.01164    895.6606
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272   4955.5586
#> 18304 714582.5 9558932 2463.25 22.194666  -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 15340 715452.5 9558852 2523.95 34.165473   0.00151    0.00899   1394.2274
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114   4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672   0.02059   -0.00628    556.0121
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284   2010.4285
#> 8288  714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 34565 712972.5 9560172 2007.72 27.947162   0.00598    0.01212    561.1948
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.803944  -0.01097    0.00156    816.6036
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259   0.01414    0.03416   1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 12460 712762.5 9559622 1906.41 24.261898   0.01047   -0.00887    366.8537
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 40738 714972.5 9557642 2745.48 42.542689  -0.00294   -0.00096    520.5570
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376   2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714    475.3120
#> 31316 715022.5 9559332 2308.28 20.063263   0.00940    0.00679    550.0344
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 28743 715202.5 9558872 2463.87 34.657071  -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117   1018.2141
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 48808 715032.5 9557792 2707.02 25.437607  -0.02329   -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.954422  -0.04708    0.00408  76318.0156
#> 39520 715342.5 9557622 2889.48 41.648875  -0.01282    0.00542    860.8840
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 7668  713372.5 9559062 2164.82 37.934963  -0.01807   -0.01883    616.1993
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.825318  -0.00104    0.01664   1898.5588
#> 28046 714882.5 9558492 2711.77 28.305834  -0.00131    0.00201    788.0621
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 37769 714412.5 9558562 2382.08 39.876717  -0.04601   -0.08389  47022.9688
#> 1973  715252.5 9557822 2861.00 15.570701   0.01865    0.00335    195.0293
#> 11612 715212.5 9557752 2820.51 39.162238   0.00017    0.01213    328.7922
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 19892 714482.5 9560562 2059.63 19.815172  -0.00332   -0.00058   1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 7     714952.5 9559862 2284.19 16.106417   0.00516    0.01574    183.3220
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 5572  715062.5 9558262 2754.60 39.911667   0.00707    0.01833    369.3771
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786  -0.01109    0.00298   1533.7925
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 10692 712712.5 9561022 2005.85 41.433443   0.01410   -0.00380    601.0843
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.324378  -0.00296    0.00426   1315.5334
#> 20426 714702.5 9559712 2208.61 43.046319  -0.08153   -0.02417  34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 37115 715612.5 9559202 2358.15  9.727677  -0.01020    0.00279  39495.1797
#> 4479  713982.5 9557812 2399.24 41.231189   0.01895    0.04744    226.7911
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566  -0.01454    0.00144 317260.6875
#> 11332 712742.5 9560072 1850.34 22.700015   0.00683   -0.00603   1562.0503
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 25212 714272.5 9559012 2433.45 37.671402  -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.343300  -0.00533    0.01493   2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 1488  714562.5 9560002 2150.93 11.283831   0.00197    0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 43243 715682.5 9558462 2693.23 27.406290  -0.00031   -0.00199    836.4196
#> 39796 714802.5 9560372 2046.48 40.155747   0.00176    0.02973    215.5101
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431    315.9084
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077   8759.9961
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 38689 714622.5 9559342 2264.90 42.648113  -0.00844    0.00514    897.9290
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 33956 714082.5 9558712 2327.95 39.684203   0.02218    0.00902    218.7467
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462    716.5762
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 29107 713172.5 9559272 2067.58 28.264008   0.02413    0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 330   714232.5 9560362 2165.00 18.060776   0.01142    0.00868    235.8511
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 43222 714472.5 9560712 2039.25 22.190655   0.01858    0.00312    254.0656
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653  -0.00070   -0.00060    383.3547
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3724  714322.5 9557402 2603.46 34.597484   0.02523    0.02247    282.1076
#> 40612 713542.5 9561152 1856.15 33.568451   0.00420    0.01960    884.2003
#> 9007  714952.5 9560902 1863.75 32.770321   0.00074   -0.02904    608.3561
#> 974   713402.5 9561062 1854.56 23.634509   0.00082    0.02578   1258.8865
#> 7986  712752.5 9560242 1857.41 39.830307   0.01852    0.00248   1432.4019
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347   1893.3645
#> 18308 714842.5 9558742 2568.20 26.017441  -0.02552    0.00532  16771.3066
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 20950 714912.5 9557782 2667.71 30.811378  -0.04770    0.00270  35950.5508
#> 16443 714472.5 9559532 2222.55 16.744119   0.03263   -0.02012    305.3624
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029    894.2672
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052    451.7209
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 24516 713802.5 9560862 1873.21 23.349749  -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 16435 49.443966   300.00   TRUE       300.00             41  2.478905
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 34512 34.639882   300.00   TRUE       195.00              2  3.224136
#> 28293 28.902283   300.00   TRUE       300.00              4  2.967869
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 24018 29.305072   300.00   TRUE       300.00              5  3.021777
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 39061 35.240915   300.00   TRUE        70.65             56  3.215230
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 33792 51.208103   300.00   TRUE       300.00              5  2.596460
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 18304 44.908432   300.00   TRUE       300.00             57  3.589919
#> 35285 42.466486   300.00   TRUE       300.00            100  2.997681
#> 30125 40.735580   300.00   TRUE       300.00            100  2.748832
#> 15340 27.975237   300.00   TRUE       300.00            100  3.144334
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 13847 20.570904   300.00   TRUE       300.00            100  3.062637
#> 7879  28.580854   300.00   TRUE       300.00            100  2.912011
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 22735 36.516892   300.00   TRUE       300.00             90  2.622310
#> 12460 36.256769   300.00   TRUE         0.00            100  2.564493
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 40738 28.960152   300.00   TRUE       300.00            100  2.716468
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 17036 24.928248    25.33   TRUE       208.61             15  3.608265
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 39599 33.134722   300.00   TRUE       300.00             25  3.007839
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 48808 33.773570   300.00   TRUE       300.00            100  4.066528
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 42941 39.063117   300.00   TRUE       300.00             17  4.882627
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 18265 31.445643   300.00   TRUE       300.00            100  3.236197
#> 7668  36.320368   300.00   TRUE       300.00              5  2.789721
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 31561 31.442778   300.00   TRUE       300.00              0  2.888200
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 28046 19.591146   300.00   TRUE       300.00             15  2.896560
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 2552  32.215698   300.00   TRUE       300.00            100  3.005436
#> 5239  24.665260   197.29   TRUE         4.67             57  3.591568
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 37769 37.425030   300.00   TRUE       300.00            100  4.672310
#> 1973  19.325866   300.00   TRUE       300.00            100  2.290100
#> 11612 35.167004   300.00   TRUE       300.00            100  2.516921
#> 44767 17.295304   300.00   TRUE       300.00             25  3.046207
#> 19892 27.026419   300.00   TRUE       300.00             92  3.259412
#> 17345 29.821307   300.00   TRUE       300.00              2  4.693266
#> 31117 26.645975   300.00   TRUE       291.23            100  3.277641
#> 7     13.220428   300.00   TRUE       300.00              0  2.263215
#> 16009 33.413180   300.00   TRUE       300.00            100  3.049449
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 10692 33.646374   300.00   TRUE        60.19             11  2.778935
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 4667  31.517262    24.78   TRUE       172.24             38  3.119102
#> 20426 33.408596   300.00   TRUE       300.00             76  4.542933
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 37115 32.187050   300.00   TRUE       300.00            100  4.596544
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 6212  31.931511   300.00   TRUE       300.00             47  5.501416
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 25584 27.146740   300.00  FALSE       300.00            100  3.105832
#> 25212 38.982902   300.00  FALSE       300.00            100  3.109374
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 1488  29.921575   300.00  FALSE       300.00            100  2.930251
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43243 27.986124   300.00  FALSE       300.00            100  2.922424
#> 39796 34.567690   300.00  FALSE       300.00             71  2.333468
#> 13286 25.838678   300.00  FALSE       300.00             76  2.499561
#> 36833 28.075505   187.46  FALSE        18.49             45  3.942504
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 40139 33.143317   300.00  FALSE         1.11             21  3.274260
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 1691  27.000063   300.00  FALSE       300.00            100  2.855262
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 14148 21.877820   300.00  FALSE       214.75            100  2.760532
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 43222 21.265329   300.00  FALSE       300.00            100  2.404946
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 34532 31.206719   300.00  FALSE       300.00             92  2.583601
#> 5861  27.612555   300.00  FALSE       300.00            100  4.032264
#> 17628 26.992042   300.00  FALSE       205.59            100  2.626915
#> 3724  25.726951   300.00  FALSE       300.00            100  2.450415
#> 40612 21.927095    55.73  FALSE       125.03            100  2.946551
#> 9007  30.964358   300.00  FALSE        14.98             35  2.784158
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 19767 42.010411   300.00  FALSE       300.00            100  2.878525
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 16443 26.538259   300.00  FALSE       300.00             75  2.484816
#> 22628 29.121153   300.00  FALSE       300.00            100  2.951467
#> 31440 24.035007   300.00  FALSE       300.00             96  2.654870
#> 6577  42.407471   300.00  FALSE       300.00             70  2.789872
#> 23808 37.533892   300.00  FALSE       300.00             72  2.692387
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491