Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 36052 714842.5 9558892 2483.27 68.789886  -0.04921   -0.12438    753.6304
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813   1130.5791
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 24051 715462.5 9559372 2319.46 32.327998  -0.00661   -0.01339   1002.5877
#> 19963 714612.5 9560552 2031.70 26.848229   0.00046    0.00324    861.3537
#> 21939 713642.5 9558712 2299.09 33.750652  -0.00503    0.02433   1475.3191
#> 44952 715352.5 9557472 2873.35 38.952536  -0.01073    0.00174   2310.4871
#> 42917 714202.5 9557412 2544.08 39.877863  -0.02104   -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073  -0.00891   -0.00949   6217.8213
#> 49372 714862.5 9560982 1863.12 20.969109  -0.00295    0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595  -0.00979   -0.01542    674.0147
#> 39629 714602.5 9558912 2474.89 40.258307  -0.06175   -0.08226   2325.2053
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526    539.8119
#> 21703 713842.5 9559672 2282.58 24.991846   0.01581    0.01189    287.5425
#> 8019  714672.5 9559272 2316.84 40.381492  -0.01966   -0.02234    849.0547
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 40307 713302.5 9560702 1845.07 39.004675   0.00129    0.02510   3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 23375 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 31576 713942.5 9558492 2387.12 46.143156  -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 2942  713232.5 9560652 1851.75 34.535031   0.00856    0.01114   4510.8833
#> 6676  715092.5 9559912 2217.44 41.707890  -0.00077   -0.00284   2010.4285
#> 32821 715022.5 9559952 2208.33 43.416450  -0.00643    0.00314   2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751   0.00108   -0.00678   1891.6584
#> 38077 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733    598.7983
#> 5749  712852.5 9560162 1934.78 23.388710   0.00071    0.00150   1274.3594
#> 28230 715182.5 9558102 2791.83 39.071711  -0.00669   -0.00950    950.3138
#> 6763  714692.5 9561052 1830.60 48.394880  -0.01996    0.00626   1014.7159
#> 3466  713232.5 9560612 1812.72 51.566202  -0.01804   -0.01596   5296.9014
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.054753  -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.160862  -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.005073   0.03696    0.02514    297.7842
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167   1938.5389
#> 37348 715632.5 9558182 2743.16 13.620353  -0.00904   -0.00866  29195.6016
#> 47884 715032.5 9557712 2738.31 28.289218   0.00006    0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520  -0.00034    0.00294    739.5768
#> 18394 715022.5 9559942 2216.16 42.137608  -0.00613    0.00903   2360.1536
#> 3944  715262.5 9558352 2654.20 33.555273  -0.04492    0.00072  54831.6289
#> 34707 713962.5 9561092 1819.49 65.686173  -0.04749    0.04779   2972.0193
#> 5084  714912.5 9559302 2359.77 47.968027   0.00107    0.02373    527.3528
#> 35194 713692.5 9559952 2185.68 28.796859  -0.00449   -0.00181   2406.1252
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092    412.5849
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595  -0.00211    0.00241    585.6182
#> 7466  712962.5 9560392 1819.03 18.402831  -0.00140   -0.02330   1456.9246
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543    352.1558
#> 29801 714392.5 9559162 2356.18 31.562526  -0.00856   -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598   0.00114    0.03026    247.0894
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224    942.0322
#> 12137 712842.5 9558902 2097.88 30.098619   0.00024   -0.00474   2343.2100
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 43859 714892.5 9557602 2722.70 28.129363   0.00370   -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 10073 715572.5 9558612 2632.78 32.161267   0.00129    0.00171    788.0004
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 43201 715712.5 9557922 2852.82 31.462258  -0.00949    0.00520   3390.5708
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 20797 713012.5 9560392 1842.25 36.649818   0.01299    0.00141   4611.0718
#> 21683 714792.5 9561052 1807.84 55.732751  -0.02070   -0.02661   2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 34615 713242.5 9560732 1879.60 27.593647   0.00384    0.00175   1026.6818
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 5911  713952.5 9561282 1801.16 33.789613   0.00466   -0.01776   1021.8267
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 17058 714032.5 9558502 2402.95 31.352824   0.01924    0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.590021  -0.00253    0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 16071 714562.5 9560062 2166.69 28.696018   0.02281    0.04759    158.9370
#> 13259 714862.5 9560992 1859.28 21.267620  -0.00317    0.00077   3056.9475
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200   1305.6691
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915   2036.0378
#> 1916  715202.5 9559512 2213.91 26.327984  -0.00292   -0.00109   2047.9031
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705    413.5702
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 10456 712442.5 9560292 2010.91 26.891774   0.00445    0.00875    597.2893
#> 49459 714692.5 9557342 2639.74 42.047081   0.00216   -0.02606    919.4241
#> 17674 713882.5 9558562 2339.20 41.036383  -0.03675   -0.01905   6487.0010
#> 43834 713092.5 9560662 1905.55 31.323603   0.00758   -0.00288   2144.8164
#> 31491 712832.5 9558892 2091.39 29.700413   0.00186    0.00063   2430.2278
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684    209.1778
#> 9984  712742.5 9560092 1848.56 23.612737   0.00614   -0.00093    118.1402
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301    338.9676
#> 46165 714392.5 9557912 2542.56 16.329870   0.00325   -0.00505    351.4078
#> 11288 712822.5 9560442 1873.42 39.609718  -0.00465   -0.00105   4432.8760
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085   2229.5105
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208    951.3401
#> 18746 713292.5 9560952 1890.51 41.885507  -0.00402   -0.01328   1066.7134
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 47005 714942.5 9559192 2333.66 29.220848  -0.04937   -0.02273  72123.6172
#> 16878 714232.5 9558172 2485.33 22.808304   0.00818    0.00502   1428.8546
#> 18627 714292.5 9559252 2330.78 18.496224  -0.00538   -0.00052   3508.5591
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131   0.00601   -0.00011    712.5228
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983    270.7949
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 13737 714272.5 9559622 2257.05 55.457158  -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.214748  -0.01338    0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312    593.7773
#> 34345 715852.5 9557902 2874.08 43.655946   0.00379   -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.074729  -0.00112   -0.00038   2017.3547
#> 45486 712762.5 9560172 1853.82 56.797306   0.01978    0.01282   2322.7944
#> 21070 712832.5 9560062 1884.42 50.933656  -0.00342    0.02552   2515.0752
#> 42424 714692.5 9560902 1921.34 25.331037   0.01154   -0.00204    618.2027
#> 44823 714202.5 9557442 2548.13 43.230239  -0.00233    0.01282    561.2366
#> 48988 714682.5 9560852 1942.97 20.290727   0.00480    0.00740    909.4791
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361   1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360   4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 14671 712672.5 9560372 1872.21 39.197189  -0.00873   -0.01417   1520.6630
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907   1281.5044
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 4269  714162.5 9560572 2087.45  7.334433   0.00293    0.00186   2751.4626
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 15412 37.592907   300.00   TRUE       300.00            100  3.318273
#> 36052 53.745160   300.00   TRUE       300.00            100  2.877158
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 46178 20.931294   300.00   TRUE       300.00             31  2.440958
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 39629 47.548622   300.00   TRUE       300.00             85  3.366461
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 30800 33.248168   300.00   TRUE       300.00            100  4.403435
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 23375 34.639882   300.00   TRUE       195.00              2  3.224136
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 6676  34.621548   300.00   TRUE       300.00            100  3.303289
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 38077 26.915265   300.00   TRUE       300.00            100  2.777281
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 6763  16.218143   125.47   TRUE         0.00              6  3.006344
#> 3466  26.794944   101.32   TRUE        95.02             19  3.724022
#> 35142 27.153043   300.00   TRUE       300.00            100  2.176525
#> 12084 34.595765   300.00   TRUE         0.00              0  3.904384
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 46756 38.961703   300.00   TRUE       300.00             37  2.473902
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 37348 39.287716   300.00   TRUE       300.00            100  4.465317
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 18394 28.610647   300.00   TRUE       300.00             57  3.372940
#> 3944  39.360482   300.00   TRUE       300.00            100  4.739031
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 5084  31.595185   300.00   TRUE       300.00             18  2.722101
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 22299 21.719111   300.00   TRUE       127.54              2  2.615513
#> 2718  36.732897   300.00   TRUE       300.00            100  2.990104
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 462    9.812475    61.17   TRUE        57.16             74  3.452385
#> 26698 34.249698   300.00   TRUE       300.00              7  2.767615
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 27875 21.267047   157.93   TRUE         0.00             57  3.709413
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 35083 43.540782   300.00   TRUE       300.00             23  2.546735
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 32735 39.358190   300.00   TRUE       300.00            100  3.245566
#> 44825 29.021458   300.00   TRUE       300.00            100  2.392854
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 12137 27.588491   300.00   TRUE       300.00            100  3.369811
#> 12772 37.920066   300.00   TRUE       300.00            100  2.936990
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 7086  34.204434   300.00   TRUE       128.61             63  3.071308
#> 24135 38.668349   300.00   TRUE       122.87            100  3.878395
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 20797 32.056989   285.07   TRUE         0.00             96  3.663802
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 27915 33.574754   300.00   TRUE       300.00              6  3.211627
#> 34615 34.189538    17.57   TRUE       142.95             16  3.011436
#> 15432 28.767638   300.00   TRUE       300.00            100  2.880810
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 30933 12.843422    99.29   TRUE         0.00              6  3.137802
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 16071 29.621345   300.00   TRUE       300.00            100  2.201225
#> 13259 23.160673   203.85   TRUE        10.45             74  3.485288
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 14192 40.586611   300.00   TRUE       300.00            100  3.115833
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 1916  29.273560   300.00   TRUE       300.00             21  3.311309
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 9606  33.823990   300.00   TRUE       300.00              0  2.979485
#> 42617 31.231929   300.00   TRUE       300.00            100  2.848895
#> 10456 21.668691    24.34   TRUE         0.00             18  2.776185
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 17674 30.482501   300.00   TRUE       300.00              1  3.812044
#> 43834 28.512099    20.21   TRUE       150.14              2  3.331390
#> 31491 27.558697   300.00   TRUE       300.00            100  3.385647
#> 34718 34.409553   300.00   TRUE       300.00            100  2.320516
#> 9984   3.946533   300.00   TRUE        36.46             76  2.072398
#> 17851 32.263253   300.00   TRUE       300.00            100  4.161853
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 46165 14.818407   300.00  FALSE       300.00            100  2.545811
#> 11288 28.537309   132.07  FALSE        85.50              1  3.646686
#> 34250 30.892738   300.00  FALSE       300.00            100  3.348210
#> 8635  27.681883   300.00  FALSE       300.00             23  2.716617
#> 14717 50.856307   300.00  FALSE       300.00            100  2.978336
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 441   25.113313   294.31  FALSE       152.09            100  3.054287
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 47005 36.135302   300.00  FALSE       300.00            100  4.858077
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 44668 30.102630     9.15  FALSE       163.30              6  2.852799
#> 8895  38.077629   300.00  FALSE       300.00             90  2.432641
#> 34105 20.077014   300.00  FALSE       300.00             26  3.059266
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 13737 40.960180   300.00  FALSE       300.00            100  2.678473
#> 3733  22.100128   300.00  FALSE       300.00            100  3.055767
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 17746 36.391414   300.00  FALSE       300.00             87  2.773624
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 12044 32.236897   300.00  FALSE       300.00            100  3.304782
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 21070 34.677125   300.00  FALSE         0.00             10  3.400551
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48988 18.958600   300.00  FALSE       144.19            100  2.958793
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 1281  35.393895   300.00  FALSE       300.00            100  3.632329
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 484   32.145797   300.00  FALSE        55.63            100  3.216864
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 44216 26.064996   300.00  FALSE       300.00            100  3.167741
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 48050 23.361208   300.00  FALSE        72.23            100  2.520806
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 46639 33.563295   300.00  FALSE       300.00            100  3.756965
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 28978  4.439850   300.00  FALSE        10.00            100  2.209674
#> 47629 31.663367   300.00  FALSE       300.00            100  3.627607