Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896  -0.00501   -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.9766069   0.00958    0.02642     671.1807
#> 42733 715112.5 9559062 2399.70 36.0545152  -0.00041    0.01741     746.4155
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339    1002.5877
#> 11088 715532.5 9558762 2549.28 38.0501272  -0.00373   -0.00057    1630.3259
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835  -0.02191   -0.00579    2246.7725
#> 39885 715062.5 9561022 1840.94 38.4357278   0.00250   -0.01340     604.1032
#> 1768  713912.5 9558552 2357.19 38.6924129  -0.00645    0.00835    2425.3115
#> 25690 715042.5 9557572 2752.67 32.7135983   0.00194    0.00526     539.8119
#> 10899 713852.5 9559652 2282.82 28.8140475   0.00952    0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304  -0.00769   -0.00281    1325.6138
#> 30800 715182.5 9557582 2772.37 39.2040642  -0.02475    0.00665   25318.3555
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392    1255.0884
#> 40307 713302.5 9560702 1845.07 39.0046749   0.00129    0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.9269342  -0.00070   -0.00530    9000.4356
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382     560.8309
#> 20290 714812.5 9561092 1775.73 34.6267043  -0.01537   -0.00203   10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214   0.00832    0.00838     238.6685
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678    1891.6584
#> 48364 715582.5 9558222 2746.23 25.0720602   0.01257    0.00593     939.0466
#> 17004 713102.5 9559032 2186.79 36.5335079  -0.01383    0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605     798.6775
#> 4329  712882.5 9559982 1911.46 26.2317267  -0.00193   -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.7801245   0.00646   -0.04086    6042.5757
#> 19128 714852.5 9557872 2681.49 20.7576880   0.00453    0.01077     300.7915
#> 37348 715632.5 9558182 2743.16 13.6203527  -0.00904   -0.00866   29195.6016
#> 19784 714982.5 9557722 2716.98 35.4362937   0.00888    0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.5357857   0.00140   -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.2770255   0.01458    0.00672     565.1107
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517    1000.7119
#> 6967  715342.5 9560192 1995.36 44.8815030  -0.00666   -0.00504    1020.2766
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907    1943.4092
#> 13134 714652.5 9559632 2265.33 49.5998104  -0.03552   -0.02058     431.4362
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.1730683   0.00359   -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.6656935   0.00290    0.00230    1056.5876
#> 16015 712582.5 9560312 1899.18 56.5469237   0.03480    0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 18892 712682.5 9560202 1839.08 40.9235105  -0.01364   -0.03357     473.4709
#> 39520 715342.5 9557622 2889.48 41.6488751  -0.01282    0.00542     860.8840
#> 10507 714892.5 9559282 2374.45 42.8423462  -0.01057    0.02427     380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.8738238   0.00090   -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.7501070  -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.9349627  -0.01807   -0.01883     616.1993
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361  -0.00099    0.02169     495.8840
#> 31561 713302.5 9559192 2092.11 36.5277783   0.00406   -0.00055     773.0359
#> 35014 713502.5 9559772 2190.48 44.2357795   0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.0591409  -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395    2011.3270
#> 2552  714942.5 9557652 2719.32 34.8736492  -0.01238   -0.00851    1012.5963
#> 41721 715182.5 9557562 2784.85 31.7899903  -0.00588    0.01448    1258.8196
#> 40216 714602.5 9560592 2015.55 34.3688733  -0.00093    0.00664    1807.2659
#> 1973  715252.5 9557822 2861.00 15.5707010   0.01865    0.00335     195.0293
#> 46461 712682.5 9560032 1892.30 22.9778994   0.01453    0.01337     486.7551
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524   10861.9482
#> 47718 712812.5 9559892 1867.55 54.9386311  -0.01491    0.01061    5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211   49347.5547
#> 18440 714862.5 9558932 2448.29 41.0524260  -0.04989   -0.02781   29995.1543
#> 16009 715122.5 9559402 2264.81 50.0186426  -0.00589    0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797    7557.7866
#> 41125 712772.5 9559412 2021.88 38.8763323  -0.00310   -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.1804123  -0.01059   -0.07431    4079.0366
#> 29228 712782.5 9560422 1870.56 32.9061121  -0.00218   -0.00372    4154.5786
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776    1021.8267
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.9597815  -0.01439   -0.01161    3065.3694
#> 14545 713562.5 9559602 2222.17 54.2390497   0.01011    0.00739     789.7573
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 37439 714602.5 9560672 1982.99 18.9064613  -0.00070    0.00010    3467.2690
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 10692 712712.5 9561022 2005.85 41.4334430   0.01410   -0.00380     601.0843
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190     264.6404
#> 49378 713252.5 9560732 1875.41 27.8354356   0.00181   -0.00001    1184.7952
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391    1794.7964
#> 48650 715592.5 9558612 2626.74 31.8765069   0.00010    0.00140    1066.3957
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.2552461   0.01257    0.00044     288.5468
#> 9606  714392.5 9559172 2353.05 30.7833035  -0.00185   -0.01185     953.8614
#> 22847 714582.5 9560382 2040.49 25.3361937   0.00181   -0.00341    1363.4000
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732    4237.6934
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 24532 715202.5 9560922 1799.62  4.4885514  -0.02312   -0.01848 1890804.2500
#> 43834 713092.5 9560662 1905.55 31.3236027   0.00758   -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.3812407  -0.00465   -0.01036    2505.8369
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 41408 713172.5 9559612 2056.81 22.2725884  -0.00166    0.01476    1404.4454
#> 23832 714252.5 9560182 2188.12 30.6475125   0.00427    0.02403     474.8613
#> 17851 715362.5 9557482 2869.26 29.1257366  -0.03251   -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021  -0.01519    0.00609    4275.3457
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230     714.7432
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463     851.6306
#> 19493 714252.5 9560372 2156.93 24.8795463   0.00825    0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 441   714312.5 9560972 1916.59 31.3029762   0.00540    0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.4949170   0.01306   -0.00626     450.7346
#> 39193 715332.5 9560852 1872.90 30.2905598  -0.00161    0.00171     756.8818
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199     836.4196
#> 36833 713322.5 9560572 1794.25 33.7540896   0.00527   -0.02077    8759.9961
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052    3508.5591
#> 32849 714302.5 9558222 2520.01 32.1572563  -0.00846    0.00216    1412.2295
#> 38689 714622.5 9559342 2264.90 42.6481135  -0.00844    0.00514     897.9290
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267   31588.1953
#> 6143  712762.5 9559022 2096.24 29.8963648   0.00358    0.02422     967.9738
#> 47059 713822.5 9558272 2413.20 38.1750320  -0.00687   -0.00223    1056.3701
#> 48869 712562.5 9560122 1972.89 24.9465824   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 3733  713762.5 9559632 2297.63 34.2147477  -0.01338    0.00518    1137.0179
#> 39440 714222.5 9557872 2508.52 39.0476467   0.02785    0.00645     234.0114
#> 17746 714752.5 9558322 2648.52 47.7365517  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196     576.1450
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506   0.01462    0.00268     252.0072
#> 7553  714642.5 9559132 2391.59 63.9678730  -0.00066    0.05257     549.3252
#> 35038 715092.5 9558152 2715.69 40.1838220   0.00270   -0.00250    2225.4426
#> 42424 714692.5 9560902 1921.34 25.3310371   0.01154   -0.00204     618.2027
#> 35826 712912.5 9560302 1891.22 50.8694849  -0.00229    0.00099     909.0560
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 26768 715092.5 9560582 1960.72 36.6561209   0.00980   -0.00710     472.0236
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105     868.7802
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417    1520.6630
#> 32430 715832.5 9558112 2824.39 23.6940967   0.00944    0.00836     250.5111
#> 23499 714272.5 9558232 2508.35 28.9676002  -0.00820   -0.00260    1484.6593
#> 38362 714402.5 9558042 2499.31 56.9010116   0.03044   -0.00964     543.0025
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.8303070   0.01852    0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532   16771.3066
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208     842.4825
#> 19767 714272.5 9558052 2431.67 55.7247292   0.00361   -0.01841     756.0055
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972     331.7462
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148     481.8292
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233     968.7780
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177     184.1042
#> 46148 715082.5 9558042 2725.39 42.8360436  -0.00181   -0.00799    1036.6606
#> 32342 714862.5 9559622 2294.74 36.9672369   0.01021    0.00109     915.4989
#> 30403 713952.5 9560752 2003.57 42.9643862  -0.00933   -0.00157     766.0220
#> 14645 715072.5 9560262 2034.04 38.1693024  -0.00716   -0.03234    1707.3130
#> 26635 715092.5 9560412 2010.59 30.5718820  -0.00463    0.00253    1237.0536
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388     180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525     162.0592
#> 40852 715392.5 9557932 2939.71 18.2974072   0.00505    0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 23178 41.6001737   300.00   TRUE       300.00             21  2.826839
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 1768  28.1855128   300.00   TRUE       300.00              7  3.384768
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 29420 28.8524357   253.02   TRUE         0.00            100  3.954264
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 48364 32.0758962   300.00   TRUE       300.00            100  2.972687
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 44612 34.6914486   300.00   TRUE       300.00            100  2.460216
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 31491 27.5586970   300.00   TRUE       300.00            100  3.385647
#> 41408 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 17746 36.3914144   300.00  FALSE       300.00             87  2.773624
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 7553  39.7712924   300.00  FALSE       300.00            100  2.739830
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 42424 22.6071957   275.21  FALSE        93.50             70  2.791131
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 32430 22.8604431   300.00  FALSE       300.00            100  2.398827
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983