Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681   -0.00029    5577.3916
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 46178 714892.5 9559312 2379.54 32.766883   -0.00266    0.02896     276.0310
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250    1413.3312
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767    3321.6113
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940    0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257     357.1436
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088     266.4355
#> 39629 714602.5 9558912 2474.89 40.258307   -0.06175   -0.08226    2325.2053
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010    1527.1028
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392    1255.0884
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131    0.00501    3889.7292
#> 36030 712852.5 9559572 1938.79 28.678829   -0.01440    0.00710     843.7961
#> 38070 715212.5 9559222 2342.41 55.677174    0.00426   -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491    0.00581    1855.1630
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173    8967.1924
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108   -0.00678    1891.6584
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629  360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898     588.3761
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694    2123.4473
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452   -0.01437    1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202   -0.01804   -0.01596    5296.9014
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353    0.00063    5035.0000
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063   18733.8848
#> 3944  715262.5 9558352 2654.20 33.555273   -0.04492    0.00072   54831.6289
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 22299 713532.5 9560362 1975.58 27.709958    0.00478    0.01092     412.5849
#> 2718  715062.5 9558222 2719.56 44.753160   -0.00119   -0.01560     977.4719
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197     711.1091
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048   -0.00332    1418.6622
#> 2595  714912.5 9557662 2700.72 24.194862   -0.02681   -0.01319    3457.7519
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458   -0.00401    1722.6481
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542    3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664    1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972    0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945   -0.00078    0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132   42970.7891
#> 41783 715572.5 9558652 2609.51 32.408212    0.00291   -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629   -0.00469    1021.5407
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399    3900.3445
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 6845  713412.5 9559172 2123.32 47.153854   -0.01556   -0.04394     587.8840
#> 18958 714192.5 9558372 2495.72 34.677125   -0.00178   -0.00822     472.8748
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370   -0.00190     640.8842
#> 1973  715252.5 9557822 2861.00 15.570701    0.01865    0.00335     195.0293
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759   -0.00400     608.6523
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129    0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025   -0.01245    6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332   -0.00058    1817.2371
#> 17345 713922.5 9559682 2227.44 34.490913   -0.06509   -0.01211   49347.5547
#> 31117 712892.5 9558922 2129.84 30.550683    0.00521   -0.00870    1895.1399
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860     695.2302
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949    0.00520    3390.5708
#> 20797 713012.5 9560392 1842.25 36.649818    0.01299    0.00141    4611.0718
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078  622888.2500
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218   -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476    1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833     369.3771
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298    1533.7925
#> 11352 713002.5 9560562 1858.95 39.803951   -0.00442   -0.00118   19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396     704.7076
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669   -0.01981    3975.3608
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190     264.6404
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075    2170.2249
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133    2397.2681
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121    1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730   -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360   -0.03000  162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 37115 715612.5 9559202 2358.15  9.727677   -0.01020    0.00279   39495.1797
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115    0.01015     790.2421
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144  317260.6875
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253     528.1924
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840    0.00189     556.5211
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585   -0.02665    1904.4769
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130   -0.01084   -0.02996    1599.3491
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548   -0.00203    0.00642    1688.1782
#> 24836 714202.5 9558272 2507.87 38.489013   -0.00226    0.00736     477.8591
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156   -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695   -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403     241.2991
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494    0.01064     683.2795
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306   -0.00626     450.7346
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464    4029.5947
#> 43243 715682.5 9558462 2693.23 27.406290   -0.00031   -0.00199     836.4196
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367    3002.8828
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759    1880.4403
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452   -0.01298     436.1459
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400    4126.7812
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462     716.5762
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032     994.9224
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759    4131.5356
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714  480535.5000
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978    0.01282    2322.7944
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948    1222.5433
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312     254.0656
#> 48988 714682.5 9560852 1942.97 20.290727    0.00480    0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361    1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050   -0.00930   -0.00360    4288.7329
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555    1882.0179
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578    1258.8865
#> 16845 712772.5 9560022 1831.76 21.399974   -0.00865   -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287    0.00253   -0.01233     968.7780
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377    1727.6055
#> 9320  713872.5 9559262 2378.76 29.858550    0.00309    0.01011     501.7867
#> 9526  714432.5 9560582 2076.13 31.080096   -0.00172   -0.00119     792.3544
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 11570 715572.5 9558482 2685.72 16.814019    0.00282    0.00458     485.7418
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962   -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288    0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156     758.4840
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855     229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 39629 47.54862150   300.00   TRUE       300.00             85  3.366461
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 36030 29.19964811   300.00   TRUE         0.00            100  2.926238
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 18958 29.98975691   300.00   TRUE       300.00              5  2.674746
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 1973  19.32586643   300.00   TRUE       300.00            100  2.290100
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 5572  26.59784677   300.00   TRUE       300.00            100  2.567470
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 37115 32.18705006   300.00   TRUE       300.00            100  4.596544
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 14779 16.52123802    55.08   TRUE         0.00              1  2.897760
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 34347 31.72008945   300.00  FALSE       300.00            100  3.477538
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491