Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.952326   -0.00123    0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023   -0.00043    1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476   -0.00075    0.00475    1001.0861
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041    0.01741     746.4155
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339    1002.5877
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236   -0.00857    0.03677    1675.4679
#> 24129 714602.5 9560542 2038.30 25.584857    0.00065    0.00005     769.3426
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174    2310.4871
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835    2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607    0.00677    1343.5006
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514     297.7842
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412    0.00898     458.6938
#> 48055 712862.5 9559972 1902.48 29.016875   -0.00720    0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382    0.00672    1114.8801
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626    4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203   10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151    0.00189    2317.1819
#> 25081 713392.5 9558452 2242.24 46.538497   -0.02716    0.03696    4143.4019
#> 4827  714962.5 9559882 2274.42 29.244912    0.00051    0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.817660    0.00713   -0.00173    8967.1924
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739     558.2286
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033   -0.03917     116.1986
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212     561.1948
#> 13847 715052.5 9559232 2337.87 26.114843   -0.01452   -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996    0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276    9614.6924
#> 2950  715042.5 9557712 2740.72 26.525081   -0.00089   -0.00051     779.0555
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077     300.7915
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154    0.00124    2114.2253
#> 40738 714972.5 9557642 2745.48 42.542689   -0.00294   -0.00096     520.5570
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606   -0.00556     660.3690
#> 19784 714982.5 9557722 2716.98 35.436294    0.00888    0.00862     874.5408
#> 36931 712852.5 9558892 2102.80 30.535786    0.00140   -0.00070    2562.7559
#> 46235 712712.5 9561042 2023.11 41.277025    0.01458    0.00672     565.1107
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940    0.00679     550.0344
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060    0.00270    4057.5637
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290    0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 20220 714052.5 9558522 2378.67 39.656128   -0.00855   -0.02175    1122.9120
#> 462   713712.5 9561182 1785.08 47.717071   -0.03260   -0.03310    2833.9001
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211    0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221    0.00121    4159.6738
#> 25700 715172.5 9559272 2324.49 35.534842   -0.00529    0.02009    1247.7061
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406   -0.00055     773.0359
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558    1504.9717
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927   -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 38197 712522.5 9560322 1942.04 62.777267   -0.01611    0.02961    1086.6963
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605   -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.098619    0.00024   -0.00474    2343.2100
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093    0.00664    1807.2659
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246    0.00534     344.1447
#> 1424  714012.5 9558902 2305.97 32.223719   -0.03986   -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129    0.00171     788.0004
#> 6217  714442.5 9559932 2151.30 36.670445    0.01489    0.00241    1752.9242
#> 41125 712772.5 9559412 2021.88 38.876332   -0.00310   -0.00860     695.2302
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661    2508.5374
#> 30933 714672.5 9561082 1798.37 54.810289    0.00542   -0.00722    1373.4147
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296    0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292   -0.00109    2047.9031
#> 750   712812.5 9560032 1847.09 54.959258   -0.00767   -0.03713    6379.7974
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153   -0.02417   34908.6562
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450    2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915   -0.02705     413.5702
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194    0.00181   -0.00341    1363.4000
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632     620.0229
#> 14140 714672.5 9559262 2326.36 45.792506   -0.01400   -0.01210     585.8992
#> 46971 713532.5 9559952 2198.03 29.723904   -0.00529    0.01230    1044.6443
#> 6212  714592.5 9560362 2034.54 12.522566   -0.01454    0.00144  317260.6875
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025   -0.05345    2417.1389
#> 39466 714942.5 9557772 2680.35 16.753286    0.00166   -0.01265    1148.6487
#> 6470  713742.5 9558802 2260.79 30.680171    0.00000    0.00320     911.5911
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519    0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797    7557.7866
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989   -0.02781   29995.1543
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336    2806.2644
#> 8635  713832.5 9557902 2397.71 47.119476    0.02295    0.02715     520.7355
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118    2273.4758
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197    0.00463     851.6306
#> 49363 712772.5 9559552 1935.91 29.036928   -0.03220   -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195   -0.00995    1026.7981
#> 441   714312.5 9560972 1916.59 31.302976    0.00540    0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945     259.7421
#> 49835 713272.5 9560372 1898.68 32.825898    0.00846    0.00794    1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502    1428.8546
#> 20547 713732.5 9560822 1851.67 36.968383   -0.01405    0.00476   11992.5273
#> 40139 712852.5 9560072 1907.89 43.704075    0.00001    0.00759    1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216    1412.2295
#> 21925 714542.5 9559232 2314.33 50.247826    0.00737    0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.085050   -0.00196   -0.00175     725.1355
#> 37029 713022.5 9559082 2163.08 36.032170    0.00490   -0.00780     825.4451
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422     967.9738
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423   -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.216858   -0.01292    0.00462     716.5762
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041   -0.00011     201.1272
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.457158   -0.03224   -0.02036     476.9504
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018     937.6911
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035    0.00955     713.4463
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310   -0.07650     766.6254
#> 35038 715092.5 9558152 2715.69 40.183822    0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204     618.2027
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858    0.00312     254.0656
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944    0.00836     250.5111
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438    1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907    1281.5044
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875    2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576    1045.1074
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278   -0.01172  111973.8906
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340    0.01809    1158.0426
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161    0.00709    1181.2555
#> 33881 712992.5 9558822 2173.26 33.197748   -0.00595    0.03064    1485.5201
#> 32342 714862.5 9559622 2294.74 36.967237    0.01021    0.00109     915.4989
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573    1490.0312
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241    0.00020     673.2861
#> 40043 713322.5 9560562 1802.09 39.852079    0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 42632 20.47808456   300.00   TRUE       300.00             10  2.566509
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 44952 34.08239444   300.00   TRUE       300.00            100  3.363704
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34565 20.38010877   300.00   TRUE        68.13            100  2.749114
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 46235 30.10835918   300.00   TRUE        41.37              0  2.752133
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 20220 32.17559090   300.00   TRUE       300.00              8  3.050346
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 3418  35.45348245   300.00   TRUE       300.00             73  4.552486
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 10073 25.76132838   300.00   TRUE       300.00            100  2.896526
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 41125 36.55585324   300.00   TRUE         0.00            100  2.842129
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 14140 35.48728696   300.00   TRUE       300.00             65  2.767823
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 6212  31.93151088   300.00   TRUE       300.00             47  5.501416
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 25140 14.48380010   104.12  FALSE        46.28             74  2.414542
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 21925 31.24854519   300.00  FALSE       300.00             95  3.537385
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 42424 22.60719572   275.21  FALSE        93.50             70  2.791131
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 18308 30.02069663   300.00  FALSE       300.00             55  4.224567
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 32618 39.38282701   300.00  FALSE       300.00             56  3.111762
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603