Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029    5577.3916
#> 37912 712802.5 9559952 1838.40 52.1013441  0.00183 -0.09203     634.3320
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393    2081.0056
#> 28223 712742.5 9560482 1927.69 34.6679573 -0.00292  0.00712    3636.5303
#> 5999  715552.5 9558812 2518.03  7.9841669 -0.01863 -0.00507  159065.0781
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292    1233.9299
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403    1787.3024
#> 38827 713412.5 9560472 1869.10 16.8713152 -0.00156  0.00406    2421.1145
#> 38993 714202.5 9558402 2475.75 33.1593594 -0.01607  0.00677    1343.5006
#> 11463 713112.5 9558902 2244.08 21.8915714  0.00160  0.00311     586.7354
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328    2235.5259
#> 10899 713852.5 9559652 2282.82 28.8140475  0.00952  0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.1555304 -0.00769 -0.00281    1325.6138
#> 33792 714652.5 9558172 2598.36 59.4586952 -0.03525 -0.11024     394.8751
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272    4955.5586
#> 15109 713872.5 9559602 2288.64 25.1855058  0.00602  0.01518     854.4269
#> 18304 714582.5 9558932 2463.25 22.1946661 -0.02131  0.00501    3889.7292
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471 -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.8152355 -0.00857  0.03677    1675.4679
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672  0.04398     267.2659
#> 14522 715012.5 9557732 2721.34 35.1813275 -0.00382  0.00672    1114.8801
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899    1394.2274
#> 6676  715092.5 9559912 2217.44 41.7078897 -0.00077 -0.00284    2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713 -0.00173    8967.1924
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678    1891.6584
#> 38077 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733     598.7983
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225  0.00325     616.8652
#> 17004 713102.5 9559032 2186.79 36.5335079 -0.01383  0.00193    1809.3838
#> 2950  715042.5 9557712 2740.72 26.5250811 -0.00089 -0.00051     779.0555
#> 4329  712882.5 9559982 1911.46 26.2317267 -0.00193 -0.00077    3559.4204
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696  0.02514     297.7842
#> 40738 714972.5 9557642 2745.48 42.5426892 -0.00294 -0.00096     520.5570
#> 11687 714552.5 9560382 2052.66 29.5755084 -0.00255 -0.00365    1536.2647
#> 37348 715632.5 9558182 2743.16 13.6203527 -0.00904 -0.00866   29195.6016
#> 27084 713632.5 9558472 2380.37 23.1354628 -0.01752 -0.00517    1000.7119
#> 6967  715342.5 9560192 1995.36 44.8815030 -0.00666 -0.00504    1020.2766
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025    4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591  0.00288  0.01452     437.3965
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925     665.6116
#> 651   714602.5 9560152 2111.23 53.8517302 -0.01149  0.01949    1182.1506
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072   54831.6289
#> 39599 714112.5 9558562 2374.25 29.1681354  0.00283  0.00117    1018.2141
#> 21952 715352.5 9560172 2013.44 40.4880626  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.3489734 -0.10372 -0.00008  825875.6875
#> 835   715052.5 9558212 2707.65 37.7252601 -0.01399 -0.00731    1384.7959
#> 2718  715062.5 9558222 2719.56 44.7531604 -0.00119 -0.01560     977.4719
#> 421   714892.5 9561022 1849.01 26.1320321  0.00002  0.00588     749.2502
#> 375   715622.5 9558922 2510.87 37.3665885 -0.01017  0.00197     711.1091
#> 19887 714662.5 9559632 2266.47 45.1794410 -0.02797 -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.8312924  0.03658  0.04602     345.8933
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690    5121.6924
#> 28845 714892.5 9559272 2374.42 45.7501070 -0.00209  0.02739     365.7626
#> 37416 713282.5 9560692 1850.61 30.2590471  0.00736  0.01684    2556.8049
#> 32735 714922.5 9559002 2414.49 40.0646468  0.00062 -0.00262    1760.2162
#> 20065 713772.5 9558852 2221.21 28.0199280 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.7539361 -0.00099  0.02169     495.8840
#> 46065 713072.5 9559162 2109.39 16.2645529 -0.00195 -0.00656    3036.7383
#> 17728 713092.5 9560652 1900.14 30.2791006  0.00645 -0.00395    2011.3270
#> 45500 712812.5 9560052 1852.16 56.7703772  0.00904 -0.02013    4229.5708
#> 21680 712802.5 9560462 1893.75 38.7342388 -0.00481 -0.00399    3900.3445
#> 5239  714832.5 9560992 1859.70 19.9601307 -0.00320  0.00180    3904.5254
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472 -0.00228    2262.3796
#> 47366 714892.5 9557612 2718.78 28.0239387  0.00538  0.00112     745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474    2343.2100
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133  0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242     507.9263
#> 8136  713242.5 9560602 1798.18 41.2403562 -0.01169 -0.03220    8936.2666
#> 47320 713422.5 9559182 2143.03 56.3251890  0.01246  0.00534     344.1447
#> 41125 712772.5 9559412 2021.88 38.8763323 -0.00310 -0.00860     695.2302
#> 42201 714642.5 9557452 2592.29 13.2255848 -0.02084 -0.02236    9755.3027
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187 -0.01175  0.04195    1901.6743
#> 29228 712782.5 9560422 1870.56 32.9061121 -0.00218 -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.2725884 -0.00166  0.01476    1404.4454
#> 7485  714362.5 9558042 2484.02 47.9588593  0.03169  0.01441     424.2680
#> 47298 714602.5 9560572 2025.50 28.5900210 -0.00253  0.00753    1337.7854
#> 39325 714142.5 9558592 2357.72 40.7928761 -0.00246  0.00396     704.7076
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.6960182  0.02281  0.04759     158.9370
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042  0.00957     495.4453
#> 22297 714642.5 9558172 2606.18 59.4690084 -0.00495 -0.03756     353.9086
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705 -0.00915    2036.0378
#> 1916  715202.5 9559512 2213.91 26.3279836 -0.00292 -0.00109    2047.9031
#> 20426 714702.5 9559712 2208.61 43.0463191 -0.08153 -0.02417   34908.6562
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185 -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.3774607  0.00730 -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526     706.1463
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.3218698 -0.00157 -0.00152     850.4633
#> 28588 715532.5 9559362 2299.84 11.1446021 -0.04360 -0.03000  162178.1562
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856    1789.1596
#> 47722 712732.5 9561022 1999.03 37.4926392  0.00004 -0.02384     947.2075
#> 35458 715242.5 9558442 2623.01 38.9548275 -0.00770 -0.03400    1982.7782
#> 43834 713092.5 9560662 1905.55 31.3236027  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.5225656 -0.01454  0.00144  317260.6875
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870    2099.7930
#> 21490 714672.5 9560382 2021.86 32.8041256 -0.00783 -0.00367    1966.2112
#> 23176 714012.5 9558892 2308.59 43.0291304 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382     100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331     699.3024
#> 29431 714882.5 9559892 2258.03 21.9179275 -0.00849  0.00150     503.5590
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325 -0.00505     351.4078
#> 30809 712532.5 9560032 1928.24 40.8994463  0.00570  0.02230     714.7432
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162 -0.00517    1060.5494
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118    2273.4758
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290    2481.5632
#> 37832 712552.5 9559192 1998.10 41.5508993 -0.00195 -0.00995    1026.7981
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031 -0.00199     836.4196
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598     186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476   11992.5273
#> 31497 714032.5 9560672 2060.77 16.2169974  0.00973 -0.00943     371.5514
#> 36126 715052.5 9558292 2759.47 29.0621382  0.00196  0.01734     335.0773
#> 34616 714522.5 9558742 2568.77 34.9979810  0.00990  0.02790     209.6952
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.5071308  0.00601 -0.00011     712.5228
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514     897.9290
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706 -0.03035    1471.9783
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298     436.1459
#> 8895  714132.5 9557692 2487.19 37.3642967  0.03423 -0.00983     270.7949
#> 24009 714042.5 9558962 2357.58 34.3619978  0.05041 -0.00011     201.1272
#> 29835 715402.5 9558512 2610.14 27.1180924 -0.01322 -0.00759    4131.5356
#> 14148 714162.5 9560352 2149.57 37.1333947  0.00424  0.00196     576.1450
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009 -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119  0.00361    1011.0248
#> 26768 715092.5 9560582 1960.72 36.6561209  0.00980 -0.00710     472.0236
#> 24819 714032.5 9560882 1981.86 33.2842642  0.00804 -0.00165     805.1403
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600 -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.1722787 -0.03757 -0.00003 5426606.0000
#> 40750 712962.5 9558702 2126.06 34.7040537 -0.00264  0.00624    1870.8307
#> 4943  712452.5 9559172 1927.46 34.4651303  0.00105 -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.7218013 -0.00598  0.02208     842.4825
#> 4269  714162.5 9560572 2087.45  7.3344327  0.00293  0.00186    2751.4626
#> 10981 713832.5 9560022 2125.33 34.8484390 -0.06909 -0.01151  143213.1562
#> 18504 713372.5 9560672 1786.89  6.4738501 -0.00278 -0.01172  111973.8906
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.5342555 -0.00163 -0.00377    1727.6055
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884     672.2466
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858     298.4490
#> 49812 713392.5 9559222 2133.99 35.6626120 -0.00983  0.00803    1058.5630
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181 -0.00799    1036.6606
#> 6577  714142.5 9559902 2244.30 57.8905097 -0.06986 -0.05013     616.4128
#> 11570 715572.5 9558482 2685.72 16.8140195  0.00282  0.00458     485.7418
#> 23808 715172.5 9558702 2533.50 22.3459270  0.00811 -0.02281     492.4785
#> 32618 715132.5 9559642 2266.36 34.5327393 -0.00305 -0.01135    1293.4855
#> 11025 714402.5 9559252 2325.64 49.6204369 -0.00241  0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.1245906 -0.00166  0.00156     758.4840
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586     941.0288
#> 40852 715392.5 9557932 2939.71 18.2974072  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 38287 14.8969027    41.43   TRUE         1.90            100  3.349380
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 33792 51.2081029   300.00   TRUE       300.00              5  2.596460
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 4329  34.5619601   300.00   TRUE        16.35              2  3.551379
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 40738 28.9601518   300.00   TRUE       300.00            100  2.716468
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 37348 39.2877160   300.00   TRUE       300.00            100  4.465317
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 45500 34.9561551   300.00   TRUE         0.00              2  3.626296
#> 21680 28.3482328   104.12   TRUE       110.76              4  3.591103
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 7485  39.4905431   300.00   TRUE       300.00            100  2.627640
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 1916  29.2735597   300.00   TRUE       300.00             21  3.311309
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 17825 30.7334561    73.66   TRUE       137.63             18  3.469116
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 30378 28.5693945   300.00   TRUE       300.00             57  3.322176
#> 21490 34.0600491   300.00   TRUE       300.00              1  3.293630
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 28253  2.9576081   300.00  FALSE       300.00            100  2.000000
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 49363 32.9576783   300.00  FALSE         0.00            100  3.394725
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 24009 38.4970343   300.00  FALSE       300.00            100  2.303471
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 4943  37.7716697   300.00  FALSE       300.00            100  3.019161
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 4269  19.1167368   300.00  FALSE       194.53            100  3.439564
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 37254  4.1911863   131.59  FALSE        80.55            100  2.827529
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 49812 33.3581758   300.00  FALSE       300.00             65  3.024717
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 32618 39.3828270   300.00  FALSE       300.00             56  3.111762
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 41373 31.1287970   300.00  FALSE       120.90            100  2.879946
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983