Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634     500.5027
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921   -0.12438     753.6304
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357    1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021   -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292    0.00712    3636.5303
#> 1300  714042.5 9558482 2408.27 24.148452  0.00659    0.01041     772.9746
#> 8059  713292.5 9559122 2136.06 29.662025  0.02240   -0.00250    1413.3312
#> 9669  713852.5 9558612 2308.97 52.180412 -0.01059   -0.07431    4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863 -0.02104   -0.02046    1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891   -0.00949    6217.8213
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284    0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156    0.00406    2421.1145
#> 125   713992.5 9561062 1878.28 49.971087  0.00183    0.03387    1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327 -0.01407    0.00547     519.4228
#> 1393  714832.5 9557902 2664.72 31.295528  0.00550   -0.00120     512.9740
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378     440.9104
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414   -0.00104    1641.4587
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169   -0.00160    9344.2930
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270   -0.00769    1032.7177
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518     854.4269
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677    1675.4679
#> 35942 712862.5 9558902 2109.97 31.400379  0.00151    0.00189    2317.1819
#> 2942  713232.5 9560652 1851.75 34.535031  0.00856    0.01114    4510.8833
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716    0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733     598.7983
#> 18391 715112.5 9559432 2235.02 22.257691 -0.02651   -0.03629  360379.1875
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917     116.1986
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382    0.00088    2322.1746
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669   -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.009098  0.01937    0.00083     994.7876
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895  0.01275    0.00605     798.6775
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696    0.02514     297.7842
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167    1938.5389
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353    0.00063    5035.0000
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140   -0.00070    2562.7559
#> 38686 714292.5 9557832 2496.22 38.480991 -0.00509    0.00499     885.0064
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316    0.00714     475.3120
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321    0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036     516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843    0.04407     344.8395
#> 17423 715592.5 9558562 2654.69 33.219520 -0.00034    0.00294     739.5768
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613    0.00903    2360.1536
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359   -0.00279     930.0871
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283    0.00117    1018.2141
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373     527.3528
#> 421   714892.5 9561022 1849.01 26.132032  0.00002    0.00588     749.2502
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797   -0.00063     576.2819
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658    0.04602     345.8933
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681   -0.01319    3457.7519
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529    0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209    0.02739     365.7626
#> 47100 712752.5 9560462 1910.00 41.911863 -0.00032    0.00542    3614.7334
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807   -0.01883     616.1993
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099    0.02169     495.8840
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558    1504.9717
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972    0.03512   13178.3223
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099    0.00240    1201.4724
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605   -0.03594   16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393    8301.6094
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242     507.9263
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188    0.00832    3022.3462
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394     587.8840
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376   -0.01644 1057765.6250
#> 47718 712812.5 9559892 1867.55 54.938631 -0.01491    0.01061    5744.7583
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574     183.3220
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175    0.04195    1901.6743
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007   -0.02132    1580.1621
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379    0.00801     172.7491
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428   -0.00522    1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070  0.00511   -0.01801    1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546    0.01484     495.0292
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410   -0.00380     601.0843
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296    0.00426    1315.5334
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044     288.5468
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730   -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708   -0.02538     807.8237
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875     597.2893
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216   -0.02606     919.4241
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770   -0.03400    1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983    0.01253     528.1924
#> 48461 714672.5 9559272 2316.84 40.381492 -0.01966   -0.02234     849.0547
#> 36744 714922.5 9557722 2692.39 25.980771  0.00316   -0.00427     866.9398
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781   29995.1543
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084   -0.02996    1599.3491
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325   -0.00505     351.4078
#> 48601 713692.5 9560152 2058.14 59.557244 -0.01433   -0.02177     997.3089
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230     714.7432
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390   -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156   -0.00116    1817.5222
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363   -0.00363     389.2691
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132     737.5982
#> 8066  715842.5 9558732 2636.38 33.717420 -0.00494    0.01064     683.2795
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445     355.5078
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309   -0.03409    1575.9425
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171     756.8818
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502   -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031   -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846    0.00794    1363.5321
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176    0.02973     215.5101
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405    0.00476   11992.5273
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167    0.01133     828.8604
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507     504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065     704.4843
#> 40139 712852.5 9560072 1907.89 43.704075  0.00001    0.00759    1880.4403
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065    0.00475     515.1661
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432   -0.00362     283.7090
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193   -0.00267   31588.1953
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022    0.00662    1269.2650
#> 33175 715822.5 9558832 2563.13 36.622889 -0.00674   -0.00376     629.8236
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223    1056.3701
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322   -0.00759    4131.5356
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978    0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305   -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418    0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270   -0.00250    2225.4426
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947     842.1514
#> 1281  715102.5 9560032 2120.64 34.649050 -0.00930   -0.00360    4288.7329
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262   -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537   -0.00577     423.5598
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420    0.01960     884.2003
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820   -0.00260    1484.6593
#> 974   713402.5 9561062 1854.56 23.634509  0.00082    0.02578    1258.8865
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070   -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600   -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757   -0.00003 5426606.0000
#> 26283 715932.5 9557592 3075.48 32.766883 -0.00375    0.00215     917.5296
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105   -0.01576    1045.1074
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598    0.02208     842.4825
#> 48050 714022.5 9560572 2099.82 18.713948  0.01182   -0.00972     331.7462
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129    0.00491     261.7371
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186    2751.4626
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284   -0.00886   43249.3945
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948    1538.1741
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253   -0.01233     968.7780
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196   -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012     305.3624
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298    0.00222     991.8300
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288    0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020     673.2861
#> 41373 715332.5 9560862 1873.10 29.124591 -0.00166    0.00156     758.4840
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208    0.00492     772.2691
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456   -0.01586     941.0288
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 1393  23.41220142   300.00   TRUE       300.00              1  2.710095
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 18391 34.50752913   300.00   TRUE       300.00            100  5.556760
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 8073  36.07571461   138.71   TRUE        76.41             35  2.997730
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 38686 31.18895758   300.00   TRUE       300.00             75  2.946946
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 7668  36.32036759   300.00   TRUE       300.00              5  2.789721
#> 30886 39.18515657   300.00   TRUE       300.00            100  2.695380
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 28312 41.64257255   300.00   TRUE       300.00            100  3.198702
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10692 33.64637356   300.00   TRUE        60.19             11  2.778935
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 36744 29.40476700   300.00   TRUE       300.00            100  2.937989
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 43243 27.98612350   300.00  FALSE       300.00            100  2.922424
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 39685 47.87348857   300.00  FALSE       300.00            100  2.918481
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 1577  38.92961739   300.00  FALSE       300.00             63  4.635980
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603