Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv       carea
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261    323.8441
#> 29391 712842.5 9560152 1930.16 28.0955584   0.00753    0.01487   1077.7585
#> 24051 715462.5 9559372 2319.46 32.3279977  -0.00661   -0.01339   1002.5877
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767   3321.6113
#> 24072 715282.5 9557602 2837.46 34.3940835  -0.02191   -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.2854102   0.00118    0.00761    928.6858
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491   1844.4353
#> 37270 714852.5 9561032 1838.96 35.1589821  -0.00854    0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.6567008   0.00730    0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.7033271  -0.01407    0.00547    519.4228
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088    266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.1593594  -0.01607    0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.8654321   0.00102    0.00388   1051.4222
#> 11463 713112.5 9558902 2244.08 21.8915714   0.00160    0.00311    586.7354
#> 2990  713542.5 9560372 1970.07 31.8879661   0.01903    0.01447    331.4833
#> 43964 714802.5 9558452 2684.33 37.9825182  -0.02123    0.01164    895.6606
#> 48498 714972.5 9557972 2681.66 38.3469193   0.00528    0.01392   1255.0884
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898    458.6938
#> 48055 712862.5 9559972 1902.48 29.0168746  -0.00720    0.01250   3686.1240
#> 38279 714192.5 9558612 2325.14 28.2525489  -0.04430   -0.06070 138869.3906
#> 26215 714332.5 9560262 2138.08 25.7212213  -0.00491    0.00581   1855.1630
#> 13585 714552.5 9560092 2123.69 57.5289733  -0.03613   -0.01898    588.3761
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437   1155.1469
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156    816.6036
#> 23141 715632.5 9558942 2495.99 34.3768948   0.01275    0.00605    798.6775
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114    150.1498
#> 29667 713892.5 9559712 2239.25 41.5927252   0.00804    0.00346    543.0749
#> 5728  713502.5 9559662 2140.27 34.3075669  -0.00154    0.00124   2114.2253
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063   5035.0000
#> 27084 713632.5 9558472 2380.37 23.1354628  -0.01752   -0.00517   1000.7119
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796   4729.6382
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025   4425.5151
#> 5578  714962.5 9557612 2751.95 39.1914591   0.00288    0.01452    437.3965
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175    804.5679
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626   1014.7159
#> 651   714602.5 9560152 2111.23 53.8517302  -0.01149    0.01949   1182.1506
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560    977.4719
#> 42941 713252.5 9559742 1990.10 28.9544222  -0.04708    0.00408  76318.0156
#> 27875 714832.5 9561032 1838.22 41.2122812  -0.01160    0.01690   5121.6924
#> 13029 713542.5 9560242 2019.80 29.5451417   0.00576   -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.9118627  -0.00032    0.00542   3614.7334
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684   2556.8049
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473   2930.2791
#> 44650 715612.5 9558142 2764.88 39.3604817  -0.00550   -0.00941   1290.7709
#> 10364 714762.5 9561002 1856.21 25.2748872  -0.00845    0.00454   3476.0818
#> 5103  714612.5 9559892 2220.36 49.0079450  -0.00078    0.00219    420.7032
#> 26010 712852.5 9560542 1928.25 30.9781091   0.00165    0.00555   4439.0967
#> 5239  714832.5 9560992 1859.70 19.9601307  -0.00320    0.00180   3904.5254
#> 7385  713752.5 9560122 2064.61 31.7710827  -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112    745.3315
#> 39722 715022.5 9557712 2735.53 30.2607659   0.00133    0.00367    806.0966
#> 30268 714172.5 9558612 2328.84 54.9506633   0.03743   -0.05242    507.9263
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008 825875.6875
#> 43859 714892.5 9557602 2722.70 28.1293630   0.00370   -0.00190    640.8842
#> 32879 713912.5 9559672 2238.00 31.9767745  -0.00278   -0.00412   1350.3611
#> 17345 713922.5 9559682 2227.44 34.4909134  -0.06509   -0.01211  49347.5547
#> 30287 713992.5 9557822 2403.92 18.2154742   0.03890   -0.00911    173.2806
#> 34887 714662.5 9557392 2609.02 51.0207457  -0.01878   -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195   1901.6743
#> 15981 713302.5 9559172 2106.91 35.9977924   0.01354    0.00245    629.5983
#> 5911  713952.5 9561282 1801.16 33.7896130   0.00466   -0.01776   1021.8267
#> 47298 714602.5 9560572 2025.50 28.5900210  -0.00253    0.00753   1337.7854
#> 10617 713232.5 9560722 1879.95 26.4970062   0.00258    0.00422   1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340   0.01003    0.01196   1052.5341
#> 2801  715252.5 9559612 2199.28 24.4034184   0.03051    0.02190    264.6404
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077   3056.9475
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391   1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576  -0.00767   -0.03713   6379.7974
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450   2589.9910
#> 9606  714392.5 9559172 2353.05 30.7833035  -0.00185   -0.01185    953.8614
#> 10456 712442.5 9560292 2010.91 26.8917741   0.00445    0.00875    597.2893
#> 22847 714582.5 9560382 2040.49 25.3361937   0.00181   -0.00341   1363.4000
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.1803772   0.00124    0.01166   1813.7841
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732   4237.6934
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856   1789.1596
#> 10213 715542.5 9558782 2532.10 35.4133754  -0.00893   -0.00867   2247.8748
#> 14140 714672.5 9559262 2326.36 45.7925059  -0.01400   -0.01210    585.8992
#> 43834 713092.5 9560662 1905.55 31.3236027   0.00758   -0.00288   2144.8164
#> 9864  714952.5 9557592 2752.40 31.9893796  -0.00983    0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234    849.0547
#> 11332 712742.5 9560072 1850.34 22.7000149   0.00683   -0.00603   1562.0503
#> 39466 714942.5 9557772 2680.35 16.7532859   0.00166   -0.01265   1148.6487
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797   7557.7866
#> 23176 714012.5 9558892 2308.59 43.0291304  -0.01084   -0.02996   1599.3491
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150    503.5590
#> 30809 712532.5 9560032 1928.24 40.8994463   0.00570    0.02230    714.7432
#> 40459 712542.5 9559192 1992.60 39.8171290  -0.00162   -0.00517   1060.5494
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463   1050.7473
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538    265.4677
#> 8066  715842.5 9558732 2636.38 33.7174203  -0.00494    0.01064    683.2795
#> 441   714312.5 9560972 1916.59 31.3029762   0.00540    0.00650   1133.1501
#> 25140 713332.5 9560692 1809.95 52.2571887   0.00075   -0.01945    259.7421
#> 16542 713772.5 9559622 2304.85 30.8102325  -0.00712    0.00923    811.5580
#> 43243 715682.5 9558462 2693.23 27.4062902  -0.00031   -0.00199    836.4196
#> 29125 714992.5 9558792 2584.41 24.6658331   0.01253   -0.00333    422.8840
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.7842537   0.02327    0.01833    300.0125
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507    504.3085
#> 18627 714292.5 9559252 2330.78 18.4962235  -0.00538   -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333  -0.00007   -0.00853   1354.6001
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759   1880.4403
#> 43403 712632.5 9559262 2072.53 34.4863297  -0.00069   -0.00900    614.5898
#> 29944 713562.5 9558952 2209.55 17.3399947   0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011    712.5228
#> 14800 712952.5 9558682 2120.89 38.2735807  -0.00022    0.00662   1269.2650
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376    629.8236
#> 29527 714332.5 9557502 2567.02 48.9048126  -0.00452   -0.01298    436.1459
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098    164.0802
#> 33370 714202.5 9558962 2472.93 32.6414056  -0.00307    0.02997    230.8641
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.8507379  -0.00782   -0.00048   1550.6704
#> 29107 713172.5 9559272 2067.58 28.2640080   0.02413    0.00528    355.3475
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196    576.1450
#> 34345 715852.5 9557902 2874.08 43.6559462   0.00379   -0.01018    937.6911
#> 13908 715022.5 9559372 2290.81 27.2429972  -0.03004    0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654   1576.2065
#> 33410 715342.5 9560452 1942.88 29.0930780  -0.00305   -0.00705    968.9738
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599  12873.8018
#> 44823 714202.5 9557442 2548.13 43.2302386  -0.00233    0.01282    561.2366
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911    100.0000
#> 24819 714032.5 9560882 1981.86 33.2842642   0.00804   -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.1971887  -0.00873   -0.01417   1520.6630
#> 28286 713702.5 9559842 2245.07 27.9689348   0.00511    0.01499    617.3855
#> 34336 713792.5 9559072 2288.95 36.9328595  -0.00952   -0.00438   1276.5448
#> 2999  714592.5 9558492 2515.83 47.0318136   0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.0174405  -0.02552    0.00532  16771.3066
#> 40750 712962.5 9558702 2126.06 34.7040537  -0.00264    0.00624   1870.8307
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972    331.7462
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148    481.8292
#> 11394 713472.5 9561232 1889.48 34.3144424   0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871   0.00253   -0.01233    968.7780
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012    305.3624
#> 41873 714832.5 9559942 2253.59 10.5980003   0.01508   -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.0235474   0.00340    0.01809   1158.0426
#> 46639 715232.5 9560042 2107.15 39.6572738  -0.01305    0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.8014354   0.00992   -0.01052    451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678   0.00732    0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.2771580  -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.8140195   0.00282    0.00458    485.7418
#> 30403 713952.5 9560752 2003.57 42.9643862  -0.00933   -0.00157    766.0220
#> 13548 715812.5 9558122 2821.65 15.5082486   0.02751    0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.2293858   0.00366    0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.6210169  -0.02110   -0.02931   1812.3820
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057    722.3952
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.8520794   0.04456   -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 14254 15.6652391   300.00   TRUE       300.00             10  2.510336
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 28293 28.9022830   300.00   TRUE       300.00              4  2.967869
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 16811 35.2523743   300.00   TRUE       300.00             59  2.715521
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 48498 33.6182986   300.00   TRUE       300.00            100  3.098674
#> 2107  35.4236886   300.00   TRUE       300.00             45  2.661523
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 13585 47.0673370   300.00   TRUE       300.00            100  2.769655
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 35142 27.1530429   300.00   TRUE       300.00            100  2.176525
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 46171 29.0713056   300.00   TRUE       270.22            100  3.645964
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 2718  36.7328972   300.00   TRUE       300.00            100  2.990104
#> 42941 39.0631166   300.00   TRUE       300.00             17  4.882627
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 32879 29.7382284   300.00   TRUE       300.00              6  3.130450
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 15981 30.0035079   300.00   TRUE       300.00              0  2.799064
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 13259 23.1606730   203.85   TRUE        10.45             74  3.485288
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 10456 21.6686909    24.34   TRUE         0.00             18  2.776185
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 7771  16.5791068   300.00   TRUE        35.59            100  3.258586
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 12074 23.2855778   300.00   TRUE       300.00             63  3.252649
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 39466 27.2515916   300.00   TRUE       300.00            100  3.060187
#> 31253 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 8066  23.5273659   300.00  FALSE       300.00            100  2.834598
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 33175 34.7836948   300.00  FALSE       300.00            100  2.799219
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 14148 21.8778204   300.00  FALSE       214.75            100  2.760532
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 44823 36.1072273   300.00  FALSE       300.00            100  2.749146
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 34336 32.4511836   300.00  FALSE       300.00            100  3.106036
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 46639 33.5632947   300.00  FALSE       300.00            100  3.756965
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 11570 16.6071817   300.00  FALSE       300.00            100  2.686405
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 28978  4.4398500   300.00  FALSE        10.00            100  2.209674
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 15363 26.9421944   300.00  FALSE       300.00            100  2.858775
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603