Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.24719   -0.00501  -0.00649    1399.2329
#> 31357 714752.5 9561022 1848.22 33.44641   -0.00347   0.02357    1752.0375
#> 44302 714932.5 9557982 2649.56 37.30070    0.01633  -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.53351   -0.01021  -0.01579    1318.7560
#> 1300  714042.5 9558482 2408.27 24.14845    0.00659   0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.32800   -0.00661  -0.01339    1002.5877
#> 8598  714002.5 9558872 2310.02 39.58394   -0.00498  -0.00752    1321.5997
#> 21939 713642.5 9558712 2299.09 33.75065   -0.00503   0.02433    1475.3191
#> 9669  713852.5 9558612 2308.97 52.18041   -0.01059  -0.07431    4079.0366
#> 37104 713962.5 9561282 1795.56 25.74185    0.00273  -0.02403    1787.3024
#> 38827 713412.5 9560472 1869.10 16.87132   -0.00156   0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068   0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783     377.8918
#> 27375 714292.5 9558982 2458.96 43.60094    0.00214   0.00466     724.0713
#> 10899 713852.5 9559652 2282.82 28.81405    0.00952   0.01378     440.9104
#> 44328 713222.5 9558932 2227.94 29.15553   -0.00769  -0.00281    1325.6138
#> 21703 713842.5 9559672 2282.58 24.99185    0.01581   0.01189     287.5425
#> 22283 713592.5 9560762 1832.60 36.01040    0.00448  -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.63795   -0.00270  -0.00769    1032.7177
#> 38279 714192.5 9558612 2325.14 28.25255   -0.04430  -0.06070  138869.3906
#> 27978 713552.5 9558532 2373.05 17.04492    0.00832   0.00838     238.6685
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440   0.00710     843.7961
#> 15340 715452.5 9558852 2523.95 34.16547    0.00151   0.00899    1394.2274
#> 2942  713232.5 9560652 1851.75 34.53503    0.00856   0.01114    4510.8833
#> 4827  714962.5 9559882 2274.42 29.24491    0.00051   0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713  -0.00173    8967.1924
#> 32821 715022.5 9559952 2208.33 43.41645   -0.00643   0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.12575    0.00108  -0.00678    1891.6584
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669  -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.00910    0.01937   0.00083     994.7876
#> 6763  714692.5 9561052 1830.60 48.39488   -0.01996   0.00626    1014.7159
#> 17004 713102.5 9559032 2186.79 36.53351   -0.01383   0.00193    1809.3838
#> 27013 713062.5 9559782 1976.81 41.63226    0.01414   0.03416    1204.8629
#> 28577 712812.5 9559962 1849.82 59.78012    0.00646  -0.04086    6042.5757
#> 45104 715062.5 9559242 2345.02 24.15647   -0.01743   0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.59273    0.00804   0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.63433   -0.00669  -0.00991    1403.0193
#> 34359 712862.5 9558912 2111.37 32.81215    0.00463   0.00167    1938.5389
#> 6310  713802.5 9559752 2238.38 19.65131   -0.02379   0.01338   10787.6299
#> 5578  714962.5 9557612 2751.95 39.19146    0.00288   0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.48766   -0.00093  -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.65707   -0.01267  -0.05063   18733.8848
#> 13134 714652.5 9559632 2265.33 49.59981   -0.03552  -0.02058     431.4362
#> 37341 714412.5 9559142 2377.24 40.59979   -0.00735  -0.00925     665.6116
#> 44550 714842.5 9557642 2696.38 15.17307    0.00359  -0.00279     930.0871
#> 2976  715122.5 9559142 2390.05 33.16795    0.01008   0.00102     483.4184
#> 33240 714972.5 9557652 2741.04 39.43439   -0.00286  -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.68617   -0.04749   0.04779    2972.0193
#> 16015 712582.5 9560312 1899.18 56.54692    0.03480   0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.14313   -0.00759  -0.00400     608.6523
#> 5084  714912.5 9559302 2359.77 47.96803    0.00107   0.02373     527.3528
#> 40733 712862.5 9558892 2108.60 29.99491    0.00175   0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.95061   -0.00762   0.01252    5761.3486
#> 38862 712952.5 9559892 1950.86 37.72182   -0.00048  -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.64888   -0.01282   0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953  -0.02283     590.1536
#> 33350 713132.5 9560672 1896.89 24.69505    0.00306  -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.53484   -0.00529   0.02009    1247.7061
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209   0.02739     365.7626
#> 23213 715592.5 9558562 2654.69 33.21952   -0.00034   0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.04953    0.00126  -0.00485    2240.4185
#> 32280 712572.5 9560182 1956.86 50.67812   -0.02554   0.03304    1569.7313
#> 21013 712742.5 9560032 1845.53 29.35263    0.00568  -0.01558    1504.9717
#> 13146 713792.5 9558552 2341.40 29.82532   -0.00104   0.01664    1898.5588
#> 2552  714942.5 9557652 2719.32 34.87365   -0.01238  -0.00851    1012.5963
#> 42765 713552.5 9559632 2186.95 44.01118    0.00394  -0.00224     942.0322
#> 34192 712702.5 9560102 1867.01 35.47468    0.00629  -0.00469    1021.5407
#> 7385  713752.5 9560122 2064.61 31.77108   -0.02605  -0.03594   16587.9980
#> 12137 712842.5 9558902 2097.88 30.09862    0.00024  -0.00474    2343.2100
#> 6845  713412.5 9559172 2123.32 47.15385   -0.01556  -0.04394     587.8840
#> 1424  714012.5 9558902 2305.97 32.22372   -0.03986  -0.02524   10861.9482
#> 33910 715202.5 9559572 2195.66 13.79167   -0.04376  -0.01644 1057765.6250
#> 10933 712702.5 9560132 1854.55 36.29115   -0.00204  -0.00376    2048.3198
#> 18440 714862.5 9558932 2448.29 41.05243   -0.04989  -0.02781   29995.1543
#> 31117 712892.5 9558922 2129.84 30.55068    0.00521  -0.00870    1895.1399
#> 6217  714442.5 9559932 2151.30 36.67044    0.01489   0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.10642    0.00516   0.01574     183.3220
#> 30287 713992.5 9557822 2403.92 18.21547    0.03890  -0.00911     173.2806
#> 41125 712772.5 9559412 2021.88 38.87633   -0.00310  -0.00860     695.2302
#> 4588  713852.5 9558612 2308.97 52.18041   -0.01059  -0.07431    4079.0366
#> 20797 713012.5 9560392 1842.25 36.64982    0.01299   0.00141    4611.0718
#> 34615 713242.5 9560732 1879.60 27.59365    0.00384   0.00175    1026.6818
#> 22907 713172.5 9559612 2056.81 22.27259   -0.00166   0.01476    1404.4454
#> 5911  713952.5 9561282 1801.16 33.78961    0.00466  -0.01776    1021.8267
#> 26428 713042.5 9560332 1895.15 33.95978   -0.01439  -0.01161    3065.3694
#> 17058 714032.5 9558502 2402.95 31.35282    0.01924   0.02606     564.6876
#> 47298 714602.5 9560572 2025.50 28.59002   -0.00253   0.00753    1337.7854
#> 28173 714562.5 9560382 2048.26 27.68589   -0.00124  -0.00456    1531.4274
#> 19756 715572.5 9558972 2485.95 38.81273    0.00546   0.01484     495.0292
#> 2801  715252.5 9559612 2199.28 24.40342    0.03051   0.02190     264.6404
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281   0.04759     158.9370
#> 49378 713252.5 9560732 1875.41 27.83544    0.00181  -0.00001    1184.7952
#> 34984 715042.5 9559312 2320.49 42.85782   -0.01106  -0.04634     500.5027
#> 48574 712692.5 9560412 1901.97 41.97833    0.01364  -0.00094    1392.6008
#> 22847 714582.5 9560382 2040.49 25.33619    0.00181  -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.44550    0.02451  -0.03921     525.3245
#> 17227 714812.5 9558892 2545.43 54.07175    0.00458   0.04291     293.4231
#> 17674 713882.5 9558562 2339.20 41.03638   -0.03675  -0.01905    6487.0010
#> 10213 715542.5 9558782 2532.10 35.41338   -0.00893  -0.00867    2247.8748
#> 43834 713092.5 9560662 1905.55 31.32360    0.00758  -0.00288    2144.8164
#> 33142 715652.5 9558212 2740.94 14.38124   -0.00465  -0.01036    2505.8369
#> 34718 714802.5 9558802 2599.22 44.65003    0.01996   0.04684     209.1778
#> 6470  713742.5 9558802 2260.79 30.68017    0.00000   0.00320     911.5911
#> 4785  713792.5 9560472 2055.34 33.05279   -0.00586   0.00067    1123.7318
#> 10145 713782.5 9560932 1865.62 23.59555   -0.00203   0.00642    1688.1782
#> 7317  713682.5 9558582 2347.46 28.34766   -0.02196   0.00336    2806.2644
#> 36561 715292.5 9560802 1844.64 31.93724    0.00002  -0.01232     977.7320
#> 27884 714072.5 9559202 2446.27 19.75673   -0.00009   0.00109     605.1908
#> 30809 712532.5 9560032 1928.24 40.89945    0.00570   0.02230     714.7432
#> 24836 714202.5 9558272 2507.87 38.48901   -0.00226   0.00736     477.8591
#> 35852 715112.5 9559962 2173.61 39.31579   -0.00738   0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.34330   -0.00533   0.01493    2212.0525
#> 20067 714242.5 9560442 2125.23 38.64027    0.00363  -0.00363     389.2691
#> 441   714312.5 9560972 1916.59 31.30298    0.00540   0.00650    1133.1501
#> 29125 714992.5 9558792 2584.41 24.66583    0.01253  -0.00333     422.8840
#> 36126 715052.5 9558292 2759.47 29.06214    0.00196   0.01734     335.0773
#> 31884 715782.5 9558612 2632.60 30.27967    0.00084  -0.00083    1309.2347
#> 21925 714542.5 9559232 2314.33 50.24783    0.00737   0.01033    3446.5520
#> 41753 715722.5 9558192 2771.39 26.08505   -0.00196  -0.00175     725.1355
#> 42934 713832.5 9558672 2270.74 27.61198   -0.03193  -0.00267   31588.1953
#> 6143  712762.5 9559022 2096.24 29.89636    0.00358   0.02422     967.9738
#> 36245 713142.5 9560572 1827.82 43.67600   -0.00410   0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.31557    0.04152   0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.95631   -0.00301  -0.01099    1146.2140
#> 20135 714772.5 9560232 2107.52 37.85074   -0.00782  -0.00048    1550.6704
#> 34810 714112.5 9559572 2337.51 10.28803   -0.00393   0.00032     994.9224
#> 39440 714222.5 9557872 2508.52 39.04765    0.02785   0.00645     234.0114
#> 33410 715342.5 9560452 1942.88 29.09308   -0.00305  -0.00705     968.9738
#> 33666 713632.5 9561012 1775.88 21.17881   -0.00950  -0.05130     107.5831
#> 49831 715262.5 9557332 2884.30 22.29493   -0.00097   0.01487    1769.3392
#> 40909 712722.5 9560272 1827.15 29.29304    0.00009  -0.00599   12873.8018
#> 35826 712912.5 9560302 1891.22 50.86948   -0.00229   0.00099     909.0560
#> 34532 713812.5 9559502 2328.80 44.93765   -0.00070  -0.00060     383.3547
#> 48988 714682.5 9560852 1942.97 20.29073    0.00480   0.00740     909.4791
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199     107.2986
#> 26768 715092.5 9560582 1960.72 36.65612    0.00980  -0.00710     472.0236
#> 3868  714542.5 9560622 2024.15 29.44717    0.00360   0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.96760   -0.00820  -0.00260    1484.6593
#> 7986  712752.5 9560242 1857.41 39.83031    0.01852   0.00248    1432.4019
#> 18308 714842.5 9558742 2568.20 26.01744   -0.02552   0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.00410    0.00745   0.00875    2156.9358
#> 48050 714022.5 9560572 2099.82 18.71395    0.01182  -0.00972     331.7462
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909  -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.56749   -0.03284  -0.00886   43249.3945
#> 16443 714472.5 9559532 2222.55 16.74412    0.03263  -0.02012     305.3624
#> 40653 714262.5 9561262 1791.94 12.22234    0.00161   0.00709    1181.2555
#> 9320  713872.5 9559262 2378.76 29.85855    0.00309   0.01011     501.7867
#> 23808 715172.5 9558702 2533.50 22.34593    0.00811  -0.02281     492.4785
#> 26635 715092.5 9560412 2010.59 30.57188   -0.00463   0.00253    1237.0536
#> 48157 713142.5 9560992 1978.48 44.62138    0.00208   0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 31357 23.8172826   158.92   TRUE         0.00              5  3.243543
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 24051 32.7388084   300.00   TRUE       300.00             11  3.001122
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 9669  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 2942  29.8207344    68.13   TRUE       116.32             10  3.654262
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 28577 34.3912187   300.00   TRUE         0.56              0  3.781222
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37341 31.4519452   300.00   TRUE       300.00              1  2.823221
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 33240 29.3549197   300.00   TRUE       300.00            100  2.756740
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 33350 28.9527033    10.00   TRUE       166.26              2  3.415431
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 28845 29.8425068   300.00   TRUE       300.00             54  2.563199
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 21013 24.9631982   300.00   TRUE        13.47             64  3.177528
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 10933 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 16071 29.6213451   300.00   TRUE       300.00            100  2.201225
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 17674 30.4825006   300.00   TRUE       300.00              1  3.812044
#> 10213 31.5986224   300.00   TRUE       300.00            100  3.351772
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 33142 26.9060981   300.00   TRUE       300.00            100  3.398953
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 7317  27.2258085   300.00  FALSE       300.00             61  3.448129
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 35852 36.3621935   300.00  FALSE       300.00            100  3.356690
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 41753 25.0073159   300.00  FALSE       300.00            100  2.860419
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34105 20.0770141   300.00  FALSE       300.00             26  3.059266
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 35826 44.2277581   298.23  FALSE        16.35             96  2.958591
#> 34532 31.2067193   300.00  FALSE       300.00             92  2.583601
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 7986  32.0031306   275.06  FALSE         0.00             65  3.156065
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 9320  18.1392708   300.00  FALSE       300.00             93  2.700519
#> 23808 37.5338922   300.00  FALSE       300.00             72  2.692387
#> 26635 31.3946494   300.00  FALSE       300.00            100  3.092389
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769