Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv      vcurv       carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106   -0.04634    500.5027
#> 23178 715382.5 9560142 2021.07 41.976607  0.00958    0.02642    671.1807
#> 31357 714752.5 9561022 1848.22 33.446411 -0.00347    0.02357   1752.0375
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023   -0.00043   1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075    0.00475   1001.0861
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479    0.01261    323.8441
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514   -0.00644   1179.0695
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118    0.00761    928.6858
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645    0.00835   2425.3115
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951   -0.00529   3282.1186
#> 7349  714032.5 9558502 2402.95 31.352824  0.01924    0.02606    564.6876
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894   -0.03416    947.4878
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214    0.00466    724.0713
#> 10899 713852.5 9559652 2282.82 28.814048  0.00952    0.01378    440.9104
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979   -0.00269    491.8017
#> 33792 714652.5 9558172 2598.36 59.458695 -0.03525   -0.11024    394.8751
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696    0.02514    297.7842
#> 48498 714972.5 9557972 2681.66 38.346919  0.00528    0.01392   1255.0884
#> 2107  714142.5 9558562 2382.67 35.553750  0.01412    0.00898    458.6938
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602    0.01518    854.4269
#> 23375 713162.5 9559632 2041.52 46.815236 -0.00857    0.03677   1675.4679
#> 20290 714812.5 9561092 1775.73 34.626704 -0.01537   -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.044921  0.00832    0.00838    238.6685
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043    0.00973    514.5488
#> 47022 715362.5 9560102 2059.29 49.119672  0.02059   -0.00628    556.0121
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713   -0.00173   8967.1924
#> 36230 713322.5 9560662 1798.99 40.128245 -0.00033   -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356    0.00774    527.1068
#> 20194 714132.5 9558452 2447.22 24.022401 -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620 -0.00664    0.00694   2123.4473
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205    0.02205    629.4165
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097    0.00156    816.6036
#> 17004 713102.5 9559032 2186.79 36.533508 -0.01383    0.00193   1809.3838
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414    0.03416   1204.8629
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646   -0.04086   6042.5757
#> 15990 712852.5 9559982 1894.77 37.694320 -0.01545    0.02405   5681.2412
#> 2858  714292.5 9559332 2335.04 42.355905  0.04547    0.03323    183.1836
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463    0.00167   1938.5389
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066   -0.00036    516.4709
#> 46171 712822.5 9558932 2089.30 29.854539 -0.01174    0.00025   4425.5151
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006    0.00175    804.5679
#> 28743 715202.5 9558872 2463.87 34.657071 -0.01267   -0.05063  18733.8848
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617    0.00307 147470.5312
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492    0.00072  54831.6289
#> 41037 715532.5 9558792 2529.58 32.846524 -0.00585   -0.00974   2136.6939
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290    0.00230   1056.5876
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759   -0.00400    608.6523
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107    0.02373    527.3528
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762    0.01252   5761.3486
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048   -0.00332   1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282    0.00542    860.8840
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576   -0.00166    650.6246
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027   -0.01473   2930.2791
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548    0.05543    352.1558
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195   -0.00656   3036.7383
#> 31121 713462.5 9559132 2135.56 61.596974 -0.03800   -0.06020   1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628  0.00568   -0.01558   1504.9717
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904   -0.02013   4229.5708
#> 17431 712802.5 9560022 1832.71 41.525116 -0.00478   -0.07132  42970.7891
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394   -0.00224    942.0322
#> 21680 712802.5 9560462 1893.75 38.734239 -0.00481   -0.00399   3900.3445
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538    0.00112    745.3315
#> 18897 712802.5 9559882 1851.71 50.234648 -0.02627   -0.05393   8301.6094
#> 40216 714602.5 9560592 2015.55 34.368873 -0.00093    0.00664   1807.2659
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743   -0.05242    507.9263
#> 12772 715172.5 9557642 2759.42 34.805467 -0.00567   -0.03793    864.9472
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169   -0.03220   8936.2666
#> 6845  713412.5 9559172 2123.32 47.153854 -0.01556   -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370   -0.00190    640.8842
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025   -0.01245   6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509   -0.01211  49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489    0.00241   1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516    0.01574    183.3220
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038   -0.04212  46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084   -0.02236   9755.3027
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070   -0.02661   2508.5374
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166    0.01476   1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542   -0.00722   1373.4147
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011    0.00739    789.7573
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213   -0.01363    565.5207
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124   -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.906461 -0.00070    0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.829930 -0.03669   -0.01981   3975.3608
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042    0.00957    495.4453
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153   -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507  0.00010    0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364   -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881    0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246  0.01257    0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.106682  0.00915   -0.02705    413.5702
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445    0.00875    597.2893
#> 46792 713862.5 9559672 2272.03 35.464942  0.00560    0.00610    454.8103
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622    0.00152   1530.5657
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360   -0.03000 162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176 -0.00212    0.02732   4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118   -0.00252   3631.7017
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004   -0.02384    947.2075
#> 46971 713532.5 9559952 2198.03 29.723904 -0.00529    0.01230   1044.6443
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758   -0.00288   2144.8164
#> 30378 714852.5 9557752 2651.70 41.298225 -0.00010   -0.00870   2099.7930
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585   -0.02665   1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989   -0.02781  29995.1543
#> 7317  713682.5 9558582 2347.46 28.347660 -0.02196    0.00336   2806.2644
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002   -0.01232    977.7320
#> 25584 715062.5 9558532 2675.93 37.688591  0.00534    0.01476   1275.9446
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570    0.02230    714.7432
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007   -0.00463   1050.7473
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815   -0.03345 665381.5625
#> 32869 715332.5 9558892 2533.17 20.635075 -0.00242    0.01132    737.5982
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220   -0.03290   2481.5632
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345   -0.00075   8141.7549
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161    0.00171    756.8818
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937   -0.02273  72123.6172
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987   -0.00507    504.3085
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174   -0.03065    704.4843
#> 12974 714662.5 9557332 2667.61 41.941657 -0.00065    0.00475    515.1661
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236    0.00483   1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601   -0.00011    712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180    0.01170    253.5129
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423   -0.00983    270.7949
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687   -0.00223   1056.3701
#> 48869 712562.5 9560122 1972.89 24.946582  0.00185    0.01655    560.8554
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586    0.00654   1576.2065
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462    0.00268    252.0072
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342    0.02552   2515.0752
#> 46509 712512.5 9559532 1860.09 89.988751 -0.00270 1018.51947    842.1514
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858    0.00312    254.0656
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360    0.00750   1892.8978
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511    0.01499    617.3855
#> 34336 713792.5 9559072 2288.95 36.932859 -0.00952   -0.00438   1276.5448
#> 40750 712962.5 9558702 2126.06 34.704054 -0.00264    0.00624   1870.8307
#> 15138 713702.5 9561122 1780.68 49.221658  0.00607   -0.04566  12862.4922
#> 19767 714272.5 9558052 2431.67 55.724729  0.00361   -0.01841    756.0055
#> 4269  714162.5 9560572 2087.45  7.334433  0.00293    0.00186   2751.4626
#> 11394 713472.5 9561232 1889.48 34.314442  0.00108   -0.00948   1538.1741
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203    0.00238    787.0053
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263   -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.534255 -0.00163   -0.00377   1727.6055
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986   -0.05013    616.4128
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933   -0.00157    766.0220
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087   -0.00573   1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716   -0.03234   1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962   -0.00172    351.5812
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241    0.00020    673.2861
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 28293 28.90228302   300.00   TRUE       300.00              4  2.967869
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 43007 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 2858  35.31941032   300.00   TRUE       300.00              2  2.262887
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 18751 38.07591028   300.00   TRUE       300.00            100  2.713046
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 28743 34.65306041   300.00   TRUE       300.00             71  4.272628
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 21680 28.34823283   104.12   TRUE       110.76              4  3.591103
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 8136  25.42901286   115.18   TRUE        84.59             30  3.951156
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 14545 39.31407207   300.00   TRUE       300.00             37  2.897494
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 46038 29.73077999   300.00   TRUE       243.86            100  3.184852
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 28286 17.04835919   300.00  FALSE       300.00            100  2.790556
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 4269  19.11673683   300.00  FALSE       194.53            100  3.439564
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 42990 22.23706499   205.00  FALSE         0.00            100  3.173195
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200