Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 25090 715362.5 9560102 2059.29 49.119672  0.02059 -0.00628     556.0121
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475    1001.0861
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 24072 715282.5 9557602 2837.46 34.394083 -0.02191 -0.00579    2246.7725
#> 34512 713162.5 9559632 2041.52 46.815236 -0.00857  0.03677    1675.4679
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 39885 715062.5 9561022 1840.94 38.435728  0.00250 -0.01340     604.1032
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529    3282.1186
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 22814 713862.5 9558582 2327.48 48.612031 -0.02894 -0.03416     947.4878
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 42598 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518 -0.02123  0.01164     895.6606
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769    1032.7177
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032     994.6740
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38032 714762.5 9560442 1988.33 35.169868 -0.00905 -0.01385     674.0721
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 40828 714842.5 9557632 2698.65 17.536519  0.00490 -0.00450     873.0448
#> 13585 714552.5 9560092 2123.69 57.528973 -0.03613 -0.01898     588.3761
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 4329  712882.5 9559982 1911.46 26.231727 -0.00193 -0.00077    3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125  0.00646 -0.04086    6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 48434 713352.5 9559042 2176.63 35.702146 -0.01771 -0.01939     703.7817
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 2718  715062.5 9558222 2719.56 44.753160 -0.00119 -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565    2232.6133
#> 48808 715032.5 9557792 2707.02 25.437607 -0.02329 -0.01181   11655.4316
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 29801 714392.5 9559162 2356.18 31.562526 -0.00856 -0.01625    1063.9150
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 24501 713302.5 9559142 2125.66 27.108925  0.01784  0.00596     842.0994
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 37769 714412.5 9558562 2382.08 39.876717 -0.04601 -0.08389   47022.9688
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 46461 712682.5 9560032 1892.30 22.977899  0.01453  0.01337     486.7551
#> 19892 714482.5 9560562 2059.63 19.815172 -0.00332 -0.00058    1817.2371
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 43201 715712.5 9557922 2852.82 31.462258 -0.00949  0.00520    3390.5708
#> 4588  713852.5 9558612 2308.97 52.180412 -0.01059 -0.07431    4079.0366
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 10689 713542.5 9559132 2211.95 31.981358  0.05379  0.00801     172.7491
#> 26101 715562.5 9558682 2594.87 32.411077  0.00079  0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 11352 713002.5 9560562 1858.95 39.803951 -0.00442 -0.00118   19557.6621
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109    2047.9031
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 23626 713742.5 9560032 2137.92 39.304905 -0.00605  0.00075    2170.2249
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.397535  0.00143  0.00477    2945.2100
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 18498 714842.5 9557712 2679.04 28.939525  0.00902  0.01358     802.7622
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 41408 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476    1404.4454
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 31253 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996    1599.3491
#> 28253 714102.5 9559892 2288.88  2.957608  0.00378  0.02382     100.0000
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 748   715102.5 9559692 2301.39 27.021836 -0.00759  0.00158     837.3730
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493    2212.0525
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 49363 712772.5 9559552 1935.91 29.036928 -0.03220 -0.03290    2481.5632
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 42329 715852.5 9558702 2651.92 22.870756  0.00321  0.00548     532.8210
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780     825.4451
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 33410 715342.5 9560452 1942.88 29.093078 -0.00305 -0.00705     968.9738
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 44823 714202.5 9557442 2548.13 43.230239 -0.00233  0.01282     561.2366
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 44434 713022.5 9558762 2161.07 31.454237 -0.00273 -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 14763 713462.5 9560852 1830.40 47.739989  0.01354  0.00907    1281.5044
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576    1045.1074
#> 6732  715482.5 9559232 2376.03 29.905532 -0.01137 -0.01523     888.2811
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 16845 712772.5 9560022 1831.76 21.399974 -0.00865 -0.00805    7489.7417
#> 34334 714642.5 9557912 2582.25 30.991287  0.00253 -0.01233     968.7780
#> 40653 714262.5 9561262 1791.94 12.222336  0.00161  0.00709    1181.2555
#> 22628 714942.5 9557732 2696.19 26.211673  0.00788 -0.01029     894.2672
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 46639 715232.5 9560042 2107.15 39.657274 -0.01305  0.00585    5714.3267
#> 49812 713392.5 9559222 2133.99 35.662612 -0.00983  0.00803    1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 30403 713952.5 9560752 2003.57 42.964386 -0.00933 -0.00157     766.0220
#> 126   715042.5 9561262 1733.49 30.110651 -0.00288  0.00448     941.8503
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 28978 713982.5 9560462 2133.10  5.229386  0.00366  0.00525     162.0592
#> 40043 713322.5 9560562 1802.09 39.852079  0.04456 -0.01586     941.0288
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 22814 34.91203733   300.00   TRUE       300.00              6  2.976574
#> 10974 32.46550755   300.00   TRUE       300.00            100  4.101499
#> 42598 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 3835  35.13892862   300.00   TRUE       300.00             39  2.691790
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 48055 33.82570935   300.00   TRUE         1.67              0  3.566570
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 38279 37.62843024   300.00   TRUE       300.00             24  5.142607
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 25081 34.87536803   300.00   TRUE       300.00            100  3.617357
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 13585 47.06733695   300.00   TRUE       300.00            100  2.769655
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 5578  25.74184782   300.00   TRUE       300.00            100  2.640875
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 49569 41.84941031    25.00   TRUE        35.00            100  3.348814
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 27695 27.36790204   300.00   TRUE       300.00             41  3.042574
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 24501 27.15246991   300.00   TRUE       300.00              0  2.925363
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39722 26.54169690   300.00   TRUE       300.00            100  2.906387
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 7     13.22042816   300.00   TRUE       300.00              0  2.263215
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34887 42.93573829   300.00   TRUE       300.00            100  3.159244
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17227 43.59750455   300.00   TRUE       300.00            100  2.467494
#> 24532 32.54629459   300.00   TRUE         7.49            100  6.276647
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 3771   0.32945073    44.23   TRUE         1.67             94  2.117095
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 30809 36.36620421   181.84  FALSE         0.00            100  2.854150
#> 748   38.13263310   300.00  FALSE       300.00             12  2.922919
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 11111 33.38395889   300.00  FALSE       300.00            100  3.344795
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 47059 25.75960951   300.00  FALSE       300.00            100  3.023816
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 21070 34.67712463   300.00  FALSE         0.00             10  3.400551
#> 44823 36.10722729   300.00  FALSE       300.00            100  2.749146
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491