Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649    1399.2329
#> 37912 712802.5 9559952 1838.40 52.101344    0.00183   -0.09203     634.3320
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633   -0.01813    1130.5791
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661   -0.01339    1002.5877
#> 8059  713292.5 9559122 2136.06 29.662025    0.02240   -0.00250    1413.3312
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752    1321.5997
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433    1475.3191
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858   -0.00292    1233.9299
#> 44952 715352.5 9557472 2873.35 38.952536   -0.01073    0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.435728    0.00250   -0.01340     604.1032
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387    1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770     474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407    0.00547     519.4228
#> 23817 714842.5 9557832 2677.58 26.657434    0.00922    0.01088     266.4355
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269     491.8017
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234     849.0547
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514     297.7842
#> 48498 714972.5 9557972 2681.66 38.346919    0.00528    0.01392    1255.0884
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129    0.02510    3342.5518
#> 29420 713592.5 9560772 1826.82 34.926934   -0.00070   -0.00530    9000.4356
#> 38279 714192.5 9558612 2325.14 28.252549   -0.04430   -0.06070  138869.3906
#> 34482 715152.5 9557642 2753.52  4.995619   -0.01126   -0.02044  129412.3984
#> 38032 714762.5 9560442 1988.33 35.169868   -0.00905   -0.01385     674.0721
#> 27978 713552.5 9558532 2373.05 17.044921    0.00832    0.00838     238.6685
#> 2942  713232.5 9560652 1851.75 34.535031    0.00856    0.01114    4510.8833
#> 47210 714952.5 9557612 2744.35 40.884549   -0.00493    0.00733     598.7983
#> 8288  714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490   -0.00450     873.0448
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629  360379.1875
#> 13585 714552.5 9560092 2123.69 57.528973   -0.03613   -0.01898     588.3761
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521   -0.00739     558.2286
#> 5749  712852.5 9560162 1934.78 23.388710    0.00071    0.00150    1274.3594
#> 3133  715042.5 9559252 2344.13 29.020885    0.00422    0.02357     275.7474
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276    9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077    3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 45104 715062.5 9559242 2345.02 24.156474   -0.01743    0.01313     783.4548
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323     183.1836
#> 34563 715022.5 9558162 2675.70 18.889846   -0.01993    0.01373   16031.7500
#> 46201 714832.5 9561022 1845.94 34.573992   -0.00856    0.01796    4729.6382
#> 43092 715532.5 9558192 2770.31 36.833165   -0.00442   -0.00178     800.1933
#> 47091 713382.5 9559182 2122.71 48.037927    0.04527    0.04613     289.6113
#> 28743 715202.5 9558872 2463.87 34.657071   -0.01267   -0.05063   18733.8848
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294     739.5768
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749    0.04779    2972.0193
#> 421   714892.5 9561022 1849.01 26.132032    0.00002    0.00588     749.2502
#> 49569 713232.5 9561122 1868.73 37.312158   -0.00405   -0.01565    2232.6133
#> 38862 712952.5 9559892 1950.86 37.721822   -0.00048   -0.00332    1418.6622
#> 39520 715342.5 9557622 2889.48 41.648875   -0.01282    0.00542     860.8840
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292    0.03658    0.04602     345.8933
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.010397    0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.019928   -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941    1290.7709
#> 664   714452.5 9559852 2174.83 44.860876    0.00282    0.00109     767.4097
#> 16355 712962.5 9560212 1992.88 39.526066    0.01927   -0.00316     454.9674
#> 27695 714142.5 9558452 2449.28 20.490117   -0.00539    0.00809    1102.9956
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141   -0.00410    0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101    0.00645   -0.00395    2011.3270
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013    4229.5708
#> 45592 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132   42970.7891
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165    0.00555    4439.0967
#> 21680 712802.5 9560462 1893.75 38.734239   -0.00481   -0.00399    3900.3445
#> 41721 715182.5 9557562 2784.85 31.789990   -0.00588    0.01448    1258.8196
#> 37714 713192.5 9560692 1885.05 18.034420    0.00188    0.00832    3022.3462
#> 43859 714892.5 9557602 2722.70 28.129363    0.00370   -0.00190     640.8842
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 18348 714042.5 9558882 2341.41 44.063892   -0.00338    0.00728    1137.2450
#> 43201 715712.5 9557922 2852.82 31.462258   -0.00949    0.00520    3390.5708
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195    1901.6743
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466   -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833     369.3771
#> 26428 713042.5 9560332 1895.15 33.959781   -0.01439   -0.01161    3065.3694
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321    1170.5904
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.603786   -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.958859    0.03169    0.01441     424.2680
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190     264.6404
#> 22301 713722.5 9560712 1913.37 39.271100    0.00705   -0.00915    2036.0378
#> 20171 715732.5 9557962 2815.53 48.657868   -0.01341    0.00171    4425.5410
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881    0.00121    1783.0658
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915   -0.02705     413.5702
#> 17825 712732.5 9560462 1915.51 40.397535    0.00143    0.00477    2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560    0.00610     454.8103
#> 49459 714692.5 9557342 2639.74 42.047081    0.00216   -0.02606     919.4241
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124    0.01166    1813.7841
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732    4237.6934
#> 18498 714842.5 9557712 2679.04 28.939525    0.00902    0.01358     802.7622
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856    1789.1596
#> 24532 715202.5 9560922 1799.62  4.488551   -0.02312   -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601    0.00115    0.01015     790.2421
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983    0.01253     528.1924
#> 33142 715652.5 9558212 2740.94 14.381241   -0.00465   -0.01036    2505.8369
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195   -0.00205    1112.2626
#> 31253 712612.5 9559492 1903.67 46.756794   -0.06953   -0.01797    7557.7866
#> 36744 714922.5 9557722 2692.39 25.980771    0.00316   -0.00427     866.9398
#> 4785  713792.5 9560472 2055.34 33.052789   -0.00586    0.00067    1123.7318
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559     866.5775
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328    1066.7134
#> 26023 715832.5 9557632 3097.56 41.710182    0.00133   -0.00452     578.5847
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363     389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815   -0.03345  665381.5625
#> 37832 712552.5 9559192 1998.10 41.550899   -0.00195   -0.00995    1026.7981
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712    0.00923     811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028    2394.7261
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943     371.5514
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216    1412.2295
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084   -0.00083    1309.2347
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035    1471.9783
#> 33956 714082.5 9558712 2327.95 39.684203    0.02218    0.00902     218.7467
#> 36245 713142.5 9560572 1827.82 43.676000   -0.00410    0.00400    4126.7812
#> 34345 715852.5 9557902 2874.08 43.655946    0.00379   -0.01018     937.6911
#> 13908 715022.5 9559372 2290.81 27.242997   -0.03004    0.00714  480535.5000
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586    0.00654    1576.2065
#> 45486 712762.5 9560172 1853.82 56.797306    0.01978    0.01282    2322.7944
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 42424 714692.5 9560902 1921.34 25.331037    0.01154   -0.00204     618.2027
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105     868.7802
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438    1276.5448
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852    0.00248    1432.4019
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757   -0.00003 5426606.0000
#> 4943  712452.5 9559172 1927.46 34.465130    0.00105   -0.01576    1045.1074
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452    0.00368    1071.8734
#> 20950 714912.5 9557782 2667.71 30.811378   -0.04770    0.00270   35950.5508
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196   -0.01935     903.6397
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177     184.1042
#> 28526 715152.5 9560732 1871.53 29.534255   -0.00163   -0.00377    1727.6055
#> 39883 712732.5 9560842 2082.69 25.090968    0.00732    0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.836044   -0.00181   -0.00799    1036.6606
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298    0.00222     991.8300
#> 42990 712432.5 9560692 2102.11 32.450038   -0.00087   -0.00573    1490.0312
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234    1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860   -0.00011   -0.00419     323.0790
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253    1237.0536
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855     229.6061
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 37912 30.294570   300.00   TRUE         9.15              2  2.802317
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 24051 32.738808   300.00   TRUE       300.00             11  3.001122
#> 8059  26.977145   300.00   TRUE       300.00              4  3.150244
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 15843 42.221833   300.00   TRUE       300.00            100  3.091290
#> 44952 34.082394   300.00   TRUE       300.00            100  3.363704
#> 39885  6.084812   210.57   TRUE         0.00            100  2.781111
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 23817 21.611395   300.00   TRUE       300.00              4  2.425592
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 48498 33.618299   300.00   TRUE       300.00            100  3.098674
#> 22283 19.942369   256.35   TRUE         0.00            100  2.772566
#> 40307 29.513629    77.82   TRUE        76.41             54  3.524078
#> 29420 28.852436   253.02   TRUE         0.00            100  3.954264
#> 38279 37.628430   300.00   TRUE       300.00             24  5.142607
#> 34482 32.971429   300.00   TRUE       300.00            100  5.111976
#> 38032 35.008294   300.00   TRUE       300.00             60  2.828706
#> 27978 14.328529   300.00   TRUE       300.00             12  2.377795
#> 2942  29.820734    68.13   TRUE       116.32             10  3.654262
#> 47210 26.915265   300.00   TRUE       300.00            100  2.777281
#> 8288  22.717777   300.00   TRUE       300.00             83  2.904587
#> 40828 25.599754   300.00   TRUE       300.00            100  2.941037
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 13585 47.067337   300.00   TRUE       300.00            100  2.769655
#> 4477  32.169861   300.00   TRUE       300.00             11  2.746812
#> 5749  28.233068   300.00   TRUE         0.00            100  3.105292
#> 3133  18.834269   300.00   TRUE       300.00            100  2.440511
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 45104 17.992594   300.00   TRUE       300.00            100  2.894014
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 34563 38.156697   300.00   TRUE       300.00            100  4.204981
#> 46201 22.183780   167.75   TRUE         0.00             55  3.674828
#> 43092 39.478511   300.00   TRUE       300.00            100  2.903195
#> 47091 40.151736   300.00   TRUE       300.00             25  2.461815
#> 28743 34.653060   300.00   TRUE       300.00             71  4.272628
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 34707 27.702509   165.13   TRUE        39.56            100  3.473052
#> 421   10.713738   182.02   TRUE         0.00             83  2.874627
#> 49569 41.849410    25.00   TRUE        35.00            100  3.348814
#> 38862 34.900005   300.00   TRUE       104.48             92  3.151879
#> 39520 35.894087   300.00   TRUE       300.00            100  2.934945
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 34005 41.733673   300.00   TRUE       300.00            100  2.538942
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 46783 19.942369   256.35   TRUE         0.00            100  2.772566
#> 20065 30.389681   300.00   TRUE       300.00              5  3.243142
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 664   30.842318   300.00   TRUE       300.00             12  2.885027
#> 16355 28.246819   300.00   TRUE        58.52            100  2.657980
#> 27695 27.367902   300.00   TRUE       300.00             41  3.042574
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 11618 33.985565   300.00   TRUE       300.00              1  2.578737
#> 17728 28.552206    30.14   TRUE       140.15              2  3.303483
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 45592 15.648623   300.00   TRUE       300.00              2  2.460907
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 26010 25.229624    69.52   TRUE       166.56              2  3.647295
#> 21680 28.348233   104.12   TRUE       110.76              4  3.591103
#> 41721 29.979444   300.00   TRUE       300.00            100  3.099963
#> 37714 30.731737    13.57   TRUE       170.00              4  3.480344
#> 43859 29.557747   300.00   TRUE       300.00            100  2.806780
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 18348 41.010600   300.00   TRUE       300.00            100  3.055854
#> 43201 41.828211   300.00   TRUE       300.00            100  3.530273
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26428 35.475255   300.00   TRUE         0.00            100  3.486483
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 28009 27.017825   300.00   TRUE       300.00             29  3.185767
#> 7485  39.490543   300.00   TRUE       300.00            100  2.627640
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22301 33.615434   300.00   TRUE         0.00            100  3.308786
#> 20171 42.130160   300.00   TRUE       300.00            100  3.645966
#> 1223  40.458269   300.00   TRUE       300.00              4  3.251167
#> 14950 41.874047   300.00   TRUE       300.00             15  2.616549
#> 17825 30.733456    73.66   TRUE       137.63             18  3.469116
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46792 27.601096   300.00   TRUE       300.00             18  2.657830
#> 49459 41.856859   300.00   TRUE       300.00            100  2.963516
#> 7771  16.579107   300.00   TRUE        35.59            100  3.258586
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 18498 22.717777   300.00   TRUE       300.00             83  2.904587
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 24532 32.546295   300.00   TRUE         7.49            100  6.276647
#> 14779 16.521238    55.08   TRUE         0.00              1  2.897760
#> 9864  20.820713   300.00   TRUE       300.00            100  2.722792
#> 33142 26.906098   300.00   TRUE       300.00            100  3.398953
#> 41310 17.295304   300.00   TRUE       300.00             25  3.046207
#> 31253 38.668349   300.00   TRUE       122.87            100  3.878395
#> 36744 29.404767   300.00   TRUE       300.00            100  2.937989
#> 4785  26.486120   300.00  FALSE         0.00            100  3.050663
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 26023 45.072871   300.00  FALSE       300.00            100  2.762367
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 7517  22.984775   144.37  FALSE        60.16            100  5.823071
#> 37832 36.446418   300.00  FALSE       207.18              2  3.011485
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 31884 24.137566   300.00  FALSE       300.00            100  3.117018
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 33956 38.595583   300.00  FALSE       300.00            100  2.339942
#> 36245 25.853575   110.19  FALSE        81.50             20  3.615611
#> 34345 44.395062   300.00  FALSE       300.00            100  2.972060
#> 13908 34.377468   300.00  FALSE       300.00             96  5.681725
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 45486 33.421774   300.00  FALSE         0.00             68  3.366011
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 42424 22.607196   275.21  FALSE        93.50             70  2.791131
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 7986  32.003131   275.06  FALSE         0.00             65  3.156065
#> 42638 23.421369    86.41  FALSE       172.46            100  6.734528
#> 4943  37.771670   300.00  FALSE       300.00            100  3.019161
#> 7132  34.049163   300.00  FALSE       225.26             12  3.030143
#> 20950 32.711879   300.00  FALSE       300.00             81  4.555706
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 28526 31.814627   300.00  FALSE        46.28            100  3.237445
#> 39883 17.044348   233.81  FALSE       219.23            100  2.474870
#> 46148 38.910710   300.00  FALSE       300.00            100  3.015637
#> 24962 29.050106   300.00  FALSE       300.00            100  2.996437
#> 42990 22.237065   205.00  FALSE         0.00            100  3.173195
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 47245 34.976209   300.00  FALSE       150.03            100  2.509309
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983