Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.56423   -0.00681  -0.00029   5577.3916
#> 16435 713512.5 9559092 2166.13 55.97397    0.02056  -0.06976    301.2347
#> 15762 714922.5 9558862 2535.97 54.91056    0.01305  -0.00555    430.4604
#> 30059 713832.5 9559662 2288.02 18.27793    0.01917   0.00333    258.6537
#> 46178 714892.5 9559312 2379.54 32.76688   -0.00266   0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.66203    0.02240  -0.00250   1413.3312
#> 11088 715532.5 9558762 2549.28 38.05013   -0.00373  -0.00057   1630.3259
#> 24072 715282.5 9557602 2837.46 34.39408   -0.02191  -0.00579   2246.7725
#> 34512 713162.5 9559632 2041.52 46.81524   -0.00857   0.03677   1675.4679
#> 21939 713642.5 9558712 2299.09 33.75065   -0.00503   0.02433   1475.3191
#> 20799 714792.5 9561002 1857.29 23.12916    0.00090   0.00890   2301.0454
#> 26531 714462.5 9560452 2104.74 19.52411    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.40193    0.00310   0.01440   1088.8639
#> 43327 713962.5 9561082 1839.11 63.36168   -0.04951  -0.00529   3282.1186
#> 49372 714862.5 9560982 1863.12 20.96911   -0.00295   0.00035   2697.5974
#> 125   713992.5 9561062 1878.28 49.97109    0.00183   0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.65670    0.00730   0.01770    474.5059
#> 23817 714842.5 9557832 2677.58 26.65743    0.00922   0.01088    266.4355
#> 31023 714712.5 9561042 1838.66 44.12004   -0.00223   0.01783    377.8918
#> 23925 715202.5 9557652 2782.99 41.95713   -0.00542   0.01842    868.6137
#> 25690 715042.5 9557572 2752.67 32.71360    0.00194   0.00526    539.8119
#> 10899 713852.5 9559652 2282.82 28.81405    0.00952   0.01378    440.9104
#> 21703 713842.5 9559672 2282.58 24.99185    0.01581   0.01189    287.5425
#> 43964 714802.5 9558452 2684.33 37.98252   -0.02123   0.01164    895.6606
#> 40307 713302.5 9560702 1845.07 39.00467    0.00129   0.02510   3342.5518
#> 18304 714582.5 9558932 2463.25 22.19467   -0.02131   0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.10784   -0.00017  -0.00032    994.6740
#> 29472 714792.5 9561072 1786.08 36.17312   -0.01029  -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.12347   -0.00746   0.00096   3210.5781
#> 23375 713162.5 9559632 2041.52 46.81524   -0.00857   0.03677   1675.4679
#> 22566 712832.5 9559842 1897.49 33.06024    0.00672   0.04398    267.2659
#> 14522 715012.5 9557732 2721.34 35.18133   -0.00382   0.00672   1114.8801
#> 20290 714812.5 9561092 1775.73 34.62670   -0.01537  -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.04492    0.00832   0.00838    238.6685
#> 36030 712852.5 9559572 1938.79 28.67883   -0.01440   0.00710    843.7961
#> 15340 715452.5 9558852 2523.95 34.16547    0.00151   0.00899   1394.2274
#> 35942 712862.5 9558902 2109.97 31.40038    0.00151   0.00189   2317.1819
#> 26215 714332.5 9560262 2138.08 25.72122   -0.00491   0.00581   1855.1630
#> 40828 714842.5 9557632 2698.65 17.53652    0.00490  -0.00450    873.0448
#> 38077 714952.5 9557612 2744.35 40.88455   -0.00493   0.00733    598.7983
#> 3466  713232.5 9560612 1812.72 51.56620   -0.01804  -0.01596   5296.9014
#> 17004 713102.5 9559032 2186.79 36.53351   -0.01383   0.00193   1809.3838
#> 1635  713102.5 9559052 2172.87 31.48403   -0.00064  -0.00916   2009.1686
#> 2950  715042.5 9557712 2740.72 26.52508   -0.00089  -0.00051    779.0555
#> 12084 712832.5 9559992 1872.46 56.05475   -0.03184   0.03924   8023.8657
#> 15990 712852.5 9559982 1894.77 37.69432   -0.01545   0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.59273    0.00804   0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.63433   -0.00669  -0.00991   1403.0193
#> 47429 714962.5 9557832 2705.12 32.05756    0.02224   0.00177    291.5937
#> 11687 714552.5 9560382 2052.66 29.57551   -0.00255  -0.00365   1536.2647
#> 6967  715342.5 9560192 1995.36 44.88150   -0.00666  -0.00504   1020.2766
#> 651   714602.5 9560152 2111.23 53.85173   -0.01149   0.01949   1182.1506
#> 34707 713962.5 9561092 1819.49 65.68617   -0.04749   0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.54692    0.03480   0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.34897   -0.10372  -0.00008 825875.6875
#> 835   715052.5 9558212 2707.65 37.72526   -0.01399  -0.00731   1384.7959
#> 4617  714292.5 9559312 2342.95 32.39217    0.03504   0.05825    162.6688
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953  -0.02283    590.1536
#> 37416 713282.5 9560692 1850.61 30.25905    0.00736   0.01684   2556.8049
#> 46783 713592.5 9560762 1832.60 36.01040    0.00448  -0.00328    592.3326
#> 44650 715612.5 9558142 2764.88 39.36048   -0.00550  -0.00941   1290.7709
#> 3418  715372.5 9558492 2601.00 22.62725   -0.01168  -0.04102  35685.0000
#> 31121 713462.5 9559132 2135.56 61.59697   -0.03800  -0.06020   1563.5549
#> 41721 715182.5 9557562 2784.85 31.78999   -0.00588   0.01448   1258.8196
#> 40216 714602.5 9560592 2015.55 34.36887   -0.00093   0.00664   1807.2659
#> 18958 714192.5 9558372 2495.72 34.67712   -0.00178  -0.00822    472.8748
#> 43859 714892.5 9557602 2722.70 28.12936    0.00370  -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.14047   -0.05458  -0.14813 287033.4688
#> 1424  714012.5 9558902 2305.97 32.22372   -0.03986  -0.02524  10861.9482
#> 10467 714642.5 9559652 2249.12 37.14313   -0.00759  -0.00400    608.6523
#> 17345 713922.5 9559682 2227.44 34.49091   -0.06509  -0.01211  49347.5547
#> 31117 712892.5 9558922 2129.84 30.55068    0.00521  -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.01864   -0.00589   0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.21547    0.03890  -0.00911    173.2806
#> 33875 712832.5 9559852 1892.81 37.55452   -0.01175   0.04195   1901.6743
#> 22907 713172.5 9559612 2056.81 22.27259   -0.00166   0.01476   1404.4454
#> 17058 714032.5 9558502 2402.95 31.35282    0.01924   0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.05925    0.02213  -0.01363    565.5207
#> 47298 714602.5 9560572 2025.50 28.59002   -0.00253   0.00753   1337.7854
#> 39325 714142.5 9558592 2357.72 40.79288   -0.00246   0.00396    704.7076
#> 37439 714602.5 9560672 1982.99 18.90646   -0.00070   0.00010   3467.2690
#> 10363 713252.5 9559722 1999.55 41.82993   -0.03669  -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.40342    0.03051   0.02190    264.6404
#> 49378 713252.5 9560732 1875.41 27.83544    0.00181  -0.00001   1184.7952
#> 4667  712652.5 9560472 1957.89 36.32438   -0.00296   0.00426   1315.5334
#> 48650 715592.5 9558612 2626.74 31.87651    0.00010   0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.97833    0.01364  -0.00094   1392.6008
#> 14950 713382.5 9559172 2111.91 48.10668    0.00915  -0.02705    413.5702
#> 26058 713402.5 9560442 1880.12 31.37746    0.00730  -0.02610   1872.9276
#> 28260 715312.5 9558302 2681.75 34.48404   -0.01125  -0.02685   2568.4631
#> 22847 714582.5 9560382 2040.49 25.33619    0.00181  -0.00341   1363.4000
#> 13278 712752.5 9560452 1901.25 42.11354   -0.00118  -0.00252   3631.7017
#> 14140 714672.5 9559262 2326.36 45.79251   -0.01400  -0.01210    585.8992
#> 3771  714842.5 9561152 1724.51 45.53983    0.02405  -0.05655    130.9468
#> 34718 714802.5 9558802 2599.22 44.65003    0.01996   0.04684    209.1778
#> 36765 715302.5 9560262 1952.73 27.20346   -0.00585  -0.02665   1904.4769
#> 17851 715362.5 9557482 2869.26 29.12574   -0.03251  -0.00389  14516.2002
#> 36744 714922.5 9557722 2692.39 25.98077    0.00316  -0.00427    866.9398
#> 10145 713782.5 9560932 1865.62 23.59555   -0.00203   0.00642   1688.1782
#> 36561 715292.5 9560802 1844.64 31.93724    0.00002  -0.01232    977.7320
#> 27884 714072.5 9559202 2446.27 19.75673   -0.00009   0.00109    605.1908
#> 15186 713322.5 9561212 1936.88 36.07973    0.00165   0.00565   1410.8372
#> 748   715102.5 9559692 2301.39 27.02184   -0.00759   0.00158    837.3730
#> 25212 714272.5 9559012 2433.45 37.67140   -0.01909  -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.97550    0.00872   0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.18588   -0.00298  -0.00712   1010.5097
#> 8635  713832.5 9557902 2397.71 47.11948    0.02295   0.02715    520.7355
#> 11111 715142.5 9558482 2650.65 42.34330   -0.00533   0.01493   2212.0525
#> 25140 713332.5 9560692 1809.95 52.25719    0.00075  -0.01945    259.7421
#> 43243 715682.5 9558462 2693.23 27.40629   -0.00031  -0.00199    836.4196
#> 47005 714942.5 9559192 2333.66 29.22085   -0.04937  -0.02273  72123.6172
#> 7538  714722.5 9558582 2660.25 15.99469    0.00544  -0.01014    271.4219
#> 39796 714802.5 9560372 2046.48 40.15575    0.00176   0.02973    215.5101
#> 1085  715722.5 9557532 3097.22 28.78425    0.02327   0.01833    300.0125
#> 44396 715512.5 9558102 2845.31 26.59842    0.01263   0.03598    186.3527
#> 39685 715332.5 9558102 2797.04 47.78812    0.00167   0.01133    828.8604
#> 31374 714652.5 9560592 1999.69 27.10663   -0.00007  -0.00853   1354.6001
#> 31497 714032.5 9560672 2060.77 16.21700    0.00973  -0.00943    371.5514
#> 12974 714662.5 9557332 2667.61 41.94166   -0.00065   0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.99798    0.00990   0.02790    209.6952
#> 21925 714542.5 9559232 2314.33 50.24783    0.00737   0.01033   3446.5520
#> 44668 712612.5 9560472 1972.24 27.50713    0.00601  -0.00011    712.5228
#> 42934 713832.5 9558672 2270.74 27.61198   -0.03193  -0.00267  31588.1953
#> 33956 714082.5 9558712 2327.95 39.68420    0.02218   0.00902    218.7467
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224  -0.02036    476.9504
#> 29835 715402.5 9558512 2610.14 27.11809   -0.01322  -0.00759   4131.5356
#> 21070 712832.5 9560062 1884.42 50.93366   -0.00342   0.02552   2515.0752
#> 7983  713512.5 9560432 1907.38 50.01922    0.03310  -0.07650    766.6254
#> 40909 712722.5 9560272 1827.15 29.29304    0.00009  -0.00599  12873.8018
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649 100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.75618    0.00119   0.00361   1011.0248
#> 26768 715092.5 9560582 1960.72 36.65612    0.00980  -0.00710    472.0236
#> 29716 715272.5 9559992 2144.94 25.71778    0.00262  -0.00502    520.5218
#> 28286 713702.5 9559842 2245.07 27.96893    0.00511   0.01499    617.3855
#> 38362 714402.5 9558042 2499.31 56.90101    0.03044  -0.00964    543.0025
#> 18246 713902.5 9561342 1818.52 19.11043   -0.01502   0.00282   1659.6211
#> 48050 714022.5 9560572 2099.82 18.71395    0.01182  -0.00972    331.7462
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909  -0.01151 143213.1562
#> 16845 712772.5 9560022 1831.76 21.39997   -0.00865  -0.00805   7489.7417
#> 11394 713472.5 9561232 1889.48 34.31444    0.00108  -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.99129    0.00253  -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.47385   -0.00278  -0.01172 111973.8906
#> 1434  714672.5 9559902 2194.92 42.45216   -0.01196  -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.59800    0.01508  -0.00177    184.1042
#> 43972 713642.5 9558702 2304.16 24.02355    0.00340   0.01809   1158.0426
#> 28526 715152.5 9560732 1871.53 29.53426   -0.00163  -0.00377   1727.6055
#> 9526  714432.5 9560582 2076.13 31.08010   -0.00172  -0.00119    792.3544
#> 35257 713822.5 9557982 2350.80 29.30794    0.01549  -0.00169    649.9975
#> 48157 713142.5 9560992 1978.48 44.62138    0.00208   0.00492    772.2691
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456  -0.01586    941.0288
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 16435 49.4439659   300.00   TRUE       300.00             41  2.478905
#> 15762 46.3018653   300.00   TRUE       300.00            100  2.633933
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 24072 37.6668184   300.00   TRUE       300.00            100  3.351559
#> 34512 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 26531 15.9878143   300.00   TRUE       300.00              2  2.552843
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43327 29.0587005   173.40   TRUE        33.81            100  3.516154
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 26447 33.0367465   300.00   TRUE       300.00            100  2.676242
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 10899 24.7741221   300.00   TRUE       300.00             20  2.644350
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 43964 28.7573247   300.00   TRUE       300.00             46  2.952143
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 35285 42.4664859   300.00   TRUE       300.00            100  2.997681
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 23375 34.6398824   300.00   TRUE       195.00              2  3.224136
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 14522 28.3986531   300.00   TRUE       300.00            100  3.047228
#> 20290 15.4566824    95.53   TRUE         0.00             49  4.011286
#> 27978 14.3285285   300.00   TRUE       300.00             12  2.377795
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 15340 27.9752373   300.00   TRUE       300.00            100  3.144334
#> 35942 27.1645020   300.00   TRUE       300.00            100  3.364960
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 17004 27.1891392   300.00   TRUE       300.00            100  3.257531
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 2950  28.2038475   300.00   TRUE       300.00            100  2.891568
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 11687 23.1939045   300.00   TRUE       300.00             68  3.186466
#> 6967  41.1194621   300.00   TRUE       300.00             85  3.008718
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 4617  27.9517460   300.00   TRUE       300.00              6  2.211304
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 3418  35.4534824   300.00   TRUE       300.00             73  4.552486
#> 31121 38.1601351   300.00   TRUE       300.00             45  3.194113
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 40216 26.3640800   300.00   TRUE       300.00            100  3.257022
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 8125  33.1874344   300.00   TRUE       300.00             48  5.457933
#> 1424  39.1542168   300.00   TRUE       300.00            100  4.035908
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 22907 33.8136772   300.00   TRUE       205.28              1  3.147505
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 10363 45.0934974   300.00   TRUE       297.43              6  3.599377
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 48650 27.0522023   300.00   TRUE       300.00            100  3.027918
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14950 41.8740475   300.00   TRUE       300.00             15  2.616549
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 15186 38.5486005    60.00  FALSE         0.00            100  3.149477
#> 748   38.1326331   300.00  FALSE       300.00             12  2.922919
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 8635  27.6818829   300.00  FALSE       300.00             23  2.716617
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 12974 37.3385136   300.00  FALSE       300.00            100  2.711947
#> 34616 31.4995007   300.00  FALSE       300.00            100  2.321589
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 42934 32.2271571   300.00  FALSE       300.00             22  4.499525
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 7983  38.5062016   300.00  FALSE        57.15             22  2.884583
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26768 32.2821611   300.00  FALSE       207.22            100  2.673964
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 48050 23.3612082   300.00  FALSE        72.23            100  2.520806
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 16845 22.8827884   300.00  FALSE        25.55             33  3.874467
#> 11394 40.6038001    46.04  FALSE        55.00            100  3.187005
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 28526 31.8146275   300.00  FALSE        46.28            100  3.237445
#> 9526  27.5695832   300.00  FALSE       300.00            100  2.898919
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769
#> 40043 23.9966184   194.79  FALSE        24.89             54  2.973603