Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv      carea
#> 31358 712882.5 9560002 1911.52 25.564231  -0.00681   -0.00029  5577.3916
#> 16435 713512.5 9559092 2166.13 55.973966   0.02056   -0.06976   301.2347
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357  1752.0375
#> 44302 714932.5 9557982 2649.56 37.300698   0.01633   -0.01813  1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579  1318.7560
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043  1806.4552
#> 40756 714022.5 9558862 2331.20 45.085476  -0.00075    0.00475  1001.0861
#> 30059 713832.5 9559662 2288.02 18.277927   0.01917    0.00333   258.6537
#> 37265 712812.5 9559942 1860.77 63.337047   0.00514   -0.00644  1179.0695
#> 15843 714042.5 9558902 2332.30 50.097711  -0.00858   -0.00292  1233.9299
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010  1527.1028
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388  1051.4222
#> 11463 713112.5 9558902 2244.08 21.891571   0.00160    0.00311   586.7354
#> 25690 715042.5 9557572 2752.67 32.713598   0.00194    0.00526   539.8119
#> 10899 713852.5 9559652 2282.82 28.814048   0.00952    0.01378   440.9104
#> 39061 712922.5 9559912 1931.50 36.315211   0.00414   -0.00104  1641.4587
#> 38410 713712.5 9561172 1776.08 36.319222  -0.02678   -0.04272  4955.5586
#> 16843 714882.5 9558512 2701.78 26.637954  -0.00270   -0.00769  1032.7177
#> 2107  714142.5 9558562 2382.67 35.553750   0.01412    0.00898   458.6938
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518   854.4269
#> 35285 714602.5 9560112 2142.75 51.107835  -0.00017   -0.00032   994.6740
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973   514.5488
#> 47210 714952.5 9557612 2744.35 40.884549  -0.00493    0.00733   598.7983
#> 45575 715432.5 9559452 2305.28 24.234396   0.00993   -0.00413   364.9788
#> 48364 715582.5 9558222 2746.23 25.072060   0.01257    0.00593   939.0466
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088  2322.1746
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118  2359.5059
#> 1635  713102.5 9559052 2172.87 31.484031  -0.00064   -0.00916  2009.1686
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605   798.6775
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642   613.6390
#> 2950  715042.5 9557712 2740.72 26.525081  -0.00089   -0.00051   779.0555
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405  5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346   543.0749
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991  1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562   0.02224    0.00177   291.5937
#> 34359 712862.5 9558912 2111.37 32.812147   0.00463    0.00167  1938.5389
#> 20001 712702.5 9560132 1854.55 36.291147  -0.00204   -0.00376  2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314   0.00316    0.00714   475.3120
#> 4605  714852.5 9557892 2677.78 26.933600   0.00321    0.01169   350.1799
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407   344.8395
#> 651   714602.5 9560152 2111.23 53.851730  -0.01149    0.01949  1182.1506
#> 41037 715532.5 9558792 2529.58 32.846524  -0.00585   -0.00974  2136.6939
#> 39599 714112.5 9558562 2374.25 29.168135   0.00283    0.00117  1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400   608.6523
#> 21952 715352.5 9560172 2013.44 40.488063   0.02409    0.02771   674.3859
#> 22299 713532.5 9560362 1975.58 27.709958   0.00478    0.01092   412.5849
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588   749.2502
#> 40733 712862.5 9558892 2108.60 29.994914   0.00175    0.00236  2715.4138
#> 36120 714622.5 9559942 2166.51 32.986326  -0.02917   -0.01993 10308.6406
#> 19887 714662.5 9559632 2266.47 45.179441  -0.02797   -0.00063   576.2819
#> 2595  714912.5 9557662 2700.72 24.194862  -0.02681   -0.01319  3457.7519
#> 10507 714892.5 9559282 2374.45 42.842346  -0.01057    0.02427   380.9472
#> 13029 713542.5 9560242 2019.80 29.545142   0.00576   -0.00166   650.6246
#> 6317  715042.5 9557702 2745.36 27.873824   0.00090   -0.00299   688.0706
#> 37416 713282.5 9560692 1850.61 30.259047   0.00736    0.01684  2556.8049
#> 35083 714612.5 9558152 2596.12 60.598308   0.04548    0.05543   352.1558
#> 19966 715362.5 9559572 2235.06 26.049526   0.00126   -0.00485  2240.4185
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304  1569.7313
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620   379.0855
#> 42765 713552.5 9559632 2186.95 44.011180   0.00394   -0.00224   942.0322
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555  4439.0967
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599  1378.5760
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180  3904.5254
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594 16587.9980
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112   745.3315
#> 40216 714602.5 9560592 2015.55 34.368873  -0.00093    0.00664  1807.2659
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793   864.9472
#> 10467 714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400   608.6523
#> 44767 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205  1112.2626
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245  6353.3179
#> 17345 713922.5 9559682 2227.44 34.490913  -0.06509   -0.01211 49347.5547
#> 6217  714442.5 9559932 2151.30 36.670445   0.01489    0.00241  1752.9242
#> 24135 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797  7557.7866
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195  1901.6743
#> 22907 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476  1404.4454
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722  1373.4147
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801  1071.6141
#> 11352 713002.5 9560562 1858.95 39.803951  -0.00442   -0.00118 19557.6621
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396   704.7076
#> 19756 715572.5 9558972 2485.95 38.812734   0.00546    0.01484   495.0292
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422  1159.2100
#> 49133 714162.5 9558462 2451.13 21.890426   0.00042    0.00957   495.4453
#> 22301 713722.5 9560712 1913.37 39.271100   0.00705   -0.00915  2036.0378
#> 48574 712692.5 9560412 1901.97 41.978326   0.01364   -0.00094  1392.6008
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044   288.5468
#> 39405 713122.5 9559052 2169.30 37.646192  -0.02110   -0.02450  2589.9910
#> 14950 713382.5 9559172 2111.91 48.106682   0.00915   -0.02705   413.5702
#> 24723 712942.5 9560202 1983.18 33.071697   0.00285   -0.01325   595.0426
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921   525.3245
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358   802.7622
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856  1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632   620.0229
#> 46971 713532.5 9559952 2198.03 29.723904  -0.00529    0.01230  1044.6443
#> 33142 715652.5 9558212 2740.94 14.381241  -0.00465   -0.01036  2505.8369
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476  1404.4454
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189   556.5211
#> 34718 714802.5 9558802 2599.22 44.650028   0.01996    0.04684   209.1778
#> 36765 715302.5 9560262 1952.73 27.203463  -0.00585   -0.02665  1904.4769
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320   911.5911
#> 17374 715712.5 9557952 2832.22 37.874802  -0.01519    0.00609  4275.3457
#> 23176 714012.5 9558892 2308.59 43.029130  -0.01084   -0.02996  1599.3491
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842   246.5161
#> 32405 713022.5 9558982 2208.85 15.004619   0.00599    0.00301   338.9676
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232   977.7320
#> 34250 713822.5 9559142 2316.36 33.759819  -0.02284    0.01085  2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496   0.00872    0.05419   761.5427
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463  1050.7473
#> 20067 714242.5 9560442 2125.23 38.640274   0.00363   -0.00363   389.2691
#> 26963 715702.5 9557882 2880.32 43.733296  -0.01174   -0.01466  2633.9666
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064   683.2795
#> 19493 714252.5 9560372 2156.93 24.879546   0.00825    0.00445   355.5078
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995  1026.7981
#> 16542 713772.5 9559622 2304.85 30.810232  -0.00712    0.00923   811.5580
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028  2394.7261
#> 49835 713272.5 9560372 1898.68 32.825898   0.00846    0.00794  1363.5321
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598   186.3527
#> 13286 713662.5 9559702 2268.14 41.996660  -0.01341    0.04431   315.9084
#> 36833 713322.5 9560572 1794.25 33.754090   0.00527   -0.02077  8759.9961
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853  1354.6001
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759  1880.4403
#> 48183 713642.5 9558262 2314.44 50.023226   0.01338   -0.00399  1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138   0.00196    0.01734   335.0773
#> 20319 715282.5 9560482 1901.16 27.676726   0.00706   -0.03035  1471.9783
#> 8895  714132.5 9557692 2487.19 37.364297   0.03423   -0.00983   270.7949
#> 1691  714222.5 9559022 2448.69 38.216858  -0.01292    0.00462   716.5762
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366   267.5057
#> 17746 714752.5 9558322 2648.52 47.736552  -0.00068   -0.00312   593.7773
#> 35038 715092.5 9558152 2715.69 40.183822   0.00270   -0.00250  2225.4426
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547   128.2613
#> 41612 713952.5 9560002 2198.00 28.756179   0.00119    0.00361  1011.0248
#> 1281  715102.5 9560032 2120.64 34.649050  -0.00930   -0.00360  4288.7329
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577   423.5598
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079  1647.6451
#> 14763 713462.5 9560852 1830.40 47.739989   0.01354    0.00907  1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396   0.00537   -0.00347  1893.3645
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050  1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814   0.01600   -0.01849   764.8193
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972   331.7462
#> 16845 712772.5 9560022 1831.76 21.399974  -0.00865   -0.00805  7489.7417
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238   787.0053
#> 14881 715482.5 9559232 2376.03 29.905532  -0.01137   -0.01523   888.2811
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011   501.7867
#> 22628 714942.5 9557732 2696.19 26.211673   0.00788   -0.01029   894.2672
#> 9526  714432.5 9560582 2076.13 31.080096  -0.00172   -0.00119   792.3544
#> 30403 713952.5 9560752 2003.57 42.964386  -0.00933   -0.00157   766.0220
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419   323.0790
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448   941.8503
#> 39234 712902.5 9559652 1953.39 29.731926   0.07379   -0.00249   148.1857
#> 15363 714742.5 9559212 2373.13 33.583348  -0.00247    0.01057   722.3952
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149  4242.3579
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 16435 49.44396589   300.00   TRUE       300.00             41  2.478905
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 44302 35.10111340   300.00   TRUE       300.00            100  3.053301
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 11463 20.13373692   300.00   TRUE       300.00            100  2.768442
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 10899 24.77412210   300.00   TRUE       300.00             20  2.644350
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 48364 32.07589624   300.00   TRUE       300.00            100  2.972687
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 34359 27.41603050   300.00   TRUE       294.31            100  3.287475
#> 20001 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 41037 31.57054747   300.00   TRUE       300.00            100  3.329742
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 22299 21.71911114   300.00   TRUE       127.54              2  2.615513
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 36120 31.56309902   300.00   TRUE       300.00             94  4.013201
#> 19887 38.06215929   300.00   TRUE       300.00             63  2.760635
#> 2595  30.43952878   300.00   TRUE       300.00            100  3.538794
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 17345 29.82130732   300.00   TRUE       300.00              2  4.693266
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 24135 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 22301 33.61543384   300.00   TRUE         0.00            100  3.308786
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 23176 41.55834775   300.00   TRUE       300.00            100  3.203943
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 19493 18.57242693   300.00  FALSE       291.23            100  2.550849
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 16542 21.11234883   300.00  FALSE       300.00            100  2.909320
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 49835 32.36122923   300.00  FALSE         0.00            100  3.134665
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 13286 25.83867769   300.00  FALSE       300.00             76  2.499561
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 35038 40.20387553   300.00  FALSE       300.00            100  3.347416
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 31098 34.65764407   300.00  FALSE       300.00             45  3.277234
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 9526  27.56958319   300.00  FALSE       300.00            100  2.898919
#> 30403 36.13071856   300.00  FALSE       209.91            100  2.884241
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 39234 33.48594538   300.00  FALSE         0.00            100  2.170806
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607