Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 46178 714892.5 9559312 2379.54 32.766883 -0.00266  0.02896     276.0310
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 37104 713962.5 9561282 1795.56 25.741848  0.00273 -0.02403    1787.3024
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 20151 714602.5 9559922 2184.66 48.341595 -0.00979 -0.01542     674.0147
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677    1343.5006
#> 39629 714602.5 9558912 2474.89 40.258307 -0.06175 -0.08226    2325.2053
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 10974 715312.5 9557502 2844.23 26.664883 -0.00831 -0.00149   12632.7793
#> 21703 713842.5 9559672 2282.58 24.991846  0.01581  0.01189     287.5425
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.010397  0.00448 -0.00328     592.3326
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 38279 714192.5 9558612 2325.14 28.252549 -0.04430 -0.06070  138869.3906
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 26215 714332.5 9560262 2138.08 25.721221 -0.00491  0.00581    1855.1630
#> 27708 715222.5 9557782 2844.65 29.542850  0.03206  0.02773     179.3077
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 46756 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 12460 712762.5 9559622 1906.41 24.261898  0.01047 -0.00887     366.8537
#> 34359 712862.5 9558912 2111.37 32.812147  0.00463  0.00167    1938.5389
#> 6310  713802.5 9559752 2238.38 19.651306 -0.02379  0.01338   10787.6299
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 4031  715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 39599 714112.5 9558562 2374.25 29.168135  0.00283  0.00117    1018.2141
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 40733 712862.5 9558892 2108.60 29.994914  0.00175  0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607 -0.00762  0.01252    5761.3486
#> 18892 712682.5 9560202 1839.08 40.923511 -0.01364 -0.03357     473.4709
#> 26698 713382.5 9559082 2155.56 34.264595 -0.00211  0.00241     585.6182
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063     576.2819
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 37416 713282.5 9560692 1850.61 30.259047  0.00736  0.01684    2556.8049
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 35014 713502.5 9559772 2190.48 44.235780  0.00927  0.01643     670.2289
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219     420.7032
#> 7064  715592.5 9558662 2597.69 34.222196  0.00099  0.00240    1201.4724
#> 31834 713492.5 9560682 1800.85 34.237093 -0.00472 -0.00228    2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083 -0.02605 -0.03594   16587.9980
#> 39722 715022.5 9557712 2735.53 30.260766  0.00133  0.00367     806.0966
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 39166 712832.5 9560142 1922.46 39.660139  0.01469  0.02940     987.8085
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220    8936.2666
#> 1973  715252.5 9557822 2861.00 15.570701  0.01865  0.00335     195.0293
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 10073 715572.5 9558612 2632.78 32.161267  0.00129  0.00171     788.0004
#> 20368 715252.5 9560622 1865.83 33.572462  0.00025 -0.01245    6353.3179
#> 6217  714442.5 9559932 2151.30 36.670445  0.01489  0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.106417  0.00516  0.01574     183.3220
#> 47702 714692.5 9559712 2200.51 33.098626 -0.13038 -0.04212   46310.0195
#> 42201 714642.5 9557452 2592.29 13.225585 -0.02084 -0.02236    9755.3027
#> 34887 714662.5 9557392 2609.02 51.020746 -0.01878 -0.04351    1442.9247
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722    1373.4147
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161    3065.3694
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 39325 714142.5 9558592 2357.72 40.792876 -0.00246  0.00396     704.7076
#> 38668 715582.5 9558242 2735.55 35.446034  0.01003  0.01196    1052.5341
#> 49378 713252.5 9560732 1875.41 27.835436  0.00181 -0.00001    1184.7952
#> 22301 713722.5 9560712 1913.37 39.271100  0.00705 -0.00915    2036.0378
#> 23783 713062.5 9560252 1983.09 36.429803  0.00774 -0.00224    1137.3363
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 48574 712692.5 9560412 1901.97 41.978326  0.01364 -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 16065 715242.5 9560622 1871.61 34.306421 -0.00019  0.00119    6589.5796
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 22847 714582.5 9560382 2040.49 25.336194  0.00181 -0.00341    1363.4000
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921     525.3245
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 14140 714672.5 9559262 2326.36 45.792506 -0.01400 -0.01210     585.8992
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 36765 715302.5 9560262 1952.73 27.203463 -0.00585 -0.02665    1904.4769
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.625209  0.00159 -0.01559     866.5775
#> 46165 714392.5 9557912 2542.56 16.329870  0.00325 -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.756731 -0.00009  0.00109     605.1908
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230     714.7432
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403     241.2991
#> 18746 713292.5 9560952 1890.51 41.885507 -0.00402 -0.01328    1066.7134
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 16542 713772.5 9559622 2304.85 30.810232 -0.00712  0.00923     811.5580
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 47005 714942.5 9559192 2333.66 29.220848 -0.04937 -0.02273   72123.6172
#> 13286 713662.5 9559702 2268.14 41.996660 -0.01341  0.04431     315.9084
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.062138  0.00196  0.01734     335.0773
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 8831  714372.5 9559852 2200.77 48.760427 -0.00457  0.01697     412.1185
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 31201 714672.5 9560022 2224.47 38.863727  0.01180  0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983     270.7949
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 20135 714772.5 9560232 2107.52 37.850738 -0.00782 -0.00048    1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 41855 713182.5 9560452 1851.53 45.178295  0.00667  0.00863    1098.3462
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499     617.3855
#> 974   713402.5 9561062 1854.56 23.634509  0.00082  0.02578    1258.8865
#> 12300 712372.5 9560062 1903.64 50.036977  0.00343  0.01177    1558.2231
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 31440 714842.5 9558392 2704.41 25.801435  0.00992 -0.01052     451.7209
#> 32342 714862.5 9559622 2294.74 36.967237  0.01021  0.00109     915.4989
#> 24962 715152.5 9560472 1967.37 28.845560  0.00298  0.00222     991.8300
#> 14645 715072.5 9560262 2034.04 38.169302 -0.00716 -0.03234    1707.3130
#> 11025 714402.5 9559252 2325.64 49.620437 -0.00241  0.00020     673.2861
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.272623 -0.01181 -0.01149    4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 20151 44.7955593   300.00   TRUE       300.00            100  2.828669
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 38993 29.5485794   300.00   TRUE       300.00              2  3.128238
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 10974 32.4655075   300.00   TRUE       300.00            100  4.101499
#> 21703 22.2691506   300.00   TRUE       300.00             34  2.458702
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 38279 37.6284302   300.00   TRUE       300.00             24  5.142607
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 12460 36.2567693   300.00   TRUE         0.00            100  2.564493
#> 34359 27.4160305   300.00   TRUE       294.31            100  3.287475
#> 6310  23.8264499   300.00   TRUE       300.00            100  4.032926
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 40733 27.0138141   300.00   TRUE       300.00            100  3.433836
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 18892 27.0630885   235.02   TRUE        27.24              0  2.675293
#> 26698 34.2496981   300.00   TRUE       300.00              7  2.767615
#> 19887 38.0621593   300.00   TRUE       300.00             63  2.760635
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 37416 28.7716486    72.23   TRUE        87.56             31  3.407698
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 35014 32.5302518   300.00   TRUE       300.00             57  2.826223
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 7064  29.1326121   300.00   TRUE       300.00            100  3.079714
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 39166 29.8184425   300.00   TRUE         1.90             90  2.994673
#> 8136  25.4290129   115.18   TRUE        84.59             30  3.951156
#> 1973  19.3258664   300.00   TRUE       300.00            100  2.290100
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 20368 30.3444178   300.00   TRUE       166.26            100  3.803001
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 47702 33.6664271   300.00   TRUE       300.00             86  4.665675
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 26428 35.4752548   300.00   TRUE         0.00            100  3.486483
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 39325 37.4485215   300.00   TRUE       300.00             60  2.848009
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49378 33.9941589    22.68   TRUE       134.63             25  3.073643
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 23783 27.6005229   300.00   TRUE         0.00            100  3.055889
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 48574 34.4244502    96.21   TRUE       101.32             10  3.143827
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 22847 24.7477660   300.00   TRUE       300.00             41  3.134623
#> 46627 53.6832806   300.00   TRUE       300.00            100  2.720428
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 36765 36.1851496   300.00   TRUE       300.00            100  3.279776
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 30809 36.3662042   181.84  FALSE         0.00            100  2.854150
#> 27965 32.2953391   300.00  FALSE       300.00            100  2.382556
#> 18746 39.1364552     4.48  FALSE       205.00             27  3.028048
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 47005 36.1353022   300.00  FALSE       300.00            100  4.858077
#> 13286 25.8386777   300.00  FALSE       300.00             76  2.499561
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 36126 19.2954997   300.00  FALSE       300.00            100  2.525145
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 8831  32.5554619   300.00  FALSE       300.00             24  2.615022
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 31201 33.7225769   300.00  FALSE       300.00            100  2.404000
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 20135 32.3325813   300.00  FALSE       300.00            100  3.190519
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 41855 28.7017478   234.40  FALSE        47.05            100  3.040739
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 974   22.7584566    80.00  FALSE        90.05            100  3.099987
#> 12300 38.4173931   124.35  FALSE         0.00            100  3.192630
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 32342 37.6198359   300.00  FALSE       300.00             81  2.961658
#> 24962 29.0501061   300.00  FALSE       300.00            100  2.996437
#> 14645 45.2132455   300.00  FALSE       300.00            100  3.232313
#> 11025 36.8199868   300.00  FALSE       300.00             50  2.828200
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 47629 31.6633666   300.00  FALSE       300.00            100  3.627607
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983