Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 15412 715272.5 9557702 2813.17 30.952326 -0.00123  0.00393    2081.0056
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507  159065.0781
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644    1179.0695
#> 21939 713642.5 9558712 2299.09 33.750652 -0.00503  0.02433    1475.3191
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890    2301.0454
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 31023 714712.5 9561042 1838.66 44.120042 -0.00223  0.01783     377.8918
#> 8019  714672.5 9559272 2316.84 40.381492 -0.01966 -0.02234     849.0547
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441 -0.02169 -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073  0.03696  0.02514     297.7842
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501    3889.7292
#> 22933 712712.5 9560432 1902.17 38.123466 -0.00746  0.00096    3210.5781
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 34275 715332.5 9558442 2650.77 45.536967 -0.00043  0.00973     514.5488
#> 4827  714962.5 9559882 2274.42 29.244912  0.00051  0.00889     362.8363
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088    2322.1746
#> 5749  712852.5 9560162 1934.78 23.388710  0.00071  0.00150    1274.3594
#> 27957 715012.5 9559332 2306.35 20.480376  0.00225  0.00325     616.8652
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124    2114.2253
#> 17594 714832.5 9561052 1818.04 46.167220 -0.01353  0.00063    5035.0000
#> 24100 715132.5 9558992 2420.18 29.213972 -0.03229 -0.00511  361979.2812
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 46235 712712.5 9561042 2023.11 41.277025  0.01458  0.00672     565.1107
#> 18564 715552.5 9558592 2648.36 26.291314  0.00316  0.00714     475.3120
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.851730 -0.01149  0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 34707 713962.5 9561092 1819.49 65.686173 -0.04749  0.04779    2972.0193
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 835   715052.5 9558212 2707.65 37.725260 -0.01399 -0.00731    1384.7959
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588     749.2502
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 2595  714912.5 9557662 2700.72 24.194862 -0.02681 -0.01319    3457.7519
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690    5121.6924
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299     688.0706
#> 25700 715172.5 9559272 2324.49 35.534842 -0.00529  0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.726972 -0.02027 -0.01473    2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526  0.00126 -0.00485    2240.4185
#> 20065 713772.5 9558852 2221.21 28.019928 -0.00568 -0.02553    1750.4175
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112     745.3315
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 33910 715202.5 9559572 2195.66 13.791667 -0.04376 -0.01644 1057765.6250
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 24135 712612.5 9559492 1903.67 46.756794 -0.06953 -0.01797    7557.7866
#> 43245 713722.5 9560012 2153.38 34.191256 -0.01263  0.01322    4605.1411
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 19832 715182.5 9557662 2761.27 38.432863 -0.00666 -0.03583    1898.0105
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 1036  713042.5 9558892 2205.38 31.509814 -0.00428 -0.00522    1355.7436
#> 28009 715002.5 9559922 2237.68 34.603786 -0.01109  0.00298    1533.7925
#> 47298 714602.5 9560572 2025.50 28.590021 -0.00253  0.00753    1337.7854
#> 332   714852.5 9561102 1787.11 45.358204  0.00540  0.00879    1200.4210
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 4667  712652.5 9560472 1957.89 36.324378 -0.00296  0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.897497  0.00401 -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319 -0.08153 -0.02417   34908.6562
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121    1783.0658
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325  0.04864  0.04226     338.3474
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 13278 712752.5 9560452 1901.25 42.113544 -0.00118 -0.00252    3631.7017
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 11057 714962.5 9557642 2737.48 42.685929 -0.00488 -0.00632     620.0229
#> 4479  713982.5 9557812 2399.24 41.231189  0.01895  0.04744     226.7911
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 43834 713092.5 9560662 1905.55 31.323603  0.00758 -0.00288    2144.8164
#> 6212  714592.5 9560362 2034.54 12.522566 -0.01454  0.00144  317260.6875
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 3771  714842.5 9561152 1724.51 45.539831  0.02405 -0.05655     130.9468
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684     209.1778
#> 9984  712742.5 9560092 1848.56 23.612737  0.00614 -0.00093     118.1402
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 13072 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781   29995.1543
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 25212 714272.5 9559012 2433.45 37.671402 -0.01909 -0.01181    1286.3949
#> 40459 712542.5 9559192 1992.60 39.817129 -0.00162 -0.00517    1060.5494
#> 14694 715032.5 9559482 2307.44 44.173327 -0.00007 -0.00463    1050.7473
#> 14717 715382.5 9558062 2799.05 50.590454 -0.00812 -0.02208     951.3401
#> 26023 715832.5 9557632 3097.56 41.710182  0.00133 -0.00452     578.5847
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875    1165.3953
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 9665  713362.5 9558812 2275.28 31.494917  0.01306 -0.00626     450.7346
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028    2394.7261
#> 43243 715682.5 9558462 2693.23 27.406290 -0.00031 -0.00199     836.4196
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 4007  714132.5 9558412 2464.44 43.720118  0.00174 -0.03065     704.4843
#> 32849 714302.5 9558222 2520.01 32.157256 -0.00846  0.00216    1412.2295
#> 43403 712632.5 9559262 2072.53 34.486330 -0.00069 -0.00900     614.5898
#> 29944 713562.5 9558952 2209.55 17.339995  0.01432 -0.00362     283.7090
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 3733  713762.5 9559632 2297.63 34.214748 -0.01338  0.00518    1137.0179
#> 29835 715402.5 9558512 2610.14 27.118092 -0.01322 -0.00759    4131.5356
#> 39440 714222.5 9557872 2508.52 39.047647  0.02785  0.00645     234.0114
#> 12044 713612.5 9559292 2247.76 23.074729 -0.00112 -0.00038    2017.3547
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948    1222.5433
#> 49831 715262.5 9557332 2884.30 22.294934 -0.00097  0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.060776  0.01142  0.00868     235.8511
#> 40909 712722.5 9560272 1827.15 29.293040  0.00009 -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740     909.4791
#> 24819 714032.5 9560882 1981.86 33.284264  0.00804 -0.00165     805.1403
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 29716 715272.5 9559992 2144.94 25.717784  0.00262 -0.00502     520.5218
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451  0.00420  0.01960     884.2003
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849     764.8193
#> 24446 714742.5 9560082 2200.60 28.721801 -0.00598  0.02208     842.4825
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 10981 713832.5 9560022 2125.33 34.848439 -0.06909 -0.01151  143213.1562
#> 1577  713322.5 9559742 2027.47 29.567487 -0.03284 -0.00886   43249.3945
#> 1434  714672.5 9559902 2194.92 42.452162 -0.01196 -0.01935     903.6397
#> 16443 714472.5 9559532 2222.55 16.744119  0.03263 -0.02012     305.3624
#> 39883 712732.5 9560842 2082.69 25.090968  0.00732  0.01858     298.4490
#> 6577  714142.5 9559902 2244.30 57.890510 -0.06986 -0.05013     616.4128
#> 42990 712432.5 9560692 2102.11 32.450038 -0.00087 -0.00573    1490.0312
#> 15880 715892.5 9557572 3113.36 53.043287  0.00962 -0.00172     351.5812
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 39234 712902.5 9559652 1953.39 29.731926  0.07379 -0.00249     148.1857
#> 30296 713322.5 9559782 2042.96 52.621017 -0.02110 -0.02931    1812.3820
#> 24516 713802.5 9560862 1873.21 23.349749 -0.00771  0.00661    5642.7544
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 15412 37.5929069   300.00   TRUE       300.00            100  3.318273
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 5999  27.1421567   300.00   TRUE       300.00            100  5.201575
#> 37265 36.5174651   300.00   TRUE         2.45              0  3.071539
#> 21939 23.9324471   300.00   TRUE       300.00            100  3.168886
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 31023 16.6776555   135.00   TRUE         0.00              2  2.577367
#> 8019  35.7800684   300.00   TRUE       300.00             61  2.928936
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 45869 23.4546003   300.00   TRUE       300.00             12  3.970546
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 18304 44.9084320   300.00   TRUE       300.00             57  3.589919
#> 22933 33.0631025    87.56   TRUE       115.40              2  3.506583
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 34280 33.6211634   300.00   TRUE         0.00            100  3.365895
#> 5749  28.2330683   300.00   TRUE         0.00            100  3.105292
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 5728  40.4084215   300.00   TRUE       300.00              2  3.325151
#> 17594 19.2319014   138.35   TRUE         0.00             60  3.701999
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 46235 30.1083592   300.00   TRUE        41.37              0  2.752133
#> 18564 21.7002035   300.00   TRUE       300.00            100  2.676979
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 835   36.6503913   300.00   TRUE       300.00            100  3.141386
#> 421   10.7137378   182.02   TRUE         0.00             83  2.874627
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 2595  30.4395288   300.00   TRUE       300.00            100  3.538794
#> 27875 21.2670474   157.93   TRUE         0.00             57  3.709413
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 6317  27.3232113   300.00   TRUE       300.00            100  2.837633
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 31462 33.7546626   300.00   TRUE       300.00              1  3.466909
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 20065 30.3896815   300.00   TRUE       300.00              5  3.243142
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 33910 32.6734912   300.00   TRUE       300.00              6  6.024389
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 18440 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 24135 38.6683486   300.00   TRUE       122.87            100  3.878395
#> 43245 24.4458173   300.00   TRUE       300.00            100  3.663243
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 1036  25.7269509   300.00   TRUE       300.00            100  3.132178
#> 28009 27.0178248   300.00   TRUE       300.00             29  3.185767
#> 47298 25.3361937   300.00   TRUE       300.00            100  3.126386
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 20426 33.4085961   300.00   TRUE       300.00             76  4.542933
#> 1223  40.4582688   300.00   TRUE       300.00              4  3.251167
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 11057 31.2250539   300.00   TRUE       300.00            100  2.792408
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 43834 28.5120988    20.21   TRUE       150.14              2  3.331390
#> 6212  31.9315109   300.00   TRUE       300.00             47  5.501416
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 9984   3.9465333   300.00   TRUE        36.46             76  2.072398
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 13072 34.9446959   300.00   TRUE       300.00            100  4.477051
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 14694 43.5115609   300.00  FALSE       300.00            100  3.021498
#> 14717 50.8563069   300.00  FALSE       300.00            100  2.978336
#> 26023 45.0728709   300.00  FALSE       300.00            100  2.762367
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 12716 24.3449767   300.00  FALSE       300.00             25  2.731444
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 9665  34.0812485   300.00  FALSE       300.00             64  2.653921
#> 45166 23.4477248   300.00  FALSE       300.00              1  3.379256
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 4007  32.0925120   300.00  FALSE       300.00             68  2.847871
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 43403 31.0617609   300.00  FALSE       111.88             86  2.788585
#> 29944 18.6280038   300.00  FALSE       300.00             48  2.452873
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 29835 28.8530086   300.00  FALSE       300.00             45  3.616112
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 12519 30.4859384   300.00  FALSE       120.00            100  3.087264
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 24819 29.9370448   300.00  FALSE       150.00            100  2.905872
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 40612 21.9270948    55.73  FALSE       125.03            100  2.946551
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 24446 31.3631367   300.00  FALSE       300.00            100  2.925561
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 10981 29.1664166   300.00  FALSE       300.00            100  5.155983
#> 1577  38.9296174   300.00  FALSE       300.00             63  4.635980
#> 1434  39.6927972   300.00  FALSE       300.00             35  2.955995
#> 16443 26.5382592   300.00  FALSE       300.00             75  2.484816
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 6577  42.4074712   300.00  FALSE       300.00             70  2.789872
#> 42990 22.2370650   205.00  FALSE         0.00            100  3.173195
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 39234 33.4859454   300.00  FALSE         0.00            100  2.170806
#> 30296 40.1586119   300.00  FALSE       300.00             68  3.258250
#> 24516 36.2544774   300.00  FALSE       122.13            100  3.751491