Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv       carea
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969    368.5606
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305   -0.00555    430.4604
#> 25090 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628    556.0121
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507 159065.0781
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324    861.3537
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514   -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191   -0.00579   2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503    0.02433   1475.3191
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431   4079.0366
#> 42917 714202.5 9557412 2544.08 39.877863   -0.02104   -0.02046   1024.1573
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949   6217.8213
#> 14456 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010   1527.1028
#> 1768  713912.5 9558552 2357.19 38.692413   -0.00645    0.00835   2425.3115
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035   2697.5974
#> 37270 714852.5 9561032 1838.96 35.158982   -0.00854    0.01264   5057.1431
#> 125   713992.5 9561062 1878.28 49.971087    0.00183    0.03387   1989.3012
#> 26447 715632.5 9559102 2410.77 39.656701    0.00730    0.01770    474.5059
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407    0.00547    519.4228
#> 42598 713542.5 9559972 2184.39 35.241488   -0.00841   -0.00010   1527.1028
#> 8019  714672.5 9559272 2316.84 40.381492   -0.01966   -0.02234    849.0547
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123    0.01164    895.6606
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514    297.7842
#> 38410 713712.5 9561172 1776.08 36.319222   -0.02678   -0.04272   4955.5586
#> 29472 714792.5 9561072 1786.08 36.173117   -0.01029   -0.02312   2657.6936
#> 22933 712712.5 9560432 1902.17 38.123466   -0.00746    0.00096   3210.5781
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634   -0.02626   4316.0977
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203  10263.2734
#> 34275 715332.5 9558442 2650.77 45.536967   -0.00043    0.00973    514.5488
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643    0.00314   2360.1501
#> 18391 715112.5 9559432 2235.02 22.257691   -0.02651   -0.03629 360379.1875
#> 27708 715222.5 9557782 2844.65 29.542850    0.03206    0.02773    179.3077
#> 20194 714132.5 9558452 2447.22 24.022401   -0.00987    0.02107   1392.5880
#> 47220 715052.5 9557752 2726.93 32.721620   -0.00664    0.00694   2123.4473
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605    798.6775
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193   -0.00077   3559.4204
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416   1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077    300.7915
#> 15990 712852.5 9559982 1894.77 37.694320   -0.01545    0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725    0.00804    0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991   1403.0193
#> 2858  714292.5 9559332 2335.04 42.355905    0.04547    0.03323    183.1836
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771   -0.01939    703.7817
#> 6310  713802.5 9559752 2238.38 19.651306   -0.02379    0.01338  10787.6299
#> 47884 715032.5 9557712 2738.31 28.289218    0.00006    0.00175    804.5679
#> 17423 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 39278 714692.5 9561052 1830.60 48.394880   -0.01996    0.00626   1014.7159
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964    571.1364
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449   -0.00181   2406.1252
#> 36120 714622.5 9559942 2166.51 32.986326   -0.02917   -0.01993  10308.6406
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252   5761.3486
#> 375   715622.5 9558922 2510.87 37.366589   -0.01017    0.00197    711.1091
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063    576.2819
#> 32909 712812.5 9560452 1883.85 38.251235   -0.00221    0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057    0.02427    380.9472
#> 39732 715362.5 9559582 2230.97 23.346884   -0.00161   -0.00409   2312.8413
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576   -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032    0.00542   3614.7334
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473   2930.2791
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856   -0.01625   1063.9150
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034    0.00294    739.5768
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568   -0.01558   1504.9717
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104    0.01664   1898.5588
#> 16370 713852.5 9558632 2296.97 34.624413   -0.01972    0.03512  13178.3223
#> 45500 712812.5 9560052 1852.16 56.770377    0.00904   -0.02013   4229.5708
#> 42765 713552.5 9559632 2186.95 44.011180    0.00394   -0.00224    942.0322
#> 31834 713492.5 9560682 1800.85 34.237093   -0.00472   -0.00228   2262.3796
#> 18897 712802.5 9559882 1851.71 50.234648   -0.02627   -0.05393   8301.6094
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458   -0.14813 287033.4688
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556   -0.11694 102301.4609
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084   -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143   -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519   -0.01175    0.04195   1901.6743
#> 43245 713722.5 9560012 2153.38 34.191256   -0.01263    0.01322   4605.1411
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 5911  713952.5 9561282 1801.16 33.789613    0.00466   -0.01776   1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833    369.3771
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079    0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.509814   -0.00428   -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193    0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.352824    0.01924    0.02606    564.6876
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213   -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876   -0.00246    0.00396    704.7076
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546    0.01484    495.0292
#> 49072 715472.5 9558902 2482.92 30.689338   -0.01102   -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051    0.02190    264.6404
#> 22297 714642.5 9558172 2606.18 59.469008   -0.00495   -0.03756    353.9086
#> 1997  714752.5 9561012 1853.19 26.566334   -0.00693    0.01133   2397.2681
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152   1530.5657
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144   -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488   -0.00632    620.0229
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010   -0.00870   2099.7930
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476   1404.4454
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842    246.5161
#> 32405 713022.5 9558982 2208.85 15.004619    0.00599    0.00301    338.9676
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159   -0.01559    866.5775
#> 48601 713692.5 9560152 2058.14 59.557244   -0.01433   -0.02177    997.3089
#> 16464 715252.5 9559982 2143.24 23.185883   -0.00298   -0.00712   1010.5097
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390   -0.02831    454.9113
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336    0.01403    241.2991
#> 18746 713292.5 9560952 1890.51 41.885507   -0.00402   -0.01328   1066.7134
#> 20067 714242.5 9560442 2125.23 38.640274    0.00363   -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466   2633.9666
#> 32869 715332.5 9558892 2533.17 20.635075   -0.00242    0.01132    737.5982
#> 19493 714252.5 9560372 2156.93 24.879546    0.00825    0.00445    355.5078
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480    0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345   -0.00075   8141.7549
#> 13512 715152.5 9557992 2778.81 43.389521   -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.810232   -0.00712    0.00923    811.5580
#> 45166 715262.5 9558722 2531.59 13.553317   -0.00502   -0.00028   2394.7261
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538   -0.00052   3508.5591
#> 31497 714032.5 9560672 2060.77 16.216997    0.00973   -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846    0.00216   1412.2295
#> 8413  713242.5 9561232 1967.45 38.194513    0.00236    0.00483   1097.3082
#> 38689 714622.5 9559342 2264.90 42.648113   -0.00844    0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.611982   -0.03193   -0.00267  31588.1953
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706   -0.03035   1471.9783
#> 34228 715932.5 9558592 2690.97  9.535164  197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393    0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.264008    0.02413    0.00528    355.3475
#> 29835 715402.5 9558512 2610.14 27.118092   -0.01322   -0.00759   4131.5356
#> 36721 714092.5 9559752 2301.81 28.933223    0.03418    0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310   -0.07650    766.6254
#> 49831 715262.5 9557332 2884.30 22.294934   -0.00097    0.01487   1769.3392
#> 44715 714932.5 9560132 2195.78 30.393692    0.03479    0.02251    168.1733
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229    0.00099    909.0560
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199    107.2986
#> 41612 713952.5 9560002 2198.00 28.756179    0.00119    0.00361   1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353    0.00165    0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165    805.1403
#> 14671 712672.5 9560372 1872.21 39.197189   -0.00873   -0.01417   1520.6630
#> 34336 713792.5 9559072 2288.95 36.932859   -0.00952   -0.00438   1276.5448
#> 14763 713462.5 9560852 1830.40 47.739989    0.01354    0.00907   1281.5044
#> 31098 714352.5 9559042 2410.15 30.336396    0.00537   -0.00347   1893.3645
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600   -0.01849    764.8193
#> 18308 714842.5 9558742 2568.20 26.017441   -0.02552    0.00532  16771.3066
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282   1659.6211
#> 4269  714162.5 9560572 2087.45  7.334433    0.00293    0.00186   2751.4626
#> 36016 715262.5 9557682 2814.30 37.002187    0.00203    0.00238    787.0053
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196   -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508   -0.00177    184.1042
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169    649.9975
#> 36141 714222.5 9559422 2316.64 16.277158   -0.07627   -0.00124  25826.6914
#> 26635 715092.5 9560412 2010.59 30.571882   -0.00463    0.00253   1237.0536
#> 32618 715132.5 9559642 2266.36 34.532739   -0.00305   -0.01135   1293.4855
#> 11025 714402.5 9559252 2325.64 49.620437   -0.00241    0.00020    673.2861
#> 41373 715332.5 9560862 1873.10 29.124591   -0.00166    0.00156    758.4840
#> 39234 712902.5 9559652 1953.39 29.731926    0.07379   -0.00249    148.1857
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023   -0.00227    841.2025
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505    0.00855    229.6061
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661   5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 15762 46.301865   300.00   TRUE       300.00            100  2.633933
#> 25090 43.531614   300.00   TRUE       300.00             26  2.745084
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 37265 36.517465   300.00   TRUE         2.45              0  3.071539
#> 24072 37.666818   300.00   TRUE       300.00            100  3.351559
#> 21939 23.932447   300.00   TRUE       300.00            100  3.168886
#> 9669  31.664513   300.00   TRUE       300.00              1  3.610558
#> 42917 38.803567   300.00   TRUE       300.00            100  3.010367
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 14456 20.068993   300.00   TRUE       247.02            100  3.183868
#> 1768  28.185513   300.00   TRUE       300.00              7  3.384768
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 37270 18.946568   162.65   TRUE         0.00             77  3.703905
#> 125   29.596708   207.04   TRUE        21.03            100  3.298701
#> 26447 33.036746   300.00   TRUE       300.00            100  2.676242
#> 16811 35.252374   300.00   TRUE       300.00             59  2.715521
#> 42598 20.068993   300.00   TRUE       247.02            100  3.183868
#> 8019  35.780068   300.00   TRUE       300.00             61  2.928936
#> 43964 28.757325   300.00   TRUE       300.00             46  2.952143
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 38410  8.761098    69.52   TRUE        47.61             65  3.695093
#> 29472 18.648630   111.09   TRUE         0.00             25  3.424505
#> 22933 33.063103    87.56   TRUE       115.40              2  3.506583
#> 31576 26.935319   300.00   TRUE       300.00              4  3.635091
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 34275 36.829154   300.00   TRUE       300.00            100  2.711427
#> 32821 29.483262   300.00   TRUE       300.00             65  3.372940
#> 18391 34.507529   300.00   TRUE       300.00            100  5.556760
#> 27708 25.710908   300.00   TRUE       300.00            100  2.253599
#> 20194 28.117331   300.00   TRUE       300.00             41  3.143823
#> 47220 36.480223   300.00   TRUE       300.00            100  3.327041
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 4329  34.561960   300.00   TRUE        16.35              2  3.551379
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 15990 34.148858   300.00   TRUE         0.00              0  3.754443
#> 29667 34.251990   300.00   TRUE       300.00             27  2.734860
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 2858  35.319410   300.00   TRUE       300.00              2  2.262887
#> 48434 29.935899   300.00   TRUE       300.00              6  2.847438
#> 6310  23.826450   300.00   TRUE       300.00            100  4.032926
#> 47884 27.054494   300.00   TRUE       300.00            100  2.905563
#> 17423 25.460526   300.00   TRUE       300.00            100  2.868983
#> 39278 16.218143   125.47   TRUE         0.00              6  3.006344
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 35194 23.449444   300.00   TRUE       300.00            100  3.381318
#> 36120 31.563099   300.00   TRUE       300.00             94  4.013201
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 375   27.974091   300.00   TRUE       300.00            100  2.851936
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 32909 28.379173   118.07   TRUE        98.15              0  3.619059
#> 10507 27.162210   300.00   TRUE       300.00             46  2.580865
#> 39732 29.168135   300.00   TRUE       300.00             25  3.364146
#> 13029 34.187819   300.00   TRUE        96.65              4  2.813330
#> 47100 29.645982    82.80   TRUE       129.73             36  3.558076
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 29801 33.241865   300.00   TRUE       300.00              2  3.026907
#> 23213 25.460526   300.00   TRUE       300.00            100  2.868983
#> 21013 24.963198   300.00   TRUE        13.47             64  3.177528
#> 13146 33.576473   300.00   TRUE       300.00              2  3.278424
#> 16370 31.158591   300.00   TRUE       300.00              8  4.119860
#> 45500 34.956155   300.00   TRUE         0.00              2  3.626296
#> 42765 42.353040   300.00   TRUE       300.00             12  2.974066
#> 31834 24.646925   215.68   TRUE         0.00            100  3.354565
#> 18897 31.788271   300.00   TRUE        20.00              5  3.919162
#> 8125  33.187434   300.00   TRUE       300.00             48  5.457933
#> 27244 37.775107   300.00   TRUE       300.00             54  5.009882
#> 42201 39.558725   300.00   TRUE       300.00            100  3.989241
#> 46138 32.694691   300.00   TRUE       300.00              8  5.794410
#> 33875 27.569583   300.00   TRUE         4.67              0  3.279136
#> 43245 24.445817   300.00   TRUE       300.00            100  3.663243
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5911   4.138474    42.88   TRUE         0.00             13  3.009377
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 26101 28.646171   300.00   TRUE       300.00            100  3.068405
#> 1036  25.726951   300.00   TRUE       300.00            100  3.132178
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 17058 26.322827   300.00   TRUE       300.00              8  2.751808
#> 47244 28.038836   300.00   TRUE       300.00              2  2.752448
#> 39325 37.448521   300.00   TRUE       300.00             60  2.848009
#> 19756 30.116381   300.00   TRUE       300.00            100  2.694631
#> 49072 28.987654   300.00   TRUE       300.00            100  3.662232
#> 2801  25.746431   300.00   TRUE       300.00             26  2.422656
#> 22297 46.649651   300.00   TRUE       300.00             11  2.548891
#> 1997  25.235926   168.72   TRUE         0.00              5  3.379717
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 12074 23.285578   300.00   TRUE       300.00             63  3.252649
#> 11057 31.225054   300.00   TRUE       300.00            100  2.792408
#> 30378 28.569395   300.00   TRUE       300.00             57  3.322176
#> 41408 33.813677   300.00   TRUE       205.28              1  3.147505
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 32405 15.762069   300.00  FALSE       300.00            100  2.530158
#> 16208 38.373848   256.60  FALSE         0.00             67  2.937807
#> 48601 35.857991   300.00  FALSE       245.19            100  2.998830
#> 16464 35.046109   300.00  FALSE       300.00            100  3.004540
#> 15558 50.344083   300.00  FALSE       300.00             45  2.657927
#> 27965 32.295339   300.00  FALSE       300.00            100  2.382556
#> 18746 39.136455     4.48  FALSE       205.00             27  3.028048
#> 20067 29.823599   300.00  FALSE       268.03            100  2.590250
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 32869 14.600111   300.00  FALSE       300.00             78  2.867820
#> 19493 18.572427   300.00  FALSE       291.23            100  2.550849
#> 12716 24.344977   300.00  FALSE       300.00             25  2.731444
#> 43426 35.462650   300.00  FALSE        82.11            100  3.910718
#> 13512 35.934767   300.00  FALSE       300.00            100  3.605261
#> 42329 19.227891   300.00  FALSE       300.00            100  2.726581
#> 16542 21.112349   300.00  FALSE       300.00            100  2.909320
#> 45166 23.447725   300.00  FALSE       300.00              1  3.379256
#> 18627 31.257140   300.00  FALSE       300.00             20  3.545129
#> 31497 21.270485   300.00  FALSE       154.49            100  2.570019
#> 32849 29.552017   300.00  FALSE       300.00             45  3.149905
#> 8413  39.889895    80.73  FALSE         0.00            100  3.040329
#> 38689 36.084309   300.00  FALSE       300.00             18  2.953242
#> 42934 32.227157   300.00  FALSE       300.00             22  4.499525
#> 20319 32.072458   300.00  FALSE       300.00            100  3.167901
#> 34228 10.748688   300.00  FALSE       300.00            100  2.427333
#> 34810 15.554658   300.00  FALSE       300.00            100  2.997789
#> 29107 25.230769   300.00  FALSE       300.00            100  2.550653
#> 29835 28.853009   300.00  FALSE       300.00             45  3.616112
#> 36721 26.668321   300.00  FALSE       300.00            100  2.209482
#> 7983  38.506202   300.00  FALSE        57.15             22  2.884583
#> 49831 29.837923   300.00  FALSE       300.00            100  3.247811
#> 44715 28.450219   300.00  FALSE       300.00            100  2.225757
#> 35826 44.227758   298.23  FALSE        16.35             96  2.958591
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 41612 25.587149   300.00  FALSE       300.00            100  3.004762
#> 26724 21.777553   300.00  FALSE       185.65            100  3.274624
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 14671 37.882251   123.04  FALSE        75.11              0  3.182033
#> 34336 32.451184   300.00  FALSE       300.00            100  3.106036
#> 14763 10.144791   110.04  FALSE        60.35            100  3.107720
#> 31098 34.657644   300.00  FALSE       300.00             45  3.277234
#> 2999  45.583949   300.00  FALSE       300.00            100  2.883559
#> 18308 30.020697   300.00  FALSE       300.00             55  4.224567
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 4269  19.116737   300.00  FALSE       194.53            100  3.439564
#> 36016 39.523775   300.00  FALSE       300.00            100  2.895978
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 41873 11.038605   300.00  FALSE       300.00            100  2.265064
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 36141 27.306023   300.00  FALSE       300.00             85  4.412069
#> 26635 31.394649   300.00  FALSE       300.00            100  3.092389
#> 32618 39.382827   300.00  FALSE       300.00             56  3.111762
#> 11025 36.819987   300.00  FALSE       300.00             50  2.828200
#> 41373 31.128797   300.00  FALSE       120.90            100  2.879946
#> 39234 33.485945   300.00  FALSE         0.00            100  2.170806
#> 6206  32.463216   294.31  FALSE        35.90            100  2.924901
#> 40852 21.518003   300.00  FALSE       300.00            100  2.360983
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491