Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv       carea   cslope
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649   1399.2329 30.71569
#> 36052 714842.5 9558892 2483.27 68.789886 -0.04921 -0.12438    753.6304 53.74516
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969    368.5606 20.47808
#> 40756 714022.5 9558862 2331.20 45.085476 -0.00075  0.00475   1001.0861 39.33527
#> 14254 714852.5 9557882 2680.00 23.335425  0.00479  0.01261    323.8441 15.66524
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333    258.6537 20.19218
#> 5999  715552.5 9558812 2518.03  7.984167 -0.01863 -0.00507 159065.0781 27.14216
#> 19963 714612.5 9560552 2031.70 26.848229  0.00046  0.00324    861.3537 24.43780
#> 37265 712812.5 9559942 1860.77 63.337047  0.00514 -0.00644   1179.0695 36.51747
#> 20799 714792.5 9561002 1857.29 23.129160  0.00090  0.00890   2301.0454 24.80793
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292   1233.9299 42.22183
#> 28293 713762.5 9558812 2250.98 33.285410  0.00118  0.00761    928.6858 28.90228
#> 1768  713912.5 9558552 2357.19 38.692413 -0.00645  0.00835   2425.3115 28.18551
#> 43327 713962.5 9561082 1839.11 63.361684 -0.04951 -0.00529   3282.1186 29.05870
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035   2697.5974 23.54054
#> 26447 715632.5 9559102 2410.77 39.656701  0.00730  0.01770    474.5059 33.03675
#> 38993 714202.5 9558402 2475.75 33.159359 -0.01607  0.00677   1343.5006 29.54858
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388   1051.4222 29.30507
#> 27375 714292.5 9558982 2458.96 43.600942  0.00214  0.00466    724.0713 38.77034
#> 11463 713112.5 9558902 2244.08 21.891571  0.00160  0.00311    586.7354 20.13374
#> 38287 714872.5 9561162 1719.41 48.235598 -0.03021 -0.09328   2235.5259 14.89690
#> 30800 715182.5 9557582 2772.37 39.204064 -0.02475  0.00665  25318.3555 33.24817
#> 16843 714882.5 9558512 2701.78 26.637954 -0.00270 -0.00769   1032.7177 21.39825
#> 18304 714582.5 9558932 2463.25 22.194666 -0.02131  0.00501   3889.7292 44.90843
#> 35285 714602.5 9560112 2142.75 51.107835 -0.00017 -0.00032    994.6740 42.46649
#> 18005 714832.5 9557912 2660.76 31.912030  0.00191 -0.00261    554.9243 24.90590
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044 129412.3984 32.97143
#> 15340 715452.5 9558852 2523.95 34.165473  0.00151  0.00899   1394.2274 27.97524
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733    598.7983 26.91527
#> 8288  714842.5 9557712 2679.04 28.939525  0.00902  0.01358    802.7622 22.71778
#> 32821 715022.5 9559952 2208.33 43.416450 -0.00643  0.00314   2360.1501 29.48326
#> 10320 714912.5 9558482 2724.21 25.282909  0.00356  0.00774    527.1068 16.45592
#> 34280 713652.5 9560702 1896.37 41.273588  0.00382  0.00088   2322.1746 33.62116
#> 23031 714042.5 9558492 2404.94 28.496056  0.01205  0.02205    629.4165 26.74109
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596   5296.9014 26.79494
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642    613.6390 38.47870
#> 35142 714792.5 9558862 2575.17 25.664499  0.06765  0.00114    150.1498 27.15304
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077    300.7915 14.65970
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313    783.4548 17.99259
#> 5728  713502.5 9559662 2140.27 34.307567 -0.00154  0.00124   2114.2253 40.40842
#> 27084 713632.5 9558472 2380.37 23.135463 -0.01752 -0.00517   1000.7119 21.67270
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169    350.1799 17.63564
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679    550.0344 30.00981
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907   1943.4092 34.78942
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407    344.8395 32.09480
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270   4057.5637 24.92825
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058    431.4362 37.64390
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072  54831.6289 39.36048
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373    527.3528 31.59518
#> 22299 713532.5 9560362 1975.58 27.709958  0.00478  0.01092    412.5849 21.71911
#> 421   714892.5 9561022 1849.01 26.132032  0.00002  0.00588    749.2502 10.71374
#> 49569 713232.5 9561122 1868.73 37.312158 -0.00405 -0.01565   2232.6133 41.84941
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197    711.1091 27.97409
#> 19887 714662.5 9559632 2266.47 45.179441 -0.02797 -0.00063    576.2819 38.06216
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602    345.8933 41.73367
#> 27875 714832.5 9561032 1838.22 41.212281 -0.01160  0.01690   5121.6924 21.26705
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427    380.9472 27.16221
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401   1722.6481 31.44564
#> 6317  715042.5 9557702 2745.36 27.873824  0.00090 -0.00299    688.0706 27.32321
#> 28845 714892.5 9559272 2374.42 45.750107 -0.00209  0.02739    365.7626 29.84251
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026    247.0894 29.02146
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656   3036.7383 29.77089
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809   1102.9956 27.36790
#> 5103  714612.5 9559892 2220.36 49.007945 -0.00078  0.00219    420.7032 34.91318
#> 28046 714882.5 9558492 2711.77 28.305834 -0.00131  0.00201    788.0621 19.59115
#> 45500 712812.5 9560052 1852.16 56.770377  0.00904 -0.02013   4229.5708 34.95616
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318    289.0063 15.64862
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081    940.6002 28.18609
#> 47366 714892.5 9557612 2718.78 28.023939  0.00538  0.00112    745.3315 28.72753
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242    507.9263 47.27246
#> 37714 713192.5 9560692 1885.05 18.034420  0.00188  0.00832   3022.3462 30.73174
#> 8136  713242.5 9560602 1798.18 41.240356 -0.01169 -0.03220   8936.2666 25.42901
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822    472.8748 29.98976
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534    344.1447 45.66416
#> 1424  714012.5 9558902 2305.97 32.223719 -0.03986 -0.02524  10861.9482 39.15422
#> 44767 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205   1112.2626 17.29530
#> 18440 714862.5 9558932 2448.29 41.052426 -0.04989 -0.02781  29995.1543 34.94470
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849   1120.5961 33.41318
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261   1627.8989 33.57475
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017    759.9937 28.76764
#> 28312 715712.5 9558032 2783.23 24.252731 -0.00007 -0.02132   1580.1621 41.64257
#> 22907 713172.5 9559612 2056.81 22.272588 -0.00166  0.01476   1404.4454 33.81368
#> 30933 714672.5 9561082 1798.37 54.810289  0.00542 -0.00722   1373.4147 12.84342
#> 26428 713042.5 9560332 1895.15 33.959781 -0.01439 -0.01161   3065.3694 35.47525
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812   1423.7578 33.59481
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261   3353.5422 23.38699
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190    264.6404 25.74643
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957    495.4453 23.51934
#> 1916  715202.5 9559512 2213.91 26.327984 -0.00292 -0.00109   2047.9031 29.27356
#> 1223  713632.5 9559632 2226.22 50.808751 -0.01881  0.00121   1783.0658 40.45827
#> 9606  714392.5 9559172 2353.05 30.783303 -0.00185 -0.01185    953.8614 33.82399
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325    595.0426 28.17348
#> 10456 712442.5 9560292 2010.91 26.891774  0.00445  0.00875    597.2893 21.66869
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152    850.4633 29.80297
#> 46627 714852.5 9558932 2459.24 53.445503  0.02451 -0.03921    525.3245 53.68328
#> 49459 714692.5 9557342 2639.74 42.047081  0.00216 -0.02606    919.4241 41.85686
#> 46038 713042.5 9559102 2146.38 29.497586 -0.00622  0.00152   1530.5657 29.73078
#> 30626 713302.5 9559172 2106.91 35.997792  0.01354  0.00245    629.5983 30.00351
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279  39495.1797 32.18705
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384    947.2075 33.11581
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400   1982.7782 37.47316
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253    528.1924 20.82071
#> 34718 714802.5 9558802 2599.22 44.650028  0.01996  0.04684    209.1778 34.40955
#> 21490 714672.5 9560382 2021.86 32.804126 -0.00783 -0.00367   1966.2112 34.06005
#> 23176 714012.5 9558892 2308.59 43.029130 -0.01084 -0.02996   1599.3491 41.55835
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067   1123.7318 26.48612
#> 31983 714302.5 9559912 2204.77 23.464914 -0.01332  0.01842    246.5161 19.77564
#> 15211 715772.5 9558082 2801.07 36.296876 -0.00079 -0.00331    699.3024 30.83487
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301    338.9676 15.76207
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232    977.7320 25.80487
#> 30809 712532.5 9560032 1928.24 40.899446  0.00570  0.02230    714.7432 36.36620
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116   1817.5222 31.73785
#> 27965 715042.5 9560342 2063.39 29.197356  0.04336  0.01403    241.2991 32.29534
#> 11111 715142.5 9558482 2650.65 42.343300 -0.00533  0.01493   2212.0525 33.38396
#> 46930 713642.5 9560602 1945.03 34.509821  0.00165  0.00875   1165.3953 31.92062
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479    604.0975 25.27030
#> 12716 713962.5 9557762 2402.60 18.932244  0.03480  0.01500    538.8209 24.34498
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409   1575.9425 18.66983
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464   4029.5947 35.93477
#> 45166 715262.5 9558722 2531.59 13.553317 -0.00502 -0.00028   2394.7261 23.44772
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502   1428.8546 25.38891
#> 20547 713732.5 9560822 1851.67 36.968383 -0.01405  0.00476  11992.5273 34.26116
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943    371.5514 21.27049
#> 34616 714522.5 9558742 2568.77 34.997981  0.00990  0.02790    209.6952 31.49950
#> 37029 713022.5 9559082 2163.08 36.032170  0.00490 -0.00780    825.4451 28.89312
#> 42934 713832.5 9558672 2270.74 27.611982 -0.03193 -0.00267  31588.1953 32.22716
#> 8895  714132.5 9557692 2487.19 37.364297  0.03423 -0.00983    270.7949 38.07763
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223   1056.3701 25.75961
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462    716.5762 27.00006
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098    164.0802 23.68436
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528    355.3475 25.23077
#> 14148 714162.5 9560352 2149.57 37.133395  0.00424  0.00196    576.1450 21.87782
#> 26389 713002.5 9559712 1926.02 24.037298  0.00586  0.00654   1576.2065 36.12212
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268    252.0072 19.73782
#> 12519 713982.5 9560912 1956.98 28.824934 -0.00452 -0.00948   1222.5433 30.48594
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552   2515.0752 34.67712
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312    254.0656 21.26533
#> 35826 712912.5 9560302 1891.22 50.869485 -0.00229  0.00099    909.0560 44.22776
#> 34532 713812.5 9559502 2328.80 44.937653 -0.00070 -0.00060    383.3547 31.20672
#> 48988 714682.5 9560852 1942.97 20.290727  0.00480  0.00740    909.4791 18.95860
#> 28286 713702.5 9559842 2245.07 27.968935  0.00511  0.01499    617.3855 17.04836
#> 7986  712752.5 9560242 1857.41 39.830307  0.01852  0.00248   1432.4019 32.00313
#> 44216 715772.5 9558402 2710.44 30.444685 -0.00070 -0.00050   1471.4346 26.06500
#> 2999  714592.5 9558492 2515.83 47.031814  0.01600 -0.01849    764.8193 45.58395
#> 4943  712452.5 9559172 1927.46 34.465130  0.00105 -0.01576   1045.1074 37.77167
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491    261.7371 18.61941
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064   1485.5201 29.18590
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169    649.9975 30.88185
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124  25826.6914 27.30602
#> 6206  714122.5 9560992 1917.28 37.969913 -0.01023 -0.00227    841.2025 32.46322
#>       distroad slides distdeforest distslidespast log.carea
#> 36710   300.00   TRUE       300.00             21  3.145890
#> 36052   300.00   TRUE       300.00            100  2.877158
#> 42632   300.00   TRUE       300.00             10  2.566509
#> 40756   300.00   TRUE       300.00            100  3.000471
#> 14254   300.00   TRUE       300.00             10  2.510336
#> 30059   300.00   TRUE       300.00             41  2.412719
#> 5999    300.00   TRUE       300.00            100  5.201575
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 37265   300.00   TRUE         2.45              0  3.071539
#> 20799   180.67   TRUE         0.00             16  3.361925
#> 15843   300.00   TRUE       300.00            100  3.091290
#> 28293   300.00   TRUE       300.00              4  2.967869
#> 1768    300.00   TRUE       300.00              7  3.384768
#> 43327   173.40   TRUE        33.81            100  3.516154
#> 49372   213.54   TRUE        20.21             68  3.430977
#> 26447   300.00   TRUE       300.00            100  2.676242
#> 38993   300.00   TRUE       300.00              2  3.128238
#> 24018   300.00   TRUE       300.00              5  3.021777
#> 27375   300.00   TRUE       300.00            100  2.859781
#> 11463   300.00   TRUE       300.00            100  2.768442
#> 38287    41.43   TRUE         1.90            100  3.349380
#> 30800   300.00   TRUE       300.00            100  4.403435
#> 16843   300.00   TRUE       300.00             12  3.013982
#> 18304   300.00   TRUE       300.00             57  3.589919
#> 35285   300.00   TRUE       300.00            100  2.997681
#> 18005   300.00   TRUE       300.00              0  2.744234
#> 34482   300.00   TRUE       300.00            100  5.111976
#> 15340   300.00   TRUE       300.00            100  3.144334
#> 47210   300.00   TRUE       300.00            100  2.777281
#> 8288    300.00   TRUE       300.00             83  2.904587
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 10320   300.00   TRUE       300.00             30  2.721899
#> 34280   300.00   TRUE         0.00            100  3.365895
#> 23031   300.00   TRUE       300.00              2  2.798938
#> 3466    101.32   TRUE        95.02             19  3.724022
#> 19730   300.00   TRUE       300.00            100  2.787913
#> 35142   300.00   TRUE       300.00            100  2.176525
#> 19128   300.00   TRUE       300.00             15  2.478266
#> 45104   300.00   TRUE       300.00            100  2.894014
#> 5728    300.00   TRUE       300.00              2  3.325151
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 4605    300.00   TRUE       300.00             10  2.544291
#> 31316   300.00   TRUE       300.00             85  2.740390
#> 13699   300.00   TRUE        53.95             43  3.288564
#> 29926   300.00   TRUE       300.00            100  2.537617
#> 17036    25.33   TRUE       208.61             15  3.608265
#> 13134   300.00   TRUE       300.00             67  2.634917
#> 3944    300.00   TRUE       300.00            100  4.739031
#> 5084    300.00   TRUE       300.00             18  2.722101
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 421     182.02   TRUE         0.00             83  2.874627
#> 49569    25.00   TRUE        35.00            100  3.348814
#> 375     300.00   TRUE       300.00            100  2.851936
#> 19887   300.00   TRUE       300.00             63  2.760635
#> 34005   300.00   TRUE       300.00            100  2.538942
#> 27875   157.93   TRUE         0.00             57  3.709413
#> 10507   300.00   TRUE       300.00             46  2.580865
#> 18265   300.00   TRUE       300.00            100  3.236197
#> 6317    300.00   TRUE       300.00            100  2.837633
#> 28845   300.00   TRUE       300.00             54  2.563199
#> 44825   300.00   TRUE       300.00            100  2.392854
#> 46065   300.00   TRUE       251.11            100  3.482407
#> 27695   300.00   TRUE       300.00             41  3.042574
#> 5103    300.00   TRUE       300.00             91  2.623976
#> 28046   300.00   TRUE       300.00             15  2.896560
#> 45500   300.00   TRUE         0.00              2  3.626296
#> 45592   300.00   TRUE       300.00              2  2.460907
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 47366   300.00   TRUE       300.00            100  2.872349
#> 30268   300.00   TRUE       300.00             41  2.705801
#> 37714    13.57   TRUE       170.00              4  3.480344
#> 8136    115.18   TRUE        84.59             30  3.951156
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 47320   300.00   TRUE       300.00             43  2.536741
#> 1424    300.00   TRUE       300.00            100  4.035908
#> 44767   300.00   TRUE       300.00             25  3.046207
#> 18440   300.00   TRUE       300.00            100  4.477051
#> 16009   300.00   TRUE       300.00            100  3.049449
#> 27915   300.00   TRUE       300.00              6  3.211627
#> 15432   300.00   TRUE       300.00            100  2.880810
#> 28312   300.00   TRUE       300.00            100  3.198702
#> 22907   300.00   TRUE       205.28              1  3.147505
#> 30933    99.29   TRUE         0.00              6  3.137802
#> 26428   300.00   TRUE         0.00            100  3.486483
#> 43074   300.00   TRUE       205.02              1  3.153436
#> 47086   300.00   TRUE       300.00            100  3.525504
#> 2801    300.00   TRUE       300.00             26  2.422656
#> 49133   300.00   TRUE       300.00             38  2.694996
#> 1916    300.00   TRUE       300.00             21  3.311309
#> 1223    300.00   TRUE       300.00              4  3.251167
#> 9606    300.00   TRUE       300.00              0  2.979485
#> 24723   300.00   TRUE        36.63            100  2.774548
#> 10456    24.34   TRUE         0.00             18  2.776185
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 46627   300.00   TRUE       300.00            100  2.720428
#> 49459   300.00   TRUE       300.00            100  2.963516
#> 46038   300.00   TRUE       243.86            100  3.184852
#> 30626   300.00   TRUE       300.00              0  2.799064
#> 37115   300.00   TRUE       300.00            100  4.596544
#> 47722   300.00   TRUE        65.02              1  2.976445
#> 35458   300.00   TRUE       300.00            100  3.297274
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 34718   300.00   TRUE       300.00            100  2.320516
#> 21490   300.00   TRUE       300.00              1  3.293630
#> 23176   300.00   TRUE       300.00            100  3.203943
#> 4785    300.00  FALSE         0.00            100  3.050663
#> 31983   300.00  FALSE       300.00             65  2.391845
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 32405   300.00  FALSE       300.00            100  2.530158
#> 36561   300.00  FALSE        70.00            100  2.990220
#> 30809   181.84  FALSE         0.00            100  2.854150
#> 31071   300.00  FALSE       300.00             76  3.259480
#> 27965   300.00  FALSE       300.00            100  2.382556
#> 11111   300.00  FALSE       300.00            100  3.344795
#> 46930   300.00  FALSE         0.00            100  3.066473
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 12716   300.00  FALSE       300.00             25  2.731444
#> 28162   130.00  FALSE        75.00            100  3.197540
#> 13512   300.00  FALSE       300.00            100  3.605261
#> 45166   300.00  FALSE       300.00              1  3.379256
#> 16878   300.00  FALSE       300.00             18  3.154988
#> 20547   300.00  FALSE        42.56            100  4.078911
#> 31497   300.00  FALSE       154.49            100  2.570019
#> 34616   300.00  FALSE       300.00            100  2.321589
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 42934   300.00  FALSE       300.00             22  4.499525
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 47059   300.00  FALSE       300.00            100  3.023816
#> 1691    300.00  FALSE       300.00            100  2.855262
#> 22545   300.00  FALSE       300.00             96  2.215056
#> 29107   300.00  FALSE       300.00            100  2.550653
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 26389   300.00  FALSE        76.38             26  3.197613
#> 44267   300.00  FALSE       226.30            100  2.401413
#> 12519   300.00  FALSE       120.00            100  3.087264
#> 21070   300.00  FALSE         0.00             10  3.400551
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 35826   298.23  FALSE        16.35             96  2.958591
#> 34532   300.00  FALSE       300.00             92  2.583601
#> 48988   300.00  FALSE       144.19            100  2.958793
#> 28286   300.00  FALSE       300.00            100  2.790556
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 44216   300.00  FALSE       300.00            100  3.167741
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 10127   300.00  FALSE       300.00             49  2.417865
#> 33881   300.00  FALSE       300.00            100  3.171879
#> 35257   300.00  FALSE       300.00             79  2.812912
#> 36141   300.00  FALSE       300.00             85  4.412069
#> 6206    294.31  FALSE        35.90            100  2.924901