Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope    hcurv    vcurv       carea    cslope
#> 37912 712802.5 9559952 1838.40 52.10134  0.00183 -0.09203    634.3320 30.294570
#> 36052 714842.5 9558892 2483.27 68.78989 -0.04921 -0.12438    753.6304 53.745160
#> 15762 714922.5 9558862 2535.97 54.91056  0.01305 -0.00555    430.4604 46.301865
#> 19963 714612.5 9560552 2031.70 26.84823  0.00046  0.00324    861.3537 24.437796
#> 8059  713292.5 9559122 2136.06 29.66203  0.02240 -0.00250   1413.3312 26.977145
#> 37265 712812.5 9559942 1860.77 63.33705  0.00514 -0.00644   1179.0695 36.517465
#> 24072 715282.5 9557602 2837.46 34.39408 -0.02191 -0.00579   2246.7725 37.666818
#> 9669  713852.5 9558612 2308.97 52.18041 -0.01059 -0.07431   4079.0366 31.664513
#> 16811 714902.5 9559262 2362.67 50.70333 -0.01407  0.00547    519.4228 35.252374
#> 31023 714712.5 9561042 1838.66 44.12004 -0.00223  0.01783    377.8918 16.677656
#> 39629 714602.5 9558912 2474.89 40.25831 -0.06175 -0.08226   2325.2053 47.548622
#> 22814 713862.5 9558582 2327.48 48.61203 -0.02894 -0.03416    947.4878 34.912037
#> 42598 713542.5 9559972 2184.39 35.24149 -0.00841 -0.00010   1527.1028 20.068993
#> 25690 715042.5 9557572 2752.67 32.71360  0.00194  0.00526    539.8119 21.747759
#> 2990  713542.5 9560372 1970.07 31.88797  0.01903  0.01447    331.4833 25.321297
#> 45869 714342.5 9560282 2127.11 27.62344 -0.02169 -0.00160   9344.2930 23.454600
#> 38410 713712.5 9561172 1776.08 36.31922 -0.02678 -0.04272   4955.5586  8.761098
#> 40307 713302.5 9560702 1845.07 39.00467  0.00129  0.02510   3342.5518 29.513629
#> 48055 712862.5 9559972 1902.48 29.01687 -0.00720  0.01250   3686.1240 33.825709
#> 23375 713162.5 9559632 2041.52 46.81524 -0.00857  0.03677   1675.4679 34.639882
#> 14522 715012.5 9557732 2721.34 35.18133 -0.00382  0.00672   1114.8801 28.398653
#> 36030 712852.5 9559572 1938.79 28.67883 -0.01440  0.00710    843.7961 29.199648
#> 47022 715362.5 9560102 2059.29 49.11967  0.02059 -0.00628    556.0121 43.531614
#> 47210 714952.5 9557612 2744.35 40.88455 -0.00493  0.00733    598.7983 26.915265
#> 45575 715432.5 9559452 2305.28 24.23440  0.00993 -0.00413    364.9788 25.728097
#> 40828 714842.5 9557632 2698.65 17.53652  0.00490 -0.00450    873.0448 25.599754
#> 32821 715022.5 9559952 2208.33 43.41645 -0.00643  0.00314   2360.1501 29.483262
#> 3453  713482.5 9559642 2145.11 35.12575  0.00108 -0.00678   1891.6584 40.084700
#> 20194 714132.5 9558452 2447.22 24.02240 -0.00987  0.02107   1392.5880 28.117331
#> 15390 713132.5 9560622 1873.26 38.48042  0.00252  0.01118   2359.5059 29.322261
#> 6763  714692.5 9561052 1830.60 48.39488 -0.01996  0.00626   1014.7159 16.218143
#> 46125 713732.5 9560062 2115.18 37.59348 -0.04616  0.01276   9614.6924 25.657050
#> 7879  714722.5 9559232 2356.19 36.80394 -0.01097  0.00156    816.6036 28.580854
#> 22735 714492.5 9559772 2170.30 39.68879 -0.00440 -0.02651    419.0927 36.516892
#> 12868 712532.5 9560332 1913.92 61.93731 -0.03543 -0.05146   1347.6224 37.588323
#> 15990 712852.5 9559982 1894.77 37.69432 -0.01545  0.02405   5681.2412 34.148858
#> 5728  713502.5 9559662 2140.27 34.30757 -0.00154  0.00124   2114.2253 40.408421
#> 12460 712762.5 9559622 1906.41 24.26190  0.01047 -0.00887    366.8537 36.256769
#> 47011 713262.5 9559722 2003.72 37.19298 -0.01846 -0.00034   4495.6929 44.534291
#> 2858  714292.5 9559332 2335.04 42.35591  0.04547  0.03323    183.1836 35.319410
#> 37828 713482.5 9558652 2330.58 20.44543  0.00606 -0.00556    660.3690 18.763795
#> 27084 713632.5 9558472 2380.37 23.13546 -0.01752 -0.00517   1000.7119 21.672702
#> 47091 713382.5 9559182 2122.71 48.03793  0.04527  0.04613    289.6113 40.151736
#> 28743 715202.5 9558872 2463.87 34.65707 -0.01267 -0.05063  18733.8848 34.653060
#> 13134 714652.5 9559632 2265.33 49.59981 -0.03552 -0.02058    431.4362 37.643900
#> 22332 715482.5 9558922 2475.28 19.93664 -0.02617  0.00307 147470.5312 27.068818
#> 23029 713502.5 9559102 2174.07 56.45754  0.03395  0.01495    266.5888 46.601522
#> 16015 712582.5 9560312 1899.18 56.54692  0.03480  0.02530    937.1306 38.832788
#> 9704  714642.5 9559652 2249.12 37.14313 -0.00759 -0.00400    608.6523 38.996080
#> 21952 715352.5 9560172 2013.44 40.48806  0.02409  0.02771    674.3859 40.091576
#> 835   715052.5 9558212 2707.65 37.72526 -0.01399 -0.00731   1384.7959 36.650391
#> 22299 713532.5 9560362 1975.58 27.70996  0.00478  0.01092    412.5849 21.719111
#> 40733 712862.5 9558892 2108.60 29.99491  0.00175  0.00236   2715.4138 27.013814
#> 15766 715562.5 9558592 2646.16 29.02089  0.00194  0.00516    576.0335 23.337144
#> 32909 712812.5 9560452 1883.85 38.25124 -0.00221  0.00121   4159.6738 28.379173
#> 39732 715362.5 9559582 2230.97 23.34688 -0.00161 -0.00409   2312.8413 29.168135
#> 25700 715172.5 9559272 2324.49 35.53484 -0.00529  0.02009   1247.7061 32.541138
#> 37416 713282.5 9560692 1850.61 30.25905  0.00736  0.01684   2556.8049 28.771649
#> 31462 714082.5 9558542 2367.84 34.72697 -0.02027 -0.01473   2930.2791 33.754663
#> 7668  713372.5 9559062 2164.82 37.93496 -0.01807 -0.01883    616.1993 36.320368
#> 32735 714922.5 9559002 2414.49 40.06465  0.00062 -0.00262   1760.2162 39.358190
#> 3418  715372.5 9558492 2601.00 22.62725 -0.01168 -0.04102  35685.0000 35.453482
#> 664   714452.5 9559852 2174.83 44.86088  0.00282  0.00109    767.4097 30.842318
#> 46065 713072.5 9559162 2109.39 16.26455 -0.00195 -0.00656   3036.7383 29.770887
#> 31121 713462.5 9559132 2135.56 61.59697 -0.03800 -0.06020   1563.5549 38.160135
#> 27695 714142.5 9558452 2449.28 20.49012 -0.00539  0.00809   1102.9956 27.367902
#> 13146 713792.5 9558552 2341.40 29.82532 -0.00104  0.01664   1898.5588 33.576473
#> 16370 713852.5 9558632 2296.97 34.62441 -0.01972  0.03512  13178.3223 31.158591
#> 45500 712812.5 9560052 1852.16 56.77038  0.00904 -0.02013   4229.5708 34.956155
#> 26010 712852.5 9560542 1928.25 30.97811  0.00165  0.00555   4439.0967 25.229624
#> 41783 715572.5 9558652 2609.51 32.40821  0.00291 -0.00081    940.6002 28.186086
#> 40270 713312.5 9559012 2186.56 30.93571 -0.00771 -0.00599   1378.5760 27.597658
#> 21680 712802.5 9560462 1893.75 38.73424 -0.00481 -0.00399   3900.3445 28.348233
#> 31834 713492.5 9560682 1800.85 34.23709 -0.00472 -0.00228   2262.3796 24.646925
#> 39722 715022.5 9557712 2735.53 30.26077  0.00133  0.00367    806.0966 26.541697
#> 18958 714192.5 9558372 2495.72 34.67712 -0.00178 -0.00822    472.8748 29.989757
#> 47320 713422.5 9559182 2143.03 56.32519  0.01246  0.00534    344.1447 45.664163
#> 18440 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 31117 712892.5 9558922 2129.84 30.55068  0.00521 -0.00870   1895.1399 26.645975
#> 18348 714042.5 9558882 2341.41 44.06389 -0.00338  0.00728   1137.2450 41.010600
#> 7086  712762.5 9560962 2022.00 44.91359 -0.01259 -0.00281   1178.4420 34.204434
#> 47702 714692.5 9559712 2200.51 33.09863 -0.13038 -0.04212  46310.0195 33.666427
#> 43201 715712.5 9557922 2852.82 31.46226 -0.00949  0.00520   3390.5708 41.828211
#> 42201 714642.5 9557452 2592.29 13.22558 -0.02084 -0.02236   9755.3027 39.558725
#> 34887 714662.5 9557392 2609.02 51.02075 -0.01878 -0.04351   1442.9247 42.935738
#> 46138 715202.5 9559582 2194.40 18.86005 -0.02143 -0.05078 622888.2500 32.694691
#> 33875 712832.5 9559852 1892.81 37.55452 -0.01175  0.04195   1901.6743 27.569583
#> 34615 713242.5 9560732 1879.60 27.59365  0.00384  0.00175   1026.6818 34.189538
#> 15432 715632.5 9558922 2509.67 35.82762 -0.00444 -0.00017    759.9937 28.767638
#> 19832 715182.5 9557662 2761.27 38.43286 -0.00666 -0.03583   1898.0105 36.633776
#> 5572  715062.5 9558262 2754.60 39.91167  0.00707  0.01833    369.3771 26.597847
#> 26428 713042.5 9560332 1895.15 33.95978 -0.01439 -0.01161   3065.3694 35.475255
#> 36292 712862.5 9559602 1915.77 37.97507  0.00511 -0.01801   1071.6141 31.515544
#> 7485  714362.5 9558042 2484.02 47.95886  0.03169  0.01441    424.2680 39.490543
#> 47244 714262.5 9559302 2350.47 17.05925  0.02213 -0.01363    565.5207 28.038836
#> 11352 713002.5 9560562 1858.95 39.80395 -0.00442 -0.00118  19557.6621 27.998156
#> 37439 714602.5 9560672 1982.99 18.90646 -0.00070  0.00010   3467.2690 26.592690
#> 332   714852.5 9561102 1787.11 45.35820  0.00540  0.00879   1200.4210  0.556915
#> 38668 715582.5 9558242 2735.55 35.44603  0.01003  0.01196   1052.5341 32.922728
#> 16071 714562.5 9560062 2166.69 28.69602  0.02281  0.04759    158.9370 29.621345
#> 22297 714642.5 9558172 2606.18 59.46901 -0.00495 -0.03756    353.9086 46.649651
#> 4667  712652.5 9560472 1957.89 36.32438 -0.00296  0.00426   1315.5334 31.517262
#> 48650 715592.5 9558612 2626.74 31.87651  0.00010  0.00140   1066.3957 27.052202
#> 14494 715332.5 9558792 2557.22 20.72045  0.01267  0.00483    343.1384 21.185114
#> 16065 715242.5 9560622 1871.61 34.30642 -0.00019  0.00119   6589.5796 30.320354
#> 30363 715192.5 9557542 2796.31 25.32187 -0.00157 -0.00152    850.4633 29.802973
#> 28588 715532.5 9559362 2299.84 11.14460 -0.04360 -0.03000 162178.1562 31.625551
#> 13278 712752.5 9560452 1901.25 42.11354 -0.00118 -0.00252   3631.7017 30.225243
#> 14779 712502.5 9560202 2004.61 27.52260  0.00115  0.01015    790.2421 16.521238
#> 43834 713092.5 9560662 1905.55 31.32360  0.00758 -0.00288   2144.8164 28.512099
#> 9864  714952.5 9557592 2752.40 31.98938 -0.00983  0.01253    528.1924 20.820713
#> 23832 714252.5 9560182 2188.12 30.64751  0.00427  0.02403    474.8613 25.485163
#> 39466 714942.5 9557772 2680.35 16.75329  0.00166 -0.01265   1148.6487 27.251592
#> 13072 714862.5 9558932 2448.29 41.05243 -0.04989 -0.02781  29995.1543 34.944696
#> 15211 715772.5 9558082 2801.07 36.29688 -0.00079 -0.00331    699.3024 30.834870
#> 24836 714202.5 9558272 2507.87 38.48901 -0.00226  0.00736    477.8591 25.583711
#> 16464 715252.5 9559982 2143.24 23.18588 -0.00298 -0.00712   1010.5097 35.046109
#> 30267 714442.5 9557512 2612.30 28.76592 -0.00458  0.00479    604.0975 25.270304
#> 1488  714562.5 9560002 2150.93 11.28383  0.00197  0.00463    851.6306 29.921575
#> 43243 715682.5 9558462 2693.23 27.40629 -0.00031 -0.00199    836.4196 27.986124
#> 7538  714722.5 9558582 2660.25 15.99469  0.00544 -0.01014    271.4219 17.264937
#> 13286 713662.5 9559702 2268.14 41.99666 -0.01341  0.04431    315.9084 25.838678
#> 4007  714132.5 9558412 2464.44 43.72012  0.00174 -0.03065    704.4843 32.092512
#> 34616 714522.5 9558742 2568.77 34.99798  0.00990  0.02790    209.6952 31.499501
#> 43403 712632.5 9559262 2072.53 34.48633 -0.00069 -0.00900    614.5898 31.061761
#> 37029 713022.5 9559082 2163.08 36.03217  0.00490 -0.00780    825.4451 28.893116
#> 8895  714132.5 9557692 2487.19 37.36430  0.03423 -0.00983    270.7949 38.077629
#> 47059 713822.5 9558272 2413.20 38.17503 -0.00687 -0.00223   1056.3701 25.759610
#> 34105 715432.5 9558592 2621.73 18.95631 -0.00301 -0.01099   1146.2140 20.077014
#> 14148 714162.5 9560352 2149.57 37.13339  0.00424  0.00196    576.1450 21.877820
#> 12044 713612.5 9559292 2247.76 23.07473 -0.00112 -0.00038   2017.3547 32.236897
#> 21070 712832.5 9560062 1884.42 50.93366 -0.00342  0.02552   2515.0752 34.677125
#> 43222 714472.5 9560712 2039.25 22.19066  0.01858  0.00312    254.0656 21.265329
#> 34532 713812.5 9559502 2328.80 44.93765 -0.00070 -0.00060    383.3547 31.206719
#> 15008 714492.5 9559012 2449.93 29.51535  0.00972  0.00308    422.5022 26.907817
#> 28286 713702.5 9559842 2245.07 27.96893  0.00511  0.01499    617.3855 17.048359
#> 32430 715832.5 9558112 2824.39 23.69410  0.00944  0.00836    250.5111 22.860443
#> 7986  712752.5 9560242 1857.41 39.83031  0.01852  0.00248   1432.4019 32.003131
#> 2999  714592.5 9558492 2515.83 47.03181  0.01600 -0.01849    764.8193 45.583949
#> 18308 714842.5 9558742 2568.20 26.01744 -0.02552  0.00532  16771.3066 30.020697
#> 4943  712452.5 9559172 1927.46 34.46513  0.00105 -0.01576   1045.1074 37.771670
#> 18246 713902.5 9561342 1818.52 19.11043 -0.01502  0.00282   1659.6211  7.987605
#> 6732  715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811 26.856251
#> 10981 713832.5 9560022 2125.33 34.84844 -0.06909 -0.01151 143213.1562 29.166417
#> 1434  714672.5 9559902 2194.92 42.45216 -0.01196 -0.01935    903.6397 39.692797
#> 41873 714832.5 9559942 2253.59 10.59800  0.01508 -0.00177    184.1042 11.038605
#> 14881 715482.5 9559232 2376.03 29.90553 -0.01137 -0.01523    888.2811 26.856251
#> 9526  714432.5 9560582 2076.13 31.08010 -0.00172 -0.00119    792.3544 27.569583
#> 39883 712732.5 9560842 2082.69 25.09097  0.00732  0.01858    298.4490 17.044348
#> 47245 715372.5 9560792 1900.65 34.47086 -0.00011 -0.00419    323.0790 34.976209
#> 47629 714472.5 9558482 2437.33 37.27262 -0.01181 -0.01149   4242.3579 31.663367
#>       distroad slides distdeforest distslidespast log.carea
#> 37912   300.00   TRUE         9.15              2  2.802317
#> 36052   300.00   TRUE       300.00            100  2.877158
#> 15762   300.00   TRUE       300.00            100  2.633933
#> 19963   300.00   TRUE       300.00             89  2.935182
#> 8059    300.00   TRUE       300.00              4  3.150244
#> 37265   300.00   TRUE         2.45              0  3.071539
#> 24072   300.00   TRUE       300.00            100  3.351559
#> 9669    300.00   TRUE       300.00              1  3.610558
#> 16811   300.00   TRUE       300.00             59  2.715521
#> 31023   135.00   TRUE         0.00              2  2.577367
#> 39629   300.00   TRUE       300.00             85  3.366461
#> 22814   300.00   TRUE       300.00              6  2.976574
#> 42598   300.00   TRUE       247.02            100  3.183868
#> 25690   300.00   TRUE       300.00            100  2.732242
#> 2990    300.00   TRUE       123.56              5  2.520462
#> 45869   300.00   TRUE       300.00             12  3.970546
#> 38410    69.52   TRUE        47.61             65  3.695093
#> 40307    77.82   TRUE        76.41             54  3.524078
#> 48055   300.00   TRUE         1.67              0  3.566570
#> 23375   300.00   TRUE       195.00              2  3.224136
#> 14522   300.00   TRUE       300.00            100  3.047228
#> 36030   300.00   TRUE         0.00            100  2.926238
#> 47022   300.00   TRUE       300.00             26  2.745084
#> 47210   300.00   TRUE       300.00            100  2.777281
#> 45575   300.00   TRUE       300.00             59  2.562268
#> 40828   300.00   TRUE       300.00            100  2.941037
#> 32821   300.00   TRUE       300.00             65  3.372940
#> 3453    300.00   TRUE       300.00              2  3.276843
#> 20194   300.00   TRUE       300.00             41  3.143823
#> 15390    60.00   TRUE       118.92              2  3.372821
#> 6763    125.47   TRUE         0.00              6  3.006344
#> 46125   300.00   TRUE       300.00            100  3.982935
#> 7879    300.00   TRUE       300.00            100  2.912011
#> 22735   300.00   TRUE       300.00             90  2.622310
#> 12868   119.34   TRUE        79.16             61  3.129568
#> 15990   300.00   TRUE         0.00              0  3.754443
#> 5728    300.00   TRUE       300.00              2  3.325151
#> 12460   300.00   TRUE         0.00            100  2.564493
#> 47011   300.00   TRUE       300.00             10  3.652797
#> 2858    300.00   TRUE       300.00              2  2.262887
#> 37828   300.00   TRUE       300.00              2  2.819787
#> 27084   300.00   TRUE       300.00             81  3.000309
#> 47091   300.00   TRUE       300.00             25  2.461815
#> 28743   300.00   TRUE       300.00             71  4.272628
#> 13134   300.00   TRUE       300.00             67  2.634917
#> 22332   300.00   TRUE       300.00            100  5.168705
#> 23029   300.00   TRUE       300.00             34  2.425842
#> 16015   152.09   TRUE        90.23             45  2.971800
#> 9704    300.00   TRUE       300.00             90  2.784369
#> 21952   300.00   TRUE       300.00             64  2.828908
#> 835     300.00   TRUE       300.00            100  3.141386
#> 22299   300.00   TRUE       127.54              2  2.615513
#> 40733   300.00   TRUE       300.00            100  3.433836
#> 15766   300.00   TRUE       300.00            100  2.760448
#> 32909   118.07   TRUE        98.15              0  3.619059
#> 39732   300.00   TRUE       300.00             25  3.364146
#> 25700   300.00   TRUE       300.00            100  3.096112
#> 37416    72.23   TRUE        87.56             31  3.407698
#> 31462   300.00   TRUE       300.00              1  3.466909
#> 7668    300.00   TRUE       300.00              5  2.789721
#> 32735   300.00   TRUE       300.00            100  3.245566
#> 3418    300.00   TRUE       300.00             73  4.552486
#> 664     300.00   TRUE       300.00             12  2.885027
#> 46065   300.00   TRUE       251.11            100  3.482407
#> 31121   300.00   TRUE       300.00             45  3.194113
#> 27695   300.00   TRUE       300.00             41  3.042574
#> 13146   300.00   TRUE       300.00              2  3.278424
#> 16370   300.00   TRUE       300.00              8  4.119860
#> 45500   300.00   TRUE         0.00              2  3.626296
#> 26010    69.52   TRUE       166.56              2  3.647295
#> 41783   300.00   TRUE       300.00            100  2.973405
#> 40270   300.00   TRUE       300.00              5  3.139431
#> 21680   104.12   TRUE       110.76              4  3.591103
#> 31834   215.68   TRUE         0.00            100  3.354565
#> 39722   300.00   TRUE       300.00            100  2.906387
#> 18958   300.00   TRUE       300.00              5  2.674746
#> 47320   300.00   TRUE       300.00             43  2.536741
#> 18440   300.00   TRUE       300.00            100  4.477051
#> 31117   300.00   TRUE       291.23            100  3.277641
#> 18348   300.00   TRUE       300.00            100  3.055854
#> 7086    300.00   TRUE       128.61             63  3.071308
#> 47702   300.00   TRUE       300.00             86  4.665675
#> 43201   300.00   TRUE       300.00            100  3.530273
#> 42201   300.00   TRUE       300.00            100  3.989241
#> 34887   300.00   TRUE       300.00            100  3.159244
#> 46138   300.00   TRUE       300.00              8  5.794410
#> 33875   300.00   TRUE         4.67              0  3.279136
#> 34615    17.57   TRUE       142.95             16  3.011436
#> 15432   300.00   TRUE       300.00            100  2.880810
#> 19832   300.00   TRUE       300.00            100  3.278299
#> 5572    300.00   TRUE       300.00            100  2.567470
#> 26428   300.00   TRUE         0.00            100  3.486483
#> 36292   300.00   TRUE        16.03            100  3.030038
#> 7485    300.00   TRUE       300.00            100  2.627640
#> 47244   300.00   TRUE       300.00              2  2.752448
#> 11352   135.89   TRUE        82.11              4  4.291317
#> 37439   300.00   TRUE       300.00            100  3.539988
#> 332      94.82   TRUE         0.00             90  3.079334
#> 38668   300.00   TRUE       300.00            100  3.022236
#> 16071   300.00   TRUE       300.00            100  2.201225
#> 22297   300.00   TRUE       300.00             11  2.548891
#> 4667     24.78   TRUE       172.24             38  3.119102
#> 48650   300.00   TRUE       300.00            100  3.027918
#> 14494   300.00   TRUE       300.00             28  2.535469
#> 16065   300.00   TRUE       162.92            100  3.818858
#> 30363   300.00   TRUE       300.00            100  2.929656
#> 28588   300.00   TRUE       300.00             78  5.209992
#> 13278    91.73   TRUE       120.78             34  3.560110
#> 14779    55.08   TRUE         0.00              1  2.897760
#> 43834    20.21   TRUE       150.14              2  3.331390
#> 9864    300.00   TRUE       300.00            100  2.722792
#> 23832   300.00   TRUE       300.00              8  2.676567
#> 39466   300.00   TRUE       300.00            100  3.060187
#> 13072   300.00   TRUE       300.00            100  4.477051
#> 15211   300.00  FALSE       300.00            100  2.844665
#> 24836   300.00  FALSE       300.00             95  2.679300
#> 16464   300.00  FALSE       300.00            100  3.004540
#> 30267   300.00  FALSE       300.00            100  2.781107
#> 1488    300.00  FALSE       300.00            100  2.930251
#> 43243   300.00  FALSE       300.00            100  2.922424
#> 7538    300.00  FALSE       300.00             87  2.433645
#> 13286   300.00  FALSE       300.00             76  2.499561
#> 4007    300.00  FALSE       300.00             68  2.847871
#> 34616   300.00  FALSE       300.00            100  2.321589
#> 43403   300.00  FALSE       111.88             86  2.788585
#> 37029   300.00  FALSE       237.21            100  2.916688
#> 8895    300.00  FALSE       300.00             90  2.432641
#> 47059   300.00  FALSE       300.00            100  3.023816
#> 34105   300.00  FALSE       300.00             26  3.059266
#> 14148   300.00  FALSE       214.75            100  2.760532
#> 12044   300.00  FALSE       300.00            100  3.304782
#> 21070   300.00  FALSE         0.00             10  3.400551
#> 43222   300.00  FALSE       300.00            100  2.404946
#> 34532   300.00  FALSE       300.00             92  2.583601
#> 15008   300.00  FALSE       300.00             48  2.625829
#> 28286   300.00  FALSE       300.00            100  2.790556
#> 32430   300.00  FALSE       300.00            100  2.398827
#> 7986    275.06  FALSE         0.00             65  3.156065
#> 2999    300.00  FALSE       300.00            100  2.883559
#> 18308   300.00  FALSE       300.00             55  4.224567
#> 4943    300.00  FALSE       300.00            100  3.019161
#> 18246    18.49  FALSE         0.00             29  3.220009
#> 6732    300.00  FALSE       300.00            100  2.948550
#> 10981   300.00  FALSE       300.00            100  5.155983
#> 1434    300.00  FALSE       300.00             35  2.955995
#> 41873   300.00  FALSE       300.00            100  2.265064
#> 14881   300.00  FALSE       300.00            100  2.948550
#> 9526    300.00  FALSE       300.00            100  2.898919
#> 39883   233.81  FALSE       219.23            100  2.474870
#> 47245   300.00  FALSE       150.03            100  2.509309
#> 47629   300.00  FALSE       300.00            100  3.627607