Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem    slope      hcurv      vcurv       carea
#> 37912 712802.5 9559952 1838.40 52.10134    0.00183   -0.09203    634.3320
#> 36052 714842.5 9558892 2483.27 68.78989   -0.04921   -0.12438    753.6304
#> 15762 714922.5 9558862 2535.97 54.91056    0.01305   -0.00555    430.4604
#> 30059 713832.5 9559662 2288.02 18.27793    0.01917    0.00333    258.6537
#> 19963 714612.5 9560552 2031.70 26.84823    0.00046    0.00324    861.3537
#> 48187 712722.5 9560152 1835.29 29.32799   -0.00333   -0.01767   3321.6113
#> 24129 714602.5 9560542 2038.30 25.58486    0.00065    0.00005    769.3426
#> 38411 713552.5 9560652 1851.63 46.00507   -0.00891   -0.00949   6217.8213
#> 32622 714972.5 9557762 2687.38 30.55641   -0.00730   -0.01491   1844.4353
#> 43669 712392.5 9560162 2001.90 47.28392   -0.00284    0.01804    628.5153
#> 20151 714602.5 9559922 2184.66 48.34160   -0.00979   -0.01542    674.0147
#> 47255 713552.5 9559962 2191.77 33.57876   -0.00068    0.00588   1295.1610
#> 37270 714852.5 9561032 1838.96 35.15898   -0.00854    0.01264   5057.1431
#> 7349  714032.5 9558502 2402.95 31.35282    0.01924    0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.15936   -0.01607    0.00677   1343.5006
#> 38410 713712.5 9561172 1776.08 36.31922   -0.02678   -0.04272   4955.5586
#> 48498 714972.5 9557972 2681.66 38.34692    0.00528    0.01392   1255.0884
#> 22283 713592.5 9560762 1832.60 36.01040    0.00448   -0.00328    592.3326
#> 18304 714582.5 9558932 2463.25 22.19467   -0.02131    0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.10784   -0.00017   -0.00032    994.6740
#> 30125 715162.5 9558992 2438.18 39.85666    0.00471   -0.01382    560.8309
#> 38070 715212.5 9559222 2342.41 55.67717    0.00426   -0.00367    793.9665
#> 47210 714952.5 9557612 2744.35 40.88455   -0.00493    0.00733    598.7983
#> 6808  715192.5 9559492 2219.03 21.81766    0.00713   -0.00173   8967.1924
#> 8288  714842.5 9557712 2679.04 28.93953    0.00902    0.01358    802.7622
#> 40828 714842.5 9557632 2698.65 17.53652    0.00490   -0.00450    873.0448
#> 32821 715022.5 9559952 2208.33 43.41645   -0.00643    0.00314   2360.1501
#> 4477  715432.5 9559392 2325.25 37.02912   -0.00521   -0.00739    558.2286
#> 47220 715052.5 9557752 2726.93 32.72162   -0.00664    0.00694   2123.4473
#> 28230 715182.5 9558102 2791.83 39.07171   -0.00669   -0.00950    950.3138
#> 27957 715012.5 9559332 2306.35 20.48038    0.00225    0.00325    616.8652
#> 1635  713102.5 9559052 2172.87 31.48403   -0.00064   -0.00916   2009.1686
#> 27013 713062.5 9559782 1976.81 41.63226    0.01414    0.03416   1204.8629
#> 12084 712832.5 9559992 1872.46 56.05475   -0.03184    0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.68879   -0.00440   -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.15647   -0.01743    0.01313    783.4548
#> 46756 713372.5 9559192 2120.47 46.00507    0.03696    0.02514    297.7842
#> 12868 712532.5 9560332 1913.92 61.93731   -0.03543   -0.05146   1347.6224
#> 12460 712762.5 9559622 1906.41 24.26190    0.01047   -0.00887    366.8537
#> 40738 714972.5 9557642 2745.48 42.54269   -0.00294   -0.00096    520.5570
#> 24100 715132.5 9558992 2420.18 29.21397   -0.03229   -0.00511 361979.2812
#> 34563 715022.5 9558162 2675.70 18.88985   -0.01993    0.01373  16031.7500
#> 47091 713382.5 9559182 2122.71 48.03793    0.04527    0.04613    289.6113
#> 17036 712832.5 9560582 1951.38 27.64579    0.00060    0.00270   4057.5637
#> 17423 715592.5 9558562 2654.69 33.21952   -0.00034    0.00294    739.5768
#> 21952 715352.5 9560172 2013.44 40.48806    0.02409    0.02771    674.3859
#> 2718  715062.5 9558222 2719.56 44.75316   -0.00119   -0.01560    977.4719
#> 4617  714292.5 9559312 2342.95 32.39217    0.03504    0.05825    162.6688
#> 462   713712.5 9561182 1785.08 47.71707   -0.03260   -0.03310   2833.9001
#> 375   715622.5 9558922 2510.87 37.36659   -0.01017    0.00197    711.1091
#> 39520 715342.5 9557622 2889.48 41.64888   -0.01282    0.00542    860.8840
#> 30406 715312.5 9560152 2029.10 34.13740    0.00953   -0.02283    590.1536
#> 27875 714832.5 9561032 1838.22 41.21228   -0.01160    0.01690   5121.6924
#> 32909 712812.5 9560452 1883.85 38.25124   -0.00221    0.00121   4159.6738
#> 10507 714892.5 9559282 2374.45 42.84235   -0.01057    0.02427    380.9472
#> 33350 713132.5 9560672 1896.89 24.69505    0.00306   -0.00436   2602.7390
#> 25700 715172.5 9559272 2324.49 35.53484   -0.00529    0.02009   1247.7061
#> 28845 714892.5 9559272 2374.42 45.75011   -0.00209    0.02739    365.7626
#> 46065 713072.5 9559162 2109.39 16.26455   -0.00195   -0.00656   3036.7383
#> 16370 713852.5 9558632 2296.97 34.62441   -0.01972    0.03512  13178.3223
#> 17431 712802.5 9560022 1832.71 41.52512   -0.00478   -0.07132  42970.7891
#> 34192 712702.5 9560102 1867.01 35.47468    0.00629   -0.00469   1021.5407
#> 40270 713312.5 9559012 2186.56 30.93571   -0.00771   -0.00599   1378.5760
#> 7385  713752.5 9560122 2064.61 31.77108   -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.02394    0.00538    0.00112    745.3315
#> 37769 714412.5 9558562 2382.08 39.87672   -0.04601   -0.08389  47022.9688
#> 37714 713192.5 9560692 1885.05 18.03442    0.00188    0.00832   3022.3462
#> 7041  713472.5 9559092 2137.03 18.34897   -0.10372   -0.00008 825875.6875
#> 18348 714042.5 9558882 2341.41 44.06389   -0.00338    0.00728   1137.2450
#> 47702 714692.5 9559712 2200.51 33.09863   -0.13038   -0.04212  46310.0195
#> 46138 715202.5 9559582 2194.40 18.86005   -0.02143   -0.05078 622888.2500
#> 37720 715722.5 9557962 2820.42 45.07631   -0.01474    0.01334   4905.2002
#> 33875 712832.5 9559852 1892.81 37.55452   -0.01175    0.04195   1901.6743
#> 26101 715562.5 9558682 2594.87 32.41108    0.00079    0.00321   1170.5904
#> 1036  713042.5 9558892 2205.38 31.50981   -0.00428   -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.52758   -0.00193    0.02812   1423.7578
#> 17058 714032.5 9558502 2402.95 31.35282    0.01924    0.02606    564.6876
#> 36292 712862.5 9559602 1915.77 37.97507    0.00511   -0.01801   1071.6141
#> 11352 713002.5 9560562 1858.95 39.80395   -0.00442   -0.00118  19557.6621
#> 12629 714712.5 9558872 2533.15 37.02568    0.04250   -0.00970   1443.9200
#> 28173 714562.5 9560382 2048.26 27.68589   -0.00124   -0.00456   1531.4274
#> 38668 715582.5 9558242 2735.55 35.44603    0.01003    0.01196   1052.5341
#> 16071 714562.5 9560062 2166.69 28.69602    0.02281    0.04759    158.9370
#> 20426 714702.5 9559712 2208.61 43.04632   -0.08153   -0.02417  34908.6562
#> 20171 715732.5 9557962 2815.53 48.65787   -0.01341    0.00171   4425.5410
#> 46792 713862.5 9559672 2272.03 35.46494    0.00560    0.00610    454.8103
#> 42640 714332.5 9558522 2403.10 38.87118   -0.00212    0.02732   4237.6934
#> 11057 714962.5 9557642 2737.48 42.68593   -0.00488   -0.00632    620.0229
#> 35458 715242.5 9558442 2623.01 38.95483   -0.00770   -0.03400   1982.7782
#> 48461 714672.5 9559272 2316.84 40.38149   -0.01966   -0.02234    849.0547
#> 33142 715652.5 9558212 2740.94 14.38124   -0.00465   -0.01036   2505.8369
#> 31491 712832.5 9558892 2091.39 29.70041    0.00186    0.00063   2430.2278
#> 41310 714452.5 9560402 2092.27 12.46298   -0.00195   -0.00205   1112.2626
#> 39466 714942.5 9557772 2680.35 16.75329    0.00166   -0.01265   1148.6487
#> 16208 712882.5 9560332 1846.34 48.62521    0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.34766   -0.02196    0.00336   2806.2644
#> 25584 715062.5 9558532 2675.93 37.68859    0.00534    0.01476   1275.9446
#> 34250 713822.5 9559142 2316.36 33.75982   -0.02284    0.01085   2229.5105
#> 14717 715382.5 9558062 2799.05 50.59045   -0.00812   -0.02208    951.3401
#> 20067 714242.5 9560442 2125.23 38.64027    0.00363   -0.00363    389.2691
#> 26963 715702.5 9557882 2880.32 43.73330   -0.01174   -0.01466   2633.9666
#> 46930 713642.5 9560602 1945.03 34.50982    0.00165    0.00875   1165.3953
#> 12716 713962.5 9557762 2402.60 18.93224    0.03480    0.01500    538.8209
#> 9665  713362.5 9558812 2275.28 31.49492    0.01306   -0.00626    450.7346
#> 31497 714032.5 9560672 2060.77 16.21700    0.00973   -0.00943    371.5514
#> 32849 714302.5 9558222 2520.01 32.15726   -0.00846    0.00216   1412.2295
#> 29944 713562.5 9558952 2209.55 17.33999    0.01432   -0.00362    283.7090
#> 44668 712612.5 9560472 1972.24 27.50713    0.00601   -0.00011    712.5228
#> 41753 715722.5 9558192 2771.39 26.08505   -0.00196   -0.00175    725.1355
#> 38689 714622.5 9559342 2264.90 42.64811   -0.00844    0.00514    897.9290
#> 42934 713832.5 9558672 2270.74 27.61198   -0.03193   -0.00267  31588.1953
#> 6143  712762.5 9559022 2096.24 29.89636    0.00358    0.02422    967.9738
#> 36245 713142.5 9560572 1827.82 43.67600   -0.00410    0.00400   4126.7812
#> 33370 714202.5 9558962 2472.93 32.64141   -0.00307    0.02997    230.8641
#> 13737 714272.5 9559622 2257.05 55.45716   -0.03224   -0.02036    476.9504
#> 3733  713762.5 9559632 2297.63 34.21475   -0.01338    0.00518   1137.0179
#> 33410 715342.5 9560452 1942.88 29.09308   -0.00305   -0.00705    968.9738
#> 36721 714092.5 9559752 2301.81 28.93322    0.03418    0.00513    161.9878
#> 330   714232.5 9560362 2165.00 18.06078    0.01142    0.00868    235.8511
#> 46509 712512.5 9559532 1860.09 89.98875   -0.00270 1018.51947    842.1514
#> 44715 714932.5 9560132 2195.78 30.39369    0.03479    0.02251    168.1733
#> 36889 714302.5 9559802 2243.60 21.16621    0.01923    0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.93765   -0.00070   -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.91241 -100.76649  100.79199    107.2986
#> 1281  715102.5 9560032 2120.64 34.64905   -0.00930   -0.00360   4288.7329
#> 14671 712672.5 9560372 1872.21 39.19719   -0.00873   -0.01417   1520.6630
#> 9007  714952.5 9560902 1863.75 32.77032    0.00074   -0.02904    608.3561
#> 14763 713462.5 9560852 1830.40 47.73999    0.01354    0.00907   1281.5044
#> 26283 715932.5 9557592 3075.48 32.76688   -0.00375    0.00215    917.5296
#> 24446 714742.5 9560082 2200.60 28.72180   -0.00598    0.02208    842.4825
#> 10981 713832.5 9560022 2125.33 34.84844   -0.06909   -0.01151 143213.1562
#> 11394 713472.5 9561232 1889.48 34.31444    0.00108   -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.99129    0.00253   -0.01233    968.7780
#> 18504 713372.5 9560672 1786.89  6.47385   -0.00278   -0.01172 111973.8906
#> 16443 714472.5 9559532 2222.55 16.74412    0.03263   -0.02012    305.3624
#> 28526 715152.5 9560732 1871.53 29.53426   -0.00163   -0.00377   1727.6055
#> 35257 713822.5 9557982 2350.80 29.30794    0.01549   -0.00169    649.9975
#> 39883 712732.5 9560842 2082.69 25.09097    0.00732    0.01858    298.4490
#> 36141 714222.5 9559422 2316.64 16.27716   -0.07627   -0.00124  25826.6914
#> 11570 715572.5 9558482 2685.72 16.81402    0.00282    0.00458    485.7418
#> 32342 714862.5 9559622 2294.74 36.96724    0.01021    0.00109    915.4989
#> 14645 715072.5 9560262 2034.04 38.16930   -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.47086   -0.00011   -0.00419    323.0790
#> 126   715042.5 9561262 1733.49 30.11065   -0.00288    0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.12459   -0.00166    0.00156    758.4840
#> 13548 715812.5 9558122 2821.65 15.50825    0.02751    0.00388    180.4414
#> 46567 714472.5 9560352 2080.13 21.38164    0.01336    0.01403    609.8665
#> 30296 713322.5 9559782 2042.96 52.62102   -0.02110   -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.85208    0.04456   -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.34975   -0.00771    0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 37912 30.29457046   300.00   TRUE         9.15              2  2.802317
#> 36052 53.74516006   300.00   TRUE       300.00            100  2.877158
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 30059 20.19217862   300.00   TRUE       300.00             41  2.412719
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 24129 24.12897163   300.00   TRUE       300.00             89  2.886120
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43669 28.47657538    24.22   TRUE         0.00             89  2.798316
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 48498 33.61829863   300.00   TRUE       300.00            100  3.098674
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 6808  28.56595679   300.00   TRUE       300.00             40  3.952656
#> 8288  22.71777658   300.00   TRUE       300.00             83  2.904587
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 27013 30.91852150   300.00   TRUE       168.05             18  3.080938
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 40738 28.96015175   300.00   TRUE       300.00            100  2.716468
#> 24100 34.29782657   300.00   TRUE       300.00            100  5.558684
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 17423 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 39520 35.89408699   300.00   TRUE       300.00            100  2.934945
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 46065 29.77088703   300.00   TRUE       251.11            100  3.482407
#> 16370 31.15859081   300.00   TRUE       300.00              8  4.119860
#> 17431 25.32530750   300.00   TRUE         1.11              5  4.633173
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 18348 41.01060010   300.00   TRUE       300.00            100  3.055854
#> 47702 33.66642708   300.00   TRUE       300.00             86  4.665675
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 11352 27.99815562   135.89   TRUE        82.11              4  4.291317
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 38668 32.92272787   300.00   TRUE       300.00            100  3.022236
#> 16071 29.62134505   300.00   TRUE       300.00            100  2.201225
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 34250 30.89273840   300.00  FALSE       300.00            100  3.348210
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 36245 25.85357459   110.19  FALSE        81.50             20  3.615611
#> 33370 26.41908393   300.00  FALSE       300.00            100  2.363356
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 46509 45.57879260   300.00  FALSE       162.53            100  2.925390
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 1281  35.39389484   300.00  FALSE       300.00            100  3.632329
#> 14671 37.88225054   123.04  FALSE        75.11              0  3.182033
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 26283 41.51079226   300.00  FALSE       300.00            100  2.962620
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 35257 30.88185220   300.00  FALSE       300.00             79  2.812912
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 11570 16.60718169   300.00  FALSE       300.00            100  2.686405
#> 32342 37.61983587   300.00  FALSE       300.00             81  2.961658
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491