Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 21788 715392.5 9560172 1988.71 40.476603  -0.01919   -0.04051 351155.1250
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 27864 712992.5 9560672 1926.38 27.224090  -0.00199    0.00659   3553.7166
#> 13748 715232.5 9559542 2212.12 26.123438   0.00023   -0.00043   1806.4552
#> 48187 712722.5 9560152 1835.29 29.327991  -0.00333   -0.01767   3321.6113
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 9669  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 14456 713542.5 9559972 2184.39 35.241488  -0.00841   -0.00010   1527.1028
#> 43327 713962.5 9561082 1839.11 63.361684  -0.04951   -0.00529   3282.1186
#> 16811 714902.5 9559262 2362.67 50.703327  -0.01407    0.00547    519.4228
#> 33792 714652.5 9558172 2598.36 59.458695  -0.03525   -0.11024    394.8751
#> 29420 713592.5 9560772 1826.82 34.926934  -0.00070   -0.00530   9000.4356
#> 15109 713872.5 9559602 2288.64 25.185506   0.00602    0.01518    854.4269
#> 30125 715162.5 9558992 2438.18 39.856663   0.00471   -0.01382    560.8309
#> 4827  714962.5 9559882 2274.42 29.244912   0.00051    0.00889    362.8363
#> 40828 714842.5 9557632 2698.65 17.536519   0.00490   -0.00450    873.0448
#> 27708 715222.5 9557782 2844.65 29.542850   0.03206    0.02773    179.3077
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739    558.2286
#> 34280 713652.5 9560702 1896.37 41.273588   0.00382    0.00088   2322.1746
#> 23031 714042.5 9558492 2404.94 28.496056   0.01205    0.02205    629.4165
#> 46125 713732.5 9560062 2115.18 37.593480  -0.04616    0.01276   9614.6924
#> 35142 714792.5 9558862 2575.17 25.664499   0.06765    0.00114    150.1498
#> 22735 714492.5 9559772 2170.30 39.688786  -0.00440   -0.02651    419.0927
#> 15990 712852.5 9559982 1894.77 37.694320  -0.01545    0.02405   5681.2412
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 6557  713312.5 9559022 2181.69 27.634327  -0.00669   -0.00991   1403.0193
#> 17594 714832.5 9561052 1818.04 46.167220  -0.01353    0.00063   5035.0000
#> 6310  713802.5 9559752 2238.38 19.651306  -0.02379    0.01338  10787.6299
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 9704  714642.5 9559652 2249.12 37.143135  -0.00759   -0.00400    608.6523
#> 2718  715062.5 9558222 2719.56 44.753160  -0.00119   -0.01560    977.4719
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 21053 712862.5 9559992 1901.01 30.950607  -0.00762    0.01252   5761.3486
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 38862 712952.5 9559892 1950.86 37.721822  -0.00048   -0.00332   1418.6622
#> 32909 712812.5 9560452 1883.85 38.251235  -0.00221    0.00121   4159.6738
#> 18265 715522.5 9558782 2539.23 35.856272  -0.00458   -0.00401   1722.6481
#> 32280 712572.5 9560182 1956.86 50.678117  -0.02554    0.03304   1569.7313
#> 10364 714762.5 9561002 1856.21 25.274887  -0.00845    0.00454   3476.0818
#> 31561 713302.5 9559192 2092.11 36.527778   0.00406   -0.00055    773.0359
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.235780   0.00927    0.01643    670.2289
#> 7064  715592.5 9558662 2597.69 34.222196   0.00099    0.00240   1201.4724
#> 45500 712812.5 9560052 1852.16 56.770377   0.00904   -0.02013   4229.5708
#> 26010 712852.5 9560542 1928.25 30.978109   0.00165    0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 34192 712702.5 9560102 1867.01 35.474682   0.00629   -0.00469   1021.5407
#> 5239  714832.5 9560992 1859.70 19.960131  -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.237093  -0.00472   -0.00228   2262.3796
#> 41721 715182.5 9557562 2784.85 31.789990  -0.00588    0.01448   1258.8196
#> 7385  713752.5 9560122 2064.61 31.771083  -0.02605   -0.03594  16587.9980
#> 18897 712802.5 9559882 1851.71 50.234648  -0.02627   -0.05393   8301.6094
#> 37714 713192.5 9560692 1885.05 18.034420   0.00188    0.00832   3022.3462
#> 39166 712832.5 9560142 1922.46 39.660139   0.01469    0.02940    987.8085
#> 12772 715172.5 9557642 2759.42 34.805467  -0.00567   -0.03793    864.9472
#> 20368 715252.5 9560622 1865.83 33.572462   0.00025   -0.01245   6353.3179
#> 31117 712892.5 9558922 2129.84 30.550683   0.00521   -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.018643  -0.00589    0.00849   1120.5961
#> 7086  712762.5 9560962 2022.00 44.913589  -0.01259   -0.00281   1178.4420
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 27915 715232.5 9559552 2208.93 25.480579   0.00171   -0.00261   1627.8989
#> 43245 713722.5 9560012 2153.38 34.191256  -0.01263    0.01322   4605.1411
#> 30933 714672.5 9561082 1798.37 54.810289   0.00542   -0.00722   1373.4147
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 36292 712862.5 9559602 1915.77 37.975070   0.00511   -0.01801   1071.6141
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 39325 714142.5 9558592 2357.72 40.792876  -0.00246    0.00396    704.7076
#> 28173 714562.5 9560382 2048.26 27.685894  -0.00124   -0.00456   1531.4274
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 2801  715252.5 9559612 2199.28 24.403418   0.03051    0.02190    264.6404
#> 22297 714642.5 9558172 2606.18 59.469008  -0.00495   -0.03756    353.9086
#> 14192 715162.5 9558122 2772.18 44.462671  -0.02970    0.00200   1305.6691
#> 750   712812.5 9560032 1847.09 54.959258  -0.00767   -0.03713   6379.7974
#> 1997  714752.5 9561012 1853.19 26.566334  -0.00693    0.01133   2397.2681
#> 7409  714772.5 9559122 2404.90 30.215502  -0.01016   -0.00294    734.9908
#> 48650 715592.5 9558612 2626.74 31.876507   0.00010    0.00140   1066.3957
#> 14494 715332.5 9558792 2557.22 20.720446   0.01267    0.00483    343.1384
#> 44612 714932.5 9558822 2578.80 38.255246   0.01257    0.00044    288.5468
#> 9606  714392.5 9559172 2353.05 30.783303  -0.00185   -0.01185    953.8614
#> 42617 714092.5 9558392 2469.21 42.799947  -0.01397    0.00526    706.1463
#> 46627 714852.5 9558932 2459.24 53.445503   0.02451   -0.03921    525.3245
#> 7771  715232.5 9560922 1816.88 37.180377   0.00124    0.01166   1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 12074 714562.5 9560372 2045.17 26.287304  -0.00144   -0.00856   1789.1596
#> 41408 713172.5 9559612 2056.81 22.272588  -0.00166    0.01476   1404.4454
#> 41310 714452.5 9560402 2092.27 12.462978  -0.00195   -0.00205   1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122   0.00840    0.00189    556.5211
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 4785  713792.5 9560472 2055.34 33.052789  -0.00586    0.00067   1123.7318
#> 10145 713782.5 9560932 1865.62 23.595548  -0.00203    0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.625209   0.00159   -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.347660  -0.02196    0.00336   2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927  -0.00849    0.00150    503.5590
#> 25584 715062.5 9558532 2675.93 37.688591   0.00534    0.01476   1275.9446
#> 16464 715252.5 9559982 2143.24 23.185883  -0.00298   -0.00712   1010.5097
#> 31071 713472.5 9558462 2298.33 29.153239   0.00156   -0.00116   1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695  -0.00124    0.00204    853.2902
#> 14717 715382.5 9558062 2799.05 50.590454  -0.00812   -0.02208    951.3401
#> 32869 715332.5 9558892 2533.17 20.635075  -0.00242    0.01132    737.5982
#> 30267 714442.5 9557512 2612.30 28.765919  -0.00458    0.00479    604.0975
#> 8066  715842.5 9558732 2636.38 33.717420  -0.00494    0.01064    683.2795
#> 12716 713962.5 9557762 2402.60 18.932244   0.03480    0.01500    538.8209
#> 37832 712552.5 9559192 1998.10 41.550899  -0.00195   -0.00995   1026.7981
#> 28162 714062.5 9561152 1785.47 48.444727   0.02309   -0.03409   1575.9425
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 42329 715852.5 9558702 2651.92 22.870756   0.00321    0.00548    532.8210
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 40139 712852.5 9560072 1907.89 43.704075   0.00001    0.00759   1880.4403
#> 32849 714302.5 9558222 2520.01 32.157256  -0.00846    0.00216   1412.2295
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 29944 713562.5 9558952 2209.55 17.339995   0.01432   -0.00362    283.7090
#> 8413  713242.5 9561232 1967.45 38.194513   0.00236    0.00483   1097.3082
#> 14800 712952.5 9558682 2120.89 38.273581  -0.00022    0.00662   1269.2650
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 20135 714772.5 9560232 2107.52 37.850738  -0.00782   -0.00048   1550.6704
#> 29835 715402.5 9558512 2610.14 27.118092  -0.01322   -0.00759   4131.5356
#> 26389 713002.5 9559712 1926.02 24.037298   0.00586    0.00654   1576.2065
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 7983  713512.5 9560432 1907.38 50.019216   0.03310   -0.07650    766.6254
#> 44715 714932.5 9560132 2195.78 30.393692   0.03479    0.02251    168.1733
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 24819 714032.5 9560882 1981.86 33.284264   0.00804   -0.00165    805.1403
#> 5861  713882.5 9559182 2350.43 27.393685  -0.02608    0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570   0.00537   -0.00577    423.5598
#> 3868  714542.5 9560622 2024.15 29.447166   0.00360    0.00750   1892.8978
#> 12300 712372.5 9560062 1903.64 50.036977   0.00343    0.01177   1558.2231
#> 44216 715772.5 9558402 2710.44 30.444685  -0.00070   -0.00050   1471.4346
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 24446 714742.5 9560082 2200.60 28.721801  -0.00598    0.02208    842.4825
#> 19767 714272.5 9558052 2431.67 55.724729   0.00361   -0.01841    756.0055
#> 37783 713832.5 9557892 2405.28 40.886841   0.01572    0.03148    481.8292
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.002187   0.00203    0.00238    787.0053
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709   1181.2555
#> 9320  713872.5 9559262 2378.76 29.858550   0.00309    0.01011    501.7867
#> 37254 713482.5 9560892 1823.73 39.882446   0.01014   -0.00884    672.2466
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 31440 714842.5 9558392 2704.41 25.801435   0.00992   -0.01052    451.7209
#> 36141 714222.5 9559422 2316.64 16.277158  -0.07627   -0.00124  25826.6914
#> 24962 715152.5 9560472 1967.37 28.845560   0.00298    0.00222    991.8300
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 41373 715332.5 9560862 1873.10 29.124591  -0.00166    0.00156    758.4840
#> 30296 713322.5 9559782 2042.96 52.621017  -0.02110   -0.02931   1812.3820
#> 40043 713322.5 9560562 1802.09 39.852079   0.04456   -0.01586    941.0288
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 14456 20.06899269   300.00   TRUE       247.02            100  3.183868
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 15109 22.63527065   300.00   TRUE       300.00             40  2.931675
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 46125 25.65705007   300.00   TRUE       300.00            100  3.982935
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 15990 34.14885755   300.00   TRUE         0.00              0  3.754443
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 6310  23.82644991   300.00   TRUE       300.00            100  4.032926
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 38862 34.90000522   300.00   TRUE       104.48             92  3.151879
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 7064  29.13261205   300.00   TRUE       300.00            100  3.079714
#> 45500 34.95615508   300.00   TRUE         0.00              2  3.626296
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 41721 29.97944367   300.00   TRUE       300.00            100  3.099963
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 18897 31.78827143   300.00   TRUE        20.00              5  3.919162
#> 37714 30.73173726    13.57   TRUE       170.00              4  3.480344
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 12772 37.92006576   300.00   TRUE       300.00            100  2.936990
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 43245 24.44581729   300.00   TRUE       300.00            100  3.663243
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 22297 46.64965072   300.00   TRUE       300.00             11  2.548891
#> 14192 40.58661133   300.00   TRUE       300.00            100  3.115833
#> 750   34.34710094   300.00   TRUE         0.00              0  3.804807
#> 1997  25.23592609   168.72   TRUE         0.00              5  3.379717
#> 7409  25.37057117   300.00   TRUE       300.00            100  2.866282
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 10145 27.41603050   276.16  FALSE        91.33            100  3.227418
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 25584 27.14674033   300.00  FALSE       300.00            100  3.105832
#> 16464 35.04610945   300.00  FALSE       300.00            100  3.004540
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 32869 14.60011054   300.00  FALSE       300.00             78  2.867820
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 28162 18.66982975   130.00  FALSE        75.00            100  3.197540
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 8413  39.88989465    80.73  FALSE         0.00            100  3.040329
#> 14800 31.52299197   300.00  FALSE       300.00            100  3.103552
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 44715 28.45021932   300.00  FALSE       300.00            100  2.225757
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 3868  26.14750194   300.00  FALSE       300.00            100  3.277127
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 19767 42.01041145   300.00  FALSE       300.00            100  2.878525
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 9320  18.13927084   300.00  FALSE       300.00             93  2.700519
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 31440 24.03500655   300.00  FALSE       300.00             96  2.654870
#> 36141 27.30602260   300.00  FALSE       300.00             85  4.412069
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 30296 40.15861186   300.00  FALSE       300.00             68  3.258250
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983