Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190 -0.00501 -0.00649    1399.2329
#> 44302 714932.5 9557982 2649.56 37.300698  0.01633 -0.01813    1130.5791
#> 47223 715302.5 9557472 2857.45 39.533515 -0.01021 -0.01579    1318.7560
#> 28223 712742.5 9560482 1927.69 34.667957 -0.00292  0.00712    3636.5303
#> 16749 715072.5 9558242 2745.27 49.053782  0.00097  0.01483     566.7674
#> 30059 713832.5 9559662 2288.02 18.277927  0.01917  0.00333     258.6537
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 11088 715532.5 9558762 2549.28 38.050127 -0.00373 -0.00057    1630.3259
#> 15843 714042.5 9558902 2332.30 50.097711 -0.00858 -0.00292    1233.9299
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 8869  715062.5 9559392 2278.56 41.401930  0.00310  0.01440    1088.8639
#> 14456 713542.5 9559972 2184.39 35.241488 -0.00841 -0.00010    1527.1028
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 43669 712392.5 9560162 2001.90 47.283915 -0.00284  0.01804     628.5153
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 37270 714852.5 9561032 1838.96 35.158982 -0.00854  0.01264    5057.1431
#> 25690 715042.5 9557572 2752.67 32.713598  0.00194  0.00526     539.8119
#> 38410 713712.5 9561172 1776.08 36.319222 -0.02678 -0.04272    4955.5586
#> 48055 712862.5 9559972 1902.48 29.016875 -0.00720  0.01250    3686.1240
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 34482 715152.5 9557642 2753.52  4.995619 -0.01126 -0.02044  129412.3984
#> 38070 715212.5 9559222 2342.41 55.677174  0.00426 -0.00367     793.9665
#> 6808  715192.5 9559492 2219.03 21.817660  0.00713 -0.00173    8967.1924
#> 45575 715432.5 9559452 2305.28 24.234396  0.00993 -0.00413     364.9788
#> 38077 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 13847 715052.5 9559232 2337.87 26.114843 -0.01452 -0.01437    1155.1469
#> 3466  713232.5 9560612 1812.72 51.566202 -0.01804 -0.01596    5296.9014
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 47429 714962.5 9557832 2705.12 32.057562  0.02224  0.00177     291.5937
#> 37828 713482.5 9558652 2330.58 20.445426  0.00606 -0.00556     660.3690
#> 36931 712852.5 9558892 2102.80 30.535786  0.00140 -0.00070    2562.7559
#> 46201 714832.5 9561022 1845.94 34.573992 -0.00856  0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.933600  0.00321  0.01169     350.1799
#> 18751 714882.5 9559182 2381.81 39.299175  0.00066 -0.00036     516.4709
#> 13699 712902.5 9559902 1917.21 30.487657 -0.00093 -0.00907    1943.4092
#> 47884 715032.5 9557712 2738.31 28.289218  0.00006  0.00175     804.5679
#> 13134 714652.5 9559632 2265.33 49.599810 -0.03552 -0.02058     431.4362
#> 37141 714382.5 9559822 2219.36 10.106403  0.01661  0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.137608 -0.00613  0.00903    2360.1536
#> 22332 715482.5 9558922 2475.28 19.936639 -0.02617  0.00307  147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 5084  714912.5 9559302 2359.77 47.968027  0.00107  0.02373     527.3528
#> 38862 712952.5 9559892 1950.86 37.721822 -0.00048 -0.00332    1418.6622
#> 7466  712962.5 9560392 1819.03 18.402831 -0.00140 -0.02330    1456.9246
#> 39520 715342.5 9557622 2889.48 41.648875 -0.01282  0.00542     860.8840
#> 18265 715522.5 9558782 2539.23 35.856272 -0.00458 -0.00401    1722.6481
#> 13029 713542.5 9560242 2019.80 29.545142  0.00576 -0.00166     650.6246
#> 32735 714922.5 9559002 2414.49 40.064647  0.00062 -0.00262    1760.2162
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 31561 713302.5 9559192 2092.11 36.527778  0.00406 -0.00055     773.0359
#> 46065 713072.5 9559162 2109.39 16.264553 -0.00195 -0.00656    3036.7383
#> 16355 712962.5 9560212 1992.88 39.526066  0.01927 -0.00316     454.9674
#> 13146 713792.5 9558552 2341.40 29.825318 -0.00104  0.01664    1898.5588
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 16370 713852.5 9558632 2296.97 34.624413 -0.01972  0.03512   13178.3223
#> 2552  714942.5 9557652 2719.32 34.873649 -0.01238 -0.00851    1012.5963
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 41783 715572.5 9558652 2609.51 32.408212  0.00291 -0.00081     940.6002
#> 34192 712702.5 9560102 1867.01 35.474682  0.00629 -0.00469    1021.5407
#> 41721 715182.5 9557562 2784.85 31.789990 -0.00588  0.01448    1258.8196
#> 12137 712842.5 9558902 2097.88 30.098619  0.00024 -0.00474    2343.2100
#> 30268 714172.5 9558612 2328.84 54.950663  0.03743 -0.05242     507.9263
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 27244 714162.5 9558622 2316.00 36.751232 -0.02556 -0.11694  102301.4609
#> 31117 712892.5 9558922 2129.84 30.550683  0.00521 -0.00870    1895.1399
#> 7086  712762.5 9560962 2022.00 44.913589 -0.01259 -0.00281    1178.4420
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 15432 715632.5 9558922 2509.67 35.827624 -0.00444 -0.00017     759.9937
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 5572  715062.5 9558262 2754.60 39.911667  0.00707  0.01833     369.3771
#> 34164 714212.5 9558492 2421.45 51.329570  0.02247  0.01123     364.2903
#> 17058 714032.5 9558502 2402.95 31.352824  0.01924  0.02606     564.6876
#> 14545 713562.5 9559602 2222.17 54.239050  0.01011  0.00739     789.7573
#> 19756 715572.5 9558972 2485.95 38.812734  0.00546  0.01484     495.0292
#> 10617 713232.5 9560722 1879.95 26.497006  0.00258  0.00422    1159.2100
#> 2801  715252.5 9559612 2199.28 24.403418  0.03051  0.02190     264.6404
#> 49133 714162.5 9558462 2451.13 21.890426  0.00042  0.00957     495.4453
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 7409  714772.5 9559122 2404.90 30.215502 -0.01016 -0.00294     734.9908
#> 26058 713402.5 9560442 1880.12 31.377461  0.00730 -0.02610    1872.9276
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 28260 715312.5 9558302 2681.75 34.484038 -0.01125 -0.02685    2568.4631
#> 28588 715532.5 9559362 2299.84 11.144602 -0.04360 -0.03000  162178.1562
#> 37115 715612.5 9559202 2358.15  9.727677 -0.01020  0.00279   39495.1797
#> 17227 714812.5 9558892 2545.43 54.071746  0.00458  0.04291     293.4231
#> 24532 715202.5 9560922 1799.62  4.488551 -0.02312 -0.01848 1890804.2500
#> 14779 712502.5 9560202 2004.61 27.522601  0.00115  0.01015     790.2421
#> 35458 715242.5 9558442 2623.01 38.954828 -0.00770 -0.03400    1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 23832 714252.5 9560182 2188.12 30.647512  0.00427  0.02403     474.8613
#> 6470  713742.5 9558802 2260.79 30.680171  0.00000  0.00320     911.5911
#> 10145 713782.5 9560932 1865.62 23.595548 -0.00203  0.00642    1688.1782
#> 29431 714882.5 9559892 2258.03 21.917927 -0.00849  0.00150     503.5590
#> 36561 715292.5 9560802 1844.64 31.937240  0.00002 -0.01232     977.7320
#> 11288 712822.5 9560442 1873.42 39.609718 -0.00465 -0.00105    4432.8760
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 18286 714792.5 9557882 2650.57 28.253695 -0.00124  0.00204     853.2902
#> 20067 714242.5 9560442 2125.23 38.640274  0.00363 -0.00363     389.2691
#> 7517  714002.5 9561142 1772.85 26.189328  0.00815 -0.03345  665381.5625
#> 19493 714252.5 9560372 2156.93 24.879546  0.00825  0.00445     355.5078
#> 43426 713782.5 9560822 1876.35 23.823012 -0.00345 -0.00075    8141.7549
#> 37832 712552.5 9559192 1998.10 41.550899 -0.00195 -0.00995    1026.7981
#> 13512 715152.5 9557992 2778.81 43.389521 -0.00724  0.01464    4029.5947
#> 39193 715332.5 9560852 1872.90 30.290560 -0.00161  0.00171     756.8818
#> 39796 714802.5 9560372 2046.48 40.155747  0.00176  0.02973     215.5101
#> 1085  715722.5 9557532 3097.22 28.784254  0.02327  0.01833     300.0125
#> 39685 715332.5 9558102 2797.04 47.788118  0.00167  0.01133     828.8604
#> 36833 713322.5 9560572 1794.25 33.754090  0.00527 -0.02077    8759.9961
#> 34347 714742.5 9560142 2161.95 26.566907 -0.00978  0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633 -0.00007 -0.00853    1354.6001
#> 31497 714032.5 9560672 2060.77 16.216997  0.00973 -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.023226  0.01338 -0.00399    1601.4852
#> 31884 715782.5 9558612 2632.60 30.279674  0.00084 -0.00083    1309.2347
#> 8413  713242.5 9561232 1967.45 38.194513  0.00236  0.00483    1097.3082
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 29527 714332.5 9557502 2567.02 48.904813 -0.00452 -0.01298     436.1459
#> 33956 714082.5 9558712 2327.95 39.684203  0.02218  0.00902     218.7467
#> 6143  712762.5 9559022 2096.24 29.896365  0.00358  0.02422     967.9738
#> 1691  714222.5 9559022 2448.69 38.216858 -0.01292  0.00462     716.5762
#> 33370 714202.5 9558962 2472.93 32.641406 -0.00307  0.02997     230.8641
#> 13737 714272.5 9559622 2257.05 55.457158 -0.03224 -0.02036     476.9504
#> 29107 713172.5 9559272 2067.58 28.264008  0.02413  0.00528     355.3475
#> 45486 712762.5 9560172 1853.82 56.797306  0.01978  0.01282    2322.7944
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 21070 712832.5 9560062 1884.42 50.933656 -0.00342  0.02552    2515.0752
#> 36721 714092.5 9559752 2301.81 28.933223  0.03418  0.00513     161.9878
#> 35038 715092.5 9558152 2715.69 40.183822  0.00270 -0.00250    2225.4426
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.190655  0.01858  0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207  0.01923  0.05547     128.2613
#> 41612 713952.5 9560002 2198.00 28.756179  0.00119  0.00361    1011.0248
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 1479  715622.5 9557952 2888.67 34.953863  0.01635 -0.00105     868.7802
#> 5861  713882.5 9559182 2350.43 27.393685 -0.02608  0.00148   10771.1963
#> 17628 714032.5 9560252 2160.03 30.294570  0.00537 -0.00577     423.5598
#> 3868  714542.5 9560622 2024.15 29.447166  0.00360  0.00750    1892.8978
#> 23499 714272.5 9558232 2508.35 28.967600 -0.00820 -0.00260    1484.6593
#> 38362 714402.5 9558042 2499.31 56.901012  0.03044 -0.00964     543.0025
#> 7132  713192.5 9559612 2060.27 20.773731  0.00452  0.00368    1071.8734
#> 36016 715262.5 9557682 2814.30 37.002187  0.00203  0.00238     787.0053
#> 41873 714832.5 9559942 2253.59 10.598000  0.01508 -0.00177     184.1042
#> 35257 713822.5 9557982 2350.80 29.307937  0.01549 -0.00169     649.9975
#> 46148 715082.5 9558042 2725.39 42.836044 -0.00181 -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.277158 -0.07627 -0.00124   25826.6914
#> 40852 715392.5 9557932 2939.71 18.297407  0.00505  0.00855     229.6061
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 44302 35.1011134   300.00   TRUE       300.00            100  3.053301
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 28223 29.0226042    60.46   TRUE       152.09             35  3.560687
#> 16749 32.1830394   300.00   TRUE       300.00            100  2.753405
#> 30059 20.1921786   300.00   TRUE       300.00             41  2.412719
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 11088 30.9987356   300.00   TRUE       300.00            100  3.212274
#> 15843 42.2218329   300.00   TRUE       300.00            100  3.091290
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 14456 20.0689927   300.00   TRUE       247.02            100  3.183868
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 37270 18.9465684   162.65   TRUE         0.00             77  3.703905
#> 25690 21.7477590   300.00   TRUE       300.00            100  2.732242
#> 38410  8.7610976    69.52   TRUE        47.61             65  3.695093
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 34482 32.9714293   300.00   TRUE       300.00            100  5.111976
#> 38070 34.2330823   300.00   TRUE       300.00            100  2.899802
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 45575 25.7280968   300.00   TRUE       300.00             59  2.562268
#> 38077 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 13847 20.5709037   300.00   TRUE       300.00            100  3.062637
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 47429 32.0976686   300.00   TRUE       300.00            100  2.464778
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 36931 27.2091927   300.00   TRUE       300.00            100  3.408707
#> 46201 22.1837799   167.75   TRUE         0.00             55  3.674828
#> 4605  17.6356409   300.00   TRUE       300.00             10  2.544291
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 47884 27.0544941   300.00   TRUE       300.00            100  2.905563
#> 13134 37.6439001   300.00   TRUE       300.00             67  2.634917
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 39278 16.2181433   125.47   TRUE         0.00              6  3.006344
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 22332 27.0688181   300.00   TRUE       300.00            100  5.168705
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 5084  31.5951847   300.00   TRUE       300.00             18  2.722101
#> 38862 34.9000052   300.00   TRUE       104.48             92  3.151879
#> 7466  28.7905562   255.23   TRUE         0.00             60  3.163437
#> 39520 35.8940870   300.00   TRUE       300.00            100  2.934945
#> 18265 31.4456427   300.00   TRUE       300.00            100  3.236197
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 46065 29.7708870   300.00   TRUE       251.11            100  3.482407
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 13146 33.5764727   300.00   TRUE       300.00              2  3.278424
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 16370 31.1585908   300.00   TRUE       300.00              8  4.119860
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 34192 33.8749837   273.94   TRUE         4.48             85  3.009256
#> 41721 29.9794437   300.00   TRUE       300.00            100  3.099963
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 30268 47.2724558   300.00   TRUE       300.00             41  2.705801
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 7086  34.2044345   300.00   TRUE       128.61             63  3.071308
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 15432 28.7676379   300.00   TRUE       300.00            100  2.880810
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 14545 39.3140721   300.00   TRUE       300.00             37  2.897494
#> 19756 30.1163806   300.00   TRUE       300.00            100  2.694631
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 26058 34.4983618   300.00   TRUE        76.02             39  3.272521
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 28260 31.8335351   300.00   TRUE       300.00            100  3.409673
#> 28588 31.6255514   300.00   TRUE       300.00             78  5.209992
#> 37115 32.1870501   300.00   TRUE       300.00            100  4.596544
#> 17227 43.5975045   300.00   TRUE       300.00            100  2.467494
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14779 16.5212380    55.08   TRUE         0.00              1  2.897760
#> 35458 37.4731587   300.00   TRUE       300.00            100  3.297274
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 6470  28.0932666   300.00   TRUE       300.00              2  2.959800
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 36561 25.8048732   300.00  FALSE        70.00            100  2.990220
#> 11288 28.5373089   132.07  FALSE        85.50              1  3.646686
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 18286 24.3507063   300.00  FALSE       300.00             42  2.931097
#> 20067 29.8235992   300.00  FALSE       268.03            100  2.590250
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 19493 18.5724269   300.00  FALSE       291.23            100  2.550849
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 37832 36.4464183   300.00  FALSE       207.18              2  3.011485
#> 13512 35.9347670   300.00  FALSE       300.00            100  3.605261
#> 39193 31.4800201   300.00  FALSE       117.12            100  2.879028
#> 39796 34.5676897   300.00  FALSE       300.00             71  2.333468
#> 1085  25.8948276   300.00  FALSE       300.00            100  2.477139
#> 39685 47.8734886   300.00  FALSE       300.00            100  2.918481
#> 36833 28.0755049   187.46  FALSE        18.49             45  3.942504
#> 34347 31.7200895   300.00  FALSE       300.00            100  3.477538
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 31497 21.2704852   300.00  FALSE       154.49            100  2.570019
#> 48183 30.9414398   300.00  FALSE       300.00            100  3.204523
#> 31884 24.1375660   300.00  FALSE       300.00            100  3.117018
#> 8413  39.8898947    80.73  FALSE         0.00            100  3.040329
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 33956 38.5955830   300.00  FALSE       300.00            100  2.339942
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 33370 26.4190839   300.00  FALSE       300.00            100  2.363356
#> 13737 40.9601798   300.00  FALSE       300.00            100  2.678473
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 21070 34.6771246   300.00  FALSE         0.00             10  3.400551
#> 36721 26.6683206   300.00  FALSE       300.00            100  2.209482
#> 35038 40.2038755   300.00  FALSE       300.00            100  3.347416
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 36889 19.6450039   300.00  FALSE       300.00             48  2.108096
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 1479  43.6765727   300.00  FALSE       300.00            100  2.938910
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 17628 26.9920417   300.00  FALSE       205.59            100  2.626915
#> 3868  26.1475019   300.00  FALSE       300.00            100  3.277127
#> 23499 28.5149635   300.00  FALSE       300.00             50  3.171627
#> 38362 38.2678511   300.00  FALSE       300.00            100  2.734802
#> 7132  34.0491629   300.00  FALSE       225.26             12  3.030143
#> 36016 39.5237746   300.00  FALSE       300.00            100  2.895978
#> 41873 11.0386049   300.00  FALSE       300.00            100  2.265064
#> 35257 30.8818522   300.00  FALSE       300.00             79  2.812912
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069
#> 40852 21.5180030   300.00  FALSE       300.00            100  2.360983