Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 4965  715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 23178 715382.5 9560142 2021.07 41.9766069   0.00958    0.02642     671.1807
#> 15762 714922.5 9558862 2535.97 54.9105562   0.01305   -0.00555     430.4604
#> 14254 714852.5 9557882 2680.00 23.3354251   0.00479    0.01261     323.8441
#> 5999  715552.5 9558812 2518.03  7.9841669  -0.01863   -0.00507  159065.0781
#> 19963 714612.5 9560552 2031.70 26.8482293   0.00046    0.00324     861.3537
#> 8059  713292.5 9559122 2136.06 29.6620251   0.02240   -0.00250    1413.3312
#> 37265 712812.5 9559942 1860.77 63.3370465   0.00514   -0.00644    1179.0695
#> 42917 714202.5 9557412 2544.08 39.8778625  -0.02104   -0.02046    1024.1573
#> 23817 714842.5 9557832 2677.58 26.6574344   0.00922    0.01088     266.4355
#> 7349  714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 23925 715202.5 9557652 2782.99 41.9571264  -0.00542    0.01842     868.6137
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104    1641.4587
#> 33792 714652.5 9558172 2598.36 59.4586952  -0.03525   -0.11024     394.8751
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898     458.6938
#> 30125 715162.5 9558992 2438.18 39.8566631   0.00471   -0.01382     560.8309
#> 23375 713162.5 9559632 2041.52 46.8152355  -0.00857    0.03677    1675.4679
#> 38070 715212.5 9559222 2342.41 55.6771737   0.00426   -0.00367     793.9665
#> 2942  713232.5 9560652 1851.75 34.5350311   0.00856    0.01114    4510.8833
#> 47022 715362.5 9560102 2059.29 49.1196718   0.02059   -0.00628     556.0121
#> 4827  714962.5 9559882 2274.42 29.2449118   0.00051    0.00889     362.8363
#> 6676  715092.5 9559912 2217.44 41.7078897  -0.00077   -0.00284    2010.4285
#> 40828 714842.5 9557632 2698.65 17.5365192   0.00490   -0.00450     873.0448
#> 27708 715222.5 9557782 2844.65 29.5428498   0.03206    0.02773     179.3077
#> 36230 713322.5 9560662 1798.99 40.1282451  -0.00033   -0.03917     116.1986
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694    2123.4473
#> 28230 715182.5 9558102 2791.83 39.0717109  -0.00669   -0.00950     950.3138
#> 3466  713232.5 9560612 1812.72 51.5662016  -0.01804   -0.01596    5296.9014
#> 7879  714722.5 9559232 2356.19 36.8039440  -0.01097    0.00156     816.6036
#> 28577 712812.5 9559962 1849.82 59.7801245   0.00646   -0.04086    6042.5757
#> 12084 712832.5 9559992 1872.46 56.0547529  -0.03184    0.03924    8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736  -0.01743    0.01313     783.4548
#> 48434 713352.5 9559042 2176.63 35.7021461  -0.01771   -0.01939     703.7817
#> 43092 715532.5 9558192 2770.31 36.8331648  -0.00442   -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.0632631   0.00940    0.00679     550.0344
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407     344.8395
#> 47884 715032.5 9557712 2738.31 28.2892182   0.00006    0.00175     804.5679
#> 37141 714382.5 9559822 2219.36 10.1064025   0.01661    0.01199     196.4080
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 18394 715022.5 9559942 2216.16 42.1376081  -0.00613    0.00903    2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302  -0.01149    0.01949    1182.1506
#> 44550 714842.5 9557642 2696.38 15.1730683   0.00359   -0.00279     930.0871
#> 34707 713962.5 9561092 1819.49 65.6861735  -0.04749    0.04779    2972.0193
#> 2718  715062.5 9558222 2719.56 44.7531604  -0.00119   -0.01560     977.4719
#> 15766 715562.5 9558592 2646.16 29.0208853   0.00194    0.00516     576.0335
#> 4617  714292.5 9559312 2342.95 32.3921689   0.03504    0.05825     162.6688
#> 32909 712812.5 9560452 1883.85 38.2512354  -0.00221    0.00121    4159.6738
#> 18265 715522.5 9558782 2539.23 35.8562718  -0.00458   -0.00401    1722.6481
#> 39732 715362.5 9559582 2230.97 23.3468842  -0.00161   -0.00409    2312.8413
#> 37416 713282.5 9560692 1850.61 30.2590471   0.00736    0.01684    2556.8049
#> 35083 714612.5 9558152 2596.12 60.5983082   0.04548    0.05543     352.1558
#> 19966 715362.5 9559572 2235.06 26.0495262   0.00126   -0.00485    2240.4185
#> 44825 714322.5 9557882 2534.80 40.6885978   0.00114    0.03026     247.0894
#> 21013 712742.5 9560032 1845.53 29.3526278   0.00568   -0.01558    1504.9717
#> 17728 713092.5 9560652 1900.14 30.2791006   0.00645   -0.00395    2011.3270
#> 38197 712522.5 9560322 1942.04 62.7772667  -0.01611    0.02961    1086.6963
#> 45592 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 20601 715542.5 9558182 2773.29 32.7554242  -0.00604    0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930  -0.00472   -0.00228    2262.3796
#> 12137 712842.5 9558902 2097.88 30.0986189   0.00024   -0.00474    2343.2100
#> 1424  714012.5 9558902 2305.97 32.2237194  -0.03986   -0.02524   10861.9482
#> 10467 714642.5 9559652 2249.12 37.1431350  -0.00759   -0.00400     608.6523
#> 19892 714482.5 9560562 2059.63 19.8151724  -0.00332   -0.00058    1817.2371
#> 6217  714442.5 9559932 2151.30 36.6704448   0.01489    0.00241    1752.9242
#> 16009 715122.5 9559402 2264.81 50.0186426  -0.00589    0.00849    1120.5961
#> 7086  712762.5 9560962 2022.00 44.9135886  -0.01259   -0.00281    1178.4420
#> 43201 715712.5 9557922 2852.82 31.4622584  -0.00949    0.00520    3390.5708
#> 42201 714642.5 9557452 2592.29 13.2255848  -0.02084   -0.02236    9755.3027
#> 20797 713012.5 9560392 1842.25 36.6498183   0.01299    0.00141    4611.0718
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175    1026.6818
#> 30933 714672.5 9561082 1798.37 54.8102886   0.00542   -0.00722    1373.4147
#> 26101 715562.5 9558682 2594.87 32.4110766   0.00079    0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.3295700   0.02247    0.01123     364.2903
#> 36292 712862.5 9559602 1915.77 37.9750697   0.00511   -0.01801    1071.6141
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 39325 714142.5 9558592 2357.72 40.7928761  -0.00246    0.00396     704.7076
#> 28173 714562.5 9560382 2048.26 27.6858936  -0.00124   -0.00456    1531.4274
#> 49133 714162.5 9558462 2451.13 21.8904255   0.00042    0.00957     495.4453
#> 1916  715202.5 9559512 2213.91 26.3279836  -0.00292   -0.00109    2047.9031
#> 34984 715042.5 9559312 2320.49 42.8578160  -0.01106   -0.04634     500.5027
#> 20171 715732.5 9557962 2815.53 48.6578678  -0.01341    0.00171    4425.5410
#> 14494 715332.5 9558792 2557.22 20.7204457   0.01267    0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.0716969   0.00285   -0.01325     595.0426
#> 26058 713402.5 9560442 1880.12 31.3774607   0.00730   -0.02610    1872.9276
#> 29024 713492.5 9559112 2173.47 53.4323251   0.04864    0.04226     338.3474
#> 17825 712732.5 9560462 1915.51 40.3975353   0.00143    0.00477    2945.2100
#> 28260 715312.5 9558302 2681.75 34.4840379  -0.01125   -0.02685    2568.4631
#> 18498 714842.5 9557712 2679.04 28.9395253   0.00902    0.01358     802.7622
#> 17674 713882.5 9558562 2339.20 41.0363832  -0.03675   -0.01905    6487.0010
#> 31491 712832.5 9558892 2091.39 29.7004132   0.00186    0.00063    2430.2278
#> 41310 714452.5 9560402 2092.27 12.4629780  -0.00195   -0.00205    1112.2626
#> 21490 714672.5 9560382 2021.86 32.8041256  -0.00783   -0.00367    1966.2112
#> 36765 715302.5 9560262 1952.73 27.2034632  -0.00585   -0.02665    1904.4769
#> 17851 715362.5 9557482 2869.26 29.1257366  -0.03251   -0.00389   14516.2002
#> 6470  713742.5 9558802 2260.79 30.6801711   0.00000    0.00320     911.5911
#> 31253 712612.5 9559492 1903.67 46.7567938  -0.06953   -0.01797    7557.7866
#> 31983 714302.5 9559912 2204.77 23.4649135  -0.01332    0.01842     246.5161
#> 29431 714882.5 9559892 2258.03 21.9179275  -0.00849    0.00150     503.5590
#> 27884 714072.5 9559202 2446.27 19.7567307  -0.00009    0.00109     605.1908
#> 18286 714792.5 9557882 2650.57 28.2536948  -0.00124    0.00204     853.2902
#> 27965 715042.5 9560342 2063.39 29.1973563   0.04336    0.01403     241.2991
#> 14694 715032.5 9559482 2307.44 44.1733271  -0.00007   -0.00463    1050.7473
#> 26023 715832.5 9557632 3097.56 41.7101816   0.00133   -0.00452     578.5847
#> 26963 715702.5 9557882 2880.32 43.7332955  -0.01174   -0.01466    2633.9666
#> 1488  714562.5 9560002 2150.93 11.2838308   0.00197    0.00463     851.6306
#> 43426 713782.5 9560822 1876.35 23.8230122  -0.00345   -0.00075    8141.7549
#> 47005 714942.5 9559192 2333.66 29.2208476  -0.04937   -0.02273   72123.6172
#> 7538  714722.5 9558582 2660.25 15.9946898   0.00544   -0.01014     271.4219
#> 1085  715722.5 9557532 3097.22 28.7842537   0.02327    0.01833     300.0125
#> 45071 712652.5 9560032 1900.80 14.4734869   0.00987   -0.00507     504.3085
#> 31497 714032.5 9560672 2060.77 16.2169974   0.00973   -0.00943     371.5514
#> 48183 713642.5 9558262 2314.44 50.0232262   0.01338   -0.00399    1601.4852
#> 36126 715052.5 9558292 2759.47 29.0621382   0.00196    0.01734     335.0773
#> 44668 712612.5 9560472 1972.24 27.5071308   0.00601   -0.00011     712.5228
#> 41753 715722.5 9558192 2771.39 26.0850495  -0.00196   -0.00175     725.1355
#> 1691  714222.5 9559022 2448.69 38.2168579  -0.01292    0.00462     716.5762
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098     164.0802
#> 48869 712562.5 9560122 1972.89 24.9465824   0.00185    0.01655     560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 17746 714752.5 9558322 2648.52 47.7365517  -0.00068   -0.00312     593.7773
#> 14148 714162.5 9560352 2149.57 37.1333947   0.00424    0.00196     576.1450
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513     161.9878
#> 41855 713182.5 9560452 1851.53 45.1782951   0.00667    0.00863    1098.3462
#> 40909 712722.5 9560272 1827.15 29.2930402   0.00009   -0.00599   12873.8018
#> 48988 714682.5 9560852 1942.97 20.2907274   0.00480    0.00740     909.4791
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788   0.00119    0.00361    1011.0248
#> 24819 714032.5 9560882 1981.86 33.2842642   0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.9538632   0.01635   -0.00105     868.7802
#> 484   713812.5 9560592 2016.41 44.3010967  -0.00719    0.00079    1647.6451
#> 48184 714202.5 9558222 2489.17 36.1135298   0.02236   -0.00066     306.0146
#> 44434 713022.5 9558762 2161.07 31.4542370  -0.00273   -0.01047    2826.7700
#> 974   713402.5 9561062 1854.56 23.6345090   0.00082    0.02578    1258.8865
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 24446 714742.5 9560082 2200.60 28.7218013  -0.00598    0.02208     842.4825
#> 37783 713832.5 9557892 2405.28 40.8868412   0.01572    0.03148     481.8292
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172  111973.8906
#> 16443 714472.5 9559532 2222.55 16.7441186   0.03263   -0.02012     305.3624
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029     894.2672
#> 6577  714142.5 9559902 2244.30 57.8905097  -0.06986   -0.05013     616.4128
#> 23808 715172.5 9558702 2533.50 22.3459270   0.00811   -0.02281     492.4785
#> 11025 714402.5 9559252 2325.64 49.6204369  -0.00241    0.00020     673.2861
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.5833482  -0.00247    0.01057     722.3952
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492     772.2691
#> 24516 713802.5 9560862 1873.21 23.3497490  -0.00771    0.00661    5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 4965  33.90592344   300.00   TRUE       300.00            100  2.699406
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 5999  27.14215667   300.00   TRUE       300.00            100  5.201575
#> 19963 24.43779588   300.00   TRUE       300.00             89  2.935182
#> 8059  26.97714483   300.00   TRUE       300.00              4  3.150244
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 23817 21.61139507   300.00   TRUE       300.00              4  2.425592
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 23925 37.53331924   300.00   TRUE       300.00            100  2.938827
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 30125 40.73558036   300.00   TRUE       300.00            100  2.748832
#> 23375 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 2942  29.82073436    68.13   TRUE       116.32             10  3.654262
#> 47022 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 6676  34.62154773   300.00   TRUE       300.00            100  3.303289
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 27708 25.71090810   300.00   TRUE       300.00            100  2.253599
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 28230 37.90173111   300.00   TRUE       300.00            100  2.977867
#> 3466  26.79494425   101.32   TRUE        95.02             19  3.724022
#> 7879  28.58085369   300.00   TRUE       300.00            100  2.912011
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 31316 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 47884 27.05449413   300.00   TRUE       300.00            100  2.905563
#> 37141 12.38849345   300.00   TRUE       300.00              1  2.293159
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 18394 28.61064750   300.00   TRUE       300.00             57  3.372940
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 2718  36.73289720   300.00   TRUE       300.00            100  2.990104
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 32909 28.37917255   118.07   TRUE        98.15              0  3.619059
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 37416 28.77164864    72.23   TRUE        87.56             31  3.407698
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 19966 29.22886896   300.00   TRUE       300.00             23  3.350329
#> 44825 29.02145824   300.00   TRUE       300.00            100  2.392854
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 17728 28.55220580    30.14   TRUE       140.15              2  3.303483
#> 38197 34.34251728   105.52   TRUE        65.32             50  3.036108
#> 45592 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 20601 37.94871364   300.00   TRUE       300.00            100  3.059843
#> 31834 24.64692547   215.68   TRUE         0.00            100  3.354565
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 1424  39.15421685   300.00   TRUE       300.00            100  4.035908
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 6217  30.97180657   300.00   TRUE       300.00              1  3.243763
#> 16009 33.41317974   300.00   TRUE       300.00            100  3.049449
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 43201 41.82821088   300.00   TRUE       300.00            100  3.530273
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 20797 32.05698864   285.07   TRUE         0.00             96  3.663802
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 30933 12.84342194    99.29   TRUE         0.00              6  3.137802
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 36292 31.51554352   300.00   TRUE        16.03            100  3.030038
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 39325 37.44852149   300.00   TRUE       300.00             60  2.848009
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 49133 23.51934453   300.00   TRUE       300.00             38  2.694996
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 14494 21.18511447   300.00   TRUE       300.00             28  2.535469
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 17674 30.48250062   300.00   TRUE       300.00              1  3.812044
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 29431 20.89176008   300.00  FALSE       300.00             42  2.702050
#> 27884 15.49736244   300.00  FALSE       300.00            100  2.781892
#> 18286 24.35070629   300.00  FALSE       300.00             42  2.931097
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 26963 42.06312357   300.00  FALSE       300.00            100  3.420610
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 47005 36.13530222   300.00  FALSE       300.00            100  4.858077
#> 7538  17.26493724   300.00  FALSE       300.00             87  2.433645
#> 1085  25.89482755   300.00  FALSE       300.00            100  2.477139
#> 45071 27.80506884   264.04  FALSE         0.00            100  2.702696
#> 31497 21.27048519   300.00  FALSE       154.49            100  2.570019
#> 48183 30.94143981   300.00  FALSE       300.00            100  3.204523
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 44668 30.10262960     9.15  FALSE       163.30              6  2.852799
#> 41753 25.00731593   300.00  FALSE       300.00            100  2.860419
#> 1691  27.00006314   300.00  FALSE       300.00            100  2.855262
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 17746 36.39141436   300.00  FALSE       300.00             87  2.773624
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 40909 22.87820476   234.82  FALSE         0.00             43  4.109707
#> 48988 18.95860048   300.00  FALSE       144.19            100  2.958793
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 41612 25.58714921   300.00  FALSE       300.00            100  3.004762
#> 24819 29.93704480   300.00  FALSE       150.00            100  2.905872
#> 1479  43.67657272   300.00  FALSE       300.00            100  2.938910
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 48184 30.95633671   300.00  FALSE       300.00             75  2.485742
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 24446 31.36313675   300.00  FALSE       300.00            100  2.925561
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 16443 26.53825915   300.00  FALSE       300.00             75  2.484816
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491