Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501   -0.00649    1399.2329
#> 4965  715042.5 9559312 2320.49 42.857816   -0.01106   -0.04634     500.5027
#> 44302 714932.5 9557982 2649.56 37.300698    0.01633   -0.01813    1130.5791
#> 42632 714852.5 9557902 2674.80 30.652669    0.00221    0.00969     368.5606
#> 5999  715552.5 9558812 2518.03  7.984167   -0.01863   -0.00507  159065.0781
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659    0.01041     772.9746
#> 19963 714612.5 9560552 2031.70 26.848229    0.00046    0.00324     861.3537
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498   -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.327991   -0.00333   -0.01767    3321.6113
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891   -0.00949    6217.8213
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453    0.00257     357.1436
#> 25169 712642.5 9560232 1883.10 46.845029    0.00884    0.01345    1120.6929
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951   -0.00529    3282.1186
#> 49372 714862.5 9560982 1863.12 20.969109   -0.00295    0.00035    2697.5974
#> 20151 714602.5 9559922 2184.66 48.341595   -0.00979   -0.01542     674.0147
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156    0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765   -0.00068    0.00588    1295.1610
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924    0.02606     564.6876
#> 38287 714872.5 9561162 1719.41 48.235598   -0.03021   -0.09328    2235.5259
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194    0.00526     539.8119
#> 3835  714142.5 9558552 2388.42 32.900383    0.00979   -0.00269     491.8017
#> 21703 713842.5 9559672 2282.58 24.991846    0.01581    0.01189     287.5425
#> 2990  713542.5 9560372 1970.07 31.887966    0.01903    0.01447     331.4833
#> 45869 714342.5 9560282 2127.11 27.623441   -0.02169   -0.00160    9344.2930
#> 43007 713372.5 9559192 2120.47 46.005073    0.03696    0.02514     297.7842
#> 15109 713872.5 9559602 2288.64 25.185506    0.00602    0.01518     854.4269
#> 22566 712832.5 9559842 1897.49 33.060238    0.00672    0.04398     267.2659
#> 20290 714812.5 9561092 1775.73 34.626704   -0.01537   -0.00203   10263.2734
#> 47022 715362.5 9560102 2059.29 49.119672    0.02059   -0.00628     556.0121
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108   -0.00678    1891.6584
#> 48364 715582.5 9558222 2746.23 25.072060    0.01257    0.00593     939.0466
#> 28230 715182.5 9558102 2791.83 39.071711   -0.00669   -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162    0.00598    0.01212     561.1948
#> 46125 713732.5 9560062 2115.18 37.593480   -0.04616    0.01276    9614.6924
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383    0.00193    1809.3838
#> 23141 715632.5 9558942 2495.99 34.376895    0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.642587   -0.00211    0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259    0.01414    0.03416    1204.8629
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028    0.01318     289.0063
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453    0.01077     300.7915
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669   -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224    0.00177     291.5937
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846   -0.00034    4495.6929
#> 34359 712862.5 9558912 2111.37 32.812147    0.00463    0.00167    1938.5389
#> 20001 712702.5 9560132 1854.55 36.291147   -0.00204   -0.00376    2048.3198
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316    0.00714     475.3120
#> 31316 715022.5 9559332 2308.28 20.063263    0.00940    0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.191459    0.00288    0.01452     437.3965
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093   -0.00907    1943.4092
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617    0.00307  147470.5312
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.434393   -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.285823    0.00708   -0.02538     807.8237
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480    0.02530     937.1306
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409    0.02771     674.3859
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175    0.00236    2715.4138
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762    0.01252    5761.3486
#> 19887 714662.5 9559632 2266.47 45.179441   -0.02797   -0.00063     576.2819
#> 7466  712962.5 9560392 1819.03 18.402831   -0.00140   -0.02330    1456.9246
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953   -0.02283     590.1536
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209    0.02739     365.7626
#> 31462 714082.5 9558542 2367.84 34.726972   -0.02027   -0.01473    2930.2791
#> 19966 715362.5 9559572 2235.06 26.049526    0.00126   -0.00485    2240.4185
#> 44650 715612.5 9558142 2764.88 39.360482   -0.00550   -0.00941    1290.7709
#> 30886 714912.5 9558822 2563.86 49.753936   -0.00099    0.02169     495.8840
#> 3418  715372.5 9558492 2601.00 22.627249   -0.01168   -0.04102   35685.0000
#> 35014 713502.5 9559772 2190.48 44.235780    0.00927    0.01643     670.2289
#> 7064  715592.5 9558662 2597.69 34.222196    0.00099    0.00240    1201.4724
#> 17431 712802.5 9560022 1832.71 41.525116   -0.00478   -0.07132   42970.7891
#> 20601 715542.5 9558182 2773.29 32.755424   -0.00604    0.00354    1147.7384
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278   -0.00412    1350.3611
#> 33910 715202.5 9559572 2195.66 13.791667   -0.04376   -0.01644 1057765.6250
#> 10073 715572.5 9558612 2632.78 32.161267    0.00129    0.00171     788.0004
#> 30287 713992.5 9557822 2403.92 18.215474    0.03890   -0.00911     173.2806
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059   -0.07431    4079.0366
#> 37720 715722.5 9557962 2820.42 45.076309   -0.01474    0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751   -0.02070   -0.02661    2508.5374
#> 28312 715712.5 9558032 2783.23 24.252731   -0.00007   -0.02132    1580.1621
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166    0.01476    1404.4454
#> 5572  715062.5 9558262 2754.60 39.911667    0.00707    0.01833     369.3771
#> 10689 713542.5 9559132 2211.95 31.981358    0.05379    0.00801     172.7491
#> 43074 713172.5 9559622 2052.40 31.527576   -0.00193    0.02812    1423.7578
#> 36292 712862.5 9559602 1915.77 37.975070    0.00511   -0.01801    1071.6141
#> 47086 713692.5 9559972 2175.51 25.045704   -0.00789   -0.00261    3353.5422
#> 47298 714602.5 9560572 2025.50 28.590021   -0.00253    0.00753    1337.7854
#> 12629 714712.5 9558872 2533.15 37.025679    0.04250   -0.00970    1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006    0.00258    0.00422    1159.2100
#> 23626 713742.5 9560032 2137.92 39.304905   -0.00605    0.00075    2170.2249
#> 7409  714772.5 9559122 2404.90 30.215502   -0.01016   -0.00294     734.9908
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010    0.00140    1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364   -0.00094    1392.6008
#> 14494 715332.5 9558792 2557.22 20.720446    0.01267    0.00483     343.1384
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257    0.00044     288.5468
#> 39405 713122.5 9559052 2169.30 37.646192   -0.02110   -0.02450    2589.9910
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125   -0.02685    2568.4631
#> 30363 715192.5 9557542 2796.31 25.321870   -0.00157   -0.00152     850.4633
#> 46038 713042.5 9559102 2146.38 29.497586   -0.00622    0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212    0.02732    4237.6934
#> 30626 713302.5 9559172 2106.91 35.997792    0.01354    0.00245     629.5983
#> 4479  713982.5 9557812 2399.24 41.231189    0.01895    0.04744     226.7911
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840    0.00189     556.5211
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683   -0.00603    1562.0503
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783   -0.00367    1966.2112
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332    0.01842     246.5161
#> 7317  713682.5 9558582 2347.46 28.347660   -0.02196    0.00336    2806.2644
#> 29431 714882.5 9559892 2258.03 21.917927   -0.00849    0.00150     503.5590
#> 35852 715112.5 9559962 2173.61 39.315791   -0.00738    0.00118    2273.4758
#> 11111 715142.5 9558482 2650.65 42.343300   -0.00533    0.01493    2212.0525
#> 26963 715702.5 9557882 2880.32 43.733296   -0.01174   -0.01466    2633.9666
#> 46930 713642.5 9560602 1945.03 34.509821    0.00165    0.00875    1165.3953
#> 28162 714062.5 9561152 1785.47 48.444727    0.02309   -0.03409    1575.9425
#> 25140 713332.5 9560692 1809.95 52.257189    0.00075   -0.01945     259.7421
#> 7538  714722.5 9558582 2660.25 15.994690    0.00544   -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304    0.00818    0.00502    1428.8546
#> 34347 714742.5 9560142 2161.95 26.566907   -0.00978    0.00367    3002.8828
#> 31374 714652.5 9560592 1999.69 27.106633   -0.00007   -0.00853    1354.6001
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174   -0.03065     704.4843
#> 48183 713642.5 9558262 2314.44 50.023226    0.01338   -0.00399    1601.4852
#> 8831  714372.5 9559852 2200.77 48.760427   -0.00457    0.01697     412.1185
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180    0.01170     253.5129
#> 14800 712952.5 9558682 2120.89 38.273581   -0.00022    0.00662    1269.2650
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358    0.02422     967.9738
#> 33370 714202.5 9558962 2472.93 32.641406   -0.00307    0.02997     230.8641
#> 26389 713002.5 9559712 1926.02 24.037298    0.00586    0.00654    1576.2065
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452   -0.00948    1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078   -0.00305   -0.00705     968.9738
#> 330   714232.5 9560362 2165.00 18.060776    0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.988751   -0.00270 1018.51947     842.1514
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923    0.05547     128.2613
#> 40909 712722.5 9560272 1827.15 29.293040    0.00009   -0.00599   12873.8018
#> 44823 714202.5 9557442 2548.13 43.230239   -0.00233    0.01282     561.2366
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649  100.79199     107.2986
#> 24819 714032.5 9560882 1981.86 33.284264    0.00804   -0.00165     805.1403
#> 1479  715622.5 9557952 2888.67 34.953863    0.01635   -0.00105     868.7802
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262   -0.00502     520.5218
#> 48184 714202.5 9558222 2489.17 36.113530    0.02236   -0.00066     306.0146
#> 974   713402.5 9561062 1854.56 23.634509    0.00082    0.02578    1258.8865
#> 280   713182.5 9560632 1862.83 39.004102    0.00745    0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434   -0.01502    0.00282    1659.6211
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137   -0.01523     888.2811
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572    0.03148     481.8292
#> 1577  713322.5 9559742 2027.47 29.567487   -0.03284   -0.00886   43249.3945
#> 18504 713372.5 9560672 1786.89  6.473850   -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196   -0.01935     903.6397
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014   -0.00884     672.2466
#> 35257 713822.5 9557982 2350.80 29.307937    0.01549   -0.00169     649.9975
#> 49812 713392.5 9559222 2133.99 35.662612   -0.00983    0.00803    1058.5630
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716   -0.03234    1707.3130
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336    0.01403     609.8665
#> 24516 713802.5 9560862 1873.21 23.349749   -0.00771    0.00661    5642.7544
#>          cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.715694   300.00   TRUE       300.00             21  3.145890
#> 4965  33.905923   300.00   TRUE       300.00            100  2.699406
#> 44302 35.101113   300.00   TRUE       300.00            100  3.053301
#> 42632 20.478085   300.00   TRUE       300.00             10  2.566509
#> 5999  27.142157   300.00   TRUE       300.00            100  5.201575
#> 1300  27.530622   300.00   TRUE       300.00              6  2.888165
#> 19963 24.437796   300.00   TRUE       300.00             89  2.935182
#> 8598  39.598259   300.00   TRUE       300.00            100  3.121100
#> 48187 30.966650   279.54   TRUE        21.35             46  3.521349
#> 38411 32.758289   279.54   TRUE         0.00            100  3.793638
#> 26531 15.987814   300.00   TRUE       300.00              2  2.552843
#> 25169 36.606274   195.00   TRUE        47.05              0  3.049487
#> 43327 29.058700   173.40   TRUE        33.81            100  3.516154
#> 49372 23.540544   213.54   TRUE        20.21             68  3.430977
#> 20151 44.795559   300.00   TRUE       300.00            100  2.828669
#> 38827 31.035978   300.00   TRUE        52.52             39  3.384015
#> 47255 19.677663   300.00   TRUE       260.03            100  3.112324
#> 7349  26.322827   300.00   TRUE       300.00              8  2.751808
#> 38287 14.896903    41.43   TRUE         1.90            100  3.349380
#> 25690 21.747759   300.00   TRUE       300.00            100  2.732242
#> 3835  35.138929   300.00   TRUE       300.00             39  2.691790
#> 21703 22.269151   300.00   TRUE       300.00             34  2.458702
#> 2990  25.321297   300.00   TRUE       123.56              5  2.520462
#> 45869 23.454600   300.00   TRUE       300.00             12  3.970546
#> 43007 38.961703   300.00   TRUE       300.00             37  2.473902
#> 15109 22.635271   300.00   TRUE       300.00             40  2.931675
#> 22566 24.586765   300.00   TRUE         7.49              0  2.426944
#> 20290 15.456682    95.53   TRUE         0.00             49  4.011286
#> 47022 43.531614   300.00   TRUE       300.00             26  2.745084
#> 3453  40.084700   300.00   TRUE       300.00              2  3.276843
#> 48364 32.075896   300.00   TRUE       300.00            100  2.972687
#> 28230 37.901731   300.00   TRUE       300.00            100  2.977867
#> 34565 20.380109   300.00   TRUE        68.13            100  2.749114
#> 46125 25.657050   300.00   TRUE       300.00            100  3.982935
#> 17004 27.189139   300.00   TRUE       300.00            100  3.257531
#> 23141 30.291706   300.00   TRUE       300.00            100  2.902371
#> 19730 38.478700   300.00   TRUE       300.00            100  2.787913
#> 27013 30.918521   300.00   TRUE       168.05             18  3.080938
#> 34642 15.648623   300.00   TRUE       300.00              2  2.460907
#> 19128 14.659698   300.00   TRUE       300.00             15  2.478266
#> 6557  27.419468   300.00   TRUE       300.00              5  3.147064
#> 47429 32.097669   300.00   TRUE       300.00            100  2.464778
#> 47011 44.534291   300.00   TRUE       300.00             10  3.652797
#> 34359 27.416030   300.00   TRUE       294.31            100  3.287475
#> 20001 34.677125   264.46   TRUE        18.17             56  3.311398
#> 18564 21.700204   300.00   TRUE       300.00            100  2.676979
#> 31316 30.009810   300.00   TRUE       300.00             85  2.740390
#> 5578  25.741848   300.00   TRUE       300.00            100  2.640875
#> 13699 34.789424   300.00   TRUE        53.95             43  3.288564
#> 22332 27.068818   300.00   TRUE       300.00            100  5.168705
#> 26567 40.437069   300.00   TRUE       300.00            100  3.023906
#> 33240 29.354920   300.00   TRUE       300.00            100  2.756740
#> 4031  42.761559   300.00   TRUE       300.00             31  2.907317
#> 16015 38.832788   152.09   TRUE        90.23             45  2.971800
#> 21952 40.091576   300.00   TRUE       300.00             64  2.828908
#> 40733 27.013814   300.00   TRUE       300.00            100  3.433836
#> 21053 34.251417   300.00   TRUE         0.00              2  3.760524
#> 19887 38.062159   300.00   TRUE       300.00             63  2.760635
#> 7466  28.790556   255.23   TRUE         0.00             60  3.163437
#> 30406 43.100750   300.00   TRUE       300.00             85  2.770965
#> 28845 29.842507   300.00   TRUE       300.00             54  2.563199
#> 31462 33.754663   300.00   TRUE       300.00              1  3.466909
#> 19966 29.228869   300.00   TRUE       300.00             23  3.350329
#> 44650 37.362578   300.00   TRUE       300.00            100  3.110849
#> 30886 39.185157   300.00   TRUE       300.00            100  2.695380
#> 3418  35.453482   300.00   TRUE       300.00             73  4.552486
#> 35014 32.530252   300.00   TRUE       300.00             57  2.826223
#> 7064  29.132612   300.00   TRUE       300.00            100  3.079714
#> 17431 25.325308   300.00   TRUE         1.11              5  4.633173
#> 20601 37.948714   300.00   TRUE       300.00            100  3.059843
#> 32879 29.738228   300.00   TRUE       300.00              6  3.130450
#> 33910 32.673491   300.00   TRUE       300.00              6  6.024389
#> 10073 25.761328   300.00   TRUE       300.00            100  2.896526
#> 30287 19.848404   300.00   TRUE       300.00             65  2.238750
#> 4588  31.664513   300.00   TRUE       300.00              1  3.610558
#> 37720 42.773591   300.00   TRUE       300.00            100  3.690657
#> 21683 19.745272   130.93   TRUE         0.00             21  3.399421
#> 28312 41.642573   300.00   TRUE       300.00            100  3.198702
#> 22907 33.813677   300.00   TRUE       205.28              1  3.147505
#> 5572  26.597847   300.00   TRUE       300.00            100  2.567470
#> 10689 30.648085   300.00   TRUE       300.00             61  2.237416
#> 43074 33.594807   300.00   TRUE       205.02              1  3.153436
#> 36292 31.515544   300.00   TRUE        16.03            100  3.030038
#> 47086 23.386991   300.00   TRUE       300.00            100  3.525504
#> 47298 25.336194   300.00   TRUE       300.00            100  3.126386
#> 12629 32.180748   300.00   TRUE       300.00            100  3.159543
#> 10617 33.798207    17.89   TRUE       145.83              4  3.064162
#> 23626 26.778328   300.00   TRUE       300.00            100  3.336505
#> 7409  25.370571   300.00   TRUE       300.00            100  2.866282
#> 48650 27.052202   300.00   TRUE       300.00            100  3.027918
#> 48574 34.424450    96.21   TRUE       101.32             10  3.143827
#> 14494 21.185114   300.00   TRUE       300.00             28  2.535469
#> 44612 34.691449   300.00   TRUE       300.00            100  2.460216
#> 39405 27.720271   300.00   TRUE       300.00            100  3.413298
#> 28260 31.833535   300.00   TRUE       300.00            100  3.409673
#> 30363 29.802973   300.00   TRUE       300.00            100  2.929656
#> 46038 29.730780   300.00   TRUE       243.86            100  3.184852
#> 42640 34.909745   300.00   TRUE       300.00            100  3.627130
#> 30626 30.003508   300.00   TRUE       300.00              0  2.799064
#> 4479  32.011725   300.00   TRUE       300.00             55  2.355626
#> 43536 29.094797   300.00   TRUE       300.00              8  2.745482
#> 11332 25.681687   300.00   TRUE        27.04             68  3.193695
#> 21490 34.060049   300.00   TRUE       300.00              1  3.293630
#> 31983 19.775638   300.00  FALSE       300.00             65  2.391845
#> 7317  27.225809   300.00  FALSE       300.00             61  3.448129
#> 29431 20.891760   300.00  FALSE       300.00             42  2.702050
#> 35852 36.362194   300.00  FALSE       300.00            100  3.356690
#> 11111 33.383959   300.00  FALSE       300.00            100  3.344795
#> 26963 42.063124   300.00  FALSE       300.00            100  3.420610
#> 46930 31.920625   300.00  FALSE         0.00            100  3.066473
#> 28162 18.669830   130.00  FALSE        75.00            100  3.197540
#> 25140 14.483800   104.12  FALSE        46.28             74  2.414542
#> 7538  17.264937   300.00  FALSE       300.00             87  2.433645
#> 16878 25.388906   300.00  FALSE       300.00             18  3.154988
#> 34347 31.720089   300.00  FALSE       300.00            100  3.477538
#> 31374 26.679207   300.00  FALSE       300.00            100  3.131811
#> 4007  32.092512   300.00  FALSE       300.00             68  2.847871
#> 48183 30.941440   300.00  FALSE       300.00            100  3.204523
#> 8831  32.555462   300.00  FALSE       300.00             24  2.615022
#> 31201 33.722577   300.00  FALSE       300.00            100  2.404000
#> 14800 31.522992   300.00  FALSE       300.00            100  3.103552
#> 6143  28.757325   300.00  FALSE       185.58            100  2.985864
#> 33370 26.419084   300.00  FALSE       300.00            100  2.363356
#> 26389 36.122124   300.00  FALSE        76.38             26  3.197613
#> 12519 30.485938   300.00  FALSE       120.00            100  3.087264
#> 33410 32.879756   300.00  FALSE       300.00            100  2.986312
#> 330   13.572224   300.00  FALSE       275.17            100  2.372638
#> 46509 45.578793   300.00  FALSE       162.53            100  2.925390
#> 36889 19.645004   300.00  FALSE       300.00             48  2.108096
#> 40909 22.878205   234.82  FALSE         0.00             43  4.109707
#> 44823 36.107227   300.00  FALSE       300.00            100  2.749146
#> 48845 31.559088   169.72  FALSE         0.00            100  2.030594
#> 24819 29.937045   300.00  FALSE       150.00            100  2.905872
#> 1479  43.676573   300.00  FALSE       300.00            100  2.938910
#> 29716 32.693545   300.00  FALSE       300.00            100  2.716439
#> 48184 30.956337   300.00  FALSE       300.00             75  2.485742
#> 974   22.758457    80.00  FALSE        90.05            100  3.099987
#> 280   29.323407    60.55  FALSE       127.54             29  3.333837
#> 18246  7.987605    18.49  FALSE         0.00             29  3.220009
#> 6732  26.856251   300.00  FALSE       300.00            100  2.948550
#> 37783 24.257314   300.00  FALSE       300.00             18  2.682893
#> 1577  38.929617   300.00  FALSE       300.00             63  4.635980
#> 18504 23.095929   144.96  FALSE         4.48             89  5.049117
#> 1434  39.692797   300.00  FALSE       300.00             35  2.955995
#> 37254  4.191186   131.59  FALSE        80.55            100  2.827529
#> 35257 30.881852   300.00  FALSE       300.00             79  2.812912
#> 49812 33.358176   300.00  FALSE       300.00             65  3.024717
#> 14645 45.213246   300.00  FALSE       300.00            100  3.232313
#> 46567 16.948092   300.00  FALSE       300.00             78  2.785235
#> 24516 36.254477   300.00  FALSE       122.13            100  3.751491