Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.564231   -0.00681  -0.00029   5577.3916
#> 23178 715382.5 9560142 2021.07 41.976607    0.00958   0.02642    671.1807
#> 47223 715302.5 9557472 2857.45 39.533515   -0.01021  -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555    430.4604
#> 13748 715232.5 9559542 2212.12 26.123438    0.00023  -0.00043   1806.4552
#> 42733 715112.5 9559062 2399.70 36.054515   -0.00041   0.01741    746.4155
#> 24051 715462.5 9559372 2319.46 32.327998   -0.00661  -0.01339   1002.5877
#> 37265 712812.5 9559942 1860.77 63.337047    0.00514  -0.00644   1179.0695
#> 24072 715282.5 9557602 2837.46 34.394083   -0.02191  -0.00579   2246.7725
#> 21939 713642.5 9558712 2299.09 33.750652   -0.00503   0.02433   1475.3191
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679    550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440   1088.8639
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529   3282.1186
#> 125   713992.5 9561062 1878.28 49.971087    0.00183   0.03387   1989.3012
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547    519.4228
#> 7349  714032.5 9558502 2402.95 31.352824    0.01924   0.02606    564.6876
#> 38993 714202.5 9558402 2475.75 33.159359   -0.01607   0.00677   1343.5006
#> 25690 715042.5 9557572 2752.67 32.713598    0.00194   0.00526    539.8119
#> 33792 714652.5 9558172 2598.36 59.458695   -0.03525  -0.11024    394.8751
#> 22283 713592.5 9560762 1832.60 36.010397    0.00448  -0.00328    592.3326
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510   3342.5518
#> 2107  714142.5 9558562 2382.67 35.553750    0.01412   0.00898    458.6938
#> 18304 714582.5 9558932 2463.25 22.194666   -0.02131   0.00501   3889.7292
#> 35285 714602.5 9560112 2142.75 51.107835   -0.00017  -0.00032    994.6740
#> 18005 714832.5 9557912 2660.76 31.912030    0.00191  -0.00261    554.9243
#> 31576 713942.5 9558492 2387.12 46.143156   -0.05634  -0.02626   4316.0977
#> 15340 715452.5 9558852 2523.95 34.165473    0.00151   0.00899   1394.2274
#> 26215 714332.5 9560262 2138.08 25.721221   -0.00491   0.00581   1855.1630
#> 40828 714842.5 9557632 2698.65 17.536519    0.00490  -0.00450    873.0448
#> 36230 713322.5 9560662 1798.99 40.128245   -0.00033  -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774    527.1068
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088   2322.1746
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626   1014.7159
#> 4329  712882.5 9559982 1911.46 26.231727   -0.00193  -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086   6042.5757
#> 34642 714962.5 9559862 2282.08 22.160862   -0.00028   0.01318    289.0063
#> 22735 714492.5 9559772 2170.30 39.688786   -0.00440  -0.02651    419.0927
#> 19128 714852.5 9557872 2681.49 20.757688    0.00453   0.01077    300.7915
#> 12868 712532.5 9560332 1913.92 61.937311   -0.03543  -0.05146   1347.6224
#> 5728  713502.5 9559662 2140.27 34.307567   -0.00154   0.00124   2114.2253
#> 47351 714222.5 9558792 2420.94 28.668516   -0.09566  -0.03355  14919.6260
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177    291.5937
#> 47011 713262.5 9559722 2003.72 37.192982   -0.01846  -0.00034   4495.6929
#> 48434 713352.5 9559042 2176.63 35.702146   -0.01771  -0.01939    703.7817
#> 17594 714832.5 9561052 1818.04 46.167220   -0.01353   0.00063   5035.0000
#> 13699 712902.5 9559902 1917.21 30.487657   -0.00093  -0.00907   1943.4092
#> 23029 713502.5 9559102 2174.07 56.457542    0.03395   0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.686173   -0.04749   0.04779   2972.0193
#> 16015 712582.5 9560312 1899.18 56.546924    0.03480   0.02530    937.1306
#> 39599 714112.5 9558562 2374.25 29.168135    0.00283   0.00117   1018.2141
#> 9704  714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400    608.6523
#> 21952 715352.5 9560172 2013.44 40.488063    0.02409   0.02771    674.3859
#> 835   715052.5 9558212 2707.65 37.725260   -0.01399  -0.00731   1384.7959
#> 40733 712862.5 9558892 2108.60 29.994914    0.00175   0.00236   2715.4138
#> 4617  714292.5 9559312 2342.95 32.392169    0.03504   0.05825    162.6688
#> 21053 712862.5 9559992 1901.01 30.950607   -0.00762   0.01252   5761.3486
#> 26698 713382.5 9559082 2155.56 34.264595   -0.00211   0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408  76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283    590.1536
#> 10507 714892.5 9559282 2374.45 42.842346   -0.01057   0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.545142    0.00576  -0.00166    650.6246
#> 47100 712752.5 9560462 1910.00 41.911863   -0.00032   0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.562526   -0.00856  -0.01625   1063.9150
#> 664   714452.5 9559852 2174.83 44.860876    0.00282   0.00109    767.4097
#> 31561 713302.5 9559192 2092.11 36.527778    0.00406  -0.00055    773.0359
#> 26010 712852.5 9560542 1928.25 30.978109    0.00165   0.00555   4439.0967
#> 34192 712702.5 9560102 1867.01 35.474682    0.00629  -0.00469   1021.5407
#> 47366 714892.5 9557612 2718.78 28.023939    0.00538   0.00112    745.3315
#> 40216 714602.5 9560592 2015.55 34.368873   -0.00093   0.00664   1807.2659
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389  47022.9688
#> 39166 712832.5 9560142 1922.46 39.660139    0.01469   0.02940    987.8085
#> 32879 713912.5 9559672 2238.00 31.976775   -0.00278  -0.00412   1350.3611
#> 46461 712682.5 9560032 1892.30 22.977899    0.01453   0.01337    486.7551
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245   6353.3179
#> 19892 714482.5 9560562 2059.63 19.815172   -0.00332  -0.00058   1817.2371
#> 27244 714162.5 9558622 2316.00 36.751232   -0.02556  -0.11694 102301.4609
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236   9755.3027
#> 46138 715202.5 9559582 2194.40 18.860052   -0.02143  -0.05078 622888.2500
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017    759.9937
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363    565.5207
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981   3975.3608
#> 2801  715252.5 9559612 2199.28 24.403418    0.03051   0.02190    264.6404
#> 23783 713062.5 9560252 1983.09 36.429803    0.00774  -0.00224   1137.3363
#> 39624 715222.5 9559552 2205.54 23.897497    0.00401  -0.00391   1794.7964
#> 20426 714702.5 9559712 2208.61 43.046319   -0.08153  -0.02417  34908.6562
#> 48650 715592.5 9558612 2626.74 31.876507    0.00010   0.00140   1066.3957
#> 48574 712692.5 9560412 1901.97 41.978326    0.01364  -0.00094   1392.6008
#> 1223  713632.5 9559632 2226.22 50.808751   -0.01881   0.00121   1783.0658
#> 44612 714932.5 9558822 2578.80 38.255246    0.01257   0.00044    288.5468
#> 14950 713382.5 9559172 2111.91 48.106682    0.00915  -0.02705    413.5702
#> 24723 712942.5 9560202 1983.18 33.071697    0.00285  -0.01325    595.0426
#> 16065 715242.5 9560622 1871.61 34.306421   -0.00019   0.00119   6589.5796
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685   2568.4631
#> 46792 713862.5 9559672 2272.03 35.464942    0.00560   0.00610    454.8103
#> 12074 714562.5 9560372 2045.17 26.287304   -0.00144  -0.00856   1789.1596
#> 10213 715542.5 9558782 2532.10 35.413375   -0.00893  -0.00867   2247.8748
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400   1982.7782
#> 9864  714952.5 9557592 2752.40 31.989380   -0.00983   0.01253    528.1924
#> 48461 714672.5 9559272 2316.84 40.381492   -0.01966  -0.02234    849.0547
#> 30378 714852.5 9557752 2651.70 41.298225   -0.00010  -0.00870   2099.7930
#> 49346 715322.5 9558782 2553.74 25.395208    0.00319   0.00291    487.3438
#> 41408 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476   1404.4454
#> 11332 712742.5 9560072 1850.34 22.700015    0.00683  -0.00603   1562.0503
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403    474.8613
#> 21490 714672.5 9560382 2021.86 32.804126   -0.00783  -0.00367   1966.2112
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665   1904.4769
#> 17374 715712.5 9557952 2832.22 37.874802   -0.01519   0.00609   4275.3457
#> 31983 714302.5 9559912 2204.77 23.464914   -0.01332   0.01842    246.5161
#> 16208 712882.5 9560332 1846.34 48.625209    0.00159  -0.01559    866.5775
#> 31071 713472.5 9558462 2298.33 29.153239    0.00156  -0.00116   1817.5222
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345 665381.5625
#> 8066  715842.5 9558732 2636.38 33.717420   -0.00494   0.01064    683.2795
#> 1488  714562.5 9560002 2150.93 11.283831    0.00197   0.00463    851.6306
#> 12716 713962.5 9557762 2402.60 18.932244    0.03480   0.01500    538.8209
#> 43426 713782.5 9560822 1876.35 23.823012   -0.00345  -0.00075   8141.7549
#> 42329 715852.5 9558702 2651.92 22.870756    0.00321   0.00548    532.8210
#> 36126 715052.5 9558292 2759.47 29.062138    0.00196   0.01734    335.0773
#> 12974 714662.5 9557332 2667.61 41.941657   -0.00065   0.00475    515.1661
#> 34616 714522.5 9558742 2568.77 34.997981    0.00990   0.02790    209.6952
#> 32849 714302.5 9558222 2520.01 32.157256   -0.00846   0.00216   1412.2295
#> 20319 715282.5 9560482 1901.16 27.676726    0.00706  -0.03035   1471.9783
#> 6143  712762.5 9559022 2096.24 29.896365    0.00358   0.02422    967.9738
#> 48869 712562.5 9560122 1972.89 24.946582    0.00185   0.01655    560.8554
#> 20135 714772.5 9560232 2107.52 37.850738   -0.00782  -0.00048   1550.6704
#> 3733  713762.5 9559632 2297.63 34.214748   -0.01338   0.00518   1137.0179
#> 34810 714112.5 9559572 2337.51 10.288030   -0.00393   0.00032    994.9224
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196    576.1450
#> 12519 713982.5 9560912 1956.98 28.824934   -0.00452  -0.00948   1222.5433
#> 5656  713132.5 9560142 2042.24 33.595380   -0.00035   0.00955    713.4463
#> 7983  713512.5 9560432 1907.38 50.019216    0.03310  -0.07650    766.6254
#> 35826 712912.5 9560302 1891.22 50.869485   -0.00229   0.00099    909.0560
#> 34532 713812.5 9559502 2328.80 44.937653   -0.00070  -0.00060    383.3547
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199    107.2986
#> 29716 715272.5 9559992 2144.94 25.717784    0.00262  -0.00502    520.5218
#> 23499 714272.5 9558232 2508.35 28.967600   -0.00820  -0.00260   1484.6593
#> 12300 712372.5 9560062 1903.64 50.036977    0.00343   0.01177   1558.2231
#> 7986  712752.5 9560242 1857.41 39.830307    0.01852   0.00248   1432.4019
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050   1471.4346
#> 40750 712962.5 9558702 2126.06 34.704054   -0.00264   0.00624   1870.8307
#> 280   713182.5 9560632 1862.83 39.004102    0.00745   0.00875   2156.9358
#> 6732  715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368   1071.8734
#> 1434  714672.5 9559902 2194.92 42.452162   -0.01196  -0.01935    903.6397
#> 41873 714832.5 9559942 2253.59 10.598000    0.01508  -0.00177    184.1042
#> 40653 714262.5 9561262 1791.94 12.222336    0.00161   0.00709   1181.2555
#> 14881 715482.5 9559232 2376.03 29.905532   -0.01137  -0.01523    888.2811
#> 24962 715152.5 9560472 1967.37 28.845560    0.00298   0.00222    991.8300
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448    941.8503
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525    162.0592
#> 6206  714122.5 9560992 1917.28 37.969913   -0.01023  -0.00227    841.2025
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 13748 34.11906374   300.00   TRUE       300.00              2  3.256827
#> 42733 31.28292266   300.00   TRUE       300.00            100  2.872981
#> 24051 32.73880841   300.00   TRUE       300.00             11  3.001122
#> 37265 36.51746507   300.00   TRUE         2.45              0  3.071539
#> 24072 37.66681841   300.00   TRUE       300.00            100  3.351559
#> 21939 23.93244710   300.00   TRUE       300.00            100  3.168886
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 125   29.59670787   207.04   TRUE        21.03            100  3.298701
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 33792 51.20810294   300.00   TRUE       300.00              5  2.596460
#> 22283 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 18304 44.90843198   300.00   TRUE       300.00             57  3.589919
#> 35285 42.46648586   300.00   TRUE       300.00            100  2.997681
#> 18005 24.90590240   300.00   TRUE       300.00              0  2.744234
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 26215 23.83160653   300.00   TRUE       300.00             10  3.268382
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 19128 14.65969815   300.00   TRUE       300.00             15  2.478266
#> 12868 37.58832319   119.34   TRUE        79.16             61  3.129568
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 47351 34.87880578   300.00   TRUE       300.00            100  4.173758
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 47011 44.53429054   300.00   TRUE       300.00             10  3.652797
#> 48434 29.93589888   300.00   TRUE       300.00              6  2.847438
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 23029 46.60152227   300.00   TRUE       300.00             34  2.425842
#> 34707 27.70250939   165.13   TRUE        39.56            100  3.473052
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 39599 33.13472225   300.00   TRUE       300.00             25  3.007839
#> 9704  38.99608049   300.00   TRUE       300.00             90  2.784369
#> 21952 40.09157580   300.00   TRUE       300.00             64  2.828908
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 40733 27.01381412   300.00   TRUE       300.00            100  3.433836
#> 4617  27.95174604   300.00   TRUE       300.00              6  2.211304
#> 21053 34.25141699   300.00   TRUE         0.00              2  3.760524
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 10507 27.16221019   300.00   TRUE       300.00             46  2.580865
#> 13029 34.18781868   300.00   TRUE        96.65              4  2.813330
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 29801 33.24186536   300.00   TRUE       300.00              2  3.026907
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31561 31.44277788   300.00   TRUE       300.00              0  2.888200
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 34192 33.87498372   273.94   TRUE         4.48             85  3.009256
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 40216 26.36407999   300.00   TRUE       300.00            100  3.257022
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 46461 24.94142578   288.47   TRUE         0.00            100  2.687311
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 19892 27.02641920   300.00   TRUE       300.00             92  3.259412
#> 27244 37.77510743   300.00   TRUE       300.00             54  5.009882
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 48650 27.05220230   300.00   TRUE       300.00            100  3.027918
#> 48574 34.42445025    96.21   TRUE       101.32             10  3.143827
#> 1223  40.45826879   300.00   TRUE       300.00              4  3.251167
#> 44612 34.69144858   300.00   TRUE       300.00            100  2.460216
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46792 27.60109586   300.00   TRUE       300.00             18  2.657830
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 48461 35.78006839   300.00   TRUE       300.00             61  2.928936
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 49346 23.34573832   300.00   TRUE       300.00             29  2.687835
#> 41408 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 11332 25.68168725   300.00   TRUE        27.04             68  3.193695
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 17374 43.14544085   300.00   TRUE       300.00            100  3.630971
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 31071 31.73785115   300.00  FALSE       300.00             76  3.259480
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 8066  23.52736594   300.00  FALSE       300.00            100  2.834598
#> 1488  29.92157494   300.00  FALSE       300.00            100  2.930251
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 42329 19.22789065   300.00  FALSE       300.00            100  2.726581
#> 36126 19.29549967   300.00  FALSE       300.00            100  2.525145
#> 12974 37.33851359   300.00  FALSE       300.00            100  2.711947
#> 34616 31.49950070   300.00  FALSE       300.00            100  2.321589
#> 32849 29.55201716   300.00  FALSE       300.00             45  3.149905
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 6143  28.75732470   300.00  FALSE       185.58            100  2.985864
#> 48869 24.87553563   138.98  FALSE         0.00             63  2.748851
#> 20135 32.33258134   300.00  FALSE       300.00            100  3.190519
#> 3733  22.10012807   300.00  FALSE       300.00            100  3.055767
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 5656  31.27260942   300.00  FALSE        94.43            100  2.853361
#> 7983  38.50620158   300.00  FALSE        57.15             22  2.884583
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 29716 32.69354475   300.00  FALSE       300.00            100  2.716439
#> 23499 28.51496355   300.00  FALSE       300.00             50  3.171627
#> 12300 38.41739312   124.35  FALSE         0.00            100  3.192630
#> 7986  32.00313060   275.06  FALSE         0.00             65  3.156065
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 40750 32.09996047   300.00  FALSE       300.00            100  3.272034
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 14881 26.85625073   300.00  FALSE       300.00            100  2.948550
#> 24962 29.05010613   300.00  FALSE       300.00            100  2.996437
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 6206  32.46321571   294.31  FALSE        35.90            100  2.924901