Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope     hcurv      vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.5642309  -0.00681   -0.00029    5577.3916
#> 31357 714752.5 9561022 1848.22 33.4464113  -0.00347    0.02357    1752.0375
#> 27864 712992.5 9560672 1926.38 27.2240896  -0.00199    0.00659    3553.7166
#> 28223 712742.5 9560482 1927.69 34.6679573  -0.00292    0.00712    3636.5303
#> 40756 714022.5 9558862 2331.20 45.0854759  -0.00075    0.00475    1001.0861
#> 48187 712722.5 9560152 1835.29 29.3279907  -0.00333   -0.01767    3321.6113
#> 42917 714202.5 9557412 2544.08 39.8778625  -0.02104   -0.02046    1024.1573
#> 32622 714972.5 9557762 2687.38 30.5564122  -0.00730   -0.01491    1844.4353
#> 47255 713552.5 9559962 2191.77 33.5787645  -0.00068    0.00588    1295.1610
#> 26447 715632.5 9559102 2410.77 39.6567008   0.00730    0.01770     474.5059
#> 31023 714712.5 9561042 1838.66 44.1200421  -0.00223    0.01783     377.8918
#> 21703 713842.5 9559672 2282.58 24.9918461   0.01581    0.01189     287.5425
#> 39061 712922.5 9559912 1931.50 36.3152110   0.00414   -0.00104    1641.4587
#> 8019  714672.5 9559272 2316.84 40.3814924  -0.01966   -0.02234     849.0547
#> 45869 714342.5 9560282 2127.11 27.6234412  -0.02169   -0.00160    9344.2930
#> 16843 714882.5 9558512 2701.78 26.6379538  -0.00270   -0.00769    1032.7177
#> 2107  714142.5 9558562 2382.67 35.5537501   0.01412    0.00898     458.6938
#> 38032 714762.5 9560442 1988.33 35.1698683  -0.00905   -0.01385     674.0721
#> 31576 713942.5 9558492 2387.12 46.1431560  -0.05634   -0.02626    4316.0977
#> 35942 712862.5 9558902 2109.97 31.4003790   0.00151    0.00189    2317.1819
#> 32821 715022.5 9559952 2208.33 43.4164499  -0.00643    0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.1257506   0.00108   -0.00678    1891.6584
#> 20194 714132.5 9558452 2447.22 24.0224015  -0.00987    0.02107    1392.5880
#> 47220 715052.5 9557752 2726.93 32.7216197  -0.00664    0.00694    2123.4473
#> 13847 715052.5 9559232 2337.87 26.1148433  -0.01452   -0.01437    1155.1469
#> 6763  714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 1635  713102.5 9559052 2172.87 31.4840308  -0.00064   -0.00916    2009.1686
#> 2950  715042.5 9557712 2740.72 26.5250811  -0.00089   -0.00051     779.0555
#> 35142 714792.5 9558862 2575.17 25.6644985   0.06765    0.00114     150.1498
#> 34642 714962.5 9559862 2282.08 22.1608616  -0.00028    0.01318     289.0063
#> 22735 714492.5 9559772 2170.30 39.6887865  -0.00440   -0.02651     419.0927
#> 46756 713372.5 9559192 2120.47 46.0050732   0.03696    0.02514     297.7842
#> 29667 713892.5 9559712 2239.25 41.5927252   0.00804    0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.6343274  -0.00669   -0.00991    1403.0193
#> 17594 714832.5 9561052 1818.04 46.1672203  -0.01353    0.00063    5035.0000
#> 18564 715552.5 9558592 2648.36 26.2913143   0.00316    0.00714     475.3120
#> 46201 714832.5 9561022 1845.94 34.5739922  -0.00856    0.01796    4729.6382
#> 4605  714852.5 9557892 2677.78 26.9336000   0.00321    0.01169     350.1799
#> 47091 713382.5 9559182 2122.71 48.0379275   0.04527    0.04613     289.6113
#> 46171 712822.5 9558932 2089.30 29.8545389  -0.01174    0.00025    4425.5151
#> 13699 712902.5 9559902 1917.21 30.4876572  -0.00093   -0.00907    1943.4092
#> 29926 714702.5 9557622 2637.00 44.6425796   0.01843    0.04407     344.8395
#> 39278 714692.5 9561052 1830.60 48.3948802  -0.01996    0.00626    1014.7159
#> 26567 714022.5 9558872 2327.42 46.6656935   0.00290    0.00230    1056.5876
#> 33240 714972.5 9557652 2741.04 39.4343932  -0.00286   -0.00964     571.1364
#> 4031  715392.5 9560162 1998.16 46.2858225   0.00708   -0.02538     807.8237
#> 835   715052.5 9558212 2707.65 37.7252601  -0.01399   -0.00731    1384.7959
#> 15766 715562.5 9558592 2646.16 29.0208853   0.00194    0.00516     576.0335
#> 375   715622.5 9558922 2510.87 37.3665885  -0.01017    0.00197     711.1091
#> 26698 713382.5 9559082 2155.56 34.2645950  -0.00211    0.00241     585.6182
#> 33350 713132.5 9560672 1896.89 24.6950539   0.00306   -0.00436    2602.7390
#> 25700 715172.5 9559272 2324.49 35.5348425  -0.00529    0.02009    1247.7061
#> 31462 714082.5 9558542 2367.84 34.7269720  -0.02027   -0.01473    2930.2791
#> 46783 713592.5 9560762 1832.60 36.0103974   0.00448   -0.00328     592.3326
#> 20065 713772.5 9558852 2221.21 28.0199280  -0.00568   -0.02553    1750.4175
#> 44650 715612.5 9558142 2764.88 39.3604817  -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170  -0.02554    0.03304    1569.7313
#> 31121 713462.5 9559132 2135.56 61.5969737  -0.03800   -0.06020    1563.5549
#> 16355 712962.5 9560212 1992.88 39.5260665   0.01927   -0.00316     454.9674
#> 35014 713502.5 9559772 2190.48 44.2357795   0.00927    0.01643     670.2289
#> 42765 713552.5 9559632 2186.95 44.0111801   0.00394   -0.00224     942.0322
#> 5239  714832.5 9560992 1859.70 19.9601307  -0.00320    0.00180    3904.5254
#> 7385  713752.5 9560122 2064.61 31.7710827  -0.02605   -0.03594   16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387   0.00538    0.00112     745.3315
#> 39166 712832.5 9560142 1922.46 39.6601386   0.01469    0.02940     987.8085
#> 7041  713472.5 9559092 2137.03 18.3489734  -0.10372   -0.00008  825875.6875
#> 8125  713452.5 9559132 2120.44 42.1404729  -0.05458   -0.14813  287033.4688
#> 33910 715202.5 9559572 2195.66 13.7916671  -0.04376   -0.01644 1057765.6250
#> 30287 713992.5 9557822 2403.92 18.2154742   0.03890   -0.00911     173.2806
#> 37720 715722.5 9557962 2820.42 45.0763086  -0.01474    0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.7327506  -0.02070   -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.4805791   0.00171   -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.5545187  -0.01175    0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.5936474   0.00384    0.00175    1026.6818
#> 17058 714032.5 9558502 2402.95 31.3528235   0.01924    0.02606     564.6876
#> 28009 715002.5 9559922 2237.68 34.6037860  -0.01109    0.00298    1533.7925
#> 7485  714362.5 9558042 2484.02 47.9588593   0.03169    0.01441     424.2680
#> 12629 714712.5 9558872 2533.15 37.0256786   0.04250   -0.00970    1443.9200
#> 10363 713252.5 9559722 1999.55 41.8299297  -0.03669   -0.01981    3975.3608
#> 49072 715472.5 9558902 2482.92 30.6893384  -0.01102   -0.01778    4594.4351
#> 13259 714862.5 9560992 1859.28 21.2676204  -0.00317    0.00077    3056.9475
#> 39624 715222.5 9559552 2205.54 23.8974967   0.00401   -0.00391    1794.7964
#> 20426 714702.5 9559712 2208.61 43.0463191  -0.08153   -0.02417   34908.6562
#> 39405 713122.5 9559052 2169.30 37.6461919  -0.02110   -0.02450    2589.9910
#> 46627 714852.5 9558932 2459.24 53.4455031   0.02451   -0.03921     525.3245
#> 49459 714692.5 9557342 2639.74 42.0470808   0.00216   -0.02606     919.4241
#> 42640 714332.5 9558522 2403.10 38.8711757  -0.00212    0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.1135439  -0.00118   -0.00252    3631.7017
#> 12074 714562.5 9560372 2045.17 26.2873036  -0.00144   -0.00856    1789.1596
#> 11057 714962.5 9557642 2737.48 42.6859287  -0.00488   -0.00632     620.0229
#> 46971 713532.5 9559952 2198.03 29.7239045  -0.00529    0.01230    1044.6443
#> 43834 713092.5 9560662 1905.55 31.3236027   0.00758   -0.00288    2144.8164
#> 34718 714802.5 9558802 2599.22 44.6500280   0.01996    0.04684     209.1778
#> 10785 714052.5 9557852 2377.76 20.9840063  -0.00025   -0.05345    2417.1389
#> 21490 714672.5 9560382 2021.86 32.8041256  -0.00783   -0.00367    1966.2112
#> 9984  712742.5 9560092 1848.56 23.6127367   0.00614   -0.00093     118.1402
#> 32405 713022.5 9558982 2208.85 15.0046187   0.00599    0.00301     338.9676
#> 7317  713682.5 9558582 2347.46 28.3476599  -0.02196    0.00336    2806.2644
#> 48601 713692.5 9560152 2058.14 59.5572439  -0.01433   -0.02177     997.3089
#> 11288 712822.5 9560442 1873.42 39.6097183  -0.00465   -0.00105    4432.8760
#> 15186 713322.5 9561212 1936.88 36.0797253   0.00165    0.00565    1410.8372
#> 24836 714202.5 9558272 2507.87 38.4890128  -0.00226    0.00736     477.8591
#> 2664  713342.5 9558572 2275.86 56.9754961   0.00872    0.05419     761.5427
#> 40459 712542.5 9559192 1992.60 39.8171290  -0.00162   -0.00517    1060.5494
#> 27965 715042.5 9560342 2063.39 29.1973563   0.04336    0.01403     241.2991
#> 8635  713832.5 9557902 2397.71 47.1194761   0.02295    0.02715     520.7355
#> 14717 715382.5 9558062 2799.05 50.5904544  -0.00812   -0.02208     951.3401
#> 18746 713292.5 9560952 1890.51 41.8855067  -0.00402   -0.01328    1066.7134
#> 20458 713612.5 9559562 2288.02 33.4401088   0.00112    0.03538     265.4677
#> 20067 714242.5 9560442 2125.23 38.6402737   0.00363   -0.00363     389.2691
#> 30267 714442.5 9557512 2612.30 28.7659191  -0.00458    0.00479     604.0975
#> 49363 712772.5 9559552 1935.91 29.0369281  -0.03220   -0.03290    2481.5632
#> 12716 713962.5 9557762 2402.60 18.9322444   0.03480    0.01500     538.8209
#> 37832 712552.5 9559192 1998.10 41.5508993  -0.00195   -0.00995    1026.7981
#> 13512 715152.5 9557992 2778.81 43.3895209  -0.00724    0.01464    4029.5947
#> 39796 714802.5 9560372 2046.48 40.1557471   0.00176    0.02973     215.5101
#> 16878 714232.5 9558172 2485.33 22.8083039   0.00818    0.00502    1428.8546
#> 40139 712852.5 9560072 1907.89 43.7040747   0.00001    0.00759    1880.4403
#> 42934 713832.5 9558672 2270.74 27.6119821  -0.03193   -0.00267   31588.1953
#> 20319 715282.5 9560482 1901.16 27.6767263   0.00706   -0.03035    1471.9783
#> 33175 715822.5 9558832 2563.13 36.6228893  -0.00674   -0.00376     629.8236
#> 22545 713402.5 9559562 2195.44 25.3155672   0.04152    0.00098     164.0802
#> 34105 715432.5 9558592 2621.73 18.9563087  -0.00301   -0.01099    1146.2140
#> 34228 715932.5 9558592 2690.97  9.5351636 197.75496 -197.75366     267.5057
#> 13737 714272.5 9559622 2257.05 55.4571579  -0.03224   -0.02036     476.9504
#> 29835 715402.5 9558512 2610.14 27.1180924  -0.01322   -0.00759    4131.5356
#> 12044 713612.5 9559292 2247.76 23.0747293  -0.00112   -0.00038    2017.3547
#> 26389 713002.5 9559712 1926.02 24.0372984   0.00586    0.00654    1576.2065
#> 44267 712972.5 9559042 2191.01 22.2691506   0.01462    0.00268     252.0072
#> 33410 715342.5 9560452 1942.88 29.0930780  -0.00305   -0.00705     968.9738
#> 36721 714092.5 9559752 2301.81 28.9332227   0.03418    0.00513     161.9878
#> 49831 715262.5 9557332 2884.30 22.2949337  -0.00097    0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.0607756   0.01142    0.00868     235.8511
#> 43415 714722.5 9558642 2657.99  0.2314749   0.00799    0.00911     100.0000
#> 14763 713462.5 9560852 1830.40 47.7399894   0.01354    0.00907    1281.5044
#> 42638 714322.5 9561182 1754.22  4.1722787  -0.03757   -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.2216583   0.00607   -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.7139475   0.01182   -0.00972     331.7462
#> 6732  715482.5 9559232 2376.03 29.9055321  -0.01137   -0.01523     888.2811
#> 10127 713722.5 9558712 2311.10 19.4811380   0.01129    0.00491     261.7371
#> 18504 713372.5 9560672 1786.89  6.4738501  -0.00278   -0.01172  111973.8906
#> 1434  714672.5 9559902 2194.92 42.4521619  -0.01196   -0.01935     903.6397
#> 22628 714942.5 9557732 2696.19 26.2116732   0.00788   -0.01029     894.2672
#> 46639 715232.5 9560042 2107.15 39.6572738  -0.01305    0.00585    5714.3267
#> 47245 715372.5 9560792 1900.65 34.4708598  -0.00011   -0.00419     323.0790
#> 126   715042.5 9561262 1733.49 30.1106510  -0.00288    0.00448     941.8503
#> 41373 715332.5 9560862 1873.10 29.1245906  -0.00166    0.00156     758.4840
#> 46567 714472.5 9560352 2080.13 21.3816390   0.01336    0.01403     609.8665
#> 47629 714472.5 9558482 2437.33 37.2726234  -0.01181   -0.01149    4242.3579
#> 48157 713142.5 9560992 1978.48 44.6213801   0.00208    0.00492     772.2691
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 27864 27.81538208    30.00   TRUE       183.39             20  3.550683
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 40756 39.33527151   300.00   TRUE       300.00            100  3.000471
#> 48187 30.96664995   279.54   TRUE        21.35             46  3.521349
#> 42917 38.80356668   300.00   TRUE       300.00            100  3.010367
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 47255 19.67766252   300.00   TRUE       260.03            100  3.112324
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 31023 16.67765550   135.00   TRUE         0.00              2  2.577367
#> 21703 22.26915062   300.00   TRUE       300.00             34  2.458702
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 8019  35.78006839   300.00   TRUE       300.00             61  2.928936
#> 45869 23.45460030   300.00   TRUE       300.00             12  3.970546
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 2107  35.42368864   300.00   TRUE       300.00             45  2.661523
#> 38032 35.00829424   300.00   TRUE       300.00             60  2.828706
#> 31576 26.93531891   300.00   TRUE       300.00              4  3.635091
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 20194 28.11733084   300.00   TRUE       300.00             41  3.143823
#> 47220 36.48022282   300.00   TRUE       300.00            100  3.327041
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 2950  28.20384747   300.00   TRUE       300.00            100  2.891568
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 34642 15.64862330   300.00   TRUE       300.00              2  2.460907
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 46756 38.96170303   300.00   TRUE       300.00             37  2.473902
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 17594 19.23190135   138.35   TRUE         0.00             60  3.701999
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 46201 22.18377991   167.75   TRUE         0.00             55  3.674828
#> 4605  17.63564093   300.00   TRUE       300.00             10  2.544291
#> 47091 40.15173637   300.00   TRUE       300.00             25  2.461815
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 13699 34.78942436   300.00   TRUE        53.95             43  3.288564
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 39278 16.21814335   125.47   TRUE         0.00              6  3.006344
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 4031  42.76155912   300.00   TRUE       300.00             31  2.907317
#> 835   36.65039128   300.00   TRUE       300.00            100  3.141386
#> 15766 23.33714395   300.00   TRUE       300.00            100  2.760448
#> 375   27.97409139   300.00   TRUE       300.00            100  2.851936
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 25700 32.54113797   300.00   TRUE       300.00            100  3.096112
#> 31462 33.75466258   300.00   TRUE       300.00              1  3.466909
#> 46783 19.94236902   256.35   TRUE         0.00            100  2.772566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 44650 37.36257782   300.00   TRUE       300.00            100  3.110849
#> 32280 27.83142490   126.66   TRUE         0.00              5  3.195825
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 5239  24.66526012   197.29   TRUE         4.67             57  3.591568
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 39166 29.81844253   300.00   TRUE         1.90             90  2.994673
#> 7041  33.21894705   300.00   TRUE       300.00              6  5.916915
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 33910 32.67349123   300.00   TRUE       300.00              6  6.024389
#> 30287 19.84840394   300.00   TRUE       300.00             65  2.238750
#> 37720 42.77359124   300.00   TRUE       300.00            100  3.690657
#> 21683 19.74527154   130.93   TRUE         0.00             21  3.399421
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 17058 26.32282702   300.00   TRUE       300.00              8  2.751808
#> 28009 27.01782483   300.00   TRUE       300.00             29  3.185767
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 39624 33.29457748   300.00   TRUE       300.00              2  3.254015
#> 20426 33.40859608   300.00   TRUE       300.00             76  4.542933
#> 39405 27.72027109   300.00   TRUE       300.00            100  3.413298
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 49459 41.85685877   300.00   TRUE       300.00            100  2.963516
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 46971 16.78422565   300.00   TRUE       262.51            100  3.018968
#> 43834 28.51209876    20.21   TRUE       150.14              2  3.331390
#> 34718 34.40955334   300.00   TRUE       300.00            100  2.320516
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 21490 34.06004909   300.00   TRUE       300.00              1  3.293630
#> 9984   3.94653329   300.00   TRUE        36.46             76  2.072398
#> 32405 15.76206894   300.00  FALSE       300.00            100  2.530158
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 48601 35.85799065   300.00  FALSE       245.19            100  2.998830
#> 11288 28.53730890   132.07  FALSE        85.50              1  3.646686
#> 15186 38.54860046    60.00  FALSE         0.00            100  3.149477
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 14717 50.85630685   300.00  FALSE       300.00            100  2.978336
#> 18746 39.13645515     4.48  FALSE       205.00             27  3.028048
#> 20458 21.42976745   300.00  FALSE       300.00             28  2.424012
#> 20067 29.82359915   300.00  FALSE       268.03            100  2.590250
#> 30267 25.27030355   300.00  FALSE       300.00            100  2.781107
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 12716 24.34497672   300.00  FALSE       300.00             25  2.731444
#> 37832 36.44641831   300.00  FALSE       207.18              2  3.011485
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 16878 25.38890582   300.00  FALSE       300.00             18  3.154988
#> 40139 33.14331662   300.00  FALSE         1.11             21  3.274260
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 20319 32.07245850   300.00  FALSE       300.00            100  3.167901
#> 33175 34.78369478   300.00  FALSE       300.00            100  2.799219
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 29835 28.85300865   300.00  FALSE       300.00             45  3.616112
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 26389 36.12212419   300.00  FALSE        76.38             26  3.197613
#> 44267 19.73782308   300.00  FALSE       226.30            100  2.401413
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 49831 29.83792310   300.00  FALSE       300.00            100  3.247811
#> 330   13.57222425   300.00  FALSE       275.17            100  2.372638
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 14763 10.14479072   110.04  FALSE        60.35            100  3.107720
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 6732  26.85625073   300.00  FALSE       300.00            100  2.948550
#> 10127 18.61940947   300.00  FALSE       300.00             49  2.417865
#> 18504 23.09592872   144.96  FALSE         4.48             89  5.049117
#> 1434  39.69279717   300.00  FALSE       300.00             35  2.955995
#> 22628 29.12115290   300.00  FALSE       300.00            100  2.951467
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 41373 31.12879701   300.00  FALSE       120.90            100  2.879946
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 48157 33.52662538    57.15  FALSE       157.06             45  2.887769