Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv    vcurv       carea
#> 31358 712882.5 9560002 1911.52 25.5642309 -0.00681 -0.00029   5577.3916
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501 -0.00649   1399.2329
#> 15412 715272.5 9557702 2813.17 30.9523260 -0.00123  0.00393   2081.0056
#> 15762 714922.5 9558862 2535.97 54.9105562  0.01305 -0.00555    430.4604
#> 16749 715072.5 9558242 2745.27 49.0537816  0.00097  0.01483    566.7674
#> 25090 715362.5 9560102 2059.29 49.1196718  0.02059 -0.00628    556.0121
#> 14254 714852.5 9557882 2680.00 23.3354251  0.00479  0.01261    323.8441
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659  0.01041    772.9746
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090  0.00890   2301.0454
#> 15843 714042.5 9558902 2332.30 50.0977107 -0.00858 -0.00292   1233.9299
#> 9669  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 37104 713962.5 9561282 1795.56 25.7418478  0.00273 -0.02403   1787.3024
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250 -0.01340    604.1032
#> 1768  713912.5 9558552 2357.19 38.6924129 -0.00645  0.00835   2425.3115
#> 49372 714862.5 9560982 1863.12 20.9691094 -0.00295  0.00035   2697.5974
#> 20151 714602.5 9559922 2184.66 48.3415951 -0.00979 -0.01542    674.0147
#> 7349  714032.5 9558502 2402.95 31.3528235  0.01924  0.02606    564.6876
#> 24018 714292.5 9559282 2334.12 32.8654321  0.00102  0.00388   1051.4222
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214  0.00466    724.0713
#> 38287 714872.5 9561162 1719.41 48.2355979 -0.03021 -0.09328   2235.5259
#> 25690 715042.5 9557572 2752.67 32.7135983  0.00194  0.00526    539.8119
#> 38410 713712.5 9561172 1776.08 36.3192217 -0.02678 -0.04272   4955.5586
#> 29420 713592.5 9560772 1826.82 34.9269342 -0.00070 -0.00530   9000.4356
#> 34482 715152.5 9557642 2753.52  4.9956190 -0.01126 -0.02044 129412.3984
#> 22566 712832.5 9559842 1897.49 33.0602377  0.00672  0.04398    267.2659
#> 20290 714812.5 9561092 1775.73 34.6267043 -0.01537 -0.00203  10263.2734
#> 27978 713552.5 9558532 2373.05 17.0449214  0.00832  0.00838    238.6685
#> 15340 715452.5 9558852 2523.95 34.1654733  0.00151  0.00899   1394.2274
#> 38070 715212.5 9559222 2342.41 55.6771737  0.00426 -0.00367    793.9665
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043  0.00973    514.5488
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051  0.00889    362.8363
#> 47210 714952.5 9557612 2744.35 40.8845494 -0.00493  0.00733    598.7983
#> 45575 715432.5 9559452 2305.28 24.2343959  0.00993 -0.00413    364.9788
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490 -0.00450    873.0448
#> 3453  713482.5 9559642 2145.11 35.1257506  0.00108 -0.00678   1891.6584
#> 36230 713322.5 9560662 1798.99 40.1282451 -0.00033 -0.03917    116.1986
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356  0.00774    527.1068
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996  0.00626   1014.7159
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275  0.00605    798.6775
#> 35142 714792.5 9558862 2575.17 25.6644985  0.06765  0.00114    150.1498
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184  0.03924   8023.8657
#> 22735 714492.5 9559772 2170.30 39.6887865 -0.00440 -0.02651    419.0927
#> 45104 715062.5 9559242 2345.02 24.1564736 -0.01743  0.01313    783.4548
#> 5728  713502.5 9559662 2140.27 34.3075669 -0.00154  0.00124   2114.2253
#> 12460 712762.5 9559622 1906.41 24.2618978  0.01047 -0.00887    366.8537
#> 47429 714962.5 9557832 2705.12 32.0575616  0.02224  0.00177    291.5937
#> 46171 712822.5 9558932 2089.30 29.8545389 -0.01174  0.00025   4425.5151
#> 13134 714652.5 9559632 2265.33 49.5998104 -0.03552 -0.02058    431.4362
#> 37341 714412.5 9559142 2377.24 40.5997894 -0.00735 -0.00925    665.6116
#> 3944  715262.5 9558352 2654.20 33.5552733 -0.04492  0.00072  54831.6289
#> 33240 714972.5 9557652 2741.04 39.4343932 -0.00286 -0.00964    571.1364
#> 48808 715032.5 9557792 2707.02 25.4376072 -0.02329 -0.01181  11655.4316
#> 26698 713382.5 9559082 2155.56 34.2645950 -0.00211  0.00241    585.6182
#> 42941 713252.5 9559742 1990.10 28.9544222 -0.04708  0.00408  76318.0156
#> 27875 714832.5 9561032 1838.22 41.2122812 -0.01160  0.01690   5121.6924
#> 33350 713132.5 9560672 1896.89 24.6950539  0.00306 -0.00436   2602.7390
#> 39732 715362.5 9559582 2230.97 23.3468842 -0.00161 -0.00409   2312.8413
#> 6317  715042.5 9557702 2745.36 27.8738238  0.00090 -0.00299    688.0706
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845  0.00454   3476.0818
#> 31121 713462.5 9559132 2135.56 61.5969737 -0.03800 -0.06020   1563.5549
#> 35014 713502.5 9559772 2190.48 44.2357795  0.00927  0.01643    670.2289
#> 13146 713792.5 9558552 2341.40 29.8253180 -0.00104  0.01664   1898.5588
#> 5103  714612.5 9559892 2220.36 49.0079450 -0.00078  0.00219    420.7032
#> 42765 713552.5 9559632 2186.95 44.0111801  0.00394 -0.00224    942.0322
#> 26010 712852.5 9560542 1928.25 30.9781091  0.00165  0.00555   4439.0967
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291 -0.00081    940.6002
#> 12137 712842.5 9558902 2097.88 30.0986189  0.00024 -0.00474   2343.2100
#> 30268 714172.5 9558612 2328.84 54.9506633  0.03743 -0.05242    507.9263
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556 -0.04394    587.8840
#> 43859 714892.5 9557602 2722.70 28.1293630  0.00370 -0.00190    640.8842
#> 8125  713452.5 9559132 2120.44 42.1404729 -0.05458 -0.14813 287033.4688
#> 44767 714452.5 9560402 2092.27 12.4629780 -0.00195 -0.00205   1112.2626
#> 20368 715252.5 9560622 1865.83 33.5724620  0.00025 -0.01245   6353.3179
#> 47718 712812.5 9559892 1867.55 54.9386311 -0.01491  0.01061   5744.7583
#> 18440 714862.5 9558932 2448.29 41.0524260 -0.04989 -0.02781  29995.1543
#> 31117 712892.5 9558922 2129.84 30.5506826  0.00521 -0.00870   1895.1399
#> 7086  712762.5 9560962 2022.00 44.9135886 -0.01259 -0.00281   1178.4420
#> 4588  713852.5 9558612 2308.97 52.1804123 -0.01059 -0.07431   4079.0366
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384  0.00175   1026.6818
#> 5911  713952.5 9561282 1801.16 33.7896130  0.00466 -0.01776   1021.8267
#> 10689 713542.5 9559132 2211.95 31.9813582  0.05379  0.00801    172.7491
#> 28173 714562.5 9560382 2048.26 27.6858936 -0.00124 -0.00456   1531.4274
#> 37439 714602.5 9560672 1982.99 18.9064613 -0.00070  0.00010   3467.2690
#> 2801  715252.5 9559612 2199.28 24.4034184  0.03051  0.02190    264.6404
#> 34984 715042.5 9559312 2320.49 42.8578160 -0.01106 -0.04634    500.5027
#> 14950 713382.5 9559172 2111.91 48.1066824  0.00915 -0.02705    413.5702
#> 24723 712942.5 9560202 1983.18 33.0716969  0.00285 -0.01325    595.0426
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397  0.00526    706.1463
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019  0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.3975353  0.00143  0.00477   2945.2100
#> 10456 712442.5 9560292 2010.91 26.8917741  0.00445  0.00875    597.2893
#> 12074 714562.5 9560372 2045.17 26.2873036 -0.00144 -0.00856   1789.1596
#> 11057 714962.5 9557642 2737.48 42.6859287 -0.00488 -0.00632    620.0229
#> 9864  714952.5 9557592 2752.40 31.9893796 -0.00983  0.01253    528.1924
#> 33142 715652.5 9558212 2740.94 14.3812407 -0.00465 -0.01036   2505.8369
#> 30378 714852.5 9557752 2651.70 41.2982249 -0.00010 -0.00870   2099.7930
#> 31491 712832.5 9558892 2091.39 29.7004132  0.00186  0.00063   2430.2278
#> 39466 714942.5 9557772 2680.35 16.7532859  0.00166 -0.01265   1148.6487
#> 4785  713792.5 9560472 2055.34 33.0527893 -0.00586  0.00067   1123.7318
#> 28253 714102.5 9559892 2288.88  2.9576081  0.00378  0.02382    100.0000
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079 -0.00331    699.3024
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159 -0.01559    866.5775
#> 7317  713682.5 9558582 2347.46 28.3476599 -0.02196  0.00336   2806.2644
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226  0.00736    477.8591
#> 25212 714272.5 9559012 2433.45 37.6714021 -0.01909 -0.01181   1286.3949
#> 2664  713342.5 9558572 2275.86 56.9754961  0.00872  0.05419    761.5427
#> 35852 715112.5 9559962 2173.61 39.3157909 -0.00738  0.00118   2273.4758
#> 26023 715832.5 9557632 3097.56 41.7101816  0.00133 -0.00452    578.5847
#> 49363 712772.5 9559552 1935.91 29.0369281 -0.03220 -0.03290   2481.5632
#> 9665  713362.5 9558812 2275.28 31.4949170  0.01306 -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.3895209 -0.00724  0.01464   4029.5947
#> 39193 715332.5 9560852 1872.90 30.2905598 -0.00161  0.00171    756.8818
#> 45166 715262.5 9558722 2531.59 13.5533166 -0.00502 -0.00028   2394.7261
#> 44396 715512.5 9558102 2845.31 26.5984197  0.01263  0.03598    186.3527
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405  0.00476  11992.5273
#> 18627 714292.5 9559252 2330.78 18.4962235 -0.00538 -0.00052   3508.5591
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007 -0.00853   1354.6001
#> 4007  714132.5 9558412 2464.44 43.7201175  0.00174 -0.03065    704.4843
#> 31884 715782.5 9558612 2632.60 30.2796736  0.00084 -0.00083   1309.2347
#> 29944 713562.5 9558952 2209.55 17.3399947  0.01432 -0.00362    283.7090
#> 38689 714622.5 9559342 2264.90 42.6481135 -0.00844  0.00514    897.9290
#> 37029 713022.5 9559082 2163.08 36.0321698  0.00490 -0.00780    825.4451
#> 29527 714332.5 9557502 2567.02 48.9048126 -0.00452 -0.01298    436.1459
#> 13737 714272.5 9559622 2257.05 55.4571579 -0.03224 -0.02036    476.9504
#> 34810 714112.5 9559572 2337.51 10.2880302 -0.00393  0.00032    994.9224
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379 -0.01018    937.6911
#> 12044 713612.5 9559292 2247.76 23.0747293 -0.00112 -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972 -0.03004  0.00714 480535.5000
#> 45486 712762.5 9560172 1853.82 56.7973062  0.01978  0.01282   2322.7944
#> 41855 713182.5 9560452 1851.53 45.1782951  0.00667  0.00863   1098.3462
#> 36889 714302.5 9559802 2243.60 21.1662069  0.01923  0.05547    128.2613
#> 34532 713812.5 9559502 2328.80 44.9376528 -0.00070 -0.00060    383.3547
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799  0.00911    100.0000
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608  0.00148  10771.1963
#> 17628 714032.5 9560252 2160.03 30.2945705  0.00537 -0.00577    423.5598
#> 9007  714952.5 9560902 1863.75 32.7703211  0.00074 -0.02904    608.3561
#> 34336 713792.5 9559072 2288.95 36.9328595 -0.00952 -0.00438   1276.5448
#> 7132  713192.5 9559612 2060.27 20.7737308  0.00452  0.00368   1071.8734
#> 37783 713832.5 9557892 2405.28 40.8868412  0.01572  0.03148    481.8292
#> 16845 712772.5 9560022 1831.76 21.3999736 -0.00865 -0.00805   7489.7417
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770  0.00270  35950.5508
#> 11394 713472.5 9561232 1889.48 34.3144424  0.00108 -0.00948   1538.1741
#> 34334 714642.5 9557912 2582.25 30.9912871  0.00253 -0.01233    968.7780
#> 36016 715262.5 9557682 2814.30 37.0021874  0.00203  0.00238    787.0053
#> 41873 714832.5 9559942 2253.59 10.5980003  0.01508 -0.00177    184.1042
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161  0.00709   1181.2555
#> 37254 713482.5 9560892 1823.73 39.8824462  0.01014 -0.00884    672.2466
#> 33881 712992.5 9558822 2173.26 33.1977476 -0.00595  0.03064   1485.5201
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732  0.01858    298.4490
#> 40043 713322.5 9560562 1802.09 39.8520794  0.04456 -0.01586    941.0288
#> 24516 713802.5 9560862 1873.21 23.3497490 -0.00771  0.00661   5642.7544
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.42788799   300.00   TRUE        15.00              9  3.746431
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 16749 32.18303935   300.00   TRUE       300.00            100  2.753405
#> 25090 43.53161440   300.00   TRUE       300.00             26  2.745084
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 20799 24.80792661   180.67   TRUE         0.00             16  3.361925
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 37104  4.86613055    50.99   TRUE         0.00             21  3.252198
#> 39885  6.08481178   210.57   TRUE         0.00            100  2.781111
#> 1768  28.18551282   300.00   TRUE       300.00              7  3.384768
#> 49372 23.54054397   213.54   TRUE        20.21             68  3.430977
#> 20151 44.79555930   300.00   TRUE       300.00            100  2.828669
#> 7349  26.32282702   300.00   TRUE       300.00              8  2.751808
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 27375 38.77033512   300.00   TRUE       300.00            100  2.859781
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 25690 21.74775903   300.00   TRUE       300.00            100  2.732242
#> 38410  8.76109765    69.52   TRUE        47.61             65  3.695093
#> 29420 28.85243569   253.02   TRUE         0.00            100  3.954264
#> 34482 32.97142928   300.00   TRUE       300.00            100  5.111976
#> 22566 24.58676490   300.00   TRUE         7.49              0  2.426944
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 27978 14.32852854   300.00   TRUE       300.00             12  2.377795
#> 15340 27.97523731   300.00   TRUE       300.00            100  3.144334
#> 38070 34.23308234   300.00   TRUE       300.00            100  2.899802
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4827  21.07396066   300.00   TRUE       300.00              0  2.559711
#> 47210 26.91526538   300.00   TRUE       300.00            100  2.777281
#> 45575 25.72809683   300.00   TRUE       300.00             59  2.562268
#> 40828 25.59975429   300.00   TRUE       300.00            100  2.941037
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 36230  0.05614986   120.49   TRUE        38.30             46  2.065201
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 35142 27.15304287   300.00   TRUE       300.00            100  2.176525
#> 12084 34.59576463   300.00   TRUE         0.00              0  3.904384
#> 22735 36.51689211   300.00   TRUE       300.00             90  2.622310
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 5728  40.40842146   300.00   TRUE       300.00              2  3.325151
#> 12460 36.25676928   300.00   TRUE         0.00            100  2.564493
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 46171 29.07130557   300.00   TRUE       270.22            100  3.645964
#> 13134 37.64390010   300.00   TRUE       300.00             67  2.634917
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 3944  39.36048165   300.00   TRUE       300.00            100  4.739031
#> 33240 29.35491968   300.00   TRUE       300.00            100  2.756740
#> 48808 33.77357019   300.00   TRUE       300.00            100  4.066528
#> 26698 34.24969812   300.00   TRUE       300.00              7  2.767615
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 33350 28.95270330    10.00   TRUE       166.26              2  3.415431
#> 39732 29.16813543   300.00   TRUE       300.00             25  3.364146
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 35014 32.53025178   300.00   TRUE       300.00             57  2.826223
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 42765 42.35304022   300.00   TRUE       300.00             12  2.974066
#> 26010 25.22962355    69.52   TRUE       166.56              2  3.647295
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 12137 27.58849079   300.00   TRUE       300.00            100  3.369811
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 6845  44.59616998   300.00   TRUE       300.00             29  2.769292
#> 43859 29.55774674   300.00   TRUE       300.00            100  2.806780
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 44767 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 47718 32.32341401   300.00   TRUE        10.00              1  3.759272
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 31117 26.64597522   300.00   TRUE       291.23            100  3.277641
#> 7086  34.20443445   300.00   TRUE       128.61             63  3.071308
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 34615 34.18953755    17.57   TRUE       142.95             16  3.011436
#> 5911   4.13847415    42.88   TRUE         0.00             13  3.009377
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 28173 23.88374569   300.00   TRUE       300.00             59  3.185096
#> 37439 26.59269015   300.00   TRUE       300.00            100  3.539988
#> 2801  25.74643148   300.00   TRUE       300.00             26  2.422656
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 14950 41.87404750   300.00   TRUE       300.00             15  2.616549
#> 24723 28.17348070   300.00   TRUE        36.63            100  2.774548
#> 42617 31.23192941   300.00   TRUE       300.00            100  2.848895
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 10456 21.66869085    24.34   TRUE         0.00             18  2.776185
#> 12074 23.28557775   300.00   TRUE       300.00             63  3.252649
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 9864  20.82071332   300.00   TRUE       300.00            100  2.722792
#> 33142 26.90609806   300.00   TRUE       300.00            100  3.398953
#> 30378 28.56939454   300.00   TRUE       300.00             57  3.322176
#> 31491 27.55869699   300.00   TRUE       300.00            100  3.385647
#> 39466 27.25159161   300.00   TRUE       300.00            100  3.060187
#> 4785  26.48612000   300.00  FALSE         0.00            100  3.050663
#> 28253  2.95760814   300.00  FALSE       300.00            100  2.000000
#> 15211 30.83486966   300.00  FALSE       300.00            100  2.844665
#> 16208 38.37384833   256.60  FALSE         0.00             67  2.937807
#> 7317  27.22580851   300.00  FALSE       300.00             61  3.448129
#> 24836 25.58371147   300.00  FALSE       300.00             95  2.679300
#> 25212 38.98290247   300.00  FALSE       300.00            100  3.109374
#> 2664  39.29115376   300.00  FALSE       300.00            100  2.881694
#> 35852 36.36219351   300.00  FALSE       300.00            100  3.356690
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 49363 32.95767829   300.00  FALSE         0.00            100  3.394725
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 39193 31.48002014   300.00  FALSE       117.12            100  2.879028
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 20547 34.26115728   300.00  FALSE        42.56            100  4.078911
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 38689 36.08430898   300.00  FALSE       300.00             18  2.953242
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 13737 40.96017982   300.00  FALSE       300.00            100  2.678473
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 34345 44.39506180   300.00  FALSE       300.00            100  2.972060
#> 12044 32.23689739   300.00  FALSE       300.00            100  3.304782
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 45486 33.42177411   300.00  FALSE         0.00             68  3.366011
#> 41855 28.70174779   234.40  FALSE        47.05            100  3.040739
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 34532 31.20671927   300.00  FALSE       300.00             92  2.583601
#> 43415  0.23147495   300.00  FALSE       300.00            100  2.000000
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 17628 26.99204173   300.00  FALSE       205.59            100  2.626915
#> 9007  30.96435812   300.00  FALSE        14.98             35  2.784158
#> 34336 32.45118360   300.00  FALSE       300.00            100  3.106036
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 16845 22.88278842   300.00  FALSE        25.55             33  3.874467
#> 20950 32.71187940   300.00  FALSE       300.00             81  4.555706
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 36016 39.52377462   300.00  FALSE       300.00            100  2.895978
#> 41873 11.03860488   300.00  FALSE       300.00            100  2.265064
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 33881 29.18589713   300.00  FALSE       300.00            100  3.171879
#> 39883 17.04434849   233.81  FALSE       219.23            100  2.474870
#> 40043 23.99661838   194.79  FALSE        24.89             54  2.973603
#> 24516 36.25447744   300.00  FALSE       122.13            100  3.751491