Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope      hcurv      vcurv       carea
#> 47223 715302.5 9557472 2857.45 39.5335149   -0.01021   -0.01579   1318.7560
#> 13748 715232.5 9559542 2212.12 26.1234377    0.00023   -0.00043   1806.4552
#> 46178 714892.5 9559312 2379.54 32.7668833   -0.00266    0.02896    276.0310
#> 8059  713292.5 9559122 2136.06 29.6620251    0.02240   -0.00250   1413.3312
#> 8598  714002.5 9558872 2310.02 39.5839352   -0.00498   -0.00752   1321.5997
#> 37104 713962.5 9561282 1795.56 25.7418478    0.00273   -0.02403   1787.3024
#> 44952 715352.5 9557472 2873.35 38.9525357   -0.01073    0.00174   2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278    0.00250   -0.01340    604.1032
#> 38411 713552.5 9560652 1851.63 46.0050732   -0.00891   -0.00949   6217.8213
#> 8869  715062.5 9559392 2278.56 41.4019303    0.00310    0.01440   1088.8639
#> 7349  714032.5 9558502 2402.95 31.3528235    0.01924    0.02606    564.6876
#> 39629 714602.5 9558912 2474.89 40.2583065   -0.06175   -0.08226   2325.2053
#> 11463 713112.5 9558902 2244.08 21.8915714    0.00160    0.00311    586.7354
#> 23925 715202.5 9557652 2782.99 41.9571264   -0.00542    0.01842    868.6137
#> 40307 713302.5 9560702 1845.07 39.0046749    0.00129    0.02510   3342.5518
#> 48055 712862.5 9559972 1902.48 29.0168746   -0.00720    0.01250   3686.1240
#> 29472 714792.5 9561072 1786.08 36.1731174   -0.01029   -0.02312   2657.6936
#> 22566 712832.5 9559842 1897.49 33.0602377    0.00672    0.04398    267.2659
#> 36030 712852.5 9559572 1938.79 28.6788295   -0.01440    0.00710    843.7961
#> 26215 714332.5 9560262 2138.08 25.7212213   -0.00491    0.00581   1855.1630
#> 6676  715092.5 9559912 2217.44 41.7078897   -0.00077   -0.00284   2010.4285
#> 47210 714952.5 9557612 2744.35 40.8845494   -0.00493    0.00733    598.7983
#> 3453  713482.5 9559642 2145.11 35.1257506    0.00108   -0.00678   1891.6584
#> 4477  715432.5 9559392 2325.25 37.0291164   -0.00521   -0.00739    558.2286
#> 20194 714132.5 9558452 2447.22 24.0224015   -0.00987    0.02107   1392.5880
#> 8073  712572.5 9560302 1917.62 52.0090979    0.01937    0.00083    994.7876
#> 19730 715332.5 9558452 2640.35 47.6425866   -0.00211    0.00642    613.6390
#> 12084 712832.5 9559992 1872.46 56.0547529   -0.03184    0.03924   8023.8657
#> 34642 714962.5 9559862 2282.08 22.1608616   -0.00028    0.01318    289.0063
#> 45104 715062.5 9559242 2345.02 24.1564736   -0.01743    0.01313    783.4548
#> 15990 712852.5 9559982 1894.77 37.6943204   -0.01545    0.02405   5681.2412
#> 37828 713482.5 9558652 2330.58 20.4454260    0.00606   -0.00556    660.3690
#> 19784 714982.5 9557722 2716.98 35.4362937    0.00888    0.00862    874.5408
#> 27084 713632.5 9558472 2380.37 23.1354628   -0.01752   -0.00517   1000.7119
#> 47091 713382.5 9559182 2122.71 48.0379275    0.04527    0.04613    289.6113
#> 18751 714882.5 9559182 2381.81 39.2991752    0.00066   -0.00036    516.4709
#> 5578  714962.5 9557612 2751.95 39.1914591    0.00288    0.01452    437.3965
#> 28743 715202.5 9558872 2463.87 34.6570711   -0.01267   -0.05063  18733.8848
#> 17423 715592.5 9558562 2654.69 33.2195200   -0.00034    0.00294    739.5768
#> 37141 714382.5 9559822 2219.36 10.1064025    0.01661    0.01199    196.4080
#> 18394 715022.5 9559942 2216.16 42.1376081   -0.00613    0.00903   2360.1536
#> 651   714602.5 9560152 2111.23 53.8517302   -0.01149    0.01949   1182.1506
#> 26567 714022.5 9558872 2327.42 46.6656935    0.00290    0.00230   1056.5876
#> 23029 713502.5 9559102 2174.07 56.4575423    0.03395    0.01495    266.5888
#> 34707 713962.5 9561092 1819.49 65.6861735   -0.04749    0.04779   2972.0193
#> 39599 714112.5 9558562 2374.25 29.1681354    0.00283    0.00117   1018.2141
#> 20220 714052.5 9558522 2378.67 39.6561279   -0.00855   -0.02175   1122.9120
#> 21053 712862.5 9559992 1901.01 30.9506071   -0.00762    0.01252   5761.3486
#> 10507 714892.5 9559282 2374.45 42.8423462   -0.01057    0.02427    380.9472
#> 13029 713542.5 9560242 2019.80 29.5451417    0.00576   -0.00166    650.6246
#> 25700 715172.5 9559272 2324.49 35.5348425   -0.00529    0.02009   1247.7061
#> 47100 712752.5 9560462 1910.00 41.9118627   -0.00032    0.00542   3614.7334
#> 29801 714392.5 9559162 2356.18 31.5625261   -0.00856   -0.01625   1063.9150
#> 32735 714922.5 9559002 2414.49 40.0646468    0.00062   -0.00262   1760.2162
#> 30886 714912.5 9558822 2563.86 49.7539361   -0.00099    0.02169    495.8840
#> 664   714452.5 9559852 2174.83 44.8608765    0.00282    0.00109    767.4097
#> 16355 712962.5 9560212 1992.88 39.5260665    0.01927   -0.00316    454.9674
#> 5103  714612.5 9559892 2220.36 49.0079450   -0.00078    0.00219    420.7032
#> 24501 713302.5 9559142 2125.66 27.1089251    0.01784    0.00596    842.0994
#> 41783 715572.5 9558652 2609.51 32.4082118    0.00291   -0.00081    940.6002
#> 5239  714832.5 9560992 1859.70 19.9601307   -0.00320    0.00180   3904.5254
#> 31834 713492.5 9560682 1800.85 34.2370930   -0.00472   -0.00228   2262.3796
#> 7385  713752.5 9560122 2064.61 31.7710827   -0.02605   -0.03594  16587.9980
#> 47366 714892.5 9557612 2718.78 28.0239387    0.00538    0.00112    745.3315
#> 12137 712842.5 9558902 2097.88 30.0986189    0.00024   -0.00474   2343.2100
#> 37714 713192.5 9560692 1885.05 18.0344196    0.00188    0.00832   3022.3462
#> 18958 714192.5 9558372 2495.72 34.6771246   -0.00178   -0.00822    472.8748
#> 11612 715212.5 9557752 2820.51 39.1622383    0.00017    0.01213    328.7922
#> 10073 715572.5 9558612 2632.78 32.1612670    0.00129    0.00171    788.0004
#> 19892 714482.5 9560562 2059.63 19.8151724   -0.00332   -0.00058   1817.2371
#> 47718 712812.5 9559892 1867.55 54.9386311   -0.01491    0.01061   5744.7583
#> 17345 713922.5 9559682 2227.44 34.4909134   -0.06509   -0.01211  49347.5547
#> 27244 714162.5 9558622 2316.00 36.7512319   -0.02556   -0.11694 102301.4609
#> 31117 712892.5 9558922 2129.84 30.5506826    0.00521   -0.00870   1895.1399
#> 16009 715122.5 9559402 2264.81 50.0186426   -0.00589    0.00849   1120.5961
#> 30287 713992.5 9557822 2403.92 18.2154742    0.03890   -0.00911    173.2806
#> 41125 712772.5 9559412 2021.88 38.8763323   -0.00310   -0.00860    695.2302
#> 42201 714642.5 9557452 2592.29 13.2255848   -0.02084   -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.1804123   -0.01059   -0.07431   4079.0366
#> 34887 714662.5 9557392 2609.02 51.0207457   -0.01878   -0.04351   1442.9247
#> 33875 712832.5 9559852 1892.81 37.5545187   -0.01175    0.04195   1901.6743
#> 17058 714032.5 9558502 2402.95 31.3528235    0.01924    0.02606    564.6876
#> 37439 714602.5 9560672 1982.99 18.9064613   -0.00070    0.00010   3467.2690
#> 332   714852.5 9561102 1787.11 45.3582039    0.00540    0.00879   1200.4210
#> 2801  715252.5 9559612 2199.28 24.4034184    0.03051    0.02190    264.6404
#> 4667  712652.5 9560472 1957.89 36.3243783   -0.00296    0.00426   1315.5334
#> 750   712812.5 9560032 1847.09 54.9592576   -0.00767   -0.03713   6379.7974
#> 7409  714772.5 9559122 2404.90 30.2155023   -0.01016   -0.00294    734.9908
#> 24723 712942.5 9560202 1983.18 33.0716969    0.00285   -0.01325    595.0426
#> 29024 713492.5 9559112 2173.47 53.4323251    0.04864    0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.3064209   -0.00019    0.00119   6589.5796
#> 4479  713982.5 9557812 2399.24 41.2311889    0.01895    0.04744    226.7911
#> 43536 713782.5 9558732 2281.50 32.7491217    0.00840    0.00189    556.5211
#> 11332 712742.5 9560072 1850.34 22.7000149    0.00683   -0.00603   1562.0503
#> 34718 714802.5 9558802 2599.22 44.6500280    0.01996    0.04684    209.1778
#> 10785 714052.5 9557852 2377.76 20.9840063   -0.00025   -0.05345   2417.1389
#> 36744 714922.5 9557722 2692.39 25.9807712    0.00316   -0.00427    866.9398
#> 23176 714012.5 9558892 2308.59 43.0291304   -0.01084   -0.02996   1599.3491
#> 10145 713782.5 9560932 1865.62 23.5955479   -0.00203    0.00642   1688.1782
#> 16208 712882.5 9560332 1846.34 48.6252092    0.00159   -0.01559    866.5775
#> 29431 714882.5 9559892 2258.03 21.9179275   -0.00849    0.00150    503.5590
#> 46165 714392.5 9557912 2542.56 16.3298701    0.00325   -0.00505    351.4078
#> 25212 714272.5 9559012 2433.45 37.6714021   -0.01909   -0.01181   1286.3949
#> 34250 713822.5 9559142 2316.36 33.7598192   -0.02284    0.01085   2229.5105
#> 2664  713342.5 9558572 2275.86 56.9754961    0.00872    0.05419    761.5427
#> 16464 715252.5 9559982 2143.24 23.1858831   -0.00298   -0.00712   1010.5097
#> 20458 713612.5 9559562 2288.02 33.4401088    0.00112    0.03538    265.4677
#> 7517  714002.5 9561142 1772.85 26.1893279    0.00815   -0.03345 665381.5625
#> 43426 713782.5 9560822 1876.35 23.8230122   -0.00345   -0.00075   8141.7549
#> 42329 715852.5 9558702 2651.92 22.8707563    0.00321    0.00548    532.8210
#> 16542 713772.5 9559622 2304.85 30.8102325   -0.00712    0.00923    811.5580
#> 29125 714992.5 9558792 2584.41 24.6658331    0.01253   -0.00333    422.8840
#> 16878 714232.5 9558172 2485.33 22.8083039    0.00818    0.00502   1428.8546
#> 18627 714292.5 9559252 2330.78 18.4962235   -0.00538   -0.00052   3508.5591
#> 40139 712852.5 9560072 1907.89 43.7040747    0.00001    0.00759   1880.4403
#> 21925 714542.5 9559232 2314.33 50.2478257    0.00737    0.01033   3446.5520
#> 38689 714622.5 9559342 2264.90 42.6481135   -0.00844    0.00514    897.9290
#> 20319 715282.5 9560482 1901.16 27.6767263    0.00706   -0.03035   1471.9783
#> 29527 714332.5 9557502 2567.02 48.9048126   -0.00452   -0.01298    436.1459
#> 6143  712762.5 9559022 2096.24 29.8963648    0.00358    0.02422    967.9738
#> 8895  714132.5 9557692 2487.19 37.3642967    0.03423   -0.00983    270.7949
#> 48869 712562.5 9560122 1972.89 24.9465824    0.00185    0.01655    560.8554
#> 34228 715932.5 9558592 2690.97  9.5351636  197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.2880302   -0.00393    0.00032    994.9224
#> 29107 713172.5 9559272 2067.58 28.2640080    0.02413    0.00528    355.3475
#> 39440 714222.5 9557872 2508.52 39.0476467    0.02785    0.00645    234.0114
#> 12044 713612.5 9559292 2247.76 23.0747293   -0.00112   -0.00038   2017.3547
#> 13908 715022.5 9559372 2290.81 27.2429972   -0.03004    0.00714 480535.5000
#> 26389 713002.5 9559712 1926.02 24.0372984    0.00586    0.00654   1576.2065
#> 45486 712762.5 9560172 1853.82 56.7973062    0.01978    0.01282   2322.7944
#> 33410 715342.5 9560452 1942.88 29.0930780   -0.00305   -0.00705    968.9738
#> 330   714232.5 9560362 2165.00 18.0607756    0.01142    0.00868    235.8511
#> 43222 714472.5 9560712 2039.25 22.1906554    0.01858    0.00312    254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402    0.00009   -0.00599  12873.8018
#> 48988 714682.5 9560852 1942.97 20.2907274    0.00480    0.00740    909.4791
#> 48845 713232.5 9561322 2041.83 29.9124076 -100.76649  100.79199    107.2986
#> 43415 714722.5 9558642 2657.99  0.2314749    0.00799    0.00911    100.0000
#> 15008 714492.5 9559012 2449.93 29.5153479    0.00972    0.00308    422.5022
#> 14671 712672.5 9560372 1872.21 39.1971887   -0.00873   -0.01417   1520.6630
#> 29716 715272.5 9559992 2144.94 25.7177836    0.00262   -0.00502    520.5218
#> 28286 713702.5 9559842 2245.07 27.9689348    0.00511    0.01499    617.3855
#> 40750 712962.5 9558702 2126.06 34.7040537   -0.00264    0.00624   1870.8307
#> 18246 713902.5 9561342 1818.52 19.1104343   -0.01502    0.00282   1659.6211
#> 19767 714272.5 9558052 2431.67 55.7247292    0.00361   -0.01841    756.0055
#> 34334 714642.5 9557912 2582.25 30.9912871    0.00253   -0.01233    968.7780
#> 43972 713642.5 9558702 2304.16 24.0235474    0.00340    0.01809   1158.0426
#> 40653 714262.5 9561262 1791.94 12.2223357    0.00161    0.00709   1181.2555
#> 46148 715082.5 9558042 2725.39 42.8360436   -0.00181   -0.00799   1036.6606
#> 30403 713952.5 9560752 2003.57 42.9643862   -0.00933   -0.00157    766.0220
#> 15880 715892.5 9557572 3113.36 53.0432868    0.00962   -0.00172    351.5812
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 47223 36.2590611   300.00   TRUE       300.00            100  3.120164
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 8059  26.9771448   300.00   TRUE       300.00              4  3.150244
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 37104  4.8661306    50.99   TRUE         0.00             21  3.252198
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 7349  26.3228270   300.00   TRUE       300.00              8  2.751808
#> 39629 47.5486215   300.00   TRUE       300.00             85  3.366461
#> 11463 20.1337369   300.00   TRUE       300.00            100  2.768442
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 48055 33.8257094   300.00   TRUE         1.67              0  3.566570
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 22566 24.5867649   300.00   TRUE         7.49              0  2.426944
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 26215 23.8316065   300.00   TRUE       300.00             10  3.268382
#> 6676  34.6215477   300.00   TRUE       300.00            100  3.303289
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 20194 28.1173308   300.00   TRUE       300.00             41  3.143823
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 34642 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 15990 34.1488575   300.00   TRUE         0.00              0  3.754443
#> 37828 18.7637948   300.00   TRUE       300.00              2  2.819787
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 27084 21.6727016   300.00   TRUE       300.00             81  3.000309
#> 47091 40.1517364   300.00   TRUE       300.00             25  2.461815
#> 18751 38.0759103   300.00   TRUE       300.00            100  2.713046
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 17423 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 37141 12.3884934   300.00   TRUE       300.00              1  2.293159
#> 18394 28.6106475   300.00   TRUE       300.00             57  3.372940
#> 651   45.5243616   300.00   TRUE       300.00            100  3.072673
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 23029 46.6015223   300.00   TRUE       300.00             34  2.425842
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 39599 33.1347223   300.00   TRUE       300.00             25  3.007839
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21053 34.2514170   300.00   TRUE         0.00              2  3.760524
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 13029 34.1878187   300.00   TRUE        96.65              4  2.813330
#> 25700 32.5411380   300.00   TRUE       300.00            100  3.096112
#> 47100 29.6459822    82.80   TRUE       129.73             36  3.558076
#> 29801 33.2418654   300.00   TRUE       300.00              2  3.026907
#> 32735 39.3581898   300.00   TRUE       300.00            100  3.245566
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 5103  34.9131832   300.00   TRUE       300.00             91  2.623976
#> 24501 27.1524699   300.00   TRUE       300.00              0  2.925363
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 5239  24.6652601   197.29   TRUE         4.67             57  3.591568
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 7385  28.7103422   300.00   TRUE       300.00            100  4.219794
#> 47366 28.7275309   300.00   TRUE       300.00            100  2.872349
#> 12137 27.5884908   300.00   TRUE       300.00            100  3.369811
#> 37714 30.7317373    13.57   TRUE       170.00              4  3.480344
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 10073 25.7613284   300.00   TRUE       300.00            100  2.896526
#> 19892 27.0264192   300.00   TRUE       300.00             92  3.259412
#> 47718 32.3234140   300.00   TRUE        10.00              1  3.759272
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 27244 37.7751074   300.00   TRUE       300.00             54  5.009882
#> 31117 26.6459752   300.00   TRUE       291.23            100  3.277641
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 41125 36.5558532   300.00   TRUE         0.00            100  2.842129
#> 42201 39.5587250   300.00   TRUE       300.00            100  3.989241
#> 4588  31.6645125   300.00   TRUE       300.00              1  3.610558
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 17058 26.3228270   300.00   TRUE       300.00              8  2.751808
#> 37439 26.5926901   300.00   TRUE       300.00            100  3.539988
#> 332    0.5569150    94.82   TRUE         0.00             90  3.079334
#> 2801  25.7464315   300.00   TRUE       300.00             26  2.422656
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 7409  25.3705712   300.00   TRUE       300.00            100  2.866282
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 29024 42.5730560   300.00   TRUE       300.00             32  2.529363
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 4479  32.0117250   300.00   TRUE       300.00             55  2.355626
#> 43536 29.0947968   300.00   TRUE       300.00              8  2.745482
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 34718 34.4095533   300.00   TRUE       300.00            100  2.320516
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 36744 29.4047670   300.00   TRUE       300.00            100  2.937989
#> 23176 41.5583478   300.00   TRUE       300.00            100  3.203943
#> 10145 27.4160305   276.16  FALSE        91.33            100  3.227418
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 29431 20.8917601   300.00  FALSE       300.00             42  2.702050
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 25212 38.9829025   300.00  FALSE       300.00            100  3.109374
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 16464 35.0461095   300.00  FALSE       300.00            100  3.004540
#> 20458 21.4297675   300.00  FALSE       300.00             28  2.424012
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 43426 35.4626498   300.00  FALSE        82.11            100  3.910718
#> 42329 19.2278906   300.00  FALSE       300.00            100  2.726581
#> 16542 21.1123488   300.00  FALSE       300.00            100  2.909320
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 18627 31.2571396   300.00  FALSE       300.00             20  3.545129
#> 40139 33.1433166   300.00  FALSE         1.11             21  3.274260
#> 21925 31.2485452   300.00  FALSE       300.00             95  3.537385
#> 38689 36.0843090   300.00  FALSE       300.00             18  2.953242
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 29527 36.7265947   300.00  FALSE       300.00            100  2.639632
#> 6143  28.7573247   300.00  FALSE       185.58            100  2.985864
#> 8895  38.0776291   300.00  FALSE       300.00             90  2.432641
#> 48869 24.8755356   138.98  FALSE         0.00             63  2.748851
#> 34228 10.7486882   300.00  FALSE       300.00            100  2.427333
#> 34810 15.5546582   300.00  FALSE       300.00            100  2.997789
#> 29107 25.2307695   300.00  FALSE       300.00            100  2.550653
#> 39440 32.7061498   300.00  FALSE       300.00             92  2.369237
#> 12044 32.2368974   300.00  FALSE       300.00            100  3.304782
#> 13908 34.3774677   300.00  FALSE       300.00             96  5.681725
#> 26389 36.1221242   300.00  FALSE        76.38             26  3.197613
#> 45486 33.4217741   300.00  FALSE         0.00             68  3.366011
#> 33410 32.8797560   300.00  FALSE       300.00            100  2.986312
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 48988 18.9586005   300.00  FALSE       144.19            100  2.958793
#> 48845 31.5590883   169.72  FALSE         0.00            100  2.030594
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 15008 26.9078169   300.00  FALSE       300.00             48  2.625829
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 29716 32.6935447   300.00  FALSE       300.00            100  2.716439
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 40750 32.0999605   300.00  FALSE       300.00            100  3.272034
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 19767 42.0104115   300.00  FALSE       300.00            100  2.878525
#> 34334 25.0898218   300.00  FALSE       300.00            100  2.986224
#> 43972 22.8524217   300.00  FALSE       300.00            100  3.063725
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 30403 36.1307186   300.00  FALSE       209.91            100  2.884241
#> 15880 48.8967912   300.00  FALSE       300.00            100  2.546026