Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope      hcurv     vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.247190   -0.00501  -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.476603   -0.01919  -0.04051  351155.1250
#> 31357 714752.5 9561022 1848.22 33.446411   -0.00347   0.02357    1752.0375
#> 15762 714922.5 9558862 2535.97 54.910556    0.01305  -0.00555     430.4604
#> 28223 712742.5 9560482 1927.69 34.667957   -0.00292   0.00712    3636.5303
#> 29391 712842.5 9560152 1930.16 28.095558    0.00753   0.01487    1077.7585
#> 1300  714042.5 9558482 2408.27 24.148452    0.00659   0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.583935   -0.00498  -0.00752    1321.5997
#> 15843 714042.5 9558902 2332.30 50.097711   -0.00858  -0.00292    1233.9299
#> 9669  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 38411 713552.5 9560652 1851.63 46.005073   -0.00891  -0.00949    6217.8213
#> 27973 715022.5 9559332 2308.28 20.063263    0.00940   0.00679     550.0344
#> 26531 714462.5 9560452 2104.74 19.524110    0.00453   0.00257     357.1436
#> 8869  715062.5 9559392 2278.56 41.401930    0.00310   0.01440    1088.8639
#> 32622 714972.5 9557762 2687.38 30.556412   -0.00730  -0.01491    1844.4353
#> 43327 713962.5 9561082 1839.11 63.361684   -0.04951  -0.00529    3282.1186
#> 38827 713412.5 9560472 1869.10 16.871315   -0.00156   0.00406    2421.1145
#> 16811 714902.5 9559262 2362.67 50.703327   -0.01407   0.00547     519.4228
#> 39061 712922.5 9559912 1931.50 36.315211    0.00414  -0.00104    1641.4587
#> 43964 714802.5 9558452 2684.33 37.982518   -0.02123   0.01164     895.6606
#> 40307 713302.5 9560702 1845.07 39.004675    0.00129   0.02510    3342.5518
#> 16843 714882.5 9558512 2701.78 26.637954   -0.00270  -0.00769    1032.7177
#> 14522 715012.5 9557732 2721.34 35.181327   -0.00382   0.00672    1114.8801
#> 35942 712862.5 9558902 2109.97 31.400379    0.00151   0.00189    2317.1819
#> 32821 715022.5 9559952 2208.33 43.416450   -0.00643   0.00314    2360.1501
#> 3453  713482.5 9559642 2145.11 35.125751    0.00108  -0.00678    1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116   -0.00521  -0.00739     558.2286
#> 10320 714912.5 9558482 2724.21 25.282909    0.00356   0.00774     527.1068
#> 34280 713652.5 9560702 1896.37 41.273588    0.00382   0.00088    2322.1746
#> 23031 714042.5 9558492 2404.94 28.496056    0.01205   0.02205     629.4165
#> 6763  714692.5 9561052 1830.60 48.394880   -0.01996   0.00626    1014.7159
#> 17004 713102.5 9559032 2186.79 36.533508   -0.01383   0.00193    1809.3838
#> 1635  713102.5 9559052 2172.87 31.484031   -0.00064  -0.00916    2009.1686
#> 28577 712812.5 9559962 1849.82 59.780125    0.00646  -0.04086    6042.5757
#> 6557  713312.5 9559022 2181.69 27.634327   -0.00669  -0.00991    1403.0193
#> 47429 714962.5 9557832 2705.12 32.057562    0.02224   0.00177     291.5937
#> 11687 714552.5 9560382 2052.66 29.575508   -0.00255  -0.00365    1536.2647
#> 37828 713482.5 9558652 2330.58 20.445426    0.00606  -0.00556     660.3690
#> 18564 715552.5 9558592 2648.36 26.291314    0.00316   0.00714     475.3120
#> 6967  715342.5 9560192 1995.36 44.881503   -0.00666  -0.00504    1020.2766
#> 17036 712832.5 9560582 1951.38 27.645787    0.00060   0.00270    4057.5637
#> 22332 715482.5 9558922 2475.28 19.936639   -0.02617   0.00307  147470.5312
#> 651   714602.5 9560152 2111.23 53.851730   -0.01149   0.01949    1182.1506
#> 26567 714022.5 9558872 2327.42 46.665694    0.00290   0.00230    1056.5876
#> 25244 713472.5 9559092 2137.03 18.348973   -0.10372  -0.00008  825875.6875
#> 5084  714912.5 9559302 2359.77 47.968027    0.00107   0.02373     527.3528
#> 35194 713692.5 9559952 2185.68 28.796859   -0.00449  -0.00181    2406.1252
#> 18892 712682.5 9560202 1839.08 40.923511   -0.01364  -0.03357     473.4709
#> 42941 713252.5 9559742 1990.10 28.954422   -0.04708   0.00408   76318.0156
#> 30406 715312.5 9560152 2029.10 34.137398    0.00953  -0.02283     590.1536
#> 27875 714832.5 9561032 1838.22 41.212281   -0.01160   0.01690    5121.6924
#> 18265 715522.5 9558782 2539.23 35.856272   -0.00458  -0.00401    1722.6481
#> 6317  715042.5 9557702 2745.36 27.873824    0.00090  -0.00299     688.0706
#> 28845 714892.5 9559272 2374.42 45.750107   -0.00209   0.02739     365.7626
#> 35083 714612.5 9558152 2596.12 60.598308    0.04548   0.05543     352.1558
#> 23213 715592.5 9558562 2654.69 33.219520   -0.00034   0.00294     739.5768
#> 10364 714762.5 9561002 1856.21 25.274887   -0.00845   0.00454    3476.0818
#> 31121 713462.5 9559132 2135.56 61.596974   -0.03800  -0.06020    1563.5549
#> 21013 712742.5 9560032 1845.53 29.352628    0.00568  -0.01558    1504.9717
#> 13146 713792.5 9558552 2341.40 29.825318   -0.00104   0.01664    1898.5588
#> 2552  714942.5 9557652 2719.32 34.873649   -0.01238  -0.00851    1012.5963
#> 40270 713312.5 9559012 2186.56 30.935710   -0.00771  -0.00599    1378.5760
#> 7385  713752.5 9560122 2064.61 31.771083   -0.02605  -0.03594   16587.9980
#> 37769 714412.5 9558562 2382.08 39.876717   -0.04601  -0.08389   47022.9688
#> 30268 714172.5 9558612 2328.84 54.950663    0.03743  -0.05242     507.9263
#> 8125  713452.5 9559132 2120.44 42.140473   -0.05458  -0.14813  287033.4688
#> 47320 713422.5 9559182 2143.03 56.325189    0.01246   0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135   -0.00759  -0.00400     608.6523
#> 10933 712702.5 9560132 1854.55 36.291147   -0.00204  -0.00376    2048.3198
#> 20368 715252.5 9560622 1865.83 33.572462    0.00025  -0.01245    6353.3179
#> 18440 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 42201 714642.5 9557452 2592.29 13.225585   -0.02084  -0.02236    9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412   -0.01059  -0.07431    4079.0366
#> 27915 715232.5 9559552 2208.93 25.480579    0.00171  -0.00261    1627.8989
#> 15432 715632.5 9558922 2509.67 35.827624   -0.00444  -0.00017     759.9937
#> 29228 712782.5 9560422 1870.56 32.906112   -0.00218  -0.00372    4154.5786
#> 22907 713172.5 9559612 2056.81 22.272588   -0.00166   0.01476    1404.4454
#> 26101 715562.5 9558682 2594.87 32.411077    0.00079   0.00321    1170.5904
#> 34164 714212.5 9558492 2421.45 51.329570    0.02247   0.01123     364.2903
#> 47244 714262.5 9559302 2350.47 17.059245    0.02213  -0.01363     565.5207
#> 19756 715572.5 9558972 2485.95 38.812734    0.00546   0.01484     495.0292
#> 10363 713252.5 9559722 1999.55 41.829930   -0.03669  -0.01981    3975.3608
#> 13259 714862.5 9560992 1859.28 21.267620   -0.00317   0.00077    3056.9475
#> 49378 713252.5 9560732 1875.41 27.835436    0.00181  -0.00001    1184.7952
#> 4667  712652.5 9560472 1957.89 36.324378   -0.00296   0.00426    1315.5334
#> 1916  715202.5 9559512 2213.91 26.327984   -0.00292  -0.00109    2047.9031
#> 34984 715042.5 9559312 2320.49 42.857816   -0.01106  -0.04634     500.5027
#> 9606  714392.5 9559172 2353.05 30.783303   -0.00185  -0.01185     953.8614
#> 26058 713402.5 9560442 1880.12 31.377461    0.00730  -0.02610    1872.9276
#> 38103 715392.5 9560162 1998.16 46.285823    0.00708  -0.02538     807.8237
#> 29024 713492.5 9559112 2173.47 53.432325    0.04864   0.04226     338.3474
#> 28260 715312.5 9558302 2681.75 34.484038   -0.01125  -0.02685    2568.4631
#> 46627 714852.5 9558932 2459.24 53.445503    0.02451  -0.03921     525.3245
#> 7771  715232.5 9560922 1816.88 37.180377    0.00124   0.01166    1813.7841
#> 28588 715532.5 9559362 2299.84 11.144602   -0.04360  -0.03000  162178.1562
#> 42640 714332.5 9558522 2403.10 38.871176   -0.00212   0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.113544   -0.00118  -0.00252    3631.7017
#> 11057 714962.5 9557642 2737.48 42.685929   -0.00488  -0.00632     620.0229
#> 35458 715242.5 9558442 2623.01 38.954828   -0.00770  -0.03400    1982.7782
#> 41310 714452.5 9560402 2092.27 12.462978   -0.00195  -0.00205    1112.2626
#> 43536 713782.5 9558732 2281.50 32.749122    0.00840   0.00189     556.5211
#> 23832 714252.5 9560182 2188.12 30.647512    0.00427   0.02403     474.8613
#> 10785 714052.5 9557852 2377.76 20.984006   -0.00025  -0.05345    2417.1389
#> 36765 715302.5 9560262 1952.73 27.203463   -0.00585  -0.02665    1904.4769
#> 13072 714862.5 9558932 2448.29 41.052426   -0.04989  -0.02781   29995.1543
#> 36561 715292.5 9560802 1844.64 31.937240    0.00002  -0.01232     977.7320
#> 46165 714392.5 9557912 2542.56 16.329870    0.00325  -0.00505     351.4078
#> 15558 714732.5 9559502 2283.66 53.431752    0.00390  -0.02831     454.9113
#> 27965 715042.5 9560342 2063.39 29.197356    0.04336   0.01403     241.2991
#> 7517  714002.5 9561142 1772.85 26.189328    0.00815  -0.03345  665381.5625
#> 441   714312.5 9560972 1916.59 31.302976    0.00540   0.00650    1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917    0.01306  -0.00626     450.7346
#> 39796 714802.5 9560372 2046.48 40.155747    0.00176   0.02973     215.5101
#> 36833 713322.5 9560572 1794.25 33.754090    0.00527  -0.02077    8759.9961
#> 18627 714292.5 9559252 2330.78 18.496224   -0.00538  -0.00052    3508.5591
#> 4007  714132.5 9558412 2464.44 43.720118    0.00174  -0.03065     704.4843
#> 31884 715782.5 9558612 2632.60 30.279674    0.00084  -0.00083    1309.2347
#> 29944 713562.5 9558952 2209.55 17.339995    0.01432  -0.00362     283.7090
#> 31201 714672.5 9560022 2224.47 38.863727    0.01180   0.01170     253.5129
#> 29527 714332.5 9557502 2567.02 48.904813   -0.00452  -0.01298     436.1459
#> 8895  714132.5 9557692 2487.19 37.364297    0.03423  -0.00983     270.7949
#> 24009 714042.5 9558962 2357.58 34.361998    0.05041  -0.00011     201.1272
#> 14148 714162.5 9560352 2149.57 37.133395    0.00424   0.00196     576.1450
#> 43222 714472.5 9560712 2039.25 22.190655    0.01858   0.00312     254.0656
#> 36889 714302.5 9559802 2243.60 21.166207    0.01923   0.05547     128.2613
#> 48845 713232.5 9561322 2041.83 29.912408 -100.76649 100.79199     107.2986
#> 26768 715092.5 9560582 1960.72 36.656121    0.00980  -0.00710     472.0236
#> 5861  713882.5 9559182 2350.43 27.393685   -0.02608   0.00148   10771.1963
#> 3724  714322.5 9557402 2603.46 34.597484    0.02523   0.02247     282.1076
#> 40612 713542.5 9561152 1856.15 33.568451    0.00420   0.01960     884.2003
#> 32430 715832.5 9558112 2824.39 23.694097    0.00944   0.00836     250.5111
#> 974   713402.5 9561062 1854.56 23.634509    0.00082   0.02578    1258.8865
#> 44216 715772.5 9558402 2710.44 30.444685   -0.00070  -0.00050    1471.4346
#> 2999  714592.5 9558492 2515.83 47.031814    0.01600  -0.01849     764.8193
#> 42638 714322.5 9561182 1754.22  4.172279   -0.03757  -0.00003 5426606.0000
#> 15138 713702.5 9561122 1780.68 49.221658    0.00607  -0.04566   12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948    0.01182  -0.00972     331.7462
#> 7132  713192.5 9559612 2060.27 20.773731    0.00452   0.00368    1071.8734
#> 37783 713832.5 9557892 2405.28 40.886841    0.01572   0.03148     481.8292
#> 10981 713832.5 9560022 2125.33 34.848439   -0.06909  -0.01151  143213.1562
#> 11394 713472.5 9561232 1889.48 34.314442    0.00108  -0.00948    1538.1741
#> 43972 713642.5 9558702 2304.16 24.023547    0.00340   0.01809    1158.0426
#> 37254 713482.5 9560892 1823.73 39.882446    0.01014  -0.00884     672.2466
#> 14645 715072.5 9560262 2034.04 38.169302   -0.00716  -0.03234    1707.3130
#> 15880 715892.5 9557572 3113.36 53.043287    0.00962  -0.00172     351.5812
#> 126   715042.5 9561262 1733.49 30.110651   -0.00288   0.00448     941.8503
#> 28978 713982.5 9560462 2133.10  5.229386    0.00366   0.00525     162.0592
#> 46567 714472.5 9560352 2080.13 21.381639    0.01336   0.01403     609.8665
#> 15363 714742.5 9559212 2373.13 33.583348   -0.00247   0.01057     722.3952
#> 40852 715392.5 9557932 2939.71 18.297407    0.00505   0.00855     229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 21788 32.81443884   300.00   TRUE       300.00             40  5.545499
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 28223 29.02260415    60.46   TRUE       152.09             35  3.560687
#> 29391 28.28348860   300.00   TRUE         0.56            100  3.032521
#> 1300  27.53062206   300.00   TRUE       300.00              6  2.888165
#> 8598  39.59825914   300.00   TRUE       300.00            100  3.121100
#> 15843 42.22183288   300.00   TRUE       300.00            100  3.091290
#> 9669  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 38411 32.75828898   279.54   TRUE         0.00            100  3.793638
#> 27973 30.00981044   300.00   TRUE       300.00             85  2.740390
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 32622 30.17138453   300.00   TRUE       300.00            100  3.265863
#> 43327 29.05870050   173.40   TRUE        33.81            100  3.516154
#> 38827 31.03597785   300.00   TRUE        52.52             39  3.384015
#> 16811 35.25237426   300.00   TRUE       300.00             59  2.715521
#> 39061 35.24091511   300.00   TRUE        70.65             56  3.215230
#> 43964 28.75732470   300.00   TRUE       300.00             46  2.952143
#> 40307 29.51362898    77.82   TRUE        76.41             54  3.524078
#> 16843 21.39825477   300.00   TRUE       300.00             12  3.013982
#> 14522 28.39865312   300.00   TRUE       300.00            100  3.047228
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 32821 29.48326222   300.00   TRUE       300.00             65  3.372940
#> 3453  40.08470031   300.00   TRUE       300.00              2  3.276843
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 10320 16.45592083   300.00   TRUE       300.00             30  2.721899
#> 34280 33.62116342   300.00   TRUE         0.00            100  3.365895
#> 23031 26.74108621   300.00   TRUE       300.00              2  2.798938
#> 6763  16.21814335   125.47   TRUE         0.00              6  3.006344
#> 17004 27.18913921   300.00   TRUE       300.00            100  3.257531
#> 1635  27.82684124   300.00   TRUE       300.00            100  3.303016
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 6557  27.41946824   300.00   TRUE       300.00              5  3.147064
#> 47429 32.09766864   300.00   TRUE       300.00            100  2.464778
#> 11687 23.19390450   300.00   TRUE       300.00             68  3.186466
#> 37828 18.76379483   300.00   TRUE       300.00              2  2.819787
#> 18564 21.70020353   300.00   TRUE       300.00            100  2.676979
#> 6967  41.11946208   300.00   TRUE       300.00             85  3.008718
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 22332 27.06881807   300.00   TRUE       300.00            100  5.168705
#> 651   45.52436161   300.00   TRUE       300.00            100  3.072673
#> 26567 40.43706935   300.00   TRUE       300.00            100  3.023906
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 5084  31.59518465   300.00   TRUE       300.00             18  2.722101
#> 35194 23.44944368   300.00   TRUE       300.00            100  3.381318
#> 18892 27.06308850   235.02   TRUE        27.24              0  2.675293
#> 42941 39.06311656   300.00   TRUE       300.00             17  4.882627
#> 30406 43.10075014   300.00   TRUE       300.00             85  2.770965
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 18265 31.44564267   300.00   TRUE       300.00            100  3.236197
#> 6317  27.32321133   300.00   TRUE       300.00            100  2.837633
#> 28845 29.84250676   300.00   TRUE       300.00             54  2.563199
#> 35083 43.54078173   300.00   TRUE       300.00             23  2.546735
#> 23213 25.46052554   300.00   TRUE       300.00            100  2.868983
#> 10364 26.09307095   179.84   TRUE         0.00              0  3.541090
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 21013 24.96319818   300.00   TRUE        13.47             64  3.177528
#> 13146 33.57647271   300.00   TRUE       300.00              2  3.278424
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 7385  28.71034216   300.00   TRUE       300.00            100  4.219794
#> 37769 37.42503022   300.00   TRUE       300.00            100  4.672310
#> 30268 47.27245585   300.00   TRUE       300.00             41  2.705801
#> 8125  33.18743437   300.00   TRUE       300.00             48  5.457933
#> 47320 45.66416331   300.00   TRUE       300.00             43  2.536741
#> 10467 38.99608049   300.00   TRUE       300.00             90  2.784369
#> 10933 34.67712463   264.46   TRUE        18.17             56  3.311398
#> 20368 30.34441779   300.00   TRUE       166.26            100  3.803001
#> 18440 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 27915 33.57475384   300.00   TRUE       300.00              6  3.211627
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 22907 33.81367724   300.00   TRUE       205.28              1  3.147505
#> 26101 28.64617088   300.00   TRUE       300.00            100  3.068405
#> 34164 39.71342365   300.00   TRUE       300.00              1  2.561448
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 19756 30.11638059   300.00   TRUE       300.00            100  2.694631
#> 10363 45.09349735   300.00   TRUE       297.43              6  3.599377
#> 13259 23.16067295   203.85   TRUE        10.45             74  3.485288
#> 49378 33.99415894    22.68   TRUE       134.63             25  3.073643
#> 4667  31.51726239    24.78   TRUE       172.24             38  3.119102
#> 1916  29.27355967   300.00   TRUE       300.00             21  3.311309
#> 34984 33.90592344   300.00   TRUE       300.00            100  2.699406
#> 9606  33.82399048   300.00   TRUE       300.00              0  2.979485
#> 26058 34.49836180   300.00   TRUE        76.02             39  3.272521
#> 38103 42.76155912   300.00   TRUE       300.00             31  2.907317
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 28260 31.83353510   300.00   TRUE       300.00            100  3.409673
#> 46627 53.68328061   300.00   TRUE       300.00            100  2.720428
#> 7771  16.57910676   300.00   TRUE        35.59            100  3.258586
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 42640 34.90974550   300.00   TRUE       300.00            100  3.627130
#> 13278 30.22524257    91.73   TRUE       120.78             34  3.560110
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 35458 37.47315867   300.00   TRUE       300.00            100  3.297274
#> 41310 17.29530400   300.00   TRUE       300.00             25  3.046207
#> 43536 29.09479684   300.00   TRUE       300.00              8  2.745482
#> 23832 25.48516273   300.00   TRUE       300.00              8  2.676567
#> 10785 33.24530310   300.00   TRUE       300.00            100  3.383302
#> 36765 36.18514955   300.00   TRUE       300.00            100  3.279776
#> 13072 34.94469593   300.00   TRUE       300.00            100  4.477051
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 46165 14.81840746   300.00  FALSE       300.00            100  2.545811
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 27965 32.29533908   300.00  FALSE       300.00            100  2.382556
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 39796 34.56768970   300.00  FALSE       300.00             71  2.333468
#> 36833 28.07550492   187.46  FALSE        18.49             45  3.942504
#> 18627 31.25713956   300.00  FALSE       300.00             20  3.545129
#> 4007  32.09251202   300.00  FALSE       300.00             68  2.847871
#> 31884 24.13756599   300.00  FALSE       300.00            100  3.117018
#> 29944 18.62800384   300.00  FALSE       300.00             48  2.452873
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 29527 36.72659467   300.00  FALSE       300.00            100  2.639632
#> 8895  38.07762915   300.00  FALSE       300.00             90  2.432641
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 43222 21.26532857   300.00  FALSE       300.00            100  2.404946
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 48845 31.55908831   169.72  FALSE         0.00            100  2.030594
#> 26768 32.28216105   300.00  FALSE       207.22            100  2.673964
#> 5861  27.61255502   300.00  FALSE       300.00            100  4.032264
#> 3724  25.72695092   300.00  FALSE       300.00            100  2.450415
#> 40612 21.92709482    55.73  FALSE       125.03            100  2.946551
#> 32430 22.86044307   300.00  FALSE       300.00            100  2.398827
#> 974   22.75845658    80.00  FALSE        90.05            100  3.099987
#> 44216 26.06499602   300.00  FALSE       300.00            100  3.167741
#> 2999  45.58394922   300.00  FALSE       300.00            100  2.883559
#> 42638 23.42136875    86.41  FALSE       172.46            100  6.734528
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 37783 24.25731417   300.00  FALSE       300.00             18  2.682893
#> 10981 29.16641656   300.00  FALSE       300.00            100  5.155983
#> 11394 40.60380007    46.04  FALSE        55.00            100  3.187005
#> 43972 22.85242166   300.00  FALSE       300.00            100  3.063725
#> 37254  4.19118627   131.59  FALSE        80.55            100  2.827529
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 15363 26.94219440   300.00  FALSE       300.00            100  2.858775
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983