Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)resamplingobjects. Each of these resampling objects is a list of length nfoldcorresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope    hcurv    vcurv        carea
#> 31358 712882.5 9560002 1911.52 25.564231 -0.00681 -0.00029    5577.3916
#> 37912 712802.5 9559952 1838.40 52.101344  0.00183 -0.09203     634.3320
#> 27864 712992.5 9560672 1926.38 27.224090 -0.00199  0.00659    3553.7166
#> 42632 714852.5 9557902 2674.80 30.652669  0.00221  0.00969     368.5606
#> 13748 715232.5 9559542 2212.12 26.123438  0.00023 -0.00043    1806.4552
#> 29391 712842.5 9560152 1930.16 28.095558  0.00753  0.01487    1077.7585
#> 42733 715112.5 9559062 2399.70 36.054515 -0.00041  0.01741     746.4155
#> 38411 713552.5 9560652 1851.63 46.005073 -0.00891 -0.00949    6217.8213
#> 32622 714972.5 9557762 2687.38 30.556412 -0.00730 -0.01491    1844.4353
#> 49372 714862.5 9560982 1863.12 20.969109 -0.00295  0.00035    2697.5974
#> 38827 713412.5 9560472 1869.10 16.871315 -0.00156  0.00406    2421.1145
#> 47255 713552.5 9559962 2191.77 33.578765 -0.00068  0.00588    1295.1610
#> 125   713992.5 9561062 1878.28 49.971087  0.00183  0.03387    1989.3012
#> 23817 714842.5 9557832 2677.58 26.657434  0.00922  0.01088     266.4355
#> 24018 714292.5 9559282 2334.12 32.865432  0.00102  0.00388    1051.4222
#> 23925 715202.5 9557652 2782.99 41.957126 -0.00542  0.01842     868.6137
#> 44328 713222.5 9558932 2227.94 29.155530 -0.00769 -0.00281    1325.6138
#> 3835  714142.5 9558552 2388.42 32.900383  0.00979 -0.00269     491.8017
#> 39061 712922.5 9559912 1931.50 36.315211  0.00414 -0.00104    1641.4587
#> 2990  713542.5 9560372 1970.07 31.887966  0.01903  0.01447     331.4833
#> 40307 713302.5 9560702 1845.07 39.004675  0.00129  0.02510    3342.5518
#> 15109 713872.5 9559602 2288.64 25.185506  0.00602  0.01518     854.4269
#> 30125 715162.5 9558992 2438.18 39.856663  0.00471 -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.173117 -0.01029 -0.02312    2657.6936
#> 25081 713392.5 9558452 2242.24 46.538497 -0.02716  0.03696    4143.4019
#> 47210 714952.5 9557612 2744.35 40.884549 -0.00493  0.00733     598.7983
#> 3453  713482.5 9559642 2145.11 35.125751  0.00108 -0.00678    1891.6584
#> 4477  715432.5 9559392 2325.25 37.029116 -0.00521 -0.00739     558.2286
#> 28230 715182.5 9558102 2791.83 39.071711 -0.00669 -0.00950     950.3138
#> 34565 712972.5 9560172 2007.72 27.947162  0.00598  0.01212     561.1948
#> 6763  714692.5 9561052 1830.60 48.394880 -0.01996  0.00626    1014.7159
#> 7879  714722.5 9559232 2356.19 36.803944 -0.01097  0.00156     816.6036
#> 19730 715332.5 9558452 2640.35 47.642587 -0.00211  0.00642     613.6390
#> 27013 713062.5 9559782 1976.81 41.632259  0.01414  0.03416    1204.8629
#> 19128 714852.5 9557872 2681.49 20.757688  0.00453  0.01077     300.7915
#> 45104 715062.5 9559242 2345.02 24.156474 -0.01743  0.01313     783.4548
#> 29667 713892.5 9559712 2239.25 41.592725  0.00804  0.00346     543.0749
#> 6557  713312.5 9559022 2181.69 27.634327 -0.00669 -0.00991    1403.0193
#> 20001 712702.5 9560132 1854.55 36.291147 -0.00204 -0.00376    2048.3198
#> 43092 715532.5 9558192 2770.31 36.833165 -0.00442 -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.063263  0.00940  0.00679     550.0344
#> 5578  714962.5 9557612 2751.95 39.191459  0.00288  0.01452     437.3965
#> 29926 714702.5 9557622 2637.00 44.642580  0.01843  0.04407     344.8395
#> 17036 712832.5 9560582 1951.38 27.645787  0.00060  0.00270    4057.5637
#> 3944  715262.5 9558352 2654.20 33.555273 -0.04492  0.00072   54831.6289
#> 44550 714842.5 9557642 2696.38 15.173068  0.00359 -0.00279     930.0871
#> 26567 714022.5 9558872 2327.42 46.665694  0.00290  0.00230    1056.5876
#> 16015 712582.5 9560312 1899.18 56.546924  0.03480  0.02530     937.1306
#> 9704  714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 20220 714052.5 9558522 2378.67 39.656128 -0.00855 -0.02175    1122.9120
#> 21952 715352.5 9560172 2013.44 40.488063  0.02409  0.02771     674.3859
#> 25244 713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 35194 713692.5 9559952 2185.68 28.796859 -0.00449 -0.00181    2406.1252
#> 15766 715562.5 9558592 2646.16 29.020885  0.00194  0.00516     576.0335
#> 462   713712.5 9561182 1785.08 47.717071 -0.03260 -0.03310    2833.9001
#> 375   715622.5 9558922 2510.87 37.366589 -0.01017  0.00197     711.1091
#> 30406 715312.5 9560152 2029.10 34.137398  0.00953 -0.02283     590.1536
#> 34005 714922.5 9558842 2559.23 52.831292  0.03658  0.04602     345.8933
#> 10507 714892.5 9559282 2374.45 42.842346 -0.01057  0.02427     380.9472
#> 39732 715362.5 9559582 2230.97 23.346884 -0.00161 -0.00409    2312.8413
#> 7668  713372.5 9559062 2164.82 37.934963 -0.01807 -0.01883     616.1993
#> 35083 714612.5 9558152 2596.12 60.598308  0.04548  0.05543     352.1558
#> 44825 714322.5 9557882 2534.80 40.688598  0.00114  0.03026     247.0894
#> 30886 714912.5 9558822 2563.86 49.753936 -0.00099  0.02169     495.8840
#> 664   714452.5 9559852 2174.83 44.860876  0.00282  0.00109     767.4097
#> 27695 714142.5 9558452 2449.28 20.490117 -0.00539  0.00809    1102.9956
#> 11618 713852.5 9558402 2432.14 54.059141 -0.00410  0.03620     379.0855
#> 17728 713092.5 9560652 1900.14 30.279101  0.00645 -0.00395    2011.3270
#> 45592 714962.5 9559862 2282.08 22.160862 -0.00028  0.01318     289.0063
#> 42765 713552.5 9559632 2186.95 44.011180  0.00394 -0.00224     942.0322
#> 26010 712852.5 9560542 1928.25 30.978109  0.00165  0.00555    4439.0967
#> 18958 714192.5 9558372 2495.72 34.677125 -0.00178 -0.00822     472.8748
#> 7041  713472.5 9559092 2137.03 18.348973 -0.10372 -0.00008  825875.6875
#> 43859 714892.5 9557602 2722.70 28.129363  0.00370 -0.00190     640.8842
#> 11612 715212.5 9557752 2820.51 39.162238  0.00017  0.01213     328.7922
#> 47320 713422.5 9559182 2143.03 56.325189  0.01246  0.00534     344.1447
#> 10467 714642.5 9559652 2249.12 37.143135 -0.00759 -0.00400     608.6523
#> 17345 713922.5 9559682 2227.44 34.490913 -0.06509 -0.01211   49347.5547
#> 16009 715122.5 9559402 2264.81 50.018643 -0.00589  0.00849    1120.5961
#> 30287 713992.5 9557822 2403.92 18.215474  0.03890 -0.00911     173.2806
#> 18348 714042.5 9558882 2341.41 44.063892 -0.00338  0.00728    1137.2450
#> 37720 715722.5 9557962 2820.42 45.076309 -0.01474  0.01334    4905.2002
#> 21683 714792.5 9561052 1807.84 55.732751 -0.02070 -0.02661    2508.5374
#> 27915 715232.5 9559552 2208.93 25.480579  0.00171 -0.00261    1627.8989
#> 33875 712832.5 9559852 1892.81 37.554519 -0.01175  0.04195    1901.6743
#> 34615 713242.5 9560732 1879.60 27.593647  0.00384  0.00175    1026.6818
#> 29228 712782.5 9560422 1870.56 32.906112 -0.00218 -0.00372    4154.5786
#> 5911  713952.5 9561282 1801.16 33.789613  0.00466 -0.01776    1021.8267
#> 43074 713172.5 9559622 2052.40 31.527576 -0.00193  0.02812    1423.7578
#> 47086 713692.5 9559972 2175.51 25.045704 -0.00789 -0.00261    3353.5422
#> 47244 714262.5 9559302 2350.47 17.059245  0.02213 -0.01363     565.5207
#> 28173 714562.5 9560382 2048.26 27.685894 -0.00124 -0.00456    1531.4274
#> 10692 712712.5 9561022 2005.85 41.433443  0.01410 -0.00380     601.0843
#> 49072 715472.5 9558902 2482.92 30.689338 -0.01102 -0.01778    4594.4351
#> 22297 714642.5 9558172 2606.18 59.469008 -0.00495 -0.03756     353.9086
#> 750   712812.5 9560032 1847.09 54.959258 -0.00767 -0.03713    6379.7974
#> 34984 715042.5 9559312 2320.49 42.857816 -0.01106 -0.04634     500.5027
#> 14494 715332.5 9558792 2557.22 20.720446  0.01267  0.00483     343.1384
#> 39405 713122.5 9559052 2169.30 37.646192 -0.02110 -0.02450    2589.9910
#> 24723 712942.5 9560202 1983.18 33.071697  0.00285 -0.01325     595.0426
#> 42617 714092.5 9558392 2469.21 42.799947 -0.01397  0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.285823  0.00708 -0.02538     807.8237
#> 30363 715192.5 9557542 2796.31 25.321870 -0.00157 -0.00152     850.4633
#> 47722 712732.5 9561022 1999.03 37.492639  0.00004 -0.02384     947.2075
#> 9864  714952.5 9557592 2752.40 31.989380 -0.00983  0.01253     528.1924
#> 41310 714452.5 9560402 2092.27 12.462978 -0.00195 -0.00205    1112.2626
#> 11332 712742.5 9560072 1850.34 22.700015  0.00683 -0.00603    1562.0503
#> 10785 714052.5 9557852 2377.76 20.984006 -0.00025 -0.05345    2417.1389
#> 17851 715362.5 9557482 2869.26 29.125737 -0.03251 -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.874802 -0.01519  0.00609    4275.3457
#> 4785  713792.5 9560472 2055.34 33.052789 -0.00586  0.00067    1123.7318
#> 32405 713022.5 9558982 2208.85 15.004619  0.00599  0.00301     338.9676
#> 34250 713822.5 9559142 2316.36 33.759819 -0.02284  0.01085    2229.5105
#> 2664  713342.5 9558572 2275.86 56.975496  0.00872  0.05419     761.5427
#> 15558 714732.5 9559502 2283.66 53.431752  0.00390 -0.02831     454.9113
#> 31071 713472.5 9558462 2298.33 29.153239  0.00156 -0.00116    1817.5222
#> 26963 715702.5 9557882 2880.32 43.733296 -0.01174 -0.01466    2633.9666
#> 30267 714442.5 9557512 2612.30 28.765919 -0.00458  0.00479     604.0975
#> 441   714312.5 9560972 1916.59 31.302976  0.00540  0.00650    1133.1501
#> 28162 714062.5 9561152 1785.47 48.444727  0.02309 -0.03409    1575.9425
#> 49835 713272.5 9560372 1898.68 32.825898  0.00846  0.00794    1363.5321
#> 29125 714992.5 9558792 2584.41 24.665833  0.01253 -0.00333     422.8840
#> 7538  714722.5 9558582 2660.25 15.994690  0.00544 -0.01014     271.4219
#> 16878 714232.5 9558172 2485.33 22.808304  0.00818  0.00502    1428.8546
#> 44396 715512.5 9558102 2845.31 26.598420  0.01263  0.03598     186.3527
#> 45071 712652.5 9560032 1900.80 14.473487  0.00987 -0.00507     504.3085
#> 44668 712612.5 9560472 1972.24 27.507131  0.00601 -0.00011     712.5228
#> 14800 712952.5 9558682 2120.89 38.273581 -0.00022  0.00662    1269.2650
#> 47059 713822.5 9558272 2413.20 38.175032 -0.00687 -0.00223    1056.3701
#> 36245 713142.5 9560572 1827.82 43.676000 -0.00410  0.00400    4126.7812
#> 22545 713402.5 9559562 2195.44 25.315567  0.04152  0.00098     164.0802
#> 34345 715852.5 9557902 2874.08 43.655946  0.00379 -0.01018     937.6911
#> 44267 712972.5 9559042 2191.01 22.269151  0.01462  0.00268     252.0072
#> 33666 713632.5 9561012 1775.88 21.178812 -0.00950 -0.05130     107.5831
#> 44715 714932.5 9560132 2195.78 30.393692  0.03479  0.02251     168.1733
#> 26724 714252.5 9560862 1975.16 28.286353  0.00165  0.00555    1882.0179
#> 14671 712672.5 9560372 1872.21 39.197189 -0.00873 -0.01417    1520.6630
#> 3724  714322.5 9557402 2603.46 34.597484  0.02523  0.02247     282.1076
#> 48184 714202.5 9558222 2489.17 36.113530  0.02236 -0.00066     306.0146
#> 42638 714322.5 9561182 1754.22  4.172279 -0.03757 -0.00003 5426606.0000
#> 18308 714842.5 9558742 2568.20 26.017441 -0.02552  0.00532   16771.3066
#> 280   713182.5 9560632 1862.83 39.004102  0.00745  0.00875    2156.9358
#> 18246 713902.5 9561342 1818.52 19.110434 -0.01502  0.00282    1659.6211
#> 10127 713722.5 9558712 2311.10 19.481138  0.01129  0.00491     261.7371
#> 37783 713832.5 9557892 2405.28 40.886841  0.01572  0.03148     481.8292
#> 18504 713372.5 9560672 1786.89  6.473850 -0.00278 -0.01172  111973.8906
#> 33881 712992.5 9558822 2173.26 33.197748 -0.00595  0.03064    1485.5201
#> 13548 715812.5 9558122 2821.65 15.508249  0.02751  0.00388     180.4414
#> 46567 714472.5 9560352 2080.13 21.381639  0.01336  0.01403     609.8665
#> 48157 713142.5 9560992 1978.48 44.621380  0.00208  0.00492     772.2691
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 31358 34.4278880   300.00   TRUE        15.00              9  3.746431
#> 37912 30.2945705   300.00   TRUE         9.15              2  2.802317
#> 27864 27.8153821    30.00   TRUE       183.39             20  3.550683
#> 42632 20.4780846   300.00   TRUE       300.00             10  2.566509
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 29391 28.2834886   300.00   TRUE         0.56            100  3.032521
#> 42733 31.2829227   300.00   TRUE       300.00            100  2.872981
#> 38411 32.7582890   279.54   TRUE         0.00            100  3.793638
#> 32622 30.1713845   300.00   TRUE       300.00            100  3.265863
#> 49372 23.5405440   213.54   TRUE        20.21             68  3.430977
#> 38827 31.0359778   300.00   TRUE        52.52             39  3.384015
#> 47255 19.6776625   300.00   TRUE       260.03            100  3.112324
#> 125   29.5967079   207.04   TRUE        21.03            100  3.298701
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 24018 29.3050723   300.00   TRUE       300.00              5  3.021777
#> 23925 37.5333192   300.00   TRUE       300.00            100  2.938827
#> 44328 25.5040703   300.00   TRUE       300.00            100  3.122417
#> 3835  35.1389286   300.00   TRUE       300.00             39  2.691790
#> 39061 35.2409151   300.00   TRUE        70.65             56  3.215230
#> 2990  25.3212968   300.00   TRUE       123.56              5  2.520462
#> 40307 29.5136290    77.82   TRUE        76.41             54  3.524078
#> 15109 22.6352707   300.00   TRUE       300.00             40  2.931675
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 25081 34.8753680   300.00   TRUE       300.00            100  3.617357
#> 47210 26.9152654   300.00   TRUE       300.00            100  2.777281
#> 3453  40.0847003   300.00   TRUE       300.00              2  3.276843
#> 4477  32.1698613   300.00   TRUE       300.00             11  2.746812
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 34565 20.3801088   300.00   TRUE        68.13            100  2.749114
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 7879  28.5808537   300.00   TRUE       300.00            100  2.912011
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 27013 30.9185215   300.00   TRUE       168.05             18  3.080938
#> 19128 14.6596981   300.00   TRUE       300.00             15  2.478266
#> 45104 17.9925936   300.00   TRUE       300.00            100  2.894014
#> 29667 34.2519900   300.00   TRUE       300.00             27  2.734860
#> 6557  27.4194682   300.00   TRUE       300.00              5  3.147064
#> 20001 34.6771246   264.46   TRUE        18.17             56  3.311398
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 5578  25.7418478   300.00   TRUE       300.00            100  2.640875
#> 29926 32.0948039   300.00   TRUE       300.00            100  2.537617
#> 17036 24.9282478    25.33   TRUE       208.61             15  3.608265
#> 3944  39.3604817   300.00   TRUE       300.00            100  4.739031
#> 44550 24.8262613   300.00   TRUE       300.00            100  2.968524
#> 26567 40.4370693   300.00   TRUE       300.00            100  3.023906
#> 16015 38.8327875   152.09   TRUE        90.23             45  2.971800
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 20220 32.1755909   300.00   TRUE       300.00              8  3.050346
#> 21952 40.0915758   300.00   TRUE       300.00             64  2.828908
#> 25244 33.2189470   300.00   TRUE       300.00              6  5.916915
#> 35194 23.4494437   300.00   TRUE       300.00            100  3.381318
#> 15766 23.3371440   300.00   TRUE       300.00            100  2.760448
#> 462    9.8124752    61.17   TRUE        57.16             74  3.452385
#> 375   27.9740914   300.00   TRUE       300.00            100  2.851936
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 34005 41.7336728   300.00   TRUE       300.00            100  2.538942
#> 10507 27.1622102   300.00   TRUE       300.00             46  2.580865
#> 39732 29.1681354   300.00   TRUE       300.00             25  3.364146
#> 7668  36.3203676   300.00   TRUE       300.00              5  2.789721
#> 35083 43.5407817   300.00   TRUE       300.00             23  2.546735
#> 44825 29.0214582   300.00   TRUE       300.00            100  2.392854
#> 30886 39.1851566   300.00   TRUE       300.00            100  2.695380
#> 664   30.8423181   300.00   TRUE       300.00             12  2.885027
#> 27695 27.3679020   300.00   TRUE       300.00             41  3.042574
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 17728 28.5522058    30.14   TRUE       140.15              2  3.303483
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 42765 42.3530402   300.00   TRUE       300.00             12  2.974066
#> 26010 25.2296236    69.52   TRUE       166.56              2  3.647295
#> 18958 29.9897569   300.00   TRUE       300.00              5  2.674746
#> 7041  33.2189470   300.00   TRUE       300.00              6  5.916915
#> 43859 29.5577467   300.00   TRUE       300.00            100  2.806780
#> 11612 35.1670035   300.00   TRUE       300.00            100  2.516921
#> 47320 45.6641633   300.00   TRUE       300.00             43  2.536741
#> 10467 38.9960805   300.00   TRUE       300.00             90  2.784369
#> 17345 29.8213073   300.00   TRUE       300.00              2  4.693266
#> 16009 33.4131797   300.00   TRUE       300.00            100  3.049449
#> 30287 19.8484039   300.00   TRUE       300.00             65  2.238750
#> 18348 41.0106001   300.00   TRUE       300.00            100  3.055854
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 21683 19.7452715   130.93   TRUE         0.00             21  3.399421
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 33875 27.5695832   300.00   TRUE         4.67              0  3.279136
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 29228 30.3994217   131.25   TRUE        80.55              2  3.618527
#> 5911   4.1384742    42.88   TRUE         0.00             13  3.009377
#> 43074 33.5948074   300.00   TRUE       205.02              1  3.153436
#> 47086 23.3869913   300.00   TRUE       300.00            100  3.525504
#> 47244 28.0388356   300.00   TRUE       300.00              2  2.752448
#> 28173 23.8837457   300.00   TRUE       300.00             59  3.185096
#> 10692 33.6463736   300.00   TRUE        60.19             11  2.778935
#> 49072 28.9876537   300.00   TRUE       300.00            100  3.662232
#> 22297 46.6496507   300.00   TRUE       300.00             11  2.548891
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 34984 33.9059234   300.00   TRUE       300.00            100  2.699406
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 39405 27.7202711   300.00   TRUE       300.00            100  3.413298
#> 24723 28.1734807   300.00   TRUE        36.63            100  2.774548
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 30363 29.8029727   300.00   TRUE       300.00            100  2.929656
#> 47722 33.1158146   300.00   TRUE        65.02              1  2.976445
#> 9864  20.8207133   300.00   TRUE       300.00            100  2.722792
#> 41310 17.2953040   300.00   TRUE       300.00             25  3.046207
#> 11332 25.6816873   300.00   TRUE        27.04             68  3.193695
#> 10785 33.2453031   300.00   TRUE       300.00            100  3.383302
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 4785  26.4861200   300.00  FALSE         0.00            100  3.050663
#> 32405 15.7620689   300.00  FALSE       300.00            100  2.530158
#> 34250 30.8927384   300.00  FALSE       300.00            100  3.348210
#> 2664  39.2911538   300.00  FALSE       300.00            100  2.881694
#> 15558 50.3440826   300.00  FALSE       300.00             45  2.657927
#> 31071 31.7378511   300.00  FALSE       300.00             76  3.259480
#> 26963 42.0631236   300.00  FALSE       300.00            100  3.420610
#> 30267 25.2703036   300.00  FALSE       300.00            100  2.781107
#> 441   25.1133131   294.31  FALSE       152.09            100  3.054287
#> 28162 18.6698298   130.00  FALSE        75.00            100  3.197540
#> 49835 32.3612292   300.00  FALSE         0.00            100  3.134665
#> 29125 23.9450522   300.00  FALSE       300.00            100  2.626221
#> 7538  17.2649372   300.00  FALSE       300.00             87  2.433645
#> 16878 25.3889058   300.00  FALSE       300.00             18  3.154988
#> 44396 21.4423725   300.00  FALSE       300.00            100  2.270336
#> 45071 27.8050688   264.04  FALSE         0.00            100  2.702696
#> 44668 30.1026296     9.15  FALSE       163.30              6  2.852799
#> 14800 31.5229920   300.00  FALSE       300.00            100  3.103552
#> 47059 25.7596095   300.00  FALSE       300.00            100  3.023816
#> 36245 25.8535746   110.19  FALSE        81.50             20  3.615611
#> 22545 23.6843564   300.00  FALSE       300.00             96  2.215056
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 44267 19.7378231   300.00  FALSE       226.30            100  2.401413
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 26724 21.7775528   300.00  FALSE       185.65            100  3.274624
#> 14671 37.8822505   123.04  FALSE        75.11              0  3.182033
#> 3724  25.7269509   300.00  FALSE       300.00            100  2.450415
#> 48184 30.9563367   300.00  FALSE       300.00             75  2.485742
#> 42638 23.4213687    86.41  FALSE       172.46            100  6.734528
#> 18308 30.0206966   300.00  FALSE       300.00             55  4.224567
#> 280   29.3234070    60.55  FALSE       127.54             29  3.333837
#> 18246  7.9876046    18.49  FALSE         0.00             29  3.220009
#> 10127 18.6194095   300.00  FALSE       300.00             49  2.417865
#> 37783 24.2573142   300.00  FALSE       300.00             18  2.682893
#> 18504 23.0959287   144.96  FALSE         4.48             89  5.049117
#> 33881 29.1858971   300.00  FALSE       300.00            100  3.171879
#> 13548 16.9784583   300.00  FALSE       300.00            100  2.256336
#> 46567 16.9480916   300.00  FALSE       300.00             78  2.785235
#> 48157 33.5266254    57.15  FALSE       157.06             45  2.887769