Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem      slope    hcurv      vcurv        carea
#> 36710 715232.5 9559582 2198.66 23.2471896 -0.00501   -0.00649    1399.2329
#> 21788 715392.5 9560172 1988.71 40.4766034 -0.01919   -0.04051  351155.1250
#> 4965  715042.5 9559312 2320.49 42.8578160 -0.01106   -0.04634     500.5027
#> 13748 715232.5 9559542 2212.12 26.1234377  0.00023   -0.00043    1806.4552
#> 46178 714892.5 9559312 2379.54 32.7668833 -0.00266    0.02896     276.0310
#> 1300  714042.5 9558482 2408.27 24.1484522  0.00659    0.01041     772.9746
#> 8598  714002.5 9558872 2310.02 39.5839352 -0.00498   -0.00752    1321.5997
#> 48187 712722.5 9560152 1835.29 29.3279907 -0.00333   -0.01767    3321.6113
#> 20799 714792.5 9561002 1857.29 23.1291603  0.00090    0.00890    2301.0454
#> 44952 715352.5 9557472 2873.35 38.9525357 -0.01073    0.00174    2310.4871
#> 39885 715062.5 9561022 1840.94 38.4357278  0.00250   -0.01340     604.1032
#> 8869  715062.5 9559392 2278.56 41.4019303  0.00310    0.01440    1088.8639
#> 43669 712392.5 9560162 2001.90 47.2839150 -0.00284    0.01804     628.5153
#> 23817 714842.5 9557832 2677.58 26.6574344  0.00922    0.01088     266.4355
#> 27375 714292.5 9558982 2458.96 43.6009423  0.00214    0.00466     724.0713
#> 30800 715182.5 9557582 2772.37 39.2040642 -0.02475    0.00665   25318.3555
#> 43007 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514     297.7842
#> 22283 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328     592.3326
#> 16843 714882.5 9558512 2701.78 26.6379538 -0.00270   -0.00769    1032.7177
#> 30125 715162.5 9558992 2438.18 39.8566631  0.00471   -0.01382     560.8309
#> 29472 714792.5 9561072 1786.08 36.1731174 -0.01029   -0.02312    2657.6936
#> 36030 712852.5 9559572 1938.79 28.6788295 -0.01440    0.00710     843.7961
#> 34275 715332.5 9558442 2650.77 45.5369667 -0.00043    0.00973     514.5488
#> 4827  714962.5 9559882 2274.42 29.2449118  0.00051    0.00889     362.8363
#> 6808  715192.5 9559492 2219.03 21.8176599  0.00713   -0.00173    8967.1924
#> 8288  714842.5 9557712 2679.04 28.9395253  0.00902    0.01358     802.7622
#> 40828 714842.5 9557632 2698.65 17.5365192  0.00490   -0.00450     873.0448
#> 32821 715022.5 9559952 2208.33 43.4164499 -0.00643    0.00314    2360.1501
#> 27708 715222.5 9557782 2844.65 29.5428498  0.03206    0.02773     179.3077
#> 10320 714912.5 9558482 2724.21 25.2829086  0.00356    0.00774     527.1068
#> 28230 715182.5 9558102 2791.83 39.0717109 -0.00669   -0.00950     950.3138
#> 8073  712572.5 9560302 1917.62 52.0090979  0.01937    0.00083     994.7876
#> 27957 715012.5 9559332 2306.35 20.4803764  0.00225    0.00325     616.8652
#> 3133  715042.5 9559252 2344.13 29.0208853  0.00422    0.02357     275.7474
#> 6763  714692.5 9561052 1830.60 48.3948802 -0.01996    0.00626    1014.7159
#> 46125 713732.5 9560062 2115.18 37.5934798 -0.04616    0.01276    9614.6924
#> 3466  713232.5 9560612 1812.72 51.5662016 -0.01804   -0.01596    5296.9014
#> 1635  713102.5 9559052 2172.87 31.4840308 -0.00064   -0.00916    2009.1686
#> 23141 715632.5 9558942 2495.99 34.3768948  0.01275    0.00605     798.6775
#> 19730 715332.5 9558452 2640.35 47.6425866 -0.00211    0.00642     613.6390
#> 12084 712832.5 9559992 1872.46 56.0547529 -0.03184    0.03924    8023.8657
#> 46756 713372.5 9559192 2120.47 46.0050732  0.03696    0.02514     297.7842
#> 19784 714982.5 9557722 2716.98 35.4362937  0.00888    0.00862     874.5408
#> 24100 715132.5 9558992 2420.18 29.2139721 -0.03229   -0.00511  361979.2812
#> 38686 714292.5 9557832 2496.22 38.4809914 -0.00509    0.00499     885.0064
#> 43092 715532.5 9558192 2770.31 36.8331648 -0.00442   -0.00178     800.1933
#> 31316 715022.5 9559332 2308.28 20.0632631  0.00940    0.00679     550.0344
#> 13699 712902.5 9559902 1917.21 30.4876572 -0.00093   -0.00907    1943.4092
#> 28743 715202.5 9558872 2463.87 34.6570711 -0.01267   -0.05063   18733.8848
#> 2976  715122.5 9559142 2390.05 33.1679538  0.01008    0.00102     483.4184
#> 34707 713962.5 9561092 1819.49 65.6861735 -0.04749    0.04779    2972.0193
#> 4031  715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 9704  714642.5 9559652 2249.12 37.1431350 -0.00759   -0.00400     608.6523
#> 36120 714622.5 9559942 2166.51 32.9863262 -0.02917   -0.01993   10308.6406
#> 30406 715312.5 9560152 2029.10 34.1373984  0.00953   -0.02283     590.1536
#> 23213 715592.5 9558562 2654.69 33.2195200 -0.00034    0.00294     739.5768
#> 19966 715362.5 9559572 2235.06 26.0495262  0.00126   -0.00485    2240.4185
#> 46783 713592.5 9560762 1832.60 36.0103974  0.00448   -0.00328     592.3326
#> 44650 715612.5 9558142 2764.88 39.3604817 -0.00550   -0.00941    1290.7709
#> 32280 712572.5 9560182 1956.86 50.6781170 -0.02554    0.03304    1569.7313
#> 10364 714762.5 9561002 1856.21 25.2748872 -0.00845    0.00454    3476.0818
#> 31561 713302.5 9559192 2092.11 36.5277783  0.00406   -0.00055     773.0359
#> 16355 712962.5 9560212 1992.88 39.5260665  0.01927   -0.00316     454.9674
#> 11618 713852.5 9558402 2432.14 54.0591409 -0.00410    0.03620     379.0855
#> 2552  714942.5 9557652 2719.32 34.8736492 -0.01238   -0.00851    1012.5963
#> 45592 714962.5 9559862 2282.08 22.1608616 -0.00028    0.01318     289.0063
#> 41783 715572.5 9558652 2609.51 32.4082118  0.00291   -0.00081     940.6002
#> 20601 715542.5 9558182 2773.29 32.7554242 -0.00604    0.00354    1147.7384
#> 31834 713492.5 9560682 1800.85 34.2370930 -0.00472   -0.00228    2262.3796
#> 39722 715022.5 9557712 2735.53 30.2607659  0.00133    0.00367     806.0966
#> 6845  713412.5 9559172 2123.32 47.1538536 -0.01556   -0.04394     587.8840
#> 46461 712682.5 9560032 1892.30 22.9778994  0.01453    0.01337     486.7551
#> 6217  714442.5 9559932 2151.30 36.6704448  0.01489    0.00241    1752.9242
#> 7     714952.5 9559862 2284.19 16.1064166  0.00516    0.01574     183.3220
#> 43201 715712.5 9557922 2852.82 31.4622584 -0.00949    0.00520    3390.5708
#> 34887 714662.5 9557392 2609.02 51.0207457 -0.01878   -0.04351    1442.9247
#> 20797 713012.5 9560392 1842.25 36.6498183  0.01299    0.00141    4611.0718
#> 37720 715722.5 9557962 2820.42 45.0763086 -0.01474    0.01334    4905.2002
#> 27915 715232.5 9559552 2208.93 25.4805791  0.00171   -0.00261    1627.8989
#> 34615 713242.5 9560732 1879.60 27.5936474  0.00384    0.00175    1026.6818
#> 19832 715182.5 9557662 2761.27 38.4328630 -0.00666   -0.03583    1898.0105
#> 5572  715062.5 9558262 2754.60 39.9116671  0.00707    0.01833     369.3771
#> 30933 714672.5 9561082 1798.37 54.8102886  0.00542   -0.00722    1373.4147
#> 34164 714212.5 9558492 2421.45 51.3295700  0.02247    0.01123     364.2903
#> 10617 713232.5 9560722 1879.95 26.4970062  0.00258    0.00422    1159.2100
#> 38668 715582.5 9558242 2735.55 35.4460340  0.01003    0.01196    1052.5341
#> 49133 714162.5 9558462 2451.13 21.8904255  0.00042    0.00957     495.4453
#> 22301 713722.5 9560712 1913.37 39.2711002  0.00705   -0.00915    2036.0378
#> 4667  712652.5 9560472 1957.89 36.3243783 -0.00296    0.00426    1315.5334
#> 39624 715222.5 9559552 2205.54 23.8974967  0.00401   -0.00391    1794.7964
#> 750   712812.5 9560032 1847.09 54.9592576 -0.00767   -0.03713    6379.7974
#> 14494 715332.5 9558792 2557.22 20.7204457  0.01267    0.00483     343.1384
#> 9606  714392.5 9559172 2353.05 30.7833035 -0.00185   -0.01185     953.8614
#> 42617 714092.5 9558392 2469.21 42.7999473 -0.01397    0.00526     706.1463
#> 38103 715392.5 9560162 1998.16 46.2858225  0.00708   -0.02538     807.8237
#> 16065 715242.5 9560622 1871.61 34.3064209 -0.00019    0.00119    6589.5796
#> 49459 714692.5 9557342 2639.74 42.0470808  0.00216   -0.02606     919.4241
#> 46038 713042.5 9559102 2146.38 29.4975862 -0.00622    0.00152    1530.5657
#> 42640 714332.5 9558522 2403.10 38.8711757 -0.00212    0.02732    4237.6934
#> 13278 712752.5 9560452 1901.25 42.1135439 -0.00118   -0.00252    3631.7017
#> 24532 715202.5 9560922 1799.62  4.4885514 -0.02312   -0.01848 1890804.2500
#> 14140 714672.5 9559262 2326.36 45.7925059 -0.01400   -0.01210     585.8992
#> 48461 714672.5 9559272 2316.84 40.3814924 -0.01966   -0.02234     849.0547
#> 49346 715322.5 9558782 2553.74 25.3952084  0.00319    0.00291     487.3438
#> 23832 714252.5 9560182 2188.12 30.6475125  0.00427    0.02403     474.8613
#> 3771  714842.5 9561152 1724.51 45.5398315  0.02405   -0.05655     130.9468
#> 17851 715362.5 9557482 2869.26 29.1257366 -0.03251   -0.00389   14516.2002
#> 17374 715712.5 9557952 2832.22 37.8748021 -0.01519    0.00609    4275.3457
#> 31983 714302.5 9559912 2204.77 23.4649135 -0.01332    0.01842     246.5161
#> 15211 715772.5 9558082 2801.07 36.2968763 -0.00079   -0.00331     699.3024
#> 16208 712882.5 9560332 1846.34 48.6252092  0.00159   -0.01559     866.5775
#> 46165 714392.5 9557912 2542.56 16.3298701  0.00325   -0.00505     351.4078
#> 27884 714072.5 9559202 2446.27 19.7567307 -0.00009    0.00109     605.1908
#> 48601 713692.5 9560152 2058.14 59.5572439 -0.01433   -0.02177     997.3089
#> 24836 714202.5 9558272 2507.87 38.4890128 -0.00226    0.00736     477.8591
#> 40459 712542.5 9559192 1992.60 39.8171290 -0.00162   -0.00517    1060.5494
#> 11111 715142.5 9558482 2650.65 42.3432999 -0.00533    0.01493    2212.0525
#> 7517  714002.5 9561142 1772.85 26.1893279  0.00815   -0.03345  665381.5625
#> 46930 713642.5 9560602 1945.03 34.5098210  0.00165    0.00875    1165.3953
#> 32869 715332.5 9558892 2533.17 20.6350750 -0.00242    0.01132     737.5982
#> 1488  714562.5 9560002 2150.93 11.2838308  0.00197    0.00463     851.6306
#> 25140 713332.5 9560692 1809.95 52.2571887  0.00075   -0.01945     259.7421
#> 43243 715682.5 9558462 2693.23 27.4062902 -0.00031   -0.00199     836.4196
#> 20547 713732.5 9560822 1851.67 36.9683829 -0.01405    0.00476   11992.5273
#> 31374 714652.5 9560592 1999.69 27.1066333 -0.00007   -0.00853    1354.6001
#> 32849 714302.5 9558222 2520.01 32.1572563 -0.00846    0.00216    1412.2295
#> 20319 715282.5 9560482 1901.16 27.6767263  0.00706   -0.03035    1471.9783
#> 1691  714222.5 9559022 2448.69 38.2168579 -0.01292    0.00462     716.5762
#> 3733  713762.5 9559632 2297.63 34.2147477 -0.01338    0.00518    1137.0179
#> 34345 715852.5 9557902 2874.08 43.6559462  0.00379   -0.01018     937.6911
#> 5656  713132.5 9560142 2042.24 33.5953803 -0.00035    0.00955     713.4463
#> 33666 713632.5 9561012 1775.88 21.1788119 -0.00950   -0.05130     107.5831
#> 20764 714542.5 9558152 2602.18 52.0789988  0.02118    0.00162     290.2031
#> 49831 715262.5 9557332 2884.30 22.2949337 -0.00097    0.01487    1769.3392
#> 330   714232.5 9560362 2165.00 18.0607756  0.01142    0.00868     235.8511
#> 46509 712512.5 9559532 1860.09 89.9887513 -0.00270 1018.51947     842.1514
#> 44715 714932.5 9560132 2195.78 30.3936922  0.03479    0.02251     168.1733
#> 43222 714472.5 9560712 2039.25 22.1906554  0.01858    0.00312     254.0656
#> 40909 712722.5 9560272 1827.15 29.2930402  0.00009   -0.00599   12873.8018
#> 43415 714722.5 9558642 2657.99  0.2314749  0.00799    0.00911     100.0000
#> 41612 713952.5 9560002 2198.00 28.7561788  0.00119    0.00361    1011.0248
#> 5861  713882.5 9559182 2350.43 27.3936851 -0.02608    0.00148   10771.1963
#> 28286 713702.5 9559842 2245.07 27.9689348  0.00511    0.01499     617.3855
#> 2999  714592.5 9558492 2515.83 47.0318136  0.01600   -0.01849     764.8193
#> 20950 714912.5 9557782 2667.71 30.8113784 -0.04770    0.00270   35950.5508
#> 40653 714262.5 9561262 1791.94 12.2223357  0.00161    0.00709    1181.2555
#> 31440 714842.5 9558392 2704.41 25.8014354  0.00992   -0.01052     451.7209
#> 39883 712732.5 9560842 2082.69 25.0909678  0.00732    0.01858     298.4490
#> 46148 715082.5 9558042 2725.39 42.8360436 -0.00181   -0.00799    1036.6606
#> 36141 714222.5 9559422 2316.64 16.2771580 -0.07627   -0.00124   25826.6914
#>           cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.7156944   300.00   TRUE       300.00             21  3.145890
#> 21788 32.8144388   300.00   TRUE       300.00             40  5.545499
#> 4965  33.9059234   300.00   TRUE       300.00            100  2.699406
#> 13748 34.1190637   300.00   TRUE       300.00              2  3.256827
#> 46178 20.9312942   300.00   TRUE       300.00             31  2.440958
#> 1300  27.5306221   300.00   TRUE       300.00              6  2.888165
#> 8598  39.5982591   300.00   TRUE       300.00            100  3.121100
#> 48187 30.9666500   279.54   TRUE        21.35             46  3.521349
#> 20799 24.8079266   180.67   TRUE         0.00             16  3.361925
#> 44952 34.0823944   300.00   TRUE       300.00            100  3.363704
#> 39885  6.0848118   210.57   TRUE         0.00            100  2.781111
#> 8869  29.1102667   300.00   TRUE       300.00            100  3.036974
#> 43669 28.4765754    24.22   TRUE         0.00             89  2.798316
#> 23817 21.6113951   300.00   TRUE       300.00              4  2.425592
#> 27375 38.7703351   300.00   TRUE       300.00            100  2.859781
#> 30800 33.2481679   300.00   TRUE       300.00            100  4.403435
#> 43007 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 22283 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 16843 21.3982548   300.00   TRUE       300.00             12  3.013982
#> 30125 40.7355804   300.00   TRUE       300.00            100  2.748832
#> 29472 18.6486303   111.09   TRUE         0.00             25  3.424505
#> 36030 29.1996481   300.00   TRUE         0.00            100  2.926238
#> 34275 36.8291541   300.00   TRUE       300.00            100  2.711427
#> 4827  21.0739607   300.00   TRUE       300.00              0  2.559711
#> 6808  28.5659568   300.00   TRUE       300.00             40  3.952656
#> 8288  22.7177766   300.00   TRUE       300.00             83  2.904587
#> 40828 25.5997543   300.00   TRUE       300.00            100  2.941037
#> 32821 29.4832622   300.00   TRUE       300.00             65  3.372940
#> 27708 25.7109081   300.00   TRUE       300.00            100  2.253599
#> 10320 16.4559208   300.00   TRUE       300.00             30  2.721899
#> 28230 37.9017311   300.00   TRUE       300.00            100  2.977867
#> 8073  36.0757146   138.71   TRUE        76.41             35  2.997730
#> 27957 30.0602307   300.00   TRUE       300.00             75  2.790190
#> 3133  18.8342686   300.00   TRUE       300.00            100  2.440511
#> 6763  16.2181433   125.47   TRUE         0.00              6  3.006344
#> 46125 25.6570501   300.00   TRUE       300.00            100  3.982935
#> 3466  26.7949442   101.32   TRUE        95.02             19  3.724022
#> 1635  27.8268412   300.00   TRUE       300.00            100  3.303016
#> 23141 30.2917057   300.00   TRUE       300.00            100  2.902371
#> 19730 38.4786996   300.00   TRUE       300.00            100  2.787913
#> 12084 34.5957646   300.00   TRUE         0.00              0  3.904384
#> 46756 38.9617030   300.00   TRUE       300.00             37  2.473902
#> 19784 27.9557567   300.00   TRUE       300.00            100  2.941780
#> 24100 34.2978266   300.00   TRUE       300.00            100  5.558684
#> 38686 31.1889576   300.00   TRUE       300.00             75  2.946946
#> 43092 39.4785110   300.00   TRUE       300.00            100  2.903195
#> 31316 30.0098104   300.00   TRUE       300.00             85  2.740390
#> 13699 34.7894244   300.00   TRUE        53.95             43  3.288564
#> 28743 34.6530604   300.00   TRUE       300.00             71  4.272628
#> 2976  26.8820338   300.00   TRUE       300.00            100  2.684323
#> 34707 27.7025094   165.13   TRUE        39.56            100  3.473052
#> 4031  42.7615591   300.00   TRUE       300.00             31  2.907317
#> 9704  38.9960805   300.00   TRUE       300.00             90  2.784369
#> 36120 31.5630990   300.00   TRUE       300.00             94  4.013201
#> 30406 43.1007501   300.00   TRUE       300.00             85  2.770965
#> 23213 25.4605255   300.00   TRUE       300.00            100  2.868983
#> 19966 29.2288690   300.00   TRUE       300.00             23  3.350329
#> 46783 19.9423690   256.35   TRUE         0.00            100  2.772566
#> 44650 37.3625778   300.00   TRUE       300.00            100  3.110849
#> 32280 27.8314249   126.66   TRUE         0.00              5  3.195825
#> 10364 26.0930709   179.84   TRUE         0.00              0  3.541090
#> 31561 31.4427779   300.00   TRUE       300.00              0  2.888200
#> 16355 28.2468193   300.00   TRUE        58.52            100  2.657980
#> 11618 33.9855646   300.00   TRUE       300.00              1  2.578737
#> 2552  32.2156979   300.00   TRUE       300.00            100  3.005436
#> 45592 15.6486233   300.00   TRUE       300.00              2  2.460907
#> 41783 28.1860858   300.00   TRUE       300.00            100  2.973405
#> 20601 37.9487136   300.00   TRUE       300.00            100  3.059843
#> 31834 24.6469255   215.68   TRUE         0.00            100  3.354565
#> 39722 26.5416969   300.00   TRUE       300.00            100  2.906387
#> 6845  44.5961700   300.00   TRUE       300.00             29  2.769292
#> 46461 24.9414258   288.47   TRUE         0.00            100  2.687311
#> 6217  30.9718066   300.00   TRUE       300.00              1  3.243763
#> 7     13.2204282   300.00   TRUE       300.00              0  2.263215
#> 43201 41.8282109   300.00   TRUE       300.00            100  3.530273
#> 34887 42.9357383   300.00   TRUE       300.00            100  3.159244
#> 20797 32.0569886   285.07   TRUE         0.00             96  3.663802
#> 37720 42.7735912   300.00   TRUE       300.00            100  3.690657
#> 27915 33.5747538   300.00   TRUE       300.00              6  3.211627
#> 34615 34.1895376    17.57   TRUE       142.95             16  3.011436
#> 19832 36.6337755   300.00   TRUE       300.00            100  3.278299
#> 5572  26.5978468   300.00   TRUE       300.00            100  2.567470
#> 30933 12.8434219    99.29   TRUE         0.00              6  3.137802
#> 34164 39.7134237   300.00   TRUE       300.00              1  2.561448
#> 10617 33.7982074    17.89   TRUE       145.83              4  3.064162
#> 38668 32.9227279   300.00   TRUE       300.00            100  3.022236
#> 49133 23.5193445   300.00   TRUE       300.00             38  2.694996
#> 22301 33.6154338   300.00   TRUE         0.00            100  3.308786
#> 4667  31.5172624    24.78   TRUE       172.24             38  3.119102
#> 39624 33.2945775   300.00   TRUE       300.00              2  3.254015
#> 750   34.3471009   300.00   TRUE         0.00              0  3.804807
#> 14494 21.1851145   300.00   TRUE       300.00             28  2.535469
#> 9606  33.8239905   300.00   TRUE       300.00              0  2.979485
#> 42617 31.2319294   300.00   TRUE       300.00            100  2.848895
#> 38103 42.7615591   300.00   TRUE       300.00             31  2.907317
#> 16065 30.3203536   300.00   TRUE       162.92            100  3.818858
#> 49459 41.8568588   300.00   TRUE       300.00            100  2.963516
#> 46038 29.7307800   300.00   TRUE       243.86            100  3.184852
#> 42640 34.9097455   300.00   TRUE       300.00            100  3.627130
#> 13278 30.2252426    91.73   TRUE       120.78             34  3.560110
#> 24532 32.5462946   300.00   TRUE         7.49            100  6.276647
#> 14140 35.4872870   300.00   TRUE       300.00             65  2.767823
#> 48461 35.7800684   300.00   TRUE       300.00             61  2.928936
#> 49346 23.3457383   300.00   TRUE       300.00             29  2.687835
#> 23832 25.4851627   300.00   TRUE       300.00              8  2.676567
#> 3771   0.3294507    44.23   TRUE         1.67             94  2.117095
#> 17851 32.2632534   300.00   TRUE       300.00            100  4.161853
#> 17374 43.1454408   300.00   TRUE       300.00            100  3.630971
#> 31983 19.7756383   300.00  FALSE       300.00             65  2.391845
#> 15211 30.8348697   300.00  FALSE       300.00            100  2.844665
#> 16208 38.3738483   256.60  FALSE         0.00             67  2.937807
#> 46165 14.8184075   300.00  FALSE       300.00            100  2.545811
#> 27884 15.4973624   300.00  FALSE       300.00            100  2.781892
#> 48601 35.8579907   300.00  FALSE       245.19            100  2.998830
#> 24836 25.5837115   300.00  FALSE       300.00             95  2.679300
#> 40459 36.1868684   300.00  FALSE       216.66             10  3.025531
#> 11111 33.3839589   300.00  FALSE       300.00            100  3.344795
#> 7517  22.9847749   144.37  FALSE        60.16            100  5.823071
#> 46930 31.9206247   300.00  FALSE         0.00            100  3.066473
#> 32869 14.6001105   300.00  FALSE       300.00             78  2.867820
#> 1488  29.9215749   300.00  FALSE       300.00            100  2.930251
#> 25140 14.4838001   104.12  FALSE        46.28             74  2.414542
#> 43243 27.9861235   300.00  FALSE       300.00            100  2.922424
#> 20547 34.2611573   300.00  FALSE        42.56            100  4.078911
#> 31374 26.6792068   300.00  FALSE       300.00            100  3.131811
#> 32849 29.5520172   300.00  FALSE       300.00             45  3.149905
#> 20319 32.0724585   300.00  FALSE       300.00            100  3.167901
#> 1691  27.0000631   300.00  FALSE       300.00            100  2.855262
#> 3733  22.1001281   300.00  FALSE       300.00            100  3.055767
#> 34345 44.3950618   300.00  FALSE       300.00            100  2.972060
#> 5656  31.2726094   300.00  FALSE        94.43            100  2.853361
#> 33666  0.0217724   212.17  FALSE       103.16            100  2.031744
#> 20764 39.8503606   300.00  FALSE       300.00             91  2.462702
#> 49831 29.8379231   300.00  FALSE       300.00            100  3.247811
#> 330   13.5722243   300.00  FALSE       275.17            100  2.372638
#> 46509 45.5787926   300.00  FALSE       162.53            100  2.925390
#> 44715 28.4502193   300.00  FALSE       300.00            100  2.225757
#> 43222 21.2653286   300.00  FALSE       300.00            100  2.404946
#> 40909 22.8782048   234.82  FALSE         0.00             43  4.109707
#> 43415  0.2314749   300.00  FALSE       300.00            100  2.000000
#> 41612 25.5871492   300.00  FALSE       300.00            100  3.004762
#> 5861  27.6125550   300.00  FALSE       300.00            100  4.032264
#> 28286 17.0483592   300.00  FALSE       300.00            100  2.790556
#> 2999  45.5839492   300.00  FALSE       300.00            100  2.883559
#> 20950 32.7118794   300.00  FALSE       300.00             81  4.555706
#> 40653  0.1420935    18.02  FALSE       140.15            100  3.072344
#> 31440 24.0350065   300.00  FALSE       300.00             96  2.654870
#> 39883 17.0443485   233.81  FALSE       219.23            100  2.474870
#> 46148 38.9107098   300.00  FALSE       300.00            100  3.015637
#> 36141 27.3060226   300.00  FALSE       300.00             85  4.412069