Partition the data for a (non-spatial) cross-validation

partition_cv creates a represampling object for length(repetition)-repeated nfold-fold cross-validation.

partition_cv(
  data,
  coords = c("x", "y"),
  nfold = 10,
  repetition = 1,
  seed1 = NULL,
  return_factor = FALSE
)

Arguments

data

data.frame containing at least the columns specified by coords

coords

(ignored by partition_cv)

nfold

number of partitions (folds) in nfold-fold cross-validation partitioning

repetition

numeric vector: cross-validation repetitions to be generated. Note that this is not the number of repetitions, but the indices of these repetitions. E.g., use repetition = c(1:100) to obtain (the 'first') 100 repetitions, and repetition = c(101:200) to obtain a different set of 100 repetitions.

seed1

seed1+i is the random seed that will be used by set.seed in repetition i (i in repetition) to initialize the random number generator before sampling from the data set.

return_factor

if FALSE (default), return a represampling object; if TRUE (used internally by other sperrorest functions), return a list containing factor vectors (see Value)

Value

If return_factor = FALSE (the default), a represampling object. Specifically, this is a (named) list of length(repetition)

resampling

objects. Each of these resampling objects is a list of length nfold

corresponding to the folds. Each fold is represented by a list of containing the components train and test, specifying the indices of training and test samples (row indices for data). If return_factor = TRUE (mainly used internally), a (named) list of length length(repetition). Each component of this list is a vector of length nrow(data) of type factor, specifying for each sample the fold to which it belongs. The factor levels are factor(1:nfold).

Details

This function does not actually perform a cross-validation or partition the data set itself; it simply creates a data structure containing the indices of training and test samples.

See also

Examples

data(ecuador)
## non-spatial cross-validation:
resamp <- partition_cv(ecuador, nfold = 5, repetition = 5)
# plot(resamp, ecuador)
# first repetition, second fold, test set indices:
idx <- resamp[["1"]][[2]]$test
# test sample used in this particular repetition and fold:
ecuador[idx, ]
#>              x       y     dem     slope     hcurv      vcurv       carea
#> 36710 715232.5 9559582 2198.66 23.247190  -0.00501   -0.00649   1399.2329
#> 23178 715382.5 9560142 2021.07 41.976607   0.00958    0.02642    671.1807
#> 15412 715272.5 9557702 2813.17 30.952326  -0.00123    0.00393   2081.0056
#> 31357 714752.5 9561022 1848.22 33.446411  -0.00347    0.02357   1752.0375
#> 47223 715302.5 9557472 2857.45 39.533515  -0.01021   -0.01579   1318.7560
#> 15762 714922.5 9558862 2535.97 54.910556   0.01305   -0.00555    430.4604
#> 14254 714852.5 9557882 2680.00 23.335425   0.00479    0.01261    323.8441
#> 46178 714892.5 9559312 2379.54 32.766883  -0.00266    0.02896    276.0310
#> 34512 713162.5 9559632 2041.52 46.815236  -0.00857    0.03677   1675.4679
#> 26531 714462.5 9560452 2104.74 19.524110   0.00453    0.00257    357.1436
#> 8869  715062.5 9559392 2278.56 41.401930   0.00310    0.01440   1088.8639
#> 37270 714852.5 9561032 1838.96 35.158982  -0.00854    0.01264   5057.1431
#> 26447 715632.5 9559102 2410.77 39.656701   0.00730    0.01770    474.5059
#> 38993 714202.5 9558402 2475.75 33.159359  -0.01607    0.00677   1343.5006
#> 24018 714292.5 9559282 2334.12 32.865432   0.00102    0.00388   1051.4222
#> 38287 714872.5 9561162 1719.41 48.235598  -0.03021   -0.09328   2235.5259
#> 30800 715182.5 9557582 2772.37 39.204064  -0.02475    0.00665  25318.3555
#> 20290 714812.5 9561092 1775.73 34.626704  -0.01537   -0.00203  10263.2734
#> 35942 712862.5 9558902 2109.97 31.400379   0.00151    0.00189   2317.1819
#> 34275 715332.5 9558442 2650.77 45.536967  -0.00043    0.00973    514.5488
#> 4477  715432.5 9559392 2325.25 37.029116  -0.00521   -0.00739    558.2286
#> 27957 715012.5 9559332 2306.35 20.480376   0.00225    0.00325    616.8652
#> 3133  715042.5 9559252 2344.13 29.020885   0.00422    0.02357    275.7474
#> 15390 713132.5 9560622 1873.26 38.480418   0.00252    0.01118   2359.5059
#> 13847 715052.5 9559232 2337.87 26.114843  -0.01452   -0.01437   1155.1469
#> 23141 715632.5 9558942 2495.99 34.376895   0.01275    0.00605    798.6775
#> 19730 715332.5 9558452 2640.35 47.642587  -0.00211    0.00642    613.6390
#> 4329  712882.5 9559982 1911.46 26.231727  -0.00193   -0.00077   3559.4204
#> 28577 712812.5 9559962 1849.82 59.780125   0.00646   -0.04086   6042.5757
#> 45104 715062.5 9559242 2345.02 24.156474  -0.01743    0.01313    783.4548
#> 29667 713892.5 9559712 2239.25 41.592725   0.00804    0.00346    543.0749
#> 19784 714982.5 9557722 2716.98 35.436294   0.00888    0.00862    874.5408
#> 36931 712852.5 9558892 2102.80 30.535786   0.00140   -0.00070   2562.7559
#> 34563 715022.5 9558162 2675.70 18.889846  -0.01993    0.01373  16031.7500
#> 43092 715532.5 9558192 2770.31 36.833165  -0.00442   -0.00178    800.1933
#> 29926 714702.5 9557622 2637.00 44.642580   0.01843    0.04407    344.8395
#> 17036 712832.5 9560582 1951.38 27.645787   0.00060    0.00270   4057.5637
#> 37341 714412.5 9559142 2377.24 40.599789  -0.00735   -0.00925    665.6116
#> 44550 714842.5 9557642 2696.38 15.173068   0.00359   -0.00279    930.0871
#> 16015 712582.5 9560312 1899.18 56.546924   0.03480    0.02530    937.1306
#> 25244 713472.5 9559092 2137.03 18.348973  -0.10372   -0.00008 825875.6875
#> 421   714892.5 9561022 1849.01 26.132032   0.00002    0.00588    749.2502
#> 462   713712.5 9561182 1785.08 47.717071  -0.03260   -0.03310   2833.9001
#> 34005 714922.5 9558842 2559.23 52.831292   0.03658    0.04602    345.8933
#> 27875 714832.5 9561032 1838.22 41.212281  -0.01160    0.01690   5121.6924
#> 47100 712752.5 9560462 1910.00 41.911863  -0.00032    0.00542   3614.7334
#> 32735 714922.5 9559002 2414.49 40.064647   0.00062   -0.00262   1760.2162
#> 20065 713772.5 9558852 2221.21 28.019928  -0.00568   -0.02553   1750.4175
#> 664   714452.5 9559852 2174.83 44.860876   0.00282    0.00109    767.4097
#> 31121 713462.5 9559132 2135.56 61.596974  -0.03800   -0.06020   1563.5549
#> 16355 712962.5 9560212 1992.88 39.526066   0.01927   -0.00316    454.9674
#> 11618 713852.5 9558402 2432.14 54.059141  -0.00410    0.03620    379.0855
#> 5103  714612.5 9559892 2220.36 49.007945  -0.00078    0.00219    420.7032
#> 2552  714942.5 9557652 2719.32 34.873649  -0.01238   -0.00851   1012.5963
#> 41783 715572.5 9558652 2609.51 32.408212   0.00291   -0.00081    940.6002
#> 40270 713312.5 9559012 2186.56 30.935710  -0.00771   -0.00599   1378.5760
#> 47366 714892.5 9557612 2718.78 28.023939   0.00538    0.00112    745.3315
#> 32879 713912.5 9559672 2238.00 31.976775  -0.00278   -0.00412   1350.3611
#> 42201 714642.5 9557452 2592.29 13.225585  -0.02084   -0.02236   9755.3027
#> 4588  713852.5 9558612 2308.97 52.180412  -0.01059   -0.07431   4079.0366
#> 46138 715202.5 9559582 2194.40 18.860052  -0.02143   -0.05078 622888.2500
#> 33875 712832.5 9559852 1892.81 37.554519  -0.01175    0.04195   1901.6743
#> 15432 715632.5 9558922 2509.67 35.827624  -0.00444   -0.00017    759.9937
#> 29228 712782.5 9560422 1870.56 32.906112  -0.00218   -0.00372   4154.5786
#> 10689 713542.5 9559132 2211.95 31.981358   0.05379    0.00801    172.7491
#> 1036  713042.5 9558892 2205.38 31.509814  -0.00428   -0.00522   1355.7436
#> 43074 713172.5 9559622 2052.40 31.527576  -0.00193    0.02812   1423.7578
#> 7485  714362.5 9558042 2484.02 47.958859   0.03169    0.01441    424.2680
#> 47244 714262.5 9559302 2350.47 17.059245   0.02213   -0.01363    565.5207
#> 12629 714712.5 9558872 2533.15 37.025679   0.04250   -0.00970   1443.9200
#> 10617 713232.5 9560722 1879.95 26.497006   0.00258    0.00422   1159.2100
#> 49072 715472.5 9558902 2482.92 30.689338  -0.01102   -0.01778   4594.4351
#> 23783 713062.5 9560252 1983.09 36.429803   0.00774   -0.00224   1137.3363
#> 23626 713742.5 9560032 2137.92 39.304905  -0.00605    0.00075   2170.2249
#> 20171 715732.5 9557962 2815.53 48.657868  -0.01341    0.00171   4425.5410
#> 29024 713492.5 9559112 2173.47 53.432325   0.04864    0.04226    338.3474
#> 16065 715242.5 9560622 1871.61 34.306421  -0.00019    0.00119   6589.5796
#> 17825 712732.5 9560462 1915.51 40.397535   0.00143    0.00477   2945.2100
#> 30363 715192.5 9557542 2796.31 25.321870  -0.00157   -0.00152    850.4633
#> 22847 714582.5 9560382 2040.49 25.336194   0.00181   -0.00341   1363.4000
#> 28588 715532.5 9559362 2299.84 11.144602  -0.04360   -0.03000 162178.1562
#> 18498 714842.5 9557712 2679.04 28.939525   0.00902    0.01358    802.7622
#> 11057 714962.5 9557642 2737.48 42.685929  -0.00488   -0.00632    620.0229
#> 47722 712732.5 9561022 1999.03 37.492639   0.00004   -0.02384    947.2075
#> 10213 715542.5 9558782 2532.10 35.413375  -0.00893   -0.00867   2247.8748
#> 17851 715362.5 9557482 2869.26 29.125737  -0.03251   -0.00389  14516.2002
#> 6470  713742.5 9558802 2260.79 30.680171   0.00000    0.00320    911.5911
#> 31253 712612.5 9559492 1903.67 46.756794  -0.06953   -0.01797   7557.7866
#> 31983 714302.5 9559912 2204.77 23.464914  -0.01332    0.01842    246.5161
#> 36561 715292.5 9560802 1844.64 31.937240   0.00002   -0.01232    977.7320
#> 40459 712542.5 9559192 1992.60 39.817129  -0.00162   -0.00517   1060.5494
#> 15558 714732.5 9559502 2283.66 53.431752   0.00390   -0.02831    454.9113
#> 14694 715032.5 9559482 2307.44 44.173327  -0.00007   -0.00463   1050.7473
#> 8635  713832.5 9557902 2397.71 47.119476   0.02295    0.02715    520.7355
#> 26023 715832.5 9557632 3097.56 41.710182   0.00133   -0.00452    578.5847
#> 7517  714002.5 9561142 1772.85 26.189328   0.00815   -0.03345 665381.5625
#> 46930 713642.5 9560602 1945.03 34.509821   0.00165    0.00875   1165.3953
#> 43426 713782.5 9560822 1876.35 23.823012  -0.00345   -0.00075   8141.7549
#> 441   714312.5 9560972 1916.59 31.302976   0.00540    0.00650   1133.1501
#> 9665  713362.5 9558812 2275.28 31.494917   0.01306   -0.00626    450.7346
#> 13512 715152.5 9557992 2778.81 43.389521  -0.00724    0.01464   4029.5947
#> 45166 715262.5 9558722 2531.59 13.553317  -0.00502   -0.00028   2394.7261
#> 29125 714992.5 9558792 2584.41 24.665833   0.01253   -0.00333    422.8840
#> 44396 715512.5 9558102 2845.31 26.598420   0.01263    0.03598    186.3527
#> 31374 714652.5 9560592 1999.69 27.106633  -0.00007   -0.00853   1354.6001
#> 8831  714372.5 9559852 2200.77 48.760427  -0.00457    0.01697    412.1185
#> 37029 713022.5 9559082 2163.08 36.032170   0.00490   -0.00780    825.4451
#> 31201 714672.5 9560022 2224.47 38.863727   0.01180    0.01170    253.5129
#> 42934 713832.5 9558672 2270.74 27.611982  -0.03193   -0.00267  31588.1953
#> 22545 713402.5 9559562 2195.44 25.315567   0.04152    0.00098    164.0802
#> 34105 715432.5 9558592 2621.73 18.956309  -0.00301   -0.01099   1146.2140
#> 24009 714042.5 9558962 2357.58 34.361998   0.05041   -0.00011    201.1272
#> 34228 715932.5 9558592 2690.97  9.535164 197.75496 -197.75366    267.5057
#> 34810 714112.5 9559572 2337.51 10.288030  -0.00393    0.00032    994.9224
#> 39440 714222.5 9557872 2508.52 39.047647   0.02785    0.00645    234.0114
#> 14148 714162.5 9560352 2149.57 37.133395   0.00424    0.00196    576.1450
#> 13908 715022.5 9559372 2290.81 27.242997  -0.03004    0.00714 480535.5000
#> 12519 713982.5 9560912 1956.98 28.824934  -0.00452   -0.00948   1222.5433
#> 33410 715342.5 9560452 1942.88 29.093078  -0.00305   -0.00705    968.9738
#> 33666 713632.5 9561012 1775.88 21.178812  -0.00950   -0.05130    107.5831
#> 36721 714092.5 9559752 2301.81 28.933223   0.03418    0.00513    161.9878
#> 36889 714302.5 9559802 2243.60 21.166207   0.01923    0.05547    128.2613
#> 35826 712912.5 9560302 1891.22 50.869485  -0.00229    0.00099    909.0560
#> 26724 714252.5 9560862 1975.16 28.286353   0.00165    0.00555   1882.0179
#> 484   713812.5 9560592 2016.41 44.301097  -0.00719    0.00079   1647.6451
#> 44434 713022.5 9558762 2161.07 31.454237  -0.00273   -0.01047   2826.7700
#> 280   713182.5 9560632 1862.83 39.004102   0.00745    0.00875   2156.9358
#> 4943  712452.5 9559172 1927.46 34.465130   0.00105   -0.01576   1045.1074
#> 18246 713902.5 9561342 1818.52 19.110434  -0.01502    0.00282   1659.6211
#> 15138 713702.5 9561122 1780.68 49.221658   0.00607   -0.04566  12862.4922
#> 48050 714022.5 9560572 2099.82 18.713948   0.01182   -0.00972    331.7462
#> 7132  713192.5 9559612 2060.27 20.773731   0.00452    0.00368   1071.8734
#> 34334 714642.5 9557912 2582.25 30.991287   0.00253   -0.01233    968.7780
#> 28526 715152.5 9560732 1871.53 29.534255  -0.00163   -0.00377   1727.6055
#> 40653 714262.5 9561262 1791.94 12.222336   0.00161    0.00709   1181.2555
#> 46639 715232.5 9560042 2107.15 39.657274  -0.01305    0.00585   5714.3267
#> 49812 713392.5 9559222 2133.99 35.662612  -0.00983    0.00803   1058.5630
#> 6577  714142.5 9559902 2244.30 57.890510  -0.06986   -0.05013    616.4128
#> 23808 715172.5 9558702 2533.50 22.345927   0.00811   -0.02281    492.4785
#> 14645 715072.5 9560262 2034.04 38.169302  -0.00716   -0.03234   1707.3130
#> 47245 715372.5 9560792 1900.65 34.470860  -0.00011   -0.00419    323.0790
#> 15880 715892.5 9557572 3113.36 53.043287   0.00962   -0.00172    351.5812
#> 26635 715092.5 9560412 2010.59 30.571882  -0.00463    0.00253   1237.0536
#> 126   715042.5 9561262 1733.49 30.110651  -0.00288    0.00448    941.8503
#> 11025 714402.5 9559252 2325.64 49.620437  -0.00241    0.00020    673.2861
#> 13548 715812.5 9558122 2821.65 15.508249   0.02751    0.00388    180.4414
#> 28978 713982.5 9560462 2133.10  5.229386   0.00366    0.00525    162.0592
#> 46567 714472.5 9560352 2080.13 21.381639   0.01336    0.01403    609.8665
#> 47629 714472.5 9558482 2437.33 37.272623  -0.01181   -0.01149   4242.3579
#> 40852 715392.5 9557932 2939.71 18.297407   0.00505    0.00855    229.6061
#>            cslope distroad slides distdeforest distslidespast log.carea
#> 36710 30.71569444   300.00   TRUE       300.00             21  3.145890
#> 23178 41.60017367   300.00   TRUE       300.00             21  2.826839
#> 15412 37.59290685   300.00   TRUE       300.00            100  3.318273
#> 31357 23.81728259   158.92   TRUE         0.00              5  3.243543
#> 47223 36.25906111   300.00   TRUE       300.00            100  3.120164
#> 15762 46.30186534   300.00   TRUE       300.00            100  2.633933
#> 14254 15.66523908   300.00   TRUE       300.00             10  2.510336
#> 46178 20.93129417   300.00   TRUE       300.00             31  2.440958
#> 34512 34.63988238   300.00   TRUE       195.00              2  3.224136
#> 26531 15.98781432   300.00   TRUE       300.00              2  2.552843
#> 8869  29.11026670   300.00   TRUE       300.00            100  3.036974
#> 37270 18.94656837   162.65   TRUE         0.00             77  3.703905
#> 26447 33.03674647   300.00   TRUE       300.00            100  2.676242
#> 38993 29.54857941   300.00   TRUE       300.00              2  3.128238
#> 24018 29.30507235   300.00   TRUE       300.00              5  3.021777
#> 38287 14.89690267    41.43   TRUE         1.90            100  3.349380
#> 30800 33.24816789   300.00   TRUE       300.00            100  4.403435
#> 20290 15.45668244    95.53   TRUE         0.00             49  4.011286
#> 35942 27.16450202   300.00   TRUE       300.00            100  3.364960
#> 34275 36.82915411   300.00   TRUE       300.00            100  2.711427
#> 4477  32.16986132   300.00   TRUE       300.00             11  2.746812
#> 27957 30.06023072   300.00   TRUE       300.00             75  2.790190
#> 3133  18.83426864   300.00   TRUE       300.00            100  2.440511
#> 15390 29.32226108    60.00   TRUE       118.92              2  3.372821
#> 13847 20.57090372   300.00   TRUE       300.00            100  3.062637
#> 23141 30.29170567   300.00   TRUE       300.00            100  2.902371
#> 19730 38.47869961   300.00   TRUE       300.00            100  2.787913
#> 4329  34.56196012   300.00   TRUE        16.35              2  3.551379
#> 28577 34.39121869   300.00   TRUE         0.56              0  3.781222
#> 45104 17.99259364   300.00   TRUE       300.00            100  2.894014
#> 29667 34.25198995   300.00   TRUE       300.00             27  2.734860
#> 19784 27.95575674   300.00   TRUE       300.00            100  2.941780
#> 36931 27.20919273   300.00   TRUE       300.00            100  3.408707
#> 34563 38.15669732   300.00   TRUE       300.00            100  4.204981
#> 43092 39.47851096   300.00   TRUE       300.00            100  2.903195
#> 29926 32.09480385   300.00   TRUE       300.00            100  2.537617
#> 17036 24.92824775    25.33   TRUE       208.61             15  3.608265
#> 37341 31.45194521   300.00   TRUE       300.00              1  2.823221
#> 44550 24.82626126   300.00   TRUE       300.00            100  2.968524
#> 16015 38.83278752   152.09   TRUE        90.23             45  2.971800
#> 25244 33.21894705   300.00   TRUE       300.00              6  5.916915
#> 421   10.71373781   182.02   TRUE         0.00             83  2.874627
#> 462    9.81247520    61.17   TRUE        57.16             74  3.452385
#> 34005 41.73367284   300.00   TRUE       300.00            100  2.538942
#> 27875 21.26704744   157.93   TRUE         0.00             57  3.709413
#> 47100 29.64598224    82.80   TRUE       129.73             36  3.558076
#> 32735 39.35818982   300.00   TRUE       300.00            100  3.245566
#> 20065 30.38968145   300.00   TRUE       300.00              5  3.243142
#> 664   30.84231811   300.00   TRUE       300.00             12  2.885027
#> 31121 38.16013507   300.00   TRUE       300.00             45  3.194113
#> 16355 28.24681930   300.00   TRUE        58.52            100  2.657980
#> 11618 33.98556458   300.00   TRUE       300.00              1  2.578737
#> 5103  34.91318325   300.00   TRUE       300.00             91  2.623976
#> 2552  32.21569795   300.00   TRUE       300.00            100  3.005436
#> 41783 28.18608577   300.00   TRUE       300.00            100  2.973405
#> 40270 27.59765812   300.00   TRUE       300.00              5  3.139431
#> 47366 28.72753089   300.00   TRUE       300.00            100  2.872349
#> 32879 29.73822844   300.00   TRUE       300.00              6  3.130450
#> 42201 39.55872505   300.00   TRUE       300.00            100  3.989241
#> 4588  31.66451255   300.00   TRUE       300.00              1  3.610558
#> 46138 32.69469066   300.00   TRUE       300.00              8  5.794410
#> 33875 27.56958319   300.00   TRUE         4.67              0  3.279136
#> 15432 28.76763794   300.00   TRUE       300.00            100  2.880810
#> 29228 30.39942174   131.25   TRUE        80.55              2  3.618527
#> 10689 30.64808542   300.00   TRUE       300.00             61  2.237416
#> 1036  25.72695092   300.00   TRUE       300.00            100  3.132178
#> 43074 33.59480736   300.00   TRUE       205.02              1  3.153436
#> 7485  39.49054307   300.00   TRUE       300.00            100  2.627640
#> 47244 28.03883562   300.00   TRUE       300.00              2  2.752448
#> 12629 32.18074752   300.00   TRUE       300.00            100  3.159543
#> 10617 33.79820738    17.89   TRUE       145.83              4  3.064162
#> 49072 28.98765373   300.00   TRUE       300.00            100  3.662232
#> 23783 27.60052291   300.00   TRUE         0.00            100  3.055889
#> 23626 26.77832847   300.00   TRUE       300.00            100  3.336505
#> 20171 42.13015963   300.00   TRUE       300.00            100  3.645966
#> 29024 42.57305601   300.00   TRUE       300.00             32  2.529363
#> 16065 30.32035356   300.00   TRUE       162.92            100  3.818858
#> 17825 30.73345613    73.66   TRUE       137.63             18  3.469116
#> 30363 29.80297267   300.00   TRUE       300.00            100  2.929656
#> 22847 24.74776605   300.00   TRUE       300.00             41  3.134623
#> 28588 31.62555142   300.00   TRUE       300.00             78  5.209992
#> 18498 22.71777658   300.00   TRUE       300.00             83  2.904587
#> 11057 31.22505392   300.00   TRUE       300.00            100  2.792408
#> 47722 33.11581464   300.00   TRUE        65.02              1  2.976445
#> 10213 31.59862240   300.00   TRUE       300.00            100  3.351772
#> 17851 32.26325344   300.00   TRUE       300.00            100  4.161853
#> 6470  28.09326661   300.00   TRUE       300.00              2  2.959800
#> 31253 38.66834864   300.00   TRUE       122.87            100  3.878395
#> 31983 19.77563830   300.00  FALSE       300.00             65  2.391845
#> 36561 25.80487318   300.00  FALSE        70.00            100  2.990220
#> 40459 36.18686842   300.00  FALSE       216.66             10  3.025531
#> 15558 50.34408258   300.00  FALSE       300.00             45  2.657927
#> 14694 43.51156088   300.00  FALSE       300.00            100  3.021498
#> 8635  27.68188291   300.00  FALSE       300.00             23  2.716617
#> 26023 45.07287087   300.00  FALSE       300.00            100  2.762367
#> 7517  22.98477491   144.37  FALSE        60.16            100  5.823071
#> 46930 31.92062468   300.00  FALSE         0.00            100  3.066473
#> 43426 35.46264977   300.00  FALSE        82.11            100  3.910718
#> 441   25.11331312   294.31  FALSE       152.09            100  3.054287
#> 9665  34.08124853   300.00  FALSE       300.00             64  2.653921
#> 13512 35.93476700   300.00  FALSE       300.00            100  3.605261
#> 45166 23.44772481   300.00  FALSE       300.00              1  3.379256
#> 29125 23.94505217   300.00  FALSE       300.00            100  2.626221
#> 44396 21.44237252   300.00  FALSE       300.00            100  2.270336
#> 31374 26.67920677   300.00  FALSE       300.00            100  3.131811
#> 8831  32.55546192   300.00  FALSE       300.00             24  2.615022
#> 37029 28.89311569   300.00  FALSE       237.21            100  2.916688
#> 31201 33.72257695   300.00  FALSE       300.00            100  2.404000
#> 42934 32.22715710   300.00  FALSE       300.00             22  4.499525
#> 22545 23.68435638   300.00  FALSE       300.00             96  2.215056
#> 34105 20.07701410   300.00  FALSE       300.00             26  3.059266
#> 24009 38.49703425   300.00  FALSE       300.00            100  2.303471
#> 34228 10.74868824   300.00  FALSE       300.00            100  2.427333
#> 34810 15.55465822   300.00  FALSE       300.00            100  2.997789
#> 39440 32.70614982   300.00  FALSE       300.00             92  2.369237
#> 14148 21.87782045   300.00  FALSE       214.75            100  2.760532
#> 13908 34.37746771   300.00  FALSE       300.00             96  5.681725
#> 12519 30.48593836   300.00  FALSE       120.00            100  3.087264
#> 33410 32.87975603   300.00  FALSE       300.00            100  2.986312
#> 33666  0.02177240   212.17  FALSE       103.16            100  2.031744
#> 36721 26.66832057   300.00  FALSE       300.00            100  2.209482
#> 36889 19.64500392   300.00  FALSE       300.00             48  2.108096
#> 35826 44.22775812   298.23  FALSE        16.35             96  2.958591
#> 26724 21.77755284   300.00  FALSE       185.65            100  3.274624
#> 484   32.14579710   300.00  FALSE        55.63            100  3.216864
#> 44434 36.31979463   300.00  FALSE       300.00            100  3.451290
#> 280   29.32340700    60.55  FALSE       127.54             29  3.333837
#> 4943  37.77166969   300.00  FALSE       300.00            100  3.019161
#> 18246  7.98760462    18.49  FALSE         0.00             29  3.220009
#> 15138 20.64596119   116.69  FALSE        19.61             40  4.109325
#> 48050 23.36120818   300.00  FALSE        72.23            100  2.520806
#> 7132  34.04916289   300.00  FALSE       225.26             12  3.030143
#> 34334 25.08982185   300.00  FALSE       300.00            100  2.986224
#> 28526 31.81462749   300.00  FALSE        46.28            100  3.237445
#> 40653  0.14209353    18.02  FALSE       140.15            100  3.072344
#> 46639 33.56329468   300.00  FALSE       300.00            100  3.756965
#> 49812 33.35817579   300.00  FALSE       300.00             65  3.024717
#> 6577  42.40747121   300.00  FALSE       300.00             70  2.789872
#> 23808 37.53389220   300.00  FALSE       300.00             72  2.692387
#> 14645 45.21324553   300.00  FALSE       300.00            100  3.232313
#> 47245 34.97620860   300.00  FALSE       150.03            100  2.509309
#> 15880 48.89679119   300.00  FALSE       300.00            100  2.546026
#> 26635 31.39464943   300.00  FALSE       300.00            100  3.092389
#> 126    0.02521014     9.90  FALSE        40.52            100  2.973982
#> 11025 36.81998679   300.00  FALSE       300.00             50  2.828200
#> 13548 16.97845834   300.00  FALSE       300.00            100  2.256336
#> 28978  4.43984995   300.00  FALSE        10.00            100  2.209674
#> 46567 16.94809158   300.00  FALSE       300.00             78  2.785235
#> 47629 31.66336663   300.00  FALSE       300.00            100  3.627607
#> 40852 21.51800295   300.00  FALSE       300.00            100  2.360983